Grid with angles between points \((x, y)\) and the rotor's vertical axis

Given grids of vertical and horizontal distances from rotor axes at points \((x, y)\) inside the rotor circle, get_phi_grid calculates the associated radial angles, relative to the rotor vertical axis.

Returned grid represents the left-half of the rotor's circle.

Usage

get_phi_grid(x_grid, y_grid)

Arguments

x_grid

A 2D array, with horizontal distances \(x\) from the rotor's vertical axis, expressed as the proportion of rotor radius, for the left-half of the rotor circle area.

y_grid

A 2D array, with vertical distances (\(y\) from the rotor's horizontal axis, expressed as the proportion of rotor radius, for the left-half of the rotor circle area.

Value

A 2D array, giving a grid of angles (in radians) between points \((x, y)\) inside the rotor circle and the rotor center, for the left-half of the rotor circle area.

See also

Examples

 x_grid <- get_x_grid(yinc=0.05, xinc=0.05)
 y_grid <- get_y_grid(x_grid,yinc=0.05)
 get_phi_grid(x_grid,y_grid)
#>           [,1]       [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 3.141593         NA        NA        NA        NA        NA        NA
#>  [2,] 3.141593 3.08900959 3.0367157 2.9849908 2.9340964 2.8842689 2.8357138
#>  [3,] 3.141593 3.08609415 3.0309354 2.9764440 2.9229237 2.8706458 2.8198421
#>  [4,] 3.141593 3.08283683 3.0244839 2.9669205 2.9105020 2.8555412 2.8023000
#>  [5,] 3.141593 3.07917384 3.0172377 2.9562447 2.8966140 2.8387078 2.7828220
#>  [6,] 3.141593 3.07502449 3.0090411 2.9441971 2.8809903 2.8198421 2.7610863
#>  [7,] 3.141593 3.07028519 2.9996956 2.9304993 2.8632930 2.7985687 2.7367009
#>  [8,] 3.141593 3.06482076 2.9889433 2.9147938 2.8430937 2.7744188 2.7091849
#>  [9,] 3.141593 3.05845142 2.9764440 2.8966140 2.8198421 2.7468015 2.6779450
#> [10,] 3.141593 3.05093277 2.9617392 2.8753406 2.7928217 2.7149652 2.6422459
#> [11,] 3.141593 3.04192400 2.9441971 2.8501359 2.7610863 2.6779450 2.6011732
#> [12,] 3.141593 3.03093543 2.9229237 2.8198421 2.7233683 2.6344941 2.5535901
#> [13,] 3.141593 3.01723766 2.8966140 2.7828220 2.6779450 2.5829933 2.4980915
#> [14,] 3.141593 2.99969560 2.8632930 2.7367009 2.6224465 2.5213432 2.4329664
#> [15,] 3.141593 2.97644398 2.8198421 2.6779450 2.5535901 2.4468544 2.3561945
#> [16,] 3.141593 2.94419709 2.7610863 2.6011732 2.4668517 2.3561945 2.2655346
#> [17,] 3.141593 2.89661399 2.6779450 2.4980915 2.3561945 2.2455373 2.1587989
#> [18,] 3.141593 2.81984210 2.5535901 2.3561945 2.2142974 2.1112158 2.0344439
#> [19,] 3.141593 2.67794504 2.3561945 2.1587989 2.0344439 1.9513027 1.8925469
#> [20,] 3.141593 2.35619449 2.0344439 1.8925469 1.8157750 1.7681919 1.7359450
#> [21,]      NaN 1.57079633 1.5707963 1.5707963 1.5707963 1.5707963 1.5707963
#> [22,] 0.000000 0.78539816 1.1071487 1.2490458 1.3258177 1.3734008 1.4056476
#> [23,] 0.000000 0.46364761 0.7853982 0.9827937 1.1071487 1.1902899 1.2490458
#> [24,] 0.000000 0.32175055 0.5880026 0.7853982 0.9272952 1.0303768 1.1071487
#> [25,] 0.000000 0.24497866 0.4636476 0.6435011 0.7853982 0.8960554 0.9827937
#> [26,] 0.000000 0.19739556 0.3805064 0.5404195 0.6747409 0.7853982 0.8760581
#> [27,] 0.000000 0.16514868 0.3217506 0.4636476 0.5880026 0.6947383 0.7853982
#> [28,] 0.000000 0.14189705 0.2782997 0.4048918 0.5191461 0.6202495 0.7086263
#> [29,] 0.000000 0.12435499 0.2449787 0.3587707 0.4636476 0.5585993 0.6435011
#> [30,] 0.000000 0.11065722 0.2186689 0.3217506 0.4182243 0.5070985 0.5880026
#> [31,] 0.000000 0.09966865 0.1973956 0.2914568 0.3805064 0.4636476 0.5404195
#> [32,] 0.000000 0.09065989 0.1798535 0.2662520 0.3487710 0.4266275 0.4993467
#> [33,] 0.000000 0.08314123 0.1651487 0.2449787 0.3217506 0.3947911 0.4636476
#> [34,] 0.000000 0.07677189 0.1526493 0.2267988 0.2984989 0.3671738 0.4324078
#> [35,] 0.000000 0.07130746 0.1418971 0.2110933 0.2782997 0.3430239 0.4048918
#> [36,] 0.000000 0.06656816 0.1325515 0.1973956 0.2606024 0.3217506 0.3805064
#> [37,] 0.000000 0.06241881 0.1243550 0.1853479 0.2449787 0.3028849 0.3587707
#> [38,] 0.000000 0.05875582 0.1171087 0.1746722 0.2310907 0.2860514 0.3392926
#> [39,] 0.000000 0.05549851 0.1106572 0.1651487 0.2186689 0.2709469 0.3217506
#> [40,] 0.000000 0.05258306 0.1048769 0.1566019 0.2074962 0.2573237 0.3058789
#> [41,] 0.000000         NA        NA        NA        NA        NA        NA
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 2.8240322        NA        NA        NA        NA        NA        NA
#>  [3,] 2.7707014 2.7233683 2.6905658        NA        NA        NA        NA
#>  [4,] 2.7509856 2.7017501 2.6546934 2.6098686 2.5867816        NA        NA
#>  [5,] 2.7291822 2.6779450 2.6292032 2.5829933 2.5393053 2.4980915        NA
#>  [6,] 2.7049655 2.6516353 2.6011732 2.5535901 2.5088438 2.4668517 2.4275020
#>  [7,] 2.6779450 2.6224465 2.5702552 2.5213432 2.4756234 2.4329664 2.3932146
#>  [8,] 2.6476513 2.5899377 2.5360480 2.4858970 2.4393357 2.3961732 2.3561945
#>  [9,] 2.6135182 2.5535901 2.4980915 2.4468544 2.3996454 2.3561945 2.3162158
#> [10,] 2.5748634 2.5127964 2.4558631 2.4037776 2.3561945 2.3127436 2.2730533
#> [11,] 2.5308667 2.4668517 2.4087776 2.3561945 2.3086114 2.2655346 2.2264920
#> [12,] 2.4805495 2.4149503 2.3561945 2.3036114 2.2565258 2.2142974 2.1763410
#> [13,] 2.4227627 2.3561945 2.2974387 2.2455373 2.1995926 2.1587989 2.1224513
#> [14,] 2.3561945 2.2896263 2.2318395 2.1815223 2.1375255 2.0988708 2.0647377
#> [15,] 2.2794226 2.2142974 2.1587989 2.1112158 2.0701430 2.0344439 2.0032041
#> [16,] 2.1910458 2.1293956 2.0778948 2.0344439 1.9974238 1.9655874 1.9379702
#> [17,] 2.0899424 2.0344439 1.9890207 1.9513027 1.9195673 1.8925469 1.8692953
#> [18,] 1.9756881 1.9295670 1.8925469 1.8622531 1.8370484 1.8157750 1.7975952
#> [19,] 1.8490960 1.8157750 1.7894653 1.7681919 1.7506498 1.7359450 1.7234457
#> [20,] 1.7126934 1.6951513 1.6814535 1.6704650 1.6614562 1.6539376 1.6475682
#> [21,] 1.5707963 1.5707963 1.5707963 1.5707963 1.5707963 1.5707963 1.5707963
#> [22,] 1.4288993 1.4464413 1.4601391 1.4711277 1.4801364 1.4876551 1.4940244
#> [23,] 1.2924967 1.3258177 1.3521274 1.3734008 1.3909428 1.4056476 1.4181470
#> [24,] 1.1659045 1.2120257 1.2490458 1.2793395 1.3045443 1.3258177 1.3439975
#> [25,] 1.0516502 1.1071487 1.1525720 1.1902899 1.2220253 1.2490458 1.2722974
#> [26,] 0.9505468 1.0121970 1.0636978 1.1071487 1.1441688 1.1760052 1.2036225
#> [27,] 0.8621701 0.9272952 0.9827937 1.0303768 1.0714496 1.1071487 1.1383886
#> [28,] 0.7853982 0.8519663 0.9097532 0.9600704 1.0040671 1.0427219 1.0768550
#> [29,] 0.7188300 0.7853982 0.8441540 0.8960554 0.9420000 0.9827937 1.0191413
#> [30,] 0.6610432 0.7266423 0.7853982 0.8379812 0.8850668 0.9272952 0.9652517
#> [31,] 0.6107260 0.6747409 0.7328151 0.7853982 0.8329813 0.8760581 0.9151007
#> [32,] 0.5667292 0.6287963 0.6857295 0.7378151 0.7853982 0.8288491 0.8685394
#> [33,] 0.5280744 0.5880026 0.6435011 0.6947383 0.7419473 0.7853982 0.8253769
#> [34,] 0.4939414 0.5516550 0.6055447 0.6556956 0.7022569 0.7454195 0.7853982
#> [35,] 0.4636476 0.5191461 0.5713375 0.6202495 0.6659692 0.7086263 0.7483780
#> [36,] 0.4366272 0.4899573 0.5404195 0.5880026 0.6327488 0.6747409 0.7140907
#> [37,] 0.4124104 0.4636476 0.5123895 0.5585993 0.6022873 0.6435011        NA
#> [38,] 0.3906070 0.4398426 0.4868992 0.5317241 0.5548110        NA        NA
#> [39,] 0.3708913 0.4182243 0.4510268        NA        NA        NA        NA
#> [40,] 0.3175604        NA        NA        NA        NA        NA        NA
#> [41,]        NA        NA        NA        NA        NA        NA        NA
#>           [,15]     [,16]     [,17]     [,18]    [,19]    [,20]    [,21]
#>  [1,]        NA        NA        NA        NA       NA       NA       NA
#>  [2,]        NA        NA        NA        NA       NA       NA       NA
#>  [3,]        NA        NA        NA        NA       NA       NA       NA
#>  [4,]        NA        NA        NA        NA       NA       NA       NA
#>  [5,]        NA        NA        NA        NA       NA       NA       NA
#>  [6,] 2.4188584        NA        NA        NA       NA       NA       NA
#>  [7,] 2.3561945 2.3461938        NA        NA       NA       NA       NA
#>  [8,] 2.3191744 2.2848870 2.2783808        NA       NA       NA       NA
#>  [9,] 2.2794226 2.2455373 2.2142974 2.2142974       NA       NA       NA
#> [10,] 2.2367656 2.2035452 2.1730837 2.1531606       NA       NA       NA
#> [11,] 2.1910458 2.1587989 2.1293956 2.1025204 2.094395       NA       NA
#> [12,] 2.1421338 2.1112158 2.0831858 2.0576956 2.037562       NA       NA
#> [13,] 2.0899424 2.0607537 2.0344439 2.0106389 1.989021 1.982313       NA
#> [14,] 2.0344439 2.0074235 1.9832068 1.9614034 1.941688 1.928367       NA
#> [15,] 1.9756881 1.9513027 1.9295670 1.9100889 1.892547 1.876675 1.875489
#> [16,] 1.9138203 1.8925469 1.8736812 1.8568478 1.841743 1.828120 1.823477
#> [17,] 1.8490960 1.8313987 1.8157750 1.8018870 1.789465 1.778293 1.772154
#> [18,] 1.7818897 1.7681919 1.7561443 1.7454685 1.735945 1.727398 1.721365
#> [19,] 1.7126934 1.7033479 1.6951513 1.6879051 1.681454 1.675673 1.670964
#> [20,] 1.6421038 1.6373645 1.6332151 1.6295521 1.626295 1.623379 1.620817
#> [21,] 1.5707963 1.5707963 1.5707963 1.5707963 1.570796 1.570796 1.570796
#> [22,] 1.4994889 1.5042282 1.5083775 1.5120405 1.515298 1.518213 1.520775
#> [23,] 1.4288993 1.4382448 1.4464413 1.4536876 1.460139 1.465919 1.470629
#> [24,] 1.3597030 1.3734008 1.3854484 1.3961241 1.405648 1.414194 1.420228
#> [25,] 1.2924967 1.3101939 1.3258177 1.3397057 1.352127 1.363300 1.369438
#> [26,] 1.2277724 1.2490458 1.2679115 1.2847449 1.299849 1.313473 1.318116
#> [27,] 1.1659045 1.1902899 1.2120257 1.2315037 1.249046 1.264917 1.266104
#> [28,] 1.1071487 1.1341692 1.1583859 1.1801893 1.199905 1.213225       NA
#> [29,] 1.0516502 1.0808390 1.1071487 1.1309537 1.152572 1.159279       NA
#> [30,] 0.9994588 1.0303768 1.0584069 1.0838971 1.104031       NA       NA
#> [31,] 0.9505468 0.9827937 1.0121970 1.0390723 1.047198       NA       NA
#> [32,] 0.9048271 0.9380475 0.9685090 0.9884321       NA       NA       NA
#> [33,] 0.8621701 0.8960554 0.9272952        NA       NA       NA       NA
#> [34,] 0.8224183 0.8567056 0.8632119        NA       NA       NA       NA
#> [35,] 0.7853982 0.7953988        NA        NA       NA       NA       NA
#> [36,] 0.7227342        NA        NA        NA       NA       NA       NA
#> [37,]        NA        NA        NA        NA       NA       NA       NA
#> [38,]        NA        NA        NA        NA       NA       NA       NA
#> [39,]        NA        NA        NA        NA       NA       NA       NA
#> [40,]        NA        NA        NA        NA       NA       NA       NA
#> [41,]        NA        NA        NA        NA       NA       NA       NA