Grid of horizontal distances from points in the rotor circle to its vertical axis

Taking the center of the rotor circle as the origin, get_x_grid generates a grid containing horizontal distances \(x\) (by xinc increments) from the y-axis up to the outer edge of the rotor circle, at equidistant height bands \(y\) (by yinc increments) between minimum and maximum rotor height.

Distances are expressed as proportion of rotor radius (i.e. \(x\) is dimensionless).

Returned grid represents the left-half of the rotor's circle.

Usage

get_x_grid(xinc = 0.05, yinc = 0.05)

Arguments

xinc, yinc

numeric values, the grid increments along the y-axis and x-axis (i.e. grid cell dimensions)

Value

A 2D array giving a grid of horizontal distances \(x\) from rotor's vertical axis, expressed as the proportion of rotor radius (i.e. \([0, 1]\)), for the left-half of the rotor circle area.

Examples

get_x_grid(xinc=0.05,yinc=0.05)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7]      [,8] [,9]     [,10] [,11]
#>  [1,]    0   NA   NA   NA   NA   NA   NA        NA   NA        NA    NA
#>  [2,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3122499   NA        NA    NA
#>  [3,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4358899    NA
#>  [4,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#>  [5,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#>  [6,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#>  [7,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#>  [8,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#>  [9,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [10,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [11,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [12,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [13,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [14,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [15,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [16,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [17,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [18,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [19,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [20,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [21,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [22,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [23,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [24,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [25,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [26,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [27,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [28,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [29,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [30,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [31,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [32,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [33,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [34,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [35,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [36,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [37,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [38,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4500000   0.5
#> [39,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3500000  0.4 0.4358899    NA
#> [40,]    0 0.05  0.1 0.15  0.2 0.25  0.3 0.3122499   NA        NA    NA
#> [41,]    0   NA   NA   NA   NA   NA   NA        NA   NA        NA    NA
#>           [,12] [,13] [,14]     [,15]     [,16]     [,17]     [,18]     [,19]
#>  [1,]        NA    NA    NA        NA        NA        NA        NA        NA
#>  [2,]        NA    NA    NA        NA        NA        NA        NA        NA
#>  [3,]        NA    NA    NA        NA        NA        NA        NA        NA
#>  [4,] 0.5267827    NA    NA        NA        NA        NA        NA        NA
#>  [5,] 0.5500000   0.6    NA        NA        NA        NA        NA        NA
#>  [6,] 0.5500000   0.6  0.65 0.6614378        NA        NA        NA        NA
#>  [7,] 0.5500000   0.6  0.65 0.7000000 0.7141428        NA        NA        NA
#>  [8,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.7599342        NA        NA
#>  [9,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8000000        NA
#> [10,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8351647        NA
#> [11,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.8660254
#> [12,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.8930286
#> [13,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [14,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [15,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [16,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [17,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [18,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [19,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [20,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [21,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [22,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [23,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [24,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [25,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [26,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [27,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [28,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [29,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.9000000
#> [30,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.8930286
#> [31,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8500000 0.8660254
#> [32,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000 0.8351647        NA
#> [33,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.8000000        NA        NA
#> [34,] 0.5500000   0.6  0.65 0.7000000 0.7500000 0.7599342        NA        NA
#> [35,] 0.5500000   0.6  0.65 0.7000000 0.7141428        NA        NA        NA
#> [36,] 0.5500000   0.6  0.65 0.6614378        NA        NA        NA        NA
#> [37,] 0.5500000   0.6    NA        NA        NA        NA        NA        NA
#> [38,] 0.5267827    NA    NA        NA        NA        NA        NA        NA
#> [39,]        NA    NA    NA        NA        NA        NA        NA        NA
#> [40,]        NA    NA    NA        NA        NA        NA        NA        NA
#> [41,]        NA    NA    NA        NA        NA        NA        NA        NA
#>           [,20]     [,21]
#>  [1,]        NA        NA
#>  [2,]        NA        NA
#>  [3,]        NA        NA
#>  [4,]        NA        NA
#>  [5,]        NA        NA
#>  [6,]        NA        NA
#>  [7,]        NA        NA
#>  [8,]        NA        NA
#>  [9,]        NA        NA
#> [10,]        NA        NA
#> [11,]        NA        NA
#> [12,]        NA        NA
#> [13,] 0.9165151        NA
#> [14,] 0.9367497        NA
#> [15,] 0.9500000 0.9539392
#> [16,] 0.9500000 0.9682458
#> [17,] 0.9500000 0.9797959
#> [18,] 0.9500000 0.9886860
#> [19,] 0.9500000 0.9949874
#> [20,] 0.9500000 0.9987492
#> [21,] 0.9500000 1.0000000
#> [22,] 0.9500000 0.9987492
#> [23,] 0.9500000 0.9949874
#> [24,] 0.9500000 0.9886860
#> [25,] 0.9500000 0.9797959
#> [26,] 0.9500000 0.9682458
#> [27,] 0.9500000 0.9539392
#> [28,] 0.9367497        NA
#> [29,] 0.9165151        NA
#> [30,]        NA        NA
#> [31,]        NA        NA
#> [32,]        NA        NA
#> [33,]        NA        NA
#> [34,]        NA        NA
#> [35,]        NA        NA
#> [36,]        NA        NA
#> [37,]        NA        NA
#> [38,]        NA        NA
#> [39,]        NA        NA
#> [40,]        NA        NA
#> [41,]        NA        NA