Grid of vertical distances from points in the rotor circle to its horizontal axis

Taking the center of the rotor circle as the origin, get_y_grid generates a grid of vertical distances \(y\) (by yinc increments) from the x-axis to the outer edge of the rotor circle, across width intervals \(x\) (by xinc increments) between the center and maximum rotor width.

Returned grid represents the left-half of the rotor's circle.

Distances are expressed as proportion of rotor radius (i.e. \(y\) is dimensionless).

Usage

get_y_grid(x_grid, yinc = 0.05)

Arguments

x_grid

A 2D array, with horizontal distances from the rotor's vertical axis, expressed as the proportion of rotor radius, for the left-half of the rotor circle area.

yinc

a numeric value, the grid increment along the y-axis

Value

2D array giving a grid of vertical distances \(y\) from the rotor's horizontal axis, expressed as the proportion of rotor radius, for the left-half of the rotor circle area. Negative values represent distances from the bottom half of the rotor circle.

Examples

 x_grid <- get_x_grid(yinc=0.05, xinc=0.05)
 get_y_grid(x_grid,yinc=0.05)
#>        [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10] [,11] [,12]
#>  [1,] -1.00    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,] -0.95 -0.95 -0.95 -0.95 -0.95 -0.95 -0.95 -0.95    NA    NA    NA    NA
#>  [3,] -0.90 -0.90 -0.90 -0.90 -0.90 -0.90 -0.90 -0.90 -0.90 -0.90    NA    NA
#>  [4,] -0.85 -0.85 -0.85 -0.85 -0.85 -0.85 -0.85 -0.85 -0.85 -0.85 -0.85 -0.85
#>  [5,] -0.80 -0.80 -0.80 -0.80 -0.80 -0.80 -0.80 -0.80 -0.80 -0.80 -0.80 -0.80
#>  [6,] -0.75 -0.75 -0.75 -0.75 -0.75 -0.75 -0.75 -0.75 -0.75 -0.75 -0.75 -0.75
#>  [7,] -0.70 -0.70 -0.70 -0.70 -0.70 -0.70 -0.70 -0.70 -0.70 -0.70 -0.70 -0.70
#>  [8,] -0.65 -0.65 -0.65 -0.65 -0.65 -0.65 -0.65 -0.65 -0.65 -0.65 -0.65 -0.65
#>  [9,] -0.60 -0.60 -0.60 -0.60 -0.60 -0.60 -0.60 -0.60 -0.60 -0.60 -0.60 -0.60
#> [10,] -0.55 -0.55 -0.55 -0.55 -0.55 -0.55 -0.55 -0.55 -0.55 -0.55 -0.55 -0.55
#> [11,] -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50
#> [12,] -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45
#> [13,] -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40
#> [14,] -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35
#> [15,] -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30
#> [16,] -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25
#> [17,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#> [18,] -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15
#> [19,] -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10
#> [20,] -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05
#> [21,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> [22,]  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05
#> [23,]  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10
#> [24,]  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15
#> [25,]  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20
#> [26,]  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25
#> [27,]  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30
#> [28,]  0.35  0.35  0.35  0.35  0.35  0.35  0.35  0.35  0.35  0.35  0.35  0.35
#> [29,]  0.40  0.40  0.40  0.40  0.40  0.40  0.40  0.40  0.40  0.40  0.40  0.40
#> [30,]  0.45  0.45  0.45  0.45  0.45  0.45  0.45  0.45  0.45  0.45  0.45  0.45
#> [31,]  0.50  0.50  0.50  0.50  0.50  0.50  0.50  0.50  0.50  0.50  0.50  0.50
#> [32,]  0.55  0.55  0.55  0.55  0.55  0.55  0.55  0.55  0.55  0.55  0.55  0.55
#> [33,]  0.60  0.60  0.60  0.60  0.60  0.60  0.60  0.60  0.60  0.60  0.60  0.60
#> [34,]  0.65  0.65  0.65  0.65  0.65  0.65  0.65  0.65  0.65  0.65  0.65  0.65
#> [35,]  0.70  0.70  0.70  0.70  0.70  0.70  0.70  0.70  0.70  0.70  0.70  0.70
#> [36,]  0.75  0.75  0.75  0.75  0.75  0.75  0.75  0.75  0.75  0.75  0.75  0.75
#> [37,]  0.80  0.80  0.80  0.80  0.80  0.80  0.80  0.80  0.80  0.80  0.80  0.80
#> [38,]  0.85  0.85  0.85  0.85  0.85  0.85  0.85  0.85  0.85  0.85  0.85  0.85
#> [39,]  0.90  0.90  0.90  0.90  0.90  0.90  0.90  0.90  0.90  0.90    NA    NA
#> [40,]  0.95  0.95  0.95  0.95  0.95  0.95  0.95  0.95    NA    NA    NA    NA
#> [41,]  1.00    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,] -0.80    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,] -0.75 -0.75 -0.75    NA    NA    NA    NA    NA    NA
#>  [7,] -0.70 -0.70 -0.70 -0.70    NA    NA    NA    NA    NA
#>  [8,] -0.65 -0.65 -0.65 -0.65 -0.65    NA    NA    NA    NA
#>  [9,] -0.60 -0.60 -0.60 -0.60 -0.60 -0.60    NA    NA    NA
#> [10,] -0.55 -0.55 -0.55 -0.55 -0.55 -0.55    NA    NA    NA
#> [11,] -0.50 -0.50 -0.50 -0.50 -0.50 -0.50 -0.50    NA    NA
#> [12,] -0.45 -0.45 -0.45 -0.45 -0.45 -0.45 -0.45    NA    NA
#> [13,] -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40 -0.40    NA
#> [14,] -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35    NA
#> [15,] -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30 -0.30
#> [16,] -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25 -0.25
#> [17,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#> [18,] -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15 -0.15
#> [19,] -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10 -0.10
#> [20,] -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05 -0.05
#> [21,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> [22,]  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05  0.05
#> [23,]  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10  0.10
#> [24,]  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15  0.15
#> [25,]  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20  0.20
#> [26,]  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25  0.25
#> [27,]  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30  0.30
#> [28,]  0.35  0.35  0.35  0.35  0.35  0.35  0.35  0.35    NA
#> [29,]  0.40  0.40  0.40  0.40  0.40  0.40  0.40  0.40    NA
#> [30,]  0.45  0.45  0.45  0.45  0.45  0.45  0.45    NA    NA
#> [31,]  0.50  0.50  0.50  0.50  0.50  0.50  0.50    NA    NA
#> [32,]  0.55  0.55  0.55  0.55  0.55  0.55    NA    NA    NA
#> [33,]  0.60  0.60  0.60  0.60  0.60    NA    NA    NA    NA
#> [34,]  0.65  0.65  0.65  0.65  0.65    NA    NA    NA    NA
#> [35,]  0.70  0.70  0.70  0.70    NA    NA    NA    NA    NA
#> [36,]  0.75  0.75  0.75    NA    NA    NA    NA    NA    NA
#> [37,]  0.80    NA    NA    NA    NA    NA    NA    NA    NA
#> [38,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [39,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [40,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [41,]    NA    NA    NA    NA    NA    NA    NA    NA    NA