Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 88.96718
#> 2   Feb 92.40276
#> 3   Mar 97.61841
#> 4   Apr 85.22997
#> 5   May 95.25568
#> 6   Jun 86.12851

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 6.520788  2
#> 2   Feb 9.509807  2
#> 3   Mar 6.783755  2
#> 4   Apr 9.292077  2
#> 5   May 6.862559  2
#> 6   Jun 8.088855  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [11ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [21ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [64ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [695ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [18ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1985508    0.1978378  0.1947346
#>    [2,]    0.2053335    0.2045962  0.2013869
#>    [3,]    0.2286637    0.2278426  0.2242687
#>    [4,]    0.1923027    0.1916122  0.1886066
#>    [5,]    0.2100510    0.2092967  0.2060137
#>    [6,]    0.2755150    0.2745257  0.2702197
#>    [7,]    0.2280643    0.2272454  0.2236809
#>    [8,]    0.1772574    0.1766209  0.1738504
#>    [9,]    0.2496844    0.2487878  0.2448855
#>   [10,]    0.3861305    0.3847440  0.3787094
#>   [11,]    0.3070199    0.3059174  0.3011191
#>   [12,]    0.2401927    0.2393302  0.2355762
#>   [13,]    0.1975981    0.1968885  0.1938002
#>   [14,]    0.1841368    0.1834756  0.1805976
#>   [15,]    0.2699139    0.2689447  0.2647262
#>   [16,]    0.2205065    0.2197147  0.2162683
#>   [17,]    0.2496128    0.2487165  0.2448153
#>   [18,]    0.2372296    0.2363777  0.2326700
#>   [19,]    0.1483622    0.1478295  0.1455106
#>   [20,]    0.2541283    0.2532157  0.2492439
#>   [21,]    0.1814741    0.1808224  0.1779861
#>   [22,]    0.2499348    0.2490373  0.2451310
#>   [23,]    0.2139629    0.2131946  0.2098505
#>   [24,]    0.2071170    0.2063733  0.2031361
#>   [25,]    0.2291334    0.2283106  0.2247294
#>   [26,]    0.1903040    0.1896206  0.1866463
#>   [27,]    0.2762870    0.2752949  0.2709768
#>   [28,]    0.2419087    0.2410400  0.2372592
#>   [29,]    0.2054612    0.2047234  0.2015122
#>   [30,]    0.1733158    0.1726935  0.1699847
#>   [31,]    0.2595555    0.2586235  0.2545669
#>   [32,]    0.2457567    0.2448743  0.2410332
#>   [33,]    0.2497232    0.2488265  0.2449235
#>   [34,]    0.3152511    0.3141191  0.3091921
#>   [35,]    0.2078944    0.2071479  0.2038986
#>   [36,]    0.2215371    0.2207416  0.2172791
#>   [37,]    0.2246993    0.2238924  0.2203805
#>   [38,]    0.1517551    0.1512101  0.1488383
#>   [39,]    0.2339594    0.2331193  0.2294627
#>   [40,]    0.3271988    0.3260240  0.3209103
#>   [41,]    0.2277177    0.2269000  0.2233409
#>   [42,]    0.2025193    0.2017921  0.1986268
#>   [43,]    0.3565627    0.3552824  0.3497098
#>   [44,]    0.1893931    0.1887130  0.1857529
#>   [45,]    0.2554381    0.2545209  0.2505286
#>   [46,]    0.1496542    0.1491168  0.1467778
#>   [47,]    0.2960017    0.2949388  0.2903126
#>   [48,]    0.1779895    0.1773504  0.1745685
#>   [49,]    0.2194090    0.2186212  0.2151920
#>   [50,]    0.1965872    0.1958813  0.1928087
#>   [51,]    0.2050683    0.2043320  0.2011268
#>   [52,]    0.2234035    0.2226013  0.2191096
#>   [53,]    0.2266733    0.2258594  0.2223167
#>   [54,]    0.2644331    0.2634836  0.2593507
#>   [55,]    0.2690642    0.2680981  0.2638928
#>   [56,]    0.2957348    0.2946729  0.2900509
#>   [57,]    0.2631739    0.2622289  0.2581158
#>   [58,]    0.1857247    0.1850578  0.1821550
#>   [59,]    0.2416216    0.2407540  0.2369776
#>   [60,]    0.1739244    0.1732998  0.1705815
#>   [61,]    0.2626064    0.2616634  0.2575591
#>   [62,]    0.2045760    0.2038414  0.2006440
#>   [63,]    0.1708642    0.1702506  0.1675801
#>   [64,]    0.2523026    0.2513966  0.2474533
#>   [65,]    0.1974382    0.1967292  0.1936434
#>   [66,]    0.2438060    0.2429305  0.2391200
#>   [67,]    0.2354742    0.2346286  0.2309483
#>   [68,]    0.2361647    0.2353167  0.2316256
#>   [69,]    0.1983765    0.1976642  0.1945636
#>   [70,]    0.2568263    0.2559041  0.2518901
#>   [71,]    0.1882388    0.1875629  0.1846208
#>   [72,]    0.1887460    0.1880682  0.1851182
#>   [73,]    0.2347909    0.2339478  0.2302782
#>   [74,]    0.2346334    0.2337909  0.2301237
#>   [75,]    0.2766775    0.2756840  0.2713597
#>   [76,]    0.1966550    0.1959489  0.1928753
#>   [77,]    0.2536029    0.2526923  0.2487287
#>   [78,]    0.1885935    0.1879163  0.1849687
#>   [79,]    0.2168402    0.2160615  0.2126725
#>   [80,]    0.2000800    0.1993616  0.1962344
#>   [81,]    0.1915346    0.1908468  0.1878532
#>   [82,]    0.2871214    0.2860905  0.2816030
#>   [83,]    0.3164771    0.3153407  0.3103945
#>   [84,]    0.1883403    0.1876640  0.1847203
#>   [85,]    0.1898810    0.1891991  0.1862314
#>   [86,]    0.1794795    0.1788350  0.1760298
#>   [87,]    0.2449564    0.2440769  0.2402484
#>   [88,]    0.1940300    0.1933332  0.1903007
#>   [89,]    0.3016605    0.3005773  0.2958626
#>   [90,]    0.1428538    0.1423408  0.1401080
#>   [91,]    0.3134800    0.3123543  0.3074550
#>   [92,]    0.1564854    0.1559235  0.1534777
#>   [93,]    0.2294507    0.2286268  0.2250406
#>   [94,]    0.2464144    0.2455296  0.2416784
#>   [95,]    0.2491477    0.2482531  0.2443591
#>   [96,]    0.2264829    0.2256696  0.2221299
#>   [97,]    0.1998232    0.1991056  0.1959825
#>   [98,]    0.2474726    0.2465840  0.2427162
#>   [99,]    0.2061283    0.2053882  0.2021665
#>  [100,]    0.2201516    0.2193610  0.2159202
#>  [101,]    0.2333727    0.2325347  0.2288872
#>  [102,]    0.1917036    0.1910152  0.1880190
#>  [103,]    0.2137669    0.2129993  0.2096584
#>  [104,]    0.1292695    0.1288053  0.1267849
#>  [105,]    0.2640932    0.2631449  0.2590173
#>  [106,]    0.2221428    0.2213452  0.2178732
#>  [107,]    0.1889588    0.1882803  0.1853270
#>  [108,]    0.1155202    0.1151054  0.1132998
#>  [109,]    0.2612082    0.2602703  0.2561878
#>  [110,]    0.2244517    0.2236457  0.2201377
#>  [111,]    0.2944206    0.2933634  0.2887619
#>  [112,]    0.3548204    0.3535464  0.3480011
#>  [113,]    0.2138696    0.2131017  0.2097590
#>  [114,]    0.2134968    0.2127302  0.2093934
#>  [115,]    0.2005927    0.1998725  0.1967373
#>  [116,]    0.1437555    0.1432393  0.1409925
#>  [117,]    0.1936246    0.1929293  0.1899031
#>  [118,]    0.1709185    0.1703047  0.1676333
#>  [119,]    0.2000162    0.1992980  0.1961719
#>  [120,]    0.1827211    0.1820650  0.1792091
#>  [121,]    0.2184108    0.2176266  0.2142129
#>  [122,]    0.1782620    0.1776219  0.1748358
#>  [123,]    0.1591929    0.1586213  0.1561331
#>  [124,]    0.2621023    0.2611612  0.2570648
#>  [125,]    0.3137958    0.3126690  0.3077647
#>  [126,]    0.2628915    0.2619475  0.2578387
#>  [127,]    0.2218076    0.2210112  0.2175445
#>  [128,]    0.2760885    0.2750971  0.2707821
#>  [129,]    0.5101034    0.5082719  0.5002998
#>  [130,]    0.2262060    0.2253938  0.2218583
#>  [131,]    0.2188105    0.2180248  0.2146050
#>  [132,]    0.2172449    0.2164648  0.2130694
#>  [133,]    0.1825622    0.1819067  0.1790533
#>  [134,]    0.1554653    0.1549071  0.1524772
#>  [135,]    0.2367314    0.2358813  0.2321813
#>  [136,]    0.2209410    0.2201477  0.2166945
#>  [137,]    0.2672049    0.2662454  0.2620692
#>  [138,]    0.2202804    0.2194894  0.2160465
#>  [139,]    0.1740104    0.1733856  0.1706659
#>  [140,]    0.1601375    0.1595624  0.1570596
#>  [141,]    0.2197802    0.2189910  0.2155560
#>  [142,]    0.1538205    0.1532681  0.1508640
#>  [143,]    0.1508635    0.1503218  0.1479639
#>  [144,]    0.1631432    0.1625574  0.1600075
#>  [145,]    0.3014932    0.3004106  0.2956986
#>  [146,]    0.2528129    0.2519051  0.2479539
#>  [147,]    0.2468551    0.2459687  0.2421106
#>  [148,]    0.1836273    0.1829679  0.1800979
#>  [149,]    0.2012547    0.2005320  0.1973865
#>  [150,]    0.2912604    0.2902145  0.2856624
#>  [151,]    0.2292848    0.2284615  0.2248779
#>  [152,]    0.1796206    0.1789756  0.1761682
#>  [153,]    0.2103917    0.2096362  0.2063479
#>  [154,]    0.2875278    0.2864953  0.2820015
#>  [155,]    0.2120030    0.2112417  0.2079282
#>  [156,]    0.3417289    0.3405019  0.3351611
#>  [157,]    0.1428463    0.1423333  0.1401007
#>  [158,]    0.2140204    0.2132519  0.2099069
#>  [159,]    0.2002731    0.1995540  0.1964238
#>  [160,]    0.2484039    0.2475119  0.2436295
#>  [161,]    0.1582279    0.1576597  0.1551866
#>  [162,]    0.2581114    0.2571846  0.2531505
#>  [163,]    0.3223091    0.3211518  0.3161145
#>  [164,]    0.2082369    0.2074892  0.2042345
#>  [165,]    0.1873672    0.1866944  0.1837659
#>  [166,]    0.3166486    0.3155116  0.3105627
#>  [167,]    0.2424747    0.2416040  0.2378143
#>  [168,]    0.2210916    0.2202977  0.2168421
#>  [169,]    0.2452874    0.2444066  0.2405730
#>  [170,]    0.2011813    0.2004589  0.1973145
#>  [171,]    0.2873661    0.2863343  0.2818429
#>  [172,]    0.1871372    0.1864652  0.1835404
#>  [173,]    0.2341167    0.2332760  0.2296169
#>  [174,]    0.2373672    0.2365149  0.2328050
#>  [175,]    0.2457404    0.2448580  0.2410173
#>  [176,]    0.2733948    0.2724131  0.2681402
#>  [177,]    0.2349285    0.2340849  0.2304131
#>  [178,]    0.3024643    0.3013782  0.2966510
#>  [179,]    0.2083917    0.2076434  0.2043864
#>  [180,]    0.1990650    0.1983502  0.1952389
#>  [181,]    0.1929446    0.1922517  0.1892361
#>  [182,]    0.1993298    0.1986140  0.1954986
#>  [183,]    0.2012862    0.2005634  0.1974174
#>  [184,]    0.1394507    0.1389499  0.1367703
#>  [185,]    0.2229194    0.2221190  0.2186349
#>  [186,]    0.2670463    0.2660874  0.2619137
#>  [187,]    0.1881668    0.1874911  0.1845502
#>  [188,]    0.2176458    0.2168642  0.2134626
#>  [189,]    0.2439265    0.2430506  0.2392383
#>  [190,]    0.2153788    0.2146054  0.2112392
#>  [191,]    0.1860901    0.1854219  0.1825134
#>  [192,]    0.2104384    0.2096827  0.2063937
#>  [193,]    0.1821087    0.1814548  0.1786085
#>  [194,]    0.1927895    0.1920972  0.1890840
#>  [195,]    0.2107481    0.2099914  0.2066975
#>  [196,]    0.2288941    0.2280722  0.2244947
#>  [197,]    0.2068433    0.2061006  0.2028677
#>  [198,]    0.2834930    0.2824750  0.2780442
#>  [199,]    0.2133667    0.2126005  0.2092657
#>  [200,]    0.2111235    0.2103654  0.2070656
#>  [201,]    0.2122214    0.2114594  0.2081425
#>  [202,]    0.1760426    0.1754104  0.1726590
#>  [203,]    0.2697163    0.2687478  0.2645323
#>  [204,]    0.1628016    0.1622170  0.1596725
#>  [205,]    0.2691986    0.2682319  0.2640245
#>  [206,]    0.2659341    0.2649792  0.2608229
#>  [207,]    0.2649532    0.2640018  0.2598608
#>  [208,]    0.2865919    0.2855628  0.2810836
#>  [209,]    0.2055085    0.2047706  0.2015586
#>  [210,]    0.2196200    0.2188313  0.2153988
#>  [211,]    0.2182825    0.2174987  0.2140870
#>  [212,]    0.2181172    0.2173340  0.2139249
#>  [213,]    0.1736525    0.1730289  0.1703148
#>  [214,]    0.1839892    0.1833285  0.1804528
#>  [215,]    0.2394536    0.2385938  0.2348512
#>  [216,]    0.2339475    0.2331074  0.2294510
#>  [217,]    0.2110897    0.2103317  0.2070325
#>  [218,]    0.2343682    0.2335266  0.2298636
#>  [219,]    0.2035474    0.2028165  0.1996351
#>  [220,]    0.2426558    0.2417844  0.2379919
#>  [221,]    0.2081390    0.2073916  0.2041385
#>  [222,]    0.2198834    0.2190938  0.2156572
#>  [223,]    0.2929321    0.2918802  0.2873019
#>  [224,]    0.2115865    0.2108267  0.2075198
#>  [225,]    0.2138214    0.2130536  0.2097117
#>  [226,]    0.2104486    0.2096930  0.2064038
#>  [227,]    0.2087788    0.2080291  0.2047660
#>  [228,]    0.2550540    0.2541382  0.2501519
#>  [229,]    0.2203827    0.2195913  0.2161469
#>  [230,]    0.1780174    0.1773781  0.1745958
#>  [231,]    0.2731221    0.2721414  0.2678728
#>  [232,]    0.2312649    0.2304345  0.2268201
#>  [233,]    0.2021742    0.2014482  0.1982883
#>  [234,]    0.2756767    0.2746868  0.2703782
#>  [235,]    0.2689195    0.2679538  0.2637508
#>  [236,]    0.2562314    0.2553113  0.2513066
#>  [237,]    0.3475278    0.3462799  0.3408484
#>  [238,]    0.2693619    0.2683947  0.2641847
#>  [239,]    0.3096263    0.3085145  0.3036753
#>  [240,]    0.3013583    0.3002762  0.2955663
#>  [241,]    0.2119343    0.2111733  0.2078609
#>  [242,]    0.1888091    0.1881311  0.1851801
#>  [243,]    0.2756078    0.2746182  0.2703107
#>  [244,]    0.2869369    0.2859066  0.2814221
#>  [245,]    0.1667472    0.1661484  0.1635422
#>  [246,]    0.3021760    0.3010910  0.2963682
#>  [247,]    0.2793589    0.2783558  0.2739897
#>  [248,]    0.1957249    0.1950221  0.1919630
#>  [249,]    0.2098685    0.2091149  0.2058347
#>  [250,]    0.2254494    0.2246399  0.2211162
#>  [251,]    0.2416403    0.2407726  0.2369960
#>  [252,]    0.1721283    0.1715102  0.1688199
#>  [253,]    0.1989414    0.1982270  0.1951177
#>  [254,]    0.2508742    0.2499734  0.2460524
#>  [255,]    0.2722980    0.2713203  0.2670646
#>  [256,]    0.1749794    0.1743510  0.1716162
#>  [257,]    0.2113947    0.2106356  0.2073317
#>  [258,]    0.1969132    0.1962062  0.1931285
#>  [259,]    0.2549688    0.2540532  0.2500683
#>  [260,]    0.2271615    0.2263458  0.2227954
#>  [261,]    0.2344079    0.2335662  0.2299025
#>  [262,]    0.3090034    0.3078938  0.3030644
#>  [263,]    0.3051932    0.3040973  0.2993275
#>  [264,]    0.1699381    0.1693279  0.1666718
#>  [265,]    0.1510378    0.1504955  0.1481348
#>  [266,]    0.2175609    0.2167797  0.2133793
#>  [267,]    0.1683140    0.1677096  0.1650789
#>  [268,]    0.2334884    0.2326500  0.2290007
#>  [269,]    0.2720299    0.2710531  0.2668015
#>  [270,]    0.1785512    0.1779100  0.1751194
#>  [271,]    0.1978063    0.1970961  0.1940044
#>  [272,]    0.3013920    0.3003097  0.2955992
#>  [273,]    0.1759389    0.1753071  0.1725573
#>  [274,]    0.1882478    0.1875718  0.1846296
#>  [275,]    0.2501922    0.2492938  0.2453835
#>  [276,]    0.2811188    0.2801094  0.2757158
#>  [277,]    0.2977503    0.2966812  0.2920277
#>  [278,]    0.1856878    0.1850210  0.1821188
#>  [279,]    0.1941990    0.1935016  0.1904664
#>  [280,]    0.2823295    0.2813157  0.2769032
#>  [281,]    0.1858924    0.1852249  0.1823195
#>  [282,]    0.1897403    0.1890590  0.1860934
#>  [283,]    0.1950795    0.1943790  0.1913300
#>  [284,]    0.2640103    0.2630623  0.2589360
#>  [285,]    0.2889145    0.2878771  0.2833616
#>  [286,]    0.1768245    0.1761896  0.1734259
#>  [287,]    0.2107798    0.2100229  0.2067286
#>  [288,]    0.2306054    0.2297774  0.2261732
#>  [289,]    0.2411608    0.2402948  0.2365257
#>  [290,]    0.2448815    0.2440022  0.2401749
#>  [291,]    0.2008430    0.2001218  0.1969827
#>  [292,]    0.2197516    0.2189625  0.2155279
#>  [293,]    0.2148727    0.2141012  0.2107429
#>  [294,]    0.2183551    0.2175710  0.2141583
#>  [295,]    0.2074889    0.2067438  0.2035009
#>  [296,]    0.2067409    0.2059985  0.2027672
#>  [297,]    0.2688400    0.2678746  0.2636729
#>  [298,]    0.2113279    0.2105691  0.2072662
#>  [299,]    0.2185880    0.2178031  0.2143867
#>  [300,]    0.2463141    0.2454296  0.2415799
#>  [301,]    0.3168948    0.3157569  0.3108042
#>  [302,]    0.1467957    0.1462686  0.1439742
#>  [303,]    0.2782317    0.2772326  0.2728841
#>  [304,]    0.2289096    0.2280876  0.2245099
#>  [305,]    0.1948321    0.1941325  0.1910873
#>  [306,]    0.2029526    0.2022239  0.1990518
#>  [307,]    0.2908144    0.2897702  0.2852250
#>  [308,]    0.2634888    0.2625427  0.2584246
#>  [309,]    0.2405773    0.2397134  0.2359534
#>  [310,]    0.1968750    0.1961680  0.1930910
#>  [311,]    0.2992922    0.2982175  0.2935399
#>  [312,]    0.1953533    0.1946518  0.1915986
#>  [313,]    0.2521849    0.2512793  0.2473379
#>  [314,]    0.2069214    0.2061784  0.2029443
#>  [315,]    0.1658502    0.1652546  0.1626625
#>  [316,]    0.1859226    0.1852550  0.1823491
#>  [317,]    0.2069908    0.2062475  0.2030123
#>  [318,]    0.1691016    0.1684943  0.1658514
#>  [319,]    0.1223907    0.1219512  0.1200382
#>  [320,]    0.2120519    0.2112904  0.2079762
#>  [321,]    0.2035508    0.2028199  0.1996385
#>  [322,]    0.2674260    0.2664658  0.2622861
#>  [323,]    0.2338470    0.2330073  0.2293524
#>  [324,]    0.2110867    0.2103287  0.2070296
#>  [325,]    0.2218066    0.2210101  0.2175435
#>  [326,]    0.1937303    0.1930346  0.1900067
#>  [327,]    0.1666362    0.1660379  0.1634334
#>  [328,]    0.2339714    0.2331313  0.2294745
#>  [329,]    0.2295128    0.2286887  0.2251015
#>  [330,]    0.1767565    0.1761217  0.1733591
#>  [331,]    0.1699342    0.1693240  0.1666680
#>  [332,]    0.2233016    0.2224998  0.2190098
#>  [333,]    0.3034767    0.3023869  0.2976439
#>  [334,]    0.2342448    0.2334036  0.2297425
#>  [335,]    0.2063861    0.2056451  0.2024194
#>  [336,]    0.2552546    0.2543380  0.2503486
#>  [337,]    0.1530215    0.1524720  0.1500804
#>  [338,]    0.1911087    0.1904224  0.1874355
#>  [339,]    0.1967931    0.1960865  0.1930107
#>  [340,]    0.2179436    0.2171610  0.2137546
#>  [341,]    0.1655836    0.1649890  0.1624010
#>  [342,]    0.2254398    0.2246303  0.2211069
#>  [343,]    0.2532168    0.2523075  0.2483500
#>  [344,]    0.2591155    0.2581850  0.2541353
#>  [345,]    0.2365971    0.2357475  0.2320496
#>  [346,]    0.2347305    0.2338876  0.2302189
#>  [347,]    0.2012760    0.2005533  0.1974074
#>  [348,]    0.2060459    0.2053060  0.2020856
#>  [349,]    0.2493069    0.2484117  0.2445153
#>  [350,]    0.2407235    0.2398591  0.2360968
#>  [351,]    0.1736492    0.1730256  0.1703115
#>  [352,]    0.1975892    0.1968797  0.1937915
#>  [353,]    0.2381328    0.2372777  0.2335558
#>  [354,]    0.2910698    0.2900246  0.2854754
#>  [355,]    0.2505071    0.2496076  0.2456923
#>  [356,]    0.1744943    0.1738677  0.1711405
#>  [357,]    0.2178400    0.2170578  0.2136531
#>  [358,]    0.2035450    0.2028141  0.1996328
#>  [359,]    0.1743201    0.1736941  0.1709695
#>  [360,]    0.1431312    0.1426172  0.1403801
#>  [361,]    0.2375298    0.2366769  0.2329645
#>  [362,]    0.2473197    0.2464316  0.2425662
#>  [363,]    0.2487434    0.2478502  0.2439625
#>  [364,]    0.2161408    0.2153647  0.2119865
#>  [365,]    0.2154239    0.2146504  0.2112834
#>  [366,]    0.2622214    0.2612798  0.2571815
#>  [367,]    0.1471125    0.1465843  0.1442850
#>  [368,]    0.2226774    0.2218778  0.2183975
#>  [369,]    0.1603422    0.1597664  0.1572604
#>  [370,]    0.2175929    0.2168116  0.2134107
#>  [371,]    0.2845089    0.2834873  0.2790408
#>  [372,]    0.1343223    0.1338399  0.1317405
#>  [373,]    0.2646190    0.2636688  0.2595331
#>  [374,]    0.1998180    0.1991005  0.1959774
#>  [375,]    0.2105198    0.2097638  0.2064735
#>  [376,]    0.2477283    0.2468388  0.2429670
#>  [377,]    0.2596913    0.2587588  0.2547001
#>  [378,]    0.1575060    0.1569405  0.1544787
#>  [379,]    0.2255867    0.2247766  0.2212509
#>  [380,]    0.2285529    0.2277322  0.2241601
#>  [381,]    0.2738364    0.2728531  0.2685733
#>  [382,]    0.1499528    0.1494143  0.1470706
#>  [383,]    0.2097039    0.2089509  0.2056734
#>  [384,]    0.2124332    0.2116704  0.2083503
#>  [385,]    0.3250832    0.3239159  0.3188352
#>  [386,]    0.1589338    0.1583631  0.1558790
#>  [387,]    0.3361022    0.3348954  0.3296426
#>  [388,]    0.2139310    0.2131628  0.2098192
#>  [389,]    0.2178172    0.2170351  0.2136307
#>  [390,]    0.3093260    0.3082153  0.3033807
#>  [391,]    0.2187484    0.2179630  0.2145441
#>  [392,]    0.2952655    0.2942053  0.2895905
#>  [393,]    0.2090662    0.2083155  0.2050479
#>  [394,]    0.2268596    0.2260450  0.2224993
#>  [395,]    0.1764526    0.1758190  0.1730611
#>  [396,]    0.3081195    0.3070132  0.3021976
#>  [397,]    0.2026053    0.2018778  0.1987112
#>  [398,]    0.1926971    0.1920052  0.1889934
#>  [399,]    0.2117056    0.2109454  0.2076366
#>  [400,]    0.1822053    0.1815510  0.1787032
#>  [401,]    0.2053959    0.2046584  0.2014482
#>  [402,]    0.1942639    0.1935664  0.1905302
#>  [403,]    0.1804888    0.1798407  0.1770198
#>  [404,]    0.2799974    0.2789920  0.2746159
#>  [405,]    0.2010092    0.2002874  0.1971457
#>  [406,]    0.1730243    0.1724030  0.1696987
#>  [407,]    0.2223723    0.2215738  0.2180983
#>  [408,]    0.1898404    0.1891587  0.1861915
#>  [409,]    0.2095946    0.2088420  0.2055661
#>  [410,]    0.1731115    0.1724898  0.1697842
#>  [411,]    0.1922385    0.1915482  0.1885436
#>  [412,]    0.2118478    0.2110871  0.2077761
#>  [413,]    0.2234322    0.2226299  0.2191378
#>  [414,]    0.1820746    0.1814208  0.1785751
#>  [415,]    0.3206330    0.3194817  0.3144705
#>  [416,]    0.1464495    0.1459236  0.1436347
#>  [417,]    0.1651070    0.1645141  0.1619335
#>  [418,]    0.1905607    0.1898764  0.1868981
#>  [419,]    0.2043822    0.2036483  0.2004540
#>  [420,]    0.1763380    0.1757048  0.1729487
#>  [421,]    0.2306829    0.2298545  0.2262491
#>  [422,]    0.2507815    0.2498810  0.2459615
#>  [423,]    0.2481983    0.2473071  0.2434279
#>  [424,]    0.2077416    0.2069957  0.2037488
#>  [425,]    0.2252886    0.2244796  0.2209585
#>  [426,]    0.2256603    0.2248500  0.2213231
#>  [427,]    0.1813353    0.1806841  0.1778500
#>  [428,]    0.1590990    0.1585277  0.1560411
#>  [429,]    0.2646555    0.2637052  0.2595689
#>  [430,]    0.1568097    0.1562466  0.1537958
#>  [431,]    0.3131975    0.3120729  0.3071780
#>  [432,]    0.2205402    0.2197483  0.2163014
#>  [433,]    0.3001705    0.2990927  0.2944013
#>  [434,]    0.2687066    0.2677417  0.2635421
#>  [435,]    0.2969344    0.2958682  0.2912274
#>  [436,]    0.2023566    0.2016300  0.1984673
#>  [437,]    0.2785320    0.2775319  0.2731787
#>  [438,]    0.2194380    0.2186500  0.2152203
#>  [439,]    0.2168848    0.2161060  0.2127162
#>  [440,]    0.2103221    0.2095669  0.2062797
#>  [441,]    0.2191376    0.2183508  0.2149258
#>  [442,]    0.2170757    0.2162962  0.2129035
#>  [443,]    0.2891070    0.2880689  0.2835504
#>  [444,]    0.1749007    0.1742727  0.1715390
#>  [445,]    0.2161600    0.2153838  0.2120053
#>  [446,]    0.2188621    0.2180762  0.2146555
#>  [447,]    0.2016592    0.2009351  0.1977833
#>  [448,]    0.3194654    0.3183182  0.3133253
#>  [449,]    0.3306985    0.3295111  0.3243427
#>  [450,]    0.1788586    0.1782164  0.1754209
#>  [451,]    0.1865766    0.1859067  0.1829905
#>  [452,]    0.1809398    0.1802901  0.1774620
#>  [453,]    0.1920383    0.1913487  0.1883472
#>  [454,]    0.1763534    0.1757201  0.1729638
#>  [455,]    0.2158045    0.2150296  0.2116567
#>  [456,]    0.2130290    0.2122641  0.2089346
#>  [457,]    0.1683727    0.1677681  0.1651365
#>  [458,]    0.2207395    0.2199469  0.2164969
#>  [459,]    0.2519364    0.2510318  0.2470942
#>  [460,]    0.2404442    0.2395808  0.2358228
#>  [461,]    0.2216160    0.2208203  0.2173565
#>  [462,]    0.1623545    0.1617715  0.1592340
#>  [463,]    0.1694756    0.1688670  0.1662182
#>  [464,]    0.2070100    0.2062667  0.2030313
#>  [465,]    0.2088063    0.2080565  0.2047930
#>  [466,]    0.2641980    0.2632493  0.2591201
#>  [467,]    0.2100216    0.2092674  0.2059849
#>  [468,]    0.1869612    0.1862899  0.1833678
#>  [469,]    0.1889188    0.1882404  0.1852877
#>  [470,]    0.3011182    0.3000369  0.2953307
#>  [471,]    0.2506709    0.2497708  0.2458530
#>  [472,]    0.1321781    0.1317034  0.1296375
#>  [473,]    0.1934969    0.1928021  0.1897778
#>  [474,]    0.1636417    0.1630541  0.1604965
#>  [475,]    0.2361474    0.2352995  0.2316086
#>  [476,]    0.1987832    0.1980694  0.1949625
#>  [477,]    0.2478503    0.2469603  0.2430867
#>  [478,]    0.2101363    0.2093817  0.2060974
#>  [479,]    0.1985618    0.1978488  0.1947454
#>  [480,]    0.3429655    0.3417340  0.3363738
#>  [481,]    0.1845601    0.1838973  0.1810128
#>  [482,]    0.1912384    0.1905517  0.1875628
#>  [483,]    0.2424947    0.2416240  0.2378339
#>  [484,]    0.1586834    0.1581135  0.1556334
#>  [485,]    0.1346528    0.1341693  0.1320647
#>  [486,]    0.2028862    0.2021576  0.1989866
#>  [487,]    0.1554855    0.1549272  0.1524970
#>  [488,]    0.2150337    0.2142616  0.2109008
#>  [489,]    0.2783468    0.2773473  0.2729969
#>  [490,]    0.2046903    0.2039553  0.2007561
#>  [491,]    0.1793455    0.1787015  0.1758984
#>  [492,]    0.2376977    0.2368442  0.2331291
#>  [493,]    0.1882189    0.1875430  0.1846012
#>  [494,]    0.2294949    0.2286708  0.2250840
#>  [495,]    0.2767008    0.2757073  0.2713827
#>  [496,]    0.1932537    0.1925598  0.1895393
#>  [497,]    0.2274782    0.2266614  0.2231061
#>  [498,]    0.2362496    0.2354013  0.2317089
#>  [499,]    0.2511882    0.2502863  0.2463604
#>  [500,]    0.1895987    0.1889179  0.1859545
#>  [501,]    0.2775827    0.2765860  0.2722476
#>  [502,]    0.2013708    0.2006477  0.1975003
#>  [503,]    0.2342477    0.2334066  0.2297455
#>  [504,]    0.1755124    0.1748821  0.1721389
#>  [505,]    0.1545361    0.1539812  0.1515658
#>  [506,]    0.1891580    0.1884788  0.1855224
#>  [507,]    0.2410724    0.2402068  0.2364390
#>  [508,]    0.1705869    0.1699743  0.1673081
#>  [509,]    0.1846008    0.1839379  0.1810527
#>  [510,]    0.2452925    0.2444117  0.2405780
#>  [511,]    0.2336408    0.2328018  0.2291502
#>  [512,]    0.1845175    0.1838549  0.1809710
#>  [513,]    0.2145704    0.2137999  0.2104463
#>  [514,]    0.2042629    0.2035295  0.2003370
#>  [515,]    0.1724497    0.1718305  0.1691352
#>  [516,]    0.2815514    0.2805404  0.2761400
#>  [517,]    0.2235356    0.2227329  0.2192392
#>  [518,]    0.1705482    0.1699358  0.1672702
#>  [519,]    0.2072745    0.2065302  0.2032906
#>  [520,]    0.1625565    0.1619728  0.1594321
#>  [521,]    0.3060694    0.3049704  0.3001869
#>  [522,]    0.2507943    0.2498937  0.2459740
#>  [523,]    0.2088600    0.2081100  0.2048456
#>  [524,]    0.2123418    0.2115794  0.2082606
#>  [525,]    0.2729041    0.2719242  0.2676589
#>  [526,]    0.1995906    0.1988739  0.1957544
#>  [527,]    0.2208487    0.2200557  0.2166040
#>  [528,]    0.2094186    0.2086667  0.2053936
#>  [529,]    0.1461427    0.1456179  0.1433338
#>  [530,]    0.2186808    0.2178956  0.2144777
#>  [531,]    0.1942698    0.1935722  0.1905359
#>  [532,]    0.1977958    0.1970855  0.1939941
#>  [533,]    0.1856025    0.1849360  0.1820351
#>  [534,]    0.2288206    0.2279990  0.2244227
#>  [535,]    0.2288076    0.2279860  0.2244099
#>  [536,]    0.1962897    0.1955849  0.1925170
#>  [537,]    0.1981813    0.1974697  0.1943722
#>  [538,]    0.2232604    0.2224587  0.2189693
#>  [539,]    0.1937939    0.1930981  0.1900691
#>  [540,]    0.2030222    0.2022932  0.1991201
#>  [541,]    0.1769461    0.1763107  0.1735451
#>  [542,]    0.2512413    0.2503391  0.2464124
#>  [543,]    0.2540659    0.2531536  0.2491827
#>  [544,]    0.2519139    0.2510093  0.2470721
#>  [545,]    0.2270846    0.2262692  0.2227200
#>  [546,]    0.1990048    0.1982902  0.1951799
#>  [547,]    0.2539445    0.2530326  0.2490636
#>  [548,]    0.1899285    0.1892465  0.1862780
#>  [549,]    0.2471539    0.2462664  0.2424036
#>  [550,]    0.2782968    0.2772975  0.2729479
#>  [551,]    0.3181974    0.3170548  0.3120817
#>  [552,]    0.3118846    0.3107647  0.3058903
#>  [553,]    0.1763356    0.1757024  0.1729464
#>  [554,]    0.1891677    0.1884885  0.1855319
#>  [555,]    0.2372900    0.2364379  0.2327293
#>  [556,]    0.2175831    0.2168018  0.2134011
#>  [557,]    0.2474013    0.2465129  0.2426462
#>  [558,]    0.2433254    0.2424517  0.2386488
#>  [559,]    0.2706413    0.2696695  0.2654396
#>  [560,]    0.2010492    0.2003272  0.1971850
#>  [561,]    0.2166088    0.2158310  0.2124455
#>  [562,]    0.1892672    0.1885876  0.1856294
#>  [563,]    0.1705805    0.1699680  0.1673019
#>  [564,]    0.1856813    0.1850145  0.1821124
#>  [565,]    0.2318399    0.2310074  0.2273839
#>  [566,]    0.1586208    0.1580512  0.1555720
#>  [567,]    0.2031645    0.2024350  0.1992597
#>  [568,]    0.2470156    0.2461287  0.2422680
#>  [569,]    0.2812886    0.2802786  0.2758823
#>  [570,]    0.1633045    0.1627181  0.1601657
#>  [571,]    0.2280681    0.2272492  0.2236846
#>  [572,]    0.2122543    0.2114921  0.2081748
#>  [573,]    0.1899740    0.1892918  0.1863226
#>  [574,]    0.2609128    0.2599759  0.2558980
#>  [575,]    0.2616421    0.2607026  0.2566134
#>  [576,]    0.2040519    0.2033192  0.2001300
#>  [577,]    0.2345027    0.2336606  0.2299955
#>  [578,]    0.2117060    0.2109458  0.2076369
#>  [579,]    0.2347096    0.2338668  0.2301985
#>  [580,]    0.2128494    0.2120851  0.2087584
#>  [581,]    0.1841836    0.1835223  0.1806436
#>  [582,]    0.2096344    0.2088816  0.2056052
#>  [583,]    0.2260761    0.2252643  0.2217309
#>  [584,]    0.2274211    0.2266044  0.2230500
#>  [585,]    0.1895674    0.1888867  0.1859239
#>  [586,]    0.2303430    0.2295159  0.2259158
#>  [587,]    0.2572504    0.2563266  0.2523061
#>  [588,]    0.1963712    0.1956661  0.1925969
#>  [589,]    0.2551336    0.2542174  0.2502299
#>  [590,]    0.1602203    0.1596450  0.1571408
#>  [591,]    0.2214925    0.2206972  0.2172354
#>  [592,]    0.1774541    0.1768169  0.1740433
#>  [593,]    0.2701118    0.2691419  0.2649203
#>  [594,]    0.1644104    0.1638200  0.1612504
#>  [595,]    0.1908652    0.1901798  0.1871967
#>  [596,]    0.2003543    0.1996348  0.1965034
#>  [597,]    0.2112187    0.2104602  0.2071590
#>  [598,]    0.2512059    0.2503039  0.2463777
#>  [599,]    0.1864722    0.1858026  0.1828881
#>  [600,]    0.1536171    0.1530654  0.1506645
#>  [601,]    0.1467382    0.1462113  0.1439178
#>  [602,]    0.1982241    0.1975123  0.1944142
#>  [603,]    0.2643938    0.2634444  0.2593122
#>  [604,]    0.1950101    0.1943098  0.1912619
#>  [605,]    0.1961536    0.1954492  0.1923835
#>  [606,]    0.1943122    0.1936144  0.1905774
#>  [607,]    0.2098786    0.2091250  0.2058447
#>  [608,]    0.2064431    0.2057019  0.2024753
#>  [609,]    0.2163549    0.2155780  0.2121965
#>  [610,]    0.1920491    0.1913595  0.1883578
#>  [611,]    0.1710755    0.1704612  0.1677874
#>  [612,]    0.1539696    0.1534167  0.1510102
#>  [613,]    0.2042892    0.2035556  0.2003627
#>  [614,]    0.2410389    0.2401734  0.2364061
#>  [615,]    0.2240717    0.2232671  0.2197650
#>  [616,]    0.1640744    0.1634852  0.1609208
#>  [617,]    0.1741387    0.1735134  0.1707917
#>  [618,]    0.1964467    0.1957413  0.1926709
#>  [619,]    0.1961246    0.1954204  0.1923550
#>  [620,]    0.2066757    0.2059335  0.2027033
#>  [621,]    0.2178825    0.2171001  0.2136947
#>  [622,]    0.2291630    0.2283401  0.2247585
#>  [623,]    0.2146256    0.2138549  0.2105004
#>  [624,]    0.1976735    0.1969637  0.1938742
#>  [625,]    0.2089112    0.2081611  0.2048959
#>  [626,]    0.2329216    0.2320852  0.2284448
#>  [627,]    0.2674528    0.2664925  0.2623124
#>  [628,]    0.1768317    0.1761967  0.1734329
#>  [629,]    0.2167336    0.2159553  0.2125679
#>  [630,]    0.2740726    0.2730884  0.2688049
#>  [631,]    0.2099922    0.2092382  0.2059561
#>  [632,]    0.2677809    0.2668194  0.2626342
#>  [633,]    0.2683367    0.2673732  0.2631793
#>  [634,]    0.1973992    0.1966904  0.1936052
#>  [635,]    0.1748244    0.1741966  0.1714642
#>  [636,]    0.2049225    0.2041867  0.2009838
#>  [637,]    0.2163912    0.2156142  0.2122321
#>  [638,]    0.1976389    0.1969293  0.1938403
#>  [639,]    0.2441524    0.2432757  0.2394597
#>  [640,]    0.1625512    0.1619675  0.1594269
#>  [641,]    0.2946856    0.2936275  0.2890218
#>  [642,]    0.1697784    0.1691687  0.1665152
#>  [643,]    0.1924485    0.1917575  0.1887496
#>  [644,]    0.2900554    0.2890139  0.2844806
#>  [645,]    0.2632906    0.2623452  0.2582302
#>  [646,]    0.2812900    0.2802800  0.2758837
#>  [647,]    0.1296940    0.1292283  0.1272012
#>  [648,]    0.2328397    0.2320037  0.2283645
#>  [649,]    0.1601324    0.1595574  0.1570546
#>  [650,]    0.2219832    0.2211861  0.2177166
#>  [651,]    0.1682133    0.1676092  0.1649801
#>  [652,]    0.2299902    0.2291643  0.2255697
#>  [653,]    0.2091682    0.2084172  0.2051480
#>  [654,]    0.1915884    0.1909004  0.1879060
#>  [655,]    0.1953865    0.1946849  0.1916311
#>  [656,]    0.2421982    0.2413285  0.2375432
#>  [657,]    0.2033758    0.2026455  0.1994669
#>  [658,]    0.2227598    0.2219599  0.2184783
#>  [659,]    0.2009624    0.2002407  0.1970998
#>  [660,]    0.2148927    0.2141211  0.2107625
#>  [661,]    0.1833854    0.1827269  0.1798606
#>  [662,]    0.2679662    0.2670040  0.2628160
#>  [663,]    0.1742269    0.1736013  0.1708782
#>  [664,]    0.2114255    0.2106663  0.2073619
#>  [665,]    0.1626908    0.1621066  0.1595638
#>  [666,]    0.2206248    0.2198326  0.2163844
#>  [667,]    0.2438592    0.2429836  0.2391722
#>  [668,]    0.1679192    0.1673162  0.1646917
#>  [669,]    0.2256729    0.2248626  0.2213355
#>  [670,]    0.2711197    0.2701462  0.2659089
#>  [671,]    0.2260361    0.2252245  0.2216916
#>  [672,]    0.3177038    0.3165630  0.3115976
#>  [673,]    0.2251752    0.2243667  0.2208473
#>  [674,]    0.2839089    0.2828895  0.2784523
#>  [675,]    0.2417405    0.2408724  0.2370942
#>  [676,]    0.2085533    0.2078044  0.2045449
#>  [677,]    0.2234682    0.2226658  0.2191732
#>  [678,]    0.2430822    0.2422094  0.2384103
#>  [679,]    0.1913345    0.1906474  0.1876570
#>  [680,]    0.1986303    0.1979171  0.1948126
#>  [681,]    0.2307293    0.2299008  0.2262946
#>  [682,]    0.1902750    0.1895917  0.1866178
#>  [683,]    0.2094517    0.2086996  0.2054260
#>  [684,]    0.1766569    0.1760225  0.1732614
#>  [685,]    0.2165970    0.2158193  0.2124340
#>  [686,]    0.2393053    0.2384460  0.2347058
#>  [687,]    0.2143715    0.2136017  0.2102512
#>  [688,]    0.1893789    0.1886989  0.1857390
#>  [689,]    0.2468034    0.2459172  0.2420598
#>  [690,]    0.1833123    0.1826540  0.1797889
#>  [691,]    0.1571333    0.1565690  0.1541131
#>  [692,]    0.2349474    0.2341037  0.2304316
#>  [693,]    0.2497228    0.2488261  0.2449231
#>  [694,]    0.2676123    0.2666514  0.2624689
#>  [695,]    0.2042195    0.2034861  0.2002943
#>  [696,]    0.1855728    0.1849064  0.1820060
#>  [697,]    0.1680793    0.1674757  0.1648487
#>  [698,]    0.3082865    0.3071795  0.3023613
#>  [699,]    0.1881636    0.1874880  0.1845471
#>  [700,]    0.1954349    0.1947332  0.1916786
#>  [701,]    0.1820353    0.1813816  0.1785365
#>  [702,]    0.2427905    0.2419187  0.2381240
#>  [703,]    0.1724049    0.1717858  0.1690912
#>  [704,]    0.1870757    0.1864040  0.1834800
#>  [705,]    0.1697271    0.1691177  0.1664649
#>  [706,]    0.1340485    0.1335672  0.1314721
#>  [707,]    0.1855399    0.1848736  0.1819738
#>  [708,]    0.1937013    0.1930058  0.1899784
#>  [709,]    0.1727249    0.1721046  0.1694050
#>  [710,]    0.3069077    0.3058057  0.3010090
#>  [711,]    0.2092768    0.2085253  0.2052544
#>  [712,]    0.1881655    0.1874898  0.1845489
#>  [713,]    0.1999435    0.1992256  0.1961006
#>  [714,]    0.3006039    0.2995245  0.2948263
#>  [715,]    0.2426787    0.2418073  0.2380144
#>  [716,]    0.2999424    0.2988653  0.2941776
#>  [717,]    0.1995829    0.1988663  0.1957469
#>  [718,]    0.1883378    0.1876615  0.1847179
#>  [719,]    0.2068235    0.2060808  0.2028483
#>  [720,]    0.2625953    0.2616524  0.2575483
#>  [721,]    0.1801037    0.1794570  0.1766420
#>  [722,]    0.2040949    0.2033620  0.2001721
#>  [723,]    0.2253290    0.2245198  0.2209981
#>  [724,]    0.2400468    0.2391848  0.2354331
#>  [725,]    0.1924700    0.1917788  0.1887706
#>  [726,]    0.2392925    0.2384332  0.2346933
#>  [727,]    0.2670019    0.2660432  0.2618702
#>  [728,]    0.1931215    0.1924280  0.1894097
#>  [729,]    0.2064099    0.2056687  0.2024427
#>  [730,]    0.3541318    0.3528602  0.3473255
#>  [731,]    0.2508758    0.2499750  0.2460540
#>  [732,]    0.2691292    0.2681628  0.2639566
#>  [733,]    0.2171841    0.2164043  0.2130098
#>  [734,]    0.1803673    0.1797196  0.1769006
#>  [735,]    0.1841432    0.1834820  0.1806040
#>  [736,]    0.1958072    0.1951041  0.1920437
#>  [737,]    0.1482894    0.1477569  0.1454392
#>  [738,]    0.1812911    0.1806401  0.1778067
#>  [739,]    0.1883434    0.1876670  0.1847233
#>  [740,]    0.2033298    0.2025997  0.1994218
#>  [741,]    0.2007111    0.1999904  0.1968534
#>  [742,]    0.2253761    0.2245668  0.2210443
#>  [743,]    0.1677704    0.1671680  0.1645458
#>  [744,]    0.2497267    0.2488300  0.2449269
#>  [745,]    0.2178334    0.2170512  0.2136466
#>  [746,]    0.2003982    0.1996786  0.1965465
#>  [747,]    0.2904020    0.2893592  0.2848205
#>  [748,]    0.3017781    0.3006945  0.2959780
#>  [749,]    0.1802326    0.1795854  0.1767684
#>  [750,]    0.2534382    0.2525281  0.2485671
#>  [751,]    0.1871789    0.1865068  0.1835813
#>  [752,]    0.2352127    0.2343681  0.2306918
#>  [753,]    0.1568842    0.1563208  0.1538688
#>  [754,]    0.2249673    0.2241595  0.2206434
#>  [755,]    0.1499792    0.1494407  0.1470966
#>  [756,]    0.1657544    0.1651592  0.1625685
#>  [757,]    0.2238454    0.2230416  0.2195430
#>  [758,]    0.1777719    0.1771336  0.1743551
#>  [759,]    0.1884328    0.1877562  0.1848111
#>  [760,]    0.2719402    0.2709637  0.2667136
#>  [761,]    0.1695430    0.1689342  0.1662843
#>  [762,]    0.1848899    0.1842260  0.1813363
#>  [763,]    0.2454332    0.2445519  0.2407160
#>  [764,]    0.2364287    0.2355797  0.2318845
#>  [765,]    0.1919373    0.1912481  0.1882482
#>  [766,]    0.2359299    0.2350828  0.2313954
#>  [767,]    0.2347334    0.2338906  0.2302218
#>  [768,]    0.1696459    0.1690368  0.1663853
#>  [769,]    0.1859697    0.1853019  0.1823953
#>  [770,]    0.2421374    0.2412679  0.2374835
#>  [771,]    0.2050987    0.2043622  0.2011566
#>  [772,]    0.3209753    0.3198228  0.3148063
#>  [773,]    0.1828812    0.1822245  0.1793661
#>  [774,]    0.2126484    0.2118848  0.2085613
#>  [775,]    0.1882617    0.1875857  0.1846432
#>  [776,]    0.2191612    0.2183743  0.2149490
#>  [777,]    0.2207350    0.2199424  0.2164925
#>  [778,]    0.2255126    0.2247029  0.2211782
#>  [779,]    0.1656464    0.1650516  0.1624626
#>  [780,]    0.2494606    0.2485648  0.2446660
#>  [781,]    0.1959664    0.1952627  0.1921999
#>  [782,]    0.2058255    0.2050864  0.2018695
#>  [783,]    0.2356143    0.2347682  0.2310858
#>  [784,]    0.3157003    0.3145667  0.3096326
#>  [785,]    0.1805228    0.1798746  0.1770531
#>  [786,]    0.2071531    0.2064092  0.2031715
#>  [787,]    0.3269192    0.3257453  0.3206359
#>  [788,]    0.2220741    0.2212766  0.2178058
#>  [789,]    0.2199244    0.2191347  0.2156974
#>  [790,]    0.2788014    0.2778003  0.2734428
#>  [791,]    0.1724584    0.1718392  0.1691437
#>  [792,]    0.1888677    0.1881895  0.1852376
#>  [793,]    0.1552650    0.1547074  0.1522807
#>  [794,]    0.2727899    0.2718104  0.2675469
#>  [795,]    0.2209067    0.2201135  0.2166608
#>  [796,]    0.2995476    0.2984721  0.2937905
#>  [797,]    0.1476802    0.1471499  0.1448417
#>  [798,]    0.1860580    0.1853899  0.1824819
#>  [799,]    0.1982699    0.1975580  0.1944591
#>  [800,]    0.2296187    0.2287942  0.2252054
#>  [801,]    0.2261601    0.2253480  0.2218133
#>  [802,]    0.2179253    0.2171428  0.2137367
#>  [803,]    0.2523227    0.2514167  0.2474731
#>  [804,]    0.1903981    0.1897144  0.1867386
#>  [805,]    0.1947784    0.1940790  0.1910348
#>  [806,]    0.1772356    0.1765992  0.1738291
#>  [807,]    0.2055108    0.2047728  0.2015608
#>  [808,]    0.2105797    0.2098235  0.2065323
#>  [809,]    0.1667621    0.1661633  0.1635568
#>  [810,]    0.2257154    0.2249049  0.2213772
#>  [811,]    0.2408542    0.2399893  0.2362250
#>  [812,]    0.1981043    0.1973930  0.1942967
#>  [813,]    0.1906124    0.1899279  0.1869488
#>  [814,]    0.2034463    0.2027157  0.1995360
#>  [815,]    0.2575407    0.2566159  0.2525908
#>  [816,]    0.1412451    0.1407379  0.1385303
#>  [817,]    0.2196024    0.2188139  0.2153816
#>  [818,]    0.1737184    0.1730946  0.1703795
#>  [819,]    0.2542495    0.2533365  0.2493627
#>  [820,]    0.1943755    0.1936775  0.1906395
#>  [821,]    0.2054665    0.2047287  0.2015174
#>  [822,]    0.2205510    0.2197590  0.2163119
#>  [823,]    0.1746087    0.1739817  0.1712527
#>  [824,]    0.2489711    0.2480771  0.2441858
#>  [825,]    0.2704530    0.2694819  0.2652550
#>  [826,]    0.3277764    0.3265994  0.3214766
#>  [827,]    0.1874231    0.1867501  0.1838208
#>  [828,]    0.2247542    0.2239471  0.2204344
#>  [829,]    0.1890730    0.1883940  0.1854389
#>  [830,]    0.2208932    0.2201000  0.2166476
#>  [831,]    0.2380163    0.2371617  0.2334417
#>  [832,]    0.2471464    0.2462589  0.2423962
#>  [833,]    0.2003577    0.1996383  0.1965068
#>  [834,]    0.1828864    0.1822297  0.1793712
#>  [835,]    0.1463148    0.1457894  0.1435026
#>  [836,]    0.1563795    0.1558180  0.1533739
#>  [837,]    0.1562699    0.1557088  0.1532664
#>  [838,]    0.1843866    0.1837245  0.1808426
#>  [839,]    0.2241229    0.2233181  0.2198152
#>  [840,]    0.2142298    0.2134605  0.2101122
#>  [841,]    0.2025835    0.2018561  0.1986898
#>  [842,]    0.2480640    0.2471733  0.2432963
#>  [843,]    0.2184014    0.2176172  0.2142037
#>  [844,]    0.2764640    0.2754713  0.2711505
#>  [845,]    0.3009718    0.2998911  0.2951872
#>  [846,]    0.1649546    0.1643622  0.1617841
#>  [847,]    0.2135114    0.2127447  0.2094077
#>  [848,]    0.2091940    0.2084428  0.2051732
#>  [849,]    0.2594212    0.2584897  0.2544351
#>  [850,]    0.2497401    0.2488433  0.2449401
#>  [851,]    0.2097431    0.2089900  0.2057118
#>  [852,]    0.2094115    0.2086596  0.2053866
#>  [853,]    0.1836625    0.1830030  0.1801325
#>  [854,]    0.2188747    0.2180888  0.2146678
#>  [855,]    0.1942031    0.1935058  0.1904705
#>  [856,]    0.2098397    0.2090863  0.2058066
#>  [857,]    0.2924079    0.2913580  0.2867879
#>  [858,]    0.2246895    0.2238826  0.2203709
#>  [859,]    0.3263201    0.3251483  0.3200483
#>  [860,]    0.1343073    0.1338250  0.1317258
#>  [861,]    0.2067682    0.2060257  0.2027940
#>  [862,]    0.1965664    0.1958606  0.1927884
#>  [863,]    0.1574733    0.1569079  0.1544466
#>  [864,]    0.2579787    0.2570523  0.2530203
#>  [865,]    0.2889020    0.2878646  0.2833493
#>  [866,]    0.1614814    0.1609015  0.1583777
#>  [867,]    0.2749098    0.2739227  0.2696261
#>  [868,]    0.1876354    0.1869616  0.1840290
#>  [869,]    0.1874563    0.1867832  0.1838533
#>  [870,]    0.1889536    0.1882751  0.1853219
#>  [871,]    0.1679358    0.1673328  0.1647081
#>  [872,]    0.2135040    0.2127374  0.2094004
#>  [873,]    0.2221727    0.2213749  0.2179025
#>  [874,]    0.2285615    0.2277408  0.2241685
#>  [875,]    0.1487582    0.1482240  0.1458990
#>  [876,]    0.3070277    0.3059253  0.3011267
#>  [877,]    0.2089572    0.2082069  0.2049410
#>  [878,]    0.2334560    0.2326177  0.2289690
#>  [879,]    0.2048390    0.2041034  0.2009019
#>  [880,]    0.2581383    0.2572114  0.2531769
#>  [881,]    0.2466669    0.2457812  0.2419261
#>  [882,]    0.2299953    0.2291695  0.2255748
#>  [883,]    0.2346767    0.2338340  0.2301662
#>  [884,]    0.2321579    0.2313243  0.2276958
#>  [885,]    0.1867806    0.1861099  0.1831907
#>  [886,]    0.2235118    0.2227092  0.2192158
#>  [887,]    0.2228947    0.2220943  0.2186106
#>  [888,]    0.1997226    0.1990054  0.1958839
#>  [889,]    0.1984357    0.1977232  0.1946217
#>  [890,]    0.2106539    0.2098975  0.2066051
#>  [891,]    0.2386841    0.2378270  0.2340965
#>  [892,]    0.1820736    0.1814198  0.1785740
#>  [893,]    0.2172979    0.2165176  0.2131213
#>  [894,]    0.1739050    0.1732805  0.1705624
#>  [895,]    0.2669468    0.2659882  0.2618161
#>  [896,]    0.2853449    0.2843203  0.2798606
#>  [897,]    0.2454990    0.2446174  0.2407804
#>  [898,]    0.1487571    0.1482229  0.1458978
#>  [899,]    0.1770426    0.1764068  0.1736397
#>  [900,]    0.1687501    0.1681441  0.1655066
#>  [901,]    0.1839667    0.1833061  0.1804308
#>  [902,]    0.1764988    0.1758650  0.1731064
#>  [903,]    0.2287598    0.2279384  0.2243631
#>  [904,]    0.2350282    0.2341842  0.2305109
#>  [905,]    0.2517692    0.2508651  0.2469301
#>  [906,]    0.1785596    0.1779185  0.1751277
#>  [907,]    0.2700199    0.2690504  0.2648302
#>  [908,]    0.2649723    0.2640208  0.2598795
#>  [909,]    0.1783531    0.1777127  0.1749252
#>  [910,]    0.1785966    0.1779552  0.1751639
#>  [911,]    0.2130102    0.2122454  0.2089161
#>  [912,]    0.1986558    0.1979425  0.1948376
#>  [913,]    0.3398828    0.3386624  0.3333504
#>  [914,]    0.1661972    0.1656004  0.1630028
#>  [915,]    0.1357051    0.1352178  0.1330967
#>  [916,]    0.2358648    0.2350179  0.2313315
#>  [917,]    0.1831459    0.1824883  0.1796258
#>  [918,]    0.2008907    0.2001693  0.1970295
#>  [919,]    0.2263143    0.2255017  0.2219646
#>  [920,]    0.1799924    0.1793460  0.1765328
#>  [921,]    0.2263486    0.2255358  0.2219981
#>  [922,]    0.2301681    0.2293417  0.2257444
#>  [923,]    0.1811974    0.1805468  0.1777147
#>  [924,]    0.3256337    0.3244644  0.3193751
#>  [925,]    0.2452314    0.2443508  0.2405180
#>  [926,]    0.2650201    0.2640685  0.2599264
#>  [927,]    0.2175814    0.2168001  0.2133995
#>  [928,]    0.1949401    0.1942401  0.1911933
#>  [929,]    0.2495888    0.2486926  0.2447917
#>  [930,]    0.2657056    0.2647515  0.2605988
#>  [931,]    0.2206767    0.2198843  0.2164353
#>  [932,]    0.2260498    0.2252381  0.2217051
#>  [933,]    0.2101729    0.2094182  0.2061333
#>  [934,]    0.1801089    0.1794621  0.1766471
#>  [935,]    0.2204132    0.2196218  0.2161769
#>  [936,]    0.3526587    0.3513925  0.3458809
#>  [937,]    0.2057890    0.2050501  0.2018337
#>  [938,]    0.1963727    0.1956675  0.1925984
#>  [939,]    0.1897342    0.1890529  0.1860874
#>  [940,]    0.2165680    0.2157903  0.2124055
#>  [941,]    0.2261298    0.2253178  0.2217835
#>  [942,]    0.2021019    0.2013762  0.1982174
#>  [943,]    0.2098784    0.2091248  0.2058445
#>  [944,]    0.2010393    0.2003175  0.1971753
#>  [945,]    0.2001986    0.1994797  0.1963507
#>  [946,]    0.2233960    0.2225938  0.2191023
#>  [947,]    0.2285512    0.2277305  0.2241584
#>  [948,]    0.2129871    0.2122223  0.2088935
#>  [949,]    0.2511313    0.2502295  0.2463046
#>  [950,]    0.1875637    0.1868902  0.1839586
#>  [951,]    0.1708613    0.1702478  0.1675773
#>  [952,]    0.1788875    0.1782451  0.1754492
#>  [953,]    0.1871642    0.1864922  0.1835669
#>  [954,]    0.2088993    0.2081492  0.2048842
#>  [955,]    0.3081876    0.3070809  0.3022643
#>  [956,]    0.2251360    0.2243276  0.2208089
#>  [957,]    0.2056334    0.2048950  0.2016810
#>  [958,]    0.2109252    0.2101679  0.2068712
#>  [959,]    0.1571834    0.1566189  0.1541622
#>  [960,]    0.2208154    0.2200225  0.2165714
#>  [961,]    0.2605240    0.2595885  0.2555167
#>  [962,]    0.2149841    0.2142122  0.2108522
#>  [963,]    0.1943111    0.1936134  0.1905764
#>  [964,]    0.2653273    0.2643746  0.2602278
#>  [965,]    0.2000244    0.1993061  0.1961799
#>  [966,]    0.2799453    0.2789401  0.2745648
#>  [967,]    0.2222055    0.2214076  0.2179346
#>  [968,]    0.3277806    0.3266037  0.3214810
#>  [969,]    0.2367353    0.2358853  0.2321853
#>  [970,]    0.2730699    0.2720894  0.2678215
#>  [971,]    0.1735754    0.1729521  0.1702393
#>  [972,]    0.2676936    0.2667324  0.2625485
#>  [973,]    0.1849323    0.1842682  0.1813778
#>  [974,]    0.1927316    0.1920395  0.1890272
#>  [975,]    0.2931811    0.2921284  0.2875462
#>  [976,]    0.1919396    0.1912504  0.1882505
#>  [977,]    0.2213468    0.2205520  0.2170925
#>  [978,]    0.2381854    0.2373301  0.2336074
#>  [979,]    0.2079217    0.2071751  0.2039254
#>  [980,]    0.1796219    0.1789769  0.1761695
#>  [981,]    0.1832720    0.1826139  0.1797494
#>  [982,]    0.2508984    0.2499975  0.2460761
#>  [983,]    0.3084646    0.3073570  0.3025360
#>  [984,]    0.2231037    0.2223026  0.2188156
#>  [985,]    0.2756821    0.2746922  0.2703836
#>  [986,]    0.1799676    0.1793213  0.1765085
#>  [987,]    0.3213270    0.3201732  0.3151513
#>  [988,]    0.1817291    0.1810766  0.1782363
#>  [989,]    0.1628730    0.1622881  0.1597425
#>  [990,]    0.2073679    0.2066233  0.2033822
#>  [991,]    0.2073953    0.2066505  0.2034091
#>  [992,]    0.1803185    0.1796710  0.1768528
#>  [993,]    0.2959722    0.2949094  0.2902836
#>  [994,]    0.3343825    0.3331818  0.3279558
#>  [995,]    0.2248740    0.2240665  0.2205519
#>  [996,]    0.2215136    0.2207181  0.2172560
#>  [997,]    0.1889869    0.1883083  0.1853546
#>  [998,]    0.2154335    0.2146599  0.2112929
#>  [999,]    0.2557244    0.2548062  0.2508094
#> [1000,]    0.1972213    0.1965131  0.1934306