Stochastic collision risk model for a single species and one wind farm scenario

Runs a Stochastic Collision Risk Model (SCRM) for estimating the number of in-flight collisions with offshore windfarm turbines, for given species and windfarm scenario. Core calculations follow the work developed by Masden (2015). See Background and Updates section below for more details.

Usage

stoch_crm(
  model_options = c("1", "2", "3", "4"),
  n_iter = 1000,
  flt_speed_pars,
  body_lt_pars,
  wing_span_pars,
  avoid_bsc_pars = NULL,
  avoid_ext_pars = NULL,
  noct_act_pars,
  prop_crh_pars = NULL,
  bird_dens_opt = c("tnorm", "resample", "qtiles"),
  bird_dens_dt,
  flight_type,
  prop_upwind,
  gen_fhd_boots = NULL,
  site_fhd_boots = NULL,
  n_blades,
  air_gap_pars,
  rtr_radius_pars,
  bld_width_pars,
  bld_chord_prf = chord_prof_5MW,
  rtn_pitch_opt = c("probDist", "windSpeedReltn"),
  bld_pitch_pars = NULL,
  rtn_speed_pars = NULL,
  windspd_pars = NULL,
  rtn_pitch_windspd_dt = NULL,
  trb_wind_avbl,
  trb_downtime_pars,
  wf_n_trbs,
  wf_width,
  wf_latitude,
  tidal_offset,
  lrg_arr_corr = TRUE,
  xinc = 0.05,
  yinc = 0.05,
  out_format = c("draws", "summaries"),
  out_sampled_pars = FALSE,
  out_period = c("months", "seasons", "annum"),
  season_specs = NULL,
  verbose = TRUE,
  log_file = NULL,
  seed = NULL
)

Arguments

model_options

Character vector, the model options for calculating collision risk (see Details section below).

n_iter

An integer. The number of iterations for the model simulation.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

avoid_bsc_pars, avoid_ext_pars

Single row data frames with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used in the basic model (Options 1 and 2) and extended model (Options 3 and 4) calculations (see Details section). Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

noct_act_pars

A single row data frame with columns mean and sd, The mean and standard deviation of the species nocturnal flight activity level, expressed as a proportion of daytime activity levels, and assumed to be Beta distributed.

prop_crh_pars

Required only for model Option 1, a single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height derived from site survey, assumed to be Beta distributed.

bird_dens_opt

Option for specifying the random sampling mechanism for bird densities:

"tnorm": Sampling of density estimates from a tnorm-lw0 distribution (default value),

"resample": Re-sample draws of bird density estimates (e.g. bootstrap samples),

"qtiles": Sampling from a set of quantile estimates of bird densities.

bird_dens_dt

A data frame with monthly estimates of bird density within the windfarm footprint, expressed as the number of daytime in-flight birds/km^2 per month. Data frame format requirements:

If bird_dens_opt = "tnorm", bird_dens_dt must contain the following columns:

month, (unique) month names,

mean, the mean number of birds in flight at any height per square kilometre in each month,

sd, idem, for standard deviation.

If bird_dens_opt = "resample", bird_dens_dt columns must be named as months (i.e. Jan, Feb, ...), each containing random samples of monthly density estimates.

If bird_dens_opt = "qtiles", bird_dens_dt must comply with:

First column named as p, giving reference probabilities,

Remaining columns named as months (i.e. Jan, Feb, ...), each giving the quantile estimates of bird density in a given month, for the reference probabilities in column p.

flight_type

A character string, either 'flapping' or 'gliding', indicating the species' characteristic flight type.

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

gen_fhd_boots

Required only for model Options 2 and 3, a data frame with bootstrap samples of flight height distributions (FHD) of the species derived from general (country/regional level) data. FHD provides relative frequency distribution of bird flights at 1-metre height bands, starting from sea surface. The first column must be named as height, expressing the lower bound of the height band (thus it's first element must be 0). Each remaining column should provide a bootstrap sample of the proportion of bird flights at each height band, with no column naming requirements.

NOTE: generic_fhd_bootstraps is a list object with generic FHD bootstrap estimates for 25 seabird species from Johnson et al (2014) (see usage in Example Section below).

site_fhd_boots

Required only for model Option 4, a data frame similar to gen_fhd_boots, but for FHD estimates derived from site-specific data.

n_blades

An integer, the number of blades in rotor (\(b\)).

air_gap_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the tip clearance gap, in metres, i.e. the distance between the minimum rotor tip height and the highest astronomical tide (HAT). Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

bld_chord_prf

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width.

Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

rtn_pitch_opt

a character string, the option for specifying the sampling mechanism for rotation speed and blade pitch:

"probDist": sample rotation speed and blade pitch values from a tnorm-lw0 distribution (default value).

"windSpeedReltn": generate rotation speed and blade pitch values as a function of wind speed intensity.

bld_pitch_pars

Only required if rtn_pitch_opt = "probDist", a single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtn_speed_pars

Only required if rtn_pitch_opt = "probDist", a single row data frame with columns mean and sd, the mean and standard deviation of the operational rotation speed, in revolutions per minute. Assumed to follow a tnorm-lw0 distribution.

windspd_pars

Only required if rtn_pitch_opt = "windSpeedReltn", a single row data frame with columns mean and sd, the mean and the standard deviation of wind speed at the windfarm site, in metres/sec. Assumed to follow a tnorm-lw0 distribution.

rtn_pitch_windspd_dt

Only required if rtn_pitch_opt = "windSpeedReltn", a data frame giving the relationship between wind speed, rotation speed and blade pitch values. It must contain the columns:

wind_speed, wind speed in m/s,

rtn_speed, rotation speed in rpm,

bld_pitch, blade pitch values in degrees.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

wf_n_trbs

Integer, the number of turbines on the windfarm.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

tidal_offset

A numeric value, the tidal offset, the difference between HAT and mean sea level, in metres.

lrg_arr_corr

Boolean value. If TRUE, the large array correction will be applied. This is a correction factor to account for the decay in bird density at later rows in wind farms with a large array of turbines.

yinc, xinc

numeric values, the increments along the y-axis and x-axis for numerical integration across segments of the rotor circle. Chosen values express proportion of rotor radius. By default these are set to 0.05, i.e. integration will be performed at a resolution of one twentieth of the rotor radius.

out_format

Output format specification. Possible values are:

"draws": returns stochastic draws of collisions estimates (default value),

"summaries": returns summary statistics of collisions estimates.

out_sampled_pars

Logical, whether to output summary statistics of values sampled for each stochastic model parameter.

out_period

Controls level of temporal aggregation of collision outputs. Possible values are:

"months": monthly collisions (default value),

"seasons": collisions per user-defined season,

"annum": total collisions over 12 months.

season_specs

Only required if out_period = "seasons", a data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

verbose

Logical, print model run progress on the console?

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

Value

If out_sampled_pars = FALSE, returns a list with estimates of number of collisions per chosen time periods, with elements containing the outputs for each CRM Option. If out_sampled_pars = TRUE, returns a list object with two top-level elements:

collisions, a list comprising collision estimates for each CRM Option,

sampled_pars, a list with summary statistics of values sampled for stochastic model parameters.

Details

Collision risk can be calculated under 4 options, specified by model_options:

Option 1 - Basic model with proportion at collision risk height derived from site survey (prop_crh_surv).

Option 2 - Basic model with proportion at collision risk height derived from a generic flight height distribution (gen_fhd).

Option 3 - Extended model using a generic flight height distribution (gen_fhd).

Option 4 - Extended model using a site-specific flight height distribution (site_fhd).

Where,

Basic model - assumes a uniform distribution of bird flights at collision risk height (i.e. above the minimum and below the maximum height of the rotor blade).

Extended model - takes into account the distribution of bird flight heights at collision risk height.

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------

# ------------------------------------------------------
# Setting some of the required inputs upfront
b_dens <- data.frame(
  month = month.abb,
  mean = runif(12, 0.8, 1.5),
  sd = runif(12, 0.2, 0.3)
)
head(b_dens)
#>   month      mean        sd
#> 1   Jan 1.3305092 0.2873584
#> 2   Feb 1.3417741 0.2394753
#> 3   Mar 1.4297807 0.2566189
#> 4   Apr 0.8379951 0.2750462
#> 5   May 0.8918102 0.2039552
#> 6   Jun 1.3165074 0.2536910

# Generic FHD bootstraps from Johnson et al (2014)
fhd_boots <- generic_fhd_bootstraps[[1]]
head(fhd_boots)
#> # A tibble: 6 × 201
#>   height bootId_1 bootId_2 bootId_3 bootId_4 bootId_5 bootId_6 bootId_7 bootId_8
#>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1      0   0.160    0.166    0.174    0.152    0.174    0.168    0.177    0.172 
#> 2      1   0.134    0.138    0.144    0.129    0.144    0.140    0.145    0.142 
#> 3      2   0.113    0.115    0.119    0.109    0.119    0.116    0.120    0.118 
#> 4      3   0.0948   0.0963   0.0981   0.0927   0.0981   0.0967   0.0986   0.0976
#> 5      4   0.0796   0.0803   0.0810   0.0786   0.0810   0.0805   0.0812   0.0808
#> 6      5   0.0669   0.0670   0.0669   0.0667   0.0669   0.0670   0.0669   0.0670
#> # … with 192 more variables: bootId_9 <dbl>, bootId_10 <dbl>, bootId_11 <dbl>,
#> #   bootId_12 <dbl>, bootId_13 <dbl>, bootId_14 <dbl>, bootId_15 <dbl>,
#> #   bootId_16 <dbl>, bootId_17 <dbl>, bootId_18 <dbl>, bootId_19 <dbl>,
#> #   bootId_20 <dbl>, bootId_21 <dbl>, bootId_22 <dbl>, bootId_23 <dbl>,
#> #   bootId_24 <dbl>, bootId_25 <dbl>, bootId_26 <dbl>, bootId_27 <dbl>,
#> #   bootId_28 <dbl>, bootId_29 <dbl>, bootId_30 <dbl>, bootId_31 <dbl>,
#> #   bootId_32 <dbl>, bootId_33 <dbl>, bootId_34 <dbl>, bootId_35 <dbl>, …

# wind speed vs rotation speed vs blade pitch
wind_rtn_ptch <- data.frame(
  wind_speed = seq_len(30),
  rtn_speed = 10/(30:1),
  bld_pitch = c(rep(90, 4), rep(0, 8), 5:22)
)
head(wind_rtn_ptch)
#>   wind_speed rtn_speed bld_pitch
#> 1          1 0.3333333        90
#> 2          2 0.3448276        90
#> 3          3 0.3571429        90
#> 4          4 0.3703704        90
#> 5          5 0.3846154         0
#> 6          6 0.4000000         0

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 93.77255
#> 2   Feb 87.45896
#> 3   Mar 97.66046
#> 4   Apr 93.19274
#> 5   May 90.24935
#> 6   Jun 88.77684

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 9.601738  2
#> 2   Feb 6.082292  2
#> 3   Mar 9.654360  2
#> 4   Apr 8.435265  2
#> 5   May 9.835292  2
#> 6   Jun 7.320258  2

# seasons specification
seas_dt <- data.frame(
  season_id = c("a", "b", "c"),
  start_month = c("Jan", "May", "Oct"), end_month = c("Apr", "Sep", "Dec")
  )
head(seas_dt)
#>   season_id start_month end_month
#> 1         a         Jan       Apr
#> 2         b         May       Sep
#> 3         c         Oct       Dec

# ----------------------------------------------------------
# Run stochastic CRM, treating rotor radius, air gap and
# blade width as fixed parameters (i.e. not stochastic)

stoch_crm(
  model_options = c(1, 2, 3),
  n_iter = 1000,
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  avoid_ext_pars = data.frame(mean = 0.96, sd = 0.002),
  noct_act_pars = data.frame(mean = 0.033, sd = 0.005),
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  bird_dens_opt = "tnorm",
  bird_dens_dt = b_dens,
  flight_type = "flapping",
  prop_upwind = 0.5,
  gen_fhd_boots = fhd_boots,
  n_blades = 3,
  rtr_radius_pars = data.frame(mean = 80, sd = 0), # sd = 0, rotor radius is fixed
  air_gap_pars = data.frame(mean = 36, sd = 0),    # sd = 0, air gap is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),   # sd = 0, blade width is fixed
  rtn_pitch_opt = "windSpeedReltn",
  windspd_pars = data.frame(mean = 7.74, sd = 3),
  rtn_pitch_windspd_dt = wind_rtn_ptch,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  wf_n_trbs = 200,
  wf_width = 15,
  wf_latitude = 56.9,
  tidal_offset = 2.5,
  lrg_arr_corr = TRUE,
  verbose = TRUE,
  seed = 1234,
  out_format = "summaries",
  out_sampled_pars = TRUE,
  out_period = "seasons",
  season_specs = seas_dt,
  log_file = paste0(getwd(), "scrm_example.log")
)
#> ── Stochastic CRM ──
#> 
#> ℹ Checking inputs
#> ✔ Checking inputs [20ms]
#> 
#> ℹ Preparing data
#> ✔ Preparing data [36ms]
#> 
#> ℹ Sampling parameters
#> ✔ Sampling parameters [139ms]
#> 
#> ⠙ Calculating collisions | 3/1000 iterations
#> ✔ Calculating collisions | 1000/1000 iterations [1.8s]
#> 
#> ℹ Sorting outputs
#> ✔ Sorting outputs [777ms]
#> 
#> ✔ Job done!
#> $collisions
#> $collisions$opt1
#> # A tibble: 3 × 10
#>   season_id period  mean    sd median pctl_2.5 pctl_25 pctl_75 pctl_97.5 pctl_99
#>   <chr>     <chr>  <dbl> <dbl>  <dbl>    <dbl>   <dbl>   <dbl>     <dbl>   <dbl>
#> 1 a         Jan_A…  35.2 15.9    36.2     9.08    23.4    46.3      65.1    89.3
#> 2 b         May_S…  61.8 27.6    63.9    16.4     42.3    80.4     113.    148. 
#> 3 c         Oct_D…  18.2  8.39   18.6     4.69    12.0    24.2      34.0    46.1
#> 
#> $collisions$opt2
#> # A tibble: 3 × 10
#>   season_id period  mean    sd median pctl_2.5 pctl_25 pctl_75 pctl_97.5 pctl_99
#>   <chr>     <chr>  <dbl> <dbl>  <dbl>    <dbl>   <dbl>   <dbl>     <dbl>   <dbl>
#> 1 a         Jan_A… 0.919 1.73   0.491   0.0888   0.277   0.817      5.29   16.6 
#> 2 b         May_S… 1.61  2.93   0.856   0.139    0.516   1.45      10.6    24.9 
#> 3 c         Oct_D… 0.473 0.890  0.253   0.0422   0.147   0.429      2.96    8.90
#> 
#> $collisions$opt3
#> # A tibble: 3 × 10
#>   season_id period  mean    sd median pctl_2.5 pctl_25 pctl_75 pctl_97.5 pctl_99
#>   <chr>     <chr>  <dbl> <dbl>  <dbl>    <dbl>   <dbl>   <dbl>     <dbl>   <dbl>
#> 1 a         Jan_A… 0.444 1.07  0.183    0.0418  0.115    0.316      3.80   10.8 
#> 2 b         May_S… 0.777 1.83  0.320    0.0732  0.205    0.557      6.60   15.9 
#> 3 c         Oct_D… 0.229 0.552 0.0926   0.0205  0.0591   0.163      1.90    5.33
#> 
#> 
#> $sampled_pars
#> $sampled_pars$air_gap
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1    36     0     36       36        36
#> 
#> $sampled_pars$bld_width
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1     8     0      8        8         8
#> 
#> $sampled_pars$body_lt
#> # A tibble: 1 × 5
#>    mean      sd median pctl_2.5 pctl_97.5
#>   <dbl>   <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.390 0.00499  0.390    0.380     0.400
#> 
#> $sampled_pars$flt_speed
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1  7.28  1.47   7.28     4.30      10.0
#> 
#> $sampled_pars$noct_actv
#> # A tibble: 1 × 5
#>     mean      sd median pctl_2.5 pctl_97.5
#>    <dbl>   <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.0333 0.00498 0.0333   0.0241    0.0436
#> 
#> $sampled_pars$rtr_radius
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1    80     0     80       80        80
#> 
#> $sampled_pars$wing_span
#> # A tibble: 1 × 5
#>    mean     sd median pctl_2.5 pctl_97.5
#>   <dbl>  <dbl>  <dbl>    <dbl>     <dbl>
#> 1  1.08 0.0398   1.08     1.00      1.16
#> 
#> $sampled_pars$hub_height
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1   116     0    116      116       116
#> 
#> $sampled_pars$dens_mth
#> # A tibble: 12 × 6
#>    period  mean    sd median pctl_2.5 pctl_97.5
#>    <chr>  <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#>  1 Jan    1.32  0.283  1.32     0.793      1.89
#>  2 Feb    1.35  0.232  1.34     0.919      1.81
#>  3 Mar    1.43  0.258  1.43     0.910      1.90
#>  4 Apr    0.851 0.269  0.851    0.338      1.40
#>  5 May    0.895 0.198  0.887    0.514      1.28
#>  6 Jun    1.31  0.249  1.31     0.779      1.81
#>  7 Jul    1.22  0.223  1.22     0.795      1.66
#>  8 Aug    1.16  0.249  1.17     0.669      1.67
#>  9 Sep    1.46  0.206  1.46     1.07       1.86
#> 10 Oct    1.05  0.206  1.05     0.657      1.46
#> 11 Nov    0.831 0.278  0.824    0.303      1.39
#> 12 Dec    1.07  0.212  1.07     0.651      1.48
#> 
#> $sampled_pars$prop_oper_mth
#> # A tibble: 12 × 6
#>    period  mean     sd median pctl_2.5 pctl_97.5
#>    <chr>  <dbl>  <dbl>  <dbl>    <dbl>     <dbl>
#>  1 Jan    0.843 0.0207  0.842    0.802     0.880
#>  2 Feb    0.814 0.0205  0.814    0.774     0.852
#>  3 Mar    0.880 0.0205  0.880    0.837     0.920
#>  4 Apr    0.848 0.0196  0.847    0.810     0.886
#>  5 May    0.804 0.0195  0.804    0.766     0.841
#>  6 Jun    0.814 0.0198  0.814    0.774     0.852
#>  7 Jul    0.794 0.0199  0.794    0.754     0.835
#>  8 Aug    0.890 0.0194  0.891    0.853     0.928
#>  9 Sep    0.773 0.0206  0.774    0.732     0.812
#> 10 Oct    0.854 0.0201  0.854    0.814     0.892
#> 11 Nov    0.876 0.0201  0.876    0.838     0.916
#> 12 Dec    0.886 0.0197  0.887    0.848     0.922
#> 
#> $sampled_pars$downtime
#> # A tibble: 12 × 6
#>    period  mean    sd median pctl_2.5 pctl_97.5
#>    <chr>  <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#>  1 Jan     9.52  2.07   9.56     5.82      13.5
#>  2 Feb     6.09  2.05   6.09     2.24      10.0
#>  3 Mar     9.66  2.05   9.65     5.63      13.9
#>  4 Apr     8.43  1.96   8.47     4.64      12.2
#>  5 May     9.86  1.95   9.83     6.14      13.7
#>  6 Jun     7.40  1.98   7.38     3.55      11.4
#>  7 Jul     9.57  1.99   9.55     5.51      13.6
#>  8 Aug     7.74  1.94   7.70     3.94      11.5
#>  9 Sep     9.52  2.06   9.50     5.67      13.6
#> 10 Oct     9.36  2.01   9.39     5.55      13.4
#> 11 Nov     8.57  2.01   8.55     4.60      12.4
#> 12 Dec     6.25  1.97   6.14     2.58      10.0
#> 
#> $sampled_pars$wind_speed
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1  7.74  3.00   7.77     2.15      13.7
#> 
#> $sampled_pars$rtn_speed
#> # A tibble: 1 × 5
#>    mean     sd median pctl_2.5 pctl_97.5
#>   <dbl>  <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.428 0.0575  0.417    0.345     0.556
#> 
#> $sampled_pars$bld_pitch
#> # A tibble: 1 × 5
#>    mean    sd median pctl_2.5 pctl_97.5
#>   <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.327 0.634      0        0      1.57
#> 
#> $sampled_pars$avoid_bsc
#> # A tibble: 1 × 5
#>    mean       sd median pctl_2.5 pctl_97.5
#>   <dbl>    <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.990 0.000986  0.990    0.988     0.992
#> 
#> $sampled_pars$avoid_ext
#> # A tibble: 1 × 5
#>    mean      sd median pctl_2.5 pctl_97.5
#>   <dbl>   <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.960 0.00203  0.960    0.956     0.964
#> 
#> $sampled_pars$prop_crh
#> # A tibble: 1 × 5
#>     mean      sd median pctl_2.5 pctl_97.5
#>    <dbl>   <dbl>  <dbl>    <dbl>     <dbl>
#> 1 0.0605 0.00910 0.0602   0.0441    0.0801
#> 
#> $sampled_pars$gen_fhd
#> # A tibble: 500 × 6
#>    height   mean      sd median pctl_2.5 pctl_97.5
#>     <dbl>  <dbl>   <dbl>  <dbl>    <dbl>     <dbl>
#>  1      0 0.163  0.0187  0.166    0.109     0.187 
#>  2      1 0.136  0.0129  0.138    0.0967    0.152 
#>  3      2 0.114  0.00851 0.115    0.0863    0.124 
#>  4      3 0.0950 0.00527 0.0963   0.0769    0.100 
#>  5      4 0.0794 0.00294 0.0803   0.0686    0.0816
#>  6      5 0.0664 0.00146 0.0669   0.0606    0.0670
#>  7      6 0.0556 0.00116 0.0558   0.0530    0.0567
#>  8      7 0.0465 0.00166 0.0466   0.0439    0.0490
#>  9      8 0.0390 0.00215 0.0389   0.0357    0.0431
#> 10      9 0.0327 0.00249 0.0324   0.0290    0.0386
#> # … with 490 more rows
#> 
#>