Grid of probabilities of single transit collision at points in rotor circle

Calculates single transit collision risk at grid points (X, Y) inside the left-half of the rotor circle, for a given flight direction (upwind or downwind), and assuming no avoidance action. The origin of the (X, Y) coordinates is the rotor center

Usage

get_pcoll_grid(
  rotor_grids,
  direction,
  rotor_radius,
  blade_width,
  rotor_speed,
  blade_pitch,
  flight_type,
  n_blades,
  flight_speed,
  wing_span,
  body_lt
)

Arguments

rotor_grids

A list object containing geometric attributes of the rotor at equidistant points within its unit circle. This object should be built via the function generate_rotor_grids().

direction

an integer, indicating the flight direction: 1 for upwind; -1 for downwind.

rotor_radius

Numeric value. The radius of the rotor (\(R\)), in metres.

blade_width

Numeric value, giving the maximum blade width, in metres.

rotor_speed

Numeric value. The operational rotation speed, in revolutions/min.

blade_pitch

Numeric value. The average blade pitch angle, the angle between the blade surface and the rotor plane (\(\gamma\)), in radians.

flight_type

A character string, either 'flapping' or 'gliding', indicating the species' characteristic flight type.

n_blades

An integer, the number of blades in rotor (\(b\)).

flight_speed

Numeric value. The bird flying speed (\(v\)), in metres/sec.

wing_span

Numeric value. The wingspan of the bird (\(W\)), in metres.

body_lt

Numeric value. The length of the bird (\(L\)), in metres.

Value

a matrix. The probability of collisions at different flight height bands

Details

Methodology and assumptions underpinning get_pcoll_grid are described in section "Extended Approach Taking Account Of Flight Heights" of Band (2012). Note that collision risk calculations are based on equation (3) in "Stage C" section, but using (X, Y) coordinates to reference transit points instead of the equivalent (r, phi) coordinates.

See also

Examples

  rotor_grids <- generate_rotor_grids(yinc = 0.05, xinc = 0.05, chord_prof_5MW)

  get_pcoll_grid(
    rotor_grids = rotor_grids,
    direction = 1,
    rotor_radius = 120,
    blade_width = 5,
    rotor_speed = 15,
    blade_pitch = 15,
    flight_type = "flapping",
    n_blades = 3,
    flight_speed = 13.1,
    wing_span = 1.01,
    body_lt = 0.85)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.04866412         NA         NA         NA         NA         NA
#>  [2,] 0.08952224 0.08845290 0.08525201 0.07994061 0.07255312 0.06313622
#>  [3,] 0.09947161 0.09919728 0.09837615 0.09701378 0.09511924 0.09270487
#>  [4,] 0.11096121 0.11062615 0.10962345 0.10796045 0.10564912 0.10270571
#>  [5,] 0.11724040 0.11704657 0.11646643 0.11550404 0.11416607 0.11246162
#>  [6,] 0.12668401 0.12637326 0.12544353 0.12390225 0.12176163 0.11903827
#>  [7,] 0.13423551 0.13397046 0.13317732 0.13186208 0.13003472 0.12770909
#>  [8,] 0.14314586 0.14281006 0.14180530 0.14013959 0.13782636 0.13488438
#>  [9,] 0.15167975 0.15133351 0.15029709 0.14857766 0.14618776 0.14314586
#> [10,] 0.15973451 0.15938135 0.15832309 0.15656413 0.15411369 0.15098765
#> [11,] 0.16560718 0.16533047 0.16449766 0.16310249 0.16113972 0.15832309
#> [12,] 0.17071654 0.17045781 0.16967283 0.16833969 0.16643411 0.16394052
#> [13,] 0.17477635 0.17456159 0.17389488 0.17272016 0.17097366 0.16860849
#> [14,] 0.18019619 0.17988207 0.17888884 0.17709524 0.17456159 0.17247714
#> [15,] 0.18172140 0.18171552 0.18156066 0.18093833 0.17922782 0.17595458
#> [16,] 0.17436118 0.17528390 0.17768346 0.18074417 0.18146894 0.17988207
#> [17,] 0.15482433 0.15754598 0.16472217 0.17436118 0.17963738 0.18146894
#> [18,] 0.12381116 0.12953693 0.14365920 0.16008846 0.17436118 0.18074417
#> [19,] 0.07632676 0.09001476 0.11720842 0.14365920 0.16472217 0.17768346
#> [20,] 0.16514020 0.07765600 0.09001476 0.12953693 0.15754598 0.17528390
#> [21,] 1.00000000 0.16514020 0.07632676 0.12381116 0.15482433 0.17436118
#> [22,] 0.16514020 0.07765600 0.09001476 0.12953693 0.15754598 0.17528390
#> [23,] 0.07632676 0.09001476 0.11720842 0.14365920 0.16472217 0.17768346
#> [24,] 0.12381116 0.12953693 0.14365920 0.16008846 0.17436118 0.18074417
#> [25,] 0.15482433 0.15754598 0.16472217 0.17436118 0.17963738 0.18146894
#> [26,] 0.17436118 0.17528390 0.17768346 0.18074417 0.18146894 0.17988207
#> [27,] 0.18172140 0.18171552 0.18156066 0.18093833 0.17922782 0.17595458
#> [28,] 0.18019619 0.17988207 0.17888884 0.17709524 0.17456159 0.17247714
#> [29,] 0.17477635 0.17456159 0.17389488 0.17272016 0.17097366 0.16860849
#> [30,] 0.17071654 0.17045781 0.16967283 0.16833969 0.16643411 0.16394052
#> [31,] 0.16560718 0.16533047 0.16449766 0.16310249 0.16113972 0.15832309
#> [32,] 0.15973451 0.15938135 0.15832309 0.15656413 0.15411369 0.15098765
#> [33,] 0.15167975 0.15133351 0.15029709 0.14857766 0.14618776 0.14314586
#> [34,] 0.14314586 0.14281006 0.14180530 0.14013959 0.13782636 0.13488438
#> [35,] 0.13423551 0.13397046 0.13317732 0.13186208 0.13003472 0.12770909
#> [36,] 0.12668401 0.12637326 0.12544353 0.12390225 0.12176163 0.11903827
#> [37,] 0.11724040 0.11704657 0.11646643 0.11550404 0.11416607 0.11246162
#> [38,] 0.11096121 0.11062615 0.10962345 0.10796045 0.10564912 0.10270571
#> [39,] 0.09947161 0.09919728 0.09837615 0.09701378 0.09511924 0.09270487
#> [40,] 0.08952224 0.08845290 0.08525201 0.07994061 0.07255312 0.06313622
#> [41,] 0.04866412         NA         NA         NA         NA         NA
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,]         NA         NA         NA         NA         NA         NA
#>  [2,] 0.05174757 0.04866412         NA         NA         NA         NA
#>  [3,] 0.08978589 0.07676768 0.06105577 0.04866412         NA         NA
#>  [4,] 0.09919728 0.09565903 0.09163941 0.07994061 0.06001719 0.04866412
#>  [5,] 0.10995727 0.10564912 0.10076073 0.09592938 0.09084334 0.07255312
#>  [6,] 0.11627350 0.11378577 0.11096121 0.10532051 0.09919728 0.09350701
#>  [7,] 0.12451753 0.12054728 0.11646643 0.11321698 0.10862443 0.10173151
#>  [8,] 0.13186208 0.12848134 0.12420969 0.11903827 0.11473816 0.11062615
#>  [9,] 0.13947643 0.13520948 0.13107693 0.12668401 0.12085026 0.11550404
#> [10,] 0.14720948 0.14281006 0.13782636 0.13265022 0.12770909 0.12145744
#> [11,] 0.15446280 0.14995241 0.14483146 0.13914552 0.13317732 0.12770909
#> [12,] 0.16085879 0.15656413 0.15133351 0.14550875 0.13914552 0.13265022
#> [13,] 0.16560718 0.16198185 0.15726693 0.15133351 0.14483146 0.13782636
#> [14,] 0.16967283 0.16615904 0.16198185 0.15656413 0.14995241 0.14281006
#> [15,] 0.17296068 0.16967283 0.16560718 0.16085879 0.15446280 0.14720948
#> [16,] 0.17595458 0.17247714 0.16860849 0.16394052 0.15832309 0.15098765
#> [17,] 0.17922782 0.17456159 0.17097366 0.16643411 0.16113972 0.15411369
#> [18,] 0.18093833 0.17709524 0.17272016 0.16833969 0.16310249 0.15656413
#> [19,] 0.18156066 0.17888884 0.17389488 0.16967283 0.16449766 0.15832309
#> [20,] 0.18171552 0.17988207 0.17456159 0.17045781 0.16533047 0.15938135
#> [21,] 0.18172140 0.18019619 0.17477635 0.17071654 0.16560718 0.15973451
#> [22,] 0.18171552 0.17988207 0.17456159 0.17045781 0.16533047 0.15938135
#> [23,] 0.18156066 0.17888884 0.17389488 0.16967283 0.16449766 0.15832309
#> [24,] 0.18093833 0.17709524 0.17272016 0.16833969 0.16310249 0.15656413
#> [25,] 0.17922782 0.17456159 0.17097366 0.16643411 0.16113972 0.15411369
#> [26,] 0.17595458 0.17247714 0.16860849 0.16394052 0.15832309 0.15098765
#> [27,] 0.17296068 0.16967283 0.16560718 0.16085879 0.15446280 0.14720948
#> [28,] 0.16967283 0.16615904 0.16198185 0.15656413 0.14995241 0.14281006
#> [29,] 0.16560718 0.16198185 0.15726693 0.15133351 0.14483146 0.13782636
#> [30,] 0.16085879 0.15656413 0.15133351 0.14550875 0.13914552 0.13265022
#> [31,] 0.15446280 0.14995241 0.14483146 0.13914552 0.13317732 0.12770909
#> [32,] 0.14720948 0.14281006 0.13782636 0.13265022 0.12770909 0.12145744
#> [33,] 0.13947643 0.13520948 0.13107693 0.12668401 0.12085026 0.11550404
#> [34,] 0.13186208 0.12848134 0.12420969 0.11903827 0.11473816 0.11062615
#> [35,] 0.12451753 0.12054728 0.11646643 0.11321698 0.10862443 0.10173151
#> [36,] 0.11627350 0.11378577 0.11096121 0.10532051 0.09919728 0.09350701
#> [37,] 0.10995727 0.10564912 0.10076073 0.09592938 0.09084334 0.07255312
#> [38,] 0.09919728 0.09565903 0.09163941 0.07994061 0.06001719 0.04866412
#> [39,] 0.08978589 0.07676768 0.06105577 0.04866412         NA         NA
#> [40,] 0.05174757 0.04866412         NA         NA         NA         NA
#> [41,]         NA         NA         NA         NA         NA         NA
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,]         NA         NA         NA         NA         NA         NA
#>  [2,]         NA         NA         NA         NA         NA         NA
#>  [3,]         NA         NA         NA         NA         NA         NA
#>  [4,]         NA         NA         NA         NA         NA         NA
#>  [5,] 0.04866412         NA         NA         NA         NA         NA
#>  [6,] 0.08100056 0.05484060 0.04866412         NA         NA         NA
#>  [7,] 0.09511924 0.08525201 0.05690799 0.04866412         NA         NA
#>  [8,] 0.10303121 0.09565903 0.08525201 0.05484060 0.04866412         NA
#>  [9,] 0.11114801 0.10303121 0.09511924 0.08100056 0.04866412 0.04866412
#> [10,] 0.11550404 0.11062615 0.10173151 0.09350701 0.07255312 0.04866412
#> [11,] 0.12085026 0.11473816 0.10862443 0.09919728 0.09084334 0.06001719
#> [12,] 0.12668401 0.11903827 0.11321698 0.10532051 0.09592938 0.07994061
#> [13,] 0.13107693 0.12420969 0.11646643 0.11096121 0.10076073 0.09163941
#> [14,] 0.13520948 0.12848134 0.12054728 0.11378577 0.10564912 0.09565903
#> [15,] 0.13947643 0.13186208 0.12451753 0.11627350 0.10995727 0.09919728
#> [16,] 0.14314586 0.13488438 0.12770909 0.11903827 0.11246162 0.10270571
#> [17,] 0.14618776 0.13782636 0.13003472 0.12176163 0.11416607 0.10564912
#> [18,] 0.14857766 0.14013959 0.13186208 0.12390225 0.11550404 0.10796045
#> [19,] 0.15029709 0.14180530 0.13317732 0.12544353 0.11646643 0.10962345
#> [20,] 0.15133351 0.14281006 0.13397046 0.12637326 0.11704657 0.11062615
#> [21,] 0.15167975 0.14314586 0.13423551 0.12668401 0.11724040 0.11096121
#> [22,] 0.15133351 0.14281006 0.13397046 0.12637326 0.11704657 0.11062615
#> [23,] 0.15029709 0.14180530 0.13317732 0.12544353 0.11646643 0.10962345
#> [24,] 0.14857766 0.14013959 0.13186208 0.12390225 0.11550404 0.10796045
#> [25,] 0.14618776 0.13782636 0.13003472 0.12176163 0.11416607 0.10564912
#> [26,] 0.14314586 0.13488438 0.12770909 0.11903827 0.11246162 0.10270571
#> [27,] 0.13947643 0.13186208 0.12451753 0.11627350 0.10995727 0.09919728
#> [28,] 0.13520948 0.12848134 0.12054728 0.11378577 0.10564912 0.09565903
#> [29,] 0.13107693 0.12420969 0.11646643 0.11096121 0.10076073 0.09163941
#> [30,] 0.12668401 0.11903827 0.11321698 0.10532051 0.09592938 0.07994061
#> [31,] 0.12085026 0.11473816 0.10862443 0.09919728 0.09084334 0.06001719
#> [32,] 0.11550404 0.11062615 0.10173151 0.09350701 0.07255312 0.04866412
#> [33,] 0.11114801 0.10303121 0.09511924 0.08100056 0.04866412         NA
#> [34,] 0.10303121 0.09565903 0.08525201 0.05484060 0.04866412         NA
#> [35,] 0.09511924 0.08525201 0.05690799 0.04866412         NA         NA
#> [36,] 0.08100056 0.05484060 0.04866412         NA         NA         NA
#> [37,] 0.04866412         NA         NA         NA         NA         NA
#> [38,]         NA         NA         NA         NA         NA         NA
#> [39,]         NA         NA         NA         NA         NA         NA
#> [40,]         NA         NA         NA         NA         NA         NA
#> [41,]         NA         NA         NA         NA         NA         NA
#>            [,19]      [,20]      [,21]
#>  [1,]         NA         NA         NA
#>  [2,]         NA         NA         NA
#>  [3,]         NA         NA         NA
#>  [4,]         NA         NA         NA
#>  [5,]         NA         NA         NA
#>  [6,]         NA         NA         NA
#>  [7,]         NA         NA         NA
#>  [8,]         NA         NA         NA
#>  [9,]         NA         NA         NA
#> [10,]         NA         NA         NA
#> [11,] 0.04866412         NA         NA
#> [12,] 0.04866412         NA         NA
#> [13,] 0.06105577 0.04866412         NA
#> [14,] 0.07676768 0.04866412         NA
#> [15,] 0.08978589 0.05174757 0.04866412
#> [16,] 0.09270487 0.06313622 0.04866412
#> [17,] 0.09511924 0.07255312 0.04866412
#> [18,] 0.09701378 0.07994061 0.04866412
#> [19,] 0.09837615 0.08525201 0.04866412
#> [20,] 0.09919728 0.08845290 0.04866412
#> [21,] 0.09947161 0.08952224 0.04866412
#> [22,] 0.09919728 0.08845290 0.04866412
#> [23,] 0.09837615 0.08525201 0.04866412
#> [24,] 0.09701378 0.07994061 0.04866412
#> [25,] 0.09511924 0.07255312 0.04866412
#> [26,] 0.09270487 0.06313622 0.04866412
#> [27,] 0.08978589 0.05174757 0.04866412
#> [28,] 0.07676768 0.04866412         NA
#> [29,] 0.06105577 0.04866412         NA
#> [30,] 0.04866412         NA         NA
#> [31,] 0.04866412         NA         NA
#> [32,]         NA         NA         NA
#> [33,]         NA         NA         NA
#> [34,]         NA         NA         NA
#> [35,]         NA         NA         NA
#> [36,]         NA         NA         NA
#> [37,]         NA         NA         NA
#> [38,]         NA         NA         NA
#> [39,]         NA         NA         NA
#> [40,]         NA         NA         NA
#> [41,]         NA         NA         NA