Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 89.40439
#> 2   Feb 94.42437
#> 3   Mar 88.76919
#> 4   Apr 85.08815
#> 5   May 90.65445
#> 6   Jun 89.64773

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 6.429564  2
#> 2   Feb 9.857018  2
#> 3   Mar 9.556768  2
#> 4   Apr 6.271959  2
#> 5   May 8.550287  2
#> 6   Jun 9.241991  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [11ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [21ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [63ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [696ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [19ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1887883    0.1950257  0.2046244
#>    [2,]    0.1952375    0.2016880  0.2116145
#>    [3,]    0.2174206    0.2246040  0.2356584
#>    [4,]    0.1828474    0.1888885  0.1981852
#>    [5,]    0.1997230    0.2063217  0.2164764
#>    [6,]    0.2619686    0.2706236  0.2839427
#>    [7,]    0.2168507    0.2240152  0.2350407
#>    [8,]    0.1685418    0.1741103  0.1826797
#>    [9,]    0.2374080    0.2452515  0.2573220
#>   [10,]    0.3671460    0.3792754  0.3979413
#>   [11,]    0.2919247    0.3015692  0.3164110
#>   [12,]    0.2283829    0.2359283  0.2475400
#>   [13,]    0.1878823    0.1940899  0.2036427
#>   [14,]    0.1750829    0.1808676  0.1897695
#>   [15,]    0.2566429    0.2651219  0.2781703
#>   [16,]    0.2096646    0.2165916  0.2272517
#>   [17,]    0.2373399    0.2451813  0.2572482
#>   [18,]    0.2255654    0.2330178  0.2444863
#>   [19,]    0.1410673    0.1457281  0.1529007
#>   [20,]    0.2416332    0.2496165  0.2619019
#>   [21,]    0.1725511    0.1782521  0.1870253
#>   [22,]    0.2376460    0.2454975  0.2575800
#>   [23,]    0.2034426    0.2101642  0.2205080
#>   [24,]    0.1969333    0.2034398  0.2134527
#>   [25,]    0.2178673    0.2250654  0.2361424
#>   [26,]    0.1809469    0.1869253  0.1961254
#>   [27,]    0.2627026    0.2713819  0.2847384
#>   [28,]    0.2300144    0.2376138  0.2493085
#>   [29,]    0.1953590    0.2018134  0.2117462
#>   [30,]    0.1647943    0.1702388  0.1786174
#>   [31,]    0.2467937    0.2549474  0.2674951
#>   [32,]    0.2336732    0.2413935  0.2532743
#>   [33,]    0.2374449    0.2452896  0.2573619
#>   [34,]    0.2997511    0.3096543  0.3248941
#>   [35,]    0.1976725    0.2042034  0.2142538
#>   [36,]    0.2106445    0.2176039  0.2283137
#>   [37,]    0.2136511    0.2207099  0.2315727
#>   [38,]    0.1442934    0.1490608  0.1563972
#>   [39,]    0.2224560    0.2298057  0.2411161
#>   [40,]    0.3111116    0.3213899  0.3372072
#>   [41,]    0.2165210    0.2236747  0.2346835
#>   [42,]    0.1925616    0.1989237  0.2087143
#>   [43,]    0.3390317    0.3502325  0.3674694
#>   [44,]    0.1800809    0.1860306  0.1951865
#>   [45,]    0.2428787    0.2509031  0.2632518
#>   [46,]    0.1422958    0.1469972  0.1542322
#>   [47,]    0.2814480    0.2907466  0.3050560
#>   [48,]    0.1692379    0.1748294  0.1834342
#>   [49,]    0.2086210    0.2155136  0.2261207
#>   [50,]    0.1869213    0.1930970  0.2026006
#>   [51,]    0.1949854    0.2014275  0.2113413
#>   [52,]    0.2124191    0.2194372  0.2302373
#>   [53,]    0.2155282    0.2226490  0.2336071
#>   [54,]    0.2514315    0.2597384  0.2725219
#>   [55,]    0.2558348    0.2642873  0.2772947
#>   [56,]    0.2811943    0.2904845  0.3047810
#>   [57,]    0.2502342    0.2585016  0.2712242
#>   [58,]    0.1765928    0.1824273  0.1914059
#>   [59,]    0.2297416    0.2373319  0.2490126
#>   [60,]    0.1653726    0.1708364  0.1792447
#>   [61,]    0.2496946    0.2579441  0.2706392
#>   [62,]    0.1945173    0.2009439  0.2108339
#>   [63,]    0.1624628    0.1678306  0.1760910
#>   [64,]    0.2398973    0.2478232  0.2600203
#>   [65,]    0.1877304    0.1939328  0.2034778
#>   [66,]    0.2318185    0.2394775  0.2512638
#>   [67,]    0.2238962    0.2312935  0.2426772
#>   [68,]    0.2245529    0.2319718  0.2433888
#>   [69,]    0.1886225    0.1948545  0.2044448
#>   [70,]    0.2441987    0.2522666  0.2646823
#>   [71,]    0.1789833    0.1848968  0.1939970
#>   [72,]    0.1794656    0.1853949  0.1945196
#>   [73,]    0.2232466    0.2306224  0.2419731
#>   [74,]    0.2230968    0.2304677  0.2418108
#>   [75,]    0.2630738    0.2717654  0.2851408
#>   [76,]    0.1869857    0.1931636  0.2026707
#>   [77,]    0.2411338    0.2491005  0.2613604
#>   [78,]    0.1793206    0.1852452  0.1943626
#>   [79,]    0.2061784    0.2129904  0.2234733
#>   [80,]    0.1902424    0.1965278  0.2062004
#>   [81,]    0.1821170    0.1881340  0.1973937
#>   [82,]    0.2730044    0.2820240  0.2959041
#>   [83,]    0.3009166    0.3108584  0.3261578
#>   [84,]    0.1790798    0.1849965  0.1941016
#>   [85,]    0.1805447    0.1865098  0.1956895
#>   [86,]    0.1706546    0.1762929  0.1849697
#>   [87,]    0.2329124    0.2406075  0.2524495
#>   [88,]    0.1844897    0.1905851  0.1999653
#>   [89,]    0.2868284    0.2963048  0.3108880
#>   [90,]    0.1358297    0.1403175  0.1472237
#>   [91,]    0.2980670    0.3079145  0.3230689
#>   [92,]    0.1487911    0.1537071  0.1612723
#>   [93,]    0.2181689    0.2253770  0.2364695
#>   [94,]    0.2342988    0.2420396  0.2539519
#>   [95,]    0.2368976    0.2447244  0.2567690
#>   [96,]    0.2153471    0.2224619  0.2334109
#>   [97,]    0.1899981    0.1962755  0.2059357
#>   [98,]    0.2353049    0.2430790  0.2550426
#>   [99,]    0.1959933    0.2024687  0.2124337
#>  [100,]    0.2093270    0.2162430  0.2268859
#>  [101,]    0.2218982    0.2292294  0.2405114
#>  [102,]    0.1822777    0.1883000  0.1975678
#>  [103,]    0.2032565    0.2099718  0.2203058
#>  [104,]    0.1229134    0.1269744  0.1332239
#>  [105,]    0.2511082    0.2594045  0.2721717
#>  [106,]    0.2112205    0.2181989  0.2289380
#>  [107,]    0.1796679    0.1856040  0.1947390
#>  [108,]    0.1098401    0.1134692  0.1190540
#>  [109,]    0.2483650    0.2565707  0.2691985
#>  [110,]    0.2134158    0.2204668  0.2313175
#>  [111,]    0.2799446    0.2891935  0.3034265
#>  [112,]    0.3373753    0.3485212  0.3656737
#>  [113,]    0.2033540    0.2100726  0.2204118
#>  [114,]    0.2029994    0.2097064  0.2200276
#>  [115,]    0.1907299    0.1970314  0.2067288
#>  [116,]    0.1366871    0.1412032  0.1481530
#>  [117,]    0.1841043    0.1901869  0.1995476
#>  [118,]    0.1625145    0.1678839  0.1761469
#>  [119,]    0.1901818    0.1964651  0.2061346
#>  [120,]    0.1737369    0.1794770  0.1883105
#>  [121,]    0.2076719    0.2145331  0.2250920
#>  [122,]    0.1694971    0.1750971  0.1837151
#>  [123,]    0.1513655    0.1563665  0.1640626
#>  [124,]    0.2492156    0.2574491  0.2701196
#>  [125,]    0.2983673    0.3082247  0.3233944
#>  [126,]    0.2499655    0.2582241  0.2709332
#>  [127,]    0.2109017    0.2178697  0.2285926
#>  [128,]    0.2625138    0.2711868  0.2845337
#>  [129,]    0.4850240    0.5010476  0.5257062
#>  [130,]    0.2150838    0.2221899  0.2331256
#>  [131,]    0.2080520    0.2149258  0.2255038
#>  [132,]    0.2065634    0.2133879  0.2238902
#>  [133,]    0.1735858    0.1793210  0.1881468
#>  [134,]    0.1478212    0.1527051  0.1602210
#>  [135,]    0.2250916    0.2325284  0.2439729
#>  [136,]    0.2100777    0.2170184  0.2276995
#>  [137,]    0.2540669    0.2624609  0.2753785
#>  [138,]    0.2094495    0.2163695  0.2270186
#>  [139,]    0.1654545    0.1709210  0.1793334
#>  [140,]    0.1522636    0.1572943  0.1650361
#>  [141,]    0.2089740    0.2158782  0.2265031
#>  [142,]    0.1462572    0.1510895  0.1585259
#>  [143,]    0.1434457    0.1481851  0.1554785
#>  [144,]    0.1551216    0.1602467  0.1681338
#>  [145,]    0.2866696    0.2961406  0.3107156
#>  [146,]    0.2403828    0.2483246  0.2605462
#>  [147,]    0.2347178    0.2424725  0.2544061
#>  [148,]    0.1745985    0.1803671  0.1892445
#>  [149,]    0.1913592    0.1976816  0.2074110
#>  [150,]    0.2769397    0.2860894  0.3001698
#>  [151,]    0.2180113    0.2252141  0.2362984
#>  [152,]    0.1707888    0.1764315  0.1851152
#>  [153,]    0.2000470    0.2066563  0.2168274
#>  [154,]    0.2733907    0.2824231  0.2963230
#>  [155,]    0.2015789    0.2082390  0.2184882
#>  [156,]    0.3249273    0.3356621  0.3521818
#>  [157,]    0.1358226    0.1403101  0.1472160
#>  [158,]    0.2034973    0.2102207  0.2205672
#>  [159,]    0.1904260    0.1967175  0.2063994
#>  [160,]    0.2361903    0.2439937  0.2560024
#>  [161,]    0.1504479    0.1554186  0.1630681
#>  [162,]    0.2454206    0.2535289  0.2660069
#>  [163,]    0.3064621    0.3165870  0.3321681
#>  [164,]    0.1979981    0.2045398  0.2146069
#>  [165,]    0.1781545    0.1840406  0.1930988
#>  [166,]    0.3010799    0.3110270  0.3263345
#>  [167,]    0.2305527    0.2381698  0.2498919
#>  [168,]    0.2102208    0.2171663  0.2278547
#>  [169,]    0.2332271    0.2409326  0.2527906
#>  [170,]    0.1912895    0.1976095  0.2073353
#>  [171,]    0.2732369    0.2822642  0.2961565
#>  [172,]    0.1779358    0.1838147  0.1928617
#>  [173,]    0.2226055    0.2299602  0.2412782
#>  [174,]    0.2256964    0.2331530  0.2446281
#>  [175,]    0.2336579    0.2413776  0.2532574
#>  [176,]    0.2599526    0.2685410  0.2817577
#>  [177,]    0.2233774    0.2307576  0.2421148
#>  [178,]    0.2875929    0.2970944  0.3117162
#>  [179,]    0.1981454    0.2046919  0.2147663
#>  [180,]    0.1892771    0.1955307  0.2051543
#>  [181,]    0.1834577    0.1895190  0.1988467
#>  [182,]    0.1895290    0.1957909  0.2054272
#>  [183,]    0.1913892    0.1977125  0.2074434
#>  [184,]    0.1325939    0.1369748  0.1437165
#>  [185,]    0.2119589    0.2189617  0.2297384
#>  [186,]    0.2539163    0.2623052  0.2752150
#>  [187,]    0.1789149    0.1848260  0.1939227
#>  [188,]    0.2069444    0.2137817  0.2243035
#>  [189,]    0.2319331    0.2395959  0.2513881
#>  [190,]    0.2047889    0.2115549  0.2219671
#>  [191,]    0.1769403    0.1827862  0.1917826
#>  [192,]    0.2000914    0.2067022  0.2168756
#>  [193,]    0.1731546    0.1788755  0.1876794
#>  [194,]    0.1833103    0.1893667  0.1986868
#>  [195,]    0.2003860    0.2070065  0.2171948
#>  [196,]    0.2176397    0.2248303  0.2358959
#>  [197,]    0.1966731    0.2031710  0.2131707
#>  [198,]    0.2695542    0.2784599  0.2921648
#>  [199,]    0.2028757    0.2095785  0.2198935
#>  [200,]    0.2007428    0.2073752  0.2175818
#>  [201,]    0.2017868    0.2084536  0.2187132
#>  [202,]    0.1673868    0.1729171  0.1814276
#>  [203,]    0.2564549    0.2649278  0.2779667
#>  [204,]    0.1547968    0.1599112  0.1677816
#>  [205,]    0.2559625    0.2644192  0.2774332
#>  [206,]    0.2528587    0.2612127  0.2740688
#>  [207,]    0.2519261    0.2602493  0.2730578
#>  [208,]    0.2725008    0.2815038  0.2953585
#>  [209,]    0.1954039    0.2018599  0.2117950
#>  [210,]    0.2088216    0.2157208  0.2263380
#>  [211,]    0.2075498    0.2144071  0.2249597
#>  [212,]    0.2073926    0.2142447  0.2247893
#>  [213,]    0.1651143    0.1705694  0.1789644
#>  [214,]    0.1749426    0.1807226  0.1896174
#>  [215,]    0.2276799    0.2352023  0.2467784
#>  [216,]    0.2224447    0.2297940  0.2411038
#>  [217,]    0.2007108    0.2073420  0.2175468
#>  [218,]    0.2228447    0.2302072  0.2415374
#>  [219,]    0.1935392    0.1999335  0.2097738
#>  [220,]    0.2307248    0.2383476  0.2500785
#>  [221,]    0.1979051    0.2044437  0.2145059
#>  [222,]    0.2090720    0.2159795  0.2266095
#>  [223,]    0.2785291    0.2877314  0.3018927
#>  [224,]    0.2011831    0.2078300  0.2180588
#>  [225,]    0.2033081    0.2100252  0.2203621
#>  [226,]    0.2001012    0.2067123  0.2168862
#>  [227,]    0.1985134    0.2050721  0.2151653
#>  [228,]    0.2425134    0.2505258  0.2628560
#>  [229,]    0.2095467    0.2164700  0.2271241
#>  [230,]    0.1692644    0.1748568  0.1834629
#>  [231,]    0.2596933    0.2682732  0.2814766
#>  [232,]    0.2198943    0.2271591  0.2383390
#>  [233,]    0.1922335    0.1985847  0.2083587
#>  [234,]    0.2621222    0.2707824  0.2841094
#>  [235,]    0.2556971    0.2641450  0.2771456
#>  [236,]    0.2436330    0.2516823  0.2640693
#>  [237,]    0.3304409    0.3413579  0.3581581
#>  [238,]    0.2561179    0.2645797  0.2776014
#>  [239,]    0.2944026    0.3041292  0.3190974
#>  [240,]    0.2865414    0.2960081  0.3105765
#>  [241,]    0.2015138    0.2081716  0.2184173
#>  [242,]    0.1795256    0.1854569  0.1945847
#>  [243,]    0.2620570    0.2707148  0.2840383
#>  [244,]    0.2728290    0.2818428  0.2957140
#>  [245,]    0.1585483    0.1637867  0.1718480
#>  [246,]    0.2873187    0.2968113  0.3114193
#>  [247,]    0.2656235    0.2743993  0.2879042
#>  [248,]    0.1861013    0.1922500  0.2017121
#>  [249,]    0.1995494    0.2061424  0.2162883
#>  [250,]    0.2143644    0.2214468  0.2323458
#>  [251,]    0.2297593    0.2373502  0.2490319
#>  [252,]    0.1636649    0.1690723  0.1773937
#>  [253,]    0.1891597    0.1954093  0.2050269
#>  [254,]    0.2385392    0.2464202  0.2585483
#>  [255,]    0.2589100    0.2674638  0.2806272
#>  [256,]    0.1663758    0.1718727  0.1803320
#>  [257,]    0.2010008    0.2076416  0.2178612
#>  [258,]    0.1872312    0.1934172  0.2029368
#>  [259,]    0.2424324    0.2504421  0.2627681
#>  [260,]    0.2159922    0.2231284  0.2341103
#>  [261,]    0.2228824    0.2302462  0.2415784
#>  [262,]    0.2938105    0.3035174  0.3184554
#>  [263,]    0.2901878    0.2997749  0.3145286
#>  [264,]    0.1615824    0.1669210  0.1751365
#>  [265,]    0.1436114    0.1483562  0.1556581
#>  [266,]    0.2068637    0.2136983  0.2242160
#>  [267,]    0.1600381    0.1653257  0.1734627
#>  [268,]    0.2220082    0.2293430  0.2406306
#>  [269,]    0.2586548    0.2672004  0.2803510
#>  [270,]    0.1697721    0.1753811  0.1840130
#>  [271,]    0.1880804    0.1942944  0.2038572
#>  [272,]    0.2865731    0.2960411  0.3106113
#>  [273,]    0.1672882    0.1728152  0.1813208
#>  [274,]    0.1789918    0.1849056  0.1940063
#>  [275,]    0.2378907    0.2457503  0.2578454
#>  [276,]    0.2672968    0.2761279  0.2897180
#>  [277,]    0.2831108    0.2924642  0.3068581
#>  [278,]    0.1765577    0.1823911  0.1913680
#>  [279,]    0.1846504    0.1907511  0.2001395
#>  [280,]    0.2684480    0.2773171  0.2909656
#>  [281,]    0.1767523    0.1825920  0.1915788
#>  [282,]    0.1804110    0.1863716  0.1955444
#>  [283,]    0.1854877    0.1916160  0.2010469
#>  [284,]    0.2510295    0.2593231  0.2720861
#>  [285,]    0.2747092    0.2837852  0.2977520
#>  [286,]    0.1681303    0.1736852  0.1822335
#>  [287,]    0.2004161    0.2070376  0.2172273
#>  [288,]    0.2192669    0.2265113  0.2376595
#>  [289,]    0.2293033    0.2368792  0.2485378
#>  [290,]    0.2328411    0.2405339  0.2523722
#>  [291,]    0.1909678    0.1972772  0.2069867
#>  [292,]    0.2089467    0.2158500  0.2264737
#>  [293,]    0.2043078    0.2110579  0.2214456
#>  [294,]    0.2076190    0.2144784  0.2250345
#>  [295,]    0.1972869    0.2038051  0.2138359
#>  [296,]    0.1965756    0.2030703  0.2130650
#>  [297,]    0.2556216    0.2640670  0.2770636
#>  [298,]    0.2009372    0.2075760  0.2177924
#>  [299,]    0.2078404    0.2147072  0.2252745
#>  [300,]    0.2342033    0.2419410  0.2538486
#>  [301,]    0.3013140    0.3112688  0.3265881
#>  [302,]    0.1395779    0.1441894  0.1512862
#>  [303,]    0.2645516    0.2732920  0.2867426
#>  [304,]    0.2176545    0.2248455  0.2359118
#>  [305,]    0.1852524    0.1913730  0.2007920
#>  [306,]    0.1929737    0.1993494  0.2091609
#>  [307,]    0.2765159    0.2856514  0.2997100
#>  [308,]    0.2505336    0.2588109  0.2715486
#>  [309,]    0.2287486    0.2363061  0.2479364
#>  [310,]    0.1871949    0.1933796  0.2028973
#>  [311,]    0.2845768    0.2939787  0.3084471
#>  [312,]    0.1857481    0.1918850  0.2013291
#>  [313,]    0.2397854    0.2477076  0.2598990
#>  [314,]    0.1967474    0.2032477  0.2132510
#>  [315,]    0.1576954    0.1629056  0.1709236
#>  [316,]    0.1767810    0.1826217  0.1916099
#>  [317,]    0.1968132    0.2033158  0.2133226
#>  [318,]    0.1607870    0.1660993  0.1742744
#>  [319,]    0.1163728    0.1202177  0.1261346
#>  [320,]    0.2016256    0.2082871  0.2185385
#>  [321,]    0.1935426    0.1999369  0.2097772
#>  [322,]    0.2542773    0.2626782  0.2756063
#>  [323,]    0.2223491    0.2296953  0.2410003
#>  [324,]    0.2007079    0.2073391  0.2175437
#>  [325,]    0.2109008    0.2178687  0.2285915
#>  [326,]    0.1842048    0.1902908  0.1996565
#>  [327,]    0.1584429    0.1636777  0.1717336
#>  [328,]    0.2224674    0.2298175  0.2411284
#>  [329,]    0.2182280    0.2254380  0.2365335
#>  [330,]    0.1680655    0.1736183  0.1821634
#>  [331,]    0.1615786    0.1669171  0.1751325
#>  [332,]    0.2123222    0.2193371  0.2301323
#>  [333,]    0.2885554    0.2980888  0.3127597
#>  [334,]    0.2227273    0.2300860  0.2414102
#>  [335,]    0.1962384    0.2027219  0.2126994
#>  [336,]    0.2427041    0.2507228  0.2630626
#>  [337,]    0.1454976    0.1503047  0.1577024
#>  [338,]    0.1817120    0.1877157  0.1969547
#>  [339,]    0.1871171    0.1932993  0.2028130
#>  [340,]    0.2072275    0.2140742  0.2246104
#>  [341,]    0.1574420    0.1626438  0.1706488
#>  [342,]    0.2143554    0.2214374  0.2323359
#>  [343,]    0.2407666    0.2487212  0.2609625
#>  [344,]    0.2463753    0.2545152  0.2670416
#>  [345,]    0.2249639    0.2323965  0.2438345
#>  [346,]    0.2231891    0.2305631  0.2419108
#>  [347,]    0.1913795    0.1977025  0.2074330
#>  [348,]    0.1959149    0.2023877  0.2123487
#>  [349,]    0.2370490    0.2448808  0.2569331
#>  [350,]    0.2288876    0.2364497  0.2480870
#>  [351,]    0.1651110    0.1705661  0.1789611
#>  [352,]    0.1878741    0.1940812  0.2036333
#>  [353,]    0.2264241    0.2339049  0.2454171
#>  [354,]    0.2767585    0.2859022  0.2999733
#>  [355,]    0.2381900    0.2460596  0.2581700
#>  [356,]    0.1659146    0.1713963  0.1798321
#>  [357,]    0.2071292    0.2139725  0.2245036
#>  [358,]    0.1935370    0.1999313  0.2097714
#>  [359,]    0.1657489    0.1712251  0.1796525
#>  [360,]    0.1360935    0.1405899  0.1475096
#>  [361,]    0.2258509    0.2333127  0.2447956
#>  [362,]    0.2351594    0.2429288  0.2548850
#>  [363,]    0.2365131    0.2443272  0.2563523
#>  [364,]    0.2055135    0.2123034  0.2227525
#>  [365,]    0.2048318    0.2115992  0.2220137
#>  [366,]    0.2493285    0.2575659  0.2702426
#>  [367,]    0.1398791    0.1445006  0.1516127
#>  [368,]    0.2117286    0.2187239  0.2294890
#>  [369,]    0.1524583    0.1574954  0.1652471
#>  [370,]    0.2068940    0.2137297  0.2242491
#>  [371,]    0.2705205    0.2794579  0.2932116
#>  [372,]    0.1277177    0.1319374  0.1384312
#>  [373,]    0.2516083    0.2599210  0.2727134
#>  [374,]    0.1899931    0.1962704  0.2059304
#>  [375,]    0.2001687    0.2067821  0.2169595
#>  [376,]    0.2355481    0.2433302  0.2553060
#>  [377,]    0.2469228    0.2550808  0.2676350
#>  [378,]    0.1497615    0.1547096  0.1623242
#>  [379,]    0.2144949    0.2215816  0.2324873
#>  [380,]    0.2173153    0.2244952  0.2355443
#>  [381,]    0.2603724    0.2689747  0.2822128
#>  [382,]    0.1425797    0.1472904  0.1545398
#>  [383,]    0.1993931    0.2059808  0.2161187
#>  [384,]    0.2019883    0.2086617  0.2189314
#>  [385,]    0.3090998    0.3193118  0.3350269
#>  [386,]    0.1511191    0.1561120  0.1637957
#>  [387,]    0.3195773    0.3301353  0.3463830
#>  [388,]    0.2034124    0.2101328  0.2204750
#>  [389,]    0.2071075    0.2139501  0.2244801
#>  [390,]    0.2941170    0.3038342  0.3187880
#>  [391,]    0.2079930    0.2148648  0.2254398
#>  [392,]    0.2807479    0.2900234  0.3042974
#>  [393,]    0.1987867    0.2053544  0.2154614
#>  [394,]    0.2157052    0.2228319  0.2337992
#>  [395,]    0.1677766    0.1733198  0.1818502
#>  [396,]    0.2929702    0.3026493  0.3175444
#>  [397,]    0.1926435    0.1990082  0.2088029
#>  [398,]    0.1832224    0.1892759  0.1985917
#>  [399,]    0.2012963    0.2079470  0.2181817
#>  [400,]    0.1732464    0.1789704  0.1877789
#>  [401,]    0.1952969    0.2017493  0.2116788
#>  [402,]    0.1847123    0.1908150  0.2002064
#>  [403,]    0.1716144    0.1772844  0.1860100
#>  [404,]    0.2662306    0.2750264  0.2885621
#>  [405,]    0.1911258    0.1974404  0.2071580
#>  [406,]    0.1645169    0.1699524  0.1783170
#>  [407,]    0.2114386    0.2184243  0.2291745
#>  [408,]    0.1805061    0.1864699  0.1956475
#>  [409,]    0.1992891    0.2058734  0.2160060
#>  [410,]    0.1645998    0.1700380  0.1784069
#>  [411,]    0.1827864    0.1888255  0.1981191
#>  [412,]    0.2014316    0.2080867  0.2183281
#>  [413,]    0.2124464    0.2194653  0.2302668
#>  [414,]    0.1731222    0.1788420  0.1876443
#>  [415,]    0.3048682    0.3149405  0.3304408
#>  [416,]    0.1392487    0.1438494  0.1509294
#>  [417,]    0.1569888    0.1621756  0.1701576
#>  [418,]    0.1811911    0.1871775  0.1963899
#>  [419,]    0.1943330    0.2007536  0.2106342
#>  [420,]    0.1676677    0.1732073  0.1817322
#>  [421,]    0.2193405    0.2265873  0.2377394
#>  [422,]    0.2384511    0.2463291  0.2584527
#>  [423,]    0.2359948    0.2437918  0.2557906
#>  [424,]    0.1975273    0.2040534  0.2140964
#>  [425,]    0.2142115    0.2212888  0.2321801
#>  [426,]    0.2145649    0.2216539  0.2325632
#>  [427,]    0.1724193    0.1781158  0.1868822
#>  [428,]    0.1512763    0.1562743  0.1639658
#>  [429,]    0.2516430    0.2599569  0.2727511
#>  [430,]    0.1490995    0.1540257  0.1616065
#>  [431,]    0.2977985    0.3076371  0.3227778
#>  [432,]    0.2096966    0.2166247  0.2272864
#>  [433,]    0.2854118    0.2948413  0.3093524
#>  [434,]    0.2554950    0.2639361  0.2769260
#>  [435,]    0.2823349    0.2916627  0.3060173
#>  [436,]    0.1924070    0.1987639  0.2085466
#>  [437,]    0.2648372    0.2735870  0.2870520
#>  [438,]    0.2086485    0.2155420  0.2261505
#>  [439,]    0.2062208    0.2130342  0.2235193
#>  [440,]    0.1999809    0.2065881  0.2167558
#>  [441,]    0.2083629    0.2152470  0.2258410
#>  [442,]    0.2064025    0.2132217  0.2237159
#>  [443,]    0.2748922    0.2839743  0.2979506
#>  [444,]    0.1663009    0.1717954  0.1802509
#>  [445,]    0.2055317    0.2123222  0.2227722
#>  [446,]    0.2081009    0.2149764  0.2255571
#>  [447,]    0.1917439    0.1980789  0.2078278
#>  [448,]    0.3037581    0.3137937  0.3292374
#>  [449,]    0.3144391    0.3248275  0.3408141
#>  [450,]    0.1700644    0.1756832  0.1843299
#>  [451,]    0.1774028    0.1832641  0.1922840
#>  [452,]    0.1720431    0.1777273  0.1864748
#>  [453,]    0.1825960    0.1886288  0.1979127
#>  [454,]    0.1676822    0.1732223  0.1817480
#>  [455,]    0.2051937    0.2119731  0.2224059
#>  [456,]    0.2025547    0.2092469  0.2195455
#>  [457,]    0.1600940    0.1653834  0.1735232
#>  [458,]    0.2098862    0.2168205  0.2274918
#>  [459,]    0.2395493    0.2474636  0.2596429
#>  [460,]    0.2286219    0.2361753  0.2477993
#>  [461,]    0.2107195    0.2176815  0.2283951
#>  [462,]    0.1543717    0.1594720  0.1673209
#>  [463,]    0.1611426    0.1664666  0.1746599
#>  [464,]    0.1968316    0.2033347  0.2133424
#>  [465,]    0.1985395    0.2050991  0.2151936
#>  [466,]    0.2512078    0.2595074  0.2722796
#>  [467,]    0.1996951    0.2062928  0.2164460
#>  [468,]    0.1777686    0.1836419  0.1926803
#>  [469,]    0.1796299    0.1855647  0.1946978
#>  [470,]    0.2863129    0.2957722  0.3103291
#>  [471,]    0.2383459    0.2462205  0.2583388
#>  [472,]    0.1256789    0.1298313  0.1362214
#>  [473,]    0.1839829    0.1900615  0.1994159
#>  [474,]    0.1555957    0.1607364  0.1686475
#>  [475,]    0.2245364    0.2319548  0.2433711
#>  [476,]    0.1890093    0.1952539  0.2048639
#>  [477,]    0.2356640    0.2434500  0.2554318
#>  [478,]    0.1998042    0.2064055  0.2165643
#>  [479,]    0.1887988    0.1950365  0.2046358
#>  [480,]    0.3261029    0.3368766  0.3534563
#>  [481,]    0.1754855    0.1812833  0.1902057
#>  [482,]    0.1818354    0.1878431  0.1970884
#>  [483,]    0.2305716    0.2381894  0.2499125
#>  [484,]    0.1508810    0.1558660  0.1635375
#>  [485,]    0.1280320    0.1322621  0.1387719
#>  [486,]    0.1929105    0.1992841  0.2090924
#>  [487,]    0.1478404    0.1527250  0.1602418
#>  [488,]    0.2044609    0.2112160  0.2216114
#>  [489,]    0.2646610    0.2734050  0.2868612
#>  [490,]    0.1946259    0.2010562  0.2109517
#>  [491,]    0.1705273    0.1761614  0.1848317
#>  [492,]    0.2260105    0.2334776  0.2449687
#>  [493,]    0.1789643    0.1848772  0.1939765
#>  [494,]    0.2182109    0.2254204  0.2365151
#>  [495,]    0.2630962    0.2717884  0.2851647
#>  [496,]    0.1837517    0.1898227  0.1991653
#>  [497,]    0.2162935    0.2234396  0.2344366
#>  [498,]    0.2246337    0.2320553  0.2434763
#>  [499,]    0.2388378    0.2467287  0.2588719
#>  [500,]    0.1802763    0.1862325  0.1953985
#>  [501,]    0.2639347    0.2726546  0.2860736
#>  [502,]    0.1914696    0.1977956  0.2075306
#>  [503,]    0.2227303    0.2300889  0.2414131
#>  [504,]    0.1668825    0.1723963  0.1808813
#>  [505,]    0.1469377    0.1517924  0.1592633
#>  [506,]    0.1798575    0.1857997  0.1949442
#>  [507,]    0.2292194    0.2367925  0.2484466
#>  [508,]    0.1621993    0.1675582  0.1758051
#>  [509,]    0.1755242    0.1813234  0.1902478
#>  [510,]    0.2332319    0.2409376  0.2527958
#>  [511,]    0.2221530    0.2294927  0.2407878
#>  [512,]    0.1754449    0.1812415  0.1901618
#>  [513,]    0.2040202    0.2107609  0.2211340
#>  [514,]    0.1942198    0.2006365  0.2105112
#>  [515,]    0.1639705    0.1693880  0.1777250
#>  [516,]    0.2677082    0.2765528  0.2901637
#>  [517,]    0.2125447    0.2195669  0.2303734
#>  [518,]    0.1621626    0.1675203  0.1757653
#>  [519,]    0.1970831    0.2035945  0.2136149
#>  [520,]    0.1545638    0.1596704  0.1675291
#>  [521,]    0.2910209    0.3006356  0.3154317
#>  [522,]    0.2384631    0.2463416  0.2584660
#>  [523,]    0.1985905    0.2051518  0.2152490
#>  [524,]    0.2019013    0.2085719  0.2188373
#>  [525,]    0.2594859    0.2680590  0.2812521
#>  [526,]    0.1897769    0.1960470  0.2056960
#>  [527,]    0.2099900    0.2169278  0.2276044
#>  [528,]    0.1991218    0.2057006  0.2158247
#>  [529,]    0.1389571    0.1435481  0.1506132
#>  [530,]    0.2079286    0.2147983  0.2253702
#>  [531,]    0.1847178    0.1908207  0.2002124
#>  [532,]    0.1880705    0.1942841  0.2038462
#>  [533,]    0.1764766    0.1823073  0.1912800
#>  [534,]    0.2175699    0.2247581  0.2358201
#>  [535,]    0.2175575    0.2247454  0.2358067
#>  [536,]    0.1866385    0.1928048  0.2022941
#>  [537,]    0.1884370    0.1946628  0.2042436
#>  [538,]    0.2122830    0.2192966  0.2300898
#>  [539,]    0.1842653    0.1903533  0.1997221
#>  [540,]    0.1930399    0.1994177  0.2092326
#>  [541,]    0.1682458    0.1738045  0.1823589
#>  [542,]    0.2388882    0.2467808  0.2589265
#>  [543,]    0.2415739    0.2495552  0.2618376
#>  [544,]    0.2395278    0.2474414  0.2596197
#>  [545,]    0.2159191    0.2230529  0.2340311
#>  [546,]    0.1892201    0.1954717  0.2050922
#>  [547,]    0.2414585    0.2494359  0.2617124
#>  [548,]    0.1805900    0.1865565  0.1957384
#>  [549,]    0.2350018    0.2427660  0.2547141
#>  [550,]    0.2646135    0.2733559  0.2868096
#>  [551,]    0.3025525    0.3125482  0.3279305
#>  [552,]    0.2965502    0.3063476  0.3214247
#>  [553,]    0.1676654    0.1732049  0.1817297
#>  [554,]    0.1798665    0.1858092  0.1949543
#>  [555,]    0.2256229    0.2330772  0.2445486
#>  [556,]    0.2068849    0.2137201  0.2242389
#>  [557,]    0.2352370    0.2430089  0.2549691
#>  [558,]    0.2313618    0.2390055  0.2507685
#>  [559,]    0.2573345    0.2658364  0.2789200
#>  [560,]    0.1911639    0.1974797  0.2071992
#>  [561,]    0.2059584    0.2127631  0.2232348
#>  [562,]    0.1799612    0.1859069  0.1950568
#>  [563,]    0.1621932    0.1675520  0.1757985
#>  [564,]    0.1765515    0.1823847  0.1913613
#>  [565,]    0.2204407    0.2277238  0.2389318
#>  [566,]    0.1508216    0.1558046  0.1634730
#>  [567,]    0.1931752    0.1995575  0.2093793
#>  [568,]    0.2348705    0.2426302  0.2545716
#>  [569,]    0.2674584    0.2762947  0.2898929
#>  [570,]    0.1552749    0.1604051  0.1683001
#>  [571,]    0.2168543    0.2240190  0.2350446
#>  [572,]    0.2018182    0.2084859  0.2187469
#>  [573,]    0.1806332    0.1866011  0.1957852
#>  [574,]    0.2480842    0.2562805  0.2688939
#>  [575,]    0.2487777    0.2569970  0.2696455
#>  [576,]    0.1940189    0.2004292  0.2102938
#>  [577,]    0.2229726    0.2303393  0.2416760
#>  [578,]    0.2012967    0.2079473  0.2181820
#>  [579,]    0.2231694    0.2305426  0.2418892
#>  [580,]    0.2023839    0.2090705  0.2193604
#>  [581,]    0.1751275    0.1809136  0.1898178
#>  [582,]    0.1993270    0.2059125  0.2160470
#>  [583,]    0.2149603    0.2220623  0.2329916
#>  [584,]    0.2162391    0.2233834  0.2343777
#>  [585,]    0.1802466    0.1862018  0.1953662
#>  [586,]    0.2190175    0.2262535  0.2373891
#>  [587,]    0.2446020    0.2526832  0.2651193
#>  [588,]    0.1867159    0.1928848  0.2023782
#>  [589,]    0.2425891    0.2506039  0.2629379
#>  [590,]    0.1523424    0.1573757  0.1651214
#>  [591,]    0.2106020    0.2175601  0.2282679
#>  [592,]    0.1687288    0.1743035  0.1828823
#>  [593,]    0.2568311    0.2653163  0.2783742
#>  [594,]    0.1563265    0.1614914  0.1694397
#>  [595,]    0.1814805    0.1874765  0.1967037
#>  [596,]    0.1905032    0.1967972  0.2064829
#>  [597,]    0.2008333    0.2074687  0.2176798
#>  [598,]    0.2388546    0.2467460  0.2588902
#>  [599,]    0.1773035    0.1831615  0.1921764
#>  [600,]    0.1460638    0.1508897  0.1583162
#>  [601,]    0.1395232    0.1441329  0.1512269
#>  [602,]    0.1884776    0.1947048  0.2042877
#>  [603,]    0.2513942    0.2596998  0.2724813
#>  [604,]    0.1854217    0.1915478  0.2009754
#>  [605,]    0.1865090    0.1926711  0.2021539
#>  [606,]    0.1847581    0.1908623  0.2002561
#>  [607,]    0.1995591    0.2061524  0.2162988
#>  [608,]    0.1962928    0.2027780  0.2127581
#>  [609,]    0.2057170    0.2125137  0.2229731
#>  [610,]    0.1826062    0.1886394  0.1979238
#>  [611,]    0.1626639    0.1680382  0.1763087
#>  [612,]    0.1463990    0.1512360  0.1586795
#>  [613,]    0.1942446    0.2006622  0.2105384
#>  [614,]    0.2291874    0.2367595  0.2484121
#>  [615,]    0.2130545    0.2200935  0.2309260
#>  [616,]    0.1560070    0.1611613  0.1690934
#>  [617,]    0.1655764    0.1710470  0.1794657
#>  [618,]    0.1867877    0.1929590  0.2024560
#>  [619,]    0.1864814    0.1926426  0.2021240
#>  [620,]    0.1965137    0.2030063  0.2129978
#>  [621,]    0.2071695    0.2140142  0.2245474
#>  [622,]    0.2178954    0.2250944  0.2361730
#>  [623,]    0.2040727    0.2108151  0.2211909
#>  [624,]    0.1879542    0.1941640  0.2037203
#>  [625,]    0.1986394    0.2052022  0.2153018
#>  [626,]    0.2214692    0.2287863  0.2400465
#>  [627,]    0.2543027    0.2627045  0.2756340
#>  [628,]    0.1681371    0.1736922  0.1822410
#>  [629,]    0.2060771    0.2128857  0.2233634
#>  [630,]    0.2605970    0.2692067  0.2824561
#>  [631,]    0.1996672    0.2062640  0.2164158
#>  [632,]    0.2546147    0.2630268  0.2759721
#>  [633,]    0.2551431    0.2635727  0.2765449
#>  [634,]    0.1876934    0.1938946  0.2034376
#>  [635,]    0.1662284    0.1717205  0.1801723
#>  [636,]    0.1948467    0.2012843  0.2111911
#>  [637,]    0.2057514    0.2125493  0.2230106
#>  [638,]    0.1879213    0.1941300  0.2036847
#>  [639,]    0.2321478    0.2398177  0.2516209
#>  [640,]    0.1545587    0.1596652  0.1675236
#>  [641,]    0.2801965    0.2894538  0.3036998
#>  [642,]    0.1614305    0.1667641  0.1749719
#>  [643,]    0.1829860    0.1890317  0.1983355
#>  [644,]    0.2757940    0.2849058  0.2989279
#>  [645,]    0.2503451    0.2586162  0.2713445
#>  [646,]    0.2674596    0.2762960  0.2898944
#>  [647,]    0.1233170    0.1273913  0.1336613
#>  [648,]    0.2213914    0.2287059  0.2399622
#>  [649,]    0.1522588    0.1572894  0.1650309
#>  [650,]    0.2110686    0.2180421  0.2287735
#>  [651,]    0.1599424    0.1652267  0.1733589
#>  [652,]    0.2186819    0.2259069  0.2370254
#>  [653,]    0.1988838    0.2054547  0.2155666
#>  [654,]    0.1821682    0.1881869  0.1974490
#>  [655,]    0.1857795    0.1919175  0.2013633
#>  [656,]    0.2302898    0.2378982  0.2496068
#>  [657,]    0.1933761    0.1997650  0.2095970
#>  [658,]    0.2118070    0.2188049  0.2295739
#>  [659,]    0.1910813    0.1973944  0.2071098
#>  [660,]    0.2043268    0.2110775  0.2214662
#>  [661,]    0.1743685    0.1801295  0.1889951
#>  [662,]    0.2547910    0.2632088  0.2761630
#>  [663,]    0.1656604    0.1711336  0.1795565
#>  [664,]    0.2010300    0.2076719  0.2178930
#>  [665,]    0.1546915    0.1598024  0.1676675
#>  [666,]    0.2097770    0.2167078  0.2273736
#>  [667,]    0.2318690    0.2395297  0.2513187
#>  [668,]    0.1596627    0.1649379  0.1730558
#>  [669,]    0.2145769    0.2216663  0.2325762
#>  [670,]    0.2577895    0.2663064  0.2794129
#>  [671,]    0.2149222    0.2220230  0.2329505
#>  [672,]    0.3020831    0.3120634  0.3274219
#>  [673,]    0.2141036    0.2211774  0.2320633
#>  [674,]    0.2699498    0.2788685  0.2925934
#>  [675,]    0.2298545    0.2374486  0.2491352
#>  [676,]    0.1982991    0.2048506  0.2149327
#>  [677,]    0.2124807    0.2195008  0.2303039
#>  [678,]    0.2311305    0.2387666  0.2505179
#>  [679,]    0.1819268    0.1879375  0.1971873
#>  [680,]    0.1888640    0.1951038  0.2047063
#>  [681,]    0.2193847    0.2266329  0.2377872
#>  [682,]    0.1809193    0.1868968  0.1960954
#>  [683,]    0.1991533    0.2057331  0.2158587
#>  [684,]    0.1679708    0.1735204  0.1820608
#>  [685,]    0.2059472    0.2127515  0.2232226
#>  [686,]    0.2275391    0.2350567  0.2466254
#>  [687,]    0.2038311    0.2105655  0.2209290
#>  [688,]    0.1800674    0.1860167  0.1951720
#>  [689,]    0.2346685    0.2424217  0.2543530
#>  [690,]    0.1742990    0.1800577  0.1889198
#>  [691,]    0.1494071    0.1543434  0.1619400
#>  [692,]    0.2233954    0.2307761  0.2421343
#>  [693,]    0.2374443    0.2452892  0.2573616
#>  [694,]    0.2544545    0.2628612  0.2757983
#>  [695,]    0.1941782    0.2005937  0.2104665
#>  [696,]    0.1764483    0.1822781  0.1912494
#>  [697,]    0.1598150    0.1650952  0.1732209
#>  [698,]    0.2931287    0.3028132  0.3177166
#>  [699,]    0.1789119    0.1848229  0.1939195
#>  [700,]    0.1858256    0.1919651  0.2014132
#>  [701,]    0.1730847    0.1788033  0.1876037
#>  [702,]    0.2308528    0.2384800  0.2502173
#>  [703,]    0.1639279    0.1693440  0.1776788
#>  [704,]    0.1778774    0.1837543  0.1927984
#>  [705,]    0.1613818    0.1667137  0.1749190
#>  [706,]    0.1274575    0.1316686  0.1381491
#>  [707,]    0.1764172    0.1822458  0.1912154
#>  [708,]    0.1841773    0.1902623  0.1996266
#>  [709,]    0.1642321    0.1696582  0.1780085
#>  [710,]    0.2918179    0.3014590  0.3162956
#>  [711,]    0.1989868    0.2055612  0.2156785
#>  [712,]    0.1789136    0.1848247  0.1939214
#>  [713,]    0.1901125    0.1963937  0.2060598
#>  [714,]    0.2858238    0.2952670  0.3097991
#>  [715,]    0.2307467    0.2383702  0.2501021
#>  [716,]    0.2851951    0.2946173  0.3091172
#>  [717,]    0.1897697    0.1960395  0.2056881
#>  [718,]    0.1790774    0.1849940  0.1940990
#>  [719,]    0.1966542    0.2031515  0.2131502
#>  [720,]    0.2496841    0.2579333  0.2706279
#>  [721,]    0.1712481    0.1769061  0.1856131
#>  [722,]    0.1940598    0.2004713  0.2103381
#>  [723,]    0.2142498    0.2213284  0.2322217
#>  [724,]    0.2282443    0.2357851  0.2473895
#>  [725,]    0.1830064    0.1890528  0.1983576
#>  [726,]    0.2275269    0.2350441  0.2466123
#>  [727,]    0.2538741    0.2622616  0.2751692
#>  [728,]    0.1836260    0.1896928  0.1990290
#>  [729,]    0.1962610    0.2027453  0.2127239
#>  [730,]    0.3367202    0.3478447  0.3649642
#>  [731,]    0.2385408    0.2464218  0.2585499
#>  [732,]    0.2558968    0.2643512  0.2773616
#>  [733,]    0.2065054    0.2133282  0.2238278
#>  [734,]    0.1714989    0.1771650  0.1858846
#>  [735,]    0.1750891    0.1808739  0.1897761
#>  [736,]    0.1861796    0.1923308  0.2017969
#>  [737,]    0.1409981    0.1456566  0.1528255
#>  [738,]    0.1723773    0.1780724  0.1868368
#>  [739,]    0.1790827    0.1849995  0.1941047
#>  [740,]    0.1933324    0.1997199  0.2095495
#>  [741,]    0.1908424    0.1971476  0.2068508
#>  [742,]    0.2142947    0.2213747  0.2322702
#>  [743,]    0.1595213    0.1647918  0.1729025
#>  [744,]    0.2374480    0.2452930  0.2573657
#>  [745,]    0.2071228    0.2139659  0.2244968
#>  [746,]    0.1905450    0.1968403  0.2065283
#>  [747,]    0.2761236    0.2852462  0.2992851
#>  [748,]    0.2869406    0.2964205  0.3110090
#>  [749,]    0.1713707    0.1770327  0.1857458
#>  [750,]    0.2409772    0.2489387  0.2611906
#>  [751,]    0.1779755    0.1838557  0.1929046
#>  [752,]    0.2236476    0.2310367  0.2424077
#>  [753,]    0.1491703    0.1540988  0.1616833
#>  [754,]    0.2139060    0.2209732  0.2318489
#>  [755,]    0.1426049    0.1473164  0.1545671
#>  [756,]    0.1576044    0.1628116  0.1708248
#>  [757,]    0.2128392    0.2198712  0.2306927
#>  [758,]    0.1690310    0.1746157  0.1832099
#>  [759,]    0.1791678    0.1850873  0.1941969
#>  [760,]    0.2585695    0.2671122  0.2802586
#>  [761,]    0.1612067    0.1665329  0.1747293
#>  [762,]    0.1757991    0.1816073  0.1905457
#>  [763,]    0.2333657    0.2410758  0.2529408
#>  [764,]    0.2248039    0.2322311  0.2436609
#>  [765,]    0.1824999    0.1885296  0.1978086
#>  [766,]    0.2243297    0.2317412  0.2431469
#>  [767,]    0.2231919    0.2305660  0.2419138
#>  [768,]    0.1613046    0.1666340  0.1748354
#>  [769,]    0.1768258    0.1826680  0.1916585
#>  [770,]    0.2302320    0.2378385  0.2495442
#>  [771,]    0.1950143    0.2014573  0.2113726
#>  [772,]    0.3051939    0.3152769  0.3307935
#>  [773,]    0.1738890    0.1796342  0.1884755
#>  [774,]    0.2021928    0.2088730  0.2191532
#>  [775,]    0.1790050    0.1849192  0.1940206
#>  [776,]    0.2083855    0.2152703  0.2258652
#>  [777,]    0.2098818    0.2168161  0.2274872
#>  [778,]    0.2144245    0.2215089  0.2324109
#>  [779,]    0.1575017    0.1627055  0.1707135
#>  [780,]    0.2371952    0.2450317  0.2570913
#>  [781,]    0.1863310    0.1924872  0.2019610
#>  [782,]    0.1957053    0.2021712  0.2121216
#>  [783,]    0.2240296    0.2314312  0.2428216
#>  [784,]    0.3001780    0.3100954  0.3253572
#>  [785,]    0.1716467    0.1773178  0.1860449
#>  [786,]    0.1969676    0.2034753  0.2134898
#>  [787,]    0.3108456    0.3211152  0.3369191
#>  [788,]    0.2111550    0.2181313  0.2288672
#>  [789,]    0.2091109    0.2160198  0.2266518
#>  [790,]    0.2650933    0.2738516  0.2873296
#>  [791,]    0.1639788    0.1693966  0.1777339
#>  [792,]    0.1795813    0.1855145  0.1946451
#>  [793,]    0.1476306    0.1525083  0.1600146
#>  [794,]    0.2593774    0.2679468  0.2811343
#>  [795,]    0.2100450    0.2169847  0.2276642
#>  [796,]    0.2848198    0.2942296  0.3087104
#>  [797,]    0.1404189    0.1450583  0.1521978
#>  [798,]    0.1769097    0.1827547  0.1917495
#>  [799,]    0.1885213    0.1947498  0.2043350
#>  [800,]    0.2183287    0.2255420  0.2366426
#>  [801,]    0.2150402    0.2221449  0.2330782
#>  [802,]    0.2072102    0.2140562  0.2245915
#>  [803,]    0.2399165    0.2478430  0.2600411
#>  [804,]    0.1810365    0.1870178  0.1962223
#>  [805,]    0.1852015    0.1913203  0.2007366
#>  [806,]    0.1685211    0.1740889  0.1826573
#>  [807,]    0.1954061    0.2018621  0.2117973
#>  [808,]    0.2002258    0.2068410  0.2170212
#>  [809,]    0.1585626    0.1638013  0.1718633
#>  [810,]    0.2146175    0.2217081  0.2326198
#>  [811,]    0.2290119    0.2365781  0.2482217
#>  [812,]    0.1883638    0.1945871  0.2041643
#>  [813,]    0.1812402    0.1872282  0.1964432
#>  [814,]    0.1934431    0.1998343  0.2096697
#>  [815,]    0.2448779    0.2529683  0.2654187
#>  [816,]    0.1343002    0.1387374  0.1455658
#>  [817,]    0.2088049    0.2157036  0.2263200
#>  [818,]    0.1651769    0.1706342  0.1790324
#>  [819,]    0.2417484    0.2497355  0.2620268
#>  [820,]    0.1848182    0.1909245  0.2003214
#>  [821,]    0.1953640    0.2018186  0.2117516
#>  [822,]    0.2097067    0.2166353  0.2272975
#>  [823,]    0.1660235    0.1715087  0.1799499
#>  [824,]    0.2367295    0.2445508  0.2565870
#>  [825,]    0.2571554    0.2656515  0.2787259
#>  [826,]    0.3116605    0.3219572  0.3378027
#>  [827,]    0.1782077    0.1840956  0.1931564
#>  [828,]    0.2137034    0.2207639  0.2316293
#>  [829,]    0.1797764    0.1857161  0.1948567
#>  [830,]    0.2100323    0.2169715  0.2276502
#>  [831,]    0.2263135    0.2337906  0.2452971
#>  [832,]    0.2349946    0.2427585  0.2547064
#>  [833,]    0.1905064    0.1968005  0.2064866
#>  [834,]    0.1738940    0.1796394  0.1884808
#>  [835,]    0.1391207    0.1437171  0.1507905
#>  [836,]    0.1486905    0.1536031  0.1611632
#>  [837,]    0.1485863    0.1534955  0.1610502
#>  [838,]    0.1753206    0.1811130  0.1900269
#>  [839,]    0.2131031    0.2201438  0.2309787
#>  [840,]    0.2036964    0.2104263  0.2207830
#>  [841,]    0.1926228    0.1989868  0.2087804
#>  [842,]    0.2358672    0.2436599  0.2556521
#>  [843,]    0.2076629    0.2145239  0.2250822
#>  [844,]    0.2628710    0.2715558  0.2849207
#>  [845,]    0.2861738    0.2956284  0.3101782
#>  [846,]    0.1568439    0.1620259  0.1700005
#>  [847,]    0.2030134    0.2097207  0.2200426
#>  [848,]    0.1989081    0.2054799  0.2155932
#>  [849,]    0.2466659    0.2548154  0.2673567
#>  [850,]    0.2374608    0.2453062  0.2573795
#>  [851,]    0.1994303    0.2060193  0.2161592
#>  [852,]    0.1991151    0.2056936  0.2158174
#>  [853,]    0.1746321    0.1804018  0.1892807
#>  [854,]    0.2081128    0.2149887  0.2255701
#>  [855,]    0.1846544    0.1907552  0.2001437
#>  [856,]    0.1995223    0.2061143  0.2162586
#>  [857,]    0.2780310    0.2872166  0.3013523
#>  [858,]    0.2136418    0.2207003  0.2315626
#>  [859,]    0.3102757    0.3205267  0.3363018
#>  [860,]    0.1277034    0.1319227  0.1384158
#>  [861,]    0.1966016    0.2030972  0.2130932
#>  [862,]    0.1869015    0.1930766  0.2025793
#>  [863,]    0.1497304    0.1546775  0.1622905
#>  [864,]    0.2452943    0.2533985  0.2658701
#>  [865,]    0.2746973    0.2837729  0.2977392
#>  [866,]    0.1535415    0.1586144  0.1664211
#>  [867,]    0.2613932    0.2700291  0.2833190
#>  [868,]    0.1784096    0.1843041  0.1933750
#>  [869,]    0.1782392    0.1841281  0.1931906
#>  [870,]    0.1796630    0.1855989  0.1947337
#>  [871,]    0.1596786    0.1649543  0.1730729
#>  [872,]    0.2030063    0.2097134  0.2200350
#>  [873,]    0.2112488    0.2182282  0.2289688
#>  [874,]    0.2173235    0.2245036  0.2355531
#>  [875,]    0.1414439    0.1461171  0.1533087
#>  [876,]    0.2919320    0.3015768  0.3164193
#>  [877,]    0.1986831    0.2052474  0.2153492
#>  [878,]    0.2219774    0.2293112  0.2405973
#>  [879,]    0.1947672    0.2012022  0.2111050
#>  [880,]    0.2454462    0.2535554  0.2660345
#>  [881,]    0.2345390    0.2422877  0.2542121
#>  [882,]    0.2186869    0.2259120  0.2370307
#>  [883,]    0.2231381    0.2305103  0.2418553
#>  [884,]    0.2207431    0.2280362  0.2392595
#>  [885,]    0.1775969    0.1834645  0.1924942
#>  [886,]    0.2125220    0.2195435  0.2303490
#>  [887,]    0.2119352    0.2189373  0.2297129
#>  [888,]    0.1899025    0.1961767  0.2058320
#>  [889,]    0.1886789    0.1949127  0.2045058
#>  [890,]    0.2002964    0.2069140  0.2170977
#>  [891,]    0.2269483    0.2344464  0.2459853
#>  [892,]    0.1731212    0.1788410  0.1876432
#>  [893,]    0.2066136    0.2134399  0.2239449
#>  [894,]    0.1653542    0.1708174  0.1792248
#>  [895,]    0.2538217    0.2622075  0.2751124
#>  [896,]    0.2713151    0.2802789  0.2940733
#>  [897,]    0.2334282    0.2411404  0.2530086
#>  [898,]    0.1414427    0.1461159  0.1533076
#>  [899,]    0.1683375    0.1738993  0.1824583
#>  [900,]    0.1604528    0.1657540  0.1739121
#>  [901,]    0.1749212    0.1807005  0.1895942
#>  [902,]    0.1678205    0.1733652  0.1818979
#>  [903,]    0.2175121    0.2246984  0.2357574
#>  [904,]    0.2234723    0.2308555  0.2422175
#>  [905,]    0.2393900    0.2472992  0.2594707
#>  [906,]    0.1697801    0.1753895  0.1840217
#>  [907,]    0.2567437    0.2652260  0.2782796
#>  [908,]    0.2519441    0.2602680  0.2730776
#>  [909,]    0.1695838    0.1751866  0.1838089
#>  [910,]    0.1698152    0.1754257  0.1840597
#>  [911,]    0.2025368    0.2092284  0.2195262
#>  [912,]    0.1888882    0.1951289  0.2047326
#>  [913,]    0.3231717    0.3338487  0.3502794
#>  [914,]    0.1580253    0.1632464  0.1712812
#>  [915,]    0.1290325    0.1332957  0.1398563
#>  [916,]    0.2242677    0.2316773  0.2430798
#>  [917,]    0.1741409    0.1798943  0.1887483
#>  [918,]    0.1910132    0.1973241  0.2070358
#>  [919,]    0.2151870    0.2222964  0.2332370
#>  [920,]    0.1711423    0.1767967  0.1854983
#>  [921,]    0.2152194    0.2223300  0.2332725
#>  [922,]    0.2188514    0.2260818  0.2372087
#>  [923,]    0.1722881    0.1779804  0.1867403
#>  [924,]    0.3096231    0.3198525  0.3355944
#>  [925,]    0.2331738    0.2408775  0.2527328
#>  [926,]    0.2519895    0.2603149  0.2731269
#>  [927,]    0.2068833    0.2137185  0.2242371
#>  [928,]    0.1853552    0.1914791  0.2009032
#>  [929,]    0.2373170    0.2451576  0.2572236
#>  [930,]    0.2526415    0.2609883  0.2738333
#>  [931,]    0.2098263    0.2167588  0.2274272
#>  [932,]    0.2149353    0.2220365  0.2329645
#>  [933,]    0.1998389    0.2064414  0.2166020
#>  [934,]    0.1712530    0.1769111  0.1856184
#>  [935,]    0.2095759    0.2165000  0.2271556
#>  [936,]    0.3353198    0.3463979  0.3634459
#>  [937,]    0.1956707    0.2021354  0.2120840
#>  [938,]    0.1867173    0.1928863  0.2023796
#>  [939,]    0.1804052    0.1863656  0.1955381
#>  [940,]    0.2059196    0.2127230  0.2231927
#>  [941,]    0.2150114    0.2221151  0.2330469
#>  [942,]    0.1921648    0.1985137  0.2082841
#>  [943,]    0.1995589    0.2061522  0.2162986
#>  [944,]    0.1911546    0.1974701  0.2071890
#>  [945,]    0.1903550    0.1966442  0.2063226
#>  [946,]    0.2124120    0.2194298  0.2302296
#>  [947,]    0.2173136    0.2244934  0.2355424
#>  [948,]    0.2025149    0.2092057  0.2195022
#>  [949,]    0.2387837    0.2466727  0.2588132
#>  [950,]    0.1783413    0.1842336  0.1933012
#>  [951,]    0.1624602    0.1678278  0.1760879
#>  [952,]    0.1700919    0.1757115  0.1843595
#>  [953,]    0.1779616    0.1838413  0.1928895
#>  [954,]    0.1986280    0.2051905  0.2152894
#>  [955,]    0.2930347    0.3027161  0.3176147
#>  [956,]    0.2140664    0.2211389  0.2320229
#>  [957,]    0.1955226    0.2019825  0.2119237
#>  [958,]    0.2005544    0.2071805  0.2173773
#>  [959,]    0.1494548    0.1543927  0.1619916
#>  [960,]    0.2099584    0.2168951  0.2275700
#>  [961,]    0.2477145    0.2558986  0.2684932
#>  [962,]    0.2044138    0.2111673  0.2215603
#>  [963,]    0.1847571    0.1908613  0.2002550
#>  [964,]    0.2522819    0.2606168  0.2734433
#>  [965,]    0.1901895    0.1964731  0.2061431
#>  [966,]    0.2661811    0.2749752  0.2885084
#>  [967,]    0.2112799    0.2182604  0.2290026
#>  [968,]    0.3116650    0.3219615  0.3378066
#>  [969,]    0.2250955    0.2325323  0.2439769
#>  [970,]    0.2596436    0.2682218  0.2814229
#>  [971,]    0.1650410    0.1704938  0.1788850
#>  [972,]    0.2545315    0.2629410  0.2758822
#>  [973,]    0.1758393    0.1816489  0.1905893
#>  [974,]    0.1832552    0.1893098  0.1986272
#>  [975,]    0.2787660    0.2879760  0.3021492
#>  [976,]    0.1825022    0.1885319  0.1978109
#>  [977,]    0.2104635    0.2174170  0.2281177
#>  [978,]    0.2264742    0.2339566  0.2454713
#>  [979,]    0.1976984    0.2042302  0.2142820
#>  [980,]    0.1707900    0.1764328  0.1851165
#>  [981,]    0.1742607    0.1800181  0.1888783
#>  [982,]    0.2385622    0.2464440  0.2585732
#>  [983,]    0.2932982    0.3029882  0.3179002
#>  [984,]    0.2121341    0.2191427  0.2299283
#>  [985,]    0.2621275    0.2707877  0.2841150
#>  [986,]    0.1711187    0.1767724  0.1854728
#>  [987,]    0.3055284    0.3156224  0.3311559
#>  [988,]    0.1727938    0.1785027  0.1872881
#>  [989,]    0.1548647    0.1599813  0.1678552
#>  [990,]    0.1971719    0.2036863  0.2137112
#>  [991,]    0.1971979    0.2037131  0.2137394
#>  [992,]    0.1714525    0.1771171  0.1858344
#>  [993,]    0.2814199    0.2907175  0.3050256
#>  [994,]    0.3179419    0.3284460  0.3446108
#>  [995,]    0.2138174    0.2208816  0.2317528
#>  [996,]    0.2106220    0.2175808  0.2282896
#>  [997,]    0.1796947    0.1856316  0.1947679
#>  [998,]    0.2048411    0.2116087  0.2220234
#>  [999,]    0.2431510    0.2511843  0.2635468
#> [1000,]    0.1875241    0.1937197  0.2032542