Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 95.06492
#> 2   Feb 92.07727
#> 3   Mar 90.33104
#> 4   Apr 87.89789
#> 5   May 89.36706
#> 6   Jun 96.10838

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 7.536049  2
#> 2   Feb 9.570235  2
#> 3   Mar 6.517028  2
#> 4   Apr 7.677778  2
#> 5   May 9.937656  2
#> 6   Jun 9.224622  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [13ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [25ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [73ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [791ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [21ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1935350    0.2042419  0.2019712
#>    [2,]    0.2001463    0.2112189  0.2088707
#>    [3,]    0.2228872    0.2352178  0.2326027
#>    [4,]    0.1874447    0.1978147  0.1956154
#>    [5,]    0.2047447    0.2160717  0.2136694
#>    [6,]    0.2685551    0.2834119  0.2802611
#>    [7,]    0.2223030    0.2346013  0.2319931
#>    [8,]    0.1727795    0.1823381  0.1803110
#>    [9,]    0.2433770    0.2568410  0.2539855
#>   [10,]    0.3763766    0.3971974  0.3927818
#>   [11,]    0.2992643    0.3158196  0.3123085
#>   [12,]    0.2341250    0.2470772  0.2443303
#>   [13,]    0.1926063    0.2032619  0.2010021
#>   [14,]    0.1794851    0.1894148  0.1873089
#>   [15,]    0.2630955    0.2776503  0.2745635
#>   [16,]    0.2149361    0.2268268  0.2243051
#>   [17,]    0.2433072    0.2567673  0.2539127
#>   [18,]    0.2312367    0.2440292  0.2413162
#>   [19,]    0.1446143    0.1526148  0.1509181
#>   [20,]    0.2477085    0.2614123  0.2585060
#>   [21,]    0.1768896    0.1866757  0.1846003
#>   [22,]    0.2436210    0.2570985  0.2542402
#>   [23,]    0.2085578    0.2200958  0.2176488
#>   [24,]    0.2018848    0.2130536  0.2106849
#>   [25,]    0.2233451    0.2357010  0.2330806
#>   [26,]    0.1854965    0.1957587  0.1935823
#>   [27,]    0.2693076    0.2842061  0.2810464
#>   [28,]    0.2357976    0.2488424  0.2460759
#>   [29,]    0.2002708    0.2113503  0.2090006
#>   [30,]    0.1689376    0.1782835  0.1763014
#>   [31,]    0.2529987    0.2669951  0.2640267
#>   [32,]    0.2395484    0.2528009  0.2499903
#>   [33,]    0.2434148    0.2568808  0.2540249
#>   [34,]    0.3072875    0.3242868  0.3206816
#>   [35,]    0.2026426    0.2138532  0.2114757
#>   [36,]    0.2159407    0.2278869  0.2253534
#>   [37,]    0.2190229    0.2311398  0.2285701
#>   [38,]    0.1479214    0.1561049  0.1543693
#>   [39,]    0.2280491    0.2406653  0.2379897
#>   [40,]    0.3189335    0.3365769  0.3328351
#>   [41,]    0.2219650    0.2342448  0.2316405
#>   [42,]    0.1974032    0.2083241  0.2060080
#>   [43,]    0.3475556    0.3667824  0.3627048
#>   [44,]    0.1846087    0.1948216  0.1926557
#>   [45,]    0.2489853    0.2627597  0.2598384
#>   [46,]    0.1458736    0.1539438  0.1522323
#>   [47,]    0.2885243    0.3044858  0.3011007
#>   [48,]    0.1734931    0.1830913  0.1810557
#>   [49,]    0.2138663    0.2256980  0.2231887
#>   [50,]    0.1916210    0.2022219  0.1999737
#>   [51,]    0.1998879    0.2109462  0.2086010
#>   [52,]    0.2177599    0.2298069  0.2272520
#>   [53,]    0.2209472    0.2331704  0.2305781
#>   [54,]    0.2577531    0.2720124  0.2689883
#>   [55,]    0.2622672    0.2767763  0.2736992
#>   [56,]    0.2882642    0.3042112  0.3008292
#>   [57,]    0.2565257    0.2707172  0.2677075
#>   [58,]    0.1810329    0.1910481  0.1889241
#>   [59,]    0.2355178    0.2485471  0.2457838
#>   [60,]    0.1695306    0.1789096  0.1769205
#>   [61,]    0.2559725    0.2701333  0.2671301
#>   [62,]    0.1994080    0.2104398  0.2081002
#>   [63,]    0.1665477    0.1757618  0.1738077
#>   [64,]    0.2459289    0.2595342  0.2566488
#>   [65,]    0.1924505    0.2030974  0.2008394
#>   [66,]    0.2376470    0.2507941  0.2480059
#>   [67,]    0.2295256    0.2422235  0.2395306
#>   [68,]    0.2301988    0.2429338  0.2402330
#>   [69,]    0.1933651    0.2040626  0.2017939
#>   [70,]    0.2503385    0.2641875  0.2612504
#>   [71,]    0.1834835    0.1936343  0.1914815
#>   [72,]    0.1839779    0.1941560  0.1919974
#>   [73,]    0.2288596    0.2415207  0.2388355
#>   [74,]    0.2287061    0.2413587  0.2386754
#>   [75,]    0.2696881    0.2846077  0.2814436
#>   [76,]    0.1916871    0.2022918  0.2000427
#>   [77,]    0.2471965    0.2608718  0.2579716
#>   [78,]    0.1838292    0.1939993  0.1918424
#>   [79,]    0.2113624    0.2230555  0.2205756
#>   [80,]    0.1950256    0.2058149  0.2035267
#>   [81,]    0.1866960    0.1970246  0.1948342
#>   [82,]    0.2798684    0.2953510  0.2920674
#>   [83,]    0.3084824    0.3255481  0.3219288
#>   [84,]    0.1835824    0.1937387  0.1915848
#>   [85,]    0.1850841    0.1953236  0.1931520
#>   [86,]    0.1749454    0.1846239  0.1825713
#>   [87,]    0.2387684    0.2519775  0.2491762
#>   [88,]    0.1891284    0.1995915  0.1973725
#>   [89,]    0.2940401    0.3103068  0.3068570
#>   [90,]    0.1392449    0.1469485  0.1453147
#>   [91,]    0.3055611    0.3224650  0.3188800
#>   [92,]    0.1525322    0.1609708  0.1591811
#>   [93,]    0.2236543    0.2360274  0.2334033
#>   [94,]    0.2401896    0.2534772  0.2506592
#>   [95,]    0.2428538    0.2562890  0.2534397
#>   [96,]    0.2207615    0.2329745  0.2303844
#>   [97,]    0.1947752    0.2055507  0.2032655
#>   [98,]    0.2412210    0.2545658  0.2517356
#>   [99,]    0.2009211    0.2120366  0.2096792
#>  [100,]    0.2145901    0.2264617  0.2239440
#>  [101,]    0.2274773    0.2400617  0.2373928
#>  [102,]    0.1868607    0.1971984  0.1950060
#>  [103,]    0.2083669    0.2198940  0.2174493
#>  [104,]    0.1260039    0.1329748  0.1314964
#>  [105,]    0.2574217    0.2716628  0.2686426
#>  [106,]    0.2165311    0.2285100  0.2259695
#>  [107,]    0.1841853    0.1943749  0.1922139
#>  [108,]    0.1126019    0.1188314  0.1175103
#>  [109,]    0.2546096    0.2686952  0.2657079
#>  [110,]    0.2187816    0.2308851  0.2283182
#>  [111,]    0.2869831    0.3028593  0.2994923
#>  [112,]    0.3458574    0.3649901  0.3609325
#>  [113,]    0.2084669    0.2199998  0.2175539
#>  [114,]    0.2081035    0.2196163  0.2171747
#>  [115,]    0.1955254    0.2063423  0.2040483
#>  [116,]    0.1401239    0.1478761  0.1462320
#>  [117,]    0.1887332    0.1991745  0.1969601
#>  [118,]    0.1666007    0.1758176  0.1738629
#>  [119,]    0.1949635    0.2057492  0.2034618
#>  [120,]    0.1781052    0.1879585  0.1858688
#>  [121,]    0.2128933    0.2246712  0.2221733
#>  [122,]    0.1737588    0.1833716  0.1813329
#>  [123,]    0.1551713    0.1637559  0.1619353
#>  [124,]    0.2554813    0.2696146  0.2666172
#>  [125,]    0.3058689    0.3227898  0.3192012
#>  [126,]    0.2562504    0.2704267  0.2674202
#>  [127,]    0.2162044    0.2281653  0.2256286
#>  [128,]    0.2691140    0.2840018  0.2808444
#>  [129,]    0.4972181    0.5247234  0.5188901
#>  [130,]    0.2204916    0.2326898  0.2301028
#>  [131,]    0.2132830    0.2250822  0.2225799
#>  [132,]    0.2117569    0.2234717  0.2209872
#>  [133,]    0.1779503    0.1877950  0.1857072
#>  [134,]    0.1515379    0.1599215  0.1581435
#>  [135,]    0.2307511    0.2435168  0.2408095
#>  [136,]    0.2153596    0.2272739  0.2247471
#>  [137,]    0.2604548    0.2748636  0.2718078
#>  [138,]    0.2147157    0.2265942  0.2240750
#>  [139,]    0.1696145    0.1789982  0.1770081
#>  [140,]    0.1560920    0.1647276  0.1628961
#>  [141,]    0.2142281    0.2260797  0.2235662
#>  [142,]    0.1499346    0.1582295  0.1564703
#>  [143,]    0.1470524    0.1551878  0.1534624
#>  [144,]    0.1590218    0.1678195  0.1659537
#>  [145,]    0.2938771    0.3101347  0.3066868
#>  [146,]    0.2464265    0.2600591  0.2571679
#>  [147,]    0.2406192    0.2539305  0.2511075
#>  [148,]    0.1789884    0.1888907  0.1867906
#>  [149,]    0.1961706    0.2070233  0.2047217
#>  [150,]    0.2839027    0.2996086  0.2962777
#>  [151,]    0.2234927    0.2358567  0.2332345
#>  [152,]    0.1750830    0.1847691  0.1827149
#>  [153,]    0.2050768    0.2164221  0.2140160
#>  [154,]    0.2802644    0.2957690  0.2924808
#>  [155,]    0.2066473    0.2180797  0.2156551
#>  [156,]    0.3330965    0.3515234  0.3476155
#>  [157,]    0.1392377    0.1469407  0.1453071
#>  [158,]    0.2086138    0.2201549  0.2177073
#>  [159,]    0.1952138    0.2060136  0.2037232
#>  [160,]    0.2421288    0.2555238  0.2526830
#>  [161,]    0.1542306    0.1627632  0.1609537
#>  [162,]    0.2515911    0.2655096  0.2625577
#>  [163,]    0.3141672    0.3315472  0.3278613
#>  [164,]    0.2029764    0.2142056  0.2118242
#>  [165,]    0.1826339    0.1927378  0.1905950
#>  [166,]    0.3086497    0.3257244  0.3221032
#>  [167,]    0.2363494    0.2494247  0.2466517
#>  [168,]    0.2155064    0.2274287  0.2249002
#>  [169,]    0.2390910    0.2523180  0.2495128
#>  [170,]    0.1960990    0.2069477  0.2046469
#>  [171,]    0.2801068    0.2956028  0.2923164
#>  [172,]    0.1824097    0.1925012  0.1903610
#>  [173,]    0.2282024    0.2408271  0.2381497
#>  [174,]    0.2313709    0.2441708  0.2414562
#>  [175,]    0.2395326    0.2527840  0.2499737
#>  [176,]    0.2664884    0.2812310  0.2781044
#>  [177,]    0.2289938    0.2416622  0.2389755
#>  [178,]    0.2948237    0.3111335  0.3076745
#>  [179,]    0.2031273    0.2143648  0.2119816
#>  [180,]    0.1940362    0.2047708  0.2024942
#>  [181,]    0.1880704    0.1984750  0.1962684
#>  [182,]    0.1942943    0.2050431  0.2027635
#>  [183,]    0.1962013    0.2070556  0.2047537
#>  [184,]    0.1359278    0.1434478  0.1418530
#>  [185,]    0.2172881    0.2293089  0.2267595
#>  [186,]    0.2603003    0.2747005  0.2716465
#>  [187,]    0.1834133    0.1935601  0.1914082
#>  [188,]    0.2121476    0.2238841  0.2213951
#>  [189,]    0.2377645    0.2509181  0.2481285
#>  [190,]    0.2099379    0.2215522  0.2190890
#>  [191,]    0.1813891    0.1914240  0.1892958
#>  [192,]    0.2051223    0.2164702  0.2140635
#>  [193,]    0.1775083    0.1873286  0.1852459
#>  [194,]    0.1879192    0.1983154  0.1961106
#>  [195,]    0.2054242    0.2167887  0.2143786
#>  [196,]    0.2231118    0.2354549  0.2328371
#>  [197,]    0.2016180    0.2127721  0.2104066
#>  [198,]    0.2763315    0.2916186  0.2883765
#>  [199,]    0.2079766    0.2194824  0.2170422
#>  [200,]    0.2057901    0.2171750  0.2147605
#>  [201,]    0.2068603    0.2183043  0.2158773
#>  [202,]    0.1715954    0.1810885  0.1790752
#>  [203,]    0.2629028    0.2774471  0.2743625
#>  [204,]    0.1586889    0.1674680  0.1656061
#>  [205,]    0.2623981    0.2769145  0.2738359
#>  [206,]    0.2592162    0.2735564  0.2705152
#>  [207,]    0.2582601    0.2725473  0.2695173
#>  [208,]    0.2793522    0.2948064  0.2915288
#>  [209,]    0.2003170    0.2113990  0.2090487
#>  [210,]    0.2140719    0.2259149  0.2234033
#>  [211,]    0.2127682    0.2245391  0.2220428
#>  [212,]    0.2126071    0.2243691  0.2218746
#>  [213,]    0.1692657    0.1786298  0.1766439
#>  [214,]    0.1793412    0.1892629  0.1871587
#>  [215,]    0.2334045    0.2463171  0.2435786
#>  [216,]    0.2280376    0.2406530  0.2379776
#>  [217,]    0.2057572    0.2171401  0.2147261
#>  [218,]    0.2284476    0.2410859  0.2384056
#>  [219,]    0.1984053    0.2093817  0.2070538
#>  [220,]    0.2365258    0.2496109  0.2468359
#>  [221,]    0.2028810    0.2141049  0.2117245
#>  [222,]    0.2143287    0.2261858  0.2236712
#>  [223,]    0.2855321    0.3013283  0.2979783
#>  [224,]    0.2062414    0.2176512  0.2152314
#>  [225,]    0.2084199    0.2199502  0.2175048
#>  [226,]    0.2051323    0.2164807  0.2140739
#>  [227,]    0.2035046    0.2147630  0.2123754
#>  [228,]    0.2486109    0.2623646  0.2594477
#>  [229,]    0.2148154    0.2266995  0.2241791
#>  [230,]    0.1735202    0.1831199  0.1810840
#>  [231,]    0.2662227    0.2809504  0.2778270
#>  [232,]    0.2254229    0.2378935  0.2352487
#>  [233,]    0.1970668    0.2079691  0.2056570
#>  [234,]    0.2687126    0.2835783  0.2804256
#>  [235,]    0.2621261    0.2766275  0.2735520
#>  [236,]    0.2497586    0.2635756  0.2606453
#>  [237,]    0.3387488    0.3574886  0.3535143
#>  [238,]    0.2625574    0.2770825  0.2740020
#>  [239,]    0.3018046    0.3185009  0.3149600
#>  [240,]    0.2937456    0.3099959  0.3065496
#>  [241,]    0.2065804    0.2180090  0.2155852
#>  [242,]    0.1840394    0.1942210  0.1920617
#>  [243,]    0.2686456    0.2835073  0.2803555
#>  [244,]    0.2796885    0.2951612  0.2918798
#>  [245,]    0.1625347    0.1715267  0.1696197
#>  [246,]    0.2945426    0.3108371  0.3073814
#>  [247,]    0.2723019    0.2873660  0.2841712
#>  [248,]    0.1907805    0.2013350  0.1990966
#>  [249,]    0.2045667    0.2158840  0.2134838
#>  [250,]    0.2197541    0.2319114  0.2293331
#>  [251,]    0.2355361    0.2485664  0.2458029
#>  [252,]    0.1677799    0.1770621  0.1750935
#>  [253,]    0.1939157    0.2046436  0.2023684
#>  [254,]    0.2445367    0.2580649  0.2551959
#>  [255,]    0.2654195    0.2801026  0.2769886
#>  [256,]    0.1705590    0.1799948  0.1779937
#>  [257,]    0.2060545    0.2174539  0.2150363
#>  [258,]    0.1919388    0.2025574  0.2003054
#>  [259,]    0.2485278    0.2622769  0.2593610
#>  [260,]    0.2214229    0.2336726  0.2310747
#>  [261,]    0.2284863    0.2411267  0.2384460
#>  [262,]    0.3011976    0.3178600  0.3143263
#>  [263,]    0.2974837    0.3139406  0.3104504
#>  [264,]    0.1656451    0.1748091  0.1728656
#>  [265,]    0.1472223    0.1553671  0.1536397
#>  [266,]    0.2120649    0.2237968  0.2213087
#>  [267,]    0.1640620    0.1731384  0.1712135
#>  [268,]    0.2275900    0.2401808  0.2375106
#>  [269,]    0.2651580    0.2798269  0.2767160
#>  [270,]    0.1740406    0.1836689  0.1816270
#>  [271,]    0.1928093    0.2034761  0.2012139
#>  [272,]    0.2937783    0.3100306  0.3065838
#>  [273,]    0.1714943    0.1809819  0.1789697
#>  [274,]    0.1834922    0.1936436  0.1914907
#>  [275,]    0.2438719    0.2573634  0.2545021
#>  [276,]    0.2740173    0.2891764  0.2859615
#>  [277,]    0.2902288    0.3062844  0.3028794
#>  [278,]    0.1809969    0.1910102  0.1888866
#>  [279,]    0.1892931    0.1997654  0.1975444
#>  [280,]    0.2751974    0.2904217  0.2871929
#>  [281,]    0.1811964    0.1912207  0.1890947
#>  [282,]    0.1849471    0.1951788  0.1930089
#>  [283,]    0.1901514    0.2006711  0.1984401
#>  [284,]    0.2573410    0.2715775  0.2685582
#>  [285,]    0.2816161    0.2971954  0.2938914
#>  [286,]    0.1723576    0.1818928  0.1798706
#>  [287,]    0.2054551    0.2168212  0.2144107
#>  [288,]    0.2247799    0.2372152  0.2345780
#>  [289,]    0.2350686    0.2480732  0.2453152
#>  [290,]    0.2386954    0.2519004  0.2490999
#>  [291,]    0.1957693    0.2065998  0.2043029
#>  [292,]    0.2142002    0.2260503  0.2235371
#>  [293,]    0.2094446    0.2210316  0.2185743
#>  [294,]    0.2128391    0.2246138  0.2221166
#>  [295,]    0.2022473    0.2134361  0.2110632
#>  [296,]    0.2015181    0.2126667  0.2103023
#>  [297,]    0.2620486    0.2765456  0.2734711
#>  [298,]    0.2059894    0.2173852  0.2149684
#>  [299,]    0.2130661    0.2248534  0.2223535
#>  [300,]    0.2400918    0.2533740  0.2505571
#>  [301,]    0.3088897    0.3259776  0.3223536
#>  [302,]    0.1430873    0.1510034  0.1493245
#>  [303,]    0.2712031    0.2862065  0.2830246
#>  [304,]    0.2231269    0.2354707  0.2328529
#>  [305,]    0.1899102    0.2004166  0.1981884
#>  [306,]    0.1978256    0.2087699  0.2064488
#>  [307,]    0.2834681    0.2991498  0.2958240
#>  [308,]    0.2568327    0.2710410  0.2680277
#>  [309,]    0.2344999    0.2474729  0.2447216
#>  [310,]    0.1919015    0.2025180  0.2002665
#>  [311,]    0.2917317    0.3078704  0.3044478
#>  [312,]    0.1904183    0.2009527  0.1987186
#>  [313,]    0.2458142    0.2594132  0.2565291
#>  [314,]    0.2016942    0.2128523  0.2104859
#>  [315,]    0.1616604    0.1706040  0.1687073
#>  [316,]    0.1812258    0.1912517  0.1891254
#>  [317,]    0.2017617    0.2129238  0.2105565
#>  [318,]    0.1648297    0.1739486  0.1720146
#>  [319,]    0.1192988    0.1258988  0.1244991
#>  [320,]    0.2066950    0.2181299  0.2157048
#>  [321,]    0.1984087    0.2093851  0.2070572
#>  [322,]    0.2606705    0.2750910  0.2720327
#>  [323,]    0.2279396    0.2405497  0.2378754
#>  [324,]    0.2057543    0.2171370  0.2147230
#>  [325,]    0.2162034    0.2281641  0.2256275
#>  [326,]    0.1888362    0.1992832  0.1970676
#>  [327,]    0.1624266    0.1714126  0.1695068
#>  [328,]    0.2280609    0.2406777  0.2380019
#>  [329,]    0.2237148    0.2360913  0.2334665
#>  [330,]    0.1722912    0.1818229  0.1798014
#>  [331,]    0.1656413    0.1748051  0.1728616
#>  [332,]    0.2176606    0.2297021  0.2271484
#>  [333,]    0.2958104    0.3121750  0.3087044
#>  [334,]    0.2283273    0.2409589  0.2382800
#>  [335,]    0.2011724    0.2123018  0.2099415
#>  [336,]    0.2488064    0.2625709  0.2596517
#>  [337,]    0.1491558    0.1574076  0.1556576
#>  [338,]    0.1862808    0.1965864  0.1944008
#>  [339,]    0.1918218    0.2024338  0.2001832
#>  [340,]    0.2124379    0.2241905  0.2216980
#>  [341,]    0.1614006    0.1703298  0.1684361
#>  [342,]    0.2197448    0.2319015  0.2293233
#>  [343,]    0.2468201    0.2604746  0.2575788
#>  [344,]    0.2525698    0.2665423  0.2635791
#>  [345,]    0.2306202    0.2433787  0.2406729
#>  [346,]    0.2288007    0.2414586  0.2387741
#>  [347,]    0.1961914    0.2070452  0.2047434
#>  [348,]    0.2008407    0.2119517  0.2095953
#>  [349,]    0.2430090    0.2564527  0.2536016
#>  [350,]    0.2346424    0.2476232  0.2448703
#>  [351,]    0.1692624    0.1786265  0.1766406
#>  [352,]    0.1925977    0.2032526  0.2009929
#>  [353,]    0.2321171    0.2449583  0.2422350
#>  [354,]    0.2837169    0.2994126  0.2960838
#>  [355,]    0.2441788    0.2576874  0.2548225
#>  [356,]    0.1700862    0.1794959  0.1775003
#>  [357,]    0.2123370    0.2240839  0.2215926
#>  [358,]    0.1984031    0.2093793  0.2070514
#>  [359,]    0.1699163    0.1793166  0.1773230
#>  [360,]    0.1395153    0.1472338  0.1455969
#>  [361,]    0.2315294    0.2443380  0.2416216
#>  [362,]    0.2410719    0.2544085  0.2515801
#>  [363,]    0.2424597    0.2558730  0.2530283
#>  [364,]    0.2106807    0.2223360  0.2198642
#>  [365,]    0.2099818    0.2215986  0.2191349
#>  [366,]    0.2555972    0.2697374  0.2667385
#>  [367,]    0.1433961    0.1513293  0.1496468
#>  [368,]    0.2170521    0.2290599  0.2265133
#>  [369,]    0.1562916    0.1649381  0.1631044
#>  [370,]    0.2120960    0.2238298  0.2213413
#>  [371,]    0.2773219    0.2926635  0.2894099
#>  [372,]    0.1309290    0.1381724  0.1366362
#>  [373,]    0.2579343    0.2722036  0.2691774
#>  [374,]    0.1947701    0.2055454  0.2032602
#>  [375,]    0.2052016    0.2165539  0.2141463
#>  [376,]    0.2414703    0.2548287  0.2519957
#>  [377,]    0.2531311    0.2671347  0.2641648
#>  [378,]    0.1535271    0.1620207  0.1602194
#>  [379,]    0.2198879    0.2320526  0.2294728
#>  [380,]    0.2227792    0.2351039  0.2324901
#>  [381,]    0.2669188    0.2816853  0.2785536
#>  [382,]    0.1461646    0.1542509  0.1525360
#>  [383,]    0.2044064    0.2157147  0.2133164
#>  [384,]    0.2070668    0.2185221  0.2160927
#>  [385,]    0.3168712    0.3344006  0.3306830
#>  [386,]    0.1549188    0.1634894  0.1616718
#>  [387,]    0.3276120    0.3457354  0.3418918
#>  [388,]    0.2085267    0.2200628  0.2176162
#>  [389,]    0.2123147    0.2240604  0.2215694
#>  [390,]    0.3015119    0.3181920  0.3146545
#>  [391,]    0.2132225    0.2250184  0.2225167
#>  [392,]    0.2878066    0.3037285  0.3003518
#>  [393,]    0.2037848    0.2150586  0.2126677
#>  [394,]    0.2211287    0.2333621  0.2307677
#>  [395,]    0.1719950    0.1815102  0.1794923
#>  [396,]    0.3003361    0.3169508  0.3134272
#>  [397,]    0.1974871    0.2084126  0.2060955
#>  [398,]    0.1878292    0.1982204  0.1960167
#>  [399,]    0.2063575    0.2177738  0.2153526
#>  [400,]    0.1776024    0.1874279  0.1853441
#>  [401,]    0.2002072    0.2112831  0.2089342
#>  [402,]    0.1893565    0.1998321  0.1976104
#>  [403,]    0.1759293    0.1856622  0.1835981
#>  [404,]    0.2729243    0.2880227  0.2848206
#>  [405,]    0.1959313    0.2067707  0.2044719
#>  [406,]    0.1686534    0.1779837  0.1760049
#>  [407,]    0.2167548    0.2287461  0.2262030
#>  [408,]    0.1850446    0.1952818  0.1931107
#>  [409,]    0.2042998    0.2156022  0.2132052
#>  [410,]    0.1687383    0.1780734  0.1760936
#>  [411,]    0.1873822    0.1977487  0.1955502
#>  [412,]    0.2064962    0.2179200  0.2154972
#>  [413,]    0.2177879    0.2298363  0.2272811
#>  [414,]    0.1774751    0.1872935  0.1852112
#>  [415,]    0.3125333    0.3298230  0.3261563
#>  [416,]    0.1427498    0.1506473  0.1489724
#>  [417,]    0.1609360    0.1698395  0.1679512
#>  [418,]    0.1857467    0.1960227  0.1938434
#>  [419,]    0.1992191    0.2102405  0.2079031
#>  [420,]    0.1718833    0.1813924  0.1793758
#>  [421,]    0.2248554    0.2372949  0.2346567
#>  [422,]    0.2444463    0.2579695  0.2551015
#>  [423,]    0.2419284    0.2553124  0.2524739
#>  [424,]    0.2024937    0.2136961  0.2113203
#>  [425,]    0.2195974    0.2317460  0.2291696
#>  [426,]    0.2199597    0.2321284  0.2295477
#>  [427,]    0.1767544    0.1865328  0.1844590
#>  [428,]    0.1550798    0.1636592  0.1618397
#>  [429,]    0.2579699    0.2722412  0.2692146
#>  [430,]    0.1528483    0.1613044  0.1595110
#>  [431,]    0.3052857    0.3221744  0.3185926
#>  [432,]    0.2149689    0.2268615  0.2243394
#>  [433,]    0.2925877    0.3087741  0.3053413
#>  [434,]    0.2619187    0.2764083  0.2733353
#>  [435,]    0.2894335    0.3054452  0.3020495
#>  [436,]    0.1972447    0.2081568  0.2058425
#>  [437,]    0.2714959    0.2865154  0.2833301
#>  [438,]    0.2138945    0.2257277  0.2232181
#>  [439,]    0.2114059    0.2231014  0.2206211
#>  [440,]    0.2050090    0.2163506  0.2139453
#>  [441,]    0.2136018    0.2254188  0.2229127
#>  [442,]    0.2115920    0.2232977  0.2208152
#>  [443,]    0.2818037    0.2973936  0.2940873
#>  [444,]    0.1704823    0.1799140  0.1779137
#>  [445,]    0.2106993    0.2223558  0.2198837
#>  [446,]    0.2133332    0.2251354  0.2226324
#>  [447,]    0.1965649    0.2074393  0.2051331
#>  [448,]    0.3113952    0.3286219  0.3249685
#>  [449,]    0.3223447    0.3401770  0.3363951
#>  [450,]    0.1743403    0.1839853  0.1819398
#>  [451,]    0.1818633    0.1919245  0.1897907
#>  [452,]    0.1763688    0.1861261  0.1840568
#>  [453,]    0.1871870    0.1975427  0.1953465
#>  [454,]    0.1718983    0.1814082  0.1793914
#>  [455,]    0.2103529    0.2219901  0.2195221
#>  [456,]    0.2076475    0.2191350  0.2166988
#>  [457,]    0.1641193    0.1731988  0.1712732
#>  [458,]    0.2151633    0.2270665  0.2245421
#>  [459,]    0.2455721    0.2591575  0.2562764
#>  [460,]    0.2343701    0.2473360  0.2445862
#>  [461,]    0.2160176    0.2279682  0.2254337
#>  [462,]    0.1582531    0.1670081  0.1651514
#>  [463,]    0.1651942    0.1743334  0.1723951
#>  [464,]    0.2017806    0.2129435  0.2105761
#>  [465,]    0.2035314    0.2147913  0.2124033
#>  [466,]    0.2575239    0.2717706  0.2687491
#>  [467,]    0.2047160    0.2160414  0.2136395
#>  [468,]    0.1822382    0.1923201  0.1901819
#>  [469,]    0.1841463    0.1943338  0.1921732
#>  [470,]    0.2935115    0.3097489  0.3063053
#>  [471,]    0.2443385    0.2578558  0.2549891
#>  [472,]    0.1288389    0.1359667  0.1344551
#>  [473,]    0.1886088    0.1990431  0.1968302
#>  [474,]    0.1595078    0.1683322  0.1664607
#>  [475,]    0.2301819    0.2429161  0.2402154
#>  [476,]    0.1937615    0.2044809  0.2022075
#>  [477,]    0.2415892    0.2549543  0.2521199
#>  [478,]    0.2048278    0.2161594  0.2137562
#>  [479,]    0.1935457    0.2042532  0.2019824
#>  [480,]    0.3343018    0.3527955  0.3488734
#>  [481,]    0.1798977    0.1898501  0.1877394
#>  [482,]    0.1864073    0.1967199  0.1945328
#>  [483,]    0.2363688    0.2494453  0.2466721
#>  [484,]    0.1546747    0.1632317  0.1614170
#>  [485,]    0.1312512    0.1385125  0.1369725
#>  [486,]    0.1977608    0.2087015  0.2063812
#>  [487,]    0.1515576    0.1599422  0.1581640
#>  [488,]    0.2096016    0.2211971  0.2187379
#>  [489,]    0.2713153    0.2863249  0.2831417
#>  [490,]    0.1995194    0.2105573  0.2082164
#>  [491,]    0.1748148    0.1844861  0.1824350
#>  [492,]    0.2316930    0.2445107  0.2417924
#>  [493,]    0.1834640    0.1936138  0.1914613
#>  [494,]    0.2236974    0.2360730  0.2334484
#>  [495,]    0.2697110    0.2846316  0.2814673
#>  [496,]    0.1883717    0.1987930  0.1965828
#>  [497,]    0.2217317    0.2339984  0.2313969
#>  [498,]    0.2302816    0.2430211  0.2403194
#>  [499,]    0.2448428    0.2583880  0.2555153
#>  [500,]    0.1848090    0.1950332  0.1928649
#>  [501,]    0.2705706    0.2855388  0.2823644
#>  [502,]    0.1962837    0.2071427  0.2048397
#>  [503,]    0.2283303    0.2409618  0.2382830
#>  [504,]    0.1710785    0.1805431  0.1785359
#>  [505,]    0.1506322    0.1589656  0.1571982
#>  [506,]    0.1843796    0.1945798  0.1924165
#>  [507,]    0.2349826    0.2479821  0.2452252
#>  [508,]    0.1662775    0.1754764  0.1735255
#>  [509,]    0.1799374    0.1898921  0.1877809
#>  [510,]    0.2390960    0.2523232  0.2495180
#>  [511,]    0.2277386    0.2403376  0.2376656
#>  [512,]    0.1798562    0.1898063  0.1876961
#>  [513,]    0.2091499    0.2207206  0.2182667
#>  [514,]    0.1991029    0.2101176  0.2077816
#>  [515,]    0.1680933    0.1773927  0.1754205
#>  [516,]    0.2744390    0.2896213  0.2864014
#>  [517,]    0.2178887    0.2299428  0.2273863
#>  [518,]    0.1662398    0.1754367  0.1734862
#>  [519,]    0.2020383    0.2132155  0.2108451
#>  [520,]    0.1584500    0.1672159  0.1653568
#>  [521,]    0.2983378    0.3148420  0.3113418
#>  [522,]    0.2444587    0.2579828  0.2551146
#>  [523,]    0.2035837    0.2148466  0.2124579
#>  [524,]    0.2069777    0.2184281  0.2159997
#>  [525,]    0.2660101    0.2807263  0.2776053
#>  [526,]    0.1945485    0.2053115  0.2030289
#>  [527,]    0.2152697    0.2271789  0.2246532
#>  [528,]    0.2041283    0.2154212  0.2130262
#>  [529,]    0.1424509    0.1503316  0.1486603
#>  [530,]    0.2131565    0.2249489  0.2224479
#>  [531,]    0.1893621    0.1998381  0.1976164
#>  [532,]    0.1927991    0.2034651  0.2012031
#>  [533,]    0.1809138    0.1909224  0.1887998
#>  [534,]    0.2230402    0.2353792  0.2327624
#>  [535,]    0.2230275    0.2353659  0.2327492
#>  [536,]    0.1913311    0.2019159  0.1996711
#>  [537,]    0.1931748    0.2038617  0.2015953
#>  [538,]    0.2176204    0.2296597  0.2271064
#>  [539,]    0.1888983    0.1993487  0.1971324
#>  [540,]    0.1978935    0.2088414  0.2065196
#>  [541,]    0.1724760    0.1820180  0.1799943
#>  [542,]    0.2448945    0.2584425  0.2555692
#>  [543,]    0.2476477    0.2613481  0.2584425
#>  [544,]    0.2455501    0.2591344  0.2562534
#>  [545,]    0.2213480    0.2335936  0.2309965
#>  [546,]    0.1939776    0.2047088  0.2024329
#>  [547,]    0.2475294    0.2612231  0.2583190
#>  [548,]    0.1851305    0.1953724  0.1932003
#>  [549,]    0.2409104    0.2542379  0.2514114
#>  [550,]    0.2712666    0.2862734  0.2830908
#>  [551,]    0.3101593    0.3273175  0.3236786
#>  [552,]    0.3040061    0.3208238  0.3172571
#>  [553,]    0.1718810    0.1813900  0.1793733
#>  [554,]    0.1843889    0.1945899  0.1924265
#>  [555,]    0.2312956    0.2440914  0.2413777
#>  [556,]    0.2120865    0.2238197  0.2213313
#>  [557,]    0.2411515    0.2544925  0.2516631
#>  [558,]    0.2371787    0.2502997  0.2475170
#>  [559,]    0.2638045    0.2783985  0.2753035
#>  [560,]    0.1959703    0.2068118  0.2045125
#>  [561,]    0.2111368    0.2228175  0.2203402
#>  [562,]    0.1844859    0.1946922  0.1925276
#>  [563,]    0.1662713    0.1754699  0.1735191
#>  [564,]    0.1809906    0.1910036  0.1888800
#>  [565,]    0.2259832    0.2384851  0.2358337
#>  [566,]    0.1546137    0.1631673  0.1613533
#>  [567,]    0.1980322    0.2089879  0.2066644
#>  [568,]    0.2407757    0.2540957  0.2512708
#>  [569,]    0.2741829    0.2893509  0.2861341
#>  [570,]    0.1591790    0.1679854  0.1661178
#>  [571,]    0.2223067    0.2346052  0.2319970
#>  [572,]    0.2068924    0.2183380  0.2159106
#>  [573,]    0.1851749    0.1954192  0.1932466
#>  [574,]    0.2543217    0.2683912  0.2654073
#>  [575,]    0.2550326    0.2691414  0.2661493
#>  [576,]    0.1988972    0.2099007  0.2075671
#>  [577,]    0.2285787    0.2412242  0.2385424
#>  [578,]    0.2063579    0.2177741  0.2153529
#>  [579,]    0.2287805    0.2414370  0.2387528
#>  [580,]    0.2074724    0.2189503  0.2165161
#>  [581,]    0.1795308    0.1894629  0.1873565
#>  [582,]    0.2043386    0.2156431  0.2132456
#>  [583,]    0.2203650    0.2325560  0.2299706
#>  [584,]    0.2216760    0.2339396  0.2313387
#>  [585,]    0.1847785    0.1950010  0.1928330
#>  [586,]    0.2245242    0.2369453  0.2343110
#>  [587,]    0.2507519    0.2646237  0.2616818
#>  [588,]    0.1914105    0.2019999  0.1997541
#>  [589,]    0.2486884    0.2624464  0.2595286
#>  [590,]    0.1561728    0.1648127  0.1629804
#>  [591,]    0.2158972    0.2278411  0.2253081
#>  [592,]    0.1729712    0.1825404  0.1805110
#>  [593,]    0.2632884    0.2778538  0.2747648
#>  [594,]    0.1602571    0.1691229  0.1672427
#>  [595,]    0.1860435    0.1963359  0.1941531
#>  [596,]    0.1952930    0.2060969  0.2038056
#>  [597,]    0.2058829    0.2172729  0.2148573
#>  [598,]    0.2448600    0.2584062  0.2555333
#>  [599,]    0.1817615    0.1918171  0.1896845
#>  [600,]    0.1497363    0.1580202  0.1562634
#>  [601,]    0.1430312    0.1509442  0.1492660
#>  [602,]    0.1932165    0.2039058  0.2016388
#>  [603,]    0.2577148    0.2719719  0.2689483
#>  [604,]    0.1900837    0.2005996  0.1983694
#>  [605,]    0.1911984    0.2017760  0.1995327
#>  [606,]    0.1894035    0.1998817  0.1976595
#>  [607,]    0.2045766    0.2158944  0.2134941
#>  [608,]    0.2012281    0.2123603  0.2099994
#>  [609,]    0.2108893    0.2225562  0.2200819
#>  [610,]    0.1871975    0.1975538  0.1953575
#>  [611,]    0.1667538    0.1759791  0.1740226
#>  [612,]    0.1500800    0.1583829  0.1566220
#>  [613,]    0.1991285    0.2101448  0.2078084
#>  [614,]    0.2349498    0.2479477  0.2451912
#>  [615,]    0.2184113    0.2304943  0.2279317
#>  [616,]    0.1599295    0.1687773  0.1669008
#>  [617,]    0.1697396    0.1791301  0.1771386
#>  [618,]    0.1914841    0.2020775  0.1998308
#>  [619,]    0.1911701    0.2017461  0.1995032
#>  [620,]    0.2014546    0.2125996  0.2102360
#>  [621,]    0.2123783    0.2241276  0.2216358
#>  [622,]    0.2233739    0.2357315  0.2331107
#>  [623,]    0.2092037    0.2207774  0.2183229
#>  [624,]    0.1926799    0.2033395  0.2010788
#>  [625,]    0.2036337    0.2148993  0.2125101
#>  [626,]    0.2270375    0.2395977  0.2369340
#>  [627,]    0.2606965    0.2751187  0.2720601
#>  [628,]    0.1723646    0.1819003  0.1798780
#>  [629,]    0.2112584    0.2229458  0.2204672
#>  [630,]    0.2671491    0.2819281  0.2787938
#>  [631,]    0.2046874    0.2160112  0.2136096
#>  [632,]    0.2610164    0.2754561  0.2723938
#>  [633,]    0.2615581    0.2760279  0.2729592
#>  [634,]    0.1924126    0.2030573  0.2007997
#>  [635,]    0.1704079    0.1798354  0.1778361
#>  [636,]    0.1997457    0.2107963  0.2084527
#>  [637,]    0.2109247    0.2225936  0.2201189
#>  [638,]    0.1926462    0.2033039  0.2010436
#>  [639,]    0.2379846    0.2511505  0.2483583
#>  [640,]    0.1584448    0.1672104  0.1653514
#>  [641,]    0.2872414    0.3031320  0.2997620
#>  [642,]    0.1654894    0.1746448  0.1727031
#>  [643,]    0.1875868    0.1979647  0.1957638
#>  [644,]    0.2827281    0.2983690  0.2950519
#>  [645,]    0.2566395    0.2708373  0.2678262
#>  [646,]    0.2741842    0.2893525  0.2861356
#>  [647,]    0.1264176    0.1334114  0.1319282
#>  [648,]    0.2269578    0.2395136  0.2368508
#>  [649,]    0.1560871    0.1647224  0.1628910
#>  [650,]    0.2163755    0.2283458  0.2258072
#>  [651,]    0.1639638    0.1730348  0.1711110
#>  [652,]    0.2241802    0.2365823  0.2339521
#>  [653,]    0.2038843    0.2151636  0.2127715
#>  [654,]    0.1867484    0.1970799  0.1948888
#>  [655,]    0.1904506    0.2009868  0.1987523
#>  [656,]    0.2360799    0.2491402  0.2463704
#>  [657,]    0.1982381    0.2092051  0.2068792
#>  [658,]    0.2171324    0.2291447  0.2265972
#>  [659,]    0.1958856    0.2067226  0.2044243
#>  [660,]    0.2094642    0.2210522  0.2185946
#>  [661,]    0.1787526    0.1886418  0.1865445
#>  [662,]    0.2611970    0.2756467  0.2725823
#>  [663,]    0.1698256    0.1792208  0.1772283
#>  [664,]    0.2060845    0.2174856  0.2150677
#>  [665,]    0.1585809    0.1673541  0.1654935
#>  [666,]    0.2150514    0.2269486  0.2244254
#>  [667,]    0.2376989    0.2508489  0.2480601
#>  [668,]    0.1636771    0.1727323  0.1708119
#>  [669,]    0.2199720    0.2321414  0.2295605
#>  [670,]    0.2642709    0.2788906  0.2757901
#>  [671,]    0.2203260    0.2325150  0.2299300
#>  [672,]    0.3096782    0.3268098  0.3231766
#>  [673,]    0.2194868    0.2316294  0.2290542
#>  [674,]    0.2767370    0.2920464  0.2887996
#>  [675,]    0.2356337    0.2486694  0.2459048
#>  [676,]    0.2032848    0.2145309  0.2121459
#>  [677,]    0.2178230    0.2298734  0.2273178
#>  [678,]    0.2369416    0.2500495  0.2472696
#>  [679,]    0.1865010    0.1968187  0.1946305
#>  [680,]    0.1936126    0.2043236  0.2020520
#>  [681,]    0.2249006    0.2373426  0.2347039
#>  [682,]    0.1854682    0.1957288  0.1935528
#>  [683,]    0.2041605    0.2154552  0.2130598
#>  [684,]    0.1721941    0.1817204  0.1797001
#>  [685,]    0.2111253    0.2228053  0.2203282
#>  [686,]    0.2332600    0.2461644  0.2434276
#>  [687,]    0.2089560    0.2205160  0.2180644
#>  [688,]    0.1845948    0.1948071  0.1926413
#>  [689,]    0.2405687    0.2538775  0.2510550
#>  [690,]    0.1786814    0.1885666  0.1864701
#>  [691,]    0.1531637    0.1616373  0.1598402
#>  [692,]    0.2290121    0.2416816  0.2389947
#>  [693,]    0.2434143    0.2568805  0.2540246
#>  [694,]    0.2608521    0.2752827  0.2722223
#>  [695,]    0.1990605    0.2100730  0.2077375
#>  [696,]    0.1808848    0.1908919  0.1887696
#>  [697,]    0.1638332    0.1728970  0.1709748
#>  [698,]    0.3004987    0.3171227  0.3135971
#>  [699,]    0.1834102    0.1935570  0.1914051
#>  [700,]    0.1904978    0.2010367  0.1988016
#>  [701,]    0.1774366    0.1872530  0.1851711
#>  [702,]    0.2366571    0.2497495  0.2469729
#>  [703,]    0.1680496    0.1773466  0.1753749
#>  [704,]    0.1823498    0.1924379  0.1902984
#>  [705,]    0.1654394    0.1745920  0.1726510
#>  [706,]    0.1306622    0.1378908  0.1363578
#>  [707,]    0.1808528    0.1908580  0.1887361
#>  [708,]    0.1888081    0.1992534  0.1970382
#>  [709,]    0.1683614    0.1776758  0.1757004
#>  [710,]    0.2991548    0.3157043  0.3121945
#>  [711,]    0.2039900    0.2152753  0.2128819
#>  [712,]    0.1834120    0.1935589  0.1914069
#>  [713,]    0.1948925    0.2056746  0.2033879
#>  [714,]    0.2930101    0.3092200  0.3057822
#>  [715,]    0.2365483    0.2496345  0.2468592
#>  [716,]    0.2923655    0.3085393  0.3051092
#>  [717,]    0.1945411    0.2053036  0.2030211
#>  [718,]    0.1835800    0.1937362  0.1915822
#>  [719,]    0.2015987    0.2127517  0.2103864
#>  [720,]    0.2559618    0.2701219  0.2671189
#>  [721,]    0.1755539    0.1852661  0.1832064
#>  [722,]    0.1989390    0.2099449  0.2076108
#>  [723,]    0.2196367    0.2317875  0.2292106
#>  [724,]    0.2339829    0.2469271  0.2441819
#>  [725,]    0.1876078    0.1979867  0.1957856
#>  [726,]    0.2332475    0.2461512  0.2434146
#>  [727,]    0.2602571    0.2746548  0.2716013
#>  [728,]    0.1882429    0.1986569  0.1964483
#>  [729,]    0.2011956    0.2123262  0.2099656
#>  [730,]    0.3451860    0.3642819  0.3602321
#>  [731,]    0.2445383    0.2580665  0.2551975
#>  [732,]    0.2623306    0.2768431  0.2737653
#>  [733,]    0.2116976    0.2234093  0.2209255
#>  [734,]    0.1758109    0.1855371  0.1834744
#>  [735,]    0.1794914    0.1894213  0.1873154
#>  [736,]    0.1908607    0.2014196  0.1991803
#>  [737,]    0.1445432    0.1525398  0.1508439
#>  [738,]    0.1767113    0.1864875  0.1844142
#>  [739,]    0.1835854    0.1937419  0.1915879
#>  [740,]    0.1981933    0.2091578  0.2068325
#>  [741,]    0.1956407    0.2064641  0.2041687
#>  [742,]    0.2196826    0.2318360  0.2292585
#>  [743,]    0.1635322    0.1725793  0.1706606
#>  [744,]    0.2434181    0.2568845  0.2540286
#>  [745,]    0.2123305    0.2240771  0.2215859
#>  [746,]    0.1953358    0.2061422  0.2038504
#>  [747,]    0.2830660    0.2987256  0.2954045
#>  [748,]    0.2941548    0.3104276  0.3069765
#>  [749,]    0.1756795    0.1853986  0.1833374
#>  [750,]    0.2470359    0.2607023  0.2578040
#>  [751,]    0.1824504    0.1925440  0.1904034
#>  [752,]    0.2292707    0.2419545  0.2392646
#>  [753,]    0.1529209    0.1613810  0.1595868
#>  [754,]    0.2192842    0.2314155  0.2288427
#>  [755,]    0.1461904    0.1542781  0.1525629
#>  [756,]    0.1615671    0.1705055  0.1686098
#>  [757,]    0.2181906    0.2302615  0.2277015
#>  [758,]    0.1732810    0.1828674  0.1808343
#>  [759,]    0.1836726    0.1938339  0.1916789
#>  [760,]    0.2650706    0.2797346  0.2766247
#>  [761,]    0.1652600    0.1744026  0.1724637
#>  [762,]    0.1802192    0.1901895  0.1880750
#>  [763,]    0.2392332    0.2524679  0.2496611
#>  [764,]    0.2304561    0.2432053  0.2405015
#>  [765,]    0.1870885    0.1974388  0.1952438
#>  [766,]    0.2299699    0.2426923  0.2399942
#>  [767,]    0.2288036    0.2414616  0.2387771
#>  [768,]    0.1653603    0.1745085  0.1725684
#>  [769,]    0.1812717    0.1913002  0.1891734
#>  [770,]    0.2360206    0.2490777  0.2463085
#>  [771,]    0.1999175    0.2109774  0.2086318
#>  [772,]    0.3128671    0.3301751  0.3265044
#>  [773,]    0.1782612    0.1881231  0.1860316
#>  [774,]    0.2072765    0.2187435  0.2163116
#>  [775,]    0.1835058    0.1936579  0.1915049
#>  [776,]    0.2136249    0.2254429  0.2229366
#>  [777,]    0.2151588    0.2270620  0.2245375
#>  [778,]    0.2198158    0.2319764  0.2293974
#>  [779,]    0.1614618    0.1703944  0.1685000
#>  [780,]    0.2431589    0.2566107  0.2537579
#>  [781,]    0.1910159    0.2015834  0.1993423
#>  [782,]    0.2006259    0.2117250  0.2093711
#>  [783,]    0.2296623    0.2423676  0.2396731
#>  [784,]    0.3077252    0.3247489  0.3211386
#>  [785,]    0.1759624    0.1856971  0.1836326
#>  [786,]    0.2019200    0.2130907  0.2107216
#>  [787,]    0.3186608    0.3362892  0.3325506
#>  [788,]    0.2164641    0.2284393  0.2258996
#>  [789,]    0.2143686    0.2262281  0.2237130
#>  [790,]    0.2717584    0.2867925  0.2836041
#>  [791,]    0.1681018    0.1774016  0.1754293
#>  [792,]    0.1840965    0.1942812  0.1921212
#>  [793,]    0.1513426    0.1597154  0.1579397
#>  [794,]    0.2658988    0.2806087  0.2774891
#>  [795,]    0.2153261    0.2272386  0.2247122
#>  [796,]    0.2919807    0.3081333  0.3047077
#>  [797,]    0.1439495    0.1519133  0.1502243
#>  [798,]    0.1813578    0.1913911  0.1892632
#>  [799,]    0.1932612    0.2039530  0.2016855
#>  [800,]    0.2238181    0.2362002  0.2335742
#>  [801,]    0.2204469    0.2326425  0.2300561
#>  [802,]    0.2124201    0.2241717  0.2216794
#>  [803,]    0.2459486    0.2595549  0.2566693
#>  [804,]    0.1855883    0.1958555  0.1936780
#>  [805,]    0.1898580    0.2003613  0.1981338
#>  [806,]    0.1727583    0.1823158  0.1802888
#>  [807,]    0.2003192    0.2114013  0.2090510
#>  [808,]    0.2052600    0.2166155  0.2142072
#>  [809,]    0.1625493    0.1715420  0.1696348
#>  [810,]    0.2200135    0.2321849  0.2296036
#>  [811,]    0.2347698    0.2477577  0.2450032
#>  [812,]    0.1930998    0.2037826  0.2015170
#>  [813,]    0.1857971    0.1960759  0.1938960
#>  [814,]    0.1983068    0.2092777  0.2069510
#>  [815,]    0.2510348    0.2649225  0.2619772
#>  [816,]    0.1376769    0.1452936  0.1436783
#>  [817,]    0.2140548    0.2258969  0.2233854
#>  [818,]    0.1693299    0.1786977  0.1767110
#>  [819,]    0.2478266    0.2615370  0.2586293
#>  [820,]    0.1894651    0.1999469  0.1977239
#>  [821,]    0.2002760    0.2113557  0.2090059
#>  [822,]    0.2149794    0.2268726  0.2243503
#>  [823,]    0.1701978    0.1796135  0.1776166
#>  [824,]    0.2426816    0.2561073  0.2532600
#>  [825,]    0.2636210    0.2782049  0.2751119
#>  [826,]    0.3194963    0.3371712  0.3334227
#>  [827,]    0.1826884    0.1927952  0.1906518
#>  [828,]    0.2190765    0.2311962  0.2286259
#>  [829,]    0.1842966    0.1944924  0.1923301
#>  [830,]    0.2153130    0.2272246  0.2246984
#>  [831,]    0.2320037    0.2448385  0.2421165
#>  [832,]    0.2409030    0.2542302  0.2514038
#>  [833,]    0.1952963    0.2061006  0.2038092
#>  [834,]    0.1782663    0.1881285  0.1860369
#>  [835,]    0.1426186    0.1505086  0.1488353
#>  [836,]    0.1524290    0.1608619  0.1590735
#>  [837,]    0.1523222    0.1607492  0.1589620
#>  [838,]    0.1797286    0.1896716  0.1875629
#>  [839,]    0.2184611    0.2305469  0.2279838
#>  [840,]    0.2088179    0.2203702  0.2179202
#>  [841,]    0.1974659    0.2083901  0.2060733
#>  [842,]    0.2417975    0.2551742  0.2523373
#>  [843,]    0.2128842    0.2246615  0.2221637
#>  [844,]    0.2694802    0.2843881  0.2812264
#>  [845,]    0.2933688    0.3095984  0.3061564
#>  [846,]    0.1607875    0.1696827  0.1677962
#>  [847,]    0.2081177    0.2196313  0.2171895
#>  [848,]    0.2039093    0.2151902  0.2127977
#>  [849,]    0.2528678    0.2668569  0.2638901
#>  [850,]    0.2434312    0.2568983  0.2540422
#>  [851,]    0.2044446    0.2157551  0.2133563
#>  [852,]    0.2041214    0.2154139  0.2130190
#>  [853,]    0.1790229    0.1889268  0.1868264
#>  [854,]    0.2133454    0.2251484  0.2226452
#>  [855,]    0.1892972    0.1997696  0.1975486
#>  [856,]    0.2045388    0.2158543  0.2134545
#>  [857,]    0.2850213    0.3007889  0.2974450
#>  [858,]    0.2190134    0.2311297  0.2285601
#>  [859,]    0.3180768    0.3356731  0.3319413
#>  [860,]    0.1309143    0.1381570  0.1366210
#>  [861,]    0.2015448    0.2126948  0.2103301
#>  [862,]    0.1916008    0.2022006  0.1999526
#>  [863,]    0.1534952    0.1619871  0.1601861
#>  [864,]    0.2514617    0.2653731  0.2624227
#>  [865,]    0.2816039    0.2971826  0.2938787
#>  [866,]    0.1574020    0.1661100  0.1642632
#>  [867,]    0.2679652    0.2827894  0.2796455
#>  [868,]    0.1828953    0.1930135  0.1908677
#>  [869,]    0.1827207    0.1928294  0.1906856
#>  [870,]    0.1841802    0.1943696  0.1922087
#>  [871,]    0.1636934    0.1727494  0.1708288
#>  [872,]    0.2081105    0.2196237  0.2171820
#>  [873,]    0.2165602    0.2285407  0.2259999
#>  [874,]    0.2227876    0.2351128  0.2324989
#>  [875,]    0.1450002    0.1530221  0.1513208
#>  [876,]    0.2992718    0.3158278  0.3123166
#>  [877,]    0.2036785    0.2149466  0.2125569
#>  [878,]    0.2275585    0.2401475  0.2374776
#>  [879,]    0.1996643    0.2107103  0.2083677
#>  [880,]    0.2516173    0.2655371  0.2625850
#>  [881,]    0.2404358    0.2537369  0.2509160
#>  [882,]    0.2241852    0.2365876  0.2339573
#>  [883,]    0.2287484    0.2414032  0.2387193
#>  [884,]    0.2262932    0.2388122  0.2361572
#>  [885,]    0.1820622    0.1921343  0.1899982
#>  [886,]    0.2178654    0.2299183  0.2273622
#>  [887,]    0.2172639    0.2292835  0.2267344
#>  [888,]    0.1946772    0.2054472  0.2031631
#>  [889,]    0.1934228    0.2041235  0.2018541
#>  [890,]    0.2053324    0.2166919  0.2142828
#>  [891,]    0.2326544    0.2455254  0.2427958
#>  [892,]    0.1774740    0.1872924  0.1852101
#>  [893,]    0.2118085    0.2235263  0.2210412
#>  [894,]    0.1695117    0.1788897  0.1769008
#>  [895,]    0.2602033    0.2745981  0.2715452
#>  [896,]    0.2781366    0.2935236  0.2902603
#>  [897,]    0.2392972    0.2525357  0.2497281
#>  [898,]    0.1449991    0.1530210  0.1513197
#>  [899,]    0.1725700    0.1821172  0.1800925
#>  [900,]    0.1644871    0.1735870  0.1716571
#>  [901,]    0.1793193    0.1892398  0.1871359
#>  [902,]    0.1720401    0.1815578  0.1795393
#>  [903,]    0.2229810    0.2353167  0.2327006
#>  [904,]    0.2290910    0.2417647  0.2390769
#>  [905,]    0.2454090    0.2589856  0.2561063
#>  [906,]    0.1740489    0.1836777  0.1816356
#>  [907,]    0.2631988    0.2777593  0.2746714
#>  [908,]    0.2582787    0.2725671  0.2695368
#>  [909,]    0.1738476    0.1834653  0.1814256
#>  [910,]    0.1740849    0.1837156  0.1816732
#>  [911,]    0.2076292    0.2191158  0.2166797
#>  [912,]    0.1936374    0.2043498  0.2020779
#>  [913,]    0.3312969    0.3496246  0.3457377
#>  [914,]    0.1619986    0.1709610  0.1690602
#>  [915,]    0.1322768    0.1395948  0.1380428
#>  [916,]    0.2299064    0.2426254  0.2399279
#>  [917,]    0.1785193    0.1883954  0.1863009
#>  [918,]    0.1958158    0.2066488  0.2043513
#>  [919,]    0.2205973    0.2328010  0.2302129
#>  [920,]    0.1754454    0.1851515  0.1830930
#>  [921,]    0.2206306    0.2328364  0.2302478
#>  [922,]    0.2243538    0.2367652  0.2341330
#>  [923,]    0.1766200    0.1863911  0.1843189
#>  [924,]    0.3174077    0.3349670  0.3312430
#>  [925,]    0.2390364    0.2522603  0.2494558
#>  [926,]    0.2583252    0.2726163  0.2695855
#>  [927,]    0.2120849    0.2238179  0.2213296
#>  [928,]    0.1900155    0.2005276  0.1982982
#>  [929,]    0.2432838    0.2567427  0.2538883
#>  [930,]    0.2589935    0.2733213  0.2702827
#>  [931,]    0.2151020    0.2270020  0.2244783
#>  [932,]    0.2203394    0.2325290  0.2299438
#>  [933,]    0.2048635    0.2161971  0.2137935
#>  [934,]    0.1755589    0.1852714  0.1832116
#>  [935,]    0.2148452    0.2267309  0.2242102
#>  [936,]    0.3437503    0.3627665  0.3587335
#>  [937,]    0.2005904    0.2116875  0.2093341
#>  [938,]    0.1914119    0.2020013  0.1997555
#>  [939,]    0.1849411    0.1951725  0.1930027
#>  [940,]    0.2110970    0.2227754  0.2202987
#>  [941,]    0.2204173    0.2326113  0.2300252
#>  [942,]    0.1969964    0.2078947  0.2055834
#>  [943,]    0.2045765    0.2158942  0.2134939
#>  [944,]    0.1959607    0.2068017  0.2045025
#>  [945,]    0.1951411    0.2059369  0.2036473
#>  [946,]    0.2177526    0.2297992  0.2272443
#>  [947,]    0.2227775    0.2351021  0.2324883
#>  [948,]    0.2076066    0.2190919  0.2166561
#>  [949,]    0.2447873    0.2583294  0.2554574
#>  [950,]    0.1828254    0.1929399  0.1907948
#>  [951,]    0.1665450    0.1757587  0.1738047
#>  [952,]    0.1743684    0.1840149  0.1819691
#>  [953,]    0.1824361    0.1925289  0.1903884
#>  [954,]    0.2036221    0.2148870  0.2124979
#>  [955,]    0.3004023    0.3170209  0.3134965
#>  [956,]    0.2194486    0.2315891  0.2290144
#>  [957,]    0.2004387    0.2115275  0.2091758
#>  [958,]    0.2055969    0.2169709  0.2145587
#>  [959,]    0.1532126    0.1616887  0.1598911
#>  [960,]    0.2152373    0.2271446  0.2246193
#>  [961,]    0.2539427    0.2679913  0.2650119
#>  [962,]    0.2095533    0.2211461  0.2186875
#>  [963,]    0.1894024    0.1998807  0.1976584
#>  [964,]    0.2586248    0.2729322  0.2698979
#>  [965,]    0.1949714    0.2057577  0.2034702
#>  [966,]    0.2728735    0.2879691  0.2847676
#>  [967,]    0.2165921    0.2285745  0.2260333
#>  [968,]    0.3195007    0.3371751  0.3334268
#>  [969,]    0.2307550    0.2435208  0.2408134
#>  [970,]    0.2661717    0.2808968  0.2777739
#>  [971,]    0.1691906    0.1785506  0.1765655
#>  [972,]    0.2609312    0.2753665  0.2723051
#>  [973,]    0.1802605    0.1902330  0.1881180
#>  [974,]    0.1878628    0.1982558  0.1960517
#>  [975,]    0.2857749    0.3015844  0.2982315
#>  [976,]    0.1870908    0.1974411  0.1952460
#>  [977,]    0.2157552    0.2276913  0.2251599
#>  [978,]    0.2321684    0.2450124  0.2422885
#>  [979,]    0.2026692    0.2138814  0.2115035
#>  [980,]    0.1750842    0.1847705  0.1827162
#>  [981,]    0.1786421    0.1885252  0.1864292
#>  [982,]    0.2445603    0.2580898  0.2552205
#>  [983,]    0.3006724    0.3173059  0.3137783
#>  [984,]    0.2174677    0.2294984  0.2269470
#>  [985,]    0.2687180    0.2835838  0.2804311
#>  [986,]    0.1754212    0.1851260  0.1830678
#>  [987,]    0.3132100    0.3305368  0.3268622
#>  [988,]    0.1771383    0.1869380  0.1848597
#>  [989,]    0.1587584    0.1675414  0.1656787
#>  [990,]    0.2021294    0.2133117  0.2109401
#>  [991,]    0.2021560    0.2133399  0.2109680
#>  [992,]    0.1757633    0.1854870  0.1834248
#>  [993,]    0.2884955    0.3044554  0.3010706
#>  [994,]    0.3259356    0.3439665  0.3401426
#>  [995,]    0.2191933    0.2313195  0.2287478
#>  [996,]    0.2159177    0.2278628  0.2253295
#>  [997,]    0.1842128    0.1944038  0.1922425
#>  [998,]    0.2099913    0.2216084  0.2191446
#>  [999,]    0.2492644    0.2630541  0.2601296
#> [1000,]    0.1922390    0.2028742  0.2006187