Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 90.62527
#> 2   Feb 93.82979
#> 3   Mar 92.54855
#> 4   Apr 92.85018
#> 5   May 96.68965
#> 6   Jun 90.90727

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 8.796787  2
#> 2   Feb 9.337035  2
#> 3   Mar 7.853199  2
#> 4   Apr 7.292759  2
#> 5   May 8.634666  2
#> 6   Jun 8.639852  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [13ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [25ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [79ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [731ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [21ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.2013200    0.2015430  0.2016272
#>    [2,]    0.2081972    0.2084279  0.2085150
#>    [3,]    0.2318528    0.2321096  0.2322066
#>    [4,]    0.1949847    0.1952007  0.1952823
#>    [5,]    0.2129805    0.2132165  0.2133056
#>    [6,]    0.2793575    0.2796670  0.2797838
#>    [7,]    0.2312451    0.2315012  0.2315980
#>    [8,]    0.1797296    0.1799287  0.1800039
#>    [9,]    0.2531667    0.2534471  0.2535530
#>   [10,]    0.3915154    0.3919491  0.3921129
#>   [11,]    0.3113016    0.3116465  0.3117767
#>   [12,]    0.2435426    0.2438124  0.2439143
#>   [13,]    0.2003540    0.2005760  0.2006598
#>   [14,]    0.1867050    0.1869118  0.1869899
#>   [15,]    0.2736783    0.2739815  0.2740960
#>   [16,]    0.2235819    0.2238295  0.2239231
#>   [17,]    0.2530941    0.2533744  0.2534803
#>   [18,]    0.2405381    0.2408046  0.2409052
#>   [19,]    0.1504315    0.1505981  0.1506611
#>   [20,]    0.2576725    0.2579579  0.2580658
#>   [21,]    0.1840051    0.1842089  0.1842859
#>   [22,]    0.2534205    0.2537012  0.2538073
#>   [23,]    0.2169471    0.2171874  0.2172782
#>   [24,]    0.2100056    0.2102383  0.2103261
#>   [25,]    0.2323291    0.2325864  0.2326836
#>   [26,]    0.1929582    0.1931719  0.1932526
#>   [27,]    0.2801403    0.2804506  0.2805678
#>   [28,]    0.2452825    0.2455542  0.2456568
#>   [29,]    0.2083267    0.2085575  0.2086447
#>   [30,]    0.1757330    0.1759277  0.1760012
#>   [31,]    0.2631755    0.2634670  0.2635771
#>   [32,]    0.2491843    0.2494603  0.2495646
#>   [33,]    0.2532059    0.2534864  0.2535923
#>   [34,]    0.3196477    0.3200018  0.3201355
#>   [35,]    0.2107938    0.2110273  0.2111155
#>   [36,]    0.2246268    0.2248756  0.2249696
#>   [37,]    0.2278331    0.2280855  0.2281808
#>   [38,]    0.1538716    0.1540420  0.1541064
#>   [39,]    0.2372224    0.2374852  0.2375844
#>   [40,]    0.3317620    0.3321295  0.3322683
#>   [41,]    0.2308937    0.2311494  0.2312460
#>   [42,]    0.2053438    0.2055713  0.2056572
#>   [43,]    0.3615354    0.3619359  0.3620872
#>   [44,]    0.1920345    0.1922472  0.1923276
#>   [45,]    0.2590007    0.2592876  0.2593960
#>   [46,]    0.1517415    0.1519096  0.1519730
#>   [47,]    0.3001299    0.3004623  0.3005879
#>   [48,]    0.1804719    0.1806718  0.1807474
#>   [49,]    0.2224691    0.2227156  0.2228086
#>   [50,]    0.1993289    0.1995497  0.1996331
#>   [51,]    0.2079284    0.2081587  0.2082457
#>   [52,]    0.2265193    0.2267702  0.2268650
#>   [53,]    0.2298347    0.2300893  0.2301855
#>   [54,]    0.2681211    0.2684181  0.2685302
#>   [55,]    0.2728168    0.2731190  0.2732331
#>   [56,]    0.2998593    0.3001915  0.3003169
#>   [57,]    0.2668443    0.2671399  0.2672516
#>   [58,]    0.1883149    0.1885235  0.1886023
#>   [59,]    0.2449914    0.2452628  0.2453653
#>   [60,]    0.1763501    0.1765455  0.1766193
#>   [61,]    0.2662688    0.2665638  0.2666752
#>   [62,]    0.2074292    0.2076590  0.2077458
#>   [63,]    0.1732473    0.1734392  0.1735117
#>   [64,]    0.2558214    0.2561048  0.2562118
#>   [65,]    0.2001918    0.2004136  0.2004974
#>   [66,]    0.2472063    0.2474801  0.2475836
#>   [67,]    0.2387583    0.2390228  0.2391227
#>   [68,]    0.2394584    0.2397237  0.2398239
#>   [69,]    0.2011432    0.2013661  0.2014502
#>   [70,]    0.2604081    0.2606966  0.2608055
#>   [71,]    0.1908641    0.1910756  0.1911554
#>   [72,]    0.1913784    0.1915904  0.1916704
#>   [73,]    0.2380655    0.2383292  0.2384288
#>   [74,]    0.2379058    0.2381694  0.2382689
#>   [75,]    0.2805362    0.2808469  0.2809643
#>   [76,]    0.1993978    0.1996186  0.1997021
#>   [77,]    0.2571398    0.2574247  0.2575323
#>   [78,]    0.1912239    0.1914357  0.1915157
#>   [79,]    0.2198644    0.2201080  0.2202000
#>   [80,]    0.2028705    0.2030952  0.2031801
#>   [81,]    0.1942060    0.1944211  0.1945024
#>   [82,]    0.2911258    0.2914483  0.2915701
#>   [83,]    0.3208909    0.3212463  0.3213806
#>   [84,]    0.1909671    0.1911786  0.1912585
#>   [85,]    0.1925293    0.1927426  0.1928231
#>   [86,]    0.1819827    0.1821843  0.1822604
#>   [87,]    0.2483728    0.2486479  0.2487518
#>   [88,]    0.1967361    0.1969540  0.1970364
#>   [89,]    0.3058676    0.3062065  0.3063344
#>   [90,]    0.1448462    0.1450066  0.1450672
#>   [91,]    0.3178519    0.3182040  0.3183370
#>   [92,]    0.1586679    0.1588437  0.1589100
#>   [93,]    0.2326508    0.2329085  0.2330059
#>   [94,]    0.2498510    0.2501278  0.2502323
#>   [95,]    0.2526226    0.2529024  0.2530081
#>   [96,]    0.2296416    0.2298960  0.2299921
#>   [97,]    0.2026101    0.2028345  0.2029193
#>   [98,]    0.2509240    0.2512020  0.2513070
#>   [99,]    0.2090032    0.2092347  0.2093222
#>  [100,]    0.2232220    0.2234692  0.2235626
#>  [101,]    0.2366274    0.2368896  0.2369886
#>  [102,]    0.1943773    0.1945926  0.1946739
#>  [103,]    0.2167482    0.2169883  0.2170790
#>  [104,]    0.1310724    0.1312176  0.1312725
#>  [105,]    0.2677765    0.2680731  0.2681851
#>  [106,]    0.2252410    0.2254905  0.2255847
#>  [107,]    0.1915942    0.1918064  0.1918866
#>  [108,]    0.1171314    0.1172612  0.1173102
#>  [109,]    0.2648512    0.2651446  0.2652554
#>  [110,]    0.2275821    0.2278342  0.2279294
#>  [111,]    0.2985267    0.2988574  0.2989823
#>  [112,]    0.3597688    0.3601673  0.3603178
#>  [113,]    0.2168525    0.2170927  0.2171834
#>  [114,]    0.2164744    0.2167142  0.2168048
#>  [115,]    0.2033904    0.2036157  0.2037008
#>  [116,]    0.1457605    0.1459220  0.1459830
#>  [117,]    0.1963251    0.1965426  0.1966247
#>  [118,]    0.1733023    0.1734943  0.1735668
#>  [119,]    0.2028058    0.2030305  0.2031153
#>  [120,]    0.1852695    0.1854747  0.1855523
#>  [121,]    0.2214570    0.2217023  0.2217950
#>  [122,]    0.1807483    0.1809485  0.1810241
#>  [123,]    0.1614132    0.1615920  0.1616595
#>  [124,]    0.2657576    0.2660520  0.2661632
#>  [125,]    0.3181721    0.3185245  0.3186577
#>  [126,]    0.2665580    0.2668532  0.2669648
#>  [127,]    0.2249012    0.2251503  0.2252444
#>  [128,]    0.2799390    0.2802491  0.2803662
#>  [129,]    0.5172173    0.5177902  0.5180066
#>  [130,]    0.2293609    0.2296150  0.2297109
#>  [131,]    0.2218622    0.2221080  0.2222008
#>  [132,]    0.2202747    0.2205187  0.2206109
#>  [133,]    0.1851084    0.1853135  0.1853909
#>  [134,]    0.1576336    0.1578082  0.1578742
#>  [135,]    0.2400330    0.2402989  0.2403994
#>  [136,]    0.2240225    0.2242706  0.2243644
#>  [137,]    0.2709315    0.2712316  0.2713450
#>  [138,]    0.2233526    0.2236000  0.2236935
#>  [139,]    0.1764374    0.1766328  0.1767067
#>  [140,]    0.1623709    0.1625508  0.1626187
#>  [141,]    0.2228454    0.2230923  0.2231855
#>  [142,]    0.1559658    0.1561386  0.1562039
#>  [143,]    0.1529676    0.1531371  0.1532011
#>  [144,]    0.1654186    0.1656018  0.1656710
#>  [145,]    0.3056980    0.3060366  0.3061645
#>  [146,]    0.2563388    0.2566227  0.2567300
#>  [147,]    0.2502979    0.2505751  0.2506799
#>  [148,]    0.1861884    0.1863946  0.1864725
#>  [149,]    0.2040616    0.2042877  0.2043730
#>  [150,]    0.2953225    0.2956496  0.2957732
#>  [151,]    0.2324825    0.2327401  0.2328373
#>  [152,]    0.1821258    0.1823275  0.1824037
#>  [153,]    0.2133260    0.2135623  0.2136515
#>  [154,]    0.2915378    0.2918608  0.2919827
#>  [155,]    0.2149598    0.2151979  0.2152879
#>  [156,]    0.3464948    0.3468786  0.3470235
#>  [157,]    0.1448386    0.1449990  0.1450596
#>  [158,]    0.2170053    0.2172457  0.2173365
#>  [159,]    0.2030663    0.2032913  0.2033762
#>  [160,]    0.2518683    0.2521473  0.2522527
#>  [161,]    0.1604347    0.1606124  0.1606795
#>  [162,]    0.2617112    0.2620011  0.2621106
#>  [163,]    0.3268042    0.3271662  0.3273029
#>  [164,]    0.2111412    0.2113751  0.2114634
#>  [165,]    0.1899804    0.1901909  0.1902704
#>  [166,]    0.3210647    0.3214204  0.3215547
#>  [167,]    0.2458565    0.2461288  0.2462317
#>  [168,]    0.2241751    0.2244234  0.2245172
#>  [169,]    0.2487084    0.2489839  0.2490879
#>  [170,]    0.2039871    0.2042131  0.2042984
#>  [171,]    0.2913739    0.2916967  0.2918186
#>  [172,]    0.1897472    0.1899574  0.1900368
#>  [173,]    0.2373818    0.2376448  0.2377441
#>  [174,]    0.2406777    0.2409443  0.2410450
#>  [175,]    0.2491677    0.2494437  0.2495480
#>  [176,]    0.2772077    0.2775148  0.2776308
#>  [177,]    0.2382050    0.2384689  0.2385685
#>  [178,]    0.3066826    0.3070223  0.3071506
#>  [179,]    0.2112981    0.2115322  0.2116206
#>  [180,]    0.2018413    0.2020649  0.2021494
#>  [181,]    0.1956356    0.1958523  0.1959341
#>  [182,]    0.2021098    0.2023337  0.2024182
#>  [183,]    0.2040935    0.2043196  0.2044050
#>  [184,]    0.1413956    0.1415522  0.1416114
#>  [185,]    0.2260284    0.2262788  0.2263734
#>  [186,]    0.2707707    0.2710707  0.2711839
#>  [187,]    0.1907911    0.1910024  0.1910822
#>  [188,]    0.2206812    0.2209257  0.2210180
#>  [189,]    0.2473285    0.2476025  0.2477060
#>  [190,]    0.2183827    0.2186246  0.2187159
#>  [191,]    0.1886855    0.1888945  0.1889735
#>  [192,]    0.2133733    0.2136097  0.2136990
#>  [193,]    0.1846486    0.1848532  0.1849304
#>  [194,]    0.1954783    0.1956948  0.1957766
#>  [195,]    0.2136874    0.2139241  0.2140135
#>  [196,]    0.2320864    0.2323435  0.2324406
#>  [197,]    0.2097282    0.2099605  0.2100483
#>  [198,]    0.2874468    0.2877652  0.2878855
#>  [199,]    0.2163425    0.2165821  0.2166726
#>  [200,]    0.2140681    0.2143052  0.2143948
#>  [201,]    0.2151812    0.2154196  0.2155096
#>  [202,]    0.1784978    0.1786955  0.1787702
#>  [203,]    0.2734780    0.2737809  0.2738953
#>  [204,]    0.1650721    0.1652550  0.1653241
#>  [205,]    0.2729530    0.2732554  0.2733696
#>  [206,]    0.2696430    0.2699417  0.2700545
#>  [207,]    0.2686483    0.2689459  0.2690583
#>  [208,]    0.2905889    0.2909108  0.2910324
#>  [209,]    0.2083747    0.2086056  0.2086927
#>  [210,]    0.2226830    0.2229296  0.2230228
#>  [211,]    0.2213269    0.2215720  0.2216646
#>  [212,]    0.2211592    0.2214042  0.2214968
#>  [213,]    0.1760744    0.1762694  0.1763431
#>  [214,]    0.1865552    0.1867619  0.1868400
#>  [215,]    0.2427932    0.2430622  0.2431638
#>  [216,]    0.2372103    0.2374731  0.2375723
#>  [217,]    0.2140337    0.2142708  0.2143604
#>  [218,]    0.2376369    0.2379001  0.2379996
#>  [219,]    0.2063862    0.2066148  0.2067012
#>  [220,]    0.2460400    0.2463126  0.2464155
#>  [221,]    0.2110419    0.2112757  0.2113640
#>  [222,]    0.2229500    0.2231970  0.2232903
#>  [223,]    0.2970175    0.2973465  0.2974708
#>  [224,]    0.2145375    0.2147751  0.2148649
#>  [225,]    0.2168036    0.2170437  0.2171344
#>  [226,]    0.2133837    0.2136201  0.2137094
#>  [227,]    0.2116906    0.2119251  0.2120137
#>  [228,]    0.2586112    0.2588977  0.2590059
#>  [229,]    0.2234563    0.2237039  0.2237974
#>  [230,]    0.1805002    0.1807001  0.1807756
#>  [231,]    0.2769312    0.2772380  0.2773539
#>  [232,]    0.2344902    0.2347500  0.2348481
#>  [233,]    0.2049939    0.2052210  0.2053067
#>  [234,]    0.2795215    0.2798311  0.2799480
#>  [235,]    0.2726700    0.2729721  0.2730862
#>  [236,]    0.2598050    0.2600928  0.2602015
#>  [237,]    0.3523745    0.3527649  0.3529123
#>  [238,]    0.2731186    0.2734211  0.2735354
#>  [239,]    0.3139445    0.3142922  0.3144236
#>  [240,]    0.3055612    0.3058997  0.3060275
#>  [241,]    0.2148901    0.2151282  0.2152181
#>  [242,]    0.1914424    0.1916545  0.1917346
#>  [243,]    0.2794516    0.2797611  0.2798780
#>  [244,]    0.2909387    0.2912610  0.2913827
#>  [245,]    0.1690728    0.1692601  0.1693308
#>  [246,]    0.3063903    0.3067297  0.3068579
#>  [247,]    0.2832550    0.2835688  0.2836873
#>  [248,]    0.1984547    0.1986745  0.1987576
#>  [249,]    0.2127955    0.2130312  0.2131203
#>  [250,]    0.2285937    0.2288469  0.2289426
#>  [251,]    0.2450104    0.2452818  0.2453843
#>  [252,]    0.1745290    0.1747223  0.1747953
#>  [253,]    0.2017159    0.2019394  0.2020238
#>  [254,]    0.2543731    0.2546549  0.2547613
#>  [255,]    0.2760956    0.2764014  0.2765169
#>  [256,]    0.1774198    0.1776163  0.1776906
#>  [257,]    0.2143430    0.2145804  0.2146701
#>  [258,]    0.1996596    0.1998808  0.1999643
#>  [259,]    0.2585248    0.2588111  0.2589193
#>  [260,]    0.2303297    0.2305848  0.2306812
#>  [261,]    0.2376772    0.2379404  0.2380399
#>  [262,]    0.3133129    0.3136599  0.3137910
#>  [263,]    0.3094495    0.3097923  0.3099218
#>  [264,]    0.1723082    0.1724991  0.1725712
#>  [265,]    0.1531444    0.1533140  0.1533781
#>  [266,]    0.2205952    0.2208395  0.2209318
#>  [267,]    0.1706615    0.1708505  0.1709219
#>  [268,]    0.2367448    0.2370070  0.2371061
#>  [269,]    0.2758238    0.2761293  0.2762447
#>  [270,]    0.1810414    0.1812419  0.1813177
#>  [271,]    0.2005651    0.2007873  0.2008712
#>  [272,]    0.3055954    0.3059339  0.3060617
#>  [273,]    0.1783927    0.1785903  0.1786650
#>  [274,]    0.1908733    0.1910847  0.1911646
#>  [275,]    0.2536816    0.2539626  0.2540687
#>  [276,]    0.2850395    0.2853552  0.2854745
#>  [277,]    0.3019029    0.3022373  0.3023636
#>  [278,]    0.1882776    0.1884862  0.1885649
#>  [279,]    0.1969075    0.1971256  0.1972080
#>  [280,]    0.2862670    0.2865841  0.2867039
#>  [281,]    0.1884851    0.1886938  0.1887727
#>  [282,]    0.1923866    0.1925997  0.1926802
#>  [283,]    0.1978003    0.1980194  0.1981021
#>  [284,]    0.2676924    0.2679889  0.2681009
#>  [285,]    0.2929438    0.2932683  0.2933909
#>  [286,]    0.1792907    0.1794893  0.1795643
#>  [287,]    0.2137194    0.2139562  0.2140456
#>  [288,]    0.2338216    0.2340807  0.2341785
#>  [289,]    0.2445242    0.2447951  0.2448974
#>  [290,]    0.2482967    0.2485718  0.2486757
#>  [291,]    0.2036441    0.2038697  0.2039549
#>  [292,]    0.2228164    0.2230632  0.2231564
#>  [293,]    0.2178695    0.2181109  0.2182020
#>  [294,]    0.2214005    0.2216457  0.2217383
#>  [295,]    0.2103827    0.2106157  0.2107038
#>  [296,]    0.2096243    0.2098565  0.2099442
#>  [297,]    0.2725894    0.2728913  0.2730054
#>  [298,]    0.2142753    0.2145127  0.2146023
#>  [299,]    0.2216366    0.2218821  0.2219749
#>  [300,]    0.2497493    0.2500260  0.2501305
#>  [301,]    0.3213143    0.3216702  0.3218047
#>  [302,]    0.1488431    0.1490080  0.1490702
#>  [303,]    0.2821121    0.2824246  0.2825427
#>  [304,]    0.2321021    0.2323592  0.2324563
#>  [305,]    0.1975494    0.1977682  0.1978509
#>  [306,]    0.2057832    0.2060112  0.2060973
#>  [307,]    0.2948702    0.2951969  0.2953202
#>  [308,]    0.2671636    0.2674595  0.2675713
#>  [309,]    0.2439326    0.2442028  0.2443048
#>  [310,]    0.1996208    0.1998419  0.1999254
#>  [311,]    0.3034662    0.3038023  0.3039293
#>  [312,]    0.1980779    0.1982973  0.1983802
#>  [313,]    0.2557020    0.2559853  0.2560923
#>  [314,]    0.2098072    0.2100397  0.2101274
#>  [315,]    0.1681633    0.1683496  0.1684200
#>  [316,]    0.1885156    0.1887245  0.1888034
#>  [317,]    0.2098777    0.2101101  0.2101980
#>  [318,]    0.1714600    0.1716500  0.1717217
#>  [319,]    0.1240977    0.1242351  0.1242871
#>  [320,]    0.2150094    0.2152475  0.2153375
#>  [321,]    0.2063896    0.2066183  0.2067046
#>  [322,]    0.2711557    0.2714560  0.2715695
#>  [323,]    0.2371084    0.2373711  0.2374703
#>  [324,]    0.2140307    0.2142678  0.2143573
#>  [325,]    0.2249001    0.2251492  0.2252433
#>  [326,]    0.1964323    0.1966499  0.1967320
#>  [327,]    0.1689603    0.1691475  0.1692182
#>  [328,]    0.2372345    0.2374973  0.2375966
#>  [329,]    0.2327138    0.2329716  0.2330689
#>  [330,]    0.1792217    0.1794202  0.1794952
#>  [331,]    0.1723043    0.1724951  0.1725672
#>  [332,]    0.2264160    0.2266668  0.2267615
#>  [333,]    0.3077091    0.3080500  0.3081787
#>  [334,]    0.2375117    0.2377748  0.2378742
#>  [335,]    0.2092646    0.2094964  0.2095840
#>  [336,]    0.2588145    0.2591012  0.2592095
#>  [337,]    0.1551557    0.1553276  0.1553925
#>  [338,]    0.1937741    0.1939887  0.1940698
#>  [339,]    0.1995378    0.1997588  0.1998423
#>  [340,]    0.2209832    0.2212280  0.2213205
#>  [341,]    0.1678930    0.1680790  0.1681492
#>  [342,]    0.2285840    0.2288372  0.2289328
#>  [343,]    0.2567483    0.2570327  0.2571402
#>  [344,]    0.2627292    0.2630203  0.2631302
#>  [345,]    0.2398969    0.2401627  0.2402630
#>  [346,]    0.2380043    0.2382679  0.2383675
#>  [347,]    0.2040832    0.2043093  0.2043947
#>  [348,]    0.2089195    0.2091510  0.2092384
#>  [349,]    0.2527840    0.2530640  0.2531697
#>  [350,]    0.2440808    0.2443511  0.2444533
#>  [351,]    0.1760710    0.1762661  0.1763398
#>  [352,]    0.2003449    0.2005668  0.2006506
#>  [353,]    0.2414540    0.2417214  0.2418225
#>  [354,]    0.2951292    0.2954561  0.2955796
#>  [355,]    0.2540009    0.2542823  0.2543886
#>  [356,]    0.1769280    0.1771240  0.1771980
#>  [357,]    0.2208782    0.2211228  0.2212153
#>  [358,]    0.2063839    0.2066125  0.2066988
#>  [359,]    0.1767513    0.1769471  0.1770210
#>  [360,]    0.1451275    0.1452882  0.1453489
#>  [361,]    0.2408425    0.2411093  0.2412101
#>  [362,]    0.2507690    0.2510468  0.2511517
#>  [363,]    0.2522125    0.2524919  0.2525974
#>  [364,]    0.2191553    0.2193981  0.2194898
#>  [365,]    0.2184284    0.2186704  0.2187618
#>  [366,]    0.2658785    0.2661731  0.2662843
#>  [367,]    0.1491643    0.1493296  0.1493920
#>  [368,]    0.2257830    0.2260331  0.2261276
#>  [369,]    0.1625785    0.1627586  0.1628266
#>  [370,]    0.2206277    0.2208721  0.2209644
#>  [371,]    0.2884768    0.2887963  0.2889170
#>  [372,]    0.1361957    0.1363466  0.1364035
#>  [373,]    0.2683095    0.2686067  0.2687190
#>  [374,]    0.2026048    0.2028293  0.2029140
#>  [375,]    0.2134559    0.2136923  0.2137816
#>  [376,]    0.2511832    0.2514615  0.2515666
#>  [377,]    0.2633131    0.2636048  0.2637150
#>  [378,]    0.1597028    0.1598797  0.1599465
#>  [379,]    0.2287329    0.2289863  0.2290820
#>  [380,]    0.2317405    0.2319972  0.2320942
#>  [381,]    0.2776555    0.2779631  0.2780792
#>  [382,]    0.1520442    0.1522126  0.1522762
#>  [383,]    0.2126286    0.2128642  0.2129531
#>  [384,]    0.2153960    0.2156346  0.2157247
#>  [385,]    0.3296169    0.3299820  0.3301199
#>  [386,]    0.1611505    0.1613290  0.1613965
#>  [387,]    0.3407896    0.3411671  0.3413096
#>  [388,]    0.2169146    0.2171549  0.2172457
#>  [389,]    0.2208550    0.2210997  0.2211921
#>  [390,]    0.3136400    0.3139875  0.3141187
#>  [391,]    0.2217993    0.2220450  0.2221378
#>  [392,]    0.2993835    0.2997151  0.2998403
#>  [393,]    0.2119820    0.2122168  0.2123055
#>  [394,]    0.2300236    0.2302784  0.2303747
#>  [395,]    0.1789135    0.1791117  0.1791866
#>  [396,]    0.3124167    0.3127627  0.3128934
#>  [397,]    0.2054310    0.2056586  0.2057445
#>  [398,]    0.1953847    0.1956011  0.1956829
#>  [399,]    0.2146583    0.2148961  0.2149859
#>  [400,]    0.1847465    0.1849512  0.1850285
#>  [401,]    0.2082605    0.2084912  0.2085784
#>  [402,]    0.1969733    0.1971915  0.1972739
#>  [403,]    0.1830061    0.1832088  0.1832854
#>  [404,]    0.2839023    0.2842168  0.2843356
#>  [405,]    0.2038126    0.2040384  0.2041237
#>  [406,]    0.1754374    0.1756318  0.1757052
#>  [407,]    0.2254737    0.2257234  0.2258178
#>  [408,]    0.1924880    0.1927013  0.1927818
#>  [409,]    0.2125178    0.2127532  0.2128421
#>  [410,]    0.1755258    0.1757203  0.1757937
#>  [411,]    0.1949197    0.1951356  0.1952172
#>  [412,]    0.2148024    0.2150404  0.2151302
#>  [413,]    0.2265483    0.2267993  0.2268941
#>  [414,]    0.1846140    0.1848185  0.1848958
#>  [415,]    0.3251047    0.3254648  0.3256008
#>  [416,]    0.1484921    0.1486566  0.1487187
#>  [417,]    0.1674097    0.1675952  0.1676652
#>  [418,]    0.1932184    0.1934325  0.1935133
#>  [419,]    0.2072327    0.2074623  0.2075490
#>  [420,]    0.1787974    0.1789955  0.1790703
#>  [421,]    0.2339002    0.2341593  0.2342571
#>  [422,]    0.2542791    0.2545607  0.2546671
#>  [423,]    0.2516599    0.2519387  0.2520440
#>  [424,]    0.2106390    0.2108723  0.2109605
#>  [425,]    0.2284307    0.2286837  0.2287793
#>  [426,]    0.2288076    0.2290610  0.2291568
#>  [427,]    0.1838643    0.1840680  0.1841449
#>  [428,]    0.1613179    0.1614966  0.1615641
#>  [429,]    0.2683466    0.2686438  0.2687561
#>  [430,]    0.1589967    0.1591728  0.1592394
#>  [431,]    0.3175655    0.3179172  0.3180501
#>  [432,]    0.2236160    0.2238638  0.2239573
#>  [433,]    0.3043569    0.3046940  0.3048214
#>  [434,]    0.2724541    0.2727559  0.2728699
#>  [435,]    0.3010756    0.3014091  0.3015351
#>  [436,]    0.2051789    0.2054062  0.2054920
#>  [437,]    0.2824166    0.2827294  0.2828476
#>  [438,]    0.2224984    0.2227449  0.2228380
#>  [439,]    0.2199097    0.2201533  0.2202453
#>  [440,]    0.2132555    0.2134917  0.2135809
#>  [441,]    0.2221939    0.2224401  0.2225330
#>  [442,]    0.2201032    0.2203470  0.2204391
#>  [443,]    0.2931391    0.2934638  0.2935865
#>  [444,]    0.1773401    0.1775365  0.1776107
#>  [445,]    0.2191747    0.2194175  0.2195092
#>  [446,]    0.2219146    0.2221604  0.2222532
#>  [447,]    0.2044717    0.2046982  0.2047837
#>  [448,]    0.3239208    0.3242796  0.3244151
#>  [449,]    0.3353106    0.3356820  0.3358223
#>  [450,]    0.1813532    0.1815541  0.1816299
#>  [451,]    0.1891788    0.1893884  0.1894675
#>  [452,]    0.1834634    0.1836666  0.1837434
#>  [453,]    0.1947166    0.1949323  0.1950138
#>  [454,]    0.1788130    0.1790111  0.1790859
#>  [455,]    0.2188143    0.2190567  0.2191483
#>  [456,]    0.2160001    0.2162394  0.2163298
#>  [457,]    0.1707210    0.1709101  0.1709816
#>  [458,]    0.2238181    0.2240661  0.2241597
#>  [459,]    0.2554501    0.2557331  0.2558399
#>  [460,]    0.2437976    0.2440677  0.2441697
#>  [461,]    0.2247069    0.2249558  0.2250498
#>  [462,]    0.1646189    0.1648012  0.1648701
#>  [463,]    0.1718393    0.1720296  0.1721015
#>  [464,]    0.2098972    0.2101297  0.2102175
#>  [465,]    0.2117185    0.2119530  0.2120416
#>  [466,]    0.2678826    0.2681794  0.2682915
#>  [467,]    0.2129507    0.2131866  0.2132757
#>  [468,]    0.1895688    0.1897787  0.1898581
#>  [469,]    0.1915536    0.1917658  0.1918460
#>  [470,]    0.3053177    0.3056559  0.3057837
#>  [471,]    0.2541670    0.2544485  0.2545549
#>  [472,]    0.1340216    0.1341700  0.1342261
#>  [473,]    0.1961956    0.1964129  0.1964950
#>  [474,]    0.1659240    0.1661078  0.1661773
#>  [475,]    0.2394409    0.2397062  0.2398063
#>  [476,]    0.2015556    0.2017788  0.2018632
#>  [477,]    0.2513070    0.2515854  0.2516905
#>  [478,]    0.2130670    0.2133030  0.2133922
#>  [479,]    0.2013311    0.2015542  0.2016384
#>  [480,]    0.3477486    0.3481338  0.3482793
#>  [481,]    0.1871341    0.1873414  0.1874197
#>  [482,]    0.1939056    0.1941204  0.1942016
#>  [483,]    0.2458767    0.2461491  0.2462520
#>  [484,]    0.1608965    0.1610748  0.1611421
#>  [485,]    0.1365309    0.1366821  0.1367392
#>  [486,]    0.2057158    0.2059437  0.2060298
#>  [487,]    0.1576541    0.1578287  0.1578947
#>  [488,]    0.2180327    0.2182742  0.2183654
#>  [489,]    0.2822288    0.2825414  0.2826595
#>  [490,]    0.2075451    0.2077750  0.2078618
#>  [491,]    0.1818468    0.1820483  0.1821244
#>  [492,]    0.2410128    0.2412798  0.2413806
#>  [493,]    0.1908440    0.1910554  0.1911352
#>  [494,]    0.2326957    0.2329534  0.2330508
#>  [495,]    0.2805598    0.2808706  0.2809880
#>  [496,]    0.1959490    0.1961661  0.1962481
#>  [497,]    0.2306508    0.2309063  0.2310028
#>  [498,]    0.2395445    0.2398099  0.2399101
#>  [499,]    0.2546915    0.2549736  0.2550802
#>  [500,]    0.1922430    0.1924560  0.1925364
#>  [501,]    0.2814540    0.2817658  0.2818835
#>  [502,]    0.2041793    0.2044054  0.2044909
#>  [503,]    0.2375147    0.2377778  0.2378772
#>  [504,]    0.1779602    0.1781574  0.1782318
#>  [505,]    0.1566914    0.1568650  0.1569305
#>  [506,]    0.1917962    0.1920086  0.1920889
#>  [507,]    0.2444346    0.2447053  0.2448076
#>  [508,]    0.1729660    0.1731577  0.1732300
#>  [509,]    0.1871755    0.1873828  0.1874611
#>  [510,]    0.2487135    0.2489890  0.2490931
#>  [511,]    0.2368993    0.2371618  0.2372609
#>  [512,]    0.1870909    0.1872982  0.1873765
#>  [513,]    0.2175630    0.2178040  0.2178950
#>  [514,]    0.2071117    0.2073411  0.2074278
#>  [515,]    0.1748549    0.1750486  0.1751218
#>  [516,]    0.2854780    0.2857943  0.2859137
#>  [517,]    0.2266532    0.2269043  0.2269991
#>  [518,]    0.1729269    0.1731184  0.1731908
#>  [519,]    0.2101653    0.2103981  0.2104860
#>  [520,]    0.1648237    0.1650062  0.1650752
#>  [521,]    0.3103380    0.3106818  0.3108116
#>  [522,]    0.2542921    0.2545738  0.2546802
#>  [523,]    0.2117729    0.2120075  0.2120961
#>  [524,]    0.2153033    0.2155418  0.2156319
#>  [525,]    0.2767102    0.2770168  0.2771325
#>  [526,]    0.2023743    0.2025984  0.2026831
#>  [527,]    0.2239288    0.2241769  0.2242706
#>  [528,]    0.2123394    0.2125746  0.2126634
#>  [529,]    0.1481810    0.1483451  0.1484071
#>  [530,]    0.2217307    0.2219764  0.2220691
#>  [531,]    0.1969793    0.1971975  0.1972799
#>  [532,]    0.2005544    0.2007765  0.2008605
#>  [533,]    0.1881911    0.1883995  0.1884783
#>  [534,]    0.2320119    0.2322689  0.2323660
#>  [535,]    0.2319987    0.2322557  0.2323528
#>  [536,]    0.1990273    0.1992478  0.1993311
#>  [537,]    0.2009453    0.2011679  0.2012520
#>  [538,]    0.2263742    0.2266249  0.2267197
#>  [539,]    0.1964968    0.1967144  0.1967967
#>  [540,]    0.2058537    0.2060818  0.2061679
#>  [541,]    0.1794140    0.1796127  0.1796878
#>  [542,]    0.2547452    0.2550274  0.2551340
#>  [543,]    0.2576093    0.2578946  0.2580024
#>  [544,]    0.2554272    0.2557102  0.2558170
#>  [545,]    0.2302518    0.2305068  0.2306032
#>  [546,]    0.2017802    0.2020038  0.2020882
#>  [547,]    0.2574861    0.2577714  0.2578791
#>  [548,]    0.1925774    0.1927908  0.1928713
#>  [549,]    0.2506009    0.2508785  0.2509833
#>  [550,]    0.2821781    0.2824906  0.2826087
#>  [551,]    0.3226351    0.3229924  0.3231274
#>  [552,]    0.3162343    0.3165846  0.3167169
#>  [553,]    0.1787950    0.1789931  0.1790679
#>  [554,]    0.1918060    0.1920185  0.1920988
#>  [555,]    0.2405994    0.2408660  0.2409666
#>  [556,]    0.2206177    0.2208621  0.2209544
#>  [557,]    0.2508517    0.2511296  0.2512346
#>  [558,]    0.2467190    0.2469923  0.2470955
#>  [559,]    0.2744158    0.2747198  0.2748346
#>  [560,]    0.2038532    0.2040790  0.2041643
#>  [561,]    0.2196298    0.2198731  0.2199650
#>  [562,]    0.1919069    0.1921195  0.1921998
#>  [563,]    0.1729596    0.1731512  0.1732236
#>  [564,]    0.1882710    0.1884796  0.1885584
#>  [565,]    0.2350733    0.2353337  0.2354321
#>  [566,]    0.1608330    0.1610112  0.1610785
#>  [567,]    0.2059981    0.2062263  0.2063125
#>  [568,]    0.2504607    0.2507381  0.2508429
#>  [569,]    0.2852116    0.2855275  0.2856469
#>  [570,]    0.1655822    0.1657656  0.1658349
#>  [571,]    0.2312490    0.2315051  0.2316019
#>  [572,]    0.2152145    0.2154529  0.2155429
#>  [573,]    0.1926235    0.1928369  0.1929175
#>  [574,]    0.2645516    0.2648447  0.2649554
#>  [575,]    0.2652912    0.2655850  0.2656960
#>  [576,]    0.2068978    0.2071270  0.2072136
#>  [577,]    0.2377733    0.2380366  0.2381361
#>  [578,]    0.2146586    0.2148964  0.2149862
#>  [579,]    0.2379831    0.2382467  0.2383462
#>  [580,]    0.2158180    0.2160570  0.2161473
#>  [581,]    0.1867524    0.1869593  0.1870375
#>  [582,]    0.2125581    0.2127936  0.2128825
#>  [583,]    0.2292291    0.2294830  0.2295789
#>  [584,]    0.2305928    0.2308483  0.2309448
#>  [585,]    0.1922113    0.1924242  0.1925046
#>  [586,]    0.2335556    0.2338143  0.2339120
#>  [587,]    0.2608381    0.2611270  0.2612362
#>  [588,]    0.1991100    0.1993306  0.1994139
#>  [589,]    0.2586919    0.2589784  0.2590867
#>  [590,]    0.1624549    0.1626348  0.1627028
#>  [591,]    0.2245816    0.2248304  0.2249244
#>  [592,]    0.1799290    0.1801283  0.1802036
#>  [593,]    0.2738789    0.2741823  0.2742969
#>  [594,]    0.1667034    0.1668881  0.1669579
#>  [595,]    0.1935271    0.1937415  0.1938225
#>  [596,]    0.2031485    0.2033736  0.2034586
#>  [597,]    0.2141645    0.2144018  0.2144914
#>  [598,]    0.2547094    0.2549916  0.2550982
#>  [599,]    0.1890729    0.1892824  0.1893615
#>  [600,]    0.1557596    0.1559321  0.1559973
#>  [601,]    0.1487847    0.1489496  0.1490118
#>  [602,]    0.2009887    0.2012114  0.2012955
#>  [603,]    0.2680812    0.2683781  0.2684903
#>  [604,]    0.1977299    0.1979489  0.1980316
#>  [605,]    0.1988893    0.1991097  0.1991929
#>  [606,]    0.1970222    0.1972405  0.1973229
#>  [607,]    0.2128058    0.2130415  0.2131306
#>  [608,]    0.2093223    0.2095542  0.2096418
#>  [609,]    0.2193723    0.2196153  0.2197071
#>  [610,]    0.1947276    0.1949433  0.1950248
#>  [611,]    0.1734615    0.1736537  0.1737262
#>  [612,]    0.1561170    0.1562900  0.1563553
#>  [613,]    0.2071384    0.2073679  0.2074545
#>  [614,]    0.2444006    0.2446714  0.2447736
#>  [615,]    0.2271968    0.2274485  0.2275436
#>  [616,]    0.1663627    0.1665470  0.1666166
#>  [617,]    0.1765675    0.1767631  0.1768369
#>  [618,]    0.1991865    0.1994072  0.1994905
#>  [619,]    0.1988599    0.1990802  0.1991634
#>  [620,]    0.2095582    0.2097903  0.2098780
#>  [621,]    0.2209212    0.2211660  0.2212584
#>  [622,]    0.2323591    0.2326165  0.2327137
#>  [623,]    0.2176189    0.2178600  0.2179511
#>  [624,]    0.2004305    0.2006525  0.2007364
#>  [625,]    0.2118249    0.2120596  0.2121482
#>  [626,]    0.2361700    0.2364317  0.2365305
#>  [627,]    0.2711829    0.2714833  0.2715968
#>  [628,]    0.1792980    0.1794966  0.1795716
#>  [629,]    0.2197563    0.2199998  0.2200917
#>  [630,]    0.2778949    0.2782028  0.2783190
#>  [631,]    0.2129209    0.2131568  0.2132459
#>  [632,]    0.2715155    0.2718163  0.2719299
#>  [633,]    0.2720791    0.2723805  0.2724943
#>  [634,]    0.2001523    0.2003740  0.2004578
#>  [635,]    0.1772627    0.1774590  0.1775332
#>  [636,]    0.2077806    0.2080108  0.2080977
#>  [637,]    0.2194092    0.2196522  0.2197440
#>  [638,]    0.2003954    0.2006174  0.2007012
#>  [639,]    0.2475575    0.2478317  0.2479353
#>  [640,]    0.1648183    0.1650009  0.1650699
#>  [641,]    0.2987955    0.2991265  0.2992515
#>  [642,]    0.1721463    0.1723370  0.1724090
#>  [643,]    0.1951326    0.1953488  0.1954304
#>  [644,]    0.2941006    0.2944264  0.2945495
#>  [645,]    0.2669627    0.2672584  0.2673701
#>  [646,]    0.2852131    0.2855290  0.2856483
#>  [647,]    0.1315028    0.1316485  0.1317035
#>  [648,]    0.2360871    0.2363487  0.2364474
#>  [649,]    0.1623658    0.1625457  0.1626136
#>  [650,]    0.2250791    0.2253285  0.2254226
#>  [651,]    0.1705593    0.1707483  0.1708196
#>  [652,]    0.2331978    0.2334561  0.2335537
#>  [653,]    0.2120855    0.2123204  0.2124091
#>  [654,]    0.1942604    0.1944756  0.1945569
#>  [655,]    0.1981115    0.1983310  0.1984139
#>  [656,]    0.2455761    0.2458481  0.2459508
#>  [657,]    0.2062122    0.2064407  0.2065269
#>  [658,]    0.2258666    0.2261168  0.2262113
#>  [659,]    0.2037652    0.2039909  0.2040761
#>  [660,]    0.2178898    0.2181312  0.2182223
#>  [661,]    0.1859430    0.1861490  0.1862268
#>  [662,]    0.2717034    0.2720044  0.2721181
#>  [663,]    0.1766568    0.1768525  0.1769265
#>  [664,]    0.2143743    0.2146117  0.2147014
#>  [665,]    0.1649599    0.1651426  0.1652116
#>  [666,]    0.2237018    0.2239497  0.2240432
#>  [667,]    0.2472603    0.2475342  0.2476376
#>  [668,]    0.1702612    0.1704498  0.1705210
#>  [669,]    0.2288204    0.2290738  0.2291696
#>  [670,]    0.2749009    0.2752054  0.2753204
#>  [671,]    0.2291886    0.2294425  0.2295384
#>  [672,]    0.3221346    0.3224914  0.3226262
#>  [673,]    0.2283157    0.2285686  0.2286642
#>  [674,]    0.2878685    0.2881874  0.2883078
#>  [675,]    0.2451120    0.2453835  0.2454860
#>  [676,]    0.2114619    0.2116961  0.2117846
#>  [677,]    0.2265849    0.2268359  0.2269307
#>  [678,]    0.2464724    0.2467454  0.2468485
#>  [679,]    0.1940030    0.1942179  0.1942991
#>  [680,]    0.2014006    0.2016237  0.2017079
#>  [681,]    0.2339472    0.2342064  0.2343043
#>  [682,]    0.1929287    0.1931424  0.1932232
#>  [683,]    0.2123729    0.2126081  0.2126970
#>  [684,]    0.1791207    0.1793191  0.1793941
#>  [685,]    0.2196179    0.2198611  0.2199530
#>  [686,]    0.2426428    0.2429116  0.2430131
#>  [687,]    0.2173613    0.2176021  0.2176930
#>  [688,]    0.1920202    0.1922329  0.1923132
#>  [689,]    0.2502455    0.2505227  0.2506274
#>  [690,]    0.1858689    0.1860748  0.1861526
#>  [691,]    0.1593248    0.1595013  0.1595680
#>  [692,]    0.2382241    0.2384880  0.2385877
#>  [693,]    0.2532056    0.2534861  0.2535920
#>  [694,]    0.2713446    0.2716452  0.2717587
#>  [695,]    0.2070677    0.2072971  0.2073837
#>  [696,]    0.1881609    0.1883694  0.1884481
#>  [697,]    0.1704235    0.1706123  0.1706836
#>  [698,]    0.3125860    0.3129323  0.3130630
#>  [699,]    0.1907879    0.1909993  0.1910791
#>  [700,]    0.1981607    0.1983802  0.1984631
#>  [701,]    0.1845741    0.1847786  0.1848558
#>  [702,]    0.2461766    0.2464493  0.2465523
#>  [703,]    0.1748094    0.1750031  0.1750762
#>  [704,]    0.1896849    0.1898950  0.1899744
#>  [705,]    0.1720943    0.1722849  0.1723569
#>  [706,]    0.1359181    0.1360687  0.1361256
#>  [707,]    0.1881276    0.1883360  0.1884147
#>  [708,]    0.1964029    0.1966204  0.1967026
#>  [709,]    0.1751339    0.1753279  0.1754012
#>  [710,]    0.3111879    0.3115327  0.3116629
#>  [711,]    0.2121955    0.2124306  0.2125194
#>  [712,]    0.1907898    0.1910011  0.1910810
#>  [713,]    0.2027322    0.2029567  0.2030416
#>  [714,]    0.3047963    0.3051339  0.3052615
#>  [715,]    0.2460633    0.2463359  0.2464388
#>  [716,]    0.3041255    0.3044623  0.3045896
#>  [717,]    0.2023665    0.2025907  0.2026754
#>  [718,]    0.1909645    0.1911761  0.1912560
#>  [719,]    0.2097081    0.2099404  0.2100281
#>  [720,]    0.2662576    0.2665526  0.2666640
#>  [721,]    0.1826157    0.1828180  0.1828944
#>  [722,]    0.2069414    0.2071706  0.2072572
#>  [723,]    0.2284716    0.2287247  0.2288203
#>  [724,]    0.2433946    0.2436642  0.2437661
#>  [725,]    0.1951543    0.1953705  0.1954522
#>  [726,]    0.2426298    0.2428986  0.2430001
#>  [727,]    0.2707257    0.2710255  0.2711388
#>  [728,]    0.1958149    0.1960318  0.1961138
#>  [729,]    0.2092887    0.2095205  0.2096081
#>  [730,]    0.3590706    0.3594684  0.3596186
#>  [731,]    0.2543747    0.2546565  0.2547629
#>  [732,]    0.2728826    0.2731849  0.2732991
#>  [733,]    0.2202132    0.2204571  0.2205493
#>  [734,]    0.1828828    0.1830854  0.1831620
#>  [735,]    0.1867115    0.1869183  0.1869964
#>  [736,]    0.1985381    0.1987580  0.1988411
#>  [737,]    0.1503576    0.1505241  0.1505870
#>  [738,]    0.1838196    0.1840232  0.1841001
#>  [739,]    0.1909702    0.1911817  0.1912616
#>  [740,]    0.2061656    0.2063940  0.2064802
#>  [741,]    0.2035104    0.2037358  0.2038210
#>  [742,]    0.2285194    0.2287725  0.2288681
#>  [743,]    0.1701103    0.1702988  0.1703699
#>  [744,]    0.2532096    0.2534901  0.2535960
#>  [745,]    0.2208714    0.2211161  0.2212085
#>  [746,]    0.2031931    0.2034182  0.2035032
#>  [747,]    0.2944521    0.2947783  0.2949015
#>  [748,]    0.3059868    0.3063257  0.3064537
#>  [749,]    0.1827462    0.1829487  0.1830251
#>  [750,]    0.2569728    0.2572574  0.2573649
#>  [751,]    0.1897895    0.1899997  0.1900791
#>  [752,]    0.2384931    0.2387573  0.2388571
#>  [753,]    0.1590722    0.1592485  0.1593150
#>  [754,]    0.2281049    0.2283575  0.2284530
#>  [755,]    0.1520710    0.1522394  0.1523031
#>  [756,]    0.1680662    0.1682524  0.1683227
#>  [757,]    0.2269673    0.2272188  0.2273137
#>  [758,]    0.1802513    0.1804510  0.1805264
#>  [759,]    0.1910609    0.1912725  0.1913525
#>  [760,]    0.2757328    0.2760383  0.2761536
#>  [761,]    0.1719076    0.1720980  0.1721700
#>  [762,]    0.1874686    0.1876762  0.1877547
#>  [763,]    0.2488562    0.2491318  0.2492360
#>  [764,]    0.2397261    0.2399916  0.2400919
#>  [765,]    0.1946143    0.1948298  0.1949113
#>  [766,]    0.2392204    0.2394854  0.2395855
#>  [767,]    0.2380072    0.2382709  0.2383705
#>  [768,]    0.1720120    0.1722025  0.1722745
#>  [769,]    0.1885634    0.1887723  0.1888512
#>  [770,]    0.2455144    0.2457864  0.2458891
#>  [771,]    0.2079592    0.2081895  0.2082765
#>  [772,]    0.3254518    0.3258123  0.3259484
#>  [773,]    0.1854318    0.1856372  0.1857148
#>  [774,]    0.2156142    0.2158530  0.2159432
#>  [775,]    0.1908874    0.1910989  0.1911787
#>  [776,]    0.2222178    0.2224640  0.2225569
#>  [777,]    0.2238136    0.2240615  0.2241552
#>  [778,]    0.2286578    0.2289111  0.2290068
#>  [779,]    0.1679567    0.1681427  0.1682130
#>  [780,]    0.2529397    0.2532199  0.2533257
#>  [781,]    0.1986996    0.1989197  0.1990028
#>  [782,]    0.2086961    0.2089273  0.2090146
#>  [783,]    0.2389003    0.2391650  0.2392649
#>  [784,]    0.3201032    0.3204578  0.3205917
#>  [785,]    0.1830405    0.1832433  0.1833199
#>  [786,]    0.2100422    0.2102749  0.2103628
#>  [787,]    0.3314785    0.3318456  0.3319843
#>  [788,]    0.2251713    0.2254207  0.2255149
#>  [789,]    0.2229917    0.2232387  0.2233320
#>  [790,]    0.2826897    0.2830028  0.2831211
#>  [791,]    0.1748637    0.1750574  0.1751306
#>  [792,]    0.1915018    0.1917139  0.1917940
#>  [793,]    0.1574305    0.1576049  0.1576707
#>  [794,]    0.2765944    0.2769008  0.2770165
#>  [795,]    0.2239877    0.2242358  0.2243295
#>  [796,]    0.3037253    0.3040617  0.3041888
#>  [797,]    0.1497400    0.1499058  0.1499685
#>  [798,]    0.1886530    0.1888620  0.1889409
#>  [799,]    0.2010352    0.2012579  0.2013420
#>  [800,]    0.2328211    0.2330790  0.2331764
#>  [801,]    0.2293143    0.2295683  0.2296643
#>  [802,]    0.2209647    0.2212094  0.2213019
#>  [803,]    0.2558418    0.2561252  0.2562322
#>  [804,]    0.1930536    0.1932674  0.1933482
#>  [805,]    0.1974950    0.1977137  0.1977964
#>  [806,]    0.1797075    0.1799066  0.1799818
#>  [807,]    0.2083770    0.2086079  0.2086950
#>  [808,]    0.2135166    0.2137531  0.2138424
#>  [809,]    0.1690879    0.1692752  0.1693459
#>  [810,]    0.2288633    0.2291169  0.2292126
#>  [811,]    0.2442133    0.2444838  0.2445860
#>  [812,]    0.2008673    0.2010898  0.2011738
#>  [813,]    0.1932709    0.1934849  0.1935658
#>  [814,]    0.2062837    0.2065122  0.2065986
#>  [815,]    0.2611326    0.2614218  0.2615311
#>  [816,]    0.1432150    0.1433737  0.1434336
#>  [817,]    0.2226652    0.2229118  0.2230050
#>  [818,]    0.1761413    0.1763364  0.1764101
#>  [819,]    0.2577954    0.2580810  0.2581889
#>  [820,]    0.1970864    0.1973047  0.1973872
#>  [821,]    0.2083321    0.2085628  0.2086500
#>  [822,]    0.2236269    0.2238747  0.2239682
#>  [823,]    0.1770439    0.1772401  0.1773141
#>  [824,]    0.2524434    0.2527231  0.2528287
#>  [825,]    0.2742249    0.2745287  0.2746434
#>  [826,]    0.3323477    0.3327159  0.3328549
#>  [827,]    0.1900371    0.1902476  0.1903271
#>  [828,]    0.2278888    0.2281412  0.2282366
#>  [829,]    0.1917100    0.1919224  0.1920026
#>  [830,]    0.2239739    0.2242220  0.2243158
#>  [831,]    0.2413359    0.2416032  0.2417042
#>  [832,]    0.2505932    0.2508708  0.2509757
#>  [833,]    0.2031521    0.2033771  0.2034621
#>  [834,]    0.1854371    0.1856425  0.1857201
#>  [835,]    0.1483554    0.1485198  0.1485818
#>  [836,]    0.1585606    0.1587362  0.1588026
#>  [837,]    0.1584494    0.1586250  0.1586913
#>  [838,]    0.1869582    0.1871653  0.1872435
#>  [839,]    0.2272487    0.2275004  0.2275955
#>  [840,]    0.2172176    0.2174582  0.2175491
#>  [841,]    0.2054089    0.2056365  0.2057224
#>  [842,]    0.2515237    0.2518023  0.2519076
#>  [843,]    0.2214474    0.2216928  0.2217854
#>  [844,]    0.2803197    0.2806303  0.2807475
#>  [845,]    0.3051693    0.3055074  0.3056351
#>  [846,]    0.1672552    0.1674404  0.1675104
#>  [847,]    0.2164892    0.2167290  0.2168196
#>  [848,]    0.2121116    0.2123466  0.2124353
#>  [849,]    0.2630393    0.2633306  0.2634407
#>  [850,]    0.2532231    0.2535037  0.2536096
#>  [851,]    0.2126684    0.2129040  0.2129930
#>  [852,]    0.2123322    0.2125674  0.2126562
#>  [853,]    0.1862240    0.1864303  0.1865083
#>  [854,]    0.2219274    0.2221732  0.2222661
#>  [855,]    0.1969117    0.1971298  0.1972122
#>  [856,]    0.2127663    0.2130020  0.2130910
#>  [857,]    0.2964860    0.2968144  0.2969384
#>  [858,]    0.2278231    0.2280755  0.2281708
#>  [859,]    0.3308711    0.3312376  0.3313761
#>  [860,]    0.1361805    0.1363313  0.1363883
#>  [861,]    0.2096519    0.2098842  0.2099719
#>  [862,]    0.1993079    0.1995287  0.1996121
#>  [863,]    0.1596696    0.1598465  0.1599133
#>  [864,]    0.2615766    0.2618664  0.2619759
#>  [865,]    0.2929312    0.2932557  0.2933783
#>  [866,]    0.1637336    0.1639150  0.1639835
#>  [867,]    0.2787439    0.2790526  0.2791693
#>  [868,]    0.1902523    0.1904630  0.1905426
#>  [869,]    0.1900708    0.1902813  0.1903608
#>  [870,]    0.1915890    0.1918012  0.1918813
#>  [871,]    0.1702780    0.1704666  0.1705379
#>  [872,]    0.2164817    0.2167215  0.2168121
#>  [873,]    0.2252713    0.2255208  0.2256151
#>  [874,]    0.2317492    0.2320060  0.2321029
#>  [875,]    0.1508329    0.1510000  0.1510631
#>  [876,]    0.3113097    0.3116545  0.3117848
#>  [877,]    0.2118715    0.2121062  0.2121949
#>  [878,]    0.2367120    0.2369742  0.2370732
#>  [879,]    0.2076959    0.2079260  0.2080129
#>  [880,]    0.2617384    0.2620284  0.2621379
#>  [881,]    0.2501070    0.2503841  0.2504887
#>  [882,]    0.2332030    0.2334613  0.2335589
#>  [883,]    0.2379497    0.2382133  0.2383128
#>  [884,]    0.2353958    0.2356565  0.2357550
#>  [885,]    0.1893856    0.1895954  0.1896747
#>  [886,]    0.2266291    0.2268801  0.2269750
#>  [887,]    0.2260033    0.2262537  0.2263483
#>  [888,]    0.2025081    0.2027324  0.2028172
#>  [889,]    0.2012033    0.2014262  0.2015103
#>  [890,]    0.2135919    0.2138285  0.2139179
#>  [891,]    0.2420130    0.2422810  0.2423823
#>  [892,]    0.1846130    0.1848175  0.1848947
#>  [893,]    0.2203285    0.2205726  0.2206648
#>  [894,]    0.1763304    0.1765258  0.1765996
#>  [895,]    0.2706697    0.2709696  0.2710828
#>  [896,]    0.2893245    0.2896450  0.2897660
#>  [897,]    0.2489229    0.2491986  0.2493028
#>  [898,]    0.1508318    0.1509989  0.1510620
#>  [899,]    0.1795118    0.1797107  0.1797858
#>  [900,]    0.1711036    0.1712932  0.1713648
#>  [901,]    0.1865325    0.1867391  0.1868172
#>  [902,]    0.1789604    0.1791587  0.1792335
#>  [903,]    0.2319503    0.2322072  0.2323043
#>  [904,]    0.2383061    0.2385700  0.2386698
#>  [905,]    0.2552806    0.2555634  0.2556702
#>  [906,]    0.1810500    0.1812506  0.1813263
#>  [907,]    0.2737858    0.2740891  0.2742036
#>  [908,]    0.2686678    0.2689654  0.2690778
#>  [909,]    0.1808406    0.1810409  0.1811166
#>  [910,]    0.1810874    0.1812880  0.1813638
#>  [911,]    0.2159811    0.2162203  0.2163107
#>  [912,]    0.2014264    0.2016495  0.2017338
#>  [913,]    0.3446230    0.3450047  0.3451489
#>  [914,]    0.1685151    0.1687018  0.1687723
#>  [915,]    0.1375978    0.1377502  0.1378078
#>  [916,]    0.2391544    0.2394193  0.2395194
#>  [917,]    0.1857002    0.1859059  0.1859836
#>  [918,]    0.2036925    0.2039181  0.2040033
#>  [919,]    0.2294706    0.2297248  0.2298208
#>  [920,]    0.1825027    0.1827049  0.1827812
#>  [921,]    0.2295054    0.2297597  0.2298557
#>  [922,]    0.2333781    0.2336367  0.2337343
#>  [923,]    0.1837246    0.1839281  0.1840050
#>  [924,]    0.3301751    0.3305408  0.3306790
#>  [925,]    0.2486515    0.2489270  0.2490310
#>  [926,]    0.2687163    0.2690140  0.2691264
#>  [927,]    0.2206160    0.2208604  0.2209527
#>  [928,]    0.1976589    0.1978778  0.1979605
#>  [929,]    0.2530698    0.2533501  0.2534560
#>  [930,]    0.2694113    0.2697097  0.2698224
#>  [931,]    0.2237545    0.2240024  0.2240960
#>  [932,]    0.2292024    0.2294563  0.2295522
#>  [933,]    0.2131042    0.2133402  0.2134294
#>  [934,]    0.1826209    0.1828232  0.1828996
#>  [935,]    0.2234873    0.2237349  0.2238284
#>  [936,]    0.3575770    0.3579731  0.3581227
#>  [937,]    0.2086591    0.2088903  0.2089776
#>  [938,]    0.1991114    0.1993320  0.1994153
#>  [939,]    0.1923804    0.1925935  0.1926740
#>  [940,]    0.2195884    0.2198317  0.2199235
#>  [941,]    0.2292836    0.2295375  0.2296335
#>  [942,]    0.2049206    0.2051476  0.2052333
#>  [943,]    0.2128056    0.2130413  0.2131304
#>  [944,]    0.2038432    0.2040690  0.2041543
#>  [945,]    0.2029907    0.2032156  0.2033005
#>  [946,]    0.2265117    0.2267626  0.2268574
#>  [947,]    0.2317387    0.2319954  0.2320924
#>  [948,]    0.2159576    0.2161968  0.2162871
#>  [949,]    0.2546337    0.2549158  0.2550223
#>  [950,]    0.1901796    0.1903903  0.1904699
#>  [951,]    0.1732443    0.1734362  0.1735087
#>  [952,]    0.1813824    0.1815833  0.1816592
#>  [953,]    0.1897746    0.1899848  0.1900642
#>  [954,]    0.2118128    0.2120474  0.2121360
#>  [955,]    0.3124857    0.3128319  0.3129626
#>  [956,]    0.2282760    0.2285289  0.2286244
#>  [957,]    0.2085014    0.2087323  0.2088196
#>  [958,]    0.2138670    0.2141039  0.2141934
#>  [959,]    0.1593756    0.1595522  0.1596188
#>  [960,]    0.2238951    0.2241431  0.2242368
#>  [961,]    0.2641574    0.2644501  0.2645606
#>  [962,]    0.2179825    0.2182239  0.2183151
#>  [963,]    0.1970212    0.1972394  0.1973218
#>  [964,]    0.2690277    0.2693257  0.2694383
#>  [965,]    0.2028141    0.2030388  0.2031236
#>  [966,]    0.2838495    0.2841639  0.2842827
#>  [967,]    0.2253045    0.2255541  0.2256484
#>  [968,]    0.3323518    0.3327200  0.3328590
#>  [969,]    0.2400370    0.2403029  0.2404033
#>  [970,]    0.2768783    0.2771850  0.2773009
#>  [971,]    0.1759962    0.1761912  0.1762648
#>  [972,]    0.2714271    0.2717278  0.2718413
#>  [973,]    0.1875115    0.1877192  0.1877977
#>  [974,]    0.1954196    0.1956360  0.1957178
#>  [975,]    0.2972700    0.2975993  0.2977237
#>  [976,]    0.1946165    0.1948321  0.1949135
#>  [977,]    0.2244339    0.2246825  0.2247764
#>  [978,]    0.2415073    0.2417748  0.2418759
#>  [979,]    0.2108215    0.2110551  0.2111433
#>  [980,]    0.1821271    0.1823288  0.1824050
#>  [981,]    0.1858281    0.1860340  0.1861117
#>  [982,]    0.2543976    0.2546794  0.2547859
#>  [983,]    0.3127666    0.3131131  0.3132439
#>  [984,]    0.2262152    0.2264658  0.2265605
#>  [985,]    0.2795269    0.2798366  0.2799535
#>  [986,]    0.1824776    0.1826797  0.1827561
#>  [987,]    0.3258083    0.3261692  0.3263055
#>  [988,]    0.1842637    0.1844678  0.1845449
#>  [989,]    0.1651445    0.1653275  0.1653966
#>  [990,]    0.2102600    0.2104929  0.2105809
#>  [991,]    0.2102878    0.2105207  0.2106087
#>  [992,]    0.1828334    0.1830359  0.1831124
#>  [993,]    0.3000999    0.3004324  0.3005579
#>  [994,]    0.3390459    0.3394215  0.3395633
#>  [995,]    0.2280103    0.2282629  0.2283583
#>  [996,]    0.2246030    0.2248518  0.2249458
#>  [997,]    0.1916227    0.1918350  0.1919151
#>  [998,]    0.2184381    0.2186801  0.2187715
#>  [999,]    0.2592909    0.2595782  0.2596866
#> [1000,]    0.1999719    0.2001934  0.2002771