Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 95.67564
#> 2   Feb 93.50597
#> 3   Mar 88.42397
#> 4   Apr 85.09712
#> 5   May 93.53181
#> 6   Jun 89.52200

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 6.840244  2
#> 2   Feb 9.972478  2
#> 3   Mar 8.857493  2
#> 4   Apr 8.608232  2
#> 5   May 6.438759  2
#> 6   Jun 7.217352  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [10ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [19ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [59ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [618ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [16ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1884232    0.1999688  0.1985556
#>    [2,]    0.1948599    0.2067999  0.2053384
#>    [3,]    0.2170002    0.2302967  0.2286692
#>    [4,]    0.1824938    0.1936761  0.1923073
#>    [5,]    0.1993368    0.2115511  0.2100560
#>    [6,]    0.2614620    0.2774826  0.2755217
#>    [7,]    0.2164314    0.2296931  0.2280698
#>    [8,]    0.1682159    0.1785233  0.1772616
#>    [9,]    0.2369489    0.2514675  0.2496904
#>   [10,]    0.3664361    0.3888879  0.3861397
#>   [11,]    0.2913602    0.3092124  0.3070272
#>   [12,]    0.2279413    0.2419081  0.2401985
#>   [13,]    0.1875190    0.1990093  0.1976028
#>   [14,]    0.1747444    0.1854519  0.1841412
#>   [15,]    0.2561466    0.2718415  0.2699204
#>   [16,]    0.2092591    0.2220813  0.2205118
#>   [17,]    0.2368809    0.2513955  0.2496188
#>   [18,]    0.2251292    0.2389238  0.2372353
#>   [19,]    0.1407945    0.1494218  0.1483658
#>   [20,]    0.2411659    0.2559431  0.2541344
#>   [21,]    0.1722174    0.1827701  0.1814784
#>   [22,]    0.2371865    0.2517197  0.2499408
#>   [23,]    0.2030492    0.2154910  0.2139681
#>   [24,]    0.1965524    0.2085962  0.2071220
#>   [25,]    0.2174460    0.2307698  0.2291389
#>   [26,]    0.1805970    0.1916631  0.1903086
#>   [27,]    0.2621946    0.2782602  0.2762937
#>   [28,]    0.2295696    0.2436363  0.2419145
#>   [29,]    0.1949812    0.2069285  0.2054661
#>   [30,]    0.1644756    0.1745536  0.1733200
#>   [31,]    0.2463164    0.2614092  0.2595618
#>   [32,]    0.2332213    0.2475119  0.2457626
#>   [33,]    0.2369857    0.2515066  0.2497292
#>   [34,]    0.2991715    0.3175024  0.3152587
#>   [35,]    0.1972902    0.2093791  0.2078994
#>   [36,]    0.2102371    0.2231192  0.2215424
#>   [37,]    0.2132379    0.2263040  0.2247047
#>   [38,]    0.1440144    0.1528389  0.1517587
#>   [39,]    0.2220258    0.2356302  0.2339650
#>   [40,]    0.3105100    0.3295355  0.3272067
#>   [41,]    0.2161023    0.2293440  0.2277232
#>   [42,]    0.1921893    0.2039656  0.2025242
#>   [43,]    0.3383761    0.3591091  0.3565713
#>   [44,]    0.1797327    0.1907457  0.1893976
#>   [45,]    0.2424090    0.2572624  0.2554443
#>   [46,]    0.1420207    0.1507230  0.1496578
#>   [47,]    0.2809038    0.2981156  0.2960088
#>   [48,]    0.1689106    0.1792607  0.1779938
#>   [49,]    0.2082176    0.2209760  0.2194143
#>   [50,]    0.1865598    0.1979911  0.1965919
#>   [51,]    0.1946083    0.2065329  0.2050733
#>   [52,]    0.2120083    0.2249990  0.2234089
#>   [53,]    0.2151114    0.2282922  0.2266788
#>   [54,]    0.2509452    0.2663216  0.2644395
#>   [55,]    0.2553401    0.2709858  0.2690707
#>   [56,]    0.2806505    0.2978468  0.2957419
#>   [57,]    0.2497503    0.2650534  0.2631803
#>   [58,]    0.1762513    0.1870510  0.1857291
#>   [59,]    0.2292973    0.2433472  0.2416274
#>   [60,]    0.1650528    0.1751665  0.1739286
#>   [61,]    0.2492117    0.2644818  0.2626127
#>   [62,]    0.1941411    0.2060370  0.2045809
#>   [63,]    0.1621487    0.1720845  0.1708683
#>   [64,]    0.2394333    0.2541044  0.2523086
#>   [65,]    0.1873673    0.1988482  0.1974429
#>   [66,]    0.2313702    0.2455472  0.2438119
#>   [67,]    0.2234633    0.2371559  0.2354798
#>   [68,]    0.2241187    0.2378513  0.2361704
#>   [69,]    0.1882578    0.1997933  0.1983813
#>   [70,]    0.2437265    0.2586604  0.2568324
#>   [71,]    0.1786372    0.1895831  0.1882433
#>   [72,]    0.1791185    0.1900939  0.1887505
#>   [73,]    0.2228148    0.2364677  0.2347965
#>   [74,]    0.2226654    0.2363091  0.2346391
#>   [75,]    0.2625651    0.2786534  0.2766841
#>   [76,]    0.1866241    0.1980595  0.1966598
#>   [77,]    0.2406675    0.2554141  0.2536090
#>   [78,]    0.1789738    0.1899404  0.1885981
#>   [79,]    0.2057797    0.2183888  0.2168454
#>   [80,]    0.1898745    0.2015089  0.2000848
#>   [81,]    0.1817648    0.1929025  0.1915392
#>   [82,]    0.2724765    0.2891719  0.2871283
#>   [83,]    0.3003347    0.3187372  0.3164847
#>   [84,]    0.1787335    0.1896854  0.1883448
#>   [85,]    0.1801955    0.1912371  0.1898856
#>   [86,]    0.1703246    0.1807613  0.1794838
#>   [87,]    0.2324620    0.2467058  0.2449623
#>   [88,]    0.1841330    0.1954157  0.1940346
#>   [89,]    0.2862738    0.3038148  0.3016677
#>   [90,]    0.1355670    0.1438740  0.1428572
#>   [91,]    0.2974906    0.3157187  0.3134875
#>   [92,]    0.1485034    0.1576030  0.1564891
#>   [93,]    0.2177470    0.2310893  0.2294562
#>   [94,]    0.2338457    0.2481742  0.2464203
#>   [95,]    0.2364394    0.2509271  0.2491537
#>   [96,]    0.2149307    0.2281004  0.2264883
#>   [97,]    0.1896307    0.2012503  0.1998280
#>   [98,]    0.2348498    0.2492399  0.2474785
#>   [99,]    0.1956143    0.2076005  0.2061333
#>  [100,]    0.2089222    0.2217238  0.2201568
#>  [101,]    0.2214690    0.2350393  0.2333783
#>  [102,]    0.1819252    0.1930727  0.1917082
#>  [103,]    0.2028635    0.2152936  0.2137721
#>  [104,]    0.1226757    0.1301927  0.1292726
#>  [105,]    0.2506226    0.2659793  0.2640996
#>  [106,]    0.2108121    0.2237293  0.2221482
#>  [107,]    0.1793205    0.1903083  0.1889634
#>  [108,]    0.1096277    0.1163452  0.1155230
#>  [109,]    0.2478847    0.2630737  0.2612145
#>  [110,]    0.2130031    0.2260546  0.2244571
#>  [111,]    0.2794033    0.2965232  0.2944276
#>  [112,]    0.3367229    0.3573543  0.3548290
#>  [113,]    0.2029607    0.2153970  0.2138748
#>  [114,]    0.2026069    0.2150216  0.2135019
#>  [115,]    0.1903610    0.2020253  0.2005976
#>  [116,]    0.1364228    0.1447822  0.1437590
#>  [117,]    0.1837482    0.1950074  0.1936293
#>  [118,]    0.1622002    0.1721392  0.1709226
#>  [119,]    0.1898140    0.2014447  0.2000210
#>  [120,]    0.1734009    0.1840261  0.1827255
#>  [121,]    0.2072702    0.2199707  0.2184161
#>  [122,]    0.1691693    0.1795351  0.1782663
#>  [123,]    0.1510728    0.1603298  0.1591967
#>  [124,]    0.2487337    0.2639741  0.2621086
#>  [125,]    0.2977903    0.3160367  0.3138033
#>  [126,]    0.2494821    0.2647690  0.2628978
#>  [127,]    0.2104939    0.2233917  0.2218130
#>  [128,]    0.2620062    0.2780602  0.2760951
#>  [129,]    0.4840861    0.5137462  0.5101157
#>  [130,]    0.2146678    0.2278215  0.2262115
#>  [131,]    0.2076497    0.2203732  0.2188158
#>  [132,]    0.2061639    0.2187964  0.2172501
#>  [133,]    0.1732501    0.1838660  0.1825666
#>  [134,]    0.1475354    0.1565756  0.1554691
#>  [135,]    0.2246563    0.2384220  0.2367371
#>  [136,]    0.2096714    0.2225189  0.2209464
#>  [137,]    0.2535756    0.2691131  0.2672113
#>  [138,]    0.2090445    0.2218535  0.2202857
#>  [139,]    0.1651345    0.1752532  0.1740146
#>  [140,]    0.1519692    0.1612811  0.1601413
#>  [141,]    0.2085698    0.2213498  0.2197855
#>  [142,]    0.1459743    0.1549190  0.1538242
#>  [143,]    0.1431683    0.1519410  0.1508672
#>  [144,]    0.1548216    0.1643084  0.1631471
#>  [145,]    0.2861152    0.3036463  0.3015005
#>  [146,]    0.2399179    0.2546184  0.2528190
#>  [147,]    0.2342639    0.2486180  0.2468610
#>  [148,]    0.1742608    0.1849387  0.1836317
#>  [149,]    0.1909892    0.2026920  0.2012596
#>  [150,]    0.2764042    0.2933404  0.2912674
#>  [151,]    0.2175897    0.2309222  0.2292903
#>  [152,]    0.1704585    0.1809034  0.1796249
#>  [153,]    0.1996601    0.2118942  0.2103967
#>  [154,]    0.2728620    0.2895812  0.2875347
#>  [155,]    0.2011891    0.2135171  0.2120081
#>  [156,]    0.3242989    0.3441693  0.3417371
#>  [157,]    0.1355600    0.1438665  0.1428497
#>  [158,]    0.2031038    0.2155489  0.2140256
#>  [159,]    0.1900577    0.2017034  0.2002780
#>  [160,]    0.2357335    0.2501779  0.2484098
#>  [161,]    0.1501569    0.1593579  0.1582317
#>  [162,]    0.2449460    0.2599547  0.2581176
#>  [163,]    0.3058695    0.3246109  0.3223169
#>  [164,]    0.1976152    0.2097241  0.2082419
#>  [165,]    0.1778100    0.1887054  0.1873717
#>  [166,]    0.3004976    0.3189100  0.3166562
#>  [167,]    0.2301068    0.2442064  0.2424806
#>  [168,]    0.2098143    0.2226705  0.2210969
#>  [169,]    0.2327760    0.2470392  0.2452933
#>  [170,]    0.1909196    0.2026180  0.2011861
#>  [171,]    0.2727085    0.2894184  0.2873730
#>  [172,]    0.1775917    0.1884737  0.1871417
#>  [173,]    0.2221750    0.2357887  0.2341223
#>  [174,]    0.2252599    0.2390624  0.2373729
#>  [175,]    0.2332060    0.2474954  0.2457463
#>  [176,]    0.2594499    0.2753473  0.2734014
#>  [177,]    0.2229454    0.2366063  0.2349341
#>  [178,]    0.2870368    0.3046243  0.3024715
#>  [179,]    0.1977622    0.2098800  0.2083967
#>  [180,]    0.1889111    0.2004866  0.1990698
#>  [181,]    0.1831029    0.1943225  0.1929492
#>  [182,]    0.1891625    0.2007533  0.1993346
#>  [183,]    0.1910191    0.2027237  0.2012910
#>  [184,]    0.1323375    0.1404466  0.1394540
#>  [185,]    0.2115490    0.2245114  0.2229248
#>  [186,]    0.2534252    0.2689535  0.2670527
#>  [187,]    0.1785689    0.1895106  0.1881713
#>  [188,]    0.2065442    0.2192001  0.2176510
#>  [189,]    0.2314846    0.2456686  0.2439324
#>  [190,]    0.2043929    0.2169170  0.2153840
#>  [191,]    0.1765981    0.1874191  0.1860946
#>  [192,]    0.1997045    0.2119413  0.2104434
#>  [193,]    0.1728197    0.1834093  0.1821131
#>  [194,]    0.1829558    0.1941663  0.1927941
#>  [195,]    0.1999985    0.2122532  0.2107532
#>  [196,]    0.2172188    0.2305288  0.2288996
#>  [197,]    0.1962927    0.2083206  0.2068483
#>  [198,]    0.2690329    0.2855176  0.2834998
#>  [199,]    0.2024833    0.2148905  0.2133718
#>  [200,]    0.2003546    0.2126313  0.2111286
#>  [201,]    0.2013966    0.2137370  0.2122265
#>  [202,]    0.1670631    0.1772998  0.1760468
#>  [203,]    0.2559589    0.2716425  0.2697228
#>  [204,]    0.1544974    0.1639643  0.1628055
#>  [205,]    0.2554675    0.2711211  0.2692050
#>  [206,]    0.2523697    0.2678333  0.2659405
#>  [207,]    0.2514389    0.2668453  0.2649595
#>  [208,]    0.2719738    0.2886386  0.2865988
#>  [209,]    0.1950260    0.2069762  0.2055135
#>  [210,]    0.2084178    0.2211884  0.2196252
#>  [211,]    0.2071484    0.2198414  0.2182877
#>  [212,]    0.2069916    0.2196749  0.2181224
#>  [213,]    0.1647949    0.1748926  0.1736566
#>  [214,]    0.1746043    0.1853032  0.1839936
#>  [215,]    0.2272396    0.2411637  0.2394594
#>  [216,]    0.2220145    0.2356183  0.2339531
#>  [217,]    0.2003226    0.2125972  0.2110948
#>  [218,]    0.2224137    0.2360420  0.2343738
#>  [219,]    0.1931649    0.2050010  0.2035523
#>  [220,]    0.2302786    0.2443887  0.2426616
#>  [221,]    0.1975224    0.2096255  0.2081440
#>  [222,]    0.2086677    0.2214537  0.2198887
#>  [223,]    0.2779905    0.2950241  0.2929391
#>  [224,]    0.2007941    0.2130976  0.2115916
#>  [225,]    0.2029149    0.2153485  0.2138266
#>  [226,]    0.1997142    0.2119516  0.2104537
#>  [227,]    0.1981295    0.2102698  0.2087838
#>  [228,]    0.2420444    0.2568755  0.2550602
#>  [229,]    0.2091415    0.2219566  0.2203880
#>  [230,]    0.1689370    0.1792887  0.1780216
#>  [231,]    0.2591911    0.2750726  0.2731287
#>  [232,]    0.2194690    0.2329165  0.2312705
#>  [233,]    0.1918617    0.2036181  0.2021790
#>  [234,]    0.2616153    0.2776455  0.2756833
#>  [235,]    0.2552026    0.2708400  0.2689259
#>  [236,]    0.2431618    0.2580613  0.2562375
#>  [237,]    0.3298018    0.3500096  0.3475361
#>  [238,]    0.2556226    0.2712855  0.2693683
#>  [239,]    0.2938333    0.3118375  0.3096337
#>  [240,]    0.2859873    0.3035105  0.3013656
#>  [241,]    0.2011241    0.2134479  0.2119394
#>  [242,]    0.1791784    0.1901575  0.1888136
#>  [243,]    0.2615502    0.2775761  0.2756145
#>  [244,]    0.2723014    0.2889861  0.2869438
#>  [245,]    0.1582417    0.1679380  0.1667512
#>  [246,]    0.2867631    0.3043340  0.3021833
#>  [247,]    0.2651099    0.2813540  0.2793656
#>  [248,]    0.1857414    0.1971228  0.1957296
#>  [249,]    0.1991635    0.2113673  0.2098735
#>  [250,]    0.2139499    0.2270595  0.2254548
#>  [251,]    0.2293150    0.2433660  0.2416461
#>  [252,]    0.1633483    0.1733576  0.1721324
#>  [253,]    0.1887939    0.2003621  0.1989461
#>  [254,]    0.2380779    0.2526659  0.2508802
#>  [255,]    0.2584093    0.2742426  0.2723046
#>  [256,]    0.1660540    0.1762290  0.1749836
#>  [257,]    0.2006121    0.2129044  0.2113998
#>  [258,]    0.1868691    0.1983195  0.1969180
#>  [259,]    0.2419636    0.2567897  0.2549749
#>  [260,]    0.2155745    0.2287838  0.2271669
#>  [261,]    0.2224513    0.2360820  0.2344135
#>  [262,]    0.2932424    0.3112101  0.3090108
#>  [263,]    0.2896266    0.3073727  0.3052005
#>  [264,]    0.1612700    0.1711518  0.1699422
#>  [265,]    0.1433337    0.1521165  0.1510415
#>  [266,]    0.2064637    0.2191146  0.2175661
#>  [267,]    0.1597286    0.1695160  0.1683180
#>  [268,]    0.2215788    0.2351559  0.2334940
#>  [269,]    0.2581546    0.2739726  0.2720365
#>  [270,]    0.1694438    0.1798263  0.1785555
#>  [271,]    0.1877167    0.1992190  0.1978111
#>  [272,]    0.2860190    0.3035444  0.3013992
#>  [273,]    0.1669647    0.1771954  0.1759431
#>  [274,]    0.1786456    0.1895922  0.1882523
#>  [275,]    0.2374307    0.2519790  0.2501982
#>  [276,]    0.2667799    0.2831265  0.2811256
#>  [277,]    0.2825633    0.2998767  0.2977575
#>  [278,]    0.1762163    0.1870140  0.1856923
#>  [279,]    0.1842933    0.1955859  0.1942036
#>  [280,]    0.2679289    0.2843457  0.2823363
#>  [281,]    0.1764105    0.1872200  0.1858969
#>  [282,]    0.1800621    0.1910954  0.1897449
#>  [283,]    0.1851290    0.1964727  0.1950842
#>  [284,]    0.2505440    0.2658958  0.2640166
#>  [285,]    0.2741780    0.2909778  0.2889214
#>  [286,]    0.1678052    0.1780874  0.1768288
#>  [287,]    0.2000285    0.2122851  0.2107848
#>  [288,]    0.2188429    0.2322523  0.2306110
#>  [289,]    0.2288599    0.2428831  0.2411666
#>  [290,]    0.2323909    0.2466303  0.2448874
#>  [291,]    0.1905985    0.2022774  0.2008478
#>  [292,]    0.2085426    0.2213210  0.2197568
#>  [293,]    0.2039127    0.2164073  0.2148779
#>  [294,]    0.2072175    0.2199145  0.2183604
#>  [295,]    0.1969054    0.2089707  0.2074939
#>  [296,]    0.1961955    0.2082174  0.2067458
#>  [297,]    0.2551273    0.2707599  0.2688464
#>  [298,]    0.2005486    0.2128372  0.2113330
#>  [299,]    0.2074385    0.2201491  0.2185933
#>  [300,]    0.2337504    0.2480731  0.2463200
#>  [301,]    0.3007314    0.3191579  0.3169024
#>  [302,]    0.1393079    0.1478441  0.1467992
#>  [303,]    0.2640400    0.2802187  0.2782384
#>  [304,]    0.2172336    0.2305444  0.2289151
#>  [305,]    0.1848941    0.1962235  0.1948368
#>  [306,]    0.1926005    0.2044021  0.2029575
#>  [307,]    0.2759811    0.2928913  0.2908214
#>  [308,]    0.2500491    0.2653705  0.2634951
#>  [309,]    0.2283062    0.2422954  0.2405831
#>  [310,]    0.1868328    0.1982810  0.1968797
#>  [311,]    0.2840265    0.3014295  0.2992993
#>  [312,]    0.1853889    0.1967485  0.1953580
#>  [313,]    0.2393217    0.2539859  0.2521909
#>  [314,]    0.1963669    0.2083991  0.2069263
#>  [315,]    0.1573905    0.1670347  0.1658542
#>  [316,]    0.1764392    0.1872504  0.1859271
#>  [317,]    0.1964326    0.2084690  0.2069957
#>  [318,]    0.1604760    0.1703092  0.1691056
#>  [319,]    0.1161477    0.1232648  0.1223936
#>  [320,]    0.2012356    0.2135663  0.2120570
#>  [321,]    0.1931683    0.2050045  0.2035557
#>  [322,]    0.2537856    0.2693358  0.2674324
#>  [323,]    0.2219191    0.2355171  0.2338526
#>  [324,]    0.2003198    0.2125942  0.2110918
#>  [325,]    0.2104930    0.2233907  0.2218119
#>  [326,]    0.1838485    0.1951139  0.1937350
#>  [327,]    0.1581365    0.1678263  0.1666402
#>  [328,]    0.2220372    0.2356423  0.2339770
#>  [329,]    0.2178059    0.2311519  0.2295183
#>  [330,]    0.1677405    0.1780188  0.1767607
#>  [331,]    0.1612662    0.1711478  0.1699383
#>  [332,]    0.2119116    0.2248964  0.2233070
#>  [333,]    0.2879974    0.3056440  0.3034839
#>  [334,]    0.2222966    0.2359177  0.2342504
#>  [335,]    0.1958589    0.2078601  0.2063911
#>  [336,]    0.2422348    0.2570775  0.2552607
#>  [337,]    0.1452162    0.1541144  0.1530252
#>  [338,]    0.1813606    0.1924735  0.1911133
#>  [339,]    0.1867552    0.1981986  0.1967979
#>  [340,]    0.2068268    0.2195001  0.2179488
#>  [341,]    0.1571375    0.1667662  0.1655876
#>  [342,]    0.2139408    0.2270498  0.2254452
#>  [343,]    0.2403010    0.2550252  0.2532229
#>  [344,]    0.2458989    0.2609660  0.2591217
#>  [345,]    0.2245289    0.2382868  0.2366028
#>  [346,]    0.2227575    0.2364069  0.2347361
#>  [347,]    0.1910094    0.2027135  0.2012809
#>  [348,]    0.1955360    0.2075174  0.2060508
#>  [349,]    0.2365906    0.2510874  0.2493129
#>  [350,]    0.2284449    0.2424426  0.2407293
#>  [351,]    0.1647917    0.1748893  0.1736533
#>  [352,]    0.1875108    0.1990003  0.1975939
#>  [353,]    0.2259863    0.2398334  0.2381385
#>  [354,]    0.2762233    0.2931485  0.2910768
#>  [355,]    0.2377294    0.2522962  0.2505131
#>  [356,]    0.1655937    0.1757405  0.1744985
#>  [357,]    0.2067286    0.2193957  0.2178452
#>  [358,]    0.1931627    0.2049987  0.2035499
#>  [359,]    0.1654284    0.1755650  0.1743242
#>  [360,]    0.1358303    0.1441534  0.1431346
#>  [361,]    0.2254141    0.2392261  0.2375355
#>  [362,]    0.2347046    0.2490859  0.2473256
#>  [363,]    0.2360557    0.2505198  0.2487494
#>  [364,]    0.2051161    0.2176844  0.2161460
#>  [365,]    0.2044356    0.2169624  0.2154291
#>  [366,]    0.2488463    0.2640941  0.2622277
#>  [367,]    0.1396086    0.1481632  0.1471161
#>  [368,]    0.2113192    0.2242677  0.2226827
#>  [369,]    0.1521635    0.1614873  0.1603461
#>  [370,]    0.2064939    0.2191469  0.2175981
#>  [371,]    0.2699973    0.2865407  0.2845158
#>  [372,]    0.1274707    0.1352816  0.1343255
#>  [373,]    0.2511217    0.2665088  0.2646254
#>  [374,]    0.1896257    0.2012450  0.1998228
#>  [375,]    0.1997816    0.2120232  0.2105248
#>  [376,]    0.2350926    0.2494974  0.2477342
#>  [377,]    0.2464453    0.2615459  0.2596975
#>  [378,]    0.1494719    0.1586309  0.1575098
#>  [379,]    0.2140801    0.2271977  0.2255921
#>  [380,]    0.2168951    0.2301852  0.2285584
#>  [381,]    0.2598689    0.2757920  0.2738430
#>  [382,]    0.1423040    0.1510237  0.1499564
#>  [383,]    0.1990075    0.2112016  0.2097090
#>  [384,]    0.2015977    0.2139504  0.2124383
#>  [385,]    0.3085021    0.3274047  0.3250910
#>  [386,]    0.1508269    0.1600689  0.1589377
#>  [387,]    0.3189593    0.3385025  0.3361103
#>  [388,]    0.2030190    0.2154588  0.2139361
#>  [389,]    0.2067070    0.2193728  0.2178224
#>  [390,]    0.2935483    0.3115351  0.3093334
#>  [391,]    0.2075907    0.2203107  0.2187537
#>  [392,]    0.2802050    0.2973742  0.2952726
#>  [393,]    0.1984023    0.2105593  0.2090712
#>  [394,]    0.2152880    0.2284798  0.2268651
#>  [395,]    0.1674522    0.1777128  0.1764568
#>  [396,]    0.2924037    0.3103199  0.3081269
#>  [397,]    0.1922709    0.2040523  0.2026102
#>  [398,]    0.1828681    0.1940733  0.1927018
#>  [399,]    0.2009070    0.2132176  0.2117107
#>  [400,]    0.1729114    0.1835066  0.1822097
#>  [401,]    0.1949192    0.2068628  0.2054008
#>  [402,]    0.1843551    0.1956513  0.1942686
#>  [403,]    0.1712825    0.1817778  0.1804932
#>  [404,]    0.2657158    0.2819970  0.2800041
#>  [405,]    0.1907562    0.2024447  0.2010140
#>  [406,]    0.1641988    0.1742600  0.1730285
#>  [407,]    0.2110298    0.2239604  0.2223776
#>  [408,]    0.1801570    0.1911961  0.1898449
#>  [409,]    0.1989037    0.2110915  0.2095996
#>  [410,]    0.1642815    0.1743478  0.1731156
#>  [411,]    0.1824329    0.1936115  0.1922432
#>  [412,]    0.2010421    0.2133608  0.2118529
#>  [413,]    0.2120355    0.2250278  0.2234375
#>  [414,]    0.1727874    0.1833750  0.1820790
#>  [415,]    0.3042787    0.3229228  0.3206407
#>  [416,]    0.1389794    0.1474954  0.1464530
#>  [417,]    0.1566852    0.1662861  0.1651109
#>  [418,]    0.1808407    0.1919216  0.1905653
#>  [419,]    0.1939572    0.2058419  0.2043872
#>  [420,]    0.1673434    0.1775974  0.1763423
#>  [421,]    0.2189163    0.2323303  0.2306884
#>  [422,]    0.2379899    0.2525725  0.2507875
#>  [423,]    0.2355384    0.2499709  0.2482043
#>  [424,]    0.1971453    0.2092253  0.2077466
#>  [425,]    0.2137973    0.2268976  0.2252940
#>  [426,]    0.2141500    0.2272719  0.2256657
#>  [427,]    0.1720859    0.1826303  0.1813396
#>  [428,]    0.1509837    0.1602352  0.1591028
#>  [429,]    0.2511564    0.2665456  0.2646619
#>  [430,]    0.1488112    0.1579296  0.1568135
#>  [431,]    0.2972226    0.3154342  0.3132050
#>  [432,]    0.2092910    0.2221152  0.2205455
#>  [433,]    0.2848598    0.3023142  0.3001777
#>  [434,]    0.2550009    0.2706255  0.2687130
#>  [435,]    0.2817889    0.2990550  0.2969416
#>  [436,]    0.1920349    0.2038018  0.2023615
#>  [437,]    0.2643251    0.2805212  0.2785387
#>  [438,]    0.2082450    0.2210051  0.2194432
#>  [439,]    0.2058220    0.2184338  0.2168901
#>  [440,]    0.1995942    0.2118242  0.2103272
#>  [441,]    0.2079600    0.2207027  0.2191429
#>  [442,]    0.2060033    0.2186260  0.2170809
#>  [443,]    0.2743606    0.2911717  0.2891140
#>  [444,]    0.1659793    0.1761498  0.1749049
#>  [445,]    0.2051342    0.2177037  0.2161652
#>  [446,]    0.2076985    0.2204252  0.2188674
#>  [447,]    0.1913731    0.2030994  0.2016640
#>  [448,]    0.3031706    0.3217468  0.3194730
#>  [449,]    0.3138311    0.3330602  0.3307065
#>  [450,]    0.1697355    0.1801360  0.1788629
#>  [451,]    0.1770597    0.1879091  0.1865811
#>  [452,]    0.1717104    0.1822320  0.1809442
#>  [453,]    0.1822429    0.1934098  0.1920429
#>  [454,]    0.1673579    0.1776128  0.1763576
#>  [455,]    0.2047969    0.2173458  0.2158097
#>  [456,]    0.2021630    0.2145504  0.2130341
#>  [457,]    0.1597844    0.1695752  0.1683768
#>  [458,]    0.2094803    0.2223160  0.2207449
#>  [459,]    0.2390860    0.2537357  0.2519425
#>  [460,]    0.2281798    0.2421614  0.2404500
#>  [461,]    0.2103120    0.2231987  0.2216214
#>  [462,]    0.1540732    0.1635140  0.1623584
#>  [463,]    0.1608309    0.1706860  0.1694797
#>  [464,]    0.1964510    0.2084884  0.2070150
#>  [465,]    0.1981555    0.2102975  0.2088113
#>  [466,]    0.2507220    0.2660848  0.2642043
#>  [467,]    0.1993089    0.2115215  0.2100266
#>  [468,]    0.1774248    0.1882965  0.1869657
#>  [469,]    0.1792825    0.1902680  0.1889233
#>  [470,]    0.2857592    0.3032686  0.3011254
#>  [471,]    0.2378849    0.2524611  0.2506769
#>  [472,]    0.1254359    0.1331220  0.1321812
#>  [473,]    0.1836271    0.1948788  0.1935016
#>  [474,]    0.1552948    0.1648104  0.1636457
#>  [475,]    0.2241022    0.2378339  0.2361531
#>  [476,]    0.1886437    0.2002028  0.1987879
#>  [477,]    0.2352083    0.2496204  0.2478563
#>  [478,]    0.1994178    0.2116370  0.2101413
#>  [479,]    0.1884336    0.1999799  0.1985666
#>  [480,]    0.3254723    0.3454147  0.3429737
#>  [481,]    0.1751461    0.1858781  0.1845645
#>  [482,]    0.1814838    0.1926042  0.1912430
#>  [483,]    0.2301257    0.2442265  0.2425005
#>  [484,]    0.1505892    0.1598166  0.1586872
#>  [485,]    0.1277844    0.1356145  0.1346561
#>  [486,]    0.1925374    0.2043351  0.2028911
#>  [487,]    0.1475545    0.1565959  0.1554892
#>  [488,]    0.2040656    0.2165694  0.2150389
#>  [489,]    0.2641492    0.2803346  0.2783535
#>  [490,]    0.1942496    0.2061521  0.2046952
#>  [491,]    0.1701975    0.1806263  0.1793498
#>  [492,]    0.2255735    0.2393952  0.2377034
#>  [493,]    0.1786182    0.1895631  0.1882234
#>  [494,]    0.2177889    0.2311339  0.2295004
#>  [495,]    0.2625874    0.2786769  0.2767075
#>  [496,]    0.1833963    0.1946339  0.1932584
#>  [497,]    0.2158752    0.2291028  0.2274837
#>  [498,]    0.2241993    0.2379368  0.2362553
#>  [499,]    0.2383759    0.2529821  0.2511943
#>  [500,]    0.1799276    0.1909528  0.1896033
#>  [501,]    0.2634243    0.2795650  0.2775894
#>  [502,]    0.1910993    0.2028089  0.2013756
#>  [503,]    0.2222996    0.2359206  0.2342534
#>  [504,]    0.1665598    0.1767658  0.1755166
#>  [505,]    0.1466535    0.1556397  0.1545398
#>  [506,]    0.1795097    0.1905089  0.1891626
#>  [507,]    0.2287762    0.2427941  0.2410782
#>  [508,]    0.1618856    0.1718052  0.1705910
#>  [509,]    0.1751847    0.1859192  0.1846053
#>  [510,]    0.2327809    0.2470443  0.2452984
#>  [511,]    0.2217234    0.2353094  0.2336464
#>  [512,]    0.1751057    0.1858353  0.1845219
#>  [513,]    0.2036257    0.2161028  0.2145755
#>  [514,]    0.1938442    0.2057217  0.2042679
#>  [515,]    0.1636534    0.1736813  0.1724539
#>  [516,]    0.2671905    0.2835621  0.2815581
#>  [517,]    0.2121337    0.2251320  0.2235410
#>  [518,]    0.1618490    0.1717663  0.1705523
#>  [519,]    0.1967020    0.2087547  0.2072794
#>  [520,]    0.1542649    0.1637174  0.1625604
#>  [521,]    0.2904581    0.3082552  0.3060768
#>  [522,]    0.2380019    0.2525854  0.2508003
#>  [523,]    0.1982065    0.2103516  0.2088650
#>  [524,]    0.2015109    0.2138583  0.2123469
#>  [525,]    0.2589841    0.2748531  0.2729107
#>  [526,]    0.1894099    0.2010160  0.1995954
#>  [527,]    0.2095839    0.2224260  0.2208540
#>  [528,]    0.1987367    0.2109143  0.2094237
#>  [529,]    0.1386883    0.1471864  0.1461462
#>  [530,]    0.2075265    0.2202426  0.2186861
#>  [531,]    0.1843606    0.1956572  0.1942745
#>  [532,]    0.1877068    0.1992084  0.1978005
#>  [533,]    0.1761353    0.1869280  0.1856069
#>  [534,]    0.2171492    0.2304548  0.2288261
#>  [535,]    0.2171368    0.2304417  0.2288131
#>  [536,]    0.1862776    0.1976916  0.1962944
#>  [537,]    0.1880726    0.1995966  0.1981861
#>  [538,]    0.2118725    0.2248549  0.2232658
#>  [539,]    0.1839089    0.1951780  0.1937986
#>  [540,]    0.1926666    0.2044721  0.2030271
#>  [541,]    0.1679204    0.1782098  0.1769504
#>  [542,]    0.2384263    0.2530355  0.2512473
#>  [543,]    0.2411067    0.2558803  0.2540720
#>  [544,]    0.2390646    0.2537129  0.2519199
#>  [545,]    0.2155016    0.2287064  0.2270901
#>  [546,]    0.1888542    0.2004260  0.1990096
#>  [547,]    0.2409916    0.2557580  0.2539506
#>  [548,]    0.1802407    0.1912849  0.1899331
#>  [549,]    0.2345474    0.2489190  0.2471598
#>  [550,]    0.2641018    0.2802842  0.2783035
#>  [551,]    0.3019675    0.3204698  0.3182050
#>  [552,]    0.2959768    0.3141119  0.3118921
#>  [553,]    0.1673411    0.1775950  0.1763399
#>  [554,]    0.1795187    0.1905187  0.1891723
#>  [555,]    0.2251865    0.2389847  0.2372957
#>  [556,]    0.2064848    0.2191370  0.2175883
#>  [557,]    0.2347821    0.2491681  0.2474072
#>  [558,]    0.2309144    0.2450632  0.2433313
#>  [559,]    0.2568369    0.2725741  0.2706478
#>  [560,]    0.1907942    0.2024850  0.2010540
#>  [561,]    0.2055601    0.2181558  0.2166140
#>  [562,]    0.1796132    0.1906189  0.1892718
#>  [563,]    0.1618796    0.1717988  0.1705846
#>  [564,]    0.1762101    0.1870074  0.1856858
#>  [565,]    0.2200144    0.2334956  0.2318455
#>  [566,]    0.1505299    0.1597536  0.1586246
#>  [567,]    0.1928016    0.2046155  0.2031694
#>  [568,]    0.2344163    0.2487797  0.2470216
#>  [569,]    0.2669412    0.2832974  0.2812954
#>  [570,]    0.1549746    0.1644708  0.1633085
#>  [571,]    0.2164350    0.2296969  0.2280736
#>  [572,]    0.2014279    0.2137701  0.2122594
#>  [573,]    0.1802839    0.1913307  0.1899785
#>  [574,]    0.2476045    0.2627761  0.2609190
#>  [575,]    0.2482966    0.2635107  0.2616484
#>  [576,]    0.1936437    0.2055092  0.2040568
#>  [577,]    0.2225414    0.2361774  0.2345083
#>  [578,]    0.2009074    0.2132179  0.2117110
#>  [579,]    0.2227378    0.2363858  0.2347153
#>  [580,]    0.2019926    0.2143695  0.2128545
#>  [581,]    0.1747888    0.1854990  0.1841881
#>  [582,]    0.1989415    0.2111315  0.2096394
#>  [583,]    0.2145446    0.2276906  0.2260815
#>  [584,]    0.2158210    0.2290452  0.2274265
#>  [585,]    0.1798980    0.1909212  0.1895719
#>  [586,]    0.2185939    0.2319881  0.2303486
#>  [587,]    0.2441290    0.2590875  0.2572565
#>  [588,]    0.1863548    0.1977737  0.1963760
#>  [589,]    0.2421200    0.2569556  0.2551397
#>  [590,]    0.1520478    0.1613645  0.1602241
#>  [591,]    0.2101948    0.2230743  0.2214978
#>  [592,]    0.1684025    0.1787214  0.1774583
#>  [593,]    0.2563344    0.2720408  0.2701183
#>  [594,]    0.1560242    0.1655846  0.1644144
#>  [595,]    0.1811296    0.1922283  0.1908697
#>  [596,]    0.1901348    0.2017851  0.2003591
#>  [597,]    0.2004449    0.2127272  0.2112238
#>  [598,]    0.2383926    0.2530000  0.2512120
#>  [599,]    0.1769606    0.1878039  0.1864767
#>  [600,]    0.1457813    0.1547142  0.1536207
#>  [601,]    0.1392534    0.1477862  0.1467417
#>  [602,]    0.1881131    0.1996398  0.1982289
#>  [603,]    0.2509080    0.2662820  0.2644002
#>  [604,]    0.1850631    0.1964028  0.1950148
#>  [605,]    0.1861483    0.1975545  0.1961583
#>  [606,]    0.1844008    0.1956999  0.1943168
#>  [607,]    0.1991732    0.2113775  0.2098837
#>  [608,]    0.1959132    0.2079175  0.2064481
#>  [609,]    0.2053192    0.2179000  0.2163601
#>  [610,]    0.1822531    0.1934206  0.1920537
#>  [611,]    0.1623494    0.1722973  0.1710796
#>  [612,]    0.1461159    0.1550692  0.1539733
#>  [613,]    0.1938689    0.2057482  0.2042941
#>  [614,]    0.2287442    0.2427603  0.2410447
#>  [615,]    0.2126424    0.2256720  0.2240771
#>  [616,]    0.1557053    0.1652462  0.1640783
#>  [617,]    0.1652562    0.1753824  0.1741429
#>  [618,]    0.1864264    0.1978497  0.1964514
#>  [619,]    0.1861208    0.1975253  0.1961293
#>  [620,]    0.1961336    0.2081517  0.2066806
#>  [621,]    0.2067689    0.2194385  0.2178877
#>  [622,]    0.2174740    0.2307996  0.2291685
#>  [623,]    0.2036781    0.2161584  0.2146307
#>  [624,]    0.1875907    0.1990853  0.1976783
#>  [625,]    0.1982552    0.2104032  0.2089163
#>  [626,]    0.2210409    0.2345850  0.2329271
#>  [627,]    0.2538109    0.2693629  0.2674593
#>  [628,]    0.1678119    0.1780946  0.1768360
#>  [629,]    0.2056785    0.2182814  0.2167388
#>  [630,]    0.2600931    0.2760299  0.2740791
#>  [631,]    0.1992811    0.2114919  0.2099973
#>  [632,]    0.2541224    0.2696933  0.2677873
#>  [633,]    0.2546497    0.2702531  0.2683432
#>  [634,]    0.1873304    0.1988090  0.1974040
#>  [635,]    0.1659069    0.1760729  0.1748286
#>  [636,]    0.1944698    0.2063860  0.2049274
#>  [637,]    0.2053535    0.2179366  0.2163964
#>  [638,]    0.1875578    0.1990504  0.1976437
#>  [639,]    0.2316988    0.2458960  0.2441582
#>  [640,]    0.1542598    0.1637121  0.1625551
#>  [641,]    0.2796547    0.2967901  0.2946927
#>  [642,]    0.1611183    0.1709909  0.1697825
#>  [643,]    0.1826321    0.1938229  0.1924531
#>  [644,]    0.2752607    0.2921268  0.2900623
#>  [645,]    0.2498610    0.2651710  0.2632970
#>  [646,]    0.2669424    0.2832989  0.2812968
#>  [647,]    0.1230785    0.1306202  0.1296971
#>  [648,]    0.2209632    0.2345026  0.2328453
#>  [649,]    0.1519643    0.1612761  0.1601363
#>  [650,]    0.2106604    0.2235685  0.2219885
#>  [651,]    0.1596330    0.1694146  0.1682173
#>  [652,]    0.2182590    0.2316327  0.2299957
#>  [653,]    0.1984992    0.2106620  0.2091733
#>  [654,]    0.1818159    0.1929566  0.1915930
#>  [655,]    0.1854202    0.1967819  0.1953912
#>  [656,]    0.2298445    0.2439279  0.2422040
#>  [657,]    0.1930021    0.2048282  0.2033807
#>  [658,]    0.2113974    0.2243506  0.2227651
#>  [659,]    0.1907117    0.2023976  0.2009672
#>  [660,]    0.2039317    0.2164274  0.2148979
#>  [661,]    0.1740312    0.1846951  0.1833898
#>  [662,]    0.2542982    0.2698799  0.2679727
#>  [663,]    0.1653400    0.1754712  0.1742311
#>  [664,]    0.2006413    0.2129355  0.2114306
#>  [665,]    0.1543923    0.1638527  0.1626947
#>  [666,]    0.2093714    0.2222005  0.2206301
#>  [667,]    0.2314206    0.2456008  0.2438651
#>  [668,]    0.1593539    0.1691184  0.1679232
#>  [669,]    0.2141620    0.2272846  0.2256784
#>  [670,]    0.2572910    0.2730559  0.2711263
#>  [671,]    0.2145066    0.2276504  0.2260415
#>  [672,]    0.3014990    0.3199726  0.3177114
#>  [673,]    0.2136896    0.2267834  0.2251806
#>  [674,]    0.2694278    0.2859365  0.2839157
#>  [675,]    0.2294100    0.2434669  0.2417463
#>  [676,]    0.1979156    0.2100427  0.2085583
#>  [677,]    0.2120698    0.2250641  0.2234736
#>  [678,]    0.2306835    0.2448182  0.2430881
#>  [679,]    0.1815750    0.1927009  0.1913391
#>  [680,]    0.1884988    0.2000489  0.1986351
#>  [681,]    0.2189604    0.2323771  0.2307348
#>  [682,]    0.1805695    0.1916339  0.1902795
#>  [683,]    0.1987681    0.2109475  0.2094567
#>  [684,]    0.1676460    0.1779185  0.1766611
#>  [685,]    0.2055490    0.2181439  0.2166022
#>  [686,]    0.2270990    0.2410143  0.2393110
#>  [687,]    0.2034370    0.2159025  0.2143766
#>  [688,]    0.1797192    0.1907314  0.1893835
#>  [689,]    0.2342147    0.2485660  0.2468093
#>  [690,]    0.1739619    0.1846214  0.1833167
#>  [691,]    0.1491181    0.1582555  0.1571370
#>  [692,]    0.2229633    0.2366253  0.2349530
#>  [693,]    0.2369851    0.2515062  0.2497288
#>  [694,]    0.2539624    0.2695235  0.2676188
#>  [695,]    0.1938027    0.2056779  0.2042244
#>  [696,]    0.1761071    0.1868981  0.1855772
#>  [697,]    0.1595059    0.1692797  0.1680833
#>  [698,]    0.2925619    0.3104881  0.3082939
#>  [699,]    0.1785659    0.1895075  0.1881682
#>  [700,]    0.1854663    0.1968307  0.1954396
#>  [701,]    0.1727500    0.1833353  0.1820396
#>  [702,]    0.2304064    0.2445244  0.2427963
#>  [703,]    0.1636109    0.1736362  0.1724091
#>  [704,]    0.1775334    0.1884118  0.1870802
#>  [705,]    0.1610697    0.1709393  0.1697312
#>  [706,]    0.1272110    0.1350059  0.1340518
#>  [707,]    0.1760760    0.1868649  0.1855443
#>  [708,]    0.1838211    0.1950847  0.1937060
#>  [709,]    0.1639145    0.1739584  0.1727290
#>  [710,]    0.2912536    0.3090995  0.3069151
#>  [711,]    0.1986020    0.2107714  0.2092818
#>  [712,]    0.1785676    0.1895093  0.1881700
#>  [713,]    0.1897449    0.2013715  0.1999484
#>  [714,]    0.2852711    0.3027507  0.3006111
#>  [715,]    0.2303005    0.2444118  0.2426846
#>  [716,]    0.2846435    0.3020844  0.2999496
#>  [717,]    0.1894027    0.2010083  0.1995877
#>  [718,]    0.1787311    0.1896829  0.1883423
#>  [719,]    0.1962739    0.2083006  0.2068285
#>  [720,]    0.2492013    0.2644707  0.2626016
#>  [721,]    0.1709169    0.1813900  0.1801081
#>  [722,]    0.1936845    0.2055525  0.2040998
#>  [723,]    0.2138355    0.2269382  0.2253344
#>  [724,]    0.2278029    0.2417611  0.2400526
#>  [725,]    0.1826525    0.1938445  0.1924746
#>  [726,]    0.2270869    0.2410014  0.2392982
#>  [727,]    0.2533831    0.2689087  0.2670083
#>  [728,]    0.1832709    0.1945007  0.1931261
#>  [729,]    0.1958815    0.2078840  0.2064149
#>  [730,]    0.3360690    0.3566608  0.3541403
#>  [731,]    0.2380795    0.2526675  0.2508819
#>  [732,]    0.2554019    0.2710512  0.2691357
#>  [733,]    0.2061061    0.2187352  0.2171894
#>  [734,]    0.1711672    0.1816554  0.1803716
#>  [735,]    0.1747505    0.1854583  0.1841477
#>  [736,]    0.1858196    0.1972056  0.1958119
#>  [737,]    0.1407254    0.1493484  0.1482929
#>  [738,]    0.1720439    0.1825858  0.1812955
#>  [739,]    0.1787364    0.1896885  0.1883479
#>  [740,]    0.1929586    0.2047819  0.2033347
#>  [741,]    0.1904733    0.2021445  0.2007159
#>  [742,]    0.2138803    0.2269856  0.2253815
#>  [743,]    0.1592128    0.1689686  0.1677745
#>  [744,]    0.2369888    0.2515101  0.2497327
#>  [745,]    0.2067223    0.2193890  0.2178386
#>  [746,]    0.1901765    0.2018294  0.2004030
#>  [747,]    0.2755896    0.2924759  0.2904090
#>  [748,]    0.2863857    0.3039332  0.3017853
#>  [749,]    0.1710393    0.1815197  0.1802369
#>  [750,]    0.2405111    0.2552481  0.2534443
#>  [751,]    0.1776314    0.1885157  0.1871834
#>  [752,]    0.2232151    0.2368925  0.2352183
#>  [753,]    0.1488818    0.1580046  0.1568879
#>  [754,]    0.2134923    0.2265739  0.2249727
#>  [755,]    0.1423291    0.1510503  0.1499828
#>  [756,]    0.1572996    0.1669382  0.1657584
#>  [757,]    0.2124276    0.2254440  0.2238508
#>  [758,]    0.1687041    0.1790415  0.1777762
#>  [759,]    0.1788213    0.1897785  0.1884373
#>  [760,]    0.2580695    0.2738823  0.2719467
#>  [761,]    0.1608950    0.1707538  0.1695471
#>  [762,]    0.1754591    0.1862104  0.1848944
#>  [763,]    0.2329144    0.2471860  0.2454391
#>  [764,]    0.2243692    0.2381172  0.2364344
#>  [765,]    0.1821470    0.1933081  0.1919419
#>  [766,]    0.2238959    0.2376149  0.2359356
#>  [767,]    0.2227603    0.2364098  0.2347391
#>  [768,]    0.1609927    0.1708575  0.1696500
#>  [769,]    0.1764838    0.1872979  0.1859742
#>  [770,]    0.2297867    0.2438666  0.2421432
#>  [771,]    0.1946371    0.2065634  0.2051036
#>  [772,]    0.3046037    0.3232675  0.3209830
#>  [773,]    0.1735528    0.1841873  0.1828856
#>  [774,]    0.2018018    0.2141671  0.2126535
#>  [775,]    0.1786589    0.1896062  0.1882662
#>  [776,]    0.2079826    0.2207264  0.2191665
#>  [777,]    0.2094759    0.2223115  0.2207403
#>  [778,]    0.2140099    0.2271232  0.2255180
#>  [779,]    0.1571971    0.1668294  0.1656504
#>  [780,]    0.2367365    0.2512421  0.2494666
#>  [781,]    0.1859706    0.1973660  0.1959711
#>  [782,]    0.1953269    0.2072954  0.2058304
#>  [783,]    0.2235963    0.2372969  0.2356199
#>  [784,]    0.2995975    0.3179548  0.3157078
#>  [785,]    0.1713148    0.1818120  0.1805271
#>  [786,]    0.1965867    0.2086325  0.2071581
#>  [787,]    0.3102445    0.3292538  0.3269270
#>  [788,]    0.2107467    0.2236600  0.2220794
#>  [789,]    0.2087066    0.2214950  0.2199297
#>  [790,]    0.2645807    0.2807924  0.2788081
#>  [791,]    0.1636617    0.1736901  0.1724626
#>  [792,]    0.1792340    0.1902165  0.1888722
#>  [793,]    0.1473451    0.1563738  0.1552687
#>  [794,]    0.2588758    0.2747380  0.2727964
#>  [795,]    0.2096388    0.2224844  0.2209120
#>  [796,]    0.2842690    0.3016869  0.2995548
#>  [797,]    0.1401473    0.1487350  0.1476838
#>  [798,]    0.1765676    0.1873868  0.1860625
#>  [799,]    0.1881567    0.1996859  0.1982747
#>  [800,]    0.2179065    0.2312585  0.2296242
#>  [801,]    0.2146243    0.2277753  0.2261656
#>  [802,]    0.2068095    0.2194817  0.2179305
#>  [803,]    0.2394526    0.2541247  0.2523288
#>  [804,]    0.1806864    0.1917579  0.1904027
#>  [805,]    0.1848433    0.1961695  0.1947831
#>  [806,]    0.1681952    0.1785014  0.1772399
#>  [807,]    0.1950282    0.2069785  0.2055157
#>  [808,]    0.1998386    0.2120836  0.2105847
#>  [809,]    0.1582559    0.1679530  0.1667661
#>  [810,]    0.2142025    0.2273273  0.2257208
#>  [811,]    0.2285690    0.2425743  0.2408600
#>  [812,]    0.1879995    0.1995191  0.1981091
#>  [813,]    0.1808897    0.1919737  0.1906170
#>  [814,]    0.1930690    0.2048992  0.2034512
#>  [815,]    0.2444043    0.2593800  0.2575469
#>  [816,]    0.1340404    0.1422538  0.1412485
#>  [817,]    0.2084010    0.2211707  0.2196077
#>  [818,]    0.1648575    0.1749591  0.1737226
#>  [819,]    0.2412809    0.2560652  0.2542556
#>  [820,]    0.1844608    0.1957636  0.1943801
#>  [821,]    0.1949862    0.2069338  0.2054714
#>  [822,]    0.2093012    0.2221261  0.2205563
#>  [823,]    0.1657024    0.1758557  0.1746129
#>  [824,]    0.2362717    0.2507491  0.2489770
#>  [825,]    0.2566581    0.2723845  0.2704595
#>  [826,]    0.3110578    0.3301172  0.3277843
#>  [827,]    0.1778631    0.1887616  0.1874276
#>  [828,]    0.2132901    0.2263593  0.2247596
#>  [829,]    0.1794288    0.1904233  0.1890775
#>  [830,]    0.2096261    0.2224707  0.2208985
#>  [831,]    0.2258759    0.2397162  0.2380221
#>  [832,]    0.2345402    0.2489114  0.2471523
#>  [833,]    0.1901380    0.2017886  0.2003625
#>  [834,]    0.1735578    0.1841925  0.1828908
#>  [835,]    0.1388516    0.1473597  0.1463183
#>  [836,]    0.1484029    0.1574964  0.1563833
#>  [837,]    0.1482989    0.1573860  0.1562737
#>  [838,]    0.1749815    0.1857034  0.1843910
#>  [839,]    0.2126910    0.2257235  0.2241283
#>  [840,]    0.2033025    0.2157597  0.2142349
#>  [841,]    0.1922503    0.2040303  0.2025884
#>  [842,]    0.2354111    0.2498356  0.2480700
#>  [843,]    0.2072613    0.2199612  0.2184067
#>  [844,]    0.2623627    0.2784384  0.2764707
#>  [845,]    0.2856203    0.3031212  0.3009791
#>  [846,]    0.1565406    0.1661326  0.1649585
#>  [847,]    0.2026208    0.2150363  0.2135166
#>  [848,]    0.1985235    0.2106880  0.2091990
#>  [849,]    0.2461889    0.2612739  0.2594274
#>  [850,]    0.2370016    0.2515236  0.2497461
#>  [851,]    0.1990446    0.2112411  0.2097482
#>  [852,]    0.1987300    0.2109071  0.2094166
#>  [853,]    0.1742944    0.1849742  0.1836669
#>  [854,]    0.2077104    0.2204379  0.2188800
#>  [855,]    0.1842973    0.1955901  0.1942078
#>  [856,]    0.1991365    0.2113383  0.2098448
#>  [857,]    0.2774934    0.2944961  0.2924149
#>  [858,]    0.2132287    0.2262941  0.2246948
#>  [859,]    0.3096757    0.3286505  0.3263279
#>  [860,]    0.1274564    0.1352665  0.1343105
#>  [861,]    0.1962214    0.2082448  0.2067731
#>  [862,]    0.1865401    0.1979702  0.1965711
#>  [863,]    0.1494408    0.1585980  0.1574771
#>  [864,]    0.2448200    0.2598211  0.2579849
#>  [865,]    0.2741661    0.2909652  0.2889089
#>  [866,]    0.1532446    0.1626347  0.1614853
#>  [867,]    0.2608877    0.2768731  0.2749164
#>  [868,]    0.1780646    0.1889754  0.1876399
#>  [869,]    0.1778945    0.1887951  0.1874608
#>  [870,]    0.1793155    0.1903031  0.1889582
#>  [871,]    0.1593698    0.1691352  0.1679399
#>  [872,]    0.2026137    0.2150288  0.2135092
#>  [873,]    0.2108403    0.2237593  0.2221780
#>  [874,]    0.2169032    0.2301939  0.2285670
#>  [875,]    0.1411703    0.1498206  0.1487618
#>  [876,]    0.2913674    0.3092204  0.3070351
#>  [877,]    0.1982988    0.2104495  0.2089623
#>  [878,]    0.2215481    0.2351233  0.2334616
#>  [879,]    0.1943906    0.2063019  0.2048439
#>  [880,]    0.2449716    0.2599818  0.2581445
#>  [881,]    0.2340855    0.2484285  0.2466728
#>  [882,]    0.2182639    0.2316379  0.2300009
#>  [883,]    0.2227066    0.2363527  0.2346824
#>  [884,]    0.2203162    0.2338159  0.2321635
#>  [885,]    0.1772534    0.1881146  0.1867851
#>  [886,]    0.2121110    0.2251081  0.2235172
#>  [887,]    0.2115253    0.2244865  0.2229000
#>  [888,]    0.1895352    0.2011490  0.1997274
#>  [889,]    0.1883140    0.1998529  0.1984405
#>  [890,]    0.1999091    0.2121584  0.2106590
#>  [891,]    0.2265094    0.2403887  0.2386898
#>  [892,]    0.1727864    0.1833739  0.1820779
#>  [893,]    0.2062140    0.2188497  0.2173031
#>  [894,]    0.1650344    0.1751470  0.1739092
#>  [895,]    0.2533308    0.2688532  0.2669532
#>  [896,]    0.2707904    0.2873827  0.2853517
#>  [897,]    0.2329767    0.2472522  0.2455049
#>  [898,]    0.1411692    0.1498195  0.1487606
#>  [899,]    0.1680119    0.1783070  0.1770468
#>  [900,]    0.1601425    0.1699553  0.1687541
#>  [901,]    0.1745829    0.1852806  0.1839711
#>  [902,]    0.1674960    0.1777593  0.1765030
#>  [903,]    0.2170915    0.2303936  0.2287653
#>  [904,]    0.2230401    0.2367067  0.2350338
#>  [905,]    0.2389271    0.2535672  0.2517752
#>  [906,]    0.1694517    0.1798349  0.1785639
#>  [907,]    0.2562472    0.2719483  0.2700264
#>  [908,]    0.2514569    0.2668646  0.2649787
#>  [909,]    0.1692558    0.1796269  0.1783574
#>  [910,]    0.1694868    0.1798720  0.1786008
#>  [911,]    0.2021451    0.2145315  0.2130154
#>  [912,]    0.1885229    0.2000745  0.1986606
#>  [913,]    0.3225467    0.3423101  0.3398910
#>  [914,]    0.1577197    0.1673841  0.1662012
#>  [915,]    0.1287830    0.1366742  0.1357083
#>  [916,]    0.2238340    0.2375493  0.2358705
#>  [917,]    0.1738042    0.1844539  0.1831503
#>  [918,]    0.1906438    0.2023254  0.2008955
#>  [919,]    0.2147709    0.2279306  0.2263198
#>  [920,]    0.1708113    0.1812778  0.1799967
#>  [921,]    0.2148032    0.2279651  0.2263540
#>  [922,]    0.2184282    0.2318119  0.2301737
#>  [923,]    0.1719549    0.1824915  0.1812018
#>  [924,]    0.3090244    0.3279592  0.3256415
#>  [925,]    0.2327229    0.2469827  0.2452373
#>  [926,]    0.2515021    0.2669128  0.2650265
#>  [927,]    0.2064832    0.2191353  0.2175866
#>  [928,]    0.1849967    0.1963323  0.1949448
#>  [929,]    0.2368580    0.2513713  0.2495948
#>  [930,]    0.2521529    0.2676031  0.2657120
#>  [931,]    0.2094205    0.2222528  0.2206820
#>  [932,]    0.2145196    0.2276642  0.2260552
#>  [933,]    0.1994525    0.2116739  0.2101779
#>  [934,]    0.1709218    0.1813952  0.1801132
#>  [935,]    0.2091706    0.2219874  0.2204185
#>  [936,]    0.3346714    0.3551772  0.3526672
#>  [937,]    0.1952923    0.2072587  0.2057940
#>  [938,]    0.1863563    0.1977751  0.1963774
#>  [939,]    0.1800563    0.1910892  0.1897387
#>  [940,]    0.2055214    0.2181146  0.2165732
#>  [941,]    0.2145956    0.2277447  0.2261352
#>  [942,]    0.1917932    0.2035452  0.2021067
#>  [943,]    0.1991730    0.2113773  0.2098835
#>  [944,]    0.1907849    0.2024751  0.2010442
#>  [945,]    0.1899869    0.2016283  0.2002034
#>  [946,]    0.2120012    0.2249914  0.2234014
#>  [947,]    0.2168934    0.2301834  0.2285566
#>  [948,]    0.2021232    0.2145082  0.2129922
#>  [949,]    0.2383219    0.2529248  0.2511373
#>  [950,]    0.1779964    0.1889032  0.1875682
#>  [951,]    0.1621460    0.1720815  0.1708654
#>  [952,]    0.1697629    0.1801650  0.1788918
#>  [953,]    0.1776175    0.1885009  0.1871687
#>  [954,]    0.1982439    0.2103912  0.2089043
#>  [955,]    0.2924681    0.3103885  0.3081950
#>  [956,]    0.2136524    0.2267439  0.2251415
#>  [957,]    0.1951445    0.2071020  0.2056383
#>  [958,]    0.2001666    0.2124316  0.2109303
#>  [959,]    0.1491658    0.1583059  0.1571871
#>  [960,]    0.2095523    0.2223924  0.2208207
#>  [961,]    0.2472354    0.2623845  0.2605302
#>  [962,]    0.2040185    0.2165195  0.2149893
#>  [963,]    0.1843998    0.1956988  0.1943158
#>  [964,]    0.2517941    0.2672222  0.2653337
#>  [965,]    0.1898216    0.2014529  0.2000292
#>  [966,]    0.2656663    0.2819445  0.2799520
#>  [967,]    0.2108713    0.2237924  0.2222108
#>  [968,]    0.3110623    0.3301214  0.3277885
#>  [969,]    0.2246602    0.2384260  0.2367410
#>  [970,]    0.2591415    0.2750200  0.2730765
#>  [971,]    0.1647218    0.1748150  0.1735796
#>  [972,]    0.2540393    0.2696054  0.2677001
#>  [973,]    0.1754993    0.1862530  0.1849367
#>  [974,]    0.1829008    0.1941080  0.1927362
#>  [975,]    0.2782269    0.2952749  0.2931881
#>  [976,]    0.1821493    0.1933103  0.1919442
#>  [977,]    0.2100565    0.2229276  0.2213521
#>  [978,]    0.2260362    0.2398864  0.2381911
#>  [979,]    0.1973161    0.2094066  0.2079267
#>  [980,]    0.1704597    0.1809047  0.1796262
#>  [981,]    0.1739237    0.1845809  0.1832764
#>  [982,]    0.2381009    0.2526902  0.2509044
#>  [983,]    0.2927310    0.3106675  0.3084720
#>  [984,]    0.2117239    0.2246970  0.2231090
#>  [985,]    0.2616206    0.2776509  0.2756888
#>  [986,]    0.1707878    0.1812529  0.1799719
#>  [987,]    0.3049376    0.3236217  0.3213347
#>  [988,]    0.1724596    0.1830270  0.1817335
#>  [989,]    0.1545652    0.1640362  0.1628769
#>  [990,]    0.1967906    0.2088489  0.2073729
#>  [991,]    0.1968165    0.2088764  0.2074003
#>  [992,]    0.1711209    0.1816063  0.1803229
#>  [993,]    0.2808757    0.2980859  0.2959793
#>  [994,]    0.3173271    0.3367705  0.3343905
#>  [995,]    0.2134039    0.2264800  0.2248794
#>  [996,]    0.2102147    0.2230955  0.2215189
#>  [997,]    0.1793472    0.1903366  0.1889915
#>  [998,]    0.2044449    0.2169720  0.2154387
#>  [999,]    0.2426808    0.2575507  0.2557306
#> [1000,]    0.1871615    0.1986298  0.1972260