Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 88.19898
#> 2   Feb 94.60354
#> 3   Mar 86.06503
#> 4   Apr 89.14516
#> 5   May 92.57117
#> 6   Jun 89.01784

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 9.986130  2
#> 2   Feb 6.146603  2
#> 3   Mar 6.721933  2
#> 4   Apr 6.194245  2
#> 5   May 7.279435  2
#> 6   Jun 9.694922  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [10ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [20ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [59ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [613ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [16ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1902572    0.1970124  0.1970335
#>    [2,]    0.1967566    0.2037426  0.2037644
#>    [3,]    0.2191123    0.2268920  0.2269163
#>    [4,]    0.1842701    0.1908127  0.1908332
#>    [5,]    0.2012771    0.2084235  0.2084458
#>    [6,]    0.2640069    0.2733804  0.2734097
#>    [7,]    0.2185380    0.2262973  0.2263215
#>    [8,]    0.1698532    0.1758840  0.1759028
#>    [9,]    0.2392552    0.2477499  0.2477764
#>   [10,]    0.3700026    0.3831389  0.3831799
#>   [11,]    0.2941960    0.3046412  0.3046738
#>   [12,]    0.2301599    0.2383317  0.2383572
#>   [13,]    0.1893442    0.1960671  0.1960881
#>   [14,]    0.1764452    0.1827101  0.1827297
#>   [15,]    0.2586398    0.2678227  0.2678514
#>   [16,]    0.2112959    0.2187980  0.2188215
#>   [17,]    0.2391866    0.2476789  0.2477054
#>   [18,]    0.2273205    0.2353916  0.2354167
#>   [19,]    0.1421650    0.1472127  0.1472285
#>   [20,]    0.2435133    0.2521593  0.2521863
#>   [21,]    0.1738937    0.1800680  0.1800873
#>   [22,]    0.2394951    0.2479983  0.2480249
#>   [23,]    0.2050256    0.2123052  0.2123279
#>   [24,]    0.1984656    0.2055123  0.2055342
#>   [25,]    0.2195625    0.2273581  0.2273824
#>   [26,]    0.1823548    0.1888295  0.1888497
#>   [27,]    0.2647466    0.2741464  0.2741758
#>   [28,]    0.2318041    0.2400344  0.2400601
#>   [29,]    0.1968790    0.2038693  0.2038911
#>   [30,]    0.1660765    0.1719730  0.1719914
#>   [31,]    0.2487139    0.2575445  0.2575721
#>   [32,]    0.2354914    0.2438526  0.2438787
#>   [33,]    0.2392924    0.2477884  0.2478149
#>   [34,]    0.3020834    0.3128086  0.3128421
#>   [35,]    0.1992106    0.2062836  0.2063057
#>   [36,]    0.2122835    0.2198206  0.2198442
#>   [37,]    0.2153135    0.2229583  0.2229822
#>   [38,]    0.1454161    0.1505793  0.1505954
#>   [39,]    0.2241869    0.2321467  0.2321715
#>   [40,]    0.3135322    0.3246638  0.3246986
#>   [41,]    0.2182058    0.2259533  0.2259775
#>   [42,]    0.1940599    0.2009502  0.2009717
#>   [43,]    0.3416695    0.3538002  0.3538381
#>   [44,]    0.1814821    0.1879257  0.1879458
#>   [45,]    0.2447685    0.2534590  0.2534862
#>   [46,]    0.1434030    0.1484947  0.1485106
#>   [47,]    0.2836379    0.2937083  0.2937398
#>   [48,]    0.1705547    0.1766104  0.1766293
#>   [49,]    0.2102442    0.2177091  0.2177324
#>   [50,]    0.1883757    0.1950640  0.1950849
#>   [51,]    0.1965025    0.2034795  0.2035012
#>   [52,]    0.2140719    0.2216726  0.2216963
#>   [53,]    0.2172052    0.2249171  0.2249412
#>   [54,]    0.2533878    0.2623843  0.2624124
#>   [55,]    0.2578254    0.2669796  0.2670081
#>   [56,]    0.2833822    0.2934436  0.2934750
#>   [57,]    0.2521812    0.2611349  0.2611629
#>   [58,]    0.1779669    0.1842857  0.1843054
#>   [59,]    0.2315291    0.2397496  0.2397752
#>   [60,]    0.1666594    0.1725768  0.1725953
#>   [61,]    0.2516374    0.2605717  0.2605996
#>   [62,]    0.1960308    0.2029910  0.2030127
#>   [63,]    0.1637270    0.1695403  0.1695585
#>   [64,]    0.2417639    0.2503478  0.2503745
#>   [65,]    0.1891911    0.1959084  0.1959294
#>   [66,]    0.2336222    0.2419170  0.2419429
#>   [67,]    0.2256383    0.2336497  0.2336747
#>   [68,]    0.2263001    0.2343349  0.2343600
#>   [69,]    0.1900902    0.1968395  0.1968605
#>   [70,]    0.2460987    0.2548364  0.2548637
#>   [71,]    0.1803760    0.1867803  0.1868003
#>   [72,]    0.1808620    0.1872836  0.1873036
#>   [73,]    0.2249836    0.2329718  0.2329967
#>   [74,]    0.2248327    0.2328155  0.2328404
#>   [75,]    0.2651207    0.2745338  0.2745632
#>   [76,]    0.1884406    0.1951314  0.1951522
#>   [77,]    0.2430100    0.2516381  0.2516650
#>   [78,]    0.1807158    0.1871323  0.1871523
#>   [79,]    0.2077827    0.2151601  0.2151831
#>   [80,]    0.1917226    0.1985298  0.1985511
#>   [81,]    0.1835340    0.1900506  0.1900709
#>   [82,]    0.2751286    0.2848969  0.2849274
#>   [83,]    0.3032580    0.3140251  0.3140587
#>   [84,]    0.1804732    0.1868810  0.1869010
#>   [85,]    0.1819495    0.1884098  0.1884299
#>   [86,]    0.1719824    0.1780888  0.1781079
#>   [87,]    0.2347246    0.2430586  0.2430846
#>   [88,]    0.1859252    0.1925266  0.1925472
#>   [89,]    0.2890602    0.2993233  0.2993553
#>   [90,]    0.1368866    0.1417469  0.1417621
#>   [91,]    0.3003861    0.3110512  0.3110845
#>   [92,]    0.1499489    0.1552729  0.1552896
#>   [93,]    0.2198665    0.2276729  0.2276973
#>   [94,]    0.2361218    0.2445052  0.2445314
#>   [95,]    0.2387408    0.2472174  0.2472438
#>   [96,]    0.2170227    0.2247281  0.2247522
#>   [97,]    0.1914764    0.1982749  0.1982962
#>   [98,]    0.2371357    0.2455552  0.2455815
#>   [99,]    0.1975183    0.2045313  0.2045532
#>  [100,]    0.2109558    0.2184458  0.2184692
#>  [101,]    0.2236247    0.2315645  0.2315893
#>  [102,]    0.1836960    0.1902183  0.1902386
#>  [103,]    0.2048380    0.2121107  0.2121334
#>  [104,]    0.1238698    0.1282679  0.1282817
#>  [105,]    0.2530620    0.2620471  0.2620751
#>  [106,]    0.2128640    0.2204217  0.2204453
#>  [107,]    0.1810659    0.1874948  0.1875148
#>  [108,]    0.1106948    0.1146252  0.1146374
#>  [109,]    0.2502975    0.2591844  0.2592121
#>  [110,]    0.2150763    0.2227127  0.2227365
#>  [111,]    0.2821228    0.2921395  0.2921707
#>  [112,]    0.3400002    0.3520715  0.3521091
#>  [113,]    0.2049362    0.2122126  0.2122353
#>  [114,]    0.2045790    0.2118427  0.2118653
#>  [115,]    0.1922139    0.1990386  0.1990599
#>  [116,]    0.1377507    0.1426417  0.1426570
#>  [117,]    0.1855368    0.1921244  0.1921450
#>  [118,]    0.1637790    0.1695942  0.1696124
#>  [119,]    0.1916615    0.1984665  0.1984878
#>  [120,]    0.1750887    0.1813054  0.1813248
#>  [121,]    0.2092877    0.2167186  0.2167418
#>  [122,]    0.1708159    0.1768809  0.1768998
#>  [123,]    0.1525433    0.1579595  0.1579764
#>  [124,]    0.2511546    0.2600716  0.2600995
#>  [125,]    0.3006888    0.3113646  0.3113979
#>  [126,]    0.2519105    0.2608546  0.2608825
#>  [127,]    0.2125427    0.2200891  0.2201127
#>  [128,]    0.2645563    0.2739494  0.2739787
#>  [129,]    0.4887976    0.5061515  0.5062056
#>  [130,]    0.2167573    0.2244534  0.2244774
#>  [131,]    0.2096708    0.2171152  0.2171385
#>  [132,]    0.2081706    0.2155617  0.2155848
#>  [133,]    0.1749365    0.1811477  0.1811671
#>  [134,]    0.1489714    0.1542608  0.1542773
#>  [135,]    0.2268430    0.2348972  0.2349223
#>  [136,]    0.2117123    0.2192292  0.2192527
#>  [137,]    0.2560437    0.2651346  0.2651630
#>  [138,]    0.2110792    0.2185737  0.2185971
#>  [139,]    0.1667419    0.1726622  0.1726807
#>  [140,]    0.1534484    0.1588967  0.1589137
#>  [141,]    0.2105999    0.2180774  0.2181007
#>  [142,]    0.1473952    0.1526287  0.1526450
#>  [143,]    0.1445618    0.1496947  0.1497107
#>  [144,]    0.1563286    0.1618792  0.1618965
#>  [145,]    0.2889000    0.2991573  0.2991893
#>  [146,]    0.2422531    0.2508542  0.2508811
#>  [147,]    0.2365441    0.2449425  0.2449687
#>  [148,]    0.1759570    0.1822045  0.1822240
#>  [149,]    0.1928482    0.1996954  0.1997168
#>  [150,]    0.2790945    0.2890037  0.2890347
#>  [151,]    0.2197076    0.2275083  0.2275327
#>  [152,]    0.1721177    0.1782289  0.1782480
#>  [153,]    0.2016035    0.2087616  0.2087839
#>  [154,]    0.2755178    0.2853001  0.2853306
#>  [155,]    0.2031474    0.2103604  0.2103829
#>  [156,]    0.3274554    0.3390814  0.3391176
#>  [157,]    0.1368795    0.1417395  0.1417547
#>  [158,]    0.2050807    0.2123622  0.2123849
#>  [159,]    0.1919076    0.1987214  0.1987427
#>  [160,]    0.2380280    0.2464793  0.2465056
#>  [161,]    0.1516185    0.1570019  0.1570187
#>  [162,]    0.2473301    0.2561116  0.2561390
#>  [163,]    0.3088466    0.3198120  0.3198462
#>  [164,]    0.1995387    0.2066235  0.2066456
#>  [165,]    0.1795407    0.1859155  0.1859354
#>  [166,]    0.3034224    0.3141953  0.3142289
#>  [167,]    0.2323465    0.2405961  0.2406218
#>  [168,]    0.2118565    0.2193786  0.2194020
#>  [169,]    0.2350418    0.2433870  0.2434130
#>  [170,]    0.1927779    0.1996225  0.1996439
#>  [171,]    0.2753628    0.2851397  0.2851702
#>  [172,]    0.1793203    0.1856873  0.1857072
#>  [173,]    0.2243376    0.2323028  0.2323276
#>  [174,]    0.2274524    0.2355281  0.2355533
#>  [175,]    0.2354759    0.2438365  0.2438626
#>  [176,]    0.2619752    0.2712766  0.2713056
#>  [177,]    0.2251155    0.2331083  0.2331332
#>  [178,]    0.2898306    0.3001209  0.3001530
#>  [179,]    0.1996871    0.2067771  0.2067992
#>  [180,]    0.1907499    0.1975226  0.1975438
#>  [181,]    0.1848852    0.1914497  0.1914701
#>  [182,]    0.1910037    0.1977854  0.1978066
#>  [183,]    0.1928784    0.1997266  0.1997480
#>  [184,]    0.1336256    0.1383702  0.1383850
#>  [185,]    0.2136081    0.2211923  0.2212160
#>  [186,]    0.2558919    0.2649773  0.2650057
#>  [187,]    0.1803070    0.1867089  0.1867288
#>  [188,]    0.2085546    0.2159595  0.2159826
#>  [189,]    0.2337377    0.2420366  0.2420625
#>  [190,]    0.2063824    0.2137101  0.2137329
#>  [191,]    0.1783170    0.1846483  0.1846681
#>  [192,]    0.2016483    0.2088079  0.2088303
#>  [193,]    0.1745019    0.1806978  0.1807171
#>  [194,]    0.1847366    0.1912957  0.1913162
#>  [195,]    0.2019452    0.2091153  0.2091377
#>  [196,]    0.2193331    0.2271206  0.2271449
#>  [197,]    0.1982034    0.2052407  0.2052627
#>  [198,]    0.2716515    0.2812965  0.2813266
#>  [199,]    0.2044542    0.2117135  0.2117362
#>  [200,]    0.2023047    0.2094877  0.2095102
#>  [201,]    0.2033569    0.2105771  0.2105997
#>  [202,]    0.1686892    0.1746786  0.1746973
#>  [203,]    0.2584503    0.2676266  0.2676552
#>  [204,]    0.1560013    0.1615402  0.1615575
#>  [205,]    0.2579541    0.2671129  0.2671414
#>  [206,]    0.2548261    0.2638737  0.2639019
#>  [207,]    0.2538863    0.2629004  0.2629285
#>  [208,]    0.2746210    0.2843715  0.2844019
#>  [209,]    0.1969243    0.2039163  0.2039381
#>  [210,]    0.2104464    0.2179184  0.2179417
#>  [211,]    0.2091647    0.2165913  0.2166144
#>  [212,]    0.2090063    0.2164272  0.2164504
#>  [213,]    0.1663990    0.1723070  0.1723255
#>  [214,]    0.1763038    0.1825636  0.1825832
#>  [215,]    0.2294515    0.2375983  0.2376238
#>  [216,]    0.2241755    0.2321349  0.2321597
#>  [217,]    0.2022725    0.2094542  0.2094766
#>  [218,]    0.2245786    0.2325524  0.2325772
#>  [219,]    0.1950451    0.2019703  0.2019919
#>  [220,]    0.2325200    0.2407757  0.2408015
#>  [221,]    0.1994449    0.2065263  0.2065484
#>  [222,]    0.2106988    0.2181797  0.2182031
#>  [223,]    0.2806963    0.2906625  0.2906936
#>  [224,]    0.2027485    0.2099471  0.2099696
#>  [225,]    0.2048900    0.2121647  0.2121875
#>  [226,]    0.2016581    0.2088181  0.2088404
#>  [227,]    0.2000580    0.2071612  0.2071834
#>  [228,]    0.2444004    0.2530779  0.2531050
#>  [229,]    0.2111772    0.2186752  0.2186986
#>  [230,]    0.1705814    0.1766381  0.1766570
#>  [231,]    0.2617139    0.2710060  0.2710350
#>  [232,]    0.2216052    0.2294732  0.2294977
#>  [233,]    0.1937292    0.2006077  0.2006292
#>  [234,]    0.2641617    0.2735408  0.2735701
#>  [235,]    0.2576866    0.2668359  0.2668645
#>  [236,]    0.2455286    0.2542461  0.2542734
#>  [237,]    0.3330119    0.3448352  0.3448721
#>  [238,]    0.2581107    0.2672749  0.2673035
#>  [239,]    0.2966933    0.3072273  0.3072602
#>  [240,]    0.2887708    0.2990235  0.2990555
#>  [241,]    0.2030817    0.2102923  0.2103148
#>  [242,]    0.1809224    0.1873462  0.1873663
#>  [243,]    0.2640959    0.2734725  0.2735018
#>  [244,]    0.2749518    0.2847138  0.2847443
#>  [245,]    0.1597820    0.1654552  0.1654729
#>  [246,]    0.2895543    0.2998348  0.2998669
#>  [247,]    0.2676903    0.2771945  0.2772242
#>  [248,]    0.1875493    0.1942085  0.1942292
#>  [249,]    0.2011021    0.2082424  0.2082647
#>  [250,]    0.2160323    0.2237027  0.2237266
#>  [251,]    0.2315470    0.2397681  0.2397938
#>  [252,]    0.1649383    0.1707946  0.1708129
#>  [253,]    0.1906315    0.1974000  0.1974211
#>  [254,]    0.2403952    0.2489305  0.2489571
#>  [255,]    0.2609244    0.2701884  0.2702173
#>  [256,]    0.1676703    0.1736236  0.1736422
#>  [257,]    0.2025647    0.2097569  0.2097793
#>  [258,]    0.1886880    0.1953876  0.1954085
#>  [259,]    0.2443187    0.2529933  0.2530204
#>  [260,]    0.2176728    0.2254015  0.2254256
#>  [261,]    0.2246166    0.2325917  0.2326166
#>  [262,]    0.2960966    0.3066093  0.3066421
#>  [263,]    0.2924456    0.3028287  0.3028611
#>  [264,]    0.1628397    0.1686214  0.1686395
#>  [265,]    0.1447288    0.1498676  0.1498836
#>  [266,]    0.2084733    0.2158752  0.2158983
#>  [267,]    0.1612834    0.1670099  0.1670278
#>  [268,]    0.2237355    0.2316793  0.2317041
#>  [269,]    0.2606673    0.2699223  0.2699512
#>  [270,]    0.1710930    0.1771678  0.1771867
#>  [271,]    0.1895439    0.1962738  0.1962948
#>  [272,]    0.2888029    0.2990568  0.2990888
#>  [273,]    0.1685898    0.1745757  0.1745944
#>  [274,]    0.1803845    0.1867892  0.1868092
#>  [275,]    0.2397417    0.2482538  0.2482803
#>  [276,]    0.2693766    0.2789408  0.2789706
#>  [277,]    0.2853136    0.2954435  0.2954751
#>  [278,]    0.1779315    0.1842491  0.1842688
#>  [279,]    0.1860871    0.1926943  0.1927149
#>  [280,]    0.2705367    0.2801421  0.2801720
#>  [281,]    0.1781276    0.1844521  0.1844719
#>  [282,]    0.1818147    0.1882702  0.1882903
#>  [283,]    0.1869309    0.1935680  0.1935888
#>  [284,]    0.2529827    0.2619648  0.2619928
#>  [285,]    0.2768467    0.2866760  0.2867067
#>  [286,]    0.1694385    0.1754545  0.1754733
#>  [287,]    0.2019755    0.2091467  0.2091690
#>  [288,]    0.2209730    0.2288187  0.2288432
#>  [289,]    0.2310875    0.2392923  0.2393179
#>  [290,]    0.2346528    0.2429842  0.2430102
#>  [291,]    0.1924537    0.1992869  0.1993082
#>  [292,]    0.2105724    0.2180489  0.2180723
#>  [293,]    0.2058975    0.2132079  0.2132307
#>  [294,]    0.2092344    0.2166633  0.2166865
#>  [295,]    0.1988220    0.2058813  0.2059033
#>  [296,]    0.1981051    0.2051390  0.2051610
#>  [297,]    0.2576105    0.2667570  0.2667856
#>  [298,]    0.2025007    0.2096906  0.2097130
#>  [299,]    0.2094576    0.2168944  0.2169176
#>  [300,]    0.2360256    0.2444057  0.2444318
#>  [301,]    0.3036584    0.3144396  0.3144732
#>  [302,]    0.1406639    0.1456583  0.1456739
#>  [303,]    0.2666100    0.2760760  0.2761056
#>  [304,]    0.2193480    0.2271360  0.2271603
#>  [305,]    0.1866938    0.1933225  0.1933432
#>  [306,]    0.1944752    0.2013802  0.2014017
#>  [307,]    0.2786673    0.2885613  0.2885921
#>  [308,]    0.2524830    0.2614473  0.2614753
#>  [309,]    0.2305284    0.2387133  0.2387389
#>  [310,]    0.1886514    0.1953496  0.1953705
#>  [311,]    0.2867910    0.2969733  0.2970051
#>  [312,]    0.1871933    0.1938397  0.1938605
#>  [313,]    0.2416511    0.2502310  0.2502577
#>  [314,]    0.1982783    0.2053182  0.2053401
#>  [315,]    0.1589225    0.1645652  0.1645828
#>  [316,]    0.1781565    0.1844821  0.1845018
#>  [317,]    0.1983446    0.2053870  0.2054090
#>  [318,]    0.1620380    0.1677914  0.1678093
#>  [319,]    0.1172783    0.1214424  0.1214554
#>  [320,]    0.2031944    0.2104089  0.2104314
#>  [321,]    0.1950485    0.2019737  0.2019953
#>  [322,]    0.2562558    0.2653541  0.2653825
#>  [323,]    0.2240792    0.2320352  0.2320600
#>  [324,]    0.2022696    0.2094512  0.2094736
#>  [325,]    0.2125418    0.2200881  0.2201116
#>  [326,]    0.1856380    0.1922293  0.1922499
#>  [327,]    0.1596757    0.1653451  0.1653628
#>  [328,]    0.2241984    0.2321586  0.2321835
#>  [329,]    0.2199259    0.2277345  0.2277589
#>  [330,]    0.1693732    0.1753869  0.1754057
#>  [331,]    0.1628359    0.1686175  0.1686356
#>  [332,]    0.2139743    0.2215715  0.2215952
#>  [333,]    0.2908005    0.3011254  0.3011576
#>  [334,]    0.2244603    0.2324299  0.2324547
#>  [335,]    0.1977653    0.2047871  0.2048090
#>  [336,]    0.2445926    0.2532769  0.2533040
#>  [337,]    0.1466297    0.1518359  0.1518521
#>  [338,]    0.1831259    0.1896280  0.1896483
#>  [339,]    0.1885730    0.1952684  0.1952893
#>  [340,]    0.2088399    0.2162550  0.2162781
#>  [341,]    0.1586670    0.1643007  0.1643182
#>  [342,]    0.2160232    0.2236932  0.2237171
#>  [343,]    0.2426400    0.2512549  0.2512818
#>  [344,]    0.2482923    0.2571079  0.2571354
#>  [345,]    0.2267143    0.2347640  0.2347891
#>  [346,]    0.2249257    0.2329118  0.2329368
#>  [347,]    0.1928686    0.1997166  0.1997379
#>  [348,]    0.1974392    0.2044494  0.2044713
#>  [349,]    0.2388934    0.2473754  0.2474018
#>  [350,]    0.2306685    0.2388584  0.2388840
#>  [351,]    0.1663957    0.1723037  0.1723222
#>  [352,]    0.1893359    0.1960583  0.1960793
#>  [353,]    0.2281859    0.2362877  0.2363130
#>  [354,]    0.2789119    0.2888146  0.2888455
#>  [355,]    0.2400434    0.2485662  0.2485928
#>  [356,]    0.1672055    0.1731423  0.1731609
#>  [357,]    0.2087408    0.2161522  0.2161753
#>  [358,]    0.1950428    0.2019680  0.2019896
#>  [359,]    0.1670386    0.1729694  0.1729879
#>  [360,]    0.1371524    0.1420222  0.1420374
#>  [361,]    0.2276082    0.2356894  0.2357147
#>  [362,]    0.2369891    0.2454035  0.2454297
#>  [363,]    0.2383534    0.2468161  0.2468426
#>  [364,]    0.2071125    0.2144662  0.2144891
#>  [365,]    0.2064255    0.2137548  0.2137777
#>  [366,]    0.2512684    0.2601897  0.2602176
#>  [367,]    0.1409675    0.1459727  0.1459883
#>  [368,]    0.2133761    0.2209521  0.2209757
#>  [369,]    0.1536446    0.1590999  0.1591169
#>  [370,]    0.2085039    0.2159070  0.2159301
#>  [371,]    0.2726253    0.2823046  0.2823349
#>  [372,]    0.1287115    0.1332815  0.1332958
#>  [373,]    0.2535660    0.2625688  0.2625969
#>  [374,]    0.1914715    0.1982698  0.1982910
#>  [375,]    0.2017262    0.2088887  0.2089110
#>  [376,]    0.2373808    0.2458089  0.2458353
#>  [377,]    0.2488441    0.2576793  0.2577068
#>  [378,]    0.1509268    0.1562857  0.1563024
#>  [379,]    0.2161639    0.2238389  0.2238628
#>  [380,]    0.2190062    0.2267821  0.2268064
#>  [381,]    0.2623983    0.2717148  0.2717438
#>  [382,]    0.1436891    0.1487909  0.1488069
#>  [383,]    0.2009445    0.2080791  0.2081014
#>  [384,]    0.2035599    0.2107873  0.2108099
#>  [385,]    0.3115048    0.3225645  0.3225991
#>  [386,]    0.1522950    0.1577024  0.1577193
#>  [387,]    0.3220637    0.3334983  0.3335339
#>  [388,]    0.2049951    0.2122735  0.2122962
#>  [389,]    0.2087190    0.2161296  0.2161527
#>  [390,]    0.2964055    0.3069294  0.3069622
#>  [391,]    0.2096113    0.2170536  0.2170768
#>  [392,]    0.2829323    0.2929778  0.2930092
#>  [393,]    0.2003334    0.2074464  0.2074686
#>  [394,]    0.2173836    0.2251019  0.2251260
#>  [395,]    0.1690821    0.1750854  0.1751042
#>  [396,]    0.2952497    0.3057323  0.3057650
#>  [397,]    0.1941424    0.2010355  0.2010570
#>  [398,]    0.1846480    0.1912041  0.1912246
#>  [399,]    0.2028626    0.2100654  0.2100878
#>  [400,]    0.1745944    0.1807936  0.1808129
#>  [401,]    0.1968165    0.2038045  0.2038263
#>  [402,]    0.1861495    0.1927588  0.1927794
#>  [403,]    0.1729497    0.1790904  0.1791096
#>  [404,]    0.2683021    0.2778280  0.2778578
#>  [405,]    0.1926129    0.1994518  0.1994731
#>  [406,]    0.1657970    0.1716837  0.1717021
#>  [407,]    0.2130838    0.2206494  0.2206730
#>  [408,]    0.1819106    0.1883695  0.1883896
#>  [409,]    0.2008397    0.2079707  0.2079929
#>  [410,]    0.1658805    0.1717702  0.1717886
#>  [411,]    0.1842086    0.1907491  0.1907695
#>  [412,]    0.2029989    0.2102065  0.2102290
#>  [413,]    0.2140994    0.2217010  0.2217248
#>  [414,]    0.1744693    0.1806639  0.1806833
#>  [415,]    0.3072403    0.3181488  0.3181828
#>  [416,]    0.1403322    0.1453148  0.1453304
#>  [417,]    0.1582103    0.1638277  0.1638452
#>  [418,]    0.1826009    0.1890842  0.1891045
#>  [419,]    0.1958451    0.2027987  0.2028204
#>  [420,]    0.1689723    0.1749718  0.1749905
#>  [421,]    0.2210472    0.2288956  0.2289200
#>  [422,]    0.2403064    0.2488385  0.2488651
#>  [423,]    0.2378310    0.2462753  0.2463017
#>  [424,]    0.1990642    0.2061321  0.2061541
#>  [425,]    0.2158783    0.2235431  0.2235670
#>  [426,]    0.2162344    0.2239119  0.2239359
#>  [427,]    0.1737609    0.1799303  0.1799496
#>  [428,]    0.1524534    0.1578663  0.1578832
#>  [429,]    0.2536010    0.2626050  0.2626331
#>  [430,]    0.1502596    0.1555947  0.1556114
#>  [431,]    0.3001155    0.3107709  0.3108042
#>  [432,]    0.2113282    0.2188315  0.2188549
#>  [433,]    0.2876325    0.2978449  0.2978767
#>  [434,]    0.2574829    0.2666247  0.2666533
#>  [435,]    0.2845317    0.2946339  0.2946654
#>  [436,]    0.1939041    0.2007888  0.2008103
#>  [437,]    0.2668978    0.2763740  0.2764036
#>  [438,]    0.2102720    0.2177378  0.2177611
#>  [439,]    0.2078254    0.2152044  0.2152275
#>  [440,]    0.2015369    0.2086926  0.2087149
#>  [441,]    0.2099842    0.2174398  0.2174631
#>  [442,]    0.2080084    0.2153938  0.2154169
#>  [443,]    0.2770311    0.2868671  0.2868978
#>  [444,]    0.1675949    0.1735456  0.1735641
#>  [445,]    0.2071309    0.2144852  0.2145081
#>  [446,]    0.2097201    0.2171664  0.2171896
#>  [447,]    0.1932359    0.2000968  0.2001182
#>  [448,]    0.3061215    0.3169902  0.3170241
#>  [449,]    0.3168856    0.3281364  0.3281715
#>  [450,]    0.1713876    0.1774729  0.1774918
#>  [451,]    0.1787832    0.1851310  0.1851508
#>  [452,]    0.1733818    0.1795379  0.1795571
#>  [453,]    0.1840167    0.1905504  0.1905708
#>  [454,]    0.1689869    0.1749870  0.1750057
#>  [455,]    0.2067903    0.2141325  0.2141554
#>  [456,]    0.2041307    0.2113785  0.2114011
#>  [457,]    0.1613397    0.1670682  0.1670861
#>  [458,]    0.2115192    0.2190293  0.2190527
#>  [459,]    0.2414131    0.2499845  0.2500112
#>  [460,]    0.2304008    0.2385813  0.2386068
#>  [461,]    0.2123591    0.2198990  0.2199225
#>  [462,]    0.1555729    0.1610966  0.1611138
#>  [463,]    0.1623964    0.1681625  0.1681805
#>  [464,]    0.1983632    0.2054061  0.2054281
#>  [465,]    0.2000843    0.2071885  0.2072106
#>  [466,]    0.2531624    0.2621510  0.2621790
#>  [467,]    0.2012489    0.2083943  0.2084166
#>  [468,]    0.1791518    0.1855127  0.1855325
#>  [469,]    0.1810276    0.1874551  0.1874751
#>  [470,]    0.2885406    0.2987852  0.2988171
#>  [471,]    0.2402004    0.2487288  0.2487554
#>  [472,]    0.1266568    0.1311539  0.1311680
#>  [473,]    0.1854144    0.1919977  0.1920182
#>  [474,]    0.1568063    0.1623738  0.1623912
#>  [475,]    0.2262835    0.2343178  0.2343428
#>  [476,]    0.1904799    0.1972430  0.1972641
#>  [477,]    0.2374977    0.2459300  0.2459563
#>  [478,]    0.2013588    0.2085082  0.2085305
#>  [479,]    0.1902678    0.1970234  0.1970444
#>  [480,]    0.3286402    0.3403083  0.3403447
#>  [481,]    0.1768509    0.1831301  0.1831497
#>  [482,]    0.1832503    0.1897567  0.1897770
#>  [483,]    0.2323657    0.2406159  0.2406416
#>  [484,]    0.1520550    0.1574539  0.1574707
#>  [485,]    0.1290282    0.1336095  0.1336238
#>  [486,]    0.1944115    0.2013142  0.2013358
#>  [487,]    0.1489907    0.1542808  0.1542973
#>  [488,]    0.2060518    0.2133677  0.2133905
#>  [489,]    0.2667202    0.2761902  0.2762197
#>  [490,]    0.1961403    0.2031044  0.2031261
#>  [491,]    0.1718541    0.1779559  0.1779750
#>  [492,]    0.2277691    0.2358560  0.2358813
#>  [493,]    0.1803568    0.1867605  0.1867805
#>  [494,]    0.2199088    0.2277168  0.2277412
#>  [495,]    0.2651433    0.2745570  0.2745864
#>  [496,]    0.1851814    0.1917564  0.1917769
#>  [497,]    0.2179765    0.2257158  0.2257399
#>  [498,]    0.2263815    0.2344192  0.2344443
#>  [499,]    0.2406961    0.2492421  0.2492687
#>  [500,]    0.1816790    0.1881297  0.1881498
#>  [501,]    0.2659883    0.2754321  0.2754615
#>  [502,]    0.1929594    0.1998105  0.1998319
#>  [503,]    0.2244633    0.2324328  0.2324577
#>  [504,]    0.1681810    0.1741525  0.1741711
#>  [505,]    0.1480810    0.1533387  0.1533551
#>  [506,]    0.1812569    0.1876925  0.1877126
#>  [507,]    0.2310029    0.2392047  0.2392303
#>  [508,]    0.1634614    0.1692652  0.1692833
#>  [509,]    0.1768899    0.1831705  0.1831901
#>  [510,]    0.2350467    0.2433920  0.2434180
#>  [511,]    0.2238815    0.2318306  0.2318554
#>  [512,]    0.1768101    0.1830878  0.1831074
#>  [513,]    0.2056077    0.2129079  0.2129307
#>  [514,]    0.1957309    0.2026804  0.2027020
#>  [515,]    0.1652464    0.1711136  0.1711319
#>  [516,]    0.2697912    0.2793700  0.2793999
#>  [517,]    0.2141985    0.2218037  0.2218274
#>  [518,]    0.1634243    0.1692268  0.1692449
#>  [519,]    0.1986166    0.2056685  0.2056905
#>  [520,]    0.1557664    0.1612970  0.1613143
#>  [521,]    0.2932852    0.3036981  0.3037306
#>  [522,]    0.2403185    0.2488511  0.2488778
#>  [523,]    0.2001358    0.2072417  0.2072639
#>  [524,]    0.2034723    0.2106966  0.2107192
#>  [525,]    0.2615049    0.2707897  0.2708187
#>  [526,]    0.1912535    0.1980441  0.1980653
#>  [527,]    0.2116238    0.2191376  0.2191611
#>  [528,]    0.2006711    0.2077961  0.2078183
#>  [529,]    0.1400383    0.1450104  0.1450259
#>  [530,]    0.2095464    0.2169865  0.2170097
#>  [531,]    0.1861551    0.1927646  0.1927852
#>  [532,]    0.1895338    0.1962633  0.1962843
#>  [533,]    0.1778498    0.1841644  0.1841842
#>  [534,]    0.2192628    0.2270477  0.2270720
#>  [535,]    0.2192503    0.2270348  0.2270591
#>  [536,]    0.1880907    0.1947689  0.1947897
#>  [537,]    0.1899032    0.1966458  0.1966668
#>  [538,]    0.2139347    0.2215306  0.2215543
#>  [539,]    0.1856990    0.1922924  0.1923130
#>  [540,]    0.1945419    0.2014492  0.2014708
#>  [541,]    0.1695549    0.1755751  0.1755939
#>  [542,]    0.2407470    0.2492947  0.2493214
#>  [543,]    0.2434535    0.2520974  0.2521244
#>  [544,]    0.2413915    0.2499621  0.2499888
#>  [545,]    0.2175992    0.2253252  0.2253493
#>  [546,]    0.1906924    0.1974629  0.1974841
#>  [547,]    0.2433372    0.2519769  0.2520039
#>  [548,]    0.1819951    0.1884570  0.1884771
#>  [549,]    0.2368303    0.2452390  0.2452652
#>  [550,]    0.2666724    0.2761406  0.2761701
#>  [551,]    0.3049066    0.3157321  0.3157659
#>  [552,]    0.2988576    0.3094683  0.3095014
#>  [553,]    0.1689700    0.1749694  0.1749881
#>  [554,]    0.1812661    0.1877021  0.1877222
#>  [555,]    0.2273784    0.2354515  0.2354767
#>  [556,]    0.2084946    0.2158973  0.2159204
#>  [557,]    0.2370673    0.2454845  0.2455107
#>  [558,]    0.2331619    0.2414402  0.2414661
#>  [559,]    0.2593367    0.2685445  0.2685732
#>  [560,]    0.1926513    0.1994915  0.1995128
#>  [561,]    0.2075609    0.2149305  0.2149535
#>  [562,]    0.1813614    0.1878008  0.1878209
#>  [563,]    0.1634552    0.1692588  0.1692770
#>  [564,]    0.1779252    0.1842426  0.1842624
#>  [565,]    0.2221559    0.2300436  0.2300683
#>  [566,]    0.1519951    0.1573918  0.1574086
#>  [567,]    0.1946782    0.2015904  0.2016120
#>  [568,]    0.2366979    0.2451018  0.2451280
#>  [569,]    0.2695394    0.2791093  0.2791391
#>  [570,]    0.1564831    0.1620392  0.1620566
#>  [571,]    0.2185416    0.2263011  0.2263253
#>  [572,]    0.2033885    0.2106098  0.2106323
#>  [573,]    0.1820387    0.1885021  0.1885222
#>  [574,]    0.2500145    0.2588913  0.2589190
#>  [575,]    0.2507134    0.2596150  0.2596427
#>  [576,]    0.1955286    0.2024710  0.2024926
#>  [577,]    0.2247075    0.2326858  0.2327107
#>  [578,]    0.2028629    0.2100657  0.2100882
#>  [579,]    0.2249058    0.2328911  0.2329161
#>  [580,]    0.2039587    0.2112003  0.2112229
#>  [581,]    0.1764902    0.1827566  0.1827761
#>  [582,]    0.2008779    0.2080102  0.2080324
#>  [583,]    0.2166329    0.2243245  0.2243485
#>  [584,]    0.2179216    0.2256590  0.2256832
#>  [585,]    0.1816491    0.1880986  0.1881188
#>  [586,]    0.2207216    0.2285584  0.2285828
#>  [587,]    0.2465052    0.2552573  0.2552846
#>  [588,]    0.1881687    0.1948498  0.1948706
#>  [589,]    0.2444766    0.2531568  0.2531839
#>  [590,]    0.1535278    0.1589789  0.1589959
#>  [591,]    0.2122407    0.2197764  0.2197999
#>  [592,]    0.1700417    0.1760792  0.1760980
#>  [593,]    0.2588294    0.2680190  0.2680477
#>  [594,]    0.1575429    0.1631366  0.1631540
#>  [595,]    0.1828926    0.1893863  0.1894066
#>  [596,]    0.1919855    0.1988019  0.1988232
#>  [597,]    0.2023959    0.2095822  0.2096046
#>  [598,]    0.2407130    0.2492596  0.2492863
#>  [599,]    0.1786831    0.1850274  0.1850472
#>  [600,]    0.1472003    0.1524268  0.1524431
#>  [601,]    0.1406088    0.1456013  0.1456168
#>  [602,]    0.1899441    0.1966883  0.1967093
#>  [603,]    0.2533502    0.2623453  0.2623734
#>  [604,]    0.1868644    0.1934992  0.1935199
#>  [605,]    0.1879602    0.1946338  0.1946546
#>  [606,]    0.1861957    0.1928066  0.1928273
#>  [607,]    0.2011119    0.2082525  0.2082748
#>  [608,]    0.1978201    0.2048437  0.2048656
#>  [609,]    0.2073176    0.2146786  0.2147015
#>  [610,]    0.1840271    0.1905611  0.1905815
#>  [611,]    0.1639296    0.1697500  0.1697682
#>  [612,]    0.1475381    0.1527766  0.1527930
#>  [613,]    0.1957560    0.2027064  0.2027281
#>  [614,]    0.2309707    0.2391714  0.2391970
#>  [615,]    0.2147122    0.2223357  0.2223594
#>  [616,]    0.1572209    0.1628031  0.1628205
#>  [617,]    0.1668647    0.1727895  0.1728080
#>  [618,]    0.1882410    0.1949247  0.1949455
#>  [619,]    0.1879324    0.1946050  0.1946259
#>  [620,]    0.1980427    0.2050744  0.2050963
#>  [621,]    0.2087814    0.2161943  0.2162175
#>  [622,]    0.2195908    0.2273875  0.2274118
#>  [623,]    0.2056606    0.2129627  0.2129855
#>  [624,]    0.1894166    0.1961420  0.1961630
#>  [625,]    0.2001849    0.2072926  0.2073148
#>  [626,]    0.2231924    0.2311169  0.2311416
#>  [627,]    0.2562814    0.2653807  0.2654091
#>  [628,]    0.1694453    0.1754616  0.1754804
#>  [629,]    0.2076805    0.2150543  0.2150774
#>  [630,]    0.2626247    0.2719491  0.2719782
#>  [631,]    0.2012208    0.2083652  0.2083875
#>  [632,]    0.2565958    0.2657062  0.2657346
#>  [633,]    0.2571283    0.2662577  0.2662862
#>  [634,]    0.1891538    0.1958698  0.1958908
#>  [635,]    0.1675218    0.1734698  0.1734884
#>  [636,]    0.1963627    0.2033348  0.2033565
#>  [637,]    0.2073524    0.2147146  0.2147376
#>  [638,]    0.1893834    0.1961076  0.1961286
#>  [639,]    0.2339541    0.2422607  0.2422866
#>  [640,]    0.1557613    0.1612918  0.1613090
#>  [641,]    0.2823767    0.2924024  0.2924337
#>  [642,]    0.1626866    0.1684629  0.1684810
#>  [643,]    0.1844098    0.1909574  0.1909779
#>  [644,]    0.2779399    0.2878081  0.2878389
#>  [645,]    0.2522930    0.2612507  0.2612787
#>  [646,]    0.2695406    0.2791106  0.2791405
#>  [647,]    0.1242765    0.1286891  0.1287029
#>  [648,]    0.2231140    0.2310357  0.2310605
#>  [649,]    0.1534435    0.1588917  0.1589087
#>  [650,]    0.2127109    0.2202633  0.2202868
#>  [651,]    0.1611869    0.1669099  0.1669278
#>  [652,]    0.2203834    0.2282082  0.2282326
#>  [653,]    0.2004312    0.2075476  0.2075698
#>  [654,]    0.1835856    0.1901039  0.1901243
#>  [655,]    0.1872250    0.1938726  0.1938934
#>  [656,]    0.2320816    0.2403217  0.2403474
#>  [657,]    0.1948807    0.2018000  0.2018216
#>  [658,]    0.2134550    0.2210338  0.2210575
#>  [659,]    0.1925680    0.1994053  0.1994267
#>  [660,]    0.2059166    0.2132278  0.2132506
#>  [661,]    0.1757252    0.1819645  0.1819840
#>  [662,]    0.2567734    0.2658901  0.2659185
#>  [663,]    0.1669493    0.1728770  0.1728955
#>  [664,]    0.2025942    0.2097874  0.2098099
#>  [665,]    0.1558951    0.1614303  0.1614476
#>  [666,]    0.2114093    0.2189155  0.2189389
#>  [667,]    0.2336732    0.2419698  0.2419957
#>  [668,]    0.1609050    0.1666181  0.1666360
#>  [669,]    0.2162465    0.2239245  0.2239484
#>  [670,]    0.2597953    0.2690192  0.2690480
#>  [671,]    0.2165945    0.2242848  0.2243088
#>  [672,]    0.3044335    0.3152423  0.3152760
#>  [673,]    0.2157695    0.2234306  0.2234545
#>  [674,]    0.2720502    0.2817093  0.2817394
#>  [675,]    0.2316430    0.2398675  0.2398932
#>  [676,]    0.1998420    0.2069374  0.2069596
#>  [677,]    0.2141340    0.2217368  0.2217606
#>  [678,]    0.2329288    0.2411989  0.2412247
#>  [679,]    0.1833423    0.1898520  0.1898723
#>  [680,]    0.1903335    0.1970914  0.1971124
#>  [681,]    0.2210917    0.2289416  0.2289661
#>  [682,]    0.1823271    0.1888007  0.1888209
#>  [683,]    0.2007028    0.2078289  0.2078511
#>  [684,]    0.1692778    0.1752881  0.1753069
#>  [685,]    0.2075497    0.2149188  0.2149418
#>  [686,]    0.2293095    0.2374512  0.2374766
#>  [687,]    0.2054171    0.2127106  0.2127333
#>  [688,]    0.1814685    0.1879116  0.1879317
#>  [689,]    0.2364944    0.2448912  0.2449174
#>  [690,]    0.1756552    0.1818920  0.1819114
#>  [691,]    0.1505696    0.1559158  0.1559325
#>  [692,]    0.2251336    0.2331270  0.2331520
#>  [693,]    0.2392918    0.2477879  0.2478145
#>  [694,]    0.2564343    0.2655389  0.2655674
#>  [695,]    0.1956891    0.2026372  0.2026589
#>  [696,]    0.1778213    0.1841350  0.1841547
#>  [697,]    0.1610585    0.1667770  0.1667949
#>  [698,]    0.2954095    0.3058979  0.3059307
#>  [699,]    0.1803039    0.1867058  0.1867257
#>  [700,]    0.1872715    0.1939207  0.1939415
#>  [701,]    0.1744315    0.1806249  0.1806442
#>  [702,]    0.2326491    0.2409094  0.2409351
#>  [703,]    0.1652034    0.1710691  0.1710874
#>  [704,]    0.1792614    0.1856263  0.1856461
#>  [705,]    0.1626375    0.1684121  0.1684301
#>  [706,]    0.1284492    0.1330099  0.1330242
#>  [707,]    0.1777898    0.1841023  0.1841220
#>  [708,]    0.1856104    0.1922006  0.1922211
#>  [709,]    0.1655100    0.1713866  0.1714049
#>  [710,]    0.2940884    0.3045299  0.3045624
#>  [711,]    0.2005351    0.2076553  0.2076775
#>  [712,]    0.1803057    0.1867076  0.1867275
#>  [713,]    0.1915918    0.1983944  0.1984156
#>  [714,]    0.2880477    0.2982749  0.2983068
#>  [715,]    0.2325421    0.2407985  0.2408243
#>  [716,]    0.2874140    0.2976185  0.2976503
#>  [717,]    0.1912462    0.1980366  0.1980578
#>  [718,]    0.1804708    0.1868785  0.1868985
#>  [719,]    0.1981843    0.2052210  0.2052430
#>  [720,]    0.2516268    0.2605608  0.2605887
#>  [721,]    0.1725806    0.1787083  0.1787274
#>  [722,]    0.1955697    0.2025136  0.2025352
#>  [723,]    0.2159169    0.2235831  0.2236071
#>  [724,]    0.2300202    0.2381870  0.2382125
#>  [725,]    0.1844304    0.1909787  0.1909991
#>  [726,]    0.2292972    0.2374385  0.2374639
#>  [727,]    0.2558494    0.2649332  0.2649616
#>  [728,]    0.1850548    0.1916252  0.1916457
#>  [729,]    0.1977881    0.2048107  0.2048326
#>  [730,]    0.3393401    0.3513881  0.3514257
#>  [731,]    0.2403969    0.2489321  0.2489588
#>  [732,]    0.2578878    0.2670441  0.2670726
#>  [733,]    0.2081122    0.2155014  0.2155245
#>  [734,]    0.1728333    0.1789698  0.1789890
#>  [735,]    0.1764515    0.1827165  0.1827361
#>  [736,]    0.1876283    0.1942901  0.1943109
#>  [737,]    0.1420952    0.1471404  0.1471561
#>  [738,]    0.1737185    0.1798865  0.1799057
#>  [739,]    0.1804762    0.1868841  0.1869041
#>  [740,]    0.1948367    0.2017544  0.2017760
#>  [741,]    0.1923273    0.1991560  0.1991773
#>  [742,]    0.2159621    0.2236299  0.2236538
#>  [743,]    0.1607625    0.1664705  0.1664883
#>  [744,]    0.2392955    0.2477918  0.2478183
#>  [745,]    0.2087344    0.2161456  0.2161687
#>  [746,]    0.1920276    0.1988456  0.1988668
#>  [747,]    0.2782720    0.2881520  0.2881829
#>  [748,]    0.2891731    0.2994400  0.2994720
#>  [749,]    0.1727041    0.1788361  0.1788553
#>  [750,]    0.2428521    0.2514746  0.2515015
#>  [751,]    0.1793603    0.1857286  0.1857485
#>  [752,]    0.2253878    0.2333903  0.2334152
#>  [753,]    0.1503310    0.1556686  0.1556853
#>  [754,]    0.2155704    0.2232243  0.2232482
#>  [755,]    0.1437145    0.1488172  0.1488331
#>  [756,]    0.1588307    0.1644702  0.1644878
#>  [757,]    0.2144953    0.2221111  0.2221348
#>  [758,]    0.1703462    0.1763945  0.1764134
#>  [759,]    0.1805619    0.1869728  0.1869928
#>  [760,]    0.2605814    0.2698333  0.2698622
#>  [761,]    0.1624611    0.1682294  0.1682474
#>  [762,]    0.1771670    0.1834574  0.1834770
#>  [763,]    0.2351815    0.2435316  0.2435577
#>  [764,]    0.2265531    0.2345969  0.2346220
#>  [765,]    0.1839199    0.1904502  0.1904706
#>  [766,]    0.2260752    0.2341020  0.2341270
#>  [767,]    0.2249286    0.2329148  0.2329397
#>  [768,]    0.1625597    0.1683315  0.1683495
#>  [769,]    0.1782016    0.1845288  0.1845486
#>  [770,]    0.2320233    0.2402614  0.2402871
#>  [771,]    0.1965316    0.2035096  0.2035314
#>  [772,]    0.3075685    0.3184885  0.3185226
#>  [773,]    0.1752421    0.1814642  0.1814836
#>  [774,]    0.2037660    0.2110008  0.2110234
#>  [775,]    0.1803979    0.1868031  0.1868230
#>  [776,]    0.2100069    0.2174632  0.2174865
#>  [777,]    0.2115148    0.2190248  0.2190482
#>  [778,]    0.2160929    0.2237654  0.2237893
#>  [779,]    0.1587272    0.1643630  0.1643806
#>  [780,]    0.2390407    0.2475278  0.2475543
#>  [781,]    0.1877808    0.1944481  0.1944689
#>  [782,]    0.1972281    0.2042308  0.2042526
#>  [783,]    0.2257727    0.2337888  0.2338138
#>  [784,]    0.3025136    0.3132543  0.3132878
#>  [785,]    0.1729823    0.1791241  0.1791433
#>  [786,]    0.1985002    0.2055481  0.2055701
#>  [787,]    0.3132641    0.3243863  0.3244210
#>  [788,]    0.2127980    0.2203535  0.2203770
#>  [789,]    0.2107380    0.2182204  0.2182438
#>  [790,]    0.2671559    0.2766413  0.2766709
#>  [791,]    0.1652547    0.1711222  0.1711406
#>  [792,]    0.1809786    0.1874043  0.1874244
#>  [793,]    0.1487793    0.1540619  0.1540784
#>  [794,]    0.2613955    0.2706764  0.2707053
#>  [795,]    0.2116793    0.2191951  0.2192186
#>  [796,]    0.2870358    0.2972269  0.2972587
#>  [797,]    0.1415115    0.1465360  0.1465517
#>  [798,]    0.1782863    0.1846165  0.1846362
#>  [799,]    0.1899881    0.1967338  0.1967548
#>  [800,]    0.2200275    0.2278396  0.2278640
#>  [801,]    0.2167134    0.2244079  0.2244319
#>  [802,]    0.2088225    0.2162368  0.2162600
#>  [803,]    0.2417833    0.2503678  0.2503946
#>  [804,]    0.1824451    0.1889229  0.1889432
#>  [805,]    0.1866425    0.1932693  0.1932900
#>  [806,]    0.1698324    0.1758624  0.1758812
#>  [807,]    0.1969266    0.2039185  0.2039403
#>  [808,]    0.2017837    0.2089481  0.2089705
#>  [809,]    0.1597963    0.1654700  0.1654877
#>  [810,]    0.2162874    0.2239666  0.2239906
#>  [811,]    0.2307937    0.2389881  0.2390137
#>  [812,]    0.1898294    0.1965694  0.1965905
#>  [813,]    0.1826504    0.1891355  0.1891558
#>  [814,]    0.1949482    0.2018700  0.2018916
#>  [815,]    0.2467832    0.2555453  0.2555727
#>  [816,]    0.1353451    0.1401507  0.1401657
#>  [817,]    0.2104295    0.2179009  0.2179243
#>  [818,]    0.1664621    0.1723725  0.1723909
#>  [819,]    0.2436294    0.2522796  0.2523066
#>  [820,]    0.1862563    0.1928695  0.1928901
#>  [821,]    0.1968841    0.2038745  0.2038963
#>  [822,]    0.2113384    0.2188421  0.2188656
#>  [823,]    0.1673153    0.1732559  0.1732744
#>  [824,]    0.2385715    0.2470421  0.2470685
#>  [825,]    0.2591563    0.2683576  0.2683863
#>  [826,]    0.3140854    0.3252369  0.3252717
#>  [827,]    0.1795943    0.1859710  0.1859909
#>  [828,]    0.2153662    0.2230128  0.2230367
#>  [829,]    0.1811753    0.1876080  0.1876281
#>  [830,]    0.2116665    0.2191817  0.2192052
#>  [831,]    0.2280744    0.2361722  0.2361975
#>  [832,]    0.2368231    0.2452315  0.2452578
#>  [833,]    0.1919887    0.1988054  0.1988266
#>  [834,]    0.1752471    0.1814694  0.1814888
#>  [835,]    0.1402031    0.1451811  0.1451967
#>  [836,]    0.1498474    0.1551679  0.1551845
#>  [837,]    0.1497424    0.1550592  0.1550758
#>  [838,]    0.1766847    0.1829580  0.1829775
#>  [839,]    0.2147612    0.2223864  0.2224102
#>  [840,]    0.2052813    0.2125699  0.2125927
#>  [841,]    0.1941215    0.2010139  0.2010354
#>  [842,]    0.2377024    0.2461421  0.2461684
#>  [843,]    0.2092787    0.2167093  0.2167325
#>  [844,]    0.2649163    0.2743221  0.2743514
#>  [845,]    0.2884004    0.2986400  0.2986719
#>  [846,]    0.1580643    0.1636765  0.1636940
#>  [847,]    0.2045930    0.2118572  0.2118798
#>  [848,]    0.2004558    0.2075732  0.2075954
#>  [849,]    0.2485852    0.2574112  0.2574388
#>  [850,]    0.2393084    0.2478051  0.2478317
#>  [851,]    0.2009820    0.2081181  0.2081403
#>  [852,]    0.2006643    0.2077890  0.2078113
#>  [853,]    0.1759909    0.1822395  0.1822590
#>  [854,]    0.2097321    0.2171789  0.2172021
#>  [855,]    0.1860912    0.1926985  0.1927191
#>  [856,]    0.2010747    0.2082139  0.2082362
#>  [857,]    0.2801943    0.2901424  0.2901735
#>  [858,]    0.2153041    0.2229486  0.2229724
#>  [859,]    0.3126899    0.3237918  0.3238265
#>  [860,]    0.1286971    0.1332666  0.1332809
#>  [861,]    0.1981313    0.2051661  0.2051881
#>  [862,]    0.1883558    0.1950434  0.1950643
#>  [863,]    0.1508955    0.1562532  0.1562699
#>  [864,]    0.2472029    0.2559799  0.2560073
#>  [865,]    0.2768347    0.2866636  0.2866943
#>  [866,]    0.1547362    0.1602302  0.1602474
#>  [867,]    0.2634270    0.2727799  0.2728091
#>  [868,]    0.1797978    0.1861816  0.1862015
#>  [869,]    0.1796261    0.1860039  0.1860238
#>  [870,]    0.1810609    0.1874896  0.1875097
#>  [871,]    0.1609211    0.1666347  0.1666525
#>  [872,]    0.2045859    0.2118498  0.2118725
#>  [873,]    0.2128925    0.2204513  0.2204749
#>  [874,]    0.2190144    0.2267907  0.2268149
#>  [875,]    0.1425444    0.1476056  0.1476214
#>  [876,]    0.2942034    0.3046490  0.3046816
#>  [877,]    0.2002290    0.2073382  0.2073604
#>  [878,]    0.2237045    0.2316472  0.2316720
#>  [879,]    0.1962827    0.2032519  0.2032737
#>  [880,]    0.2473560    0.2561383  0.2561657
#>  [881,]    0.2363639    0.2447558  0.2447820
#>  [882,]    0.2203884    0.2282134  0.2282378
#>  [883,]    0.2248743    0.2328585  0.2328834
#>  [884,]    0.2224606    0.2303592  0.2303838
#>  [885,]    0.1789787    0.1853335  0.1853533
#>  [886,]    0.2141756    0.2217801  0.2218038
#>  [887,]    0.2135842    0.2211677  0.2211913
#>  [888,]    0.1913801    0.1981751  0.1981964
#>  [889,]    0.1901470    0.1968983  0.1969193
#>  [890,]    0.2018549    0.2090218  0.2090442
#>  [891,]    0.2287142    0.2368348  0.2368601
#>  [892,]    0.1744682    0.1806629  0.1806822
#>  [893,]    0.2082212    0.2156143  0.2156373
#>  [894,]    0.1666408    0.1725576  0.1725760
#>  [895,]    0.2557966    0.2648785  0.2649069
#>  [896,]    0.2734261    0.2831341  0.2831644
#>  [897,]    0.2352444    0.2435969  0.2436229
#>  [898,]    0.1425433    0.1476045  0.1476203
#>  [899,]    0.1696473    0.1756708  0.1756896
#>  [900,]    0.1617013    0.1674426  0.1674605
#>  [901,]    0.1762823    0.1825413  0.1825609
#>  [902,]    0.1691263    0.1751313  0.1751500
#>  [903,]    0.2192045    0.2269874  0.2270117
#>  [904,]    0.2252111    0.2332072  0.2332322
#>  [905,]    0.2412527    0.2498185  0.2498452
#>  [906,]    0.1711011    0.1771762  0.1771951
#>  [907,]    0.2587413    0.2679279  0.2679565
#>  [908,]    0.2539044    0.2629193  0.2629475
#>  [909,]    0.1709033    0.1769713  0.1769902
#>  [910,]    0.1711365    0.1772128  0.1772318
#>  [911,]    0.2041127    0.2113598  0.2113825
#>  [912,]    0.1903579    0.1971166  0.1971377
#>  [913,]    0.3256862    0.3372495  0.3372856
#>  [914,]    0.1592549    0.1649095  0.1649271
#>  [915,]    0.1300365    0.1346536  0.1346680
#>  [916,]    0.2260127    0.2340374  0.2340624
#>  [917,]    0.1754959    0.1817269  0.1817464
#>  [918,]    0.1924994    0.1993342  0.1993556
#>  [919,]    0.2168613    0.2245609  0.2245849
#>  [920,]    0.1724739    0.1785978  0.1786169
#>  [921,]    0.2168939    0.2245949  0.2246189
#>  [922,]    0.2205542    0.2283849  0.2284093
#>  [923,]    0.1736286    0.1797935  0.1798127
#>  [924,]    0.3120322    0.3231107  0.3231453
#>  [925,]    0.2349881    0.2433314  0.2433574
#>  [926,]    0.2539501    0.2629667  0.2629949
#>  [927,]    0.2084930    0.2158956  0.2159187
#>  [928,]    0.1867974    0.1934297  0.1934504
#>  [929,]    0.2391635    0.2476550  0.2476815
#>  [930,]    0.2546072    0.2636470  0.2636752
#>  [931,]    0.2114589    0.2189670  0.2189904
#>  [932,]    0.2166077    0.2242984  0.2243224
#>  [933,]    0.2013939    0.2085445  0.2085668
#>  [934,]    0.1725855    0.1787134  0.1787325
#>  [935,]    0.2112065    0.2187055  0.2187289
#>  [936,]    0.3379288    0.3499265  0.3499639
#>  [937,]    0.1971931    0.2041946  0.2042164
#>  [938,]    0.1881702    0.1948512  0.1948721
#>  [939,]    0.1818089    0.1882641  0.1882843
#>  [940,]    0.2075219    0.2148900  0.2149130
#>  [941,]    0.2166843    0.2243778  0.2244018
#>  [942,]    0.1936600    0.2005360  0.2005575
#>  [943,]    0.2011117    0.2082523  0.2082746
#>  [944,]    0.1926419    0.1994817  0.1995031
#>  [945,]    0.1918361    0.1986474  0.1986687
#>  [946,]    0.2140647    0.2216652  0.2216889
#>  [947,]    0.2190045    0.2267804  0.2268046
#>  [948,]    0.2040906    0.2113369  0.2113595
#>  [949,]    0.2406416    0.2491856  0.2492123
#>  [950,]    0.1797290    0.1861104  0.1861304
#>  [951,]    0.1637243    0.1695375  0.1695556
#>  [952,]    0.1714153    0.1775015  0.1775205
#>  [953,]    0.1793463    0.1857141  0.1857340
#>  [954,]    0.2001735    0.2072808  0.2073029
#>  [955,]    0.2953147    0.3057998  0.3058325
#>  [956,]    0.2157320    0.2233917  0.2234156
#>  [957,]    0.1970440    0.2040402  0.2040620
#>  [958,]    0.2021149    0.2092910  0.2093134
#>  [959,]    0.1506177    0.1559655  0.1559822
#>  [960,]    0.2115920    0.2191046  0.2191280
#>  [961,]    0.2496419    0.2585055  0.2585331
#>  [962,]    0.2060043    0.2133185  0.2133413
#>  [963,]    0.1861947    0.1928056  0.1928262
#>  [964,]    0.2542448    0.2632717  0.2632998
#>  [965,]    0.1916693    0.1984746  0.1984959
#>  [966,]    0.2682521    0.2777763  0.2778061
#>  [967,]    0.2129239    0.2204838  0.2205074
#>  [968,]    0.3140899    0.3252411  0.3252759
#>  [969,]    0.2268469    0.2349011  0.2349263
#>  [970,]    0.2616638    0.2709542  0.2709832
#>  [971,]    0.1663251    0.1722306  0.1722490
#>  [972,]    0.2565120    0.2656196  0.2656480
#>  [973,]    0.1772075    0.1834994  0.1835190
#>  [974,]    0.1846811    0.1912383  0.1912587
#>  [975,]    0.2809350    0.2909096  0.2909407
#>  [976,]    0.1839222    0.1904524  0.1904728
#>  [977,]    0.2121011    0.2196318  0.2196553
#>  [978,]    0.2282363    0.2363399  0.2363652
#>  [979,]    0.1992367    0.2063107  0.2063328
#>  [980,]    0.1721189    0.1782302  0.1782493
#>  [981,]    0.1756166    0.1818520  0.1818715
#>  [982,]    0.2404184    0.2489545  0.2489811
#>  [983,]    0.2955802    0.3060747  0.3061075
#>  [984,]    0.2137846    0.2213751  0.2213988
#>  [985,]    0.2641670    0.2735462  0.2735755
#>  [986,]    0.1724502    0.1785732  0.1785923
#>  [987,]    0.3079056    0.3188375  0.3188716
#>  [988,]    0.1741382    0.1803211  0.1803404
#>  [989,]    0.1560696    0.1616110  0.1616283
#>  [990,]    0.1987061    0.2057612  0.2057832
#>  [991,]    0.1987322    0.2057884  0.2058104
#>  [992,]    0.1727865    0.1789214  0.1789406
#>  [993,]    0.2836095    0.2936790  0.2937105
#>  [994,]    0.3204157    0.3317918  0.3318273
#>  [995,]    0.2154810    0.2231317  0.2231556
#>  [996,]    0.2122609    0.2197973  0.2198208
#>  [997,]    0.1810929    0.1875227  0.1875428
#>  [998,]    0.2064349    0.2137644  0.2137872
#>  [999,]    0.2450429    0.2537431  0.2537703
#> [1000,]    0.1889832    0.1956932  0.1957141