Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 96.24138
#> 2   Feb 91.72429
#> 3   Mar 90.44844
#> 4   Apr 89.26389
#> 5   May 90.97750
#> 6   Jun 86.13931

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 6.265188  2
#> 2   Feb 6.024272  2
#> 3   Mar 6.627605  2
#> 4   Apr 8.211236  2
#> 5   May 7.682736  2
#> 6   Jun 8.465132  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [12ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [23ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [70ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [759ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [20ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1951457    0.1982647  0.2047031
#>    [2,]    0.2018120    0.2050376  0.2116960
#>    [3,]    0.2247421    0.2283342  0.2357491
#>    [4,]    0.1890047    0.1920256  0.1982615
#>    [5,]    0.2064486    0.2097483  0.2165597
#>    [6,]    0.2707901    0.2751181  0.2840520
#>    [7,]    0.2241531    0.2277357  0.2351311
#>    [8,]    0.1742174    0.1770020  0.1827500
#>    [9,]    0.2454024    0.2493247  0.2574210
#>   [10,]    0.3795088    0.3855742  0.3980944
#>   [11,]    0.3017547    0.3065775  0.3165328
#>   [12,]    0.2360735    0.2398466  0.2476352
#>   [13,]    0.1942093    0.1973134  0.2037210
#>   [14,]    0.1809788    0.1838715  0.1898426
#>   [15,]    0.2652851    0.2695251  0.2782774
#>   [16,]    0.2167249    0.2201888  0.2273391
#>   [17,]    0.2453321    0.2492532  0.2573472
#>   [18,]    0.2331612    0.2368878  0.2445803
#>   [19,]    0.1458178    0.1481485  0.1529595
#>   [20,]    0.2497700    0.2537621  0.2620027
#>   [21,]    0.1783618    0.1812126  0.1870973
#>   [22,]    0.2456485    0.2495747  0.2576791
#>   [23,]    0.2102935    0.2136547  0.2205929
#>   [24,]    0.2035649    0.2068186  0.2135348
#>   [25,]    0.2252038    0.2288033  0.2362333
#>   [26,]    0.1870403    0.1900298  0.1962008
#>   [27,]    0.2715488    0.2758890  0.2848479
#>   [28,]    0.2377600    0.2415601  0.2494044
#>   [29,]    0.2019376    0.2051652  0.2118277
#>   [30,]    0.1703435    0.1730661  0.1786862
#>   [31,]    0.2551043    0.2591816  0.2675981
#>   [32,]    0.2415420    0.2454027  0.2533718
#>   [33,]    0.2454405    0.2493634  0.2574609
#>   [34,]    0.3098448    0.3147969  0.3250192
#>   [35,]    0.2043290    0.2075948  0.2143362
#>   [36,]    0.2177378    0.2212179  0.2284016
#>   [37,]    0.2208457    0.2243755  0.2316618
#>   [38,]    0.1491525    0.1515364  0.1564574
#>   [39,]    0.2299470    0.2336223  0.2412089
#>   [40,]    0.3215877    0.3267274  0.3373370
#>   [41,]    0.2238123    0.2273896  0.2347739
#>   [42,]    0.1990461    0.2022275  0.2087946
#>   [43,]    0.3504480    0.3560490  0.3676108
#>   [44,]    0.1861450    0.1891202  0.1952616
#>   [45,]    0.2510575    0.2550701  0.2633531
#>   [46,]    0.1470876    0.1494386  0.1542915
#>   [47,]    0.2909254    0.2955752  0.3051734
#>   [48,]    0.1749370    0.1777331  0.1835048
#>   [49,]    0.2156462    0.2190929  0.2262077
#>   [50,]    0.1932158    0.1963039  0.2026786
#>   [51,]    0.2015514    0.2047729  0.2114226
#>   [52,]    0.2195722    0.2230816  0.2303259
#>   [53,]    0.2227860    0.2263468  0.2336970
#>   [54,]    0.2598982    0.2640521  0.2726268
#>   [55,]    0.2644499    0.2686766  0.2774014
#>   [56,]    0.2906632    0.2953088  0.3048983
#>   [57,]    0.2586606    0.2627948  0.2713286
#>   [58,]    0.1825395    0.1854571  0.1914796
#>   [59,]    0.2374779    0.2412735  0.2491084
#>   [60,]    0.1709416    0.1736738  0.1793137
#>   [61,]    0.2581028    0.2622280  0.2707434
#>   [62,]    0.2010676    0.2042813  0.2109151
#>   [63,]    0.1679338    0.1706180  0.1761588
#>   [64,]    0.2479756    0.2519391  0.2601204
#>   [65,]    0.1940521    0.1971537  0.2035561
#>   [66,]    0.2396248    0.2434547  0.2513605
#>   [67,]    0.2314358    0.2351349  0.2427706
#>   [68,]    0.2321146    0.2358244  0.2434824
#>   [69,]    0.1949744    0.1980907  0.2045235
#>   [70,]    0.2524218    0.2564563  0.2647842
#>   [71,]    0.1850105    0.1879676  0.1940716
#>   [72,]    0.1855090    0.1884740  0.1945945
#>   [73,]    0.2307643    0.2344526  0.2420662
#>   [74,]    0.2306095    0.2342954  0.2419038
#>   [75,]    0.2719326    0.2762788  0.2852505
#>   [76,]    0.1932824    0.1963717  0.2027486
#>   [77,]    0.2492537    0.2532376  0.2614610
#>   [78,]    0.1853592    0.1883218  0.1944374
#>   [79,]    0.2131214    0.2165278  0.2235592
#>   [80,]    0.1966487    0.1997917  0.2062797
#>   [81,]    0.1882498    0.1912587  0.1974696
#>   [82,]    0.2821975    0.2867078  0.2960180
#>   [83,]    0.3110497    0.3160211  0.3262833
#>   [84,]    0.1851103    0.1880689  0.1941763
#>   [85,]    0.1866245    0.1896074  0.1957648
#>   [86,]    0.1764014    0.1792209  0.1850409
#>   [87,]    0.2407555    0.2446035  0.2525466
#>   [88,]    0.1907024    0.1937504  0.2000423
#>   [89,]    0.2964871    0.3012259  0.3110077
#>   [90,]    0.1404038    0.1426479  0.1472804
#>   [91,]    0.3081040    0.3130283  0.3231932
#>   [92,]    0.1538017    0.1562599  0.1613343
#>   [93,]    0.2255157    0.2291201  0.2365605
#>   [94,]    0.2421885    0.2460594  0.2540496
#>   [95,]    0.2448749    0.2487888  0.2568678
#>   [96,]    0.2225988    0.2261566  0.2335007
#>   [97,]    0.1963962    0.1995353  0.2060150
#>   [98,]    0.2432285    0.2471160  0.2551407
#>   [99,]    0.2025933    0.2058314  0.2125155
#>  [100,]    0.2163760    0.2198344  0.2269732
#>  [101,]    0.2293704    0.2330364  0.2406039
#>  [102,]    0.1884159    0.1914274  0.1976438
#>  [103,]    0.2101009    0.2134589  0.2203906
#>  [104,]    0.1270525    0.1290833  0.1332751
#>  [105,]    0.2595641    0.2637127  0.2722764
#>  [106,]    0.2183332    0.2218228  0.2290261
#>  [107,]    0.1857182    0.1886866  0.1948139
#>  [108,]    0.1135390    0.1153538  0.1190998
#>  [109,]    0.2567286    0.2608319  0.2693020
#>  [110,]    0.2206024    0.2241283  0.2314065
#>  [111,]    0.2893714    0.2939964  0.3035433
#>  [112,]    0.3487356    0.3543093  0.3658144
#>  [113,]    0.2102018    0.2135615  0.2204966
#>  [114,]    0.2098354    0.2131892  0.2201123
#>  [115,]    0.1971526    0.2003037  0.2068084
#>  [116,]    0.1412901    0.1435484  0.1482101
#>  [117,]    0.1903040    0.1933456  0.1996244
#>  [118,]    0.1679872    0.1706722  0.1762147
#>  [119,]    0.1965860    0.1997281  0.2062139
#>  [120,]    0.1795874    0.1824578  0.1883830
#>  [121,]    0.2146651    0.2180962  0.2251786
#>  [122,]    0.1752049    0.1780052  0.1837857
#>  [123,]    0.1564627    0.1589635  0.1641258
#>  [124,]    0.2576075    0.2617247  0.2702235
#>  [125,]    0.3084144    0.3133437  0.3235188
#>  [126,]    0.2583830    0.2625127  0.2710374
#>  [127,]    0.2180037    0.2214881  0.2286806
#>  [128,]    0.2713537    0.2756907  0.2846432
#>  [129,]    0.5013558    0.5093685  0.5259085
#>  [130,]    0.2223266    0.2258801  0.2332153
#>  [131,]    0.2150580    0.2184953  0.2255906
#>  [132,]    0.2135192    0.2169319  0.2239764
#>  [133,]    0.1794313    0.1822992  0.1882192
#>  [134,]    0.1527991    0.1552413  0.1602827
#>  [135,]    0.2326715    0.2363903  0.2440668
#>  [136,]    0.2171519    0.2206227  0.2277872
#>  [137,]    0.2626224    0.2668199  0.2754844
#>  [138,]    0.2165026    0.2199630  0.2271060
#>  [139,]    0.1710262    0.1737597  0.1794025
#>  [140,]    0.1573911    0.1599068  0.1650996
#>  [141,]    0.2160110    0.2194635  0.2265903
#>  [142,]    0.1511824    0.1535988  0.1585869
#>  [143,]    0.1482762    0.1506462  0.1555383
#>  [144,]    0.1603453    0.1629082  0.1681985
#>  [145,]    0.2963228    0.3010589  0.3108351
#>  [146,]    0.2484773    0.2524487  0.2606464
#>  [147,]    0.2426217    0.2464994  0.2545040
#>  [148,]    0.1804781    0.1833627  0.1893173
#>  [149,]    0.1978032    0.2009648  0.2074909
#>  [150,]    0.2862654    0.2908407  0.3002853
#>  [151,]    0.2253526    0.2289544  0.2363893
#>  [152,]    0.1765401    0.1793618  0.1851864
#>  [153,]    0.2067835    0.2100885  0.2169109
#>  [154,]    0.2825968    0.2871135  0.2964370
#>  [155,]    0.2083671    0.2116975  0.2185722
#>  [156,]    0.3358686    0.3412366  0.3523173
#>  [157,]    0.1403965    0.1426405  0.1472726
#>  [158,]    0.2103500    0.2137121  0.2206521
#>  [159,]    0.1968385    0.1999846  0.2064789
#>  [160,]    0.2441438    0.2480460  0.2561009
#>  [161,]    0.1555142    0.1579999  0.1631308
#>  [162,]    0.2536849    0.2577395  0.2661092
#>  [163,]    0.3167818    0.3218448  0.3322959
#>  [164,]    0.2046657    0.2079369  0.2146894
#>  [165,]    0.1841539    0.1870973  0.1931731
#>  [166,]    0.3112183    0.3161924  0.3264601
#>  [167,]    0.2383164    0.2421254  0.2499880
#>  [168,]    0.2172999    0.2207730  0.2279423
#>  [169,]    0.2410808    0.2449340  0.2528878
#>  [170,]    0.1977310    0.2008914  0.2074151
#>  [171,]    0.2824379    0.2869521  0.2962704
#>  [172,]    0.1839278    0.1868676  0.1929360
#>  [173,]    0.2301016    0.2337794  0.2413710
#>  [174,]    0.2332965    0.2370252  0.2447222
#>  [175,]    0.2415261    0.2453864  0.2533549
#>  [176,]    0.2687062    0.2730009  0.2818661
#>  [177,]    0.2308995    0.2345900  0.2422080
#>  [178,]    0.2972772    0.3020285  0.3118362
#>  [179,]    0.2048178    0.2080915  0.2148490
#>  [180,]    0.1956510    0.1987782  0.2052333
#>  [181,]    0.1896356    0.1926666  0.1989232
#>  [182,]    0.1959113    0.1990426  0.2055062
#>  [183,]    0.1978342    0.2009962  0.2075233
#>  [184,]    0.1370591    0.1392497  0.1437718
#>  [185,]    0.2190964    0.2225983  0.2298268
#>  [186,]    0.2624666    0.2666616  0.2753209
#>  [187,]    0.1849397    0.1878957  0.1939973
#>  [188,]    0.2139132    0.2173322  0.2243898
#>  [189,]    0.2397433    0.2435751  0.2514848
#>  [190,]    0.2116851    0.2150685  0.2220526
#>  [191,]    0.1828987    0.1858220  0.1918564
#>  [192,]    0.2068294    0.2101352  0.2169591
#>  [193,]    0.1789856    0.1818464  0.1877517
#>  [194,]    0.1894832    0.1925117  0.1987633
#>  [195,]    0.2071339    0.2104445  0.2172783
#>  [196,]    0.2249686    0.2285643  0.2359866
#>  [197,]    0.2032960    0.2065453  0.2132527
#>  [198,]    0.2786312    0.2830845  0.2922772
#>  [199,]    0.2097074    0.2130592  0.2199781
#>  [200,]    0.2075028    0.2108193  0.2176655
#>  [201,]    0.2085819    0.2119157  0.2187973
#>  [202,]    0.1730234    0.1757889  0.1814975
#>  [203,]    0.2650908    0.2693277  0.2780737
#>  [204,]    0.1600095    0.1625670  0.1678462
#>  [205,]    0.2645819    0.2688107  0.2775400
#>  [206,]    0.2613734    0.2655509  0.2741742
#>  [207,]    0.2604094    0.2645714  0.2731628
#>  [208,]    0.2816770    0.2861790  0.2954722
#>  [209,]    0.2019841    0.2052124  0.2118765
#>  [210,]    0.2158535    0.2193035  0.2264251
#>  [211,]    0.2145390    0.2179680  0.2250463
#>  [212,]    0.2143765    0.2178029  0.2248758
#>  [213,]    0.1706744    0.1734023  0.1790333
#>  [214,]    0.1808338    0.1837241  0.1896903
#>  [215,]    0.2353470    0.2391086  0.2468734
#>  [216,]    0.2299354    0.2336104  0.2411966
#>  [217,]    0.2074696    0.2107856  0.2176305
#>  [218,]    0.2303489    0.2340305  0.2416304
#>  [219,]    0.2000565    0.2032541  0.2098546
#>  [220,]    0.2384943    0.2423062  0.2501747
#>  [221,]    0.2045694    0.2078391  0.2145884
#>  [222,]    0.2161124    0.2195666  0.2266967
#>  [223,]    0.2879084    0.2925100  0.3020088
#>  [224,]    0.2079578    0.2112817  0.2181427
#>  [225,]    0.2101544    0.2135134  0.2204469
#>  [226,]    0.2068395    0.2101454  0.2169696
#>  [227,]    0.2051983    0.2084780  0.2152481
#>  [228,]    0.2506799    0.2546866  0.2629571
#>  [229,]    0.2166031    0.2200652  0.2272115
#>  [230,]    0.1749643    0.1777609  0.1835335
#>  [231,]    0.2684382    0.2727286  0.2815849
#>  [232,]    0.2272989    0.2309317  0.2384307
#>  [233,]    0.1987069    0.2018829  0.2084388
#>  [234,]    0.2709489    0.2752795  0.2842187
#>  [235,]    0.2643076    0.2685320  0.2772522
#>  [236,]    0.2518371    0.2558622  0.2641709
#>  [237,]    0.3415679    0.3470271  0.3582959
#>  [238,]    0.2647424    0.2689738  0.2777082
#>  [239,]    0.3043163    0.3091802  0.3192202
#>  [240,]    0.2961902    0.3009242  0.3106960
#>  [241,]    0.2082997    0.2116290  0.2185014
#>  [242,]    0.1855710    0.1885371  0.1946596
#>  [243,]    0.2708813    0.2752108  0.2841476
#>  [244,]    0.2820162    0.2865236  0.2958278
#>  [245,]    0.1638874    0.1665069  0.1719141
#>  [246,]    0.2969939    0.3017407  0.3115391
#>  [247,]    0.2745681    0.2789565  0.2880150
#>  [248,]    0.1923682    0.1954429  0.2017898
#>  [249,]    0.2062692    0.2095661  0.2163716
#>  [250,]    0.2215830    0.2251246  0.2324352
#>  [251,]    0.2374963    0.2412922  0.2491278
#>  [252,]    0.1691763    0.1718803  0.1774620
#>  [253,]    0.1955296    0.1986547  0.2051058
#>  [254,]    0.2465718    0.2505128  0.2586478
#>  [255,]    0.2676283    0.2719057  0.2807352
#>  [256,]    0.1719785    0.1747272  0.1804014
#>  [257,]    0.2077693    0.2110901  0.2179450
#>  [258,]    0.1935362    0.1966295  0.2030149
#>  [259,]    0.2505961    0.2546014  0.2628692
#>  [260,]    0.2232657    0.2268342  0.2342004
#>  [261,]    0.2303878    0.2340702  0.2416713
#>  [262,]    0.3037042    0.3085582  0.3185779
#>  [263,]    0.2999594    0.3047535  0.3146496
#>  [264,]    0.1670237    0.1696933  0.1752039
#>  [265,]    0.1484475    0.1508202  0.1557180
#>  [266,]    0.2138298    0.2172474  0.2243022
#>  [267,]    0.1654274    0.1680715  0.1735294
#>  [268,]    0.2294841    0.2331520  0.2407232
#>  [269,]    0.2673647    0.2716380  0.2804589
#>  [270,]    0.1754890    0.1782939  0.1840838
#>  [271,]    0.1944140    0.1975214  0.2039356
#>  [272,]    0.2962232    0.3009577  0.3107308
#>  [273,]    0.1729216    0.1756854  0.1813906
#>  [274,]    0.1850193    0.1879765  0.1940809
#>  [275,]    0.2459015    0.2498318  0.2579447
#>  [276,]    0.2762978    0.2807138  0.2898295
#>  [277,]    0.2926441    0.2973214  0.3069762
#>  [278,]    0.1825033    0.1854203  0.1914416
#>  [279,]    0.1908685    0.1939192  0.2002165
#>  [280,]    0.2774877    0.2819227  0.2910776
#>  [281,]    0.1827044    0.1856246  0.1916526
#>  [282,]    0.1864863    0.1894669  0.1956196
#>  [283,]    0.1917339    0.1947984  0.2011243
#>  [284,]    0.2594826    0.2636299  0.2721908
#>  [285,]    0.2839597    0.2884982  0.2978666
#>  [286,]    0.1737920    0.1765698  0.1823036
#>  [287,]    0.2071650    0.2104761  0.2173109
#>  [288,]    0.2266506    0.2302732  0.2377510
#>  [289,]    0.2370250    0.2408133  0.2486334
#>  [290,]    0.2406819    0.2445287  0.2524693
#>  [291,]    0.1973986    0.2005536  0.2070664
#>  [292,]    0.2159828    0.2194349  0.2265608
#>  [293,]    0.2111877    0.2145632  0.2215308
#>  [294,]    0.2146104    0.2180405  0.2251210
#>  [295,]    0.2039305    0.2071899  0.2139182
#>  [296,]    0.2031953    0.2064430  0.2131470
#>  [297,]    0.2642295    0.2684526  0.2771702
#>  [298,]    0.2077037    0.2110235  0.2178762
#>  [299,]    0.2148393    0.2182731  0.2253612
#>  [300,]    0.2420899    0.2459592  0.2539463
#>  [301,]    0.3114603    0.3164382  0.3267138
#>  [302,]    0.1442781    0.1465842  0.1513444
#>  [303,]    0.2734602    0.2778309  0.2868529
#>  [304,]    0.2249838    0.2285798  0.2360026
#>  [305,]    0.1914907    0.1945514  0.2008692
#>  [306,]    0.1994720    0.2026602  0.2092414
#>  [307,]    0.2858271    0.2903954  0.2998254
#>  [308,]    0.2589701    0.2631092  0.2716531
#>  [309,]    0.2364515    0.2402307  0.2480318
#>  [310,]    0.1934986    0.1965913  0.2029754
#>  [311,]    0.2941595    0.2988610  0.3085658
#>  [312,]    0.1920030    0.1950719  0.2014066
#>  [313,]    0.2478600    0.2518215  0.2599990
#>  [314,]    0.2033727    0.2066232  0.2133330
#>  [315,]    0.1630058    0.1656112  0.1709893
#>  [316,]    0.1827341    0.1856547  0.1916837
#>  [317,]    0.2034409    0.2066925  0.2134047
#>  [318,]    0.1662015    0.1688579  0.1743414
#>  [319,]    0.1202916    0.1222143  0.1261831
#>  [320,]    0.2084152    0.2117464  0.2186226
#>  [321,]    0.2000599    0.2032575  0.2098580
#>  [322,]    0.2628398    0.2670407  0.2757123
#>  [323,]    0.2298366    0.2335101  0.2410930
#>  [324,]    0.2074666    0.2107826  0.2176274
#>  [325,]    0.2180027    0.2214870  0.2286794
#>  [326,]    0.1904078    0.1934512  0.1997333
#>  [327,]    0.1637784    0.1663961  0.1717997
#>  [328,]    0.2299589    0.2336343  0.2412212
#>  [329,]    0.2255767    0.2291821  0.2366245
#>  [330,]    0.1737251    0.1765018  0.1822335
#>  [331,]    0.1670198    0.1696893  0.1751999
#>  [332,]    0.2194721    0.2229799  0.2302209
#>  [333,]    0.2982722    0.3030394  0.3128801
#>  [334,]    0.2302275    0.2339073  0.2415031
#>  [335,]    0.2028467    0.2060888  0.2127813
#>  [336,]    0.2508770    0.2548868  0.2631639
#>  [337,]    0.1503972    0.1528010  0.1577631
#>  [338,]    0.1878312    0.1908333  0.1970304
#>  [339,]    0.1934182    0.1965096  0.2028910
#>  [340,]    0.2142059    0.2176296  0.2246969
#>  [341,]    0.1627438    0.1653450  0.1707145
#>  [342,]    0.2215736    0.2251150  0.2324253
#>  [343,]    0.2488742    0.2528520  0.2610629
#>  [344,]    0.2546717    0.2587421  0.2671443
#>  [345,]    0.2325395    0.2362562  0.2439284
#>  [346,]    0.2307049    0.2343923  0.2420039
#>  [347,]    0.1978242    0.2009860  0.2075129
#>  [348,]    0.2025122    0.2057490  0.2124304
#>  [349,]    0.2450314    0.2489478  0.2570319
#>  [350,]    0.2365952    0.2403767  0.2481825
#>  [351,]    0.1706711    0.1733990  0.1790299
#>  [352,]    0.1942006    0.1973045  0.2037116
#>  [353,]    0.2340488    0.2377897  0.2455116
#>  [354,]    0.2860781    0.2906504  0.3000888
#>  [355,]    0.2462110    0.2501462  0.2582694
#>  [356,]    0.1715017    0.1742429  0.1799013
#>  [357,]    0.2141041    0.2175262  0.2245900
#>  [358,]    0.2000543    0.2032518  0.2098521
#>  [359,]    0.1713305    0.1740689  0.1797216
#>  [360,]    0.1406764    0.1429250  0.1475664
#>  [361,]    0.2334562    0.2371876  0.2448898
#>  [362,]    0.2430782    0.2469633  0.2549831
#>  [363,]    0.2444775    0.2483850  0.2564509
#>  [364,]    0.2124340    0.2158294  0.2228382
#>  [365,]    0.2117294    0.2151135  0.2220991
#>  [366,]    0.2577244    0.2618436  0.2703466
#>  [367,]    0.1445895    0.1469006  0.1516711
#>  [368,]    0.2188585    0.2223565  0.2295773
#>  [369,]    0.1575923    0.1601112  0.1653107
#>  [370,]    0.2138612    0.2172794  0.2243354
#>  [371,]    0.2796298    0.2840990  0.2933245
#>  [372,]    0.1320186    0.1341287  0.1384845
#>  [373,]    0.2600809    0.2642377  0.2728184
#>  [374,]    0.1963911    0.1995301  0.2060096
#>  [375,]    0.2069094    0.2102164  0.2170430
#>  [376,]    0.2434799    0.2473714  0.2554042
#>  [377,]    0.2552377    0.2593172  0.2677380
#>  [378,]    0.1548048    0.1572791  0.1623867
#>  [379,]    0.2217179    0.2252617  0.2325767
#>  [380,]    0.2246333    0.2282236  0.2356349
#>  [381,]    0.2691402    0.2734419  0.2823214
#>  [382,]    0.1473811    0.1497367  0.1545993
#>  [383,]    0.2061075    0.2094018  0.2162018
#>  [384,]    0.2087901    0.2121272  0.2190157
#>  [385,]    0.3195082    0.3246148  0.3351558
#>  [386,]    0.1562081    0.1587048  0.1638587
#>  [387,]    0.3303384    0.3356180  0.3465163
#>  [388,]    0.2102621    0.2136228  0.2205598
#>  [389,]    0.2140817    0.2175034  0.2245665
#>  [390,]    0.3040212    0.3088803  0.3189107
#>  [391,]    0.2149970    0.2184333  0.2255266
#>  [392,]    0.2902018    0.2948401  0.3044145
#>  [393,]    0.2054807    0.2087650  0.2155443
#>  [394,]    0.2229690    0.2265328  0.2338892
#>  [395,]    0.1734264    0.1761983  0.1819202
#>  [396,]    0.3028355    0.3076756  0.3176666
#>  [397,]    0.1991306    0.2023134  0.2088833
#>  [398,]    0.1893924    0.1924195  0.1986681
#>  [399,]    0.2080749    0.2114006  0.2182656
#>  [400,]    0.1790805    0.1819428  0.1878512
#>  [401,]    0.2018734    0.2051000  0.2117603
#>  [402,]    0.1909324    0.1939841  0.2002834
#>  [403,]    0.1773935    0.1802288  0.1860815
#>  [404,]    0.2751956    0.2795940  0.2886732
#>  [405,]    0.1975619    0.2007196  0.2072377
#>  [406,]    0.1700570    0.1727750  0.1783857
#>  [407,]    0.2185587    0.2220519  0.2292627
#>  [408,]    0.1865846    0.1895668  0.1957228
#>  [409,]    0.2060001    0.2092926  0.2160891
#>  [410,]    0.1701426    0.1728621  0.1784756
#>  [411,]    0.1889417    0.1919616  0.1981953
#>  [412,]    0.2082147    0.2115426  0.2184121
#>  [413,]    0.2196004    0.2231103  0.2303554
#>  [414,]    0.1789521    0.1818123  0.1877165
#>  [415,]    0.3151343    0.3201710  0.3305679
#>  [416,]    0.1439379    0.1462385  0.1509875
#>  [417,]    0.1622754    0.1648691  0.1702231
#>  [418,]    0.1872926    0.1902861  0.1964655
#>  [419,]    0.2008771    0.2040878  0.2107153
#>  [420,]    0.1733138    0.1760840  0.1818021
#>  [421,]    0.2267267    0.2303505  0.2378308
#>  [422,]    0.2464807    0.2504202  0.2585522
#>  [423,]    0.2439418    0.2478407  0.2558890
#>  [424,]    0.2041789    0.2074423  0.2141788
#>  [425,]    0.2214250    0.2249640  0.2322694
#>  [426,]    0.2217903    0.2253352  0.2326527
#>  [427,]    0.1782254    0.1810740  0.1869541
#>  [428,]    0.1563705    0.1588698  0.1640289
#>  [429,]    0.2601168    0.2642742  0.2728560
#>  [430,]    0.1541204    0.1565838  0.1616687
#>  [431,]    0.3078264    0.3127463  0.3229020
#>  [432,]    0.2167580    0.2202225  0.2273739
#>  [433,]    0.2950227    0.2997380  0.3094715
#>  [434,]    0.2640984    0.2683194  0.2770325
#>  [435,]    0.2918422    0.2965066  0.3061351
#>  [436,]    0.1988862    0.2020651  0.2086269
#>  [437,]    0.2737553    0.2781307  0.2871625
#>  [438,]    0.2156746    0.2191218  0.2262375
#>  [439,]    0.2131653    0.2165724  0.2236053
#>  [440,]    0.2067152    0.2100191  0.2168392
#>  [441,]    0.2153795    0.2188219  0.2259279
#>  [442,]    0.2133529    0.2167629  0.2238020
#>  [443,]    0.2841490    0.2886905  0.2980653
#>  [444,]    0.1719011    0.1746487  0.1803203
#>  [445,]    0.2124529    0.2158485  0.2228580
#>  [446,]    0.2151087    0.2185468  0.2256438
#>  [447,]    0.1982008    0.2013687  0.2079078
#>  [448,]    0.3139867    0.3190051  0.3293641
#>  [449,]    0.3250273    0.3302221  0.3409452
#>  [450,]    0.1757912    0.1786009  0.1844008
#>  [451,]    0.1833768    0.1863078  0.1923580
#>  [452,]    0.1778367    0.1806791  0.1865465
#>  [453,]    0.1887448    0.1917616  0.1979888
#>  [454,]    0.1733289    0.1760993  0.1818179
#>  [455,]    0.2121035    0.2154936  0.2224915
#>  [456,]    0.2093756    0.2127221  0.2196300
#>  [457,]    0.1654851    0.1681301  0.1735900
#>  [458,]    0.2169539    0.2204215  0.2275794
#>  [459,]    0.2476158    0.2515735  0.2597428
#>  [460,]    0.2363206    0.2400978  0.2478947
#>  [461,]    0.2178154    0.2212967  0.2284830
#>  [462,]    0.1595702    0.1621206  0.1673853
#>  [463,]    0.1665691    0.1692314  0.1747271
#>  [464,]    0.2034598    0.2067118  0.2134245
#>  [465,]    0.2052253    0.2085054  0.2152764
#>  [466,]    0.2596671    0.2638173  0.2723843
#>  [467,]    0.2064197    0.2097190  0.2165293
#>  [468,]    0.1837549    0.1866919  0.1927545
#>  [469,]    0.1856789    0.1886466  0.1947727
#>  [470,]    0.2959541    0.3006843  0.3104485
#>  [471,]    0.2463720    0.2503098  0.2584382
#>  [472,]    0.1299112    0.1319876  0.1362738
#>  [473,]    0.1901785    0.1932181  0.1994927
#>  [474,]    0.1608353    0.1634060  0.1687124
#>  [475,]    0.2320975    0.2358072  0.2434647
#>  [476,]    0.1953741    0.1984968  0.2049427
#>  [477,]    0.2435998    0.2474932  0.2555301
#>  [478,]    0.2065325    0.2098335  0.2166476
#>  [479,]    0.1951565    0.1982757  0.2047145
#>  [480,]    0.3370839    0.3424714  0.3535923
#>  [481,]    0.1813949    0.1842942  0.1902789
#>  [482,]    0.1879587    0.1909629  0.1971642
#>  [483,]    0.2383360    0.2421453  0.2500087
#>  [484,]    0.1559619    0.1584547  0.1636004
#>  [485,]    0.1323435    0.1344588  0.1388253
#>  [486,]    0.1994067    0.2025939  0.2091729
#>  [487,]    0.1528189    0.1552615  0.1603035
#>  [488,]    0.2113460    0.2147239  0.2216967
#>  [489,]    0.2735732    0.2779458  0.2869716
#>  [490,]    0.2011799    0.2043954  0.2110329
#>  [491,]    0.1762697    0.1790871  0.1849028
#>  [492,]    0.2336213    0.2373552  0.2450630
#>  [493,]    0.1849909    0.1879477  0.1940511
#>  [494,]    0.2255591    0.2291643  0.2366061
#>  [495,]    0.2719556    0.2763022  0.2852745
#>  [496,]    0.1899394    0.1929753  0.1992419
#>  [497,]    0.2235771    0.2271505  0.2345269
#>  [498,]    0.2321980    0.2359093  0.2435700
#>  [499,]    0.2468804    0.2508263  0.2589715
#>  [500,]    0.1863471    0.1893255  0.1954737
#>  [501,]    0.2728223    0.2771828  0.2861837
#>  [502,]    0.1979173    0.2010806  0.2076105
#>  [503,]    0.2302305    0.2339103  0.2415060
#>  [504,]    0.1725023    0.1752595  0.1809509
#>  [505,]    0.1518858    0.1543134  0.1593246
#>  [506,]    0.1859140    0.1888855  0.1950192
#>  [507,]    0.2369382    0.2407251  0.2485422
#>  [508,]    0.1676613    0.1703411  0.1758728
#>  [509,]    0.1814349    0.1843349  0.1903210
#>  [510,]    0.2410858    0.2449391  0.2528930
#>  [511,]    0.2296339    0.2333042  0.2408804
#>  [512,]    0.1813530    0.1842516  0.1902350
#>  [513,]    0.2108905    0.2142612  0.2212191
#>  [514,]    0.2007599    0.2039687  0.2105922
#>  [515,]    0.1694922    0.1722013  0.1777934
#>  [516,]    0.2767229    0.2811457  0.2902754
#>  [517,]    0.2197020    0.2232135  0.2304621
#>  [518,]    0.1676234    0.1703025  0.1758329
#>  [519,]    0.2037198    0.2069758  0.2136971
#>  [520,]    0.1597687    0.1623223  0.1675935
#>  [521,]    0.3008206    0.3056285  0.3155530
#>  [522,]    0.2464932    0.2504329  0.2585654
#>  [523,]    0.2052780    0.2085591  0.2153318
#>  [524,]    0.2087002    0.2120359  0.2189215
#>  [525,]    0.2682239    0.2725109  0.2813603
#>  [526,]    0.1961676    0.1993030  0.2057752
#>  [527,]    0.2170612    0.2205305  0.2276919
#>  [528,]    0.2058271    0.2091169  0.2159078
#>  [529,]    0.1436364    0.1459322  0.1506711
#>  [530,]    0.2149305    0.2183657  0.2254569
#>  [531,]    0.1909381    0.1939899  0.2002895
#>  [532,]    0.1944036    0.1975108  0.2039247
#>  [533,]    0.1824194    0.1853351  0.1913536
#>  [534,]    0.2248964    0.2284909  0.2359108
#>  [535,]    0.2248836    0.2284780  0.2358974
#>  [536,]    0.1929234    0.1960069  0.2023719
#>  [537,]    0.1947825    0.1978958  0.2043222
#>  [538,]    0.2194315    0.2229387  0.2301784
#>  [539,]    0.1904704    0.1935147  0.1997989
#>  [540,]    0.1995404    0.2027297  0.2093131
#>  [541,]    0.1739115    0.1766912  0.1824291
#>  [542,]    0.2469326    0.2508793  0.2590261
#>  [543,]    0.2497087    0.2536998  0.2619384
#>  [544,]    0.2475937    0.2515509  0.2597196
#>  [545,]    0.2231902    0.2267575  0.2341211
#>  [546,]    0.1955919    0.1987181  0.2051711
#>  [547,]    0.2495894    0.2535786  0.2618131
#>  [548,]    0.1866713    0.1896549  0.1958137
#>  [549,]    0.2429153    0.2467978  0.2548121
#>  [550,]    0.2735241    0.2778958  0.2869200
#>  [551,]    0.3127405    0.3177389  0.3280567
#>  [552,]    0.3065360    0.3114353  0.3215483
#>  [553,]    0.1733115    0.1760816  0.1817997
#>  [554,]    0.1859235    0.1888952  0.1950294
#>  [555,]    0.2332206    0.2369481  0.2446427
#>  [556,]    0.2138516    0.2172696  0.2243252
#>  [557,]    0.2431584    0.2470449  0.2550672
#>  [558,]    0.2391526    0.2429749  0.2508650
#>  [559,]    0.2660000    0.2702514  0.2790273
#>  [560,]    0.1976012    0.2007595  0.2072789
#>  [561,]    0.2128940    0.2162967  0.2233207
#>  [562,]    0.1860213    0.1889945  0.1951319
#>  [563,]    0.1676551    0.1703347  0.1758662
#>  [564,]    0.1824969    0.1854138  0.1914349
#>  [565,]    0.2278639    0.2315059  0.2390237
#>  [566,]    0.1559004    0.1583922  0.1635359
#>  [567,]    0.1996803    0.2028718  0.2094599
#>  [568,]    0.2427794    0.2466598  0.2546695
#>  [569,]    0.2764647    0.2808834  0.2900044
#>  [570,]    0.1605038    0.1630692  0.1683648
#>  [571,]    0.2241568    0.2277395  0.2351351
#>  [572,]    0.2086142    0.2119485  0.2188311
#>  [573,]    0.1867159    0.1897003  0.1958605
#>  [574,]    0.2564382    0.2605368  0.2689973
#>  [575,]    0.2571551    0.2612652  0.2697493
#>  [576,]    0.2005525    0.2037579  0.2103748
#>  [577,]    0.2304810    0.2341648  0.2417690
#>  [578,]    0.2080752    0.2114009  0.2182659
#>  [579,]    0.2306844    0.2343715  0.2419823
#>  [580,]    0.2091991    0.2125427  0.2194448
#>  [581,]    0.1810249    0.1839183  0.1898908
#>  [582,]    0.2060392    0.2093324  0.2161301
#>  [583,]    0.2221990    0.2257504  0.2330812
#>  [584,]    0.2235208    0.2270934  0.2344679
#>  [585,]    0.1863163    0.1892943  0.1954414
#>  [586,]    0.2263927    0.2300112  0.2374804
#>  [587,]    0.2528387    0.2568797  0.2652214
#>  [588,]    0.1930035    0.1960883  0.2024561
#>  [589,]    0.2507581    0.2547660  0.2630391
#>  [590,]    0.1574725    0.1599894  0.1651850
#>  [591,]    0.2176940    0.2211734  0.2283557
#>  [592,]    0.1744107    0.1771984  0.1829527
#>  [593,]    0.2654795    0.2697226  0.2784813
#>  [594,]    0.1615908    0.1641735  0.1695049
#>  [595,]    0.1875918    0.1905902  0.1967794
#>  [596,]    0.1969183    0.2000656  0.2065624
#>  [597,]    0.2075963    0.2109144  0.2177636
#>  [598,]    0.2468978    0.2508440  0.2589898
#>  [599,]    0.1832742    0.1862035  0.1922503
#>  [600,]    0.1509825    0.1533957  0.1583771
#>  [601,]    0.1442216    0.1465268  0.1512851
#>  [602,]    0.1948246    0.1979385  0.2043663
#>  [603,]    0.2598596    0.2640129  0.2725862
#>  [604,]    0.1916657    0.1947291  0.2010527
#>  [605,]    0.1927896    0.1958710  0.2022317
#>  [606,]    0.1909797    0.1940322  0.2003331
#>  [607,]    0.2062792    0.2095762  0.2163820
#>  [608,]    0.2029027    0.2061457  0.2128399
#>  [609,]    0.2126444    0.2160431  0.2230589
#>  [610,]    0.1887554    0.1917724  0.1980000
#>  [611,]    0.1681416    0.1708291  0.1763766
#>  [612,]    0.1513290    0.1537478  0.1587406
#>  [613,]    0.2007857    0.2039949  0.2106194
#>  [614,]    0.2369052    0.2406916  0.2485077
#>  [615,]    0.2202290    0.2237489  0.2310148
#>  [616,]    0.1612605    0.1638380  0.1691585
#>  [617,]    0.1711522    0.1738878  0.1795347
#>  [618,]    0.1930777    0.1961637  0.2025339
#>  [619,]    0.1927611    0.1958420  0.2022018
#>  [620,]    0.2031312    0.2063779  0.2130798
#>  [621,]    0.2141458    0.2175685  0.2246338
#>  [622,]    0.2252329    0.2288329  0.2362639
#>  [623,]    0.2109448    0.2143164  0.2212760
#>  [624,]    0.1942835    0.1973887  0.2037987
#>  [625,]    0.2053284    0.2086103  0.2153846
#>  [626,]    0.2289270    0.2325860  0.2401388
#>  [627,]    0.2628661    0.2670675  0.2757401
#>  [628,]    0.1737991    0.1765769  0.1823111
#>  [629,]    0.2130166    0.2164213  0.2234493
#>  [630,]    0.2693723    0.2736777  0.2825648
#>  [631,]    0.2063909    0.2096897  0.2164990
#>  [632,]    0.2631886    0.2673951  0.2760782
#>  [633,]    0.2637348    0.2679501  0.2766513
#>  [634,]    0.1940139    0.1971148  0.2035159
#>  [635,]    0.1718261    0.1745725  0.1802416
#>  [636,]    0.2014081    0.2046273  0.2112724
#>  [637,]    0.2126801    0.2160794  0.2230964
#>  [638,]    0.1942495    0.1973542  0.2037631
#>  [639,]    0.2399652    0.2438006  0.2517177
#>  [640,]    0.1597635    0.1623170  0.1675881
#>  [641,]    0.2896319    0.2942610  0.3038166
#>  [642,]    0.1668667    0.1695338  0.1750393
#>  [643,]    0.1891480    0.1921712  0.1984118
#>  [644,]    0.2850811    0.2896375  0.2990429
#>  [645,]    0.2587753    0.2629113  0.2714489
#>  [646,]    0.2764660    0.2808848  0.2900060
#>  [647,]    0.1274697    0.1295071  0.1337127
#>  [648,]    0.2288466    0.2325043  0.2400546
#>  [649,]    0.1573861    0.1599017  0.1650944
#>  [650,]    0.2181762    0.2216633  0.2288615
#>  [651,]    0.1653284    0.1679709  0.1734256
#>  [652,]    0.2260459    0.2296588  0.2371166
#>  [653,]    0.2055811    0.2088669  0.2156495
#>  [654,]    0.1883026    0.1913123  0.1975250
#>  [655,]    0.1920356    0.1951050  0.2014408
#>  [656,]    0.2380446    0.2418493  0.2497029
#>  [657,]    0.1998879    0.2030828  0.2096776
#>  [658,]    0.2189395    0.2224388  0.2296622
#>  [659,]    0.1975159    0.2006728  0.2071895
#>  [660,]    0.2112074    0.2145831  0.2215514
#>  [661,]    0.1802403    0.1831211  0.1890678
#>  [662,]    0.2633707    0.2675801  0.2762693
#>  [663,]    0.1712389    0.1739759  0.1796256
#>  [664,]    0.2077996    0.2111209  0.2179768
#>  [665,]    0.1599007    0.1624564  0.1677321
#>  [666,]    0.2168412    0.2203070  0.2274611
#>  [667,]    0.2396771    0.2435079  0.2514155
#>  [668,]    0.1650393    0.1676772  0.1731224
#>  [669,]    0.2218027    0.2253478  0.2326657
#>  [670,]    0.2664702    0.2707291  0.2795204
#>  [671,]    0.2221596    0.2257104  0.2330401
#>  [672,]    0.3122554    0.3172460  0.3275479
#>  [673,]    0.2213135    0.2248508  0.2321526
#>  [674,]    0.2790400    0.2834999  0.2927060
#>  [675,]    0.2375947    0.2413922  0.2492310
#>  [676,]    0.2049766    0.2082528  0.2150155
#>  [677,]    0.2196358    0.2231463  0.2303926
#>  [678,]    0.2389135    0.2427320  0.2506143
#>  [679,]    0.1880531    0.1910588  0.1972632
#>  [680,]    0.1952239    0.1983441  0.2047851
#>  [681,]    0.2267723    0.2303969  0.2378787
#>  [682,]    0.1870118    0.1900008  0.1961709
#>  [683,]    0.2058596    0.2091499  0.2159418
#>  [684,]    0.1736272    0.1764023  0.1821308
#>  [685,]    0.2128824    0.2162850  0.2233085
#>  [686,]    0.2352013    0.2389605  0.2467204
#>  [687,]    0.2106950    0.2140626  0.2210140
#>  [688,]    0.1861311    0.1891061  0.1952471
#>  [689,]    0.2425708    0.2464478  0.2544509
#>  [690,]    0.1801685    0.1830481  0.1889925
#>  [691,]    0.1544384    0.1569069  0.1620023
#>  [692,]    0.2309181    0.2346089  0.2422274
#>  [693,]    0.2454401    0.2493630  0.2574607
#>  [694,]    0.2630229    0.2672268  0.2759044
#>  [695,]    0.2007171    0.2039252  0.2105475
#>  [696,]    0.1823902    0.1853054  0.1913230
#>  [697,]    0.1651967    0.1678371  0.1732875
#>  [698,]    0.3029995    0.3078423  0.3178389
#>  [699,]    0.1849367    0.1878925  0.1939941
#>  [700,]    0.1920832    0.1951534  0.2014908
#>  [701,]    0.1789134    0.1817730  0.1876759
#>  [702,]    0.2386267    0.2424407  0.2503136
#>  [703,]    0.1694482    0.1721565  0.1777471
#>  [704,]    0.1838674    0.1868062  0.1928725
#>  [705,]    0.1668163    0.1694826  0.1749863
#>  [706,]    0.1317496    0.1338554  0.1382023
#>  [707,]    0.1823579    0.1852726  0.1912890
#>  [708,]    0.1903794    0.1934223  0.1997034
#>  [709,]    0.1697626    0.1724760  0.1780770
#>  [710,]    0.3016444    0.3064655  0.3164173
#>  [711,]    0.2056877    0.2089752  0.2157615
#>  [712,]    0.1849384    0.1878944  0.1939960
#>  [713,]    0.1965145    0.1996555  0.2061391
#>  [714,]    0.2954486    0.3001708  0.3099183
#>  [715,]    0.2385169    0.2423291  0.2501983
#>  [716,]    0.2947986    0.2995102  0.3092361
#>  [717,]    0.1961601    0.1992954  0.2057673
#>  [718,]    0.1851078    0.1880664  0.1941737
#>  [719,]    0.2032765    0.2065255  0.2132322
#>  [720,]    0.2580919    0.2622170  0.2707320
#>  [721,]    0.1770149    0.1798442  0.1856846
#>  [722,]    0.2005947    0.2038008  0.2104190
#>  [723,]    0.2214646    0.2250043  0.2323110
#>  [724,]    0.2359301    0.2397009  0.2474847
#>  [725,]    0.1891691    0.1921927  0.1984339
#>  [726,]    0.2351887    0.2389477  0.2467072
#>  [727,]    0.2624230    0.2666172  0.2752751
#>  [728,]    0.1898095    0.1928433  0.1991056
#>  [729,]    0.2028700    0.2061125  0.2128058
#>  [730,]    0.3480587    0.3536216  0.3651046
#>  [731,]    0.2465735    0.2505144  0.2586494
#>  [732,]    0.2645138    0.2687415  0.2774683
#>  [733,]    0.2134594    0.2168712  0.2239139
#>  [734,]    0.1772740    0.1801074  0.1859562
#>  [735,]    0.1809852    0.1838779  0.1898492
#>  [736,]    0.1924491    0.1955251  0.2018745
#>  [737,]    0.1457462    0.1480757  0.1528844
#>  [738,]    0.1781820    0.1810299  0.1869087
#>  [739,]    0.1851133    0.1880720  0.1941794
#>  [740,]    0.1998428    0.2030369  0.2096302
#>  [741,]    0.1972689    0.2004219  0.2069304
#>  [742,]    0.2215109    0.2250514  0.2323596
#>  [743,]    0.1648932    0.1675287  0.1729690
#>  [744,]    0.2454439    0.2493669  0.2574647
#>  [745,]    0.2140976    0.2175195  0.2245831
#>  [746,]    0.1969614    0.2001095  0.2066077
#>  [747,]    0.2854217    0.2899836  0.2994003
#>  [748,]    0.2966028    0.3013433  0.3111287
#>  [749,]    0.1771416    0.1799729  0.1858173
#>  [750,]    0.2490918    0.2530730  0.2612911
#>  [751,]    0.1839688    0.1869092  0.1929789
#>  [752,]    0.2311788    0.2348738  0.2425010
#>  [753,]    0.1541936    0.1566581  0.1617455
#>  [754,]    0.2211092    0.2246432  0.2319381
#>  [755,]    0.1474071    0.1497631  0.1546265
#>  [756,]    0.1629117    0.1655156  0.1708906
#>  [757,]    0.2200065    0.2235229  0.2307815
#>  [758,]    0.1747231    0.1775158  0.1832804
#>  [759,]    0.1852012    0.1881613  0.1942717
#>  [760,]    0.2672766    0.2715484  0.2803664
#>  [761,]    0.1666353    0.1692987  0.1747965
#>  [762,]    0.1817191    0.1846235  0.1906190
#>  [763,]    0.2412241    0.2450796  0.2530381
#>  [764,]    0.2323740    0.2360880  0.2437546
#>  [765,]    0.1886456    0.1916608  0.1978848
#>  [766,]    0.2318838    0.2355900  0.2432404
#>  [767,]    0.2307078    0.2343953  0.2420069
#>  [768,]    0.1667365    0.1694015  0.1749027
#>  [769,]    0.1827803    0.1857018  0.1917322
#>  [770,]    0.2379848    0.2417885  0.2496402
#>  [771,]    0.2015813    0.2048032  0.2114539
#>  [772,]    0.3154708    0.3205129  0.3309208
#>  [773,]    0.1797447    0.1826177  0.1885480
#>  [774,]    0.2090015    0.2123420  0.2192375
#>  [775,]    0.1850330    0.1879905  0.1940953
#>  [776,]    0.2154027    0.2188455  0.2259521
#>  [777,]    0.2169495    0.2204170  0.2275748
#>  [778,]    0.2216451    0.2251877  0.2325004
#>  [779,]    0.1628056    0.1654077  0.1707792
#>  [780,]    0.2451825    0.2491012  0.2571903
#>  [781,]    0.1926056    0.1956841  0.2020387
#>  [782,]    0.2022956    0.2055289  0.2122032
#>  [783,]    0.2315736    0.2352748  0.2429150
#>  [784,]    0.3102862    0.3152454  0.3254824
#>  [785,]    0.1774268    0.1802627  0.1861165
#>  [786,]    0.2036004    0.2068546  0.2135720
#>  [787,]    0.3213128    0.3264482  0.3370487
#>  [788,]    0.2182655    0.2217541  0.2289552
#>  [789,]    0.2161527    0.2196075  0.2267391
#>  [790,]    0.2740201    0.2783997  0.2874402
#>  [791,]    0.1695008    0.1722100  0.1778023
#>  [792,]    0.1856286    0.1885955  0.1947200
#>  [793,]    0.1526021    0.1550413  0.1600761
#>  [794,]    0.2681117    0.2723969  0.2812425
#>  [795,]    0.2171182    0.2205884  0.2277518
#>  [796,]    0.2944106    0.2991161  0.3088292
#>  [797,]    0.1451475    0.1474675  0.1522564
#>  [798,]    0.1828671    0.1857900  0.1918233
#>  [799,]    0.1948696    0.1979843  0.2044136
#>  [800,]    0.2256808    0.2292879  0.2367336
#>  [801,]    0.2222815    0.2258343  0.2331679
#>  [802,]    0.2141879    0.2176113  0.2246780
#>  [803,]    0.2479955    0.2519592  0.2601411
#>  [804,]    0.1871328    0.1901238  0.1962978
#>  [805,]    0.1914380    0.1944978  0.2008138
#>  [806,]    0.1741960    0.1769803  0.1827276
#>  [807,]    0.2019863    0.2052147  0.2118788
#>  [808,]    0.2069683    0.2102763  0.2171047
#>  [809,]    0.1639021    0.1665218  0.1719294
#>  [810,]    0.2218445    0.2253902  0.2327093
#>  [811,]    0.2367236    0.2405072  0.2483173
#>  [812,]    0.1947069    0.1978189  0.2042429
#>  [813,]    0.1873434    0.1903378  0.1965188
#>  [814,]    0.1999572    0.2031532  0.2097503
#>  [815,]    0.2531240    0.2571697  0.2655209
#>  [816,]    0.1388227    0.1410416  0.1456218
#>  [817,]    0.2158363    0.2192860  0.2264070
#>  [818,]    0.1707392    0.1734681  0.1791013
#>  [819,]    0.2498891    0.2538831  0.2621277
#>  [820,]    0.1910419    0.1940954  0.2003985
#>  [821,]    0.2019428    0.2051704  0.2118331
#>  [822,]    0.2167685    0.2202332  0.2273850
#>  [823,]    0.1716142    0.1743571  0.1800191
#>  [824,]    0.2447013    0.2486124  0.2566857
#>  [825,]    0.2658149    0.2700634  0.2788332
#>  [826,]    0.3221552    0.3273042  0.3379326
#>  [827,]    0.1842088    0.1871531  0.1932307
#>  [828,]    0.2208997    0.2244304  0.2317184
#>  [829,]    0.1858304    0.1888006  0.1949317
#>  [830,]    0.2171049    0.2205749  0.2277378
#>  [831,]    0.2339345    0.2376734  0.2453915
#>  [832,]    0.2429079    0.2467903  0.2548044
#>  [833,]    0.1969216    0.2000691  0.2065660
#>  [834,]    0.1797499    0.1826229  0.1885534
#>  [835,]    0.1438055    0.1461040  0.1508485
#>  [836,]    0.1536976    0.1561542  0.1612252
#>  [837,]    0.1535899    0.1560448  0.1611122
#>  [838,]    0.1812244    0.1841209  0.1901000
#>  [839,]    0.2202792    0.2238000  0.2310676
#>  [840,]    0.2105558    0.2139211  0.2208679
#>  [841,]    0.1991092    0.2022916  0.2088608
#>  [842,]    0.2438098    0.2477066  0.2557505
#>  [843,]    0.2146559    0.2180868  0.2251689
#>  [844,]    0.2717228    0.2760657  0.2850304
#>  [845,]    0.2958103    0.3005382  0.3102976
#>  [846,]    0.1621256    0.1647169  0.1700659
#>  [847,]    0.2098497    0.2132038  0.2201273
#>  [848,]    0.2056063    0.2088926  0.2156762
#>  [849,]    0.2549722    0.2590474  0.2674596
#>  [850,]    0.2454571    0.2493803  0.2574785
#>  [851,]    0.2061461    0.2094410  0.2162424
#>  [852,]    0.2058202    0.2091098  0.2159004
#>  [853,]    0.1805128    0.1833979  0.1893535
#>  [854,]    0.2151210    0.2185594  0.2256569
#>  [855,]    0.1908726    0.1939233  0.2002208
#>  [856,]    0.2062411    0.2095374  0.2163418
#>  [857,]    0.2873933    0.2919866  0.3014683
#>  [858,]    0.2208361    0.2243657  0.2316517
#>  [859,]    0.3207239    0.3258499  0.3364312
#>  [860,]    0.1320039    0.1341138  0.1384690
#>  [861,]    0.2032221    0.2064703  0.2131752
#>  [862,]    0.1931953    0.1962832  0.2026572
#>  [863,]    0.1547726    0.1572464  0.1623529
#>  [864,]    0.2535544    0.2576070  0.2659724
#>  [865,]    0.2839475    0.2884858  0.2978538
#>  [866,]    0.1587120    0.1612487  0.1664851
#>  [867,]    0.2701953    0.2745138  0.2834281
#>  [868,]    0.1844175    0.1873650  0.1934494
#>  [869,]    0.1842414    0.1871862  0.1932649
#>  [870,]    0.1857131    0.1886814  0.1948086
#>  [871,]    0.1650557    0.1676939  0.1731395
#>  [872,]    0.2098425    0.2131964  0.2201197
#>  [873,]    0.2183625    0.2218526  0.2290569
#>  [874,]    0.2246417    0.2282322  0.2356438
#>  [875,]    0.1462070    0.1485439  0.1533677
#>  [876,]    0.3017624    0.3065854  0.3165411
#>  [877,]    0.2053736    0.2086562  0.2154320
#>  [878,]    0.2294523    0.2331197  0.2406899
#>  [879,]    0.2013260    0.2045439  0.2111862
#>  [880,]    0.2537113    0.2577664  0.2661368
#>  [881,]    0.2424368    0.2463115  0.2543099
#>  [882,]    0.2260510    0.2296640  0.2371220
#>  [883,]    0.2306521    0.2343386  0.2419484
#>  [884,]    0.2281765    0.2318234  0.2393515
#>  [885,]    0.1835774    0.1865115  0.1925683
#>  [886,]    0.2196786    0.2231898  0.2304376
#>  [887,]    0.2190720    0.2225735  0.2298013
#>  [888,]    0.1962974    0.1994348  0.2059113
#>  [889,]    0.1950326    0.1981498  0.2045845
#>  [890,]    0.2070413    0.2103504  0.2171813
#>  [891,]    0.2345907    0.2383402  0.2460800
#>  [892,]    0.1789510    0.1818112  0.1877154
#>  [893,]    0.2135712    0.2169848  0.2240311
#>  [894,]    0.1709225    0.1736544  0.1792937
#>  [895,]    0.2623688    0.2665622  0.2752182
#>  [896,]    0.2804514    0.2849338  0.2941865
#>  [897,]    0.2412887    0.2451453  0.2531060
#>  [898,]    0.1462058    0.1485427  0.1533666
#>  [899,]    0.1740063    0.1767875  0.1825285
#>  [900,]    0.1658560    0.1685069  0.1739791
#>  [901,]    0.1808117    0.1837016  0.1896672
#>  [902,]    0.1734719    0.1762445  0.1819679
#>  [903,]    0.2248367    0.2284303  0.2358482
#>  [904,]    0.2309975    0.2346896  0.2423108
#>  [905,]    0.2474514    0.2514064  0.2595706
#>  [906,]    0.1754974    0.1783024  0.1840926
#>  [907,]    0.2653892    0.2696309  0.2783866
#>  [908,]    0.2604281    0.2645905  0.2731827
#>  [909,]    0.1752944    0.1780962  0.1838796
#>  [910,]    0.1755337    0.1783392  0.1841306
#>  [911,]    0.2093571    0.2127033  0.2196106
#>  [912,]    0.1952489    0.1983696  0.2048113
#>  [913,]    0.3340540    0.3393932  0.3504142
#>  [914,]    0.1633469    0.1659577  0.1713471
#>  [915,]    0.1333777    0.1355095  0.1399101
#>  [916,]    0.2318198    0.2355250  0.2431733
#>  [917,]    0.1800050    0.1828821  0.1888209
#>  [918,]    0.1974455    0.2006013  0.2071155
#>  [919,]    0.2224332    0.2259883  0.2333268
#>  [920,]    0.1769055    0.1797330  0.1855697
#>  [921,]    0.2224668    0.2260225  0.2333623
#>  [922,]    0.2262209    0.2298365  0.2372999
#>  [923,]    0.1780899    0.1809364  0.1868121
#>  [924,]    0.3200492    0.3251645  0.3357235
#>  [925,]    0.2410257    0.2448780  0.2528301
#>  [926,]    0.2604751    0.2646383  0.2732320
#>  [927,]    0.2138500    0.2172679  0.2243234
#>  [928,]    0.1915969    0.1946592  0.2009805
#>  [929,]    0.2453084    0.2492292  0.2573225
#>  [930,]    0.2611489    0.2653228  0.2739386
#>  [931,]    0.2168922    0.2203588  0.2275147
#>  [932,]    0.2221731    0.2257241  0.2330542
#>  [933,]    0.2065684    0.2098701  0.2166854
#>  [934,]    0.1770200    0.1798494  0.1856898
#>  [935,]    0.2166332    0.2200957  0.2272430
#>  [936,]    0.3466110    0.3521507  0.3635857
#>  [937,]    0.2022598    0.2054925  0.2121656
#>  [938,]    0.1930049    0.1960898  0.2024575
#>  [939,]    0.1864803    0.1894608  0.1956133
#>  [940,]    0.2128539    0.2162559  0.2232786
#>  [941,]    0.2222517    0.2258040  0.2331366
#>  [942,]    0.1986359    0.2018107  0.2083642
#>  [943,]    0.2062790    0.2095760  0.2163818
#>  [944,]    0.1975916    0.2007497  0.2072688
#>  [945,]    0.1967652    0.1999101  0.2064020
#>  [946,]    0.2195648    0.2230742  0.2303182
#>  [947,]    0.2246315    0.2282219  0.2356331
#>  [948,]    0.2093344    0.2126802  0.2195867
#>  [949,]    0.2468245    0.2507695  0.2589128
#>  [950,]    0.1843470    0.1872934  0.1933756
#>  [951,]    0.1679310    0.1706151  0.1761557
#>  [952,]    0.1758196    0.1786297  0.1844305
#>  [953,]    0.1839544    0.1868946  0.1929637
#>  [954,]    0.2053167    0.2085983  0.2153723
#>  [955,]    0.3029023    0.3077435  0.3177369
#>  [956,]    0.2212750    0.2248117  0.2321122
#>  [957,]    0.2021068    0.2053371  0.2120052
#>  [958,]    0.2073079    0.2106213  0.2174610
#>  [959,]    0.1544877    0.1569569  0.1620539
#>  [960,]    0.2170285    0.2204973  0.2276576
#>  [961,]    0.2560561    0.2601486  0.2685966
#>  [962,]    0.2112973    0.2146744  0.2216456
#>  [963,]    0.1909787    0.1940312  0.2003321
#>  [964,]    0.2607771    0.2649451  0.2735486
#>  [965,]    0.1965940    0.1997362  0.2062224
#>  [966,]    0.2751444    0.2795419  0.2886195
#>  [967,]    0.2183947    0.2218853  0.2290908
#>  [968,]    0.3221596    0.3273084  0.3379366
#>  [969,]    0.2326754    0.2363942  0.2440708
#>  [970,]    0.2683869    0.2726765  0.2815312
#>  [971,]    0.1705986    0.1733253  0.1789538
#>  [972,]    0.2631027    0.2673079  0.2759884
#>  [973,]    0.1817607    0.1846658  0.1906626
#>  [974,]    0.1894262    0.1924539  0.1987036
#>  [975,]    0.2881532    0.2927587  0.3022655
#>  [976,]    0.1886479    0.1916630  0.1978870
#>  [977,]    0.2175508    0.2210279  0.2282055
#>  [978,]    0.2341006    0.2378422  0.2455658
#>  [979,]    0.2043558    0.2076221  0.2143644
#>  [980,]    0.1765414    0.1793631  0.1851878
#>  [981,]    0.1801289    0.1830080  0.1889510
#>  [982,]    0.2465956    0.2505369  0.2586727
#>  [983,]    0.3031746    0.3080202  0.3180225
#>  [984,]    0.2192775    0.2227822  0.2300167
#>  [985,]    0.2709543    0.2752849  0.2842243
#>  [986,]    0.1768811    0.1797083  0.1855441
#>  [987,]    0.3158165    0.3208641  0.3312833
#>  [988,]    0.1786125    0.1814673  0.1873602
#>  [989,]    0.1600797    0.1626383  0.1679198
#>  [990,]    0.2038116    0.2070691  0.2137934
#>  [991,]    0.2038385    0.2070965  0.2138217
#>  [992,]    0.1772261    0.1800587  0.1859059
#>  [993,]    0.2908964    0.2955457  0.3051430
#>  [994,]    0.3286481    0.3339007  0.3447434
#>  [995,]    0.2210175    0.2245500  0.2318419
#>  [996,]    0.2177146    0.2211944  0.2283774
#>  [997,]    0.1857459    0.1887147  0.1948429
#>  [998,]    0.2117389    0.2151231  0.2221089
#>  [999,]    0.2513389    0.2553560  0.2636482
#> [1000,]    0.1938389    0.1969371  0.2033324