Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 89.97523
#> 2   Feb 86.37179
#> 3   Mar 93.27985
#> 4   Apr 85.35788
#> 5   May 94.62819
#> 6   Jun 95.29579

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 8.972009  2
#> 2   Feb 9.632568  2
#> 3   Mar 8.730486  2
#> 4   Apr 7.993794  2
#> 5   May 6.400705  2
#> 6   Jun 9.821160  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [11ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [23ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [68ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [727ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [20ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1939789    0.2066115  0.1943067
#>    [2,]    0.2006054    0.2136696  0.2009444
#>    [3,]    0.2233984    0.2379469  0.2237759
#>    [4,]    0.1878746    0.2001098  0.1881921
#>    [5,]    0.2052142    0.2185786  0.2055611
#>    [6,]    0.2691711    0.2867001  0.2696259
#>    [7,]    0.2228128    0.2373232  0.2231894
#>    [8,]    0.1731757    0.1844537  0.1734684
#>    [9,]    0.2439352    0.2598209  0.2443474
#>   [10,]    0.3772398    0.4018055  0.3778773
#>   [11,]    0.2999506    0.3194836  0.3004575
#>   [12,]    0.2346620    0.2499438  0.2350586
#>   [13,]    0.1930480    0.2056203  0.1933743
#>   [14,]    0.1798967    0.1916124  0.1802008
#>   [15,]    0.2636989    0.2808716  0.2641445
#>   [16,]    0.2154290    0.2294585  0.2157931
#>   [17,]    0.2438652    0.2597464  0.2442774
#>   [18,]    0.2317671    0.2468605  0.2321588
#>   [19,]    0.1449459    0.1543855  0.1451909
#>   [20,]    0.2482766    0.2644452  0.2486962
#>   [21,]    0.1772953    0.1888416  0.1775949
#>   [22,]    0.2441798    0.2600814  0.2445924
#>   [23,]    0.2090361    0.2226494  0.2093894
#>   [24,]    0.2023478    0.2155255  0.2026898
#>   [25,]    0.2238573    0.2384356  0.2242356
#>   [26,]    0.1859219    0.1980300  0.1862361
#>   [27,]    0.2699252    0.2875035  0.2703814
#>   [28,]    0.2363384    0.2517295  0.2367378
#>   [29,]    0.2007302    0.2138025  0.2010694
#>   [30,]    0.1693251    0.1803520  0.1696112
#>   [31,]    0.2535790    0.2700928  0.2540075
#>   [32,]    0.2400978    0.2557339  0.2405036
#>   [33,]    0.2439731    0.2598611  0.2443854
#>   [34,]    0.3079922    0.3280491  0.3085127
#>   [35,]    0.2031073    0.2163344  0.2034506
#>   [36,]    0.2164359    0.2305309  0.2168017
#>   [37,]    0.2195252    0.2338216  0.2198962
#>   [38,]    0.1482607    0.1579160  0.1485112
#>   [39,]    0.2285722    0.2434576  0.2289584
#>   [40,]    0.3196649    0.3404817  0.3202051
#>   [41,]    0.2224741    0.2369626  0.2228501
#>   [42,]    0.1978560    0.2107412  0.1981903
#>   [43,]    0.3483527    0.3710378  0.3489414
#>   [44,]    0.1850320    0.1970820  0.1853447
#>   [45,]    0.2495564    0.2658083  0.2499781
#>   [46,]    0.1462082    0.1557299  0.1464553
#>   [47,]    0.2891860    0.3080184  0.2896747
#>   [48,]    0.1738910    0.1852156  0.1741849
#>   [49,]    0.2143568    0.2283166  0.2147191
#>   [50,]    0.1920605    0.2045681  0.1923851
#>   [51,]    0.2003463    0.2133937  0.2006849
#>   [52,]    0.2182593    0.2324732  0.2186282
#>   [53,]    0.2214539    0.2358757  0.2218281
#>   [54,]    0.2583442    0.2751683  0.2587808
#>   [55,]    0.2628687    0.2799875  0.2633129
#>   [56,]    0.2889253    0.3077407  0.2894136
#>   [57,]    0.2571141    0.2738580  0.2575486
#>   [58,]    0.1814481    0.1932647  0.1817547
#>   [59,]    0.2360580    0.2514307  0.2364569
#>   [60,]    0.1699195    0.1809854  0.1702066
#>   [61,]    0.2565596    0.2732674  0.2569932
#>   [62,]    0.1998654    0.2128814  0.2002031
#>   [63,]    0.1669297    0.1778011  0.1672118
#>   [64,]    0.2464930    0.2625454  0.2469095
#>   [65,]    0.1928919    0.2054538  0.1932179
#>   [66,]    0.2381921    0.2537039  0.2385946
#>   [67,]    0.2300520    0.2450339  0.2304408
#>   [68,]    0.2307267    0.2457523  0.2311167
#>   [69,]    0.1938086    0.2064302  0.1941361
#>   [70,]    0.2509126    0.2672526  0.2513366
#>   [71,]    0.1839043    0.1958809  0.1842151
#>   [72,]    0.1843998    0.1964086  0.1847114
#>   [73,]    0.2293845    0.2443229  0.2297722
#>   [74,]    0.2292307    0.2441590  0.2296181
#>   [75,]    0.2703067    0.2879098  0.2707635
#>   [76,]    0.1921267    0.2046388  0.1924514
#>   [77,]    0.2477634    0.2638985  0.2481822
#>   [78,]    0.1842509    0.1962501  0.1845622
#>   [79,]    0.2118471    0.2256434  0.2122051
#>   [80,]    0.1954729    0.2082028  0.1958033
#>   [81,]    0.1871242    0.1993106  0.1874404
#>   [82,]    0.2805102    0.2987776  0.2809843
#>   [83,]    0.3091899    0.3293251  0.3097124
#>   [84,]    0.1840035    0.1959866  0.1843144
#>   [85,]    0.1855086    0.1975899  0.1858222
#>   [86,]    0.1753466    0.1867660  0.1756430
#>   [87,]    0.2393160    0.2549010  0.2397205
#>   [88,]    0.1895621    0.2019072  0.1898825
#>   [89,]    0.2947144    0.3139071  0.2952125
#>   [90,]    0.1395643    0.1486535  0.1398001
#>   [91,]    0.3062618    0.3262062  0.3067794
#>   [92,]    0.1528820    0.1628384  0.1531404
#>   [93,]    0.2241673    0.2387659  0.2245461
#>   [94,]    0.2407405    0.2564180  0.2411473
#>   [95,]    0.2434108    0.2592625  0.2438222
#>   [96,]    0.2212678    0.2356775  0.2216418
#>   [97,]    0.1952219    0.2079356  0.1955519
#>   [98,]    0.2417743    0.2575193  0.2421828
#>   [99,]    0.2013819    0.2144967  0.2017223
#>  [100,]    0.2150823    0.2290892  0.2154457
#>  [101,]    0.2279990    0.2428470  0.2283843
#>  [102,]    0.1872893    0.1994864  0.1876058
#>  [103,]    0.2088447    0.2224452  0.2091977
#>  [104,]    0.1262929    0.1345177  0.1265063
#>  [105,]    0.2580121    0.2748147  0.2584482
#>  [106,]    0.2170277    0.2311612  0.2173945
#>  [107,]    0.1846078    0.1966302  0.1849197
#>  [108,]    0.1128601    0.1202102  0.1130509
#>  [109,]    0.2551935    0.2718127  0.2556248
#>  [110,]    0.2192834    0.2335638  0.2196540
#>  [111,]    0.2876413    0.3063731  0.2881274
#>  [112,]    0.3466506    0.3692246  0.3472364
#>  [113,]    0.2089450    0.2225523  0.2092981
#>  [114,]    0.2085807    0.2221643  0.2089332
#>  [115,]    0.1959738    0.2087364  0.1963050
#>  [116,]    0.1404453    0.1495918  0.1406826
#>  [117,]    0.1891661    0.2014854  0.1894858
#>  [118,]    0.1669828    0.1778575  0.1672650
#>  [119,]    0.1954106    0.2081364  0.1957408
#>  [120,]    0.1785136    0.1901392  0.1788153
#>  [121,]    0.2133816    0.2272779  0.2137422
#>  [122,]    0.1741573    0.1854991  0.1744516
#>  [123,]    0.1555272    0.1656559  0.1557900
#>  [124,]    0.2560673    0.2727426  0.2565000
#>  [125,]    0.3065704    0.3265348  0.3070885
#>  [126,]    0.2568381    0.2735642  0.2572721
#>  [127,]    0.2167002    0.2308125  0.2170664
#>  [128,]    0.2697312    0.2872968  0.2701871
#>  [129,]    0.4983583    0.5308110  0.4992005
#>  [130,]    0.2209973    0.2353895  0.2213708
#>  [131,]    0.2137721    0.2276937  0.2141334
#>  [132,]    0.2122426    0.2260644  0.2126013
#>  [133,]    0.1783584    0.1899739  0.1786599
#>  [134,]    0.1518855    0.1617770  0.1521422
#>  [135,]    0.2312803    0.2463422  0.2316711
#>  [136,]    0.2158536    0.2299108  0.2162184
#>  [137,]    0.2610521    0.2780527  0.2614933
#>  [138,]    0.2152081    0.2292232  0.2155718
#>  [139,]    0.1700035    0.1810750  0.1702909
#>  [140,]    0.1564500    0.1666388  0.1567144
#>  [141,]    0.2147195    0.2287027  0.2150823
#>  [142,]    0.1502785    0.1600654  0.1505324
#>  [143,]    0.1473897    0.1569884  0.1476387
#>  [144,]    0.1593865    0.1697666  0.1596559
#>  [145,]    0.2945511    0.3137329  0.2950489
#>  [146,]    0.2469917    0.2630763  0.2474091
#>  [147,]    0.2411710    0.2568767  0.2415786
#>  [148,]    0.1793990    0.1910823  0.1797021
#>  [149,]    0.1966205    0.2094253  0.1969528
#>  [150,]    0.2845538    0.3030847  0.2850347
#>  [151,]    0.2240052    0.2385931  0.2243838
#>  [152,]    0.1754845    0.1869129  0.1757811
#>  [153,]    0.2055471    0.2189331  0.2058945
#>  [154,]    0.2809072    0.2992006  0.2813819
#>  [155,]    0.2071212    0.2206100  0.2074713
#>  [156,]    0.3338605    0.3556017  0.3344247
#>  [157,]    0.1395570    0.1486456  0.1397929
#>  [158,]    0.2090923    0.2227092  0.2094457
#>  [159,]    0.1956616    0.2084038  0.1959922
#>  [160,]    0.2426841    0.2584884  0.2430942
#>  [161,]    0.1545844    0.1646517  0.1548456
#>  [162,]    0.2521681    0.2685901  0.2525942
#>  [163,]    0.3148877    0.3353937  0.3154199
#>  [164,]    0.2034419    0.2166909  0.2037858
#>  [165,]    0.1830528    0.1949740  0.1833621
#>  [166,]    0.3093575    0.3295035  0.3098803
#>  [167,]    0.2368914    0.2523186  0.2372918
#>  [168,]    0.2160006    0.2300674  0.2163657
#>  [169,]    0.2396394    0.2552454  0.2400444
#>  [170,]    0.1965488    0.2093488  0.1968810
#>  [171,]    0.2807492    0.2990324  0.2812236
#>  [172,]    0.1828281    0.1947347  0.1831371
#>  [173,]    0.2287258    0.2436212  0.2291124
#>  [174,]    0.2319016    0.2470037  0.2322935
#>  [175,]    0.2400820    0.2557168  0.2404877
#>  [176,]    0.2670996    0.2844938  0.2675510
#>  [177,]    0.2295189    0.2444660  0.2299068
#>  [178,]    0.2954998    0.3147433  0.2959992
#>  [179,]    0.2035932    0.2168520  0.2039373
#>  [180,]    0.1944812    0.2071466  0.1948099
#>  [181,]    0.1885017    0.2007777  0.1888203
#>  [182,]    0.1947399    0.2074221  0.1950691
#>  [183,]    0.1966513    0.2094580  0.1969836
#>  [184,]    0.1362395    0.1451122  0.1364698
#>  [185,]    0.2177864    0.2319694  0.2181545
#>  [186,]    0.2608973    0.2778876  0.2613382
#>  [187,]    0.1838340    0.1958059  0.1841446
#>  [188,]    0.2126342    0.2264817  0.2129935
#>  [189,]    0.2383098    0.2538293  0.2387126
#>  [190,]    0.2104194    0.2241227  0.2107750
#>  [191,]    0.1818051    0.1936450  0.1821124
#>  [192,]    0.2055927    0.2189817  0.2059402
#>  [193,]    0.1779154    0.1895020  0.1782160
#>  [194,]    0.1883502    0.2006163  0.1886685
#>  [195,]    0.2058954    0.2193040  0.2062433
#>  [196,]    0.2236235    0.2381867  0.2240014
#>  [197,]    0.2020804    0.2152408  0.2024220
#>  [198,]    0.2769652    0.2950020  0.2774333
#>  [199,]    0.2084536    0.2220289  0.2088058
#>  [200,]    0.2062621    0.2196947  0.2066106
#>  [201,]    0.2073347    0.2208371  0.2076851
#>  [202,]    0.1719889    0.1831895  0.1722796
#>  [203,]    0.2635058    0.2806661  0.2639511
#>  [204,]    0.1590528    0.1694110  0.1593216
#>  [205,]    0.2629999    0.2801274  0.2634444
#>  [206,]    0.2598107    0.2767303  0.2602497
#>  [207,]    0.2588524    0.2757094  0.2592899
#>  [208,]    0.2799928    0.2982267  0.2804660
#>  [209,]    0.2007764    0.2138517  0.2011157
#>  [210,]    0.2145629    0.2285360  0.2149255
#>  [211,]    0.2132562    0.2271443  0.2136166
#>  [212,]    0.2130947    0.2269723  0.2134548
#>  [213,]    0.1696539    0.1807023  0.1699406
#>  [214,]    0.1797525    0.1914588  0.1800563
#>  [215,]    0.2339398    0.2491749  0.2343352
#>  [216,]    0.2285606    0.2434452  0.2289468
#>  [217,]    0.2062291    0.2196594  0.2065776
#>  [218,]    0.2289716    0.2438830  0.2293585
#>  [219,]    0.1988604    0.2118110  0.1991964
#>  [220,]    0.2370683    0.2525070  0.2374689
#>  [221,]    0.2033463    0.2165890  0.2036899
#>  [222,]    0.2148203    0.2288101  0.2151833
#>  [223,]    0.2861870    0.3048244  0.2866706
#>  [224,]    0.2067144    0.2201764  0.2070638
#>  [225,]    0.2088979    0.2225021  0.2092509
#>  [226,]    0.2056028    0.2189924  0.2059502
#>  [227,]    0.2039714    0.2172548  0.2043161
#>  [228,]    0.2491811    0.2654086  0.2496022
#>  [229,]    0.2153080    0.2293298  0.2156719
#>  [230,]    0.1739182    0.1852446  0.1742121
#>  [231,]    0.2668332    0.2842100  0.2672842
#>  [232,]    0.2259399    0.2406535  0.2263217
#>  [233,]    0.1975188    0.2103821  0.1978526
#>  [234,]    0.2693289    0.2868684  0.2697841
#>  [235,]    0.2627272    0.2798370  0.2631712
#>  [236,]    0.2503314    0.2666337  0.2507544
#>  [237,]    0.3395257    0.3616361  0.3400995
#>  [238,]    0.2631595    0.2802972  0.2636042
#>  [239,]    0.3024968    0.3221961  0.3030080
#>  [240,]    0.2944193    0.3135924  0.2949169
#>  [241,]    0.2070542    0.2205384  0.2074041
#>  [242,]    0.1844615    0.1964744  0.1847732
#>  [243,]    0.2692618    0.2867965  0.2697168
#>  [244,]    0.2803300    0.2985856  0.2808037
#>  [245,]    0.1629075    0.1735168  0.1631828
#>  [246,]    0.2952181    0.3144434  0.2957170
#>  [247,]    0.2729264    0.2907000  0.2733877
#>  [248,]    0.1912180    0.2036710  0.1915412
#>  [249,]    0.2050359    0.2183888  0.2053824
#>  [250,]    0.2202581    0.2346021  0.2206304
#>  [251,]    0.2360763    0.2514503  0.2364752
#>  [252,]    0.1681647    0.1791164  0.1684489
#>  [253,]    0.1943605    0.2070179  0.1946889
#>  [254,]    0.2450975    0.2610591  0.2455117
#>  [255,]    0.2660282    0.2833523  0.2664778
#>  [256,]    0.1709502    0.1820832  0.1712391
#>  [257,]    0.2065271    0.2199768  0.2068761
#>  [258,]    0.1923790    0.2049075  0.1927041
#>  [259,]    0.2490978    0.2653199  0.2495188
#>  [260,]    0.2219308    0.2363838  0.2223058
#>  [261,]    0.2290103    0.2439244  0.2293973
#>  [262,]    0.3018883    0.3215478  0.3023985
#>  [263,]    0.2981659    0.3175829  0.2986698
#>  [264,]    0.1660250    0.1768373  0.1663056
#>  [265,]    0.1475599    0.1571697  0.1478093
#>  [266,]    0.2125512    0.2263933  0.2129105
#>  [267,]    0.1644383    0.1751472  0.1647162
#>  [268,]    0.2281120    0.2429674  0.2284975
#>  [269,]    0.2657662    0.2830735  0.2662153
#>  [270,]    0.1744398    0.1857999  0.1747346
#>  [271,]    0.1932515    0.2058369  0.1935781
#>  [272,]    0.2944521    0.3136276  0.2949497
#>  [273,]    0.1718876    0.1830817  0.1721781
#>  [274,]    0.1839130    0.1958903  0.1842239
#>  [275,]    0.2444313    0.2603494  0.2448443
#>  [276,]    0.2746458    0.2925314  0.2751099
#>  [277,]    0.2908944    0.3098379  0.2913860
#>  [278,]    0.1814121    0.1932264  0.1817186
#>  [279,]    0.1897272    0.2020831  0.1900479
#>  [280,]    0.2758286    0.2937912  0.2762947
#>  [281,]    0.1816120    0.1934393  0.1819189
#>  [282,]    0.1853712    0.1974434  0.1856845
#>  [283,]    0.1905875    0.2029993  0.1909096
#>  [284,]    0.2579312    0.2747284  0.2583671
#>  [285,]    0.2822619    0.3006435  0.2827390
#>  [286,]    0.1727529    0.1840032  0.1730449
#>  [287,]    0.2059263    0.2193368  0.2062743
#>  [288,]    0.2252954    0.2399675  0.2256762
#>  [289,]    0.2356078    0.2509514  0.2360059
#>  [290,]    0.2392428    0.2548230  0.2396471
#>  [291,]    0.1962183    0.2089968  0.1965499
#>  [292,]    0.2146915    0.2286730  0.2150543
#>  [293,]    0.2099250    0.2235961  0.2102798
#>  [294,]    0.2133272    0.2272198  0.2136877
#>  [295,]    0.2027111    0.2159125  0.2030537
#>  [296,]    0.2019803    0.2151342  0.2023217
#>  [297,]    0.2626496    0.2797541  0.2630935
#>  [298,]    0.2064618    0.2199074  0.2068107
#>  [299,]    0.2135548    0.2274622  0.2139157
#>  [300,]    0.2406424    0.2563137  0.2410491
#>  [301,]    0.3095981    0.3297595  0.3101213
#>  [302,]    0.1434155    0.1527554  0.1436578
#>  [303,]    0.2718251    0.2895271  0.2722845
#>  [304,]    0.2236386    0.2382027  0.2240166
#>  [305,]    0.1903457    0.2027419  0.1906674
#>  [306,]    0.1982793    0.2111921  0.1986144
#>  [307,]    0.2841182    0.3026205  0.2845983
#>  [308,]    0.2574217    0.2741856  0.2578567
#>  [309,]    0.2350377    0.2503441  0.2354349
#>  [310,]    0.1923416    0.2048677  0.1926667
#>  [311,]    0.2924008    0.3114423  0.2928949
#>  [312,]    0.1908550    0.2032842  0.1911776
#>  [313,]    0.2463780    0.2624229  0.2467944
#>  [314,]    0.2021567    0.2153219  0.2024984
#>  [315,]    0.1620312    0.1725834  0.1623050
#>  [316,]    0.1816415    0.1934707  0.1819484
#>  [317,]    0.2022244    0.2153942  0.2025662
#>  [318,]    0.1652077    0.1759668  0.1654869
#>  [319,]    0.1195724    0.1273595  0.1197745
#>  [320,]    0.2071691    0.2206608  0.2075192
#>  [321,]    0.1988638    0.2118144  0.1991998
#>  [322,]    0.2612683    0.2782827  0.2617098
#>  [323,]    0.2284624    0.2433406  0.2288485
#>  [324,]    0.2062262    0.2196563  0.2065747
#>  [325,]    0.2166993    0.2308113  0.2170655
#>  [326,]    0.1892693    0.2015954  0.1895892
#>  [327,]    0.1627992    0.1734014  0.1630743
#>  [328,]    0.2285839    0.2434701  0.2289702
#>  [329,]    0.2242279    0.2388305  0.2246069
#>  [330,]    0.1726863    0.1839324  0.1729782
#>  [331,]    0.1660212    0.1768332  0.1663017
#>  [332,]    0.2181598    0.2323672  0.2185285
#>  [333,]    0.2964888    0.3157969  0.2969899
#>  [334,]    0.2288510    0.2437545  0.2292377
#>  [335,]    0.2016338    0.2147650  0.2019746
#>  [336,]    0.2493770    0.2656173  0.2497985
#>  [337,]    0.1494979    0.1592339  0.1497506
#>  [338,]    0.1867081    0.1988673  0.1870236
#>  [339,]    0.1922617    0.2047825  0.1925866
#>  [340,]    0.2129251    0.2267917  0.2132849
#>  [341,]    0.1617708    0.1723060  0.1620441
#>  [342,]    0.2202488    0.2345921  0.2206210
#>  [343,]    0.2473862    0.2634967  0.2478043
#>  [344,]    0.2531491    0.2696348  0.2535769
#>  [345,]    0.2311491    0.2462025  0.2315397
#>  [346,]    0.2293255    0.2442601  0.2297130
#>  [347,]    0.1966413    0.2094475  0.1969737
#>  [348,]    0.2013014    0.2144109  0.2016416
#>  [349,]    0.2435664    0.2594282  0.2439780
#>  [350,]    0.2351805    0.2504962  0.2355780
#>  [351,]    0.1696506    0.1806990  0.1699373
#>  [352,]    0.1930394    0.2056107  0.1933657
#>  [353,]    0.2326494    0.2478004  0.2330426
#>  [354,]    0.2843676    0.3028864  0.2848482
#>  [355,]    0.2447388    0.2606772  0.2451525
#>  [356,]    0.1704763    0.1815785  0.1707644
#>  [357,]    0.2128240    0.2266838  0.2131836
#>  [358,]    0.1988581    0.2118085  0.1991942
#>  [359,]    0.1703060    0.1813971  0.1705939
#>  [360,]    0.1398353    0.1489421  0.1400716
#>  [361,]    0.2320604    0.2471729  0.2324526
#>  [362,]    0.2416248    0.2573603  0.2420332
#>  [363,]    0.2430158    0.2588417  0.2434264
#>  [364,]    0.2111639    0.2249156  0.2115207
#>  [365,]    0.2104634    0.2241697  0.2108191
#>  [366,]    0.2561834    0.2728669  0.2566164
#>  [367,]    0.1437250    0.1530851  0.1439679
#>  [368,]    0.2175499    0.2317175  0.2179176
#>  [369,]    0.1566501    0.1668518  0.1569148
#>  [370,]    0.2125825    0.2264268  0.2129417
#>  [371,]    0.2779579    0.2960590  0.2784276
#>  [372,]    0.1312292    0.1397756  0.1314510
#>  [373,]    0.2585259    0.2753617  0.2589628
#>  [374,]    0.1952168    0.2079302  0.1955468
#>  [375,]    0.2056722    0.2190664  0.2060198
#>  [376,]    0.2420241    0.2577853  0.2424331
#>  [377,]    0.2537116    0.2702340  0.2541404
#>  [378,]    0.1538792    0.1639006  0.1541392
#>  [379,]    0.2203922    0.2347450  0.2207647
#>  [380,]    0.2232902    0.2378317  0.2236675
#>  [381,]    0.2675310    0.2849534  0.2679831
#>  [382,]    0.1464998    0.1560406  0.1467474
#>  [383,]    0.2048752    0.2182174  0.2052214
#>  [384,]    0.2075417    0.2210575  0.2078925
#>  [385,]    0.3175979    0.3382803  0.3181346
#>  [386,]    0.1552741    0.1653863  0.1555365
#>  [387,]    0.3283633    0.3497466  0.3289182
#>  [388,]    0.2090049    0.2226160  0.2093582
#>  [389,]    0.2128017    0.2266600  0.2131613
#>  [390,]    0.3022034    0.3218837  0.3027141
#>  [391,]    0.2137115    0.2276291  0.2140727
#>  [392,]    0.2884667    0.3072524  0.2889542
#>  [393,]    0.2042521    0.2175538  0.2045973
#>  [394,]    0.2216358    0.2360697  0.2220104
#>  [395,]    0.1723895    0.1836162  0.1726808
#>  [396,]    0.3010249    0.3206280  0.3015336
#>  [397,]    0.1979400    0.2108307  0.1982745
#>  [398,]    0.1882600    0.2005203  0.1885781
#>  [399,]    0.2068308    0.2203005  0.2071803
#>  [400,]    0.1780097    0.1896025  0.1783106
#>  [401,]    0.2006664    0.2137345  0.2010055
#>  [402,]    0.1897907    0.2021506  0.1901115
#>  [403,]    0.1763328    0.1878163  0.1766308
#>  [404,]    0.2735502    0.2913643  0.2740125
#>  [405,]    0.1963806    0.2091698  0.1967125
#>  [406,]    0.1690402    0.1800487  0.1693258
#>  [407,]    0.2172519    0.2314001  0.2176191
#>  [408,]    0.1854690    0.1975475  0.1857824
#>  [409,]    0.2047684    0.2181037  0.2051144
#>  [410,]    0.1691253    0.1801394  0.1694111
#>  [411,]    0.1878119    0.2000430  0.1881293
#>  [412,]    0.2069698    0.2204483  0.2073195
#>  [413,]    0.2182874    0.2325030  0.2186563
#>  [414,]    0.1778821    0.1894665  0.1781827
#>  [415,]    0.3132501    0.3336496  0.3137795
#>  [416,]    0.1430772    0.1523952  0.1433190
#>  [417,]    0.1613051    0.1718101  0.1615777
#>  [418,]    0.1861728    0.1982971  0.1864874
#>  [419,]    0.1996760    0.2126797  0.2000135
#>  [420,]    0.1722776    0.1834970  0.1725687
#>  [421,]    0.2253711    0.2400481  0.2257519
#>  [422,]    0.2450070    0.2609625  0.2454210
#>  [423,]    0.2424832    0.2582746  0.2428930
#>  [424,]    0.2029581    0.2161755  0.2033011
#>  [425,]    0.2201010    0.2344348  0.2204730
#>  [426,]    0.2204642    0.2348216  0.2208368
#>  [427,]    0.1771598    0.1886970  0.1774592
#>  [428,]    0.1554355    0.1655581  0.1556982
#>  [429,]    0.2585616    0.2753998  0.2589985
#>  [430,]    0.1531989    0.1631759  0.1534578
#>  [431,]    0.3059859    0.3259122  0.3065030
#>  [432,]    0.2154620    0.2294936  0.2158261
#>  [433,]    0.2932588    0.3123565  0.2937544
#>  [434,]    0.2625194    0.2796151  0.2629630
#>  [435,]    0.2900973    0.3089890  0.2905875
#>  [436,]    0.1976971    0.2105719  0.1980312
#>  [437,]    0.2721185    0.2898396  0.2725784
#>  [438,]    0.2143851    0.2283467  0.2147474
#>  [439,]    0.2118907    0.2256900  0.2122488
#>  [440,]    0.2054792    0.2188607  0.2058264
#>  [441,]    0.2140917    0.2280342  0.2144535
#>  [442,]    0.2120773    0.2258884  0.2124357
#>  [443,]    0.2824500    0.3008440  0.2829273
#>  [444,]    0.1708733    0.1820014  0.1711621
#>  [445,]    0.2111826    0.2249356  0.2115395
#>  [446,]    0.2138225    0.2277475  0.2141838
#>  [447,]    0.1970157    0.2098461  0.1973487
#>  [448,]    0.3121094    0.3324346  0.3126368
#>  [449,]    0.3230840    0.3441236  0.3236300
#>  [450,]    0.1747402    0.1861199  0.1750355
#>  [451,]    0.1822804    0.1941513  0.1825884
#>  [452,]    0.1767733    0.1882857  0.1770721
#>  [453,]    0.1876163    0.1998346  0.1879334
#>  [454,]    0.1722925    0.1835130  0.1725837
#>  [455,]    0.2108353    0.2245658  0.2111916
#>  [456,]    0.2081237    0.2216775  0.2084755
#>  [457,]    0.1644957    0.1752083  0.1647737
#>  [458,]    0.2156567    0.2297010  0.2160212
#>  [459,]    0.2461353    0.2621643  0.2465513
#>  [460,]    0.2349076    0.2502057  0.2353046
#>  [461,]    0.2165130    0.2306131  0.2168789
#>  [462,]    0.1586160    0.1689458  0.1588841
#>  [463,]    0.1655731    0.1763561  0.1658529
#>  [464,]    0.2022433    0.2154142  0.2025851
#>  [465,]    0.2039982    0.2172834  0.2043429
#>  [466,]    0.2581145    0.2749237  0.2585507
#>  [467,]    0.2051855    0.2185480  0.2055323
#>  [468,]    0.1826562    0.1945515  0.1829649
#>  [469,]    0.1845687    0.1965885  0.1848806
#>  [470,]    0.2941846    0.3133426  0.2946818
#>  [471,]    0.2448989    0.2608475  0.2453128
#>  [472,]    0.1291344    0.1375443  0.1293527
#>  [473,]    0.1890413    0.2013525  0.1893608
#>  [474,]    0.1598736    0.1702853  0.1601438
#>  [475,]    0.2307098    0.2457345  0.2310997
#>  [476,]    0.1942059    0.2068533  0.1945341
#>  [477,]    0.2421433    0.2579123  0.2425525
#>  [478,]    0.2052976    0.2186673  0.2056445
#>  [479,]    0.1939896    0.2066230  0.1943175
#>  [480,]    0.3350685    0.3568886  0.3356347
#>  [481,]    0.1803103    0.1920528  0.1806150
#>  [482,]    0.1868349    0.1990023  0.1871506
#>  [483,]    0.2369109    0.2523394  0.2373113
#>  [484,]    0.1550294    0.1651256  0.1552914
#>  [485,]    0.1315522    0.1401196  0.1317745
#>  [486,]    0.1982144    0.2111230  0.1985494
#>  [487,]    0.1519052    0.1617979  0.1521619
#>  [488,]    0.2100823    0.2237635  0.2104374
#>  [489,]    0.2719375    0.2896469  0.2723971
#>  [490,]    0.1999770    0.2130003  0.2003150
#>  [491,]    0.1752158    0.1866266  0.1755119
#>  [492,]    0.2322244    0.2473476  0.2326169
#>  [493,]    0.1838848    0.1958602  0.1841956
#>  [494,]    0.2242104    0.2388120  0.2245894
#>  [495,]    0.2703296    0.2879339  0.2707864
#>  [496,]    0.1888038    0.2010994  0.1891228
#>  [497,]    0.2222403    0.2367133  0.2226158
#>  [498,]    0.2308097    0.2458407  0.2311998
#>  [499,]    0.2454043    0.2613858  0.2458191
#>  [500,]    0.1852329    0.1972961  0.1855459
#>  [501,]    0.2711911    0.2888516  0.2716494
#>  [502,]    0.1967339    0.2095460  0.1970664
#>  [503,]    0.2288539    0.2437575  0.2292407
#>  [504,]    0.1714709    0.1826379  0.1717607
#>  [505,]    0.1509776    0.1608100  0.1512328
#>  [506,]    0.1848025    0.1968373  0.1851148
#>  [507,]    0.2355215    0.2508593  0.2359195
#>  [508,]    0.1666589    0.1775124  0.1669405
#>  [509,]    0.1803501    0.1920953  0.1806549
#>  [510,]    0.2396443    0.2552507  0.2400493
#>  [511,]    0.2282609    0.2431261  0.2286466
#>  [512,]    0.1802687    0.1920085  0.1805733
#>  [513,]    0.2096296    0.2232815  0.2099839
#>  [514,]    0.1995596    0.2125554  0.1998968
#>  [515,]    0.1684788    0.1794509  0.1687635
#>  [516,]    0.2750684    0.2929815  0.2755333
#>  [517,]    0.2183884    0.2326106  0.2187575
#>  [518,]    0.1666211    0.1774722  0.1669027
#>  [519,]    0.2025017    0.2156893  0.2028439
#>  [520,]    0.1588134    0.1691560  0.1590818
#>  [521,]    0.2990220    0.3184947  0.2995273
#>  [522,]    0.2450194    0.2609760  0.2454335
#>  [523,]    0.2040506    0.2173393  0.2043955
#>  [524,]    0.2074524    0.2209624  0.2078030
#>  [525,]    0.2666202    0.2839833  0.2670707
#>  [526,]    0.1949947    0.2076936  0.1953242
#>  [527,]    0.2157634    0.2298146  0.2161280
#>  [528,]    0.2045965    0.2179206  0.2049422
#>  [529,]    0.1427776    0.1520758  0.1430189
#>  [530,]    0.2136454    0.2275588  0.2140064
#>  [531,]    0.1897964    0.2021567  0.1901172
#>  [532,]    0.1932412    0.2058258  0.1935678
#>  [533,]    0.1813287    0.1931376  0.1816351
#>  [534,]    0.2235517    0.2381101  0.2239295
#>  [535,]    0.2235390    0.2380966  0.2239168
#>  [536,]    0.1917699    0.2042586  0.1920940
#>  [537,]    0.1936179    0.2062270  0.1939451
#>  [538,]    0.2181195    0.2323243  0.2184881
#>  [539,]    0.1893315    0.2016616  0.1896515
#>  [540,]    0.1983473    0.2112645  0.1986825
#>  [541,]    0.1728716    0.1841298  0.1731638
#>  [542,]    0.2454562    0.2614409  0.2458710
#>  [543,]    0.2482157    0.2643803  0.2486352
#>  [544,]    0.2461133    0.2621409  0.2465292
#>  [545,]    0.2218557    0.2363038  0.2222306
#>  [546,]    0.1944225    0.2070838  0.1947510
#>  [547,]    0.2480971    0.2642539  0.2485164
#>  [548,]    0.1855551    0.1976392  0.1858687
#>  [549,]    0.2414629    0.2571876  0.2418710
#>  [550,]    0.2718887    0.2895948  0.2723482
#>  [551,]    0.3108707    0.3311150  0.3113960
#>  [552,]    0.3047033    0.3245459  0.3052182
#>  [553,]    0.1722752    0.1834945  0.1725664
#>  [554,]    0.1848118    0.1968476  0.1851242
#>  [555,]    0.2318261    0.2469234  0.2322179
#>  [556,]    0.2125729    0.2264165  0.2129322
#>  [557,]    0.2417046    0.2574451  0.2421131
#>  [558,]    0.2377227    0.2532036  0.2381244
#>  [559,]    0.2644095    0.2816285  0.2648564
#>  [560,]    0.1964197    0.2092113  0.1967517
#>  [561,]    0.2116211    0.2254027  0.2119787
#>  [562,]    0.1849090    0.1969511  0.1852215
#>  [563,]    0.1666526    0.1775058  0.1669343
#>  [564,]    0.1814057    0.1932197  0.1817122
#>  [565,]    0.2265015    0.2412521  0.2268843
#>  [566,]    0.1549683    0.1650605  0.1552302
#>  [567,]    0.1984864    0.2114126  0.1988218
#>  [568,]    0.2413279    0.2570437  0.2417357
#>  [569,]    0.2748117    0.2927080  0.2752761
#>  [570,]    0.1595441    0.1699345  0.1598138
#>  [571,]    0.2228165    0.2373272  0.2231931
#>  [572,]    0.2073669    0.2208712  0.2077173
#>  [573,]    0.1855995    0.1976865  0.1859132
#>  [574,]    0.2549049    0.2715051  0.2553357
#>  [575,]    0.2556175    0.2722641  0.2560495
#>  [576,]    0.1993533    0.2123360  0.1996902
#>  [577,]    0.2291030    0.2440229  0.2294901
#>  [578,]    0.2068311    0.2203007  0.2071807
#>  [579,]    0.2293051    0.2442382  0.2296927
#>  [580,]    0.2079483    0.2214906  0.2082997
#>  [581,]    0.1799425    0.1916611  0.1802466
#>  [582,]    0.2048073    0.2181450  0.2051534
#>  [583,]    0.2208704    0.2352542  0.2212437
#>  [584,]    0.2221844    0.2366538  0.2225599
#>  [585,]    0.1852023    0.1972634  0.1855153
#>  [586,]    0.2250391    0.2396944  0.2254194
#>  [587,]    0.2513269    0.2676939  0.2517517
#>  [588,]    0.1918495    0.2043435  0.1921737
#>  [589,]    0.2492588    0.2654913  0.2496800
#>  [590,]    0.1565309    0.1667249  0.1567955
#>  [591,]    0.2163923    0.2304846  0.2167580
#>  [592,]    0.1733679    0.1846583  0.1736609
#>  [593,]    0.2638922    0.2810775  0.2643382
#>  [594,]    0.1606246    0.1710852  0.1608961
#>  [595,]    0.1864702    0.1986139  0.1867853
#>  [596,]    0.1957409    0.2084881  0.1960717
#>  [597,]    0.2063550    0.2197938  0.2067038
#>  [598,]    0.2454216    0.2614043  0.2458363
#>  [599,]    0.1821783    0.1940427  0.1824862
#>  [600,]    0.1500798    0.1598537  0.1503334
#>  [601,]    0.1433593    0.1526955  0.1436016
#>  [602,]    0.1936597    0.2062716  0.1939870
#>  [603,]    0.2583059    0.2751274  0.2587424
#>  [604,]    0.1905197    0.2029270  0.1908417
#>  [605,]    0.1916369    0.2041170  0.1919607
#>  [606,]    0.1898379    0.2022008  0.1901587
#>  [607,]    0.2050458    0.2183993  0.2053924
#>  [608,]    0.2016896    0.2148242  0.2020304
#>  [609,]    0.2113730    0.2251384  0.2117302
#>  [610,]    0.1876268    0.1998459  0.1879439
#>  [611,]    0.1671362    0.1780209  0.1674187
#>  [612,]    0.1504242    0.1602205  0.1506784
#>  [613,]    0.1995852    0.2125829  0.1999225
#>  [614,]    0.2354887    0.2508245  0.2358867
#>  [615,]    0.2189122    0.2331685  0.2192821
#>  [616,]    0.1602963    0.1707355  0.1605672
#>  [617,]    0.1701289    0.1812085  0.1704164
#>  [618,]    0.1919232    0.2044220  0.1922476
#>  [619,]    0.1916085    0.2040869  0.1919324
#>  [620,]    0.2019167    0.2150663  0.2022579
#>  [621,]    0.2128654    0.2267280  0.2132252
#>  [622,]    0.2238862    0.2384665  0.2242646
#>  [623,]    0.2096835    0.2233389  0.2100379
#>  [624,]    0.1931218    0.2056987  0.1934482
#>  [625,]    0.2041008    0.2173926  0.2044457
#>  [626,]    0.2275582    0.2423776  0.2279428
#>  [627,]    0.2612944    0.2783107  0.2617360
#>  [628,]    0.1727599    0.1840108  0.1730518
#>  [629,]    0.2117430    0.2255325  0.2121008
#>  [630,]    0.2677618    0.2851991  0.2682143
#>  [631,]    0.2051569    0.2185174  0.2055036
#>  [632,]    0.2616150    0.2786520  0.2620571
#>  [633,]    0.2621580    0.2792304  0.2626010
#>  [634,]    0.1928539    0.2054132  0.1931798
#>  [635,]    0.1707987    0.1819220  0.1710874
#>  [636,]    0.2002038    0.2132420  0.2005422
#>  [637,]    0.2114084    0.2251763  0.2117657
#>  [638,]    0.1930880    0.2056627  0.1934143
#>  [639,]    0.2385304    0.2540644  0.2389335
#>  [640,]    0.1588082    0.1691505  0.1590766
#>  [641,]    0.2879001    0.3066490  0.2883867
#>  [642,]    0.1658689    0.1766711  0.1661493
#>  [643,]    0.1880171    0.2002616  0.1883348
#>  [644,]    0.2833766    0.3018307  0.2838555
#>  [645,]    0.2572281    0.2739796  0.2576628
#>  [646,]    0.2748130    0.2927096  0.2752775
#>  [647,]    0.1267075    0.1349593  0.1269217
#>  [648,]    0.2274783    0.2422925  0.2278627
#>  [649,]    0.1564451    0.1666336  0.1567095
#>  [650,]    0.2168717    0.2309952  0.2172382
#>  [651,]    0.1643399    0.1750424  0.1646176
#>  [652,]    0.2246943    0.2393272  0.2250741
#>  [653,]    0.2043519    0.2176600  0.2046972
#>  [654,]    0.1871767    0.1993665  0.1874931
#>  [655,]    0.1908874    0.2033188  0.1912100
#>  [656,]    0.2366213    0.2520308  0.2370212
#>  [657,]    0.1986928    0.2116324  0.1990286
#>  [658,]    0.2176304    0.2318033  0.2179982
#>  [659,]    0.1963349    0.2091210  0.1966667
#>  [660,]    0.2099446    0.2236169  0.2102994
#>  [661,]    0.1791626    0.1908305  0.1794654
#>  [662,]    0.2617960    0.2788448  0.2622385
#>  [663,]    0.1702151    0.1813002  0.1705027
#>  [664,]    0.2065572    0.2200090  0.2069062
#>  [665,]    0.1589446    0.1692958  0.1592132
#>  [666,]    0.2155446    0.2295817  0.2159089
#>  [667,]    0.2382440    0.2537593  0.2386467
#>  [668,]    0.1640525    0.1747364  0.1643298
#>  [669,]    0.2204765    0.2348347  0.2208491
#>  [670,]    0.2648770    0.2821263  0.2653246
#>  [671,]    0.2208313    0.2352127  0.2212045
#>  [672,]    0.3103884    0.3306015  0.3109129
#>  [673,]    0.2199902    0.2343169  0.2203620
#>  [674,]    0.2773717    0.2954347  0.2778404
#>  [675,]    0.2361741    0.2515545  0.2365732
#>  [676,]    0.2037511    0.2170200  0.2040954
#>  [677,]    0.2183226    0.2325404  0.2186916
#>  [678,]    0.2374850    0.2529506  0.2378864
#>  [679,]    0.1869287    0.1991022  0.1872446
#>  [680,]    0.1940566    0.2066942  0.1943845
#>  [681,]    0.2254164    0.2400963  0.2257974
#>  [682,]    0.1858936    0.1979997  0.1862077
#>  [683,]    0.2046288    0.2179549  0.2049746
#>  [684,]    0.1725890    0.1838288  0.1728807
#>  [685,]    0.2116096    0.2253904  0.2119672
#>  [686,]    0.2337950    0.2490204  0.2341901
#>  [687,]    0.2094353    0.2230745  0.2097892
#>  [688,]    0.1850182    0.1970673  0.1853309
#>  [689,]    0.2411204    0.2568230  0.2415279
#>  [690,]    0.1790912    0.1907544  0.1793939
#>  [691,]    0.1535150    0.1635127  0.1537744
#>  [692,]    0.2295374    0.2444857  0.2299253
#>  [693,]    0.2439726    0.2598609  0.2443849
#>  [694,]    0.2614503    0.2784766  0.2618921
#>  [695,]    0.1995170    0.2125103  0.1998542
#>  [696,]    0.1812997    0.1931067  0.1816061
#>  [697,]    0.1642090    0.1749031  0.1644865
#>  [698,]    0.3011879    0.3208019  0.3016969
#>  [699,]    0.1838309    0.1958027  0.1841416
#>  [700,]    0.1909347    0.2033692  0.1912574
#>  [701,]    0.1778436    0.1894256  0.1781442
#>  [702,]    0.2371999    0.2526472  0.2376008
#>  [703,]    0.1684350    0.1794042  0.1687197
#>  [704,]    0.1827680    0.1946707  0.1830769
#>  [705,]    0.1658189    0.1766177  0.1660991
#>  [706,]    0.1309618    0.1394907  0.1311832
#>  [707,]    0.1812676    0.1930724  0.1815739
#>  [708,]    0.1892411    0.2015652  0.1895609
#>  [709,]    0.1687476    0.1797372  0.1690328
#>  [710,]    0.2998409    0.3193671  0.3003476
#>  [711,]    0.2044578    0.2177730  0.2048034
#>  [712,]    0.1838327    0.1958046  0.1841433
#>  [713,]    0.1953395    0.2080609  0.1956697
#>  [714,]    0.2936821    0.3128076  0.2941785
#>  [715,]    0.2370908    0.2525308  0.2374914
#>  [716,]    0.2930360    0.3121189  0.2935312
#>  [717,]    0.1949872    0.2076856  0.1953168
#>  [718,]    0.1840010    0.1959840  0.1843120
#>  [719,]    0.2020610    0.2152202  0.2024025
#>  [720,]    0.2565488    0.2732559  0.2569824
#>  [721,]    0.1759565    0.1874157  0.1762539
#>  [722,]    0.1993953    0.2123807  0.1997323
#>  [723,]    0.2201404    0.2344768  0.2205125
#>  [724,]    0.2345195    0.2497919  0.2349158
#>  [725,]    0.1880380    0.2002838  0.1883558
#>  [726,]    0.2337825    0.2490071  0.2341776
#>  [727,]    0.2608539    0.2778413  0.2612948
#>  [728,]    0.1886746    0.2009618  0.1889935
#>  [729,]    0.2016570    0.2147897  0.2019978
#>  [730,]    0.3459777    0.3685082  0.3465624
#>  [731,]    0.2450992    0.2610607  0.2455134
#>  [732,]    0.2629323    0.2800550  0.2633766
#>  [733,]    0.2121831    0.2260014  0.2125417
#>  [734,]    0.1762141    0.1876898  0.1765119
#>  [735,]    0.1799031    0.1916191  0.1802071
#>  [736,]    0.1912985    0.2037565  0.1916218
#>  [737,]    0.1448747    0.1543097  0.1451196
#>  [738,]    0.1771166    0.1886512  0.1774159
#>  [739,]    0.1840065    0.1959897  0.1843174
#>  [740,]    0.1986479    0.2115845  0.1989836
#>  [741,]    0.1960894    0.2088595  0.1964208
#>  [742,]    0.2201865    0.2345258  0.2205586
#>  [743,]    0.1639072    0.1745816  0.1641842
#>  [744,]    0.2439764    0.2598650  0.2443887
#>  [745,]    0.2128174    0.2266769  0.2131771
#>  [746,]    0.1957838    0.2085339  0.1961147
#>  [747,]    0.2837152    0.3021915  0.2841947
#>  [748,]    0.2948294    0.3140291  0.2953277
#>  [749,]    0.1760824    0.1875496  0.1763800
#>  [750,]    0.2476025    0.2637270  0.2480209
#>  [751,]    0.1828688    0.1947780  0.1831779
#>  [752,]    0.2297966    0.2447617  0.2301849
#>  [753,]    0.1532716    0.1632534  0.1535307
#>  [754,]    0.2197871    0.2341004  0.2201586
#>  [755,]    0.1465257    0.1560681  0.1467733
#>  [756,]    0.1619376    0.1724838  0.1622113
#>  [757,]    0.2186910    0.2329330  0.2190606
#>  [758,]    0.1736784    0.1849891  0.1739719
#>  [759,]    0.1840939    0.1960828  0.1844050
#>  [760,]    0.2656785    0.2829801  0.2661275
#>  [761,]    0.1656390    0.1764261  0.1659189
#>  [762,]    0.1806325    0.1923961  0.1809378
#>  [763,]    0.2397818    0.2553971  0.2401870
#>  [764,]    0.2309846    0.2460271  0.2313750
#>  [765,]    0.1875176    0.1997296  0.1878345
#>  [766,]    0.2304974    0.2455081  0.2308869
#>  [767,]    0.2293284    0.2442631  0.2297159
#>  [768,]    0.1657396    0.1765333  0.1660197
#>  [769,]    0.1816875    0.1935197  0.1819945
#>  [770,]    0.2365619    0.2519675  0.2369617
#>  [771,]    0.2003760    0.2134252  0.2007146
#>  [772,]    0.3135846    0.3340057  0.3141146
#>  [773,]    0.1786700    0.1903058  0.1789720
#>  [774,]    0.2077519    0.2212814  0.2081030
#>  [775,]    0.1839267    0.1959048  0.1842375
#>  [776,]    0.2141148    0.2280585  0.2144767
#>  [777,]    0.2156523    0.2296964  0.2160167
#>  [778,]    0.2203199    0.2346679  0.2206922
#>  [779,]    0.1618321    0.1723714  0.1621056
#>  [780,]    0.2437165    0.2595879  0.2441284
#>  [781,]    0.1914540    0.2039223  0.1917775
#>  [782,]    0.2010861    0.2141815  0.2014259
#>  [783,]    0.2301890    0.2451796  0.2305780
#>  [784,]    0.3084310    0.3285167  0.3089522
#>  [785,]    0.1763660    0.1878517  0.1766640
#>  [786,]    0.2023831    0.2155630  0.2027251
#>  [787,]    0.3193917    0.3401908  0.3199314
#>  [788,]    0.2169605    0.2310897  0.2173272
#>  [789,]    0.2148603    0.2288529  0.2152234
#>  [790,]    0.2723817    0.2901199  0.2728420
#>  [791,]    0.1684873    0.1794599  0.1687721
#>  [792,]    0.1845187    0.1965353  0.1848305
#>  [793,]    0.1516897    0.1615685  0.1519460
#>  [794,]    0.2665086    0.2838644  0.2669590
#>  [795,]    0.2158200    0.2298751  0.2161847
#>  [796,]    0.2926504    0.3117082  0.2931449
#>  [797,]    0.1442796    0.1536759  0.1445235
#>  [798,]    0.1817737    0.1936116  0.1820809
#>  [799,]    0.1937045    0.2063193  0.1940318
#>  [800,]    0.2243314    0.2389406  0.2247105
#>  [801,]    0.2209525    0.2353417  0.2213259
#>  [802,]    0.2129073    0.2267726  0.2132671
#>  [803,]    0.2465127    0.2625663  0.2469293
#>  [804,]    0.1860139    0.1981279  0.1863283
#>  [805,]    0.1902934    0.2026860  0.1906150
#>  [806,]    0.1731545    0.1844311  0.1734471
#>  [807,]    0.2007786    0.2138541  0.2011179
#>  [808,]    0.2057308    0.2191287  0.2060785
#>  [809,]    0.1629221    0.1735323  0.1631974
#>  [810,]    0.2205181    0.2348787  0.2208907
#>  [811,]    0.2353082    0.2506322  0.2357059
#>  [812,]    0.1935427    0.2061470  0.1938698
#>  [813,]    0.1862233    0.1983509  0.1865380
#>  [814,]    0.1987616    0.2117058  0.1990975
#>  [815,]    0.2516105    0.2679962  0.2520357
#>  [816,]    0.1379927    0.1469794  0.1382259
#>  [817,]    0.2145457    0.2285178  0.2149083
#>  [818,]    0.1697183    0.1807710  0.1700051
#>  [819,]    0.2483950    0.2645714  0.2488148
#>  [820,]    0.1898997    0.2022667  0.1902206
#>  [821,]    0.2007353    0.2138079  0.2010745
#>  [822,]    0.2154724    0.2295048  0.2158366
#>  [823,]    0.1705881    0.1816974  0.1708764
#>  [824,]    0.2432382    0.2590787  0.2436492
#>  [825,]    0.2642256    0.2814326  0.2646721
#>  [826,]    0.3202291    0.3410830  0.3207702
#>  [827,]    0.1831074    0.1950321  0.1834169
#>  [828,]    0.2195789    0.2338786  0.2199500
#>  [829,]    0.1847193    0.1967490  0.1850314
#>  [830,]    0.2158068    0.2298609  0.2161716
#>  [831,]    0.2325357    0.2476792  0.2329287
#>  [832,]    0.2414555    0.2571798  0.2418636
#>  [833,]    0.1957442    0.2084918  0.1960750
#>  [834,]    0.1786751    0.1903112  0.1789771
#>  [835,]    0.1429457    0.1522549  0.1431872
#>  [836,]    0.1527786    0.1627283  0.1530368
#>  [837,]    0.1526716    0.1626142  0.1529296
#>  [838,]    0.1801408    0.1918722  0.1804453
#>  [839,]    0.2189621    0.2332218  0.2193322
#>  [840,]    0.2092968    0.2229270  0.2096505
#>  [841,]    0.1979187    0.2108079  0.1982532
#>  [842,]    0.2423521    0.2581348  0.2427616
#>  [843,]    0.2133724    0.2272681  0.2137330
#>  [844,]    0.2700982    0.2876876  0.2705547
#>  [845,]    0.2940417    0.3131904  0.2945386
#>  [846,]    0.1611562    0.1716514  0.1614286
#>  [847,]    0.2085950    0.2221795  0.2089476
#>  [848,]    0.2043770    0.2176869  0.2047224
#>  [849,]    0.2534477    0.2699530  0.2538760
#>  [850,]    0.2439895    0.2598789  0.2444018
#>  [851,]    0.2049135    0.2182583  0.2052598
#>  [852,]    0.2045895    0.2179132  0.2049353
#>  [853,]    0.1794334    0.1911188  0.1797367
#>  [854,]    0.2138348    0.2277606  0.2141961
#>  [855,]    0.1897313    0.2020874  0.1900520
#>  [856,]    0.2050079    0.2183587  0.2053544
#>  [857,]    0.2856750    0.3042787  0.2861578
#>  [858,]    0.2195157    0.2338113  0.2198867
#>  [859,]    0.3188063    0.3395676  0.3193450
#>  [860,]    0.1312146    0.1397600  0.1314363
#>  [861,]    0.2020070    0.2151625  0.2023484
#>  [862,]    0.1920402    0.2045466  0.1923648
#>  [863,]    0.1538472    0.1638666  0.1541072
#>  [864,]    0.2520384    0.2684520  0.2524643
#>  [865,]    0.2822497    0.3006306  0.2827267
#>  [866,]    0.1577630    0.1680373  0.1580296
#>  [867,]    0.2685798    0.2860703  0.2690337
#>  [868,]    0.1833148    0.1952529  0.1836246
#>  [869,]    0.1831398    0.1950667  0.1834493
#>  [870,]    0.1846027    0.1966248  0.1849146
#>  [871,]    0.1640689    0.1747536  0.1643461
#>  [872,]    0.2085878    0.2221718  0.2089403
#>  [873,]    0.2170568    0.2311923  0.2174237
#>  [874,]    0.2232986    0.2378406  0.2236760
#>  [875,]    0.1453328    0.1547975  0.1455784
#>  [876,]    0.2999582    0.3194920  0.3004651
#>  [877,]    0.2041457    0.2174405  0.2044907
#>  [878,]    0.2280804    0.2429337  0.2284658
#>  [879,]    0.2001222    0.2131551  0.2004604
#>  [880,]    0.2521944    0.2686179  0.2526206
#>  [881,]    0.2409872    0.2566807  0.2413945
#>  [882,]    0.2246994    0.2393326  0.2250791
#>  [883,]    0.2292730    0.2442040  0.2296605
#>  [884,]    0.2268122    0.2415829  0.2271955
#>  [885,]    0.1824797    0.1943635  0.1827881
#>  [886,]    0.2183651    0.2325859  0.2187342
#>  [887,]    0.2177622    0.2319437  0.2181302
#>  [888,]    0.1951237    0.2078309  0.1954534
#>  [889,]    0.1938664    0.2064918  0.1941941
#>  [890,]    0.2058034    0.2192060  0.2061512
#>  [891,]    0.2331880    0.2483741  0.2335821
#>  [892,]    0.1778810    0.1894655  0.1781816
#>  [893,]    0.2122943    0.2261197  0.2126530
#>  [894,]    0.1699005    0.1809653  0.1701876
#>  [895,]    0.2608001    0.2777839  0.2612408
#>  [896,]    0.2787745    0.2969290  0.2792456
#>  [897,]    0.2398460    0.2554656  0.2402514
#>  [898,]    0.1453316    0.1547964  0.1455772
#>  [899,]    0.1729658    0.1842302  0.1732582
#>  [900,]    0.1648643    0.1756010  0.1651430
#>  [901,]    0.1797306    0.1914354  0.1800343
#>  [902,]    0.1724346    0.1836643  0.1727260
#>  [903,]    0.2234924    0.2380469  0.2238701
#>  [904,]    0.2296164    0.2445697  0.2300044
#>  [905,]    0.2459718    0.2619905  0.2463875
#>  [906,]    0.1744480    0.1858088  0.1747429
#>  [907,]    0.2638025    0.2809819  0.2642483
#>  [908,]    0.2588710    0.2757295  0.2593085
#>  [909,]    0.1742463    0.1855939  0.1745408
#>  [910,]    0.1744841    0.1858472  0.1747790
#>  [911,]    0.2081054    0.2216580  0.2084571
#>  [912,]    0.1940815    0.2067207  0.1944095
#>  [913,]    0.3320567    0.3536809  0.3326179
#>  [914,]    0.1623702    0.1729445  0.1626446
#>  [915,]    0.1325802    0.1412145  0.1328043
#>  [916,]    0.2304337    0.2454404  0.2308231
#>  [917,]    0.1789288    0.1905812  0.1792311
#>  [918,]    0.1962649    0.2090464  0.1965966
#>  [919,]    0.2211033    0.2355019  0.2214769
#>  [920,]    0.1758477    0.1872997  0.1761449
#>  [921,]    0.2211366    0.2355378  0.2215103
#>  [922,]    0.2248683    0.2395122  0.2252483
#>  [923,]    0.1770250    0.1885537  0.1773242
#>  [924,]    0.3181357    0.3388533  0.3186733
#>  [925,]    0.2395846    0.2551871  0.2399895
#>  [926,]    0.2589176    0.2757793  0.2593552
#>  [927,]    0.2125713    0.2264147  0.2129306
#>  [928,]    0.1904513    0.2028542  0.1907732
#>  [929,]    0.2438417    0.2597215  0.2442538
#>  [930,]    0.2595875    0.2764924  0.2600262
#>  [931,]    0.2155953    0.2296358  0.2159597
#>  [932,]    0.2208447    0.2352269  0.2212179
#>  [933,]    0.2053333    0.2187055  0.2056804
#>  [934,]    0.1759615    0.1874210  0.1762589
#>  [935,]    0.2153379    0.2293615  0.2157018
#>  [936,]    0.3445387    0.3669752  0.3451209
#>  [937,]    0.2010504    0.2141436  0.2013902
#>  [938,]    0.1918509    0.2043449  0.1921752
#>  [939,]    0.1853653    0.1974370  0.1856785
#>  [940,]    0.2115812    0.2253601  0.2119388
#>  [941,]    0.2209229    0.2353101  0.2212962
#>  [942,]    0.1974482    0.2103068  0.1977819
#>  [943,]    0.2050457    0.2183991  0.2053922
#>  [944,]    0.1964102    0.2092010  0.1967421
#>  [945,]    0.1955887    0.2083262  0.1959192
#>  [946,]    0.2182520    0.2324653  0.2186209
#>  [947,]    0.2232884    0.2378298  0.2236658
#>  [948,]    0.2080828    0.2216338  0.2084344
#>  [949,]    0.2453487    0.2613265  0.2457634
#>  [950,]    0.1832447    0.1951784  0.1835544
#>  [951,]    0.1669269    0.1777979  0.1672090
#>  [952,]    0.1747683    0.1861499  0.1750637
#>  [953,]    0.1828545    0.1947627  0.1831635
#>  [954,]    0.2040891    0.2173801  0.2044340
#>  [955,]    0.3010913    0.3206990  0.3016001
#>  [956,]    0.2199519    0.2342761  0.2203237
#>  [957,]    0.2008984    0.2139817  0.2012379
#>  [958,]    0.2060684    0.2194883  0.2064167
#>  [959,]    0.1535640    0.1635647  0.1538235
#>  [960,]    0.2157309    0.2297799  0.2160955
#>  [961,]    0.2545251    0.2711006  0.2549552
#>  [962,]    0.2100339    0.2237119  0.2103888
#>  [963,]    0.1898368    0.2021997  0.1901576
#>  [964,]    0.2592180    0.2760987  0.2596560
#>  [965,]    0.1954185    0.2081450  0.1957488
#>  [966,]    0.2734993    0.2913101  0.2739615
#>  [967,]    0.2170889    0.2312265  0.2174557
#>  [968,]    0.3202334    0.3410869  0.3207746
#>  [969,]    0.2312842    0.2463461  0.2316751
#>  [970,]    0.2667822    0.2841558  0.2672330
#>  [971,]    0.1695786    0.1806222  0.1698652
#>  [972,]    0.2615296    0.2785614  0.2619716
#>  [973,]    0.1806739    0.1924401  0.1809792
#>  [974,]    0.1882936    0.2005561  0.1886118
#>  [975,]    0.2864303    0.3050834  0.2869143
#>  [976,]    0.1875199    0.1997319  0.1878368
#>  [977,]    0.2162500    0.2303330  0.2166154
#>  [978,]    0.2327008    0.2478551  0.2330941
#>  [979,]    0.2031340    0.2163629  0.2034773
#>  [980,]    0.1754858    0.1869142  0.1757824
#>  [981,]    0.1790519    0.1907126  0.1793545
#>  [982,]    0.2451212    0.2610842  0.2455354
#>  [983,]    0.3013620    0.3209873  0.3018712
#>  [984,]    0.2179665    0.2321611  0.2183348
#>  [985,]    0.2693343    0.2868740  0.2697895
#>  [986,]    0.1758235    0.1872739  0.1761207
#>  [987,]    0.3139283    0.3343716  0.3144588
#>  [988,]    0.1775446    0.1891069  0.1778446
#>  [989,]    0.1591225    0.1694853  0.1593915
#>  [990,]    0.2025930    0.2157866  0.2029353
#>  [991,]    0.2026197    0.2158151  0.2029621
#>  [992,]    0.1761664    0.1876391  0.1764641
#>  [993,]    0.2891571    0.3079877  0.2896458
#>  [994,]    0.3266831    0.3479572  0.3272352
#>  [995,]    0.2196960    0.2340033  0.2200673
#>  [996,]    0.2164129    0.2305065  0.2167786
#>  [997,]    0.1846353    0.1966594  0.1849473
#>  [998,]    0.2104729    0.2241795  0.2108286
#>  [999,]    0.2498361    0.2661061  0.2502583
#> [1000,]    0.1926799    0.2052281  0.1930056