Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 90.66125
#> 2   Feb 89.23688
#> 3   Mar 94.03480
#> 4   Apr 86.78777
#> 5   May 97.23104
#> 6   Jun 88.30479

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 9.629770  2
#> 2   Feb 9.564116  2
#> 3   Mar 8.023473  2
#> 4   Apr 8.886288  2
#> 5   May 6.352511  2
#> 6   Jun 6.394890  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [12ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [23ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [67ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [739ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [19ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1963512    0.2028188  0.2010527
#>    [2,]    0.2030587    0.2097473  0.2079209
#>    [3,]    0.2261305    0.2335790  0.2315450
#>    [4,]    0.1901723    0.1964364  0.1947259
#>    [5,]    0.2077240    0.2145662  0.2126978
#>    [6,]    0.2724629    0.2814373  0.2789867
#>    [7,]    0.2255377    0.2329667  0.2309381
#>    [8,]    0.1752936    0.1810677  0.1794910
#>    [9,]    0.2469184    0.2550515  0.2528306
#>   [10,]    0.3818530    0.3944301  0.3909957
#>   [11,]    0.3036187    0.3136192  0.3108884
#>   [12,]    0.2375318    0.2453558  0.2432193
#>   [13,]    0.1954090    0.2018457  0.2000880
#>   [14,]    0.1820968    0.1880950  0.1864571
#>   [15,]    0.2669238    0.2757158  0.2733150
#>   [16,]    0.2180637    0.2252465  0.2232851
#>   [17,]    0.2468476    0.2549784  0.2527581
#>   [18,]    0.2346015    0.2423290  0.2402188
#>   [19,]    0.1467186    0.1515515  0.1502318
#>   [20,]    0.2513130    0.2595909  0.2573305
#>   [21,]    0.1794636    0.1853750  0.1837608
#>   [22,]    0.2471659    0.2553072  0.2530841
#>   [23,]    0.2115926    0.2185623  0.2166591
#>   [24,]    0.2048225    0.2115692  0.2097269
#>   [25,]    0.2265950    0.2340588  0.2320206
#>   [26,]    0.1881957    0.1943948  0.1927020
#>   [27,]    0.2732263    0.2822259  0.2797684
#>   [28,]    0.2392287    0.2471087  0.2449569
#>   [29,]    0.2031850    0.2098778  0.2080502
#>   [30,]    0.1713958    0.1770414  0.1754997
#>   [31,]    0.2566801    0.2651348  0.2628261
#>   [32,]    0.2430342    0.2510395  0.2488535
#>   [33,]    0.2469567    0.2550910  0.2528698
#>   [34,]    0.3117587    0.3220274  0.3192234
#>   [35,]    0.2055913    0.2123632  0.2105140
#>   [36,]    0.2190828    0.2262992  0.2243286
#>   [37,]    0.2222100    0.2295294  0.2275307
#>   [38,]    0.1500739    0.1550172  0.1536673
#>   [39,]    0.2313675    0.2389885  0.2369075
#>   [40,]    0.3235742    0.3342319  0.3313216
#>   [41,]    0.2251949    0.2326127  0.2305871
#>   [42,]    0.2002757    0.2068726  0.2050712
#>   [43,]    0.3526128    0.3642270  0.3610555
#>   [44,]    0.1872949    0.1934642  0.1917796
#>   [45,]    0.2526083    0.2609290  0.2586569
#>   [46,]    0.1479963    0.1528712  0.1515400
#>   [47,]    0.2927226    0.3023643  0.2997315
#>   [48,]    0.1760177    0.1818156  0.1802323
#>   [49,]    0.2169784    0.2241254  0.2221738
#>   [50,]    0.1944093    0.2008129  0.1990643
#>   [51,]    0.2027965    0.2094765  0.2076524
#>   [52,]    0.2209286    0.2282057  0.2262186
#>   [53,]    0.2241622    0.2315458  0.2295296
#>   [54,]    0.2615037    0.2701172  0.2677651
#>   [55,]    0.2660835    0.2748479  0.2724546
#>   [56,]    0.2924587    0.3020917  0.2994612
#>   [57,]    0.2602585    0.2688310  0.2664901
#>   [58,]    0.1836671    0.1897170  0.1880649
#>   [59,]    0.2389449    0.2468154  0.2446662
#>   [60,]    0.1719976    0.1776631  0.1761160
#>   [61,]    0.2596972    0.2682512  0.2659154
#>   [62,]    0.2023097    0.2089736  0.2071539
#>   [63,]    0.1689713    0.1745372  0.1730173
#>   [64,]    0.2495075    0.2577260  0.2554818
#>   [65,]    0.1952509    0.2016823  0.1999261
#>   [66,]    0.2411051    0.2490468  0.2468781
#>   [67,]    0.2328655    0.2405359  0.2384413
#>   [68,]    0.2335484    0.2412412  0.2391405
#>   [69,]    0.1961788    0.2026408  0.2008762
#>   [70,]    0.2539811    0.2623469  0.2600624
#>   [71,]    0.1861534    0.1922852  0.1906108
#>   [72,]    0.1866550    0.1928032  0.1911243
#>   [73,]    0.2321898    0.2398379  0.2377495
#>   [74,]    0.2320341    0.2396771  0.2375900
#>   [75,]    0.2736124    0.2826248  0.2801638
#>   [76,]    0.1944764    0.2008823  0.1991331
#>   [77,]    0.2507935    0.2590543  0.2567985
#>   [78,]    0.1865042    0.1926476  0.1909700
#>   [79,]    0.2144380    0.2215014  0.2195726
#>   [80,]    0.1978635    0.2043809  0.2026012
#>   [81,]    0.1894127    0.1956519  0.1939482
#>   [82,]    0.2839407    0.2932932  0.2907393
#>   [83,]    0.3129711    0.3232799  0.3204649
#>   [84,]    0.1862538    0.1923889  0.1907136
#>   [85,]    0.1877774    0.1939627  0.1922737
#>   [86,]    0.1774911    0.1833376  0.1817411
#>   [87,]    0.2422428    0.2502219  0.2480431
#>   [88,]    0.1918804    0.1982008  0.1964749
#>   [89,]    0.2983186    0.3081448  0.3054616
#>   [90,]    0.1412711    0.1459246  0.1446539
#>   [91,]    0.3100072    0.3202183  0.3174300
#>   [92,]    0.1547518    0.1598492  0.1584573
#>   [93,]    0.2269088    0.2343829  0.2323419
#>   [94,]    0.2436846    0.2517111  0.2495193
#>   [95,]    0.2463876    0.2545033  0.2522872
#>   [96,]    0.2239739    0.2313513  0.2293367
#>   [97,]    0.1976095    0.2041186  0.2023411
#>   [98,]    0.2447310    0.2527922  0.2505909
#>   [99,]    0.2038448    0.2105593  0.2087257
#>  [100,]    0.2177126    0.2248839  0.2229256
#>  [101,]    0.2307873    0.2383891  0.2363133
#>  [102,]    0.1895798    0.1958245  0.1941193
#>  [103,]    0.2113988    0.2183619  0.2164605
#>  [104,]    0.1278374    0.1320483  0.1308984
#>  [105,]    0.2611675    0.2697701  0.2674210
#>  [106,]    0.2196819    0.2269179  0.2249420
#>  [107,]    0.1868655    0.1930207  0.1913399
#>  [108,]    0.1142404    0.1180035  0.1169759
#>  [109,]    0.2583145    0.2668231  0.2644997
#>  [110,]    0.2219652    0.2292764  0.2272799
#>  [111,]    0.2911589    0.3007492  0.2981304
#>  [112,]    0.3508898    0.3624472  0.3592912
#>  [113,]    0.2115003    0.2184669  0.2165646
#>  [114,]    0.2111316    0.2180861  0.2161871
#>  [115,]    0.1983705    0.2049047  0.2031204
#>  [116,]    0.1421629    0.1468457  0.1455670
#>  [117,]    0.1914796    0.1977868  0.1960645
#>  [118,]    0.1690250    0.1745926  0.1730723
#>  [119,]    0.1978004    0.2043157  0.2025366
#>  [120,]    0.1806968    0.1866489  0.1850236
#>  [121,]    0.2159912    0.2231058  0.2211630
#>  [122,]    0.1762872    0.1820940  0.1805083
#>  [123,]    0.1574293    0.1626149  0.1611989
#>  [124,]    0.2591988    0.2677362  0.2654049
#>  [125,]    0.3103195    0.3205409  0.3177497
#>  [126,]    0.2599791    0.2685425  0.2662041
#>  [127,]    0.2193504    0.2265756  0.2246026
#>  [128,]    0.2730299    0.2820231  0.2795673
#>  [129,]    0.5044527    0.5210677  0.5165307
#>  [130,]    0.2237000    0.2310685  0.2290564
#>  [131,]    0.2163865    0.2235140  0.2215677
#>  [132,]    0.2148382    0.2219147  0.2199823
#>  [133,]    0.1805397    0.1864866  0.1848627
#>  [134,]    0.1537430    0.1588072  0.1574243
#>  [135,]    0.2341088    0.2418201  0.2397144
#>  [136,]    0.2184934    0.2256904  0.2237251
#>  [137,]    0.2642447    0.2729486  0.2705718
#>  [138,]    0.2178400    0.2250155  0.2230561
#>  [139,]    0.1720827    0.1777510  0.1762032
#>  [140,]    0.1583634    0.1635798  0.1621554
#>  [141,]    0.2173454    0.2245045  0.2225496
#>  [142,]    0.1521164    0.1571270  0.1557588
#>  [143,]    0.1491922    0.1541065  0.1527646
#>  [144,]    0.1613358    0.1666502  0.1651990
#>  [145,]    0.2981533    0.3079739  0.3052922
#>  [146,]    0.2500123    0.2582472  0.2559985
#>  [147,]    0.2441204    0.2521613  0.2499656
#>  [148,]    0.1815930    0.1875746  0.1859412
#>  [149,]    0.1990251    0.2055809  0.2037907
#>  [150,]    0.2880337    0.2975211  0.2949304
#>  [151,]    0.2267447    0.2342134  0.2321739
#>  [152,]    0.1776307    0.1834817  0.1818840
#>  [153,]    0.2080609    0.2149142  0.2130428
#>  [154,]    0.2843425    0.2937083  0.2911508
#>  [155,]    0.2096543    0.2165602  0.2146744
#>  [156,]    0.3379433    0.3490744  0.3460348
#>  [157,]    0.1412638    0.1459169  0.1446463
#>  [158,]    0.2116494    0.2186210  0.2167173
#>  [159,]    0.1980544    0.2045782  0.2027968
#>  [160,]    0.2456520    0.2537435  0.2515339
#>  [161,]    0.1564749    0.1616292  0.1602217
#>  [162,]    0.2552520    0.2636597  0.2613638
#>  [163,]    0.3187386    0.3292372  0.3263704
#>  [164,]    0.2059300    0.2127132  0.2108609
#>  [165,]    0.1852915    0.1913949  0.1897282
#>  [166,]    0.3131408    0.3234550  0.3206385
#>  [167,]    0.2397885    0.2476869  0.2455301
#>  [168,]    0.2186422    0.2258441  0.2238775
#>  [169,]    0.2425701    0.2505600  0.2483782
#>  [170,]    0.1989525    0.2055058  0.2037163
#>  [171,]    0.2841826    0.2935432  0.2909871
#>  [172,]    0.1850640    0.1911599  0.1894953
#>  [173,]    0.2315231    0.2391492  0.2370667
#>  [174,]    0.2347376    0.2424695  0.2403582
#>  [175,]    0.2430181    0.2510228  0.2488369
#>  [176,]    0.2703661    0.2792716  0.2768398
#>  [177,]    0.2323259    0.2399785  0.2378888
#>  [178,]    0.2991136    0.3089658  0.3062755
#>  [179,]    0.2060831    0.2128713  0.2110176
#>  [180,]    0.1968596    0.2033441  0.2015734
#>  [181,]    0.1908071    0.1970921  0.1953759
#>  [182,]    0.1971215    0.2036145  0.2018415
#>  [183,]    0.1990563    0.2056130  0.2038226
#>  [184,]    0.1379057    0.1424483  0.1412079
#>  [185,]    0.2204499    0.2277112  0.2257284
#>  [186,]    0.2640880    0.2727866  0.2704113
#>  [187,]    0.1860822    0.1922115  0.1905378
#>  [188,]    0.2152346    0.2223242  0.2203883
#>  [189,]    0.2412243    0.2491699  0.2470002
#>  [190,]    0.2129928    0.2200085  0.2180928
#>  [191,]    0.1840285    0.1900903  0.1884350
#>  [192,]    0.2081071    0.2149619  0.2130901
#>  [193,]    0.1800912    0.1860234  0.1844035
#>  [194,]    0.1906537    0.1969336  0.1952188
#>  [195,]    0.2084134    0.2152783  0.2134037
#>  [196,]    0.2263583    0.2338143  0.2317783
#>  [197,]    0.2045519    0.2112897  0.2094498
#>  [198,]    0.2803524    0.2895868  0.2870652
#>  [199,]    0.2110029    0.2179532  0.2160553
#>  [200,]    0.2087846    0.2156618  0.2137839
#>  [201,]    0.2098704    0.2167833  0.2148956
#>  [202,]    0.1740923    0.1798267  0.1782608
#>  [203,]    0.2667283    0.2755140  0.2731149
#>  [204,]    0.1609980    0.1663011  0.1648530
#>  [205,]    0.2662163    0.2749852  0.2725907
#>  [206,]    0.2629880    0.2716505  0.2692850
#>  [207,]    0.2620180    0.2706484  0.2682917
#>  [208,]    0.2834170    0.2927524  0.2902032
#>  [209,]    0.2032318    0.2099261  0.2080981
#>  [210,]    0.2171869    0.2243409  0.2223873
#>  [211,]    0.2158643    0.2229747  0.2210330
#>  [212,]    0.2157008    0.2228058  0.2208657
#>  [213,]    0.1717287    0.1773852  0.1758406
#>  [214,]    0.1819508    0.1879442  0.1863076
#>  [215,]    0.2368008    0.2446009  0.2424709
#>  [216,]    0.2313558    0.2389763  0.2368954
#>  [217,]    0.2087512    0.2156272  0.2137496
#>  [218,]    0.2317718    0.2394061  0.2373214
#>  [219,]    0.2012924    0.2079228  0.2061122
#>  [220,]    0.2399676    0.2478718  0.2457134
#>  [221,]    0.2058332    0.2126131  0.2107617
#>  [222,]    0.2174474    0.2246099  0.2226541
#>  [223,]    0.2896869    0.2992288  0.2966232
#>  [224,]    0.2092425    0.2161347  0.2142527
#>  [225,]    0.2114526    0.2184177  0.2165158
#>  [226,]    0.2081172    0.2149724  0.2131005
#>  [227,]    0.2064659    0.2132667  0.2114096
#>  [228,]    0.2522285    0.2605366  0.2582679
#>  [229,]    0.2179412    0.2251200  0.2231597
#>  [230,]    0.1760452    0.1818440  0.1802606
#>  [231,]    0.2700964    0.2789930  0.2765636
#>  [232,]    0.2287030    0.2362360  0.2341790
#>  [233,]    0.1999344    0.2065201  0.2047218
#>  [234,]    0.2726227    0.2816025  0.2791504
#>  [235,]    0.2659403    0.2747001  0.2723081
#>  [236,]    0.2533928    0.2617392  0.2594601
#>  [237,]    0.3436779    0.3549979  0.3519068
#>  [238,]    0.2663778    0.2751519  0.2727560
#>  [239,]    0.3061962    0.3162818  0.3135277
#>  [240,]    0.2980199    0.3078361  0.3051556
#>  [241,]    0.2095864    0.2164900  0.2146049
#>  [242,]    0.1867174    0.1928677  0.1911883
#>  [243,]    0.2725546    0.2815321  0.2790806
#>  [244,]    0.2837583    0.2931047  0.2905525
#>  [245,]    0.1648999    0.1703316  0.1688483
#>  [246,]    0.2988285    0.3086714  0.3059836
#>  [247,]    0.2762642    0.2853639  0.2828790
#>  [248,]    0.1935566    0.1999322  0.1981913
#>  [249,]    0.2075434    0.2143798  0.2125130
#>  [250,]    0.2229518    0.2302956  0.2282903
#>  [251,]    0.2389634    0.2468345  0.2446852
#>  [252,]    0.1702214    0.1758284  0.1742973
#>  [253,]    0.1967374    0.2032177  0.2014482
#>  [254,]    0.2480950    0.2562669  0.2540354
#>  [255,]    0.2692815    0.2781511  0.2757291
#>  [256,]    0.1730409    0.1787407  0.1771843
#>  [257,]    0.2090528    0.2159388  0.2140585
#>  [258,]    0.1947317    0.2011461  0.1993945
#>  [259,]    0.2521442    0.2604495  0.2581816
#>  [260,]    0.2246449    0.2320445  0.2300239
#>  [261,]    0.2318110    0.2394467  0.2373616
#>  [262,]    0.3055802    0.3156454  0.3128969
#>  [263,]    0.3018123    0.3117533  0.3090387
#>  [264,]    0.1680555    0.1735911  0.1720795
#>  [265,]    0.1493646    0.1542846  0.1529411
#>  [266,]    0.2151507    0.2222375  0.2203023
#>  [267,]    0.1664493    0.1719321  0.1704349
#>  [268,]    0.2309017    0.2385074  0.2364305
#>  [269,]    0.2690163    0.2778773  0.2754576
#>  [270,]    0.1765731    0.1823893  0.1808011
#>  [271,]    0.1956150    0.2020584  0.2002989
#>  [272,]    0.2980531    0.3078705  0.3051897
#>  [273,]    0.1739898    0.1797209  0.1781559
#>  [274,]    0.1861623    0.1922944  0.1906199
#>  [275,]    0.2474205    0.2555703  0.2533448
#>  [276,]    0.2780046    0.2871616  0.2846611
#>  [277,]    0.2944519    0.3041505  0.3015021
#>  [278,]    0.1836307    0.1896794  0.1880277
#>  [279,]    0.1920476    0.1983735  0.1966461
#>  [280,]    0.2792018    0.2883982  0.2858870
#>  [281,]    0.1838330    0.1898884  0.1882348
#>  [282,]    0.1876383    0.1938189  0.1921312
#>  [283,]    0.1929183    0.1992729  0.1975377
#>  [284,]    0.2610856    0.2696853  0.2673370
#>  [285,]    0.2857139    0.2951248  0.2925550
#>  [286,]    0.1748656    0.1806255  0.1790527
#>  [287,]    0.2084447    0.2153106  0.2134357
#>  [288,]    0.2280507    0.2355625  0.2335112
#>  [289,]    0.2384892    0.2463447  0.2441996
#>  [290,]    0.2421686    0.2501453  0.2479671
#>  [291,]    0.1986180    0.2051603  0.2033738
#>  [292,]    0.2173171    0.2244753  0.2225206
#>  [293,]    0.2124923    0.2194916  0.2175803
#>  [294,]    0.2159361    0.2230488  0.2211065
#>  [295,]    0.2051902    0.2119490  0.2101034
#>  [296,]    0.2044505    0.2111850  0.2093460
#>  [297,]    0.2658617    0.2746188  0.2722275
#>  [298,]    0.2089868    0.2158706  0.2139908
#>  [299,]    0.2161665    0.2232867  0.2213424
#>  [300,]    0.2435854    0.2516087  0.2494178
#>  [301,]    0.3133842    0.3237064  0.3208878
#>  [302,]    0.1451694    0.1499512  0.1486455
#>  [303,]    0.2751494    0.2842125  0.2817376
#>  [304,]    0.2263737    0.2338301  0.2317940
#>  [305,]    0.1926736    0.1990202  0.1972872
#>  [306,]    0.2007043    0.2073153  0.2055100
#>  [307,]    0.2875928    0.2970655  0.2944788
#>  [308,]    0.2605698    0.2691526  0.2668089
#>  [309,]    0.2379121    0.2457487  0.2436088
#>  [310,]    0.1946939    0.2011070  0.1993558
#>  [311,]    0.2959766    0.3057255  0.3030634
#>  [312,]    0.1931891    0.1995526  0.1978149
#>  [313,]    0.2493911    0.2576057  0.2553626
#>  [314,]    0.2046290    0.2113693  0.2095287
#>  [315,]    0.1640128    0.1694153  0.1679401
#>  [316,]    0.1838629    0.1899192  0.1882654
#>  [317,]    0.2046976    0.2114402  0.2095990
#>  [318,]    0.1672282    0.1727366  0.1712324
#>  [319,]    0.1210348    0.1250216  0.1239329
#>  [320,]    0.2097027    0.2166101  0.2147239
#>  [321,]    0.2012958    0.2079262  0.2061157
#>  [322,]    0.2644634    0.2731744  0.2707957
#>  [323,]    0.2312564    0.2388737  0.2367937
#>  [324,]    0.2087482    0.2156242  0.2137466
#>  [325,]    0.2193494    0.2265745  0.2246015
#>  [326,]    0.1915841    0.1978947  0.1961715
#>  [327,]    0.1647902    0.1702183  0.1687360
#>  [328,]    0.2313794    0.2390008  0.2369196
#>  [329,]    0.2269702    0.2344464  0.2324048
#>  [330,]    0.1747983    0.1805560  0.1789837
#>  [331,]    0.1680516    0.1735871  0.1720755
#>  [332,]    0.2208278    0.2281017  0.2261154
#>  [333,]    0.3001147    0.3100000  0.3073006
#>  [334,]    0.2316497    0.2392800  0.2371964
#>  [335,]    0.2040997    0.2108226  0.2089868
#>  [336,]    0.2524268    0.2607414  0.2584710
#>  [337,]    0.1513263    0.1563109  0.1549497
#>  [338,]    0.1889915    0.1952167  0.1935168
#>  [339,]    0.1946130    0.2010234  0.1992729
#>  [340,]    0.2155291    0.2226285  0.2206899
#>  [341,]    0.1637492    0.1691430  0.1676701
#>  [342,]    0.2229424    0.2302858  0.2282805
#>  [343,]    0.2504116    0.2586598  0.2564075
#>  [344,]    0.2562449    0.2646853  0.2623805
#>  [345,]    0.2339760    0.2416830  0.2395784
#>  [346,]    0.2321301    0.2397762  0.2376883
#>  [347,]    0.1990462    0.2056027  0.2038123
#>  [348,]    0.2037632    0.2104750  0.2086422
#>  [349,]    0.2465451    0.2546660  0.2524484
#>  [350,]    0.2380567    0.2458980  0.2437568
#>  [351,]    0.1717254    0.1773819  0.1758373
#>  [352,]    0.1954002    0.2018364  0.2000789
#>  [353,]    0.2354947    0.2432516  0.2411334
#>  [354,]    0.2878453    0.2973265  0.2947374
#>  [355,]    0.2477319    0.2558920  0.2536637
#>  [356,]    0.1725612    0.1782453  0.1766931
#>  [357,]    0.2154267    0.2225226  0.2205849
#>  [358,]    0.2012901    0.2079204  0.2061099
#>  [359,]    0.1723889    0.1780672  0.1765167
#>  [360,]    0.1415455    0.1462080  0.1449348
#>  [361,]    0.2348984    0.2426356  0.2405228
#>  [362,]    0.2445798    0.2526360  0.2504361
#>  [363,]    0.2459877    0.2540903  0.2518777
#>  [364,]    0.2137463    0.2207869  0.2188644
#>  [365,]    0.2130374    0.2200546  0.2181384
#>  [366,]    0.2593165    0.2678580  0.2655256
#>  [367,]    0.1454827    0.1502749  0.1489663
#>  [368,]    0.2202105    0.2274640  0.2254833
#>  [369,]    0.1585659    0.1637889  0.1623627
#>  [370,]    0.2151823    0.2222703  0.2203348
#>  [371,]    0.2813571    0.2906245  0.2880939
#>  [372,]    0.1328342    0.1372097  0.1360149
#>  [373,]    0.2616875    0.2703071  0.2679533
#>  [374,]    0.1976043    0.2041132  0.2023359
#>  [375,]    0.2081875    0.2150451  0.2131725
#>  [376,]    0.2449839    0.2530533  0.2508498
#>  [377,]    0.2568144    0.2652735  0.2629636
#>  [378,]    0.1557611    0.1608918  0.1594908
#>  [379,]    0.2230876    0.2304358  0.2284292
#>  [380,]    0.2260210    0.2334659  0.2314329
#>  [381,]    0.2708028    0.2797227  0.2772869
#>  [382,]    0.1482915    0.1531762  0.1518423
#>  [383,]    0.2073807    0.2142117  0.2123464
#>  [384,]    0.2100799    0.2169996  0.2151100
#>  [385,]    0.3214819    0.3320708  0.3291793
#>  [386,]    0.1571731    0.1623503  0.1609366
#>  [387,]    0.3323789    0.3433267  0.3403372
#>  [388,]    0.2115610    0.2185296  0.2166267
#>  [389,]    0.2154042    0.2224993  0.2205619
#>  [390,]    0.3058992    0.3159751  0.3132237
#>  [391,]    0.2163251    0.2234506  0.2215048
#>  [392,]    0.2919945    0.3016124  0.2989860
#>  [393,]    0.2067501    0.2135602  0.2117006
#>  [394,]    0.2243464    0.2317362  0.2297183
#>  [395,]    0.1744978    0.1802456  0.1786760
#>  [396,]    0.3047062    0.3147426  0.3120019
#>  [397,]    0.2003608    0.2069605  0.2051583
#>  [398,]    0.1905623    0.1968393  0.1951253
#>  [399,]    0.2093603    0.2162565  0.2143733
#>  [400,]    0.1801867    0.1861220  0.1845013
#>  [401,]    0.2031205    0.2098110  0.2079840
#>  [402,]    0.1921118    0.1984398  0.1967118
#>  [403,]    0.1784893    0.1843686  0.1827632
#>  [404,]    0.2768955    0.2860160  0.2835255
#>  [405,]    0.1987823    0.2053301  0.2035421
#>  [406,]    0.1711075    0.1767436  0.1752046
#>  [407,]    0.2199088    0.2271523  0.2251744
#>  [408,]    0.1877372    0.1939212  0.1922325
#>  [409,]    0.2072726    0.2141000  0.2122357
#>  [410,]    0.1711937    0.1768327  0.1752928
#>  [411,]    0.1901088    0.1963709  0.1946609
#>  [412,]    0.2095009    0.2164017  0.2145173
#>  [413,]    0.2209569    0.2282350  0.2262476
#>  [414,]    0.1800576    0.1859885  0.1843690
#>  [415,]    0.3170810    0.3275251  0.3246731
#>  [416,]    0.1448271    0.1495976  0.1482949
#>  [417,]    0.1632778    0.1686561  0.1671875
#>  [418,]    0.1884496    0.1946570  0.1929619
#>  [419,]    0.2021180    0.2087756  0.2069576
#>  [420,]    0.1743845    0.1801286  0.1785601
#>  [421,]    0.2281273    0.2356416  0.2335897
#>  [422,]    0.2480033    0.2561722  0.2539415
#>  [423,]    0.2454487    0.2535335  0.2513258
#>  [424,]    0.2054402    0.2122072  0.2103594
#>  [425,]    0.2227928    0.2301314  0.2281274
#>  [426,]    0.2231604    0.2305111  0.2285038
#>  [427,]    0.1793264    0.1852332  0.1836202
#>  [428,]    0.1573364    0.1625190  0.1611038
#>  [429,]    0.2617236    0.2703444  0.2679903
#>  [430,]    0.1550725    0.1601805  0.1587856
#>  [431,]    0.3097279    0.3199297  0.3171439
#>  [432,]    0.2180970    0.2252809  0.2233192
#>  [433,]    0.2968452    0.3066228  0.3039528
#>  [434,]    0.2657298    0.2744825  0.2720924
#>  [435,]    0.2936450    0.3033171  0.3006760
#>  [436,]    0.2001148    0.2067065  0.2049065
#>  [437,]    0.2754464    0.2845192  0.2820417
#>  [438,]    0.2170070    0.2241550  0.2222031
#>  [439,]    0.2144821    0.2215470  0.2196178
#>  [440,]    0.2079921    0.2148432  0.2129724
#>  [441,]    0.2167100    0.2238482  0.2218990
#>  [442,]    0.2146709    0.2217419  0.2198110
#>  [443,]    0.2859043    0.2953216  0.2927500
#>  [444,]    0.1729631    0.1786604  0.1771046
#>  [445,]    0.2137653    0.2208065  0.2188838
#>  [446,]    0.2164375    0.2235668  0.2216200
#>  [447,]    0.1994252    0.2059940  0.2042003
#>  [448,]    0.3159263    0.3263323  0.3234908
#>  [449,]    0.3270351    0.3378069  0.3348655
#>  [450,]    0.1768772    0.1827034  0.1811124
#>  [451,]    0.1845096    0.1905873  0.1889277
#>  [452,]    0.1789353    0.1848293  0.1832198
#>  [453,]    0.1899108    0.1961663  0.1944581
#>  [454,]    0.1743996    0.1801443  0.1785756
#>  [455,]    0.2134138    0.2204434  0.2185239
#>  [456,]    0.2106690    0.2176082  0.2157134
#>  [457,]    0.1665074    0.1719921  0.1704944
#>  [458,]    0.2182941    0.2254845  0.2235210
#>  [459,]    0.2491455    0.2573519  0.2551110
#>  [460,]    0.2377805    0.2456128  0.2434740
#>  [461,]    0.2191609    0.2263798  0.2244086
#>  [462,]    0.1605559    0.1658445  0.1644003
#>  [463,]    0.1675981    0.1731187  0.1716112
#>  [464,]    0.2047167    0.2114599  0.2096185
#>  [465,]    0.2064930    0.2132948  0.2114374
#>  [466,]    0.2612711    0.2698770  0.2675270
#>  [467,]    0.2076949    0.2145361  0.2126680
#>  [468,]    0.1848900    0.1909801  0.1893171
#>  [469,]    0.1868259    0.1929798  0.1912993
#>  [470,]    0.2977823    0.3075908  0.3049124
#>  [471,]    0.2478939    0.2560593  0.2538296
#>  [472,]    0.1307137    0.1350194  0.1338436
#>  [473,]    0.1913533    0.1976563  0.1959351
#>  [474,]    0.1618289    0.1671594  0.1657038
#>  [475,]    0.2335313    0.2412236  0.2391231
#>  [476,]    0.1965810    0.2030562  0.2012880
#>  [477,]    0.2451046    0.2531780  0.2509734
#>  [478,]    0.2078083    0.2146533  0.2127842
#>  [479,]    0.1963621    0.2028301  0.2010639
#>  [480,]    0.3391661    0.3503375  0.3472870
#>  [481,]    0.1825154    0.1885273  0.1868857
#>  [482,]    0.1891198    0.1953493  0.1936482
#>  [483,]    0.2398083    0.2477073  0.2455503
#>  [484,]    0.1569254    0.1620944  0.1606829
#>  [485,]    0.1331611    0.1375474  0.1363496
#>  [486,]    0.2006385    0.2072474  0.2054427
#>  [487,]    0.1537630    0.1588278  0.1574448
#>  [488,]    0.2126515    0.2196560  0.2177433
#>  [489,]    0.2752632    0.2843300  0.2818542
#>  [490,]    0.2024227    0.2090903  0.2072696
#>  [491,]    0.1773586    0.1832007  0.1816054
#>  [492,]    0.2350644    0.2428072  0.2406929
#>  [493,]    0.1861337    0.1922648  0.1905906
#>  [494,]    0.2269525    0.2344281  0.2323868
#>  [495,]    0.2736355    0.2826486  0.2801874
#>  [496,]    0.1911128    0.1974079  0.1956889
#>  [497,]    0.2249582    0.2323680  0.2303446
#>  [498,]    0.2336324    0.2413279  0.2392265
#>  [499,]    0.2484055    0.2565877  0.2543534
#>  [500,]    0.1874982    0.1936743  0.1919878
#>  [501,]    0.2745076    0.2835494  0.2810804
#>  [502,]    0.1991399    0.2056994  0.2039082
#>  [503,]    0.2316527    0.2392830  0.2371994
#>  [504,]    0.1735680    0.1792852  0.1777240
#>  [505,]    0.1528241    0.1578580  0.1564834
#>  [506,]    0.1870625    0.1932241  0.1915416
#>  [507,]    0.2384018    0.2462544  0.2441101
#>  [508,]    0.1686971    0.1742538  0.1727364
#>  [509,]    0.1825557    0.1885690  0.1869270
#>  [510,]    0.2425751    0.2505652  0.2483833
#>  [511,]    0.2310524    0.2386631  0.2365849
#>  [512,]    0.1824733    0.1884839  0.1868426
#>  [513,]    0.2121933    0.2191827  0.2172741
#>  [514,]    0.2020001    0.2086537  0.2068368
#>  [515,]    0.1705393    0.1761567  0.1746228
#>  [516,]    0.2784323    0.2876034  0.2850991
#>  [517,]    0.2210592    0.2283407  0.2263523
#>  [518,]    0.1686588    0.1742143  0.1726973
#>  [519,]    0.2049782    0.2117300  0.2098863
#>  [520,]    0.1607556    0.1660508  0.1646049
#>  [521,]    0.3026788    0.3126484  0.3099260
#>  [522,]    0.2480159    0.2561853  0.2539545
#>  [523,]    0.2065461    0.2133496  0.2114918
#>  [524,]    0.2099895    0.2169063  0.2150175
#>  [525,]    0.2698808    0.2787704  0.2763429
#>  [526,]    0.1973794    0.2038810  0.2021056
#>  [527,]    0.2184021    0.2255960  0.2236316
#>  [528,]    0.2070986    0.2139203  0.2120575
#>  [529,]    0.1445237    0.1492842  0.1479842
#>  [530,]    0.2162582    0.2233815  0.2214364
#>  [531,]    0.1921176    0.1984458  0.1967178
#>  [532,]    0.1956045    0.2020475  0.2002881
#>  [533,]    0.1835463    0.1895922  0.1879412
#>  [534,]    0.2262857    0.2337392  0.2317039
#>  [535,]    0.2262728    0.2337260  0.2316908
#>  [536,]    0.1941152    0.2005091  0.1987631
#>  [537,]    0.1959858    0.2024414  0.2006785
#>  [538,]    0.2207871    0.2280596  0.2260737
#>  [539,]    0.1916470    0.1979597  0.1962359
#>  [540,]    0.2007731    0.2073863  0.2055805
#>  [541,]    0.1749858    0.1807498  0.1791758
#>  [542,]    0.2484580    0.2566418  0.2544071
#>  [543,]    0.2512513    0.2595272  0.2572673
#>  [544,]    0.2491231    0.2573289  0.2550882
#>  [545,]    0.2245689    0.2319660  0.2299461
#>  [546,]    0.1968002    0.2032825  0.2015124
#>  [547,]    0.2511312    0.2594031  0.2571443
#>  [548,]    0.1878244    0.1940112  0.1923218
#>  [549,]    0.2444159    0.2524666  0.2502682
#>  [550,]    0.2752138    0.2842789  0.2818035
#>  [551,]    0.3146724    0.3250370  0.3222068
#>  [552,]    0.3084296    0.3185886  0.3158145
#>  [553,]    0.1743821    0.1801262  0.1785576
#>  [554,]    0.1870721    0.1932341  0.1915514
#>  [555,]    0.2346613    0.2423907  0.2402800
#>  [556,]    0.2151726    0.2222602  0.2203248
#>  [557,]    0.2446605    0.2527193  0.2505187
#>  [558,]    0.2406299    0.2485558  0.2463915
#>  [559,]    0.2676431    0.2764589  0.2740516
#>  [560,]    0.1988219    0.2053709  0.2035825
#>  [561,]    0.2142091    0.2212650  0.2193383
#>  [562,]    0.1871704    0.1933357  0.1916521
#>  [563,]    0.1686907    0.1742473  0.1727300
#>  [564,]    0.1836242    0.1896727  0.1880211
#>  [565,]    0.2292715    0.2368235  0.2347613
#>  [566,]    0.1568635    0.1620305  0.1606195
#>  [567,]    0.2009138    0.2075318  0.2057246
#>  [568,]    0.2442792    0.2523253  0.2501282
#>  [569,]    0.2781725    0.2873350  0.2848330
#>  [570,]    0.1614953    0.1668150  0.1653624
#>  [571,]    0.2255415    0.2329706  0.2309420
#>  [572,]    0.2099029    0.2168168  0.2149288
#>  [573,]    0.1878694    0.1940576  0.1923678
#>  [574,]    0.2580223    0.2665212  0.2642004
#>  [575,]    0.2587436    0.2672663  0.2649390
#>  [576,]    0.2017914    0.2084382  0.2066232
#>  [577,]    0.2319048    0.2395435  0.2374576
#>  [578,]    0.2093606    0.2162567  0.2143736
#>  [579,]    0.2321095    0.2397548  0.2376671
#>  [580,]    0.2104914    0.2174248  0.2155315
#>  [581,]    0.1821432    0.1881428  0.1865045
#>  [582,]    0.2073120    0.2141406  0.2122759
#>  [583,]    0.2235716    0.2309357  0.2289248
#>  [584,]    0.2249016    0.2323096  0.2302867
#>  [585,]    0.1874673    0.1936423  0.1919561
#>  [586,]    0.2277912    0.2352944  0.2332455
#>  [587,]    0.2544005    0.2627800  0.2604918
#>  [588,]    0.1941958    0.2005924  0.1988457
#>  [589,]    0.2523071    0.2606178  0.2583484
#>  [590,]    0.1584453    0.1636644  0.1622392
#>  [591,]    0.2190387    0.2262537  0.2242835
#>  [592,]    0.1754882    0.1812686  0.1796901
#>  [593,]    0.2671195    0.2759179  0.2735154
#>  [594,]    0.1625890    0.1679446  0.1664821
#>  [595,]    0.1887507    0.1949680  0.1932702
#>  [596,]    0.1981347    0.2046610  0.2028789
#>  [597,]    0.2088787    0.2157590  0.2138802
#>  [598,]    0.2484230    0.2566058  0.2543713
#>  [599,]    0.1844064    0.1904806  0.1888219
#>  [600,]    0.1519152    0.1569192  0.1555528
#>  [601,]    0.1451126    0.1498925  0.1485872
#>  [602,]    0.1960281    0.2024851  0.2007219
#>  [603,]    0.2614648    0.2700770  0.2677253
#>  [604,]    0.1928497    0.1992020  0.1974674
#>  [605,]    0.1939805    0.2003701  0.1986253
#>  [606,]    0.1921595    0.1984891  0.1967607
#>  [607,]    0.2075535    0.2143902  0.2125233
#>  [608,]    0.2041561    0.2108807  0.2090445
#>  [609,]    0.2139580    0.2210056  0.2190811
#>  [610,]    0.1899215    0.1961774  0.1944691
#>  [611,]    0.1691803    0.1747530  0.1732312
#>  [612,]    0.1522639    0.1572794  0.1559098
#>  [613,]    0.2020260    0.2086806  0.2068634
#>  [614,]    0.2383686    0.2462202  0.2440762
#>  [615,]    0.2215894    0.2288883  0.2268952
#>  [616,]    0.1622567    0.1676013  0.1661419
#>  [617,]    0.1722095    0.1778820  0.1763331
#>  [618,]    0.1942704    0.2006695  0.1989221
#>  [619,]    0.1939519    0.2003405  0.1985960
#>  [620,]    0.2043860    0.2111184  0.2092800
#>  [621,]    0.2154687    0.2225660  0.2206280
#>  [622,]    0.2266243    0.2340891  0.2320507
#>  [623,]    0.2122479    0.2192391  0.2173300
#>  [624,]    0.1954836    0.2019227  0.2001644
#>  [625,]    0.2065969    0.2134020  0.2115437
#>  [626,]    0.2303412    0.2379284  0.2358565
#>  [627,]    0.2644899    0.2732019  0.2708229
#>  [628,]    0.1748727    0.1806329  0.1790600
#>  [629,]    0.2143325    0.2213924  0.2194646
#>  [630,]    0.2710363    0.2799638  0.2775260
#>  [631,]    0.2076659    0.2145061  0.2126383
#>  [632,]    0.2648144    0.2735370  0.2711551
#>  [633,]    0.2653640    0.2741047  0.2717179
#>  [634,]    0.1952124    0.2016425  0.1998866
#>  [635,]    0.1728876    0.1785824  0.1770273
#>  [636,]    0.2026523    0.2093276  0.2075047
#>  [637,]    0.2139939    0.2210427  0.2191179
#>  [638,]    0.1954494    0.2018874  0.2001294
#>  [639,]    0.2414476    0.2494006  0.2472289
#>  [640,]    0.1607504    0.1660454  0.1645995
#>  [641,]    0.2914210    0.3010200  0.2983988
#>  [642,]    0.1678975    0.1734280  0.1719178
#>  [643,]    0.1903165    0.1965854  0.1948736
#>  [644,]    0.2868421    0.2962902  0.2937102
#>  [645,]    0.2603739    0.2689502  0.2666083
#>  [646,]    0.2781739    0.2873365  0.2848344
#>  [647,]    0.1282572    0.1324819  0.1313282
#>  [648,]    0.2302603    0.2378448  0.2357737
#>  [649,]    0.1583584    0.1635747  0.1621503
#>  [650,]    0.2195240    0.2267549  0.2247803
#>  [651,]    0.1663497    0.1718292  0.1703329
#>  [652,]    0.2274423    0.2349340  0.2328882
#>  [653,]    0.2068510    0.2136645  0.2118039
#>  [654,]    0.1894659    0.1957067  0.1940025
#>  [655,]    0.1932219    0.1995865  0.1978485
#>  [656,]    0.2395151    0.2474044  0.2452501
#>  [657,]    0.2011227    0.2077475  0.2059385
#>  [658,]    0.2202920    0.2275482  0.2255667
#>  [659,]    0.1987360    0.2052822  0.2034947
#>  [660,]    0.2125121    0.2195120  0.2176006
#>  [661,]    0.1813537    0.1873274  0.1856962
#>  [662,]    0.2649977    0.2737262  0.2713427
#>  [663,]    0.1722968    0.1779721  0.1764223
#>  [664,]    0.2090833    0.2159703  0.2140897
#>  [665,]    0.1608885    0.1661880  0.1647409
#>  [666,]    0.2181807    0.2253673  0.2234049
#>  [667,]    0.2411577    0.2491012  0.2469320
#>  [668,]    0.1660589    0.1715288  0.1700351
#>  [669,]    0.2231729    0.2305240  0.2285166
#>  [670,]    0.2681163    0.2769475  0.2745360
#>  [671,]    0.2235320    0.2308950  0.2288844
#>  [672,]    0.3141842    0.3245329  0.3217070
#>  [673,]    0.2226806    0.2300156  0.2280126
#>  [674,]    0.2807638    0.2900116  0.2874863
#>  [675,]    0.2390624    0.2469369  0.2447866
#>  [676,]    0.2062429    0.2130362  0.2111812
#>  [677,]    0.2209926    0.2282718  0.2262841
#>  [678,]    0.2403894    0.2483074  0.2461452
#>  [679,]    0.1892148    0.1954474  0.1937454
#>  [680,]    0.1964298    0.2029000  0.2011332
#>  [681,]    0.2281732    0.2356890  0.2336367
#>  [682,]    0.1881670    0.1943651  0.1926726
#>  [683,]    0.2071313    0.2139540  0.2120909
#>  [684,]    0.1746998    0.1804543  0.1788829
#>  [685,]    0.2141975    0.2212530  0.2193263
#>  [686,]    0.2366542    0.2444493  0.2423207
#>  [687,]    0.2119966    0.2189796  0.2170727
#>  [688,]    0.1872809    0.1934498  0.1917653
#>  [689,]    0.2440693    0.2521086  0.2499133
#>  [690,]    0.1812814    0.1872527  0.1856222
#>  [691,]    0.1553924    0.1605110  0.1591133
#>  [692,]    0.2323445    0.2399977  0.2379079
#>  [693,]    0.2469563    0.2550907  0.2528695
#>  [694,]    0.2646477    0.2733648  0.2709844
#>  [695,]    0.2019570    0.2086093  0.2067928
#>  [696,]    0.1835169    0.1895618  0.1879112
#>  [697,]    0.1662172    0.1716924  0.1701973
#>  [698,]    0.3048712    0.3149132  0.3121710
#>  [699,]    0.1860791    0.1922084  0.1905347
#>  [700,]    0.1932698    0.1996360  0.1978976
#>  [701,]    0.1800186    0.1859483  0.1843291
#>  [702,]    0.2401008    0.2480094  0.2458498
#>  [703,]    0.1704949    0.1761109  0.1745774
#>  [704,]    0.1850032    0.1910971  0.1894331
#>  [705,]    0.1678468    0.1733756  0.1718658
#>  [706,]    0.1325635    0.1369301  0.1357377
#>  [707,]    0.1834844    0.1895282  0.1878778
#>  [708,]    0.1915555    0.1978651  0.1961421
#>  [709,]    0.1708113    0.1764378  0.1749014
#>  [710,]    0.3035078    0.3135047  0.3107749
#>  [711,]    0.2069583    0.2137754  0.2119138
#>  [712,]    0.1860809    0.1922103  0.1905365
#>  [713,]    0.1977285    0.2042415  0.2024630
#>  [714,]    0.2972737    0.3070655  0.3043917
#>  [715,]    0.2399903    0.2478952  0.2457366
#>  [716,]    0.2966196    0.3063896  0.3037218
#>  [717,]    0.1973719    0.2038732  0.2020979
#>  [718,]    0.1862513    0.1923863  0.1907110
#>  [719,]    0.2045322    0.2112694  0.2094297
#>  [720,]    0.2596863    0.2682399  0.2659042
#>  [721,]    0.1781084    0.1839753  0.1823732
#>  [722,]    0.2018338    0.2084821  0.2066667
#>  [723,]    0.2228327    0.2301726  0.2281683
#>  [724,]    0.2373875    0.2452067  0.2430715
#>  [725,]    0.1903377    0.1966073  0.1948953
#>  [726,]    0.2366415    0.2444362  0.2423077
#>  [727,]    0.2640440    0.2727412  0.2703663
#>  [728,]    0.1909821    0.1972728  0.1955550
#>  [729,]    0.2041232    0.2108469  0.2090108
#>  [730,]    0.3502087    0.3617439  0.3585940
#>  [731,]    0.2480966    0.2562685  0.2540370
#>  [732,]    0.2661478    0.2749142  0.2725204
#>  [733,]    0.2147781    0.2218527  0.2199209
#>  [734,]    0.1783691    0.1842444  0.1826401
#>  [735,]    0.1821032    0.1881016  0.1864636
#>  [736,]    0.1936380    0.2000162  0.1982745
#>  [737,]    0.1466465    0.1514770  0.1501580
#>  [738,]    0.1792827    0.1851881  0.1835755
#>  [739,]    0.1862568    0.1923920  0.1907167
#>  [740,]    0.2010773    0.2077005  0.2058919
#>  [741,]    0.1984875    0.2050255  0.2032402
#>  [742,]    0.2228793    0.2302207  0.2282160
#>  [743,]    0.1659118    0.1713768  0.1698845
#>  [744,]    0.2469601    0.2550947  0.2528734
#>  [745,]    0.2154201    0.2225158  0.2205782
#>  [746,]    0.1981782    0.2047059  0.2029234
#>  [747,]    0.2871849    0.2966443  0.2940612
#>  [748,]    0.2984350    0.3082648  0.3055806
#>  [749,]    0.1782359    0.1841068  0.1825036
#>  [750,]    0.2506305    0.2588859  0.2566316
#>  [751,]    0.1851053    0.1912025  0.1895375
#>  [752,]    0.2326069    0.2402687  0.2381765
#>  [753,]    0.1551461    0.1602566  0.1588611
#>  [754,]    0.2224750    0.2298031  0.2278020
#>  [755,]    0.1483177    0.1532032  0.1518691
#>  [756,]    0.1639181    0.1693175  0.1678431
#>  [757,]    0.2213656    0.2286571  0.2266660
#>  [758,]    0.1758025    0.1815933  0.1800120
#>  [759,]    0.1863453    0.1924833  0.1908072
#>  [760,]    0.2689276    0.2777857  0.2753668
#>  [761,]    0.1676647    0.1731875  0.1716794
#>  [762,]    0.1828416    0.1888643  0.1872197
#>  [763,]    0.2427142    0.2507089  0.2485258
#>  [764,]    0.2338095    0.2415109  0.2394078
#>  [765,]    0.1898109    0.1960632  0.1943559
#>  [766,]    0.2333163    0.2410014  0.2389028
#>  [767,]    0.2321330    0.2397792  0.2376913
#>  [768,]    0.1677665    0.1732927  0.1717836
#>  [769,]    0.1839095    0.1899673  0.1883131
#>  [770,]    0.2394550    0.2473423  0.2451885
#>  [771,]    0.2028265    0.2095074  0.2076831
#>  [772,]    0.3174196    0.3278747  0.3250197
#>  [773,]    0.1808551    0.1868124  0.1851856
#>  [774,]    0.2102926    0.2172194  0.2153279
#>  [775,]    0.1861761    0.1923086  0.1906340
#>  [776,]    0.2167333    0.2238722  0.2219228
#>  [777,]    0.2182897    0.2254799  0.2235165
#>  [778,]    0.2230143    0.2303602  0.2283543
#>  [779,]    0.1638113    0.1692072  0.1677337
#>  [780,]    0.2466970    0.2548228  0.2526039
#>  [781,]    0.1937954    0.2001789  0.1984358
#>  [782,]    0.2035453    0.2102499  0.2084191
#>  [783,]    0.2330041    0.2406790  0.2385832
#>  [784,]    0.3122029    0.3224863  0.3196782
#>  [785,]    0.1785229    0.1844033  0.1827975
#>  [786,]    0.2048582    0.2116060  0.2097634
#>  [787,]    0.3232976    0.3339463  0.3310384
#>  [788,]    0.2196139    0.2268477  0.2248724
#>  [789,]    0.2174880    0.2246519  0.2226956
#>  [790,]    0.2757128    0.2847943  0.2823144
#>  [791,]    0.1705479    0.1761656  0.1746316
#>  [792,]    0.1867753    0.1929275  0.1912476
#>  [793,]    0.1535448    0.1586026  0.1572215
#>  [794,]    0.2697679    0.2786536  0.2762272
#>  [795,]    0.2184594    0.2256553  0.2236904
#>  [796,]    0.2962293    0.3059865  0.3033221
#>  [797,]    0.1460442    0.1508548  0.1495412
#>  [798,]    0.1839968    0.1900575  0.1884025
#>  [799,]    0.1960734    0.2025319  0.2007683
#>  [800,]    0.2270749    0.2345545  0.2325120
#>  [801,]    0.2236547    0.2310216  0.2290099
#>  [802,]    0.2155111    0.2226098  0.2206713
#>  [803,]    0.2495274    0.2577465  0.2555022
#>  [804,]    0.1882888    0.1944909  0.1927973
#>  [805,]    0.1926206    0.1989653  0.1972328
#>  [806,]    0.1752721    0.1810455  0.1794690
#>  [807,]    0.2032341    0.2099284  0.2081004
#>  [808,]    0.2082468    0.2151062  0.2132331
#>  [809,]    0.1649146    0.1703468  0.1688634
#>  [810,]    0.2232149    0.2305672  0.2285595
#>  [811,]    0.2381860    0.2460315  0.2438891
#>  [812,]    0.1959097    0.2023628  0.2006006
#>  [813,]    0.1885007    0.1947098  0.1930143
#>  [814,]    0.2011924    0.2078196  0.2060099
#>  [815,]    0.2546876    0.2630767  0.2607859
#>  [816,]    0.1396803    0.1442813  0.1430249
#>  [817,]    0.2171696    0.2243229  0.2223696
#>  [818,]    0.1717939    0.1774527  0.1759074
#>  [819,]    0.2514328    0.2597147  0.2574532
#>  [820,]    0.1922221    0.1985538  0.1968248
#>  [821,]    0.2031902    0.2098831  0.2080555
#>  [822,]    0.2181076    0.2252919  0.2233301
#>  [823,]    0.1726743    0.1783620  0.1768089
#>  [824,]    0.2462129    0.2543229  0.2521083
#>  [825,]    0.2674569    0.2762665  0.2738609
#>  [826,]    0.3241453    0.3348220  0.3319065
#>  [827,]    0.1853468    0.1914520  0.1897848
#>  [828,]    0.2222643    0.2295854  0.2275862
#>  [829,]    0.1869783    0.1931373  0.1914555
#>  [830,]    0.2184461    0.2256414  0.2236766
#>  [831,]    0.2353796    0.2431326  0.2410155
#>  [832,]    0.2444084    0.2524589  0.2502606
#>  [833,]    0.1981381    0.2046646  0.2028824
#>  [834,]    0.1808603    0.1868177  0.1851909
#>  [835,]    0.1446939    0.1494600  0.1481585
#>  [836,]    0.1546471    0.1597411  0.1583501
#>  [837,]    0.1545387    0.1596291  0.1582391
#>  [838,]    0.1823439    0.1883501  0.1867100
#>  [839,]    0.2216400    0.2289406  0.2269470
#>  [840,]    0.2118565    0.2188348  0.2169292
#>  [841,]    0.2003392    0.2069382  0.2051362
#>  [842,]    0.2453159    0.2533963  0.2511898
#>  [843,]    0.2159819    0.2230962  0.2211535
#>  [844,]    0.2734014    0.2824067  0.2799476
#>  [845,]    0.2976376    0.3074413  0.3047642
#>  [846,]    0.1631271    0.1685004  0.1670331
#>  [847,]    0.2111461    0.2181010  0.2162018
#>  [848,]    0.2068765    0.2136908  0.2118300
#>  [849,]    0.2565473    0.2649976  0.2626901
#>  [850,]    0.2469734    0.2551084  0.2528870
#>  [851,]    0.2074195    0.2142518  0.2123861
#>  [852,]    0.2070916    0.2139130  0.2120503
#>  [853,]    0.1816279    0.1876105  0.1859768
#>  [854,]    0.2164499    0.2235797  0.2216327
#>  [855,]    0.1920517    0.1983777  0.1966503
#>  [856,]    0.2075151    0.2143504  0.2124839
#>  [857,]    0.2891686    0.2986933  0.2960924
#>  [858,]    0.2222003    0.2295193  0.2275207
#>  [859,]    0.3227051    0.3333344  0.3304319
#>  [860,]    0.1328193    0.1371944  0.1359997
#>  [861,]    0.2044775    0.2112128  0.2093736
#>  [862,]    0.1943888    0.2007918  0.1990433
#>  [863,]    0.1557287    0.1608584  0.1594577
#>  [864,]    0.2551207    0.2635241  0.2612294
#>  [865,]    0.2857015    0.2951121  0.2925423
#>  [866,]    0.1596924    0.1649526  0.1635162
#>  [867,]    0.2718644    0.2808191  0.2783739
#>  [868,]    0.1855567    0.1916687  0.1899997
#>  [869,]    0.1853796    0.1914859  0.1898184
#>  [870,]    0.1868603    0.1930154  0.1913346
#>  [871,]    0.1660754    0.1715457  0.1700520
#>  [872,]    0.2111388    0.2180935  0.2161943
#>  [873,]    0.2197114    0.2269484  0.2249722
#>  [874,]    0.2260295    0.2334747  0.2314416
#>  [875,]    0.1471102    0.1519559  0.1506327
#>  [876,]    0.3036265    0.3136273  0.3108964
#>  [877,]    0.2066423    0.2134490  0.2115903
#>  [878,]    0.2308697    0.2384743  0.2363977
#>  [879,]    0.2025697    0.2092422  0.2074202
#>  [880,]    0.2552786    0.2636871  0.2613910
#>  [881,]    0.2439343    0.2519690  0.2497750
#>  [882,]    0.2274474    0.2349392  0.2328934
#>  [883,]    0.2320769    0.2397212  0.2376338
#>  [884,]    0.2295860    0.2371483  0.2350833
#>  [885,]    0.1847114    0.1907957  0.1891342
#>  [886,]    0.2210357    0.2283164  0.2263282
#>  [887,]    0.2204253    0.2276860  0.2257033
#>  [888,]    0.1975100    0.2040158  0.2022393
#>  [889,]    0.1962374    0.2027013  0.2009362
#>  [890,]    0.2083203    0.2151821  0.2133083
#>  [891,]    0.2360399    0.2438148  0.2416917
#>  [892,]    0.1800565    0.1859874  0.1843679
#>  [893,]    0.2148906    0.2219689  0.2200360
#>  [894,]    0.1719784    0.1776433  0.1760964
#>  [895,]    0.2639895    0.2726849  0.2703104
#>  [896,]    0.2821838    0.2914785  0.2889404
#>  [897,]    0.2427793    0.2507761  0.2485924
#>  [898,]    0.1471090    0.1519548  0.1506316
#>  [899,]    0.1750812    0.1808483  0.1792735
#>  [900,]    0.1668806    0.1723775  0.1708765
#>  [901,]    0.1819286    0.1879213  0.1862849
#>  [902,]    0.1745435    0.1802928  0.1787228
#>  [903,]    0.2262256    0.2336772  0.2316424
#>  [904,]    0.2324245    0.2400803  0.2379897
#>  [905,]    0.2489800    0.2571812  0.2549417
#>  [906,]    0.1765815    0.1823980  0.1808097
#>  [907,]    0.2670286    0.2758241  0.2734223
#>  [908,]    0.2620369    0.2706680  0.2683111
#>  [909,]    0.1763773    0.1821870  0.1806006
#>  [910,]    0.1766180    0.1824356  0.1808470
#>  [911,]    0.2106504    0.2175891  0.2156944
#>  [912,]    0.1964550    0.2029260  0.2011590
#>  [913,]    0.3361176    0.3471887  0.3441655
#>  [914,]    0.1643559    0.1697698  0.1682914
#>  [915,]    0.1342017    0.1386222  0.1374151
#>  [916,]    0.2332518    0.2409349  0.2388369
#>  [917,]    0.1811170    0.1870828  0.1854537
#>  [918,]    0.1986652    0.2052090  0.2034221
#>  [919,]    0.2238072    0.2311790  0.2291660
#>  [920,]    0.1779983    0.1838615  0.1822604
#>  [921,]    0.2238410    0.2312141  0.2292008
#>  [922,]    0.2276183    0.2351156  0.2330683
#>  [923,]    0.1791900    0.1850925  0.1834807
#>  [924,]    0.3220263    0.3326332  0.3297368
#>  [925,]    0.2425146    0.2505028  0.2483215
#>  [926,]    0.2620841    0.2707169  0.2683596
#>  [927,]    0.2151710    0.2222585  0.2203231
#>  [928,]    0.1927805    0.1991305  0.1973965
#>  [929,]    0.2468238    0.2549539  0.2527338
#>  [930,]    0.2627621    0.2714170  0.2690536
#>  [931,]    0.2182320    0.2254204  0.2234575
#>  [932,]    0.2235455    0.2309089  0.2288982
#>  [933,]    0.2078445    0.2146908  0.2128213
#>  [934,]    0.1781135    0.1839805  0.1823784
#>  [935,]    0.2179714    0.2251512  0.2231906
#>  [936,]    0.3487521    0.3602390  0.3571023
#>  [937,]    0.2035092    0.2102126  0.2083821
#>  [938,]    0.1941972    0.2005939  0.1988471
#>  [939,]    0.1876322    0.1938127  0.1921250
#>  [940,]    0.2141688    0.2212233  0.2192969
#>  [941,]    0.2236247    0.2309906  0.2289792
#>  [942,]    0.1998629    0.2064462  0.2046485
#>  [943,]    0.2075533    0.2143900  0.2125231
#>  [944,]    0.1988122    0.2053608  0.2035726
#>  [945,]    0.1979807    0.2045020  0.2027212
#>  [946,]    0.2209212    0.2281981  0.2262110
#>  [947,]    0.2260192    0.2334640  0.2314311
#>  [948,]    0.2106276    0.2175654  0.2156709
#>  [949,]    0.2483492    0.2565295  0.2542957
#>  [950,]    0.1854858    0.1915956  0.1899272
#>  [951,]    0.1689684    0.1745341  0.1730143
#>  [952,]    0.1769057    0.1827328  0.1811416
#>  [953,]    0.1850908    0.1911875  0.1895227
#>  [954,]    0.2065850    0.2133898  0.2115316
#>  [955,]    0.3047734    0.3148121  0.3120709
#>  [956,]    0.2226419    0.2299755  0.2279730
#>  [957,]    0.2033553    0.2100537  0.2082246
#>  [958,]    0.2085886    0.2154592  0.2135831
#>  [959,]    0.1554420    0.1605622  0.1591640
#>  [960,]    0.2183692    0.2255620  0.2235979
#>  [961,]    0.2576378    0.2661241  0.2638068
#>  [962,]    0.2126025    0.2196053  0.2176931
#>  [963,]    0.1921585    0.1984880  0.1967596
#>  [964,]    0.2623880    0.2710306  0.2686706
#>  [965,]    0.1978085    0.2043241  0.2025449
#>  [966,]    0.2768440    0.2859628  0.2834727
#>  [967,]    0.2197438    0.2269820  0.2250054
#>  [968,]    0.3241496    0.3348261  0.3319106
#>  [969,]    0.2341127    0.2418241  0.2397183
#>  [970,]    0.2700448    0.2789397  0.2765108
#>  [971,]    0.1716525    0.1773066  0.1757626
#>  [972,]    0.2647280    0.2734479  0.2710668
#>  [973,]    0.1828835    0.1889075  0.1872626
#>  [974,]    0.1905964    0.1968745  0.1951601
#>  [975,]    0.2899332    0.2994832  0.2968754
#>  [976,]    0.1898132    0.1960655  0.1943582
#>  [977,]    0.2188947    0.2261048  0.2241360
#>  [978,]    0.2355467    0.2433054  0.2411867
#>  [979,]    0.2056183    0.2123912  0.2105417
#>  [980,]    0.1776320    0.1834831  0.1818853
#>  [981,]    0.1812416    0.1872117  0.1855814
#>  [982,]    0.2481189    0.2562916  0.2540599
#>  [983,]    0.3050474    0.3150952  0.3123514
#>  [984,]    0.2206321    0.2278994  0.2259149
#>  [985,]    0.2726281    0.2816080  0.2791559
#>  [986,]    0.1779738    0.1838362  0.1822353
#>  [987,]    0.3177674    0.3282339  0.3253758
#>  [988,]    0.1797159    0.1856355  0.1840191
#>  [989,]    0.1610686    0.1663741  0.1649253
#>  [990,]    0.2050706    0.2118254  0.2099809
#>  [991,]    0.2050977    0.2118534  0.2100086
#>  [992,]    0.1783209    0.1841946  0.1825907
#>  [993,]    0.2926933    0.3023342  0.2997016
#>  [994,]    0.3306782    0.3415701  0.3385958
#>  [995,]    0.2223828    0.2297078  0.2277076
#>  [996,]    0.2190596    0.2262752  0.2243048
#>  [997,]    0.1868933    0.1930493  0.1913683
#>  [998,]    0.2130469    0.2200644  0.2181481
#>  [999,]    0.2528915    0.2612214  0.2589467
#> [1000,]    0.1950364    0.2014607  0.1997064