Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 92.76043
#> 2   Feb 93.45533
#> 3   Mar 93.59607
#> 4   Apr 90.59806
#> 5   May 85.97617
#> 6   Jun 89.38952

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 8.437660  2
#> 2   Feb 8.843697  2
#> 3   Mar 9.923485  2
#> 4   Apr 8.607092  2
#> 5   May 8.915968  2
#> 6   Jun 7.579537  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [14ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [29ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [78ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [876ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [21ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.2000120    0.1946074  0.1967318
#>    [2,]    0.2068446    0.2012553  0.2034523
#>    [3,]    0.2303465    0.2241222  0.2265688
#>    [4,]    0.1937179    0.1884834  0.1905409
#>    [5,]    0.2115968    0.2058791  0.2081266
#>    [6,]    0.2775426    0.2700431  0.2729910
#>    [7,]    0.2297427    0.2235347  0.2259749
#>    [8,]    0.1785619    0.1737368  0.1756334
#>    [9,]    0.2515219    0.2447255  0.2473970
#>   [10,]    0.3889720    0.3784620  0.3825931
#>   [11,]    0.3092792    0.3009224  0.3042072
#>   [12,]    0.2419603    0.2354223  0.2379922
#>   [13,]    0.1990523    0.1936735  0.1957878
#>   [14,]    0.1854919    0.1804796  0.1824498
#>   [15,]    0.2719003    0.2645533  0.2674412
#>   [16,]    0.2221293    0.2161270  0.2184863
#>   [17,]    0.2514498    0.2446554  0.2473260
#>   [18,]    0.2389754    0.2325180  0.2350562
#>   [19,]    0.1494541    0.1454156  0.1470030
#>   [20,]    0.2559985    0.2490810  0.2518001
#>   [21,]    0.1828096    0.1778698  0.1798115
#>   [22,]    0.2517741    0.2449709  0.2476450
#>   [23,]    0.2155376    0.2097134  0.2120027
#>   [24,]    0.2086412    0.2030034  0.2052195
#>   [25,]    0.2308197    0.2245826  0.2270342
#>   [26,]    0.1917045    0.1865243  0.1885605
#>   [27,]    0.2783203    0.2707998  0.2737559
#>   [28,]    0.2436889    0.2371042  0.2396924
#>   [29,]    0.2069733    0.2013805  0.2035789
#>   [30,]    0.1745913    0.1698737  0.1717280
#>   [31,]    0.2614657    0.2544006  0.2571776
#>   [32,]    0.2475653    0.2408757  0.2435052
#>   [33,]    0.2515609    0.2447635  0.2474354
#>   [34,]    0.3175710    0.3089901  0.3123630
#>   [35,]    0.2094243    0.2037654  0.2059898
#>   [36,]    0.2231674    0.2171372  0.2195075
#>   [37,]    0.2263529    0.2202365  0.2226407
#>   [38,]    0.1528719    0.1487410  0.1503647
#>   [39,]    0.2356812    0.2293127  0.2318160
#>   [40,]    0.3296067    0.3207006  0.3242013
#>   [41,]    0.2293936    0.2231949  0.2256314
#>   [42,]    0.2040097    0.1984970  0.2006639
#>   [43,]    0.3591867    0.3494813  0.3532962
#>   [44,]    0.1907869    0.1856315  0.1876580
#>   [45,]    0.2573180    0.2503649  0.2530980
#>   [46,]    0.1507556    0.1466819  0.1482831
#>   [47,]    0.2981800    0.2901229  0.2932899
#>   [48,]    0.1792994    0.1744544  0.1763588
#>   [49,]    0.2210238    0.2150514  0.2173989
#>   [50,]    0.1980339    0.1926828  0.1947861
#>   [51,]    0.2065775    0.2009954  0.2031896
#>   [52,]    0.2250476    0.2189665  0.2213568
#>   [53,]    0.2283415    0.2221714  0.2245967
#>   [54,]    0.2663791    0.2591813  0.2620105
#>   [55,]    0.2710443    0.2637204  0.2665992
#>   [56,]    0.2979112    0.2898614  0.2930255
#>   [57,]    0.2651107    0.2579471  0.2607629
#>   [58,]    0.1870915    0.1820360  0.1840231
#>   [59,]    0.2433997    0.2368228  0.2394080
#>   [60,]    0.1752044    0.1704700  0.1723309
#>   [61,]    0.2645389    0.2573908  0.2602005
#>   [62,]    0.2060816    0.2005129  0.2027018
#>   [63,]    0.1721217    0.1674706  0.1692988
#>   [64,]    0.2541593    0.2472916  0.2499911
#>   [65,]    0.1988912    0.1935168  0.1956293
#>   [66,]    0.2456002    0.2389638  0.2415724
#>   [67,]    0.2372071    0.2307974  0.2333169
#>   [68,]    0.2379027    0.2314743  0.2340011
#>   [69,]    0.1998364    0.1944365  0.1965591
#>   [70,]    0.2587163    0.2517255  0.2544734
#>   [71,]    0.1896241    0.1845002  0.1865142
#>   [72,]    0.1901350    0.1849973  0.1870168
#>   [73,]    0.2365188    0.2301277  0.2326399
#>   [74,]    0.2363602    0.2299734  0.2324838
#>   [75,]    0.2787136    0.2711824  0.2741427
#>   [76,]    0.1981023    0.1927492  0.1948534
#>   [77,]    0.2554693    0.2485662  0.2512796
#>   [78,]    0.1899815    0.1848478  0.1868657
#>   [79,]    0.2184360    0.2125335  0.2148536
#>   [80,]    0.2015525    0.1961062  0.1982470
#>   [81,]    0.1929442    0.1877305  0.1897799
#>   [82,]    0.2892344    0.2814191  0.2844910
#>   [83,]    0.3188061    0.3101917  0.3135778
#>   [84,]    0.1897264    0.1845996  0.1866148
#>   [85,]    0.1912784    0.1861097  0.1881414
#>   [86,]    0.1808003    0.1759148  0.1778351
#>   [87,]    0.2467591    0.2400914  0.2427123
#>   [88,]    0.1954579    0.1901763  0.1922524
#>   [89,]    0.3038805    0.2956693  0.2988969
#>   [90,]    0.1439051    0.1400165  0.1415450
#>   [91,]    0.3157869    0.3072541  0.3106081
#>   [92,]    0.1576370    0.1533774  0.1550517
#>   [93,]    0.2311393    0.2248936  0.2273486
#>   [94,]    0.2482278    0.2415205  0.2441569
#>   [95,]    0.2509813    0.2441994  0.2468652
#>   [96,]    0.2281497    0.2219848  0.2244080
#>   [97,]    0.2012938    0.1958545  0.1979925
#>   [98,]    0.2492938    0.2425576  0.2452054
#>   [99,]    0.2076453    0.2020344  0.2042399
#>  [100,]    0.2217717    0.2157791  0.2181346
#>  [101,]    0.2350901    0.2287377  0.2312346
#>  [102,]    0.1931144    0.1878961  0.1899473
#>  [103,]    0.2153401    0.2095214  0.2118085
#>  [104,]    0.1302209    0.1267021  0.1280852
#>  [105,]    0.2660368    0.2588481  0.2616737
#>  [106,]    0.2237776    0.2177309  0.2201077
#>  [107,]    0.1903494    0.1852059  0.1872277
#>  [108,]    0.1163704    0.1132258  0.1144619
#>  [109,]    0.2631305    0.2560204  0.2588152
#>  [110,]    0.2261035    0.2199939  0.2223954
#>  [111,]    0.2965872    0.2885732  0.2917233
#>  [112,]    0.3574315    0.3477737  0.3515699
#>  [113,]    0.2154436    0.2096220  0.2119103
#>  [114,]    0.2150680    0.2092565  0.2115409
#>  [115,]    0.2020690    0.1966088  0.1987550
#>  [116,]    0.1448135    0.1409003  0.1424385
#>  [117,]    0.1950496    0.1897790  0.1918507
#>  [118,]    0.1721764    0.1675238  0.1693526
#>  [119,]    0.2014882    0.1960437  0.1981838
#>  [120,]    0.1840658    0.1790920  0.1810471
#>  [121,]    0.2200182    0.2140730  0.2164099
#>  [122,]    0.1795739    0.1747216  0.1766289
#>  [123,]    0.1603645    0.1560311  0.1577344
#>  [124,]    0.2640311    0.2568969  0.2597011
#>  [125,]    0.3161050    0.3075636  0.3109210
#>  [126,]    0.2648262    0.2576702  0.2604830
#>  [127,]    0.2234400    0.2174023  0.2197756
#>  [128,]    0.2781203    0.2706052  0.2735591
#>  [129,]    0.5138572    0.4999729  0.5054304
#>  [130,]    0.2278708    0.2217133  0.2241336
#>  [131,]    0.2204208    0.2144648  0.2168059
#>  [132,]    0.2188436    0.2129302  0.2152546
#>  [133,]    0.1839058    0.1789363  0.1808897
#>  [134,]    0.1566095    0.1523776  0.1540410
#>  [135,]    0.2384736    0.2320296  0.2345625
#>  [136,]    0.2225670    0.2165529  0.2189169
#>  [137,]    0.2691713    0.2618980  0.2647569
#>  [138,]    0.2219015    0.2159054  0.2182623
#>  [139,]    0.1752911    0.1705544  0.1724162
#>  [140,]    0.1613160    0.1569569  0.1586704
#>  [141,]    0.2213976    0.2154151  0.2177667
#>  [142,]    0.1549525    0.1507654  0.1524112
#>  [143,]    0.1519738    0.1478672  0.1494814
#>  [144,]    0.1643439    0.1599030  0.1616486
#>  [145,]    0.3037120    0.2955054  0.2987312
#>  [146,]    0.2546734    0.2477919  0.2504968
#>  [147,]    0.2486718    0.2419524  0.2445936
#>  [148,]    0.1849787    0.1799802  0.1819450
#>  [149,]    0.2027358    0.1972576  0.1994109
#>  [150,]    0.2934038    0.2854757  0.2885920
#>  [151,]    0.2309721    0.2247310  0.2271842
#>  [152,]    0.1809425    0.1760531  0.1779750
#>  [153,]    0.2119400    0.2062131  0.2084642
#>  [154,]    0.2896438    0.2818173  0.2848936
#>  [155,]    0.2135632    0.2077923  0.2100607
#>  [156,]    0.3442437    0.3349421  0.3385983
#>  [157,]    0.1438976    0.1400092  0.1415376
#>  [158,]    0.2155955    0.2097698  0.2120597
#>  [159,]    0.2017470    0.1962955  0.1984383
#>  [160,]    0.2502319    0.2434704  0.2461281
#>  [161,]    0.1593923    0.1550852  0.1567782
#>  [162,]    0.2600109    0.2529851  0.2557467
#>  [163,]    0.3246811    0.3159079  0.3193564
#>  [164,]    0.2097694    0.2041011  0.2063291
#>  [165,]    0.1887461    0.1836459  0.1856506
#>  [166,]    0.3189789    0.3103598  0.3137477
#>  [167,]    0.2442592    0.2376590  0.2402533
#>  [168,]    0.2227187    0.2167005  0.2190660
#>  [169,]    0.2470925    0.2404158  0.2430402
#>  [170,]    0.2026618    0.1971856  0.1993381
#>  [171,]    0.2894809    0.2816588  0.2847334
#>  [172,]    0.1885145    0.1834204  0.1854227
#>  [173,]    0.2358396    0.2294669  0.2319718
#>  [174,]    0.2391141    0.2326529  0.2351926
#>  [175,]    0.2475489    0.2408598  0.2434891
#>  [176,]    0.2754068    0.2679650  0.2708901
#>  [177,]    0.2366574    0.2302626  0.2327762
#>  [178,]    0.3046901    0.2964572  0.2996933
#>  [179,]    0.2099253    0.2042528  0.2064825
#>  [180,]    0.2005300    0.1951113  0.1972412
#>  [181,]    0.1943645    0.1891125  0.1911769
#>  [182,]    0.2007967    0.1953709  0.1975036
#>  [183,]    0.2027675    0.1972885  0.1994421
#>  [184,]    0.1404770    0.1366810  0.1381731
#>  [185,]    0.2245600    0.2184920  0.2208772
#>  [186,]    0.2690116    0.2617426  0.2645998
#>  [187,]    0.1895515    0.1844296  0.1864429
#>  [188,]    0.2192475    0.2133231  0.2156518
#>  [189,]    0.2457217    0.2390820  0.2416918
#>  [190,]    0.2169639    0.2111012  0.2134056
#>  [191,]    0.1874596    0.1823942  0.1843852
#>  [192,]    0.2119871    0.2062589  0.2085105
#>  [193,]    0.1834490    0.1784918  0.1804403
#>  [194,]    0.1942083    0.1889605  0.1910232
#>  [195,]    0.2122991    0.2065625  0.2088174
#>  [196,]    0.2305786    0.2243480  0.2267971
#>  [197,]    0.2083656    0.2027352  0.2049483
#>  [198,]    0.2855793    0.2778626  0.2808958
#>  [199,]    0.2149369    0.2091290  0.2114119
#>  [200,]    0.2126773    0.2069304  0.2091893
#>  [201,]    0.2137832    0.2080065  0.2102772
#>  [202,]    0.1773381    0.1725462  0.1744297
#>  [203,]    0.2717012    0.2643595  0.2672453
#>  [204,]    0.1639997    0.1595681  0.1613101
#>  [205,]    0.2711797    0.2638520  0.2667323
#>  [206,]    0.2678912    0.2606524  0.2634978
#>  [207,]    0.2669030    0.2596911  0.2625259
#>  [208,]    0.2887010    0.2809000  0.2839663
#>  [209,]    0.2070209    0.2014269  0.2036258
#>  [210,]    0.2212362    0.2152581  0.2176079
#>  [211,]    0.2198889    0.2139472  0.2162827
#>  [212,]    0.2197224    0.2137851  0.2161189
#>  [213,]    0.1749304    0.1702036  0.1720616
#>  [214,]    0.1853432    0.1803349  0.1823035
#>  [215,]    0.2412158    0.2346978  0.2372598
#>  [216,]    0.2356692    0.2293011  0.2318042
#>  [217,]    0.2126432    0.2068973  0.2091558
#>  [218,]    0.2360930    0.2297134  0.2322211
#>  [219,]    0.2050453    0.1995047  0.2016825
#>  [220,]    0.2444415    0.2378364  0.2404327
#>  [221,]    0.2096708    0.2040051  0.2062321
#>  [222,]    0.2215016    0.2155163  0.2178689
#>  [223,]    0.2950879    0.2871142  0.2902484
#>  [224,]    0.2131436    0.2073842  0.2096481
#>  [225,]    0.2153950    0.2095747  0.2118625
#>  [226,]    0.2119974    0.2062689  0.2085206
#>  [227,]    0.2103153    0.2046322  0.2068661
#>  [228,]    0.2569310    0.2499884  0.2527174
#>  [229,]    0.2220046    0.2160056  0.2183636
#>  [230,]    0.1793275    0.1744817  0.1763864
#>  [231,]    0.2751321    0.2676977  0.2706200
#>  [232,]    0.2329668    0.2266719  0.2291463
#>  [233,]    0.2036621    0.1981588  0.2003219
#>  [234,]    0.2777055    0.2702015  0.2731511
#>  [235,]    0.2708985    0.2635785  0.2664558
#>  [236,]    0.2581171    0.2511424  0.2538839
#>  [237,]    0.3500853    0.3406257  0.3443440
#>  [238,]    0.2713442    0.2640121  0.2668941
#>  [239,]    0.3119049    0.3034769  0.3067897
#>  [240,]    0.3035761    0.2953732  0.2985975
#>  [241,]    0.2134940    0.2077251  0.2099927
#>  [242,]    0.1901986    0.1850591  0.1870793
#>  [243,]    0.2776361    0.2701341  0.2730829
#>  [244,]    0.2890486    0.2812382  0.2843082
#>  [245,]    0.1679743    0.1634353  0.1652195
#>  [246,]    0.3043998    0.2961746  0.2994077
#>  [247,]    0.2814148    0.2738107  0.2767996
#>  [248,]    0.1971654    0.1918376  0.1939318
#>  [249,]    0.2114130    0.2057002  0.2079457
#>  [250,]    0.2271086    0.2209718  0.2233840
#>  [251,]    0.2434186    0.2368411  0.2394266
#>  [252,]    0.1733951    0.1687096  0.1705513
#>  [253,]    0.2004054    0.1949902  0.1971188
#>  [254,]    0.2527205    0.2458916  0.2485759
#>  [255,]    0.2743019    0.2668901  0.2698035
#>  [256,]    0.1762671    0.1715040  0.1733763
#>  [257,]    0.2129504    0.2071962  0.2094580
#>  [258,]    0.1983624    0.1930023  0.1951092
#>  [259,]    0.2568452    0.2499049  0.2526329
#>  [260,]    0.2288333    0.2226498  0.2250803
#>  [261,]    0.2361330    0.2297523  0.2322604
#>  [262,]    0.3112774    0.3028664  0.3061725
#>  [263,]    0.3074391    0.2991319  0.3023973
#>  [264,]    0.1711888    0.1665630  0.1683812
#>  [265,]    0.1521494    0.1480380  0.1496541
#>  [266,]    0.2191620    0.2132399  0.2155677
#>  [267,]    0.1695527    0.1649711  0.1667720
#>  [268,]    0.2352067    0.2288511  0.2313493
#>  [269,]    0.2740318    0.2666272  0.2695378
#>  [270,]    0.1798652    0.1750050  0.1769154
#>  [271,]    0.1992621    0.1938777  0.1959941
#>  [272,]    0.3036100    0.2954061  0.2986308
#>  [273,]    0.1772337    0.1724445  0.1743270
#>  [274,]    0.1896332    0.1845089  0.1865231
#>  [275,]    0.2520335    0.2452232  0.2479001
#>  [276,]    0.2831877    0.2755356  0.2785434
#>  [277,]    0.2999415    0.2918369  0.2950226
#>  [278,]    0.1870544    0.1819998  0.1839866
#>  [279,]    0.1956282    0.1903419  0.1924198
#>  [280,]    0.2844072    0.2767222  0.2797430
#>  [281,]    0.1872605    0.1822004  0.1841894
#>  [282,]    0.1911367    0.1859718  0.1880020
#>  [283,]    0.1965152    0.1912050  0.1932923
#>  [284,]    0.2659532    0.2587669  0.2615916
#>  [285,]    0.2910407    0.2831765  0.2862676
#>  [286,]    0.1781259    0.1733126  0.1752046
#>  [287,]    0.2123309    0.2065935  0.2088487
#>  [288,]    0.2323025    0.2260254  0.2284927
#>  [289,]    0.2429356    0.2363711  0.2389514
#>  [290,]    0.2466836    0.2400179  0.2426380
#>  [291,]    0.2023211    0.1968540  0.1990030
#>  [292,]    0.2213688    0.2153871  0.2177383
#>  [293,]    0.2164541    0.2106052  0.2129042
#>  [294,]    0.2199621    0.2140184  0.2163547
#>  [295,]    0.2090159    0.2033679  0.2055880
#>  [296,]    0.2082624    0.2026347  0.2048468
#>  [297,]    0.2708184    0.2635006  0.2663770
#>  [298,]    0.2128832    0.2071307  0.2093919
#>  [299,]    0.2201967    0.2142467  0.2165854
#>  [300,]    0.2481267    0.2414221  0.2440575
#>  [301,]    0.3192268    0.3106011  0.3139916
#>  [302,]    0.1478760    0.1438801  0.1454508
#>  [303,]    0.2802793    0.2727058  0.2756827
#>  [304,]    0.2305942    0.2243632  0.2268124
#>  [305,]    0.1962659    0.1909625  0.1930471
#>  [306,]    0.2044463    0.1989218  0.2010933
#>  [307,]    0.2929546    0.2850387  0.2881502
#>  [308,]    0.2654279    0.2582557  0.2610749
#>  [309,]    0.2423478    0.2357992  0.2383733
#>  [310,]    0.1983238    0.1929648  0.1950713
#>  [311,]    0.3014947    0.2933481  0.2965503
#>  [312,]    0.1967910    0.1914734  0.1935636
#>  [313,]    0.2540408    0.2471762  0.2498745
#>  [314,]    0.2084442    0.2028117  0.2050257
#>  [315,]    0.1670708    0.1625562  0.1643307
#>  [316,]    0.1872909    0.1822300  0.1842193
#>  [317,]    0.2085141    0.2028797  0.2050944
#>  [318,]    0.1703461    0.1657430  0.1675523
#>  [319,]    0.1232914    0.1199598  0.1212694
#>  [320,]    0.2136125    0.2078403  0.2101092
#>  [321,]    0.2050488    0.1995081  0.2016860
#>  [322,]    0.2693941    0.2621148  0.2649760
#>  [323,]    0.2355680    0.2292026  0.2317046
#>  [324,]    0.2126402    0.2068943  0.2091528
#>  [325,]    0.2234389    0.2174014  0.2197745
#>  [326,]    0.1951560    0.1898826  0.1919554
#>  [327,]    0.1678626    0.1633266  0.1651096
#>  [328,]    0.2356933    0.2293245  0.2318279
#>  [329,]    0.2312019    0.2249544  0.2274101
#>  [330,]    0.1780573    0.1732459  0.1751371
#>  [331,]    0.1711848    0.1665591  0.1683773
#>  [332,]    0.2249450    0.2188667  0.2212559
#>  [333,]    0.3057100    0.2974494  0.3006964
#>  [334,]    0.2359686    0.2295924  0.2320987
#>  [335,]    0.2079050    0.2022871  0.2044953
#>  [336,]    0.2571331    0.2501850  0.2529161
#>  [337,]    0.1541477    0.1499823  0.1516196
#>  [338,]    0.1925151    0.1873130  0.1893578
#>  [339,]    0.1982414    0.1928846  0.1949902
#>  [340,]    0.2195475    0.2136150  0.2159469
#>  [341,]    0.1668022    0.1622949  0.1640666
#>  [342,]    0.2270989    0.2209624  0.2233745
#>  [343,]    0.2550803    0.2481877  0.2508970
#>  [344,]    0.2610223    0.2539692  0.2567416
#>  [345,]    0.2383383    0.2318980  0.2344295
#>  [346,]    0.2364580    0.2300685  0.2325800
#>  [347,]    0.2027573    0.1972785  0.1994320
#>  [348,]    0.2075622    0.2019536  0.2041582
#>  [349,]    0.2511417    0.2443555  0.2470229
#>  [350,]    0.2424950    0.2359425  0.2385181
#>  [351,]    0.1749271    0.1702003  0.1720583
#>  [352,]    0.1990433    0.1936649  0.1957790
#>  [353,]    0.2398853    0.2334032  0.2359511
#>  [354,]    0.2932118    0.2852889  0.2884032
#>  [355,]    0.2523507    0.2455318  0.2482121
#>  [356,]    0.1757785    0.1710286  0.1728957
#>  [357,]    0.2194432    0.2135135  0.2158443
#>  [358,]    0.2050430    0.1995024  0.2016802
#>  [359,]    0.1756029    0.1708578  0.1727230
#>  [360,]    0.1441846    0.1402884  0.1418198
#>  [361,]    0.2392778    0.2328123  0.2353537
#>  [362,]    0.2491398    0.2424077  0.2450539
#>  [363,]    0.2505740    0.2438031  0.2464645
#>  [364,]    0.2177315    0.2118480  0.2141606
#>  [365,]    0.2170093    0.2111453  0.2134503
#>  [366,]    0.2641512    0.2570135  0.2598191
#>  [367,]    0.1481952    0.1441907  0.1457647
#>  [368,]    0.2243161    0.2182548  0.2206373
#>  [369,]    0.1615222    0.1571576  0.1588732
#>  [370,]    0.2191943    0.2132712  0.2155994
#>  [371,]    0.2866027    0.2788585  0.2819025
#>  [372,]    0.1353108    0.1316544  0.1330916
#>  [373,]    0.2665664    0.2593635  0.2621947
#>  [374,]    0.2012885    0.1958494  0.1979873
#>  [375,]    0.2120690    0.2063386  0.2085911
#>  [376,]    0.2495514    0.2428083  0.2454588
#>  [377,]    0.2616024    0.2545336  0.2573122
#>  [378,]    0.1586652    0.1543778  0.1560630
#>  [379,]    0.2272468    0.2211063  0.2235200
#>  [380,]    0.2302349    0.2240136  0.2264590
#>  [381,]    0.2758516    0.2683978  0.2713277
#>  [382,]    0.1510563    0.1469745  0.1485790
#>  [383,]    0.2112472    0.2055390  0.2077827
#>  [384,]    0.2139966    0.2082141  0.2104871
#>  [385,]    0.3274755    0.3186269  0.3221050
#>  [386,]    0.1601035    0.1557772  0.1574777
#>  [387,]    0.3385756    0.3294272  0.3330232
#>  [388,]    0.2155054    0.2096821  0.2119711
#>  [389,]    0.2194202    0.2134912  0.2158217
#>  [390,]    0.3116024    0.3031825  0.3064921
#>  [391,]    0.2203583    0.2144039  0.2167444
#>  [392,]    0.2974384    0.2894013  0.2925605
#>  [393,]    0.2106048    0.2049139  0.2071508
#>  [394,]    0.2285292    0.2223539  0.2247812
#>  [395,]    0.1777512    0.1729480  0.1748360
#>  [396,]    0.3103870    0.3020002  0.3052968
#>  [397,]    0.2040964    0.1985813  0.2007491
#>  [398,]    0.1941153    0.1888699  0.1909317
#>  [399,]    0.2132637    0.2075009  0.2097661
#>  [400,]    0.1835462    0.1785865  0.1805360
#>  [401,]    0.2069075    0.2013166  0.2035142
#>  [402,]    0.1956936    0.1904057  0.1924842
#>  [403,]    0.1818171    0.1769041  0.1788353
#>  [404,]    0.2820579    0.2744365  0.2774322
#>  [405,]    0.2024885    0.1970169  0.1991676
#>  [406,]    0.1742976    0.1695879  0.1714391
#>  [407,]    0.2240088    0.2179558  0.2203351
#>  [408,]    0.1912375    0.1860699  0.1881011
#>  [409,]    0.2111371    0.2054318  0.2076744
#>  [410,]    0.1743855    0.1696733  0.1715255
#>  [411,]    0.1936533    0.1884205  0.1904773
#>  [412,]    0.2134069    0.2076403  0.2099070
#>  [413,]    0.2250765    0.2189946  0.2213852
#>  [414,]    0.1834146    0.1784584  0.1804066
#>  [415,]    0.3229926    0.3142650  0.3176956
#>  [416,]    0.1475273    0.1435408  0.1451078
#>  [417,]    0.1663221    0.1618277  0.1635943
#>  [418,]    0.1919631    0.1867760  0.1888149
#>  [419,]    0.2058864    0.2003230  0.2025098
#>  [420,]    0.1776358    0.1728357  0.1747225
#>  [421,]    0.2323805    0.2261013  0.2285695
#>  [422,]    0.2526271    0.2458008  0.2484840
#>  [423,]    0.2500249    0.2432689  0.2459245
#>  [424,]    0.2092705    0.2036157  0.2058384
#>  [425,]    0.2269466    0.2208142  0.2232246
#>  [426,]    0.2273210    0.2211785  0.2235929
#>  [427,]    0.1826698    0.1777338  0.1796740
#>  [428,]    0.1602699    0.1559391  0.1576414
#>  [429,]    0.2666032    0.2593993  0.2622309
#>  [430,]    0.1579637    0.1536953  0.1553731
#>  [431,]    0.3155024    0.3069773  0.3103282
#>  [432,]    0.2221632    0.2161601  0.2185197
#>  [433,]    0.3023795    0.2942089  0.2974205
#>  [434,]    0.2706840    0.2633699  0.2662449
#>  [435,]    0.2991196    0.2910372  0.2942141
#>  [436,]    0.2038458    0.1983376  0.2005027
#>  [437,]    0.2805818    0.2730002  0.2759803
#>  [438,]    0.2210529    0.2150797  0.2174276
#>  [439,]    0.2184810    0.2125773  0.2148978
#>  [440,]    0.2118700    0.2061449  0.2083953
#>  [441,]    0.2207504    0.2147853  0.2171300
#>  [442,]    0.2186732    0.2127644  0.2150870
#>  [443,]    0.2912347    0.2833651  0.2864584
#>  [444,]    0.1761879    0.1714269  0.1732983
#>  [445,]    0.2177508    0.2118668  0.2141796
#>  [446,]    0.2204728    0.2145153  0.2168570
#>  [447,]    0.2031433    0.1976541  0.1998117
#>  [448,]    0.3218164    0.3131206  0.3165386
#>  [449,]    0.3331322    0.3241308  0.3276689
#>  [450,]    0.1801749    0.1753063  0.1772200
#>  [451,]    0.1879497    0.1828710  0.1848673
#>  [452,]    0.1822714    0.1773461  0.1792821
#>  [453,]    0.1934516    0.1882242  0.1902789
#>  [454,]    0.1776512    0.1728508  0.1747377
#>  [455,]    0.2173927    0.2115184  0.2138274
#>  [456,]    0.2145968    0.2087981  0.2110774
#>  [457,]    0.1696118    0.1650287  0.1668302
#>  [458,]    0.2223640    0.2163555  0.2187173
#>  [459,]    0.2537905    0.2469328  0.2496284
#>  [460,]    0.2422137    0.2356687  0.2382414
#>  [461,]    0.2232470    0.2172145  0.2195857
#>  [462,]    0.1635494    0.1591300  0.1608671
#>  [463,]    0.1707229    0.1661096  0.1679229
#>  [464,]    0.2085335    0.2028986  0.2051135
#>  [465,]    0.2103430    0.2046592  0.2068933
#>  [466,]    0.2661423    0.2589508  0.2617775
#>  [467,]    0.2115672    0.2058503  0.2080974
#>  [468,]    0.1883371    0.1832480  0.1852484
#>  [469,]    0.1903091    0.1851667  0.1871880
#>  [470,]    0.3033342    0.2951378  0.2983595
#>  [471,]    0.2525157    0.2456924  0.2483744
#>  [472,]    0.1331508    0.1295528  0.1309671
#>  [473,]    0.1949209    0.1896538  0.1917242
#>  [474,]    0.1648460    0.1603916  0.1621425
#>  [475,]    0.2378853    0.2314573  0.2339839
#>  [476,]    0.2002461    0.1948351  0.1969620
#>  [477,]    0.2496743    0.2429278  0.2455797
#>  [478,]    0.2116827    0.2059628  0.2082111
#>  [479,]    0.2000231    0.1946181  0.1967427
#>  [480,]    0.3454894    0.3361541  0.3398235
#>  [481,]    0.1859183    0.1808945  0.1828692
#>  [482,]    0.1926458    0.1874402  0.1894864
#>  [483,]    0.2442793    0.2376785  0.2402731
#>  [484,]    0.1598512    0.1555317  0.1572296
#>  [485,]    0.1356438    0.1319784  0.1334192
#>  [486,]    0.2043793    0.1988566  0.2010274
#>  [487,]    0.1566298    0.1523974  0.1540610
#>  [488,]    0.2166162    0.2107630  0.2130637
#>  [489,]    0.2803952    0.2728186  0.2757967
#>  [490,]    0.2061967    0.2006249  0.2028150
#>  [491,]    0.1806654    0.1757835  0.1777024
#>  [492,]    0.2394470    0.2329768  0.2355200
#>  [493,]    0.1896041    0.1844806  0.1864945
#>  [494,]    0.2311839    0.2249369  0.2273924
#>  [495,]    0.2787371    0.2712054  0.2741659
#>  [496,]    0.1946760    0.1894155  0.1914832
#>  [497,]    0.2291523    0.2229603  0.2253942
#>  [498,]    0.2379883    0.2315575  0.2340853
#>  [499,]    0.2530368    0.2461994  0.2488870
#>  [500,]    0.1909941    0.1858330  0.1878617
#>  [501,]    0.2796255    0.2720698  0.2750397
#>  [502,]    0.2028527    0.1973713  0.1995259
#>  [503,]    0.2359716    0.2295954  0.2321017
#>  [504,]    0.1768040    0.1720265  0.1739044
#>  [505,]    0.1556734    0.1514668  0.1531203
#>  [506,]    0.1905501    0.1854012  0.1874251
#>  [507,]    0.2428465    0.2362846  0.2388639
#>  [508,]    0.1718423    0.1671988  0.1690240
#>  [509,]    0.1859594    0.1809344  0.1829096
#>  [510,]    0.2470976    0.2404208  0.2430453
#>  [511,]    0.2353602    0.2290004  0.2315003
#>  [512,]    0.1858754    0.1808527  0.1828270
#>  [513,]    0.2161495    0.2103088  0.2126046
#>  [514,]    0.2057662    0.2002061  0.2023916
#>  [515,]    0.1737189    0.1690247  0.1708698
#>  [516,]    0.2836234    0.2759596  0.2789720
#>  [517,]    0.2251807    0.2190960  0.2214877
#>  [518,]    0.1718034    0.1671610  0.1689858
#>  [519,]    0.2087999    0.2031578  0.2053755
#>  [520,]    0.1637528    0.1593279  0.1610672
#>  [521,]    0.3083218    0.2999908  0.3032655
#>  [522,]    0.2526400    0.2458132  0.2484966
#>  [523,]    0.2103971    0.2047118  0.2069465
#>  [524,]    0.2139045    0.2081245  0.2103965
#>  [525,]    0.2749125    0.2674840  0.2704039
#>  [526,]    0.2010594    0.1956265  0.1977620
#>  [527,]    0.2224740    0.2164625  0.2188254
#>  [528,]    0.2109598    0.2052594  0.2075000
#>  [529,]    0.1472182    0.1432402  0.1448038
#>  [530,]    0.2202902    0.2143376  0.2166774
#>  [531,]    0.1956995    0.1904114  0.1924900
#>  [532,]    0.1992514    0.1938673  0.1959837
#>  [533,]    0.1869684    0.1819162  0.1839021
#>  [534,]    0.2305046    0.2242760  0.2267243
#>  [535,]    0.2304915    0.2242633  0.2267114
#>  [536,]    0.1977343    0.1923912  0.1944914
#>  [537,]    0.1996398    0.1942452  0.1963657
#>  [538,]    0.2249035    0.2188262  0.2212150
#>  [539,]    0.1952202    0.1899450  0.1920185
#>  [540,]    0.2045163    0.1989900  0.2011622
#>  [541,]    0.1782483    0.1734317  0.1753250
#>  [542,]    0.2530902    0.2462514  0.2489395
#>  [543,]    0.2559356    0.2490199  0.2517383
#>  [544,]    0.2537678    0.2469107  0.2496060
#>  [545,]    0.2287558    0.2225745  0.2250042
#>  [546,]    0.2004693    0.1950524  0.1971816
#>  [547,]    0.2558133    0.2489009  0.2516180
#>  [548,]    0.1913263    0.1861563  0.1881885
#>  [549,]    0.2489728    0.2422452  0.2448896
#>  [550,]    0.2803448    0.2727696  0.2757472
#>  [551,]    0.3205390    0.3118778  0.3152823
#>  [552,]    0.3141798    0.3056905  0.3090274
#>  [553,]    0.1776334    0.1728334  0.1747201
#>  [554,]    0.1905599    0.1854106  0.1874347
#>  [555,]    0.2390363    0.2325772  0.2351161
#>  [556,]    0.2191844    0.2132617  0.2155897
#>  [557,]    0.2492220    0.2424877  0.2451347
#>  [558,]    0.2451161    0.2384929  0.2410963
#>  [559,]    0.2726330    0.2652662  0.2681619
#>  [560,]    0.2025288    0.1970561  0.1992072
#>  [561,]    0.2182029    0.2123067  0.2146243
#>  [562,]    0.1906601    0.1855082  0.1875332
#>  [563,]    0.1718359    0.1671926  0.1690177
#>  [564,]    0.1870478    0.1819934  0.1839802
#>  [565,]    0.2335461    0.2272354  0.2297159
#>  [566,]    0.1597881    0.1554704  0.1571676
#>  [567,]    0.2046597    0.1991295  0.2013032
#>  [568,]    0.2488335    0.2421098  0.2447526
#>  [569,]    0.2833587    0.2757021  0.2787117
#>  [570,]    0.1645064    0.1600611  0.1618084
#>  [571,]    0.2297466    0.2235385  0.2259787
#>  [572,]    0.2138163    0.2080387  0.2103097
#>  [573,]    0.1913721    0.1862009  0.1882335
#>  [574,]    0.2628329    0.2557308  0.2585224
#>  [575,]    0.2635676    0.2564457  0.2592451
#>  [576,]    0.2055536    0.1999992  0.2021825
#>  [577,]    0.2362285    0.2298453  0.2323543
#>  [578,]    0.2132640    0.2075013  0.2097664
#>  [579,]    0.2364369    0.2300481  0.2325593
#>  [580,]    0.2144158    0.2086220  0.2108994
#>  [581,]    0.1855391    0.1805255  0.1824962
#>  [582,]    0.2111771    0.2054709  0.2077138
#>  [583,]    0.2277398    0.2215860  0.2240049
#>  [584,]    0.2290947    0.2229043  0.2253375
#>  [585,]    0.1909625    0.1858024  0.1878307
#>  [586,]    0.2320382    0.2257682  0.2282328
#>  [587,]    0.2591435    0.2521412  0.2548936
#>  [588,]    0.1978164    0.1924711  0.1945722
#>  [589,]    0.2570112    0.2500664  0.2527962
#>  [590,]    0.1613994    0.1570381  0.1587524
#>  [591,]    0.2231226    0.2170934  0.2194633
#>  [592,]    0.1787600    0.1739296  0.1758283
#>  [593,]    0.2720996    0.2647472  0.2676372
#>  [594,]    0.1656204    0.1611450  0.1629042
#>  [595,]    0.1922698    0.1870744  0.1891165
#>  [596,]    0.2018287    0.1963751  0.1985187
#>  [597,]    0.2127731    0.2070236  0.2092836
#>  [598,]    0.2530546    0.2462168  0.2489045
#>  [599,]    0.1878445    0.1827686  0.1847638
#>  [600,]    0.1547476    0.1505660  0.1522097
#>  [601,]    0.1478181    0.1438238  0.1453938
#>  [602,]    0.1996829    0.1942871  0.1964080
#>  [603,]    0.2663395    0.2591428  0.2619716
#>  [604,]    0.1964452    0.1911370  0.1932235
#>  [605,]    0.1975972    0.1922578  0.1943565
#>  [606,]    0.1957422    0.1904529  0.1925320
#>  [607,]    0.2114232    0.2057102  0.2079558
#>  [608,]    0.2079624    0.2023430  0.2045518
#>  [609,]    0.2179471    0.2120578  0.2143727
#>  [610,]    0.1934624    0.1882348  0.1902896
#>  [611,]    0.1723345    0.1676778  0.1695082
#>  [612,]    0.1551027    0.1509116  0.1525590
#>  [613,]    0.2057927    0.2002318  0.2024176
#>  [614,]    0.2428128    0.2362517  0.2388306
#>  [615,]    0.2257208    0.2196215  0.2220189
#>  [616,]    0.1652819    0.1608157  0.1625712
#>  [617,]    0.1754203    0.1706801  0.1725433
#>  [618,]    0.1978924    0.1925451  0.1946470
#>  [619,]    0.1975680    0.1922294  0.1943278
#>  [620,]    0.2081967    0.2025709  0.2047822
#>  [621,]    0.2194859    0.2135551  0.2158863
#>  [622,]    0.2308495    0.2246116  0.2270636
#>  [623,]    0.2162051    0.2103629  0.2126593
#>  [624,]    0.1991283    0.1937475  0.1958625
#>  [625,]    0.2104487    0.2047620  0.2069973
#>  [626,]    0.2346357    0.2282955  0.2307877
#>  [627,]    0.2694211    0.2621410  0.2650026
#>  [628,]    0.1781331    0.1733196  0.1752117
#>  [629,]    0.2183286    0.2124290  0.2147480
#>  [630,]    0.2760895    0.2686293  0.2715617
#>  [631,]    0.2115376    0.2058216  0.2080684
#>  [632,]    0.2697516    0.2624626  0.2653277
#>  [633,]    0.2703115    0.2630073  0.2658784
#>  [634,]    0.1988520    0.1934787  0.1955908
#>  [635,]    0.1761110    0.1713521  0.1732227
#>  [636,]    0.2064306    0.2008525  0.2030451
#>  [637,]    0.2179837    0.2120934  0.2144087
#>  [638,]    0.1990935    0.1937136  0.1958283
#>  [639,]    0.2459492    0.2393033  0.2419156
#>  [640,]    0.1637475    0.1593228  0.1610620
#>  [641,]    0.2968543    0.2888329  0.2919859
#>  [642,]    0.1710279    0.1664064  0.1682229
#>  [643,]    0.1938648    0.1886263  0.1906854
#>  [644,]    0.2921899    0.2842947  0.2873981
#>  [645,]    0.2652283    0.2580615  0.2608785
#>  [646,]    0.2833601    0.2757034  0.2787130
#>  [647,]    0.1306484    0.1271181  0.1285058
#>  [648,]    0.2345533    0.2282153  0.2307066
#>  [649,]    0.1613109    0.1569520  0.1586653
#>  [650,]    0.2236168    0.2175744  0.2199495
#>  [651,]    0.1694512    0.1648723  0.1666722
#>  [652,]    0.2316827    0.2254223  0.2278831
#>  [653,]    0.2107076    0.2050140  0.2072519
#>  [654,]    0.1929983    0.1877832  0.1898331
#>  [655,]    0.1968244    0.1915059  0.1935964
#>  [656,]    0.2439806    0.2373880  0.2399793
#>  [657,]    0.2048725    0.1993365  0.2015126
#>  [658,]    0.2243991    0.2183355  0.2207189
#>  [659,]    0.2024413    0.1969710  0.1991212
#>  [660,]    0.2164742    0.2106248  0.2129240
#>  [661,]    0.1847350    0.1797431  0.1817053
#>  [662,]    0.2699382    0.2626442  0.2655113
#>  [663,]    0.1755091    0.1707666  0.1726307
#>  [664,]    0.2129815    0.2072264  0.2094886
#>  [665,]    0.1638881    0.1594596  0.1612003
#>  [666,]    0.2222485    0.2162430  0.2186036
#>  [667,]    0.2456539    0.2390159  0.2416251
#>  [668,]    0.1691550    0.1645841  0.1663808
#>  [669,]    0.2273337    0.2211909  0.2236054
#>  [670,]    0.2731149    0.2657352  0.2686359
#>  [671,]    0.2276996    0.2215468  0.2239653
#>  [672,]    0.3200418    0.3113940  0.3147932
#>  [673,]    0.2268324    0.2207030  0.2231123
#>  [674,]    0.2859983    0.2782703  0.2813079
#>  [675,]    0.2435195    0.2369393  0.2395258
#>  [676,]    0.2100881    0.2044112  0.2066426
#>  [677,]    0.2251128    0.2190300  0.2214209
#>  [678,]    0.2448711    0.2382545  0.2408553
#>  [679,]    0.1927426    0.1875344  0.1895815
#>  [680,]    0.2000921    0.1946853  0.1968106
#>  [681,]    0.2324273    0.2261468  0.2286155
#>  [682,]    0.1916753    0.1864959  0.1885317
#>  [683,]    0.2109931    0.2052918  0.2075328
#>  [684,]    0.1779570    0.1731482  0.1750384
#>  [685,]    0.2181910    0.2122952  0.2146127
#>  [686,]    0.2410664    0.2345525  0.2371129
#>  [687,]    0.2159491    0.2101138  0.2124075
#>  [688,]    0.1907727    0.1856177  0.1876439
#>  [689,]    0.2486197    0.2419017  0.2445423
#>  [690,]    0.1846613    0.1796715  0.1816328
#>  [691,]    0.1582897    0.1540124  0.1556937
#>  [692,]    0.2366764    0.2302811  0.2327949
#>  [693,]    0.2515605    0.2447631  0.2474349
#>  [694,]    0.2695817    0.2622974  0.2651607
#>  [695,]    0.2057224    0.2001634  0.2023485
#>  [696,]    0.1869385    0.1818871  0.1838726
#>  [697,]    0.1693163    0.1647410  0.1665394
#>  [698,]    0.3105552    0.3021637  0.3054622
#>  [699,]    0.1895484    0.1844265  0.1864398
#>  [700,]    0.1968732    0.1915534  0.1936444
#>  [701,]    0.1833749    0.1784198  0.1803675
#>  [702,]    0.2445773    0.2379684  0.2405662
#>  [703,]    0.1736737    0.1689807  0.1708254
#>  [704,]    0.1884525    0.1833602  0.1853618
#>  [705,]    0.1709762    0.1663561  0.1681721
#>  [706,]    0.1350351    0.1313862  0.1328204
#>  [707,]    0.1869053    0.1818549  0.1838401
#>  [708,]    0.1951269    0.1898542  0.1919267
#>  [709,]    0.1739960    0.1692943  0.1711424
#>  [710,]    0.3091663    0.3008124  0.3040960
#>  [711,]    0.2108169    0.2051203  0.2073595
#>  [712,]    0.1895503    0.1844283  0.1864416
#>  [713,]    0.2014150    0.1959724  0.1981117
#>  [714,]    0.3028161    0.2946337  0.2978499
#>  [715,]    0.2444647    0.2378589  0.2404554
#>  [716,]    0.3021497    0.2939854  0.2971945
#>  [717,]    0.2010517    0.1956190  0.1977544
#>  [718,]    0.1897239    0.1845972  0.1866123
#>  [719,]    0.2083456    0.2027157  0.2049287
#>  [720,]    0.2645278    0.2573800  0.2601896
#>  [721,]    0.1814292    0.1765266  0.1784537
#>  [722,]    0.2055969    0.2000413  0.2022251
#>  [723,]    0.2269872    0.2208537  0.2232646
#>  [724,]    0.2418133    0.2352793  0.2378476
#>  [725,]    0.1938864    0.1886473  0.1907066
#>  [726,]    0.2410535    0.2345399  0.2371002
#>  [727,]    0.2689668    0.2616991  0.2645558
#>  [728,]    0.1945427    0.1892859  0.1913522
#>  [729,]    0.2079290    0.2023104  0.2045189
#>  [730,]    0.3567379    0.3470986  0.3508875
#>  [731,]    0.2527221    0.2458933  0.2485775
#>  [732,]    0.2711098    0.2637841  0.2666636
#>  [733,]    0.2187825    0.2128706  0.2151944
#>  [734,]    0.1816947    0.1767850  0.1787149
#>  [735,]    0.1854984    0.1804860  0.1824562
#>  [736,]    0.1972482    0.1919183  0.1940133
#>  [737,]    0.1493807    0.1453441  0.1469308
#>  [738,]    0.1826253    0.1776905  0.1796302
#>  [739,]    0.1897294    0.1846026  0.1866178
#>  [740,]    0.2048262    0.1992915  0.2014670
#>  [741,]    0.2021882    0.1967247  0.1988723
#>  [742,]    0.2270347    0.2208999  0.2233113
#>  [743,]    0.1690051    0.1644383  0.1662334
#>  [744,]    0.2515645    0.2447669  0.2474388
#>  [745,]    0.2194365    0.2135070  0.2158377
#>  [746,]    0.2018730    0.1964181  0.1985623
#>  [747,]    0.2925391    0.2846344  0.2877415
#>  [748,]    0.3039989    0.2957847  0.2990134
#>  [749,]    0.1815590    0.1766529  0.1785813
#>  [750,]    0.2553033    0.2484047  0.2511163
#>  [751,]    0.1885564    0.1834613  0.1854641
#>  [752,]    0.2369437    0.2305411  0.2330578
#>  [753,]    0.1580387    0.1537682  0.1554468
#>  [754,]    0.2266229    0.2204992  0.2229063
#>  [755,]    0.1510830    0.1470005  0.1486052
#>  [756,]    0.1669743    0.1624623  0.1642358
#>  [757,]    0.2254927    0.2193996  0.2217946
#>  [758,]    0.1790802    0.1742411  0.1761432
#>  [759,]    0.1898196    0.1846903  0.1867065
#>  [760,]    0.2739415    0.2665393  0.2694489
#>  [761,]    0.1707907    0.1661757  0.1679897
#>  [762,]    0.1862506    0.1812178  0.1831960
#>  [763,]    0.2472394    0.2405587  0.2431847
#>  [764,]    0.2381686    0.2317330  0.2342627
#>  [765,]    0.1933499    0.1881252  0.1901788
#>  [766,]    0.2376662    0.2312442  0.2337685
#>  [767,]    0.2364609    0.2300714  0.2325829
#>  [768,]    0.1708944    0.1662766  0.1680917
#>  [769,]    0.1873383    0.1822761  0.1842659
#>  [770,]    0.2439193    0.2373284  0.2399191
#>  [771,]    0.2066081    0.2010252  0.2032197
#>  [772,]    0.3233374    0.3146006  0.3180348
#>  [773,]    0.1842271    0.1792489  0.1812057
#>  [774,]    0.2142134    0.2084250  0.2107002
#>  [775,]    0.1896472    0.1845226  0.1865369
#>  [776,]    0.2207741    0.2148085  0.2171534
#>  [777,]    0.2223595    0.2163510  0.2187128
#>  [778,]    0.2271722    0.2210337  0.2234466
#>  [779,]    0.1668655    0.1623565  0.1641288
#>  [780,]    0.2512964    0.2445062  0.2471752
#>  [781,]    0.1974086    0.1920743  0.1941711
#>  [782,]    0.2073402    0.2017376  0.2039398
#>  [783,]    0.2373482    0.2309348  0.2334557
#>  [784,]    0.3180236    0.3094303  0.3128080
#>  [785,]    0.1818513    0.1769374  0.1788689
#>  [786,]    0.2086776    0.2030388  0.2052552
#>  [787,]    0.3293250    0.3204264  0.3239242
#>  [788,]    0.2237084    0.2176635  0.2200395
#>  [789,]    0.2215429    0.2155564  0.2179095
#>  [790,]    0.2808531    0.2732642  0.2762472
#>  [791,]    0.1737276    0.1690332  0.1708785
#>  [792,]    0.1902576    0.1851165  0.1871373
#>  [793,]    0.1564076    0.1521812  0.1538425
#>  [794,]    0.2747974    0.2673721  0.2702908
#>  [795,]    0.2225325    0.2165193  0.2188829
#>  [796,]    0.3017521    0.2935985  0.2968034
#>  [797,]    0.1487671    0.1447471  0.1463272
#>  [798,]    0.1874273    0.1823627  0.1843534
#>  [799,]    0.1997291    0.1943321  0.1964535
#>  [800,]    0.2313085    0.2250583  0.2275150
#>  [801,]    0.2278245    0.2216684  0.2240882
#>  [802,]    0.2195291    0.2135971  0.2159288
#>  [803,]    0.2541796    0.2473114  0.2500111
#>  [804,]    0.1917993    0.1866166  0.1886538
#>  [805,]    0.1962119    0.1909099  0.1929940
#>  [806,]    0.1785400    0.1737155  0.1756119
#>  [807,]    0.2070232    0.2014291  0.2036280
#>  [808,]    0.2121294    0.2063973  0.2086504
#>  [809,]    0.1679893    0.1634500  0.1652343
#>  [810,]    0.2273765    0.2212325  0.2236475
#>  [811,]    0.2426267    0.2360706  0.2386476
#>  [812,]    0.1995623    0.1941698  0.1962894
#>  [813,]    0.1920152    0.1868266  0.1888661
#>  [814,]    0.2049435    0.1994056  0.2015824
#>  [815,]    0.2594360    0.2524257  0.2551813
#>  [816,]    0.1422846    0.1384398  0.1399510
#>  [817,]    0.2212185    0.2152409  0.2175905
#>  [818,]    0.1749969    0.1702682  0.1721269
#>  [819,]    0.2561206    0.2491998  0.2519201
#>  [820,]    0.1958060    0.1905149  0.1925947
#>  [821,]    0.2069786    0.2013857  0.2035841
#>  [822,]    0.2221741    0.2161706  0.2185304
#>  [823,]    0.1758937    0.1711408  0.1730090
#>  [824,]    0.2508033    0.2440263  0.2466901
#>  [825,]    0.2724434    0.2650816  0.2679753
#>  [826,]    0.3301885    0.3212666  0.3247735
#>  [827,]    0.1888024    0.1837007  0.1857060
#>  [828,]    0.2264082    0.2202904  0.2226951
#>  [829,]    0.1904645    0.1853178  0.1873408
#>  [830,]    0.2225188    0.2165061  0.2188695
#>  [831,]    0.2397680    0.2332891  0.2358358
#>  [832,]    0.2489652    0.2422378  0.2448822
#>  [833,]    0.2018322    0.1963784  0.1985221
#>  [834,]    0.1842323    0.1792540  0.1812109
#>  [835,]    0.1473916    0.1434088  0.1449743
#>  [836,]    0.1575304    0.1532736  0.1549468
#>  [837,]    0.1574200    0.1531662  0.1548383
#>  [838,]    0.1857435    0.1807245  0.1826973
#>  [839,]    0.2257723    0.2196716  0.2220696
#>  [840,]    0.2158064    0.2099749  0.2122671
#>  [841,]    0.2040744    0.1985600  0.2007275
#>  [842,]    0.2498896    0.2431373  0.2457914
#>  [843,]    0.2200087    0.2140638  0.2164005
#>  [844,]    0.2784986    0.2709733  0.2739313
#>  [845,]    0.3031867    0.2949943  0.2982145
#>  [846,]    0.1661685    0.1616784  0.1634433
#>  [847,]    0.2150827    0.2092709  0.2115553
#>  [848,]    0.2107335    0.2050392  0.2072775
#>  [849,]    0.2613303    0.2542689  0.2570445
#>  [850,]    0.2515780    0.2447800  0.2474521
#>  [851,]    0.2112867    0.2055774  0.2078216
#>  [852,]    0.2109527    0.2052524  0.2074930
#>  [853,]    0.1850142    0.1800148  0.1819799
#>  [854,]    0.2204855    0.2145276  0.2168695
#>  [855,]    0.1956323    0.1903461  0.1924239
#>  [856,]    0.2113840    0.2056722  0.2079173
#>  [857,]    0.2945598    0.2866006  0.2897291
#>  [858,]    0.2263430    0.2202269  0.2226310
#>  [859,]    0.3287215    0.3198392  0.3233306
#>  [860,]    0.1352957    0.1316397  0.1330768
#>  [861,]    0.2082899    0.2026615  0.2048738
#>  [862,]    0.1980130    0.1926624  0.1947656
#>  [863,]    0.1586323    0.1543457  0.1560306
#>  [864,]    0.2598772    0.2528550  0.2556152
#>  [865,]    0.2910281    0.2831642  0.2862553
#>  [866,]    0.1626698    0.1582742  0.1600020
#>  [867,]    0.2769329    0.2694500  0.2723913
#>  [868,]    0.1890162    0.1839088  0.1859163
#>  [869,]    0.1888359    0.1837332  0.1857389
#>  [870,]    0.1903442    0.1852008  0.1872225
#>  [871,]    0.1691717    0.1646004  0.1663973
#>  [872,]    0.2150753    0.2092636  0.2115480
#>  [873,]    0.2238077    0.2177601  0.2201372
#>  [874,]    0.2302436    0.2240221  0.2264676
#>  [875,]    0.1498530    0.1458037  0.1473953
#>  [876,]    0.3092872    0.3009300  0.3042150
#>  [877,]    0.2104950    0.2048071  0.2070429
#>  [878,]    0.2351741    0.2288194  0.2313172
#>  [879,]    0.2063465    0.2007706  0.2029623
#>  [880,]    0.2600380    0.2530115  0.2557734
#>  [881,]    0.2484822    0.2417680  0.2444071
#>  [882,]    0.2316879    0.2254274  0.2278882
#>  [883,]    0.2364038    0.2300158  0.2325267
#>  [884,]    0.2338664    0.2275470  0.2300310
#>  [885,]    0.1881552    0.1830710  0.1850694
#>  [886,]    0.2251567    0.2190726  0.2214641
#>  [887,]    0.2245350    0.2184677  0.2208526
#>  [888,]    0.2011924    0.1957559  0.1978928
#>  [889,]    0.1998961    0.1944946  0.1966177
#>  [890,]    0.2122042    0.2064702  0.2087240
#>  [891,]    0.2404406    0.2339436  0.2364974
#>  [892,]    0.1834135    0.1784574  0.1804055
#>  [893,]    0.2188970    0.2129821  0.2153071
#>  [894,]    0.1751848    0.1704510  0.1723117
#>  [895,]    0.2689113    0.2616451  0.2645012
#>  [896,]    0.2874448    0.2796777  0.2827307
#>  [897,]    0.2473057    0.2406231  0.2432498
#>  [898,]    0.1498518    0.1458025  0.1473942
#>  [899,]    0.1783455    0.1735263  0.1754206
#>  [900,]    0.1699920    0.1653985  0.1672041
#>  [901,]    0.1853206    0.1803129  0.1822813
#>  [902,]    0.1777977    0.1729933  0.1748818
#>  [903,]    0.2304433    0.2242165  0.2266641
#>  [904,]    0.2367578    0.2303603  0.2328750
#>  [905,]    0.2536220    0.2467688  0.2494626
#>  [906,]    0.1798737    0.1750133  0.1769238
#>  [907,]    0.2720071    0.2646572  0.2675462
#>  [908,]    0.2669223    0.2597097  0.2625448
#>  [909,]    0.1796657    0.1748109  0.1767192
#>  [910,]    0.1799109    0.1750494  0.1769603
#>  [911,]    0.2145779    0.2087796  0.2110587
#>  [912,]    0.2001178    0.1947103  0.1968358
#>  [913,]    0.3423841    0.3331326  0.3367691
#>  [914,]    0.1674203    0.1628963  0.1646745
#>  [915,]    0.1367038    0.1330098  0.1344618
#>  [916,]    0.2376006    0.2311803  0.2337039
#>  [917,]    0.1844938    0.1795085  0.1814680
#>  [918,]    0.2023691    0.1969008  0.1990502
#>  [919,]    0.2279798    0.2218196  0.2242410
#>  [920,]    0.1813170    0.1764175  0.1783433
#>  [921,]    0.2280144    0.2218531  0.2242749
#>  [922,]    0.2318620    0.2255969  0.2280595
#>  [923,]    0.1825309    0.1775986  0.1795374
#>  [924,]    0.3280300    0.3191664  0.3226504
#>  [925,]    0.2470361    0.2403609  0.2429847
#>  [926,]    0.2669705    0.2597565  0.2625921
#>  [927,]    0.2191827    0.2132601  0.2155881
#>  [928,]    0.1963747    0.1910684  0.1931541
#>  [929,]    0.2514256    0.2446318  0.2473022
#>  [930,]    0.2676610    0.2604285  0.2632714
#>  [931,]    0.2223008    0.2162939  0.2186550
#>  [932,]    0.2277134    0.2215602  0.2239788
#>  [933,]    0.2117196    0.2059986  0.2082474
#>  [934,]    0.1814344    0.1765317  0.1784588
#>  [935,]    0.2220353    0.2160356  0.2183939
#>  [936,]    0.3552540    0.3456549  0.3494280
#>  [937,]    0.2073035    0.2017018  0.2039037
#>  [938,]    0.1978178    0.1924725  0.1945736
#>  [939,]    0.1911305    0.1859659  0.1879959
#>  [940,]    0.2181618    0.2122667  0.2145839
#>  [941,]    0.2277939    0.2216386  0.2240581
#>  [942,]    0.2035892    0.1980879  0.2002503
#>  [943,]    0.2114230    0.2057100  0.2079556
#>  [944,]    0.2025189    0.1970465  0.1991975
#>  [945,]    0.2016719    0.1962224  0.1983644
#>  [946,]    0.2250401    0.2189592  0.2213494
#>  [947,]    0.2302331    0.2240119  0.2264573
#>  [948,]    0.2145545    0.2087570  0.2110358
#>  [949,]    0.2529794    0.2461436  0.2488306
#>  [950,]    0.1889440    0.1838384  0.1858453
#>  [951,]    0.1721187    0.1674678  0.1692959
#>  [952,]    0.1802039    0.1753346  0.1772486
#>  [953,]    0.1885416    0.1834470  0.1854495
#>  [954,]    0.2104367    0.2047503  0.2069855
#>  [955,]    0.3104556    0.3020668  0.3053642
#>  [956,]    0.2267929    0.2206646  0.2230734
#>  [957,]    0.2071467    0.2015493  0.2037495
#>  [958,]    0.2124775    0.2067361  0.2089929
#>  [959,]    0.1583401    0.1540615  0.1557433
#>  [960,]    0.2224405    0.2164299  0.2187925
#>  [961,]    0.2624413    0.2553497  0.2581372
#>  [962,]    0.2165663    0.2107144  0.2130146
#>  [963,]    0.1957411    0.1904519  0.1925309
#>  [964,]    0.2672799    0.2600578  0.2628966
#>  [965,]    0.2014965    0.1960517  0.1981919
#>  [966,]    0.2820054    0.2743854  0.2773806
#>  [967,]    0.2238408    0.2177922  0.2201697
#>  [968,]    0.3301927    0.3212709  0.3247778
#>  [969,]    0.2384775    0.2320336  0.2345665
#>  [970,]    0.2750795    0.2676465  0.2705682
#>  [971,]    0.1748528    0.1701280  0.1719852
#>  [972,]    0.2696637    0.2623770  0.2652411
#>  [973,]    0.1862932    0.1812593  0.1832380
#>  [974,]    0.1941499    0.1889037  0.1909658
#>  [975,]    0.2953387    0.2873583  0.2904952
#>  [976,]    0.1933521    0.1881275  0.1901811
#>  [977,]    0.2229758    0.2169506  0.2193189
#>  [978,]    0.2399383    0.2334548  0.2360032
#>  [979,]    0.2094519    0.2037921  0.2060168
#>  [980,]    0.1809438    0.1760544  0.1779763
#>  [981,]    0.1846208    0.1796320  0.1815929
#>  [982,]    0.2527448    0.2459154  0.2485998
#>  [983,]    0.3107347    0.3023383  0.3056387
#>  [984,]    0.2247456    0.2186727  0.2210597
#>  [985,]    0.2777109    0.2702069  0.2731565
#>  [986,]    0.1812920    0.1763932  0.1783188
#>  [987,]    0.3236917    0.3149454  0.3183833
#>  [988,]    0.1830665    0.1781198  0.1800642
#>  [989,]    0.1640716    0.1596381  0.1613808
#>  [990,]    0.2088940    0.2032494  0.2054681
#>  [991,]    0.2089216    0.2032762  0.2054952
#>  [992,]    0.1816455    0.1767372  0.1786665
#>  [993,]    0.2981503    0.2900940  0.2932607
#>  [994,]    0.3368433    0.3277415  0.3313192
#>  [995,]    0.2265289    0.2204078  0.2228139
#>  [996,]    0.2231438    0.2171141  0.2194842
#>  [997,]    0.1903778    0.1852335  0.1872555
#>  [998,]    0.2170189    0.2111548  0.2134598
#>  [999,]    0.2576064    0.2506456  0.2533816
#> [1000,]    0.1986727    0.1933042  0.1954144