Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 88.05178
#> 2   Feb 91.27486
#> 3   Mar 97.88475
#> 4   Apr 87.02076
#> 5   May 93.63183
#> 6   Jun 90.02835

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 6.563877  2
#> 2   Feb 9.877408  2
#> 3   Mar 8.774643  2
#> 4   Apr 7.653677  2
#> 5   May 8.452319  2
#> 6   Jun 9.971387  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [14ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [26ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [84ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [844ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [23ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.2007844    0.1985137  0.1988272
#>    [2,]    0.2076434    0.2052951  0.2056193
#>    [3,]    0.2312360    0.2286209  0.2289820
#>    [4,]    0.1944660    0.1922668  0.1925704
#>    [5,]    0.2124140    0.2100117  0.2103434
#>    [6,]    0.2786144    0.2754635  0.2758986
#>    [7,]    0.2306299    0.2280217  0.2283818
#>    [8,]    0.1792515    0.1772243  0.1775041
#>    [9,]    0.2524932    0.2496377  0.2500320
#>   [10,]    0.3904740    0.3860583  0.3866680
#>   [11,]    0.3104735    0.3069625  0.3074472
#>   [12,]    0.2428947    0.2401478  0.2405271
#>   [13,]    0.1998210    0.1975612  0.1978732
#>   [14,]    0.1862083    0.1841024  0.1843931
#>   [15,]    0.2729503    0.2698635  0.2702897
#>   [16,]    0.2229871    0.2204653  0.2208135
#>   [17,]    0.2524208    0.2495662  0.2499603
#>   [18,]    0.2398983    0.2371852  0.2375598
#>   [19,]    0.1500313    0.1483345  0.1485688
#>   [20,]    0.2569871    0.2540808  0.2544820
#>   [21,]    0.1835156    0.1814401  0.1817267
#>   [22,]    0.2527464    0.2498881  0.2502827
#>   [23,]    0.2163700    0.2139230  0.2142608
#>   [24,]    0.2094470    0.2070783  0.2074053
#>   [25,]    0.2317110    0.2290906  0.2294524
#>   [26,]    0.1924449    0.1902684  0.1905689
#>   [27,]    0.2793951    0.2762354  0.2766716
#>   [28,]    0.2446300    0.2418635  0.2422454
#>   [29,]    0.2077726    0.2054228  0.2057472
#>   [30,]    0.1752655    0.1732835  0.1735571
#>   [31,]    0.2624754    0.2595070  0.2599169
#>   [32,]    0.2485214    0.2457108  0.2460989
#>   [33,]    0.2525324    0.2496765  0.2500708
#>   [34,]    0.3187974    0.3151922  0.3156899
#>   [35,]    0.2102331    0.2078555  0.2081838
#>   [36,]    0.2240293    0.2214957  0.2218455
#>   [37,]    0.2272270    0.2246573  0.2250121
#>   [38,]    0.1534623    0.1517267  0.1519663
#>   [39,]    0.2365913    0.2339157  0.2342851
#>   [40,]    0.3308795    0.3271377  0.3276543
#>   [41,]    0.2302794    0.2276751  0.2280347
#>   [42,]    0.2047976    0.2024814  0.2028012
#>   [43,]    0.3605737    0.3564961  0.3570591
#>   [44,]    0.1915237    0.1893577  0.1896567
#>   [45,]    0.2583117    0.2553904  0.2557937
#>   [46,]    0.1513378    0.1496263  0.1498626
#>   [47,]    0.2993315    0.2959464  0.2964137
#>   [48,]    0.1799918    0.1779562  0.1782373
#>   [49,]    0.2218773    0.2193680  0.2197145
#>   [50,]    0.1987987    0.1965504  0.1968608
#>   [51,]    0.2073753    0.2050300  0.2053538
#>   [52,]    0.2259167    0.2233618  0.2237145
#>   [53,]    0.2292233    0.2266310  0.2269889
#>   [54,]    0.2674078    0.2643837  0.2648012
#>   [55,]    0.2720910    0.2690139  0.2694388
#>   [56,]    0.2990617    0.2956796  0.2961465
#>   [57,]    0.2661345    0.2631248  0.2635403
#>   [58,]    0.1878140    0.1856899  0.1859832
#>   [59,]    0.2443397    0.2415764  0.2419579
#>   [60,]    0.1758810    0.1738919  0.1741665
#>   [61,]    0.2655605    0.2625573  0.2629719
#>   [62,]    0.2068774    0.2045378  0.2048608
#>   [63,]    0.1727864    0.1708323  0.1711021
#>   [64,]    0.2551408    0.2522554  0.2526538
#>   [65,]    0.1996593    0.1974013  0.1977130
#>   [66,]    0.2465487    0.2437604  0.2441454
#>   [67,]    0.2381232    0.2354302  0.2358020
#>   [68,]    0.2388214    0.2361206  0.2364935
#>   [69,]    0.2006082    0.1983394  0.1986527
#>   [70,]    0.2597154    0.2567783  0.2571838
#>   [71,]    0.1903564    0.1882036  0.1885008
#>   [72,]    0.1908693    0.1887107  0.1890087
#>   [73,]    0.2374322    0.2347470  0.2351178
#>   [74,]    0.2372730    0.2345896  0.2349601
#>   [75,]    0.2797899    0.2766258  0.2770626
#>   [76,]    0.1988673    0.1966183  0.1969288
#>   [77,]    0.2564558    0.2535555  0.2539560
#>   [78,]    0.1907152    0.1885583  0.1888561
#>   [79,]    0.2192796    0.2167997  0.2171420
#>   [80,]    0.2023308    0.2000426  0.2003585
#>   [81,]    0.1936893    0.1914988  0.1918013
#>   [82,]    0.2903514    0.2870678  0.2875211
#>   [83,]    0.3200373    0.3164180  0.3169177
#>   [84,]    0.1904591    0.1883051  0.1886025
#>   [85,]    0.1920171    0.1898455  0.1901453
#>   [86,]    0.1814986    0.1794459  0.1797293
#>   [87,]    0.2477121    0.2449107  0.2452974
#>   [88,]    0.1962128    0.1939937  0.1943001
#>   [89,]    0.3050540    0.3016041  0.3020804
#>   [90,]    0.1444608    0.1428271  0.1430526
#>   [91,]    0.3170064    0.3134214  0.3139163
#>   [92,]    0.1582458    0.1564561  0.1567032
#>   [93,]    0.2320319    0.2294078  0.2297701
#>   [94,]    0.2491864    0.2463683  0.2467574
#>   [95,]    0.2519505    0.2491012  0.2494946
#>   [96,]    0.2290307    0.2264406  0.2267982
#>   [97,]    0.2020711    0.1997858  0.2001013
#>   [98,]    0.2502565    0.2474263  0.2478171
#>   [99,]    0.2084472    0.2060898  0.2064153
#>  [100,]    0.2226282    0.2201104  0.2204580
#>  [101,]    0.2359980    0.2333290  0.2336975
#>  [102,]    0.1938602    0.1916678  0.1919705
#>  [103,]    0.2161717    0.2137270  0.2140645
#>  [104,]    0.1307238    0.1292454  0.1294495
#>  [105,]    0.2670641    0.2640439  0.2644609
#>  [106,]    0.2246418    0.2221013  0.2224521
#>  [107,]    0.1910845    0.1889235  0.1892219
#>  [108,]    0.1168198    0.1154986  0.1156810
#>  [109,]    0.2641467    0.2611594  0.2615718
#>  [110,]    0.2269767    0.2244097  0.2247641
#>  [111,]    0.2977326    0.2943655  0.2948304
#>  [112,]    0.3588118    0.3547541  0.3553144
#>  [113,]    0.2162756    0.2138297  0.2141674
#>  [114,]    0.2158986    0.2134569  0.2137940
#>  [115,]    0.2028493    0.2005553  0.2008720
#>  [116,]    0.1453728    0.1437287  0.1439557
#>  [117,]    0.1958028    0.1935884  0.1938942
#>  [118,]    0.1728413    0.1708866  0.1711564
#>  [119,]    0.2022663    0.1999788  0.2002947
#>  [120,]    0.1847767    0.1826870  0.1829755
#>  [121,]    0.2208679    0.2183700  0.2187149
#>  [122,]    0.1802674    0.1782287  0.1785102
#>  [123,]    0.1609838    0.1591632  0.1594145
#>  [124,]    0.2650507    0.2620533  0.2624672
#>  [125,]    0.3173257    0.3137371  0.3142326
#>  [126,]    0.2658489    0.2628423  0.2632574
#>  [127,]    0.2243029    0.2217662  0.2221164
#>  [128,]    0.2791943    0.2760369  0.2764728
#>  [129,]    0.5158415    0.5100081  0.5108135
#>  [130,]    0.2287507    0.2261637  0.2265209
#>  [131,]    0.2212720    0.2187696  0.2191151
#>  [132,]    0.2196888    0.2172043  0.2175473
#>  [133,]    0.1846160    0.1825281  0.1828164
#>  [134,]    0.1572143    0.1554363  0.1556818
#>  [135,]    0.2393945    0.2366871  0.2370609
#>  [136,]    0.2234265    0.2208998  0.2212486
#>  [137,]    0.2702108    0.2671549  0.2675768
#>  [138,]    0.2227584    0.2202392  0.2205870
#>  [139,]    0.1759680    0.1739779  0.1742527
#>  [140,]    0.1619390    0.1601076  0.1603604
#>  [141,]    0.2222526    0.2197391  0.2200861
#>  [142,]    0.1555509    0.1537917  0.1540346
#>  [143,]    0.1525607    0.1508354  0.1510736
#>  [144,]    0.1649785    0.1631127  0.1633703
#>  [145,]    0.3048848    0.3014369  0.3019129
#>  [146,]    0.2556569    0.2527657  0.2531649
#>  [147,]    0.2496321    0.2468090  0.2471987
#>  [148,]    0.1856931    0.1835930  0.1838829
#>  [149,]    0.2035188    0.2012171  0.2015349
#>  [150,]    0.2945369    0.2912059  0.2916658
#>  [151,]    0.2318641    0.2292419  0.2296040
#>  [152,]    0.1816413    0.1795870  0.1798706
#>  [153,]    0.2127585    0.2103523  0.2106845
#>  [154,]    0.2907623    0.2874741  0.2879280
#>  [155,]    0.2143880    0.2119634  0.2122981
#>  [156,]    0.3455731    0.3416651  0.3422046
#>  [157,]    0.1444533    0.1428196  0.1430451
#>  [158,]    0.2164281    0.2139804  0.2143184
#>  [159,]    0.2025261    0.2002357  0.2005519
#>  [160,]    0.2511983    0.2483574  0.2487497
#>  [161,]    0.1600079    0.1581983  0.1584481
#>  [162,]    0.2610150    0.2580632  0.2584707
#>  [163,]    0.3259349    0.3222489  0.3227578
#>  [164,]    0.2105795    0.2081980  0.2085268
#>  [165,]    0.1894751    0.1873322  0.1876281
#>  [166,]    0.3202107    0.3165894  0.3170894
#>  [167,]    0.2452025    0.2424294  0.2428123
#>  [168,]    0.2235787    0.2210502  0.2213993
#>  [169,]    0.2480468    0.2452416  0.2456289
#>  [170,]    0.2034445    0.2011437  0.2014613
#>  [171,]    0.2905988    0.2873124  0.2877662
#>  [172,]    0.1892425    0.1871023  0.1873977
#>  [173,]    0.2367504    0.2340729  0.2344426
#>  [174,]    0.2400375    0.2373229  0.2376977
#>  [175,]    0.2485049    0.2456945  0.2460825
#>  [176,]    0.2764703    0.2733437  0.2737754
#>  [177,]    0.2375713    0.2348846  0.2352555
#>  [178,]    0.3058668    0.3024077  0.3028853
#>  [179,]    0.2107360    0.2083528  0.2086818
#>  [180,]    0.2013044    0.1990278  0.1993421
#>  [181,]    0.1951151    0.1929085  0.1932132
#>  [182,]    0.2015722    0.1992925  0.1996073
#>  [183,]    0.2035506    0.2012486  0.2015664
#>  [184,]    0.1410195    0.1394246  0.1396448
#>  [185,]    0.2254272    0.2228778  0.2232298
#>  [186,]    0.2700504    0.2669964  0.2674181
#>  [187,]    0.1902835    0.1881316  0.1884287
#>  [188,]    0.2200942    0.2176051  0.2179488
#>  [189,]    0.2466706    0.2438810  0.2442661
#>  [190,]    0.2178017    0.2153386  0.2156786
#>  [191,]    0.1881836    0.1860553  0.1863492
#>  [192,]    0.2128057    0.2103991  0.2107313
#>  [193,]    0.1841574    0.1820747  0.1823623
#>  [194,]    0.1949583    0.1927534  0.1930578
#>  [195,]    0.2131189    0.2107087  0.2110415
#>  [196,]    0.2314691    0.2288513  0.2292127
#>  [197,]    0.2091703    0.2068047  0.2071313
#>  [198,]    0.2866821    0.2834400  0.2838876
#>  [199,]    0.2157670    0.2133268  0.2136637
#>  [200,]    0.2134986    0.2110841  0.2114174
#>  [201,]    0.2146088    0.2121818  0.2125169
#>  [202,]    0.1780230    0.1760097  0.1762876
#>  [203,]    0.2727505    0.2696659  0.2700918
#>  [204,]    0.1646330    0.1627711  0.1630282
#>  [205,]    0.2722269    0.2691483  0.2695733
#>  [206,]    0.2689257    0.2658844  0.2663043
#>  [207,]    0.2679337    0.2649037  0.2653220
#>  [208,]    0.2898159    0.2865384  0.2869909
#>  [209,]    0.2078204    0.2054701  0.2057946
#>  [210,]    0.2220906    0.2195789  0.2199257
#>  [211,]    0.2207381    0.2182417  0.2185864
#>  [212,]    0.2205709    0.2180764  0.2184208
#>  [213,]    0.1756060    0.1736200  0.1738942
#>  [214,]    0.1860590    0.1839548  0.1842453
#>  [215,]    0.2421474    0.2394089  0.2397869
#>  [216,]    0.2365793    0.2339038  0.2342732
#>  [217,]    0.2134644    0.2110503  0.2113836
#>  [218,]    0.2370048    0.2343244  0.2346945
#>  [219,]    0.2058372    0.2035093  0.2038307
#>  [220,]    0.2453855    0.2426104  0.2429936
#>  [221,]    0.2104805    0.2081001  0.2084287
#>  [222,]    0.2223570    0.2198423  0.2201895
#>  [223,]    0.2962274    0.2928773  0.2933399
#>  [224,]    0.2139668    0.2115470  0.2118810
#>  [225,]    0.2162268    0.2137815  0.2141191
#>  [226,]    0.2128161    0.2104093  0.2107416
#>  [227,]    0.2111275    0.2087398  0.2090695
#>  [228,]    0.2579233    0.2550064  0.2554091
#>  [229,]    0.2228619    0.2203415  0.2206895
#>  [230,]    0.1800200    0.1779841  0.1782652
#>  [231,]    0.2761946    0.2730711  0.2735023
#>  [232,]    0.2338665    0.2312217  0.2315869
#>  [233,]    0.2044486    0.2021364  0.2024556
#>  [234,]    0.2787779    0.2756252  0.2760605
#>  [235,]    0.2719447    0.2688692  0.2692938
#>  [236,]    0.2591138    0.2561835  0.2565881
#>  [237,]    0.3514372    0.3474628  0.3480115
#>  [238,]    0.2723920    0.2693115  0.2697368
#>  [239,]    0.3131094    0.3095684  0.3100573
#>  [240,]    0.3047484    0.3013020  0.3017778
#>  [241,]    0.2143185    0.2118947  0.2122294
#>  [242,]    0.1909331    0.1887738  0.1890720
#>  [243,]    0.2787082    0.2755563  0.2759915
#>  [244,]    0.2901648    0.2868833  0.2873364
#>  [245,]    0.1686230    0.1667160  0.1669793
#>  [246,]    0.3055753    0.3021195  0.3025967
#>  [247,]    0.2825015    0.2793067  0.2797478
#>  [248,]    0.1979268    0.1956884  0.1959974
#>  [249,]    0.2122294    0.2098292  0.2101606
#>  [250,]    0.2279856    0.2254073  0.2257633
#>  [251,]    0.2443587    0.2415952  0.2419767
#>  [252,]    0.1740647    0.1720961  0.1723679
#>  [253,]    0.2011794    0.1989042  0.1992183
#>  [254,]    0.2536964    0.2508273  0.2512235
#>  [255,]    0.2753612    0.2722471  0.2726771
#>  [256,]    0.1769478    0.1749467  0.1752229
#>  [257,]    0.2137728    0.2113552  0.2116890
#>  [258,]    0.1991284    0.1968764  0.1971874
#>  [259,]    0.2578371    0.2549211  0.2553237
#>  [260,]    0.2297170    0.2271190  0.2274777
#>  [261,]    0.2370449    0.2343641  0.2347342
#>  [262,]    0.3124794    0.3089456  0.3094335
#>  [263,]    0.3086264    0.3051362  0.3056180
#>  [264,]    0.1718499    0.1699064  0.1701747
#>  [265,]    0.1527370    0.1510096  0.1512481
#>  [266,]    0.2200083    0.2175202  0.2178637
#>  [267,]    0.1702075    0.1682825  0.1685483
#>  [268,]    0.2361150    0.2334447  0.2338134
#>  [269,]    0.2750901    0.2719791  0.2724086
#>  [270,]    0.1805598    0.1785178  0.1787997
#>  [271,]    0.2000316    0.1977694  0.1980817
#>  [272,]    0.3047824    0.3013356  0.3018115
#>  [273,]    0.1779182    0.1759060  0.1761838
#>  [274,]    0.1903655    0.1882126  0.1885098
#>  [275,]    0.2530068    0.2501455  0.2505405
#>  [276,]    0.2842812    0.2810663  0.2815102
#>  [277,]    0.3010998    0.2976947  0.2981648
#>  [278,]    0.1877767    0.1856531  0.1859463
#>  [279,]    0.1963837    0.1941627  0.1944693
#>  [280,]    0.2855055    0.2822767  0.2827225
#>  [281,]    0.1879836    0.1858577  0.1861512
#>  [282,]    0.1918748    0.1897048  0.1900044
#>  [283,]    0.1972741    0.1950430  0.1953511
#>  [284,]    0.2669803    0.2639610  0.2643778
#>  [285,]    0.2921646    0.2888605  0.2893167
#>  [286,]    0.1788137    0.1767915  0.1770707
#>  [287,]    0.2131509    0.2107404  0.2110732
#>  [288,]    0.2331996    0.2305623  0.2309265
#>  [289,]    0.2438738    0.2411157  0.2414965
#>  [290,]    0.2476362    0.2448357  0.2452224
#>  [291,]    0.2031024    0.2008055  0.2011226
#>  [292,]    0.2222237    0.2197105  0.2200575
#>  [293,]    0.2172900    0.2148326  0.2151719
#>  [294,]    0.2208115    0.2183143  0.2186591
#>  [295,]    0.2098230    0.2074501  0.2077777
#>  [296,]    0.2090666    0.2067022  0.2070287
#>  [297,]    0.2718643    0.2687897  0.2692142
#>  [298,]    0.2137053    0.2112884  0.2116221
#>  [299,]    0.2210470    0.2185472  0.2188923
#>  [300,]    0.2490850    0.2462680  0.2466570
#>  [301,]    0.3204596    0.3168356  0.3173359
#>  [302,]    0.1484471    0.1467683  0.1470001
#>  [303,]    0.2813617    0.2781797  0.2786190
#>  [304,]    0.2314847    0.2288668  0.2292282
#>  [305,]    0.1970239    0.1947957  0.1951033
#>  [306,]    0.2052358    0.2029147  0.2032352
#>  [307,]    0.2940859    0.2907601  0.2912192
#>  [308,]    0.2664529    0.2634395  0.2638556
#>  [309,]    0.2432837    0.2405323  0.2409122
#>  [310,]    0.1990897    0.1968382  0.1971490
#>  [311,]    0.3026590    0.2992362  0.2997088
#>  [312,]    0.1975510    0.1953168  0.1956253
#>  [313,]    0.2550218    0.2521377  0.2525359
#>  [314,]    0.2092491    0.2068827  0.2072094
#>  [315,]    0.1677160    0.1658192  0.1660811
#>  [316,]    0.1880142    0.1858879  0.1861814
#>  [317,]    0.2093193    0.2069521  0.2072789
#>  [318,]    0.1710039    0.1690700  0.1693370
#>  [319,]    0.1237675    0.1223678  0.1225610
#>  [320,]    0.2144374    0.2120123  0.2123471
#>  [321,]    0.2058406    0.2035127  0.2038341
#>  [322,]    0.2704344    0.2673760  0.2677983
#>  [323,]    0.2364777    0.2338033  0.2341725
#>  [324,]    0.2134613    0.2110473  0.2113806
#>  [325,]    0.2243018    0.2217652  0.2221154
#>  [326,]    0.1959097    0.1936941  0.1940000
#>  [327,]    0.1685108    0.1666051  0.1668682
#>  [328,]    0.2366035    0.2339277  0.2342971
#>  [329,]    0.2320947    0.2294699  0.2298323
#>  [330,]    0.1787449    0.1767234  0.1770025
#>  [331,]    0.1718459    0.1699024  0.1701708
#>  [332,]    0.2258137    0.2232599  0.2236125
#>  [333,]    0.3068906    0.3034199  0.3038991
#>  [334,]    0.2368799    0.2342010  0.2345708
#>  [335,]    0.2087079    0.2063476  0.2066735
#>  [336,]    0.2581261    0.2552069  0.2556099
#>  [337,]    0.1547430    0.1529929  0.1532345
#>  [338,]    0.1932586    0.1910730  0.1913747
#>  [339,]    0.1990070    0.1967564  0.1970671
#>  [340,]    0.2203954    0.2179028  0.2182470
#>  [341,]    0.1674464    0.1655527  0.1658141
#>  [342,]    0.2279759    0.2253977  0.2257537
#>  [343,]    0.2560653    0.2531695  0.2535693
#>  [344,]    0.2620303    0.2590670  0.2594762
#>  [345,]    0.2392587    0.2365529  0.2369265
#>  [346,]    0.2373711    0.2346866  0.2350573
#>  [347,]    0.2035403    0.2012384  0.2015562
#>  [348,]    0.2083638    0.2060073  0.2063327
#>  [349,]    0.2521115    0.2492604  0.2496540
#>  [350,]    0.2434315    0.2406785  0.2410586
#>  [351,]    0.1756027    0.1736167  0.1738909
#>  [352,]    0.1998119    0.1975522  0.1978642
#>  [353,]    0.2408117    0.2380883  0.2384643
#>  [354,]    0.2943441    0.2910154  0.2914750
#>  [355,]    0.2533252    0.2504603  0.2508559
#>  [356,]    0.1764573    0.1744617  0.1747372
#>  [357,]    0.2202906    0.2177993  0.2181433
#>  [358,]    0.2058348    0.2035070  0.2038284
#>  [359,]    0.1762811    0.1742875  0.1745627
#>  [360,]    0.1447414    0.1431044  0.1433304
#>  [361,]    0.2402019    0.2374854  0.2378604
#>  [362,]    0.2501019    0.2472735  0.2476640
#>  [363,]    0.2515416    0.2486969  0.2490896
#>  [364,]    0.2185723    0.2161004  0.2164417
#>  [365,]    0.2178473    0.2153836  0.2157238
#>  [366,]    0.2651713    0.2621724  0.2625864
#>  [367,]    0.1487675    0.1470850  0.1473173
#>  [368,]    0.2251824    0.2226357  0.2229873
#>  [369,]    0.1621460    0.1603122  0.1605654
#>  [370,]    0.2200408    0.2175522  0.2178958
#>  [371,]    0.2877094    0.2844558  0.2849050
#>  [372,]    0.1358334    0.1342972  0.1345093
#>  [373,]    0.2675958    0.2645695  0.2649874
#>  [374,]    0.2020658    0.1997806  0.2000961
#>  [375,]    0.2128880    0.2104804  0.2108128
#>  [376,]    0.2505151    0.2476820  0.2480731
#>  [377,]    0.2626127    0.2596428  0.2600528
#>  [378,]    0.1592780    0.1574766  0.1577253
#>  [379,]    0.2281244    0.2255445  0.2259007
#>  [380,]    0.2311241    0.2285102  0.2288711
#>  [381,]    0.2769169    0.2737852  0.2742176
#>  [382,]    0.1516397    0.1499247  0.1501615
#>  [383,]    0.2120630    0.2096647  0.2099958
#>  [384,]    0.2148230    0.2123935  0.2127290
#>  [385,]    0.3287401    0.3250224  0.3255357
#>  [386,]    0.1607218    0.1589041  0.1591551
#>  [387,]    0.3398831    0.3360394  0.3365701
#>  [388,]    0.2163376    0.2138910  0.2142288
#>  [389,]    0.2202675    0.2177765  0.2181204
#>  [390,]    0.3128057    0.3092682  0.3097566
#>  [391,]    0.2212093    0.2187076  0.2190530
#>  [392,]    0.2985871    0.2952103  0.2956765
#>  [393,]    0.2114181    0.2090271  0.2093572
#>  [394,]    0.2294117    0.2268172  0.2271754
#>  [395,]    0.1784376    0.1764196  0.1766982
#>  [396,]    0.3115856    0.3080619  0.3085484
#>  [397,]    0.2048845    0.2025674  0.2028874
#>  [398,]    0.1948649    0.1926611  0.1929654
#>  [399,]    0.2140873    0.2116661  0.2120004
#>  [400,]    0.1842551    0.1821712  0.1824589
#>  [401,]    0.2077065    0.2053575  0.2056818
#>  [402,]    0.1964493    0.1942276  0.1945344
#>  [403,]    0.1825193    0.1804551  0.1807401
#>  [404,]    0.2831471    0.2799450  0.2803871
#>  [405,]    0.2032705    0.2009716  0.2012890
#>  [406,]    0.1749708    0.1729920  0.1732652
#>  [407,]    0.2248739    0.2223307  0.2226819
#>  [408,]    0.1919760    0.1898049  0.1901046
#>  [409,]    0.2119525    0.2095554  0.2098864
#>  [410,]    0.1750589    0.1730791  0.1733525
#>  [411,]    0.1944012    0.1922026  0.1925062
#>  [412,]    0.2142310    0.2118082  0.2121427
#>  [413,]    0.2259457    0.2233904  0.2237432
#>  [414,]    0.1841229    0.1820406  0.1823281
#>  [415,]    0.3242399    0.3205731  0.3210793
#>  [416,]    0.1480970    0.1464221  0.1466534
#>  [417,]    0.1669644    0.1650761  0.1653368
#>  [418,]    0.1927044    0.1905251  0.1908260
#>  [419,]    0.2066815    0.2043440  0.2046668
#>  [420,]    0.1783218    0.1763051  0.1765835
#>  [421,]    0.2332779    0.2306398  0.2310040
#>  [422,]    0.2536026    0.2507346  0.2511306
#>  [423,]    0.2509904    0.2481519  0.2485438
#>  [424,]    0.2100787    0.2077028  0.2080308
#>  [425,]    0.2278230    0.2252465  0.2256022
#>  [426,]    0.2281989    0.2256182  0.2259745
#>  [427,]    0.1833752    0.1813014  0.1815877
#>  [428,]    0.1608888    0.1590693  0.1593205
#>  [429,]    0.2676327    0.2646061  0.2650239
#>  [430,]    0.1585738    0.1567804  0.1570280
#>  [431,]    0.3167207    0.3131390  0.3136335
#>  [432,]    0.2230212    0.2204990  0.2208472
#>  [433,]    0.3035472    0.3001144  0.3005884
#>  [434,]    0.2717293    0.2686564  0.2690806
#>  [435,]    0.3002747    0.2968789  0.2973478
#>  [436,]    0.2046331    0.2023188  0.2026383
#>  [437,]    0.2816653    0.2784800  0.2789198
#>  [438,]    0.2219066    0.2193970  0.2197435
#>  [439,]    0.2193247    0.2168443  0.2171868
#>  [440,]    0.2126882    0.2102828  0.2106149
#>  [441,]    0.2216029    0.2190967  0.2194427
#>  [442,]    0.2195177    0.2170351  0.2173779
#>  [443,]    0.2923593    0.2890530  0.2895095
#>  [444,]    0.1768683    0.1748680  0.1751442
#>  [445,]    0.2185917    0.2161196  0.2164609
#>  [446,]    0.2213242    0.2188212  0.2191668
#>  [447,]    0.2039278    0.2016215  0.2019399
#>  [448,]    0.3230591    0.3194056  0.3199101
#>  [449,]    0.3344186    0.3306367  0.3311589
#>  [450,]    0.1808707    0.1788252  0.1791076
#>  [451,]    0.1886756    0.1865418  0.1868364
#>  [452,]    0.1829753    0.1809060  0.1811917
#>  [453,]    0.1941986    0.1920024  0.1923056
#>  [454,]    0.1783373    0.1763204  0.1765989
#>  [455,]    0.2182323    0.2157642  0.2161050
#>  [456,]    0.2154255    0.2129892  0.2133256
#>  [457,]    0.1702669    0.1683413  0.1686071
#>  [458,]    0.2232228    0.2206983  0.2210468
#>  [459,]    0.2547706    0.2518894  0.2522872
#>  [460,]    0.2431491    0.2403993  0.2407789
#>  [461,]    0.2241091    0.2215746  0.2219245
#>  [462,]    0.1641810    0.1623242  0.1625805
#>  [463,]    0.1713822    0.1694439  0.1697115
#>  [464,]    0.2093388    0.2069713  0.2072982
#>  [465,]    0.2111553    0.2087673  0.2090970
#>  [466,]    0.2671700    0.2641486  0.2645657
#>  [467,]    0.2123842    0.2099823  0.2103139
#>  [468,]    0.1890645    0.1869263  0.1872215
#>  [469,]    0.1910441    0.1888835  0.1891818
#>  [470,]    0.3045056    0.3010619  0.3015373
#>  [471,]    0.2534908    0.2506241  0.2510199
#>  [472,]    0.1336650    0.1321534  0.1323621
#>  [473,]    0.1956737    0.1934607  0.1937663
#>  [474,]    0.1654826    0.1636112  0.1638695
#>  [475,]    0.2388040    0.2361033  0.2364762
#>  [476,]    0.2010194    0.1987460  0.1990599
#>  [477,]    0.2506385    0.2478040  0.2481953
#>  [478,]    0.2125002    0.2100970  0.2104288
#>  [479,]    0.2007956    0.1985247  0.1988382
#>  [480,]    0.3468236    0.3429014  0.3434429
#>  [481,]    0.1866363    0.1845256  0.1848170
#>  [482,]    0.1933898    0.1912027  0.1915047
#>  [483,]    0.2452227    0.2424494  0.2428323
#>  [484,]    0.1604685    0.1586537  0.1589043
#>  [485,]    0.1361677    0.1346277  0.1348403
#>  [486,]    0.2051686    0.2028483  0.2031686
#>  [487,]    0.1572347    0.1554564  0.1557020
#>  [488,]    0.2174527    0.2149935  0.2153330
#>  [489,]    0.2814780    0.2782948  0.2787343
#>  [490,]    0.2069930    0.2046520  0.2049752
#>  [491,]    0.1813631    0.1793120  0.1795952
#>  [492,]    0.2403717    0.2376533  0.2380286
#>  [493,]    0.1903363    0.1881837  0.1884809
#>  [494,]    0.2320767    0.2294520  0.2298144
#>  [495,]    0.2798135    0.2766491  0.2770860
#>  [496,]    0.1954278    0.1932176  0.1935228
#>  [497,]    0.2300373    0.2274357  0.2277949
#>  [498,]    0.2389073    0.2362055  0.2365785
#>  [499,]    0.2540140    0.2511413  0.2515379
#>  [500,]    0.1917316    0.1895633  0.1898627
#>  [501,]    0.2807053    0.2775308  0.2779691
#>  [502,]    0.2036361    0.2013331  0.2016511
#>  [503,]    0.2368829    0.2342040  0.2345738
#>  [504,]    0.1774868    0.1754796  0.1757567
#>  [505,]    0.1562746    0.1545072  0.1547512
#>  [506,]    0.1912860    0.1891227  0.1894214
#>  [507,]    0.2437843    0.2410274  0.2414080
#>  [508,]    0.1725059    0.1705550  0.1708244
#>  [509,]    0.1866775    0.1845663  0.1848578
#>  [510,]    0.2480519    0.2452466  0.2456339
#>  [511,]    0.2362691    0.2335971  0.2339660
#>  [512,]    0.1865932    0.1844830  0.1847743
#>  [513,]    0.2169842    0.2145303  0.2148691
#>  [514,]    0.2065608    0.2042248  0.2045473
#>  [515,]    0.1743897    0.1724175  0.1726898
#>  [516,]    0.2847186    0.2814988  0.2819433
#>  [517,]    0.2260503    0.2234938  0.2238468
#>  [518,]    0.1724668    0.1705164  0.1707857
#>  [519,]    0.2096062    0.2072357  0.2075630
#>  [520,]    0.1643852    0.1625261  0.1627828
#>  [521,]    0.3095125    0.3060122  0.3064955
#>  [522,]    0.2536156    0.2507474  0.2511434
#>  [523,]    0.2112096    0.2088209  0.2091507
#>  [524,]    0.2147306    0.2123022  0.2126374
#>  [525,]    0.2759741    0.2728531  0.2732840
#>  [526,]    0.2018359    0.1995533  0.1998684
#>  [527,]    0.2233332    0.2208075  0.2211562
#>  [528,]    0.2117745    0.2093795  0.2097102
#>  [529,]    0.1477868    0.1461154  0.1463462
#>  [530,]    0.2211409    0.2186399  0.2189852
#>  [531,]    0.1964553    0.1942335  0.1945402
#>  [532,]    0.2000209    0.1977588  0.1980711
#>  [533,]    0.1876904    0.1855678  0.1858609
#>  [534,]    0.2313947    0.2287778  0.2291391
#>  [535,]    0.2313816    0.2287648  0.2291261
#>  [536,]    0.1984979    0.1962530  0.1965630
#>  [537,]    0.2004108    0.1981443  0.1984572
#>  [538,]    0.2257720    0.2232187  0.2235712
#>  [539,]    0.1959741    0.1937577  0.1940637
#>  [540,]    0.2053061    0.2029843  0.2033048
#>  [541,]    0.1789367    0.1769130  0.1771924
#>  [542,]    0.2540676    0.2511943  0.2515910
#>  [543,]    0.2569240    0.2540184  0.2544196
#>  [544,]    0.2547478    0.2518668  0.2522646
#>  [545,]    0.2296392    0.2270422  0.2274008
#>  [546,]    0.2012435    0.1989676  0.1992818
#>  [547,]    0.2568012    0.2538970  0.2542980
#>  [548,]    0.1920651    0.1898930  0.1901929
#>  [549,]    0.2499342    0.2471077  0.2474979
#>  [550,]    0.2814274    0.2782448  0.2786842
#>  [551,]    0.3217768    0.3181379  0.3186403
#>  [552,]    0.3153931    0.3118263  0.3123188
#>  [553,]    0.1783194    0.1763027  0.1765811
#>  [554,]    0.1912958    0.1891324  0.1894311
#>  [555,]    0.2399594    0.2372457  0.2376203
#>  [556,]    0.2200308    0.2175424  0.2178860
#>  [557,]    0.2501844    0.2473551  0.2477457
#>  [558,]    0.2460627    0.2432800  0.2436642
#>  [559,]    0.2736859    0.2705907  0.2710181
#>  [560,]    0.2033109    0.2010116  0.2013290
#>  [561,]    0.2190456    0.2165683  0.2169103
#>  [562,]    0.1913964    0.1892318  0.1895307
#>  [563,]    0.1724995    0.1705486  0.1708180
#>  [564,]    0.1877702    0.1856466  0.1859398
#>  [565,]    0.2344480    0.2317966  0.2321627
#>  [566,]    0.1604052    0.1585911  0.1588416
#>  [567,]    0.2054501    0.2031266  0.2034474
#>  [568,]    0.2497944    0.2469695  0.2473595
#>  [569,]    0.2844529    0.2812361  0.2816802
#>  [570,]    0.1651417    0.1632740  0.1635319
#>  [571,]    0.2306338    0.2280255  0.2283856
#>  [572,]    0.2146420    0.2122146  0.2125497
#>  [573,]    0.1921111    0.1899385  0.1902384
#>  [574,]    0.2638479    0.2608640  0.2612760
#>  [575,]    0.2645854    0.2615932  0.2620063
#>  [576,]    0.2063474    0.2040138  0.2043360
#>  [577,]    0.2371407    0.2344589  0.2348291
#>  [578,]    0.2140876    0.2116664  0.2120007
#>  [579,]    0.2373500    0.2346658  0.2350364
#>  [580,]    0.2152439    0.2128096  0.2131457
#>  [581,]    0.1862556    0.1841492  0.1844400
#>  [582,]    0.2119927    0.2095952  0.2099262
#>  [583,]    0.2286193    0.2260338  0.2263908
#>  [584,]    0.2299794    0.2273786  0.2277376
#>  [585,]    0.1917000    0.1895320  0.1898313
#>  [586,]    0.2329343    0.2303000  0.2306637
#>  [587,]    0.2601442    0.2572023  0.2576085
#>  [588,]    0.1985804    0.1963346  0.1966446
#>  [589,]    0.2580037    0.2550859  0.2554887
#>  [590,]    0.1620227    0.1601903  0.1604433
#>  [591,]    0.2239842    0.2214511  0.2218009
#>  [592,]    0.1794504    0.1774209  0.1777011
#>  [593,]    0.2731504    0.2700613  0.2704878
#>  [594,]    0.1662600    0.1643797  0.1646393
#>  [595,]    0.1930123    0.1908295  0.1911309
#>  [596,]    0.2026081    0.2003168  0.2006332
#>  [597,]    0.2135948    0.2111792  0.2115127
#>  [598,]    0.2540319    0.2511590  0.2515556
#>  [599,]    0.1885700    0.1864373  0.1867318
#>  [600,]    0.1553452    0.1535883  0.1538309
#>  [601,]    0.1483890    0.1467108  0.1469425
#>  [602,]    0.2004540    0.1981870  0.1985000
#>  [603,]    0.2673681    0.2643444  0.2647619
#>  [604,]    0.1972039    0.1949736  0.1952816
#>  [605,]    0.1983603    0.1961169  0.1964267
#>  [606,]    0.1964981    0.1942758  0.1945827
#>  [607,]    0.2122397    0.2098394  0.2101708
#>  [608,]    0.2087655    0.2064046  0.2067305
#>  [609,]    0.2187888    0.2163144  0.2166561
#>  [610,]    0.1942096    0.1920132  0.1923164
#>  [611,]    0.1730001    0.1710436  0.1713137
#>  [612,]    0.1557017    0.1539408  0.1541840
#>  [613,]    0.2065874    0.2042510  0.2045736
#>  [614,]    0.2437505    0.2409939  0.2413745
#>  [615,]    0.2265924    0.2240299  0.2243837
#>  [616,]    0.1659202    0.1640437  0.1643028
#>  [617,]    0.1760977    0.1741062  0.1743811
#>  [618,]    0.1986567    0.1964100  0.1967202
#>  [619,]    0.1983309    0.1960880  0.1963976
#>  [620,]    0.2090007    0.2066371  0.2069634
#>  [621,]    0.2203335    0.2178417  0.2181858
#>  [622,]    0.2317410    0.2291202  0.2294820
#>  [623,]    0.2170400    0.2145855  0.2149244
#>  [624,]    0.1998973    0.1976366  0.1979487
#>  [625,]    0.2112614    0.2088722  0.2092021
#>  [626,]    0.2355418    0.2328780  0.2332458
#>  [627,]    0.2704616    0.2674029  0.2678252
#>  [628,]    0.1788210    0.1767987  0.1770779
#>  [629,]    0.2191717    0.2166931  0.2170353
#>  [630,]    0.2771557    0.2740213  0.2744541
#>  [631,]    0.2123545    0.2099530  0.2102845
#>  [632,]    0.2707933    0.2677309  0.2681537
#>  [633,]    0.2713553    0.2682866  0.2687103
#>  [634,]    0.1996199    0.1973623  0.1976740
#>  [635,]    0.1767911    0.1747917  0.1750678
#>  [636,]    0.2072278    0.2048842  0.2052078
#>  [637,]    0.2188255    0.2163507  0.2166924
#>  [638,]    0.1998623    0.1976020  0.1979141
#>  [639,]    0.2468990    0.2441067  0.2444922
#>  [640,]    0.1643799    0.1625208  0.1627775
#>  [641,]    0.2980006    0.2946305  0.2950958
#>  [642,]    0.1716884    0.1697467  0.1700147
#>  [643,]    0.1946135    0.1924126  0.1927164
#>  [644,]    0.2933183    0.2900012  0.2904591
#>  [645,]    0.2662525    0.2632414  0.2636572
#>  [646,]    0.2844544    0.2812375  0.2816816
#>  [647,]    0.1311530    0.1296697  0.1298745
#>  [648,]    0.2354591    0.2327962  0.2331639
#>  [649,]    0.1619339    0.1601025  0.1603553
#>  [650,]    0.2244804    0.2219417  0.2222922
#>  [651,]    0.1701056    0.1681818  0.1684474
#>  [652,]    0.2325774    0.2299472  0.2303103
#>  [653,]    0.2115213    0.2091291  0.2094594
#>  [654,]    0.1937437    0.1915526  0.1918551
#>  [655,]    0.1975845    0.1953499  0.1956585
#>  [656,]    0.2449228    0.2421529  0.2425354
#>  [657,]    0.2056637    0.2033378  0.2036589
#>  [658,]    0.2252657    0.2227181  0.2230699
#>  [659,]    0.2032231    0.2009248  0.2012421
#>  [660,]    0.2173102    0.2148526  0.2151919
#>  [661,]    0.1854484    0.1833511  0.1836407
#>  [662,]    0.2709807    0.2679161  0.2683392
#>  [663,]    0.1761869    0.1741944  0.1744695
#>  [664,]    0.2138040    0.2113860  0.2117199
#>  [665,]    0.1645211    0.1626604  0.1629173
#>  [666,]    0.2231068    0.2205836  0.2209320
#>  [667,]    0.2466025    0.2438136  0.2441987
#>  [668,]    0.1698082    0.1678878  0.1681529
#>  [669,]    0.2282117    0.2256308  0.2259871
#>  [670,]    0.2741696    0.2710691  0.2714972
#>  [671,]    0.2285789    0.2259939  0.2263508
#>  [672,]    0.3212777    0.3176444  0.3181460
#>  [673,]    0.2277084    0.2251331  0.2254887
#>  [674,]    0.2871027    0.2838559  0.2843041
#>  [675,]    0.2444599    0.2416953  0.2420770
#>  [676,]    0.2108994    0.2085143  0.2088436
#>  [677,]    0.2259821    0.2234265  0.2237793
#>  [678,]    0.2458168    0.2430368  0.2434206
#>  [679,]    0.1934869    0.1912987  0.1916008
#>  [680,]    0.2008648    0.1985932  0.1989068
#>  [681,]    0.2333249    0.2306862  0.2310505
#>  [682,]    0.1924155    0.1902394  0.1905398
#>  [683,]    0.2118079    0.2094125  0.2097433
#>  [684,]    0.1786442    0.1766238  0.1769028
#>  [685,]    0.2190336    0.2165565  0.2168985
#>  [686,]    0.2419973    0.2392605  0.2396384
#>  [687,]    0.2167831    0.2143314  0.2146699
#>  [688,]    0.1915094    0.1893435  0.1896426
#>  [689,]    0.2495798    0.2467573  0.2471470
#>  [690,]    0.1853745    0.1832780  0.1835675
#>  [691,]    0.1589010    0.1571039  0.1573520
#>  [692,]    0.2375904    0.2349034  0.2352744
#>  [693,]    0.2525320    0.2496761  0.2500704
#>  [694,]    0.2706228    0.2675623  0.2679849
#>  [695,]    0.2065168    0.2041813  0.2045038
#>  [696,]    0.1876604    0.1855381  0.1858311
#>  [697,]    0.1699702    0.1680479  0.1683133
#>  [698,]    0.3117545    0.3082289  0.3087156
#>  [699,]    0.1902804    0.1881285  0.1884256
#>  [700,]    0.1976335    0.1953984  0.1957070
#>  [701,]    0.1840831    0.1820012  0.1822887
#>  [702,]    0.2455218    0.2427451  0.2431285
#>  [703,]    0.1743444    0.1723727  0.1726449
#>  [704,]    0.1891803    0.1870408  0.1873361
#>  [705,]    0.1716365    0.1696954  0.1699634
#>  [706,]    0.1355566    0.1340235  0.1342352
#>  [707,]    0.1876271    0.1855052  0.1857982
#>  [708,]    0.1958804    0.1936651  0.1939710
#>  [709,]    0.1746680    0.1726926  0.1729653
#>  [710,]    0.3103602    0.3068503  0.3073349
#>  [711,]    0.2116311    0.2092376  0.2095681
#>  [712,]    0.1902823    0.1881303  0.1884274
#>  [713,]    0.2021929    0.1999062  0.2002219
#>  [714,]    0.3039855    0.3005477  0.3010224
#>  [715,]    0.2454087    0.2426334  0.2430166
#>  [716,]    0.3033165    0.2998863  0.3003599
#>  [717,]    0.2018282    0.1995456  0.1998608
#>  [718,]    0.1904565    0.1883026  0.1886000
#>  [719,]    0.2091502    0.2067848  0.2071114
#>  [720,]    0.2655494    0.2625462  0.2629609
#>  [721,]    0.1821299    0.1800701  0.1803545
#>  [722,]    0.2063909    0.2040567  0.2043790
#>  [723,]    0.2278638    0.2252868  0.2256426
#>  [724,]    0.2427471    0.2400019  0.2403809
#>  [725,]    0.1946352    0.1924340  0.1927379
#>  [726,]    0.2419844    0.2392478  0.2396256
#>  [727,]    0.2700055    0.2669520  0.2673736
#>  [728,]    0.1952940    0.1930854  0.1933903
#>  [729,]    0.2087319    0.2063713  0.2066973
#>  [730,]    0.3581155    0.3540656  0.3546248
#>  [731,]    0.2536980    0.2508290  0.2512251
#>  [732,]    0.2721567    0.2690789  0.2695039
#>  [733,]    0.2196274    0.2171435  0.2174865
#>  [734,]    0.1823963    0.1803336  0.1806184
#>  [735,]    0.1862148    0.1841088  0.1843996
#>  [736,]    0.1980100    0.1957706  0.1960798
#>  [737,]    0.1499576    0.1482616  0.1484958
#>  [738,]    0.1833306    0.1812572  0.1815435
#>  [739,]    0.1904622    0.1883082  0.1886056
#>  [740,]    0.2056172    0.2032918  0.2036129
#>  [741,]    0.2029690    0.2006736  0.2009905
#>  [742,]    0.2279115    0.2253340  0.2256898
#>  [743,]    0.1696578    0.1677391  0.1680040
#>  [744,]    0.2525360    0.2496800  0.2500743
#>  [745,]    0.2202839    0.2177926  0.2181366
#>  [746,]    0.2026526    0.2003608  0.2006772
#>  [747,]    0.2936688    0.2903477  0.2908063
#>  [748,]    0.3051728    0.3017217  0.3021982
#>  [749,]    0.1822601    0.1801989  0.1804835
#>  [750,]    0.2562892    0.2533908  0.2537910
#>  [751,]    0.1892846    0.1871439  0.1874395
#>  [752,]    0.2378587    0.2351687  0.2355401
#>  [753,]    0.1586491    0.1568548  0.1571026
#>  [754,]    0.2274981    0.2249252  0.2252805
#>  [755,]    0.1516664    0.1499512  0.1501880
#>  [756,]    0.1676191    0.1657234  0.1659852
#>  [757,]    0.2263636    0.2238036  0.2241570
#>  [758,]    0.1797718    0.1777387  0.1780194
#>  [759,]    0.1905526    0.1883976  0.1886951
#>  [760,]    0.2749994    0.2718894  0.2723188
#>  [761,]    0.1714503    0.1695113  0.1697790
#>  [762,]    0.1869699    0.1848554  0.1851473
#>  [763,]    0.2481942    0.2453873  0.2457749
#>  [764,]    0.2390884    0.2363845  0.2367578
#>  [765,]    0.1940965    0.1919014  0.1922045
#>  [766,]    0.2385840    0.2358858  0.2362584
#>  [767,]    0.2373741    0.2346896  0.2350602
#>  [768,]    0.1715544    0.1696142  0.1698821
#>  [769,]    0.1880618    0.1859350  0.1862286
#>  [770,]    0.2448613    0.2420921  0.2424745
#>  [771,]    0.2074060    0.2050604  0.2053842
#>  [772,]    0.3245860    0.3209153  0.3214221
#>  [773,]    0.1849385    0.1828470  0.1831358
#>  [774,]    0.2150406    0.2126087  0.2129444
#>  [775,]    0.1903796    0.1882265  0.1885238
#>  [776,]    0.2216267    0.2191203  0.2194663
#>  [777,]    0.2232182    0.2206938  0.2210423
#>  [778,]    0.2280495    0.2254705  0.2258266
#>  [779,]    0.1675099    0.1656154  0.1658770
#>  [780,]    0.2522669    0.2494140  0.2498079
#>  [781,]    0.1981710    0.1959298  0.1962392
#>  [782,]    0.2081409    0.2057870  0.2061120
#>  [783,]    0.2382648    0.2355702  0.2359423
#>  [784,]    0.3192517    0.3156413  0.3161397
#>  [785,]    0.1825536    0.1804891  0.1807741
#>  [786,]    0.2094835    0.2071144  0.2074415
#>  [787,]    0.3305967    0.3268581  0.3273742
#>  [788,]    0.2245723    0.2220326  0.2223832
#>  [789,]    0.2223985    0.2198833  0.2202306
#>  [790,]    0.2819377    0.2787493  0.2791895
#>  [791,]    0.1743985    0.1724262  0.1726985
#>  [792,]    0.1909923    0.1888324  0.1891306
#>  [793,]    0.1570117    0.1552359  0.1554811
#>  [794,]    0.2758586    0.2727389  0.2731696
#>  [795,]    0.2233918    0.2208654  0.2212142
#>  [796,]    0.3029173    0.2994917  0.2999646
#>  [797,]    0.1493416    0.1476526  0.1478858
#>  [798,]    0.1881511    0.1860233  0.1863171
#>  [799,]    0.2005004    0.1982329  0.1985460
#>  [800,]    0.2322018    0.2295758  0.2299383
#>  [801,]    0.2287043    0.2261179  0.2264750
#>  [802,]    0.2203769    0.2178846  0.2182287
#>  [803,]    0.2551612    0.2522756  0.2526740
#>  [804,]    0.1925400    0.1903626  0.1906632
#>  [805,]    0.1969696    0.1947420  0.1950496
#>  [806,]    0.1792295    0.1772025  0.1774824
#>  [807,]    0.2078227    0.2054724  0.2057969
#>  [808,]    0.2129486    0.2105403  0.2108728
#>  [809,]    0.1686381    0.1667309  0.1669942
#>  [810,]    0.2282545    0.2256732  0.2260296
#>  [811,]    0.2435637    0.2408092  0.2411895
#>  [812,]    0.2003329    0.1980673  0.1983801
#>  [813,]    0.1927567    0.1905768  0.1908778
#>  [814,]    0.2057350    0.2034083  0.2037295
#>  [815,]    0.2604379    0.2574926  0.2578992
#>  [816,]    0.1428340    0.1412187  0.1414417
#>  [817,]    0.2220728    0.2195614  0.2199081
#>  [818,]    0.1756727    0.1736860  0.1739603
#>  [819,]    0.2571096    0.2542019  0.2546034
#>  [820,]    0.1965621    0.1943391  0.1946461
#>  [821,]    0.2077779    0.2054281  0.2057525
#>  [822,]    0.2230321    0.2205097  0.2208580
#>  [823,]    0.1765730    0.1745761  0.1748518
#>  [824,]    0.2517719    0.2489245  0.2493177
#>  [825,]    0.2734955    0.2704025  0.2708295
#>  [826,]    0.3314636    0.3277151  0.3282327
#>  [827,]    0.1895316    0.1873881  0.1876840
#>  [828,]    0.2272825    0.2247122  0.2250671
#>  [829,]    0.1912000    0.1890376  0.1893362
#>  [830,]    0.2233781    0.2208519  0.2212007
#>  [831,]    0.2406939    0.2379719  0.2383477
#>  [832,]    0.2499266    0.2471002  0.2474904
#>  [833,]    0.2026117    0.2003203  0.2006366
#>  [834,]    0.1849438    0.1828522  0.1831410
#>  [835,]    0.1479608    0.1462874  0.1465185
#>  [836,]    0.1581388    0.1563503  0.1565972
#>  [837,]    0.1580279    0.1562407  0.1564875
#>  [838,]    0.1864608    0.1843521  0.1846433
#>  [839,]    0.2266442    0.2240810  0.2244349
#>  [840,]    0.2166398    0.2141897  0.2145280
#>  [841,]    0.2048625    0.2025457  0.2028655
#>  [842,]    0.2508546    0.2480177  0.2484094
#>  [843,]    0.2208584    0.2183606  0.2187055
#>  [844,]    0.2795741    0.2764124  0.2768489
#>  [845,]    0.3043576    0.3009156  0.3013908
#>  [846,]    0.1668102    0.1649237  0.1651842
#>  [847,]    0.2159133    0.2134715  0.2138087
#>  [848,]    0.2115474    0.2091549  0.2094852
#>  [849,]    0.2623395    0.2593727  0.2597823
#>  [850,]    0.2525495    0.2496934  0.2500877
#>  [851,]    0.2121027    0.2097039  0.2100351
#>  [852,]    0.2117673    0.2093724  0.2097031
#>  [853,]    0.1857287    0.1836282  0.1839182
#>  [854,]    0.2213370    0.2188338  0.2191794
#>  [855,]    0.1963878    0.1941668  0.1944735
#>  [856,]    0.2122003    0.2098005  0.2101319
#>  [857,]    0.2956973    0.2923533  0.2928150
#>  [858,]    0.2272171    0.2246475  0.2250022
#>  [859,]    0.3299910    0.3262591  0.3267743
#>  [860,]    0.1358182    0.1342822  0.1344942
#>  [861,]    0.2090942    0.2067295  0.2070560
#>  [862,]    0.1987777    0.1965297  0.1968401
#>  [863,]    0.1592449    0.1574439  0.1576926
#>  [864,]    0.2608808    0.2579305  0.2583378
#>  [865,]    0.2921520    0.2888480  0.2893042
#>  [866,]    0.1632980    0.1614512  0.1617062
#>  [867,]    0.2780024    0.2748585  0.2752925
#>  [868,]    0.1897462    0.1876003  0.1878966
#>  [869,]    0.1895651    0.1874213  0.1877173
#>  [870,]    0.1910793    0.1889183  0.1892167
#>  [871,]    0.1698250    0.1679044  0.1681696
#>  [872,]    0.2159059    0.2134641  0.2138012
#>  [873,]    0.2246720    0.2221311  0.2224820
#>  [874,]    0.2311328    0.2285188  0.2288797
#>  [875,]    0.1504317    0.1487304  0.1489653
#>  [876,]    0.3104816    0.3069703  0.3074551
#>  [877,]    0.2113079    0.2089182  0.2092481
#>  [878,]    0.2360823    0.2334124  0.2337810
#>  [879,]    0.2071434    0.2048007  0.2051242
#>  [880,]    0.2610422    0.2580900  0.2584976
#>  [881,]    0.2494417    0.2466208  0.2470103
#>  [882,]    0.2325827    0.2299523  0.2303155
#>  [883,]    0.2373167    0.2346329  0.2350034
#>  [884,]    0.2347696    0.2321145  0.2324811
#>  [885,]    0.1888819    0.1867457  0.1870407
#>  [886,]    0.2260262    0.2234700  0.2238230
#>  [887,]    0.2254021    0.2228530  0.2232049
#>  [888,]    0.2019694    0.1996853  0.2000006
#>  [889,]    0.2006680    0.1983986  0.1987120
#>  [890,]    0.2130237    0.2106146  0.2109472
#>  [891,]    0.2413692    0.2386395  0.2390163
#>  [892,]    0.1841219    0.1820395  0.1823270
#>  [893,]    0.2197424    0.2172573  0.2176004
#>  [894,]    0.1758614    0.1738725  0.1741471
#>  [895,]    0.2699497    0.2668969  0.2673184
#>  [896,]    0.2885548    0.2852916  0.2857421
#>  [897,]    0.2482607    0.2454531  0.2458407
#>  [898,]    0.1504305    0.1487293  0.1489641
#>  [899,]    0.1790343    0.1770095  0.1772890
#>  [900,]    0.1706485    0.1687185  0.1689850
#>  [901,]    0.1860363    0.1839323  0.1842228
#>  [902,]    0.1784844    0.1764658  0.1767445
#>  [903,]    0.2313333    0.2287171  0.2290783
#>  [904,]    0.2376721    0.2349843  0.2353554
#>  [905,]    0.2546015    0.2517221  0.2521197
#>  [906,]    0.1805684    0.1785263  0.1788082
#>  [907,]    0.2730575    0.2699695  0.2703958
#>  [908,]    0.2679531    0.2649228  0.2653412
#>  [909,]    0.1803596    0.1783198  0.1786014
#>  [910,]    0.1806057    0.1785632  0.1788452
#>  [911,]    0.2154065    0.2129704  0.2133068
#>  [912,]    0.2008906    0.1986187  0.1989323
#>  [913,]    0.3437063    0.3398193  0.3403560
#>  [914,]    0.1680669    0.1661661  0.1664285
#>  [915,]    0.1372317    0.1356797  0.1358940
#>  [916,]    0.2385182    0.2358207  0.2361932
#>  [917,]    0.1852062    0.1831117  0.1834009
#>  [918,]    0.2031506    0.2008532  0.2011704
#>  [919,]    0.2288602    0.2262720  0.2266294
#>  [920,]    0.1820172    0.1799587  0.1802429
#>  [921,]    0.2288949    0.2263063  0.2266637
#>  [922,]    0.2327573    0.2301251  0.2304885
#>  [923,]    0.1832358    0.1811636  0.1814497
#>  [924,]    0.3292968    0.3255728  0.3260870
#>  [925,]    0.2479901    0.2451855  0.2455728
#>  [926,]    0.2680015    0.2649706  0.2653890
#>  [927,]    0.2200291    0.2175408  0.2178843
#>  [928,]    0.1971331    0.1949036  0.1952114
#>  [929,]    0.2523966    0.2495422  0.2499363
#>  [930,]    0.2686946    0.2656559  0.2660755
#>  [931,]    0.2231593    0.2206355  0.2209840
#>  [932,]    0.2285927    0.2260075  0.2263645
#>  [933,]    0.2125373    0.2101336  0.2104655
#>  [934,]    0.1821351    0.1800752  0.1803596
#>  [935,]    0.2228928    0.2203720  0.2207201
#>  [936,]    0.3566258    0.3525928  0.3531497
#>  [937,]    0.2081041    0.2057506  0.2060755
#>  [938,]    0.1985818    0.1963360  0.1966460
#>  [939,]    0.1918686    0.1896987  0.1899983
#>  [940,]    0.2190043    0.2165275  0.2168695
#>  [941,]    0.2286736    0.2260875  0.2264446
#>  [942,]    0.2043754    0.2020641  0.2023832
#>  [943,]    0.2122395    0.2098392  0.2101706
#>  [944,]    0.2033009    0.2010018  0.2013192
#>  [945,]    0.2024507    0.2001611  0.2004773
#>  [946,]    0.2259091    0.2233543  0.2237070
#>  [947,]    0.2311223    0.2285084  0.2288693
#>  [948,]    0.2153831    0.2129473  0.2132836
#>  [949,]    0.2539564    0.2510844  0.2514809
#>  [950,]    0.1896737    0.1875286  0.1878248
#>  [951,]    0.1727834    0.1708294  0.1710991
#>  [952,]    0.1808999    0.1788540  0.1791365
#>  [953,]    0.1892698    0.1871293  0.1874248
#>  [954,]    0.2112493    0.2088603  0.2091901
#>  [955,]    0.3116545    0.3081300  0.3086166
#>  [956,]    0.2276687    0.2250940  0.2254495
#>  [957,]    0.2079467    0.2055950  0.2059197
#>  [958,]    0.2132980    0.2108858  0.2112189
#>  [959,]    0.1589516    0.1571540  0.1574022
#>  [960,]    0.2232995    0.2207742  0.2211228
#>  [961,]    0.2634547    0.2604753  0.2608867
#>  [962,]    0.2174026    0.2149440  0.2152834
#>  [963,]    0.1964970    0.1942748  0.1945816
#>  [964,]    0.2683121    0.2652777  0.2656967
#>  [965,]    0.2022746    0.1999870  0.2003029
#>  [966,]    0.2830945    0.2798930  0.2803350
#>  [967,]    0.2247052    0.2221639  0.2225148
#>  [968,]    0.3314678    0.3277193  0.3282369
#>  [969,]    0.2393985    0.2366911  0.2370649
#>  [970,]    0.2761418    0.2730189  0.2734500
#>  [971,]    0.1755281    0.1735430  0.1738171
#>  [972,]    0.2707051    0.2676436  0.2680663
#>  [973,]    0.1870127    0.1848977  0.1851897
#>  [974,]    0.1948997    0.1926955  0.1929999
#>  [975,]    0.2964792    0.2931263  0.2935892
#>  [976,]    0.1940988    0.1919037  0.1922068
#>  [977,]    0.2238369    0.2213054  0.2216550
#>  [978,]    0.2408649    0.2381409  0.2385170
#>  [979,]    0.2102607    0.2078828  0.2082111
#>  [980,]    0.1816426    0.1795883  0.1798720
#>  [981,]    0.1853338    0.1832378  0.1835272
#>  [982,]    0.2537209    0.2508515  0.2512477
#>  [983,]    0.3119346    0.3084070  0.3088940
#>  [984,]    0.2256135    0.2230620  0.2234143
#>  [985,]    0.2787834    0.2756306  0.2760659
#>  [986,]    0.1819922    0.1799339  0.1802181
#>  [987,]    0.3249417    0.3212670  0.3217743
#>  [988,]    0.1837735    0.1816952  0.1819821
#>  [989,]    0.1647052    0.1628425  0.1630997
#>  [990,]    0.2097007    0.2073291  0.2076566
#>  [991,]    0.2097284    0.2073565  0.2076840
#>  [992,]    0.1823470    0.1802848  0.1805695
#>  [993,]    0.2993017    0.2959168  0.2963842
#>  [994,]    0.3381440    0.3343200  0.3348480
#>  [995,]    0.2274037    0.2248320  0.2251871
#>  [996,]    0.2240055    0.2214722  0.2218219
#>  [997,]    0.1911130    0.1889516  0.1892500
#>  [998,]    0.2178570    0.2153932  0.2157334
#>  [999,]    0.2586012    0.2556766  0.2560804
#> [1000,]    0.1994399    0.1971844  0.1974958