Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 97.51702
#> 2   Feb 86.62384
#> 3   Mar 93.32879
#> 4   Apr 85.71288
#> 5   May 94.10399
#> 6   Jun 97.59296

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 8.941696  2
#> 2   Feb 7.959201  2
#> 3   Mar 7.006742  2
#> 4   Apr 6.622739  2
#> 5   May 9.916925  2
#> 6   Jun 7.787127  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [12ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [22ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [65ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [701ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [19ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1944175    0.2033727  0.1999294
#>    [2,]    0.2010590    0.2103201  0.2067592
#>    [3,]    0.2239035    0.2342169  0.2302514
#>    [4,]    0.1882994    0.1969729  0.1936380
#>    [5,]    0.2056782    0.2151522  0.2115095
#>    [6,]    0.2697797    0.2822059  0.2774280
#>    [7,]    0.2233166    0.2336030  0.2296479
#>    [8,]    0.1735673    0.1815622  0.1784882
#>    [9,]    0.2444867    0.2557480  0.2514181
#>   [10,]    0.3780927    0.3955073  0.3888114
#>   [11,]    0.3006288    0.3144757  0.3091516
#>   [12,]    0.2351926    0.2460258  0.2418605
#>   [13,]    0.1934845    0.2023970  0.1989702
#>   [14,]    0.1803035    0.1886087  0.1854154
#>   [15,]    0.2642951    0.2764688  0.2717881
#>   [16,]    0.2159161    0.2258616  0.2220376
#>   [17,]    0.2444166    0.2556747  0.2513460
#>   [18,]    0.2322911    0.2429908  0.2388768
#>   [19,]    0.1452737    0.1519654  0.1493924
#>   [20,]    0.2488380    0.2602998  0.2558928
#>   [21,]    0.1776962    0.1858813  0.1827341
#>   [22,]    0.2447319    0.2560044  0.2516702
#>   [23,]    0.2095088    0.2191592  0.2154486
#>   [24,]    0.2028053    0.2121470  0.2085551
#>   [25,]    0.2243635    0.2346980  0.2307244
#>   [26,]    0.1863423    0.1949257  0.1916254
#>   [27,]    0.2705355    0.2829967  0.2782054
#>   [28,]    0.2368728    0.2477835  0.2435884
#>   [29,]    0.2011840    0.2104509  0.2068878
#>   [30,]    0.1697079    0.1775248  0.1745192
#>   [31,]    0.2541523    0.2658589  0.2613577
#>   [32,]    0.2406407    0.2517251  0.2474632
#>   [33,]    0.2445247    0.2557877  0.2514571
#>   [34,]    0.3086886    0.3229069  0.3174400
#>   [35,]    0.2035666    0.2129432  0.2093379
#>   [36,]    0.2169253    0.2269172  0.2230753
#>   [37,]    0.2200216    0.2301562  0.2262595
#>   [38,]    0.1485959    0.1554406  0.1528088
#>   [39,]    0.2290890    0.2396412  0.2355839
#>   [40,]    0.3203877    0.3351447  0.3294706
#>   [41,]    0.2229772    0.2332480  0.2292989
#>   [42,]    0.1983033    0.2074376  0.2039255
#>   [43,]    0.3491403    0.3652217  0.3590384
#>   [44,]    0.1854504    0.1939926  0.1907081
#>   [45,]    0.2501206    0.2616416  0.2572118
#>   [46,]    0.1465388    0.1532887  0.1506934
#>   [47,]    0.2898398    0.3031901  0.2980569
#>   [48,]    0.1742842    0.1823121  0.1792254
#>   [49,]    0.2148415    0.2247375  0.2209325
#>   [50,]    0.1924948    0.2013613  0.1979522
#>   [51,]    0.2007993    0.2100485  0.2064922
#>   [52,]    0.2187528    0.2288290  0.2249547
#>   [53,]    0.2219546    0.2321782  0.2282472
#>   [54,]    0.2589284    0.2708549  0.2662692
#>   [55,]    0.2634631    0.2755985  0.2709325
#>   [56,]    0.2895786    0.3029167  0.2977882
#>   [57,]    0.2576954    0.2695652  0.2650013
#>   [58,]    0.1818583    0.1902351  0.1870142
#>   [59,]    0.2365917    0.2474894  0.2432993
#>   [60,]    0.1703037    0.1781483  0.1751320
#>   [61,]    0.2571397    0.2689838  0.2644297
#>   [62,]    0.2003173    0.2095443  0.2059965
#>   [63,]    0.1673072    0.1750138  0.1720506
#>   [64,]    0.2470503    0.2584298  0.2540544
#>   [65,]    0.1933280    0.2022331  0.1988091
#>   [66,]    0.2387306    0.2497269  0.2454989
#>   [67,]    0.2305722    0.2411928  0.2371092
#>   [68,]    0.2312484    0.2419000  0.2378045
#>   [69,]    0.1942468    0.2031942  0.1997539
#>   [70,]    0.2514799    0.2630633  0.2586095
#>   [71,]    0.1843201    0.1928103  0.1895458
#>   [72,]    0.1848167    0.1933298  0.1900565
#>   [73,]    0.2299032    0.2404929  0.2364212
#>   [74,]    0.2297490    0.2403316  0.2362626
#>   [75,]    0.2709178    0.2833966  0.2785986
#>   [76,]    0.1925612    0.2014309  0.1980205
#>   [77,]    0.2483236    0.2597617  0.2553638
#>   [78,]    0.1846675    0.1931737  0.1899031
#>   [79,]    0.2123261    0.2221063  0.2183458
#>   [80,]    0.1959149    0.2049390  0.2014693
#>   [81,]    0.1875473    0.1961862  0.1928646
#>   [82,]    0.2811445    0.2940942  0.2891150
#>   [83,]    0.3098890    0.3241628  0.3186745
#>   [84,]    0.1844195    0.1929143  0.1896481
#>   [85,]    0.1859281    0.1944924  0.1911995
#>   [86,]    0.1757431    0.1838383  0.1807257
#>   [87,]    0.2398571    0.2509053  0.2466573
#>   [88,]    0.1899907    0.1987421  0.1953772
#>   [89,]    0.2953808    0.3089864  0.3037551
#>   [90,]    0.1398798    0.1463231  0.1438457
#>   [91,]    0.3069543    0.3210928  0.3156566
#>   [92,]    0.1532277    0.1602858  0.1575720
#>   [93,]    0.2246741    0.2350230  0.2310439
#>   [94,]    0.2412848    0.2523986  0.2481253
#>   [95,]    0.2439612    0.2551984  0.2508777
#>   [96,]    0.2217681    0.2319831  0.2280555
#>   [97,]    0.1956633    0.2046760  0.2012107
#>   [98,]    0.2423209    0.2534825  0.2491909
#>   [99,]    0.2018373    0.2111343  0.2075596
#>  [100,]    0.2155686    0.2254980  0.2216802
#>  [101,]    0.2285145    0.2390402  0.2349931
#>  [102,]    0.1877128    0.1963593  0.1930347
#>  [103,]    0.2093169    0.2189583  0.2152512
#>  [104,]    0.1265784    0.1324089  0.1301671
#>  [105,]    0.2585955    0.2705068  0.2659269
#>  [106,]    0.2175184    0.2275376  0.2236852
#>  [107,]    0.1850252    0.1935478  0.1902709
#>  [108,]    0.1131153    0.1183258  0.1163224
#>  [109,]    0.2557706    0.2675518  0.2630219
#>  [110,]    0.2197792    0.2299026  0.2260102
#>  [111,]    0.2882916    0.3015705  0.2964648
#>  [112,]    0.3474344    0.3634370  0.3572840
#>  [113,]    0.2094174    0.2190636  0.2153547
#>  [114,]    0.2090524    0.2186817  0.2149792
#>  [115,]    0.1964169    0.2054643  0.2019856
#>  [116,]    0.1407629    0.1472468  0.1447537
#>  [117,]    0.1895938    0.1983269  0.1949691
#>  [118,]    0.1673603    0.1750694  0.1721053
#>  [119,]    0.1958524    0.2048737  0.2014050
#>  [120,]    0.1789173    0.1871586  0.1839898
#>  [121,]    0.2138641    0.2237151  0.2199274
#>  [122,]    0.1745511    0.1825913  0.1794998
#>  [123,]    0.1558789    0.1630591  0.1602983
#>  [124,]    0.2566462    0.2684673  0.2639221
#>  [125,]    0.3072635    0.3214162  0.3159746
#>  [126,]    0.2574188    0.2692759  0.2647169
#>  [127,]    0.2171902    0.2271943  0.2233478
#>  [128,]    0.2703411    0.2827933  0.2780055
#>  [129,]    0.4994851    0.5224907  0.5136451
#>  [130,]    0.2214970    0.2316996  0.2277767
#>  [131,]    0.2142555    0.2241244  0.2203298
#>  [132,]    0.2127225    0.2225207  0.2187533
#>  [133,]    0.1787617    0.1869959  0.1838299
#>  [134,]    0.1522289    0.1592409  0.1565448
#>  [135,]    0.2318032    0.2424805  0.2383751
#>  [136,]    0.2163416    0.2263067  0.2224752
#>  [137,]    0.2616424    0.2736940  0.2690602
#>  [138,]    0.2156947    0.2256300  0.2218099
#>  [139,]    0.1703879    0.1782365  0.1752187
#>  [140,]    0.1568038    0.1640266  0.1612494
#>  [141,]    0.2152050    0.2251176  0.2213062
#>  [142,]    0.1506183    0.1575562  0.1548886
#>  [143,]    0.1477229    0.1545274  0.1519111
#>  [144,]    0.1597469    0.1671053  0.1642760
#>  [145,]    0.2952171    0.3088150  0.3035866
#>  [146,]    0.2475501    0.2589525  0.2545683
#>  [147,]    0.2417163    0.2528500  0.2485691
#>  [148,]    0.1798046    0.1880869  0.1849023
#>  [149,]    0.1970651    0.2061423  0.2026521
#>  [150,]    0.2851972    0.2983337  0.2932827
#>  [151,]    0.2245117    0.2348530  0.2308768
#>  [152,]    0.1758813    0.1839828  0.1808678
#>  [153,]    0.2060119    0.2155011  0.2118525
#>  [154,]    0.2815423    0.2945104  0.2895242
#>  [155,]    0.2075896    0.2171517  0.2134751
#>  [156,]    0.3346153    0.3500276  0.3441016
#>  [157,]    0.1398726    0.1463154  0.1438382
#>  [158,]    0.2095651    0.2192180  0.2155065
#>  [159,]    0.1961040    0.2051369  0.2016637
#>  [160,]    0.2432328    0.2544364  0.2501287
#>  [161,]    0.1549339    0.1620706  0.1593266
#>  [162,]    0.2527383    0.2643797  0.2599036
#>  [163,]    0.3155997    0.3301363  0.3245470
#>  [164,]    0.2039019    0.2132941  0.2096828
#>  [165,]    0.1834667    0.1919176  0.1886682
#>  [166,]    0.3100570    0.3243384  0.3188472
#>  [167,]    0.2374271    0.2483633  0.2441583
#>  [168,]    0.2164890    0.2264609  0.2226267
#>  [169,]    0.2401812    0.2512443  0.2469906
#>  [170,]    0.1969932    0.2060671  0.2025782
#>  [171,]    0.2813840    0.2943449  0.2893614
#>  [172,]    0.1832415    0.1916820  0.1884366
#>  [173,]    0.2292430    0.2398023  0.2357423
#>  [174,]    0.2324259    0.2431317  0.2390154
#>  [175,]    0.2406248    0.2517083  0.2474467
#>  [176,]    0.2677035    0.2800342  0.2752931
#>  [177,]    0.2300379    0.2406338  0.2365597
#>  [178,]    0.2961679    0.3098096  0.3045644
#>  [179,]    0.2040535    0.2134526  0.2098387
#>  [180,]    0.1949209    0.2038994  0.2004472
#>  [181,]    0.1889280    0.1976304  0.1942843
#>  [182,]    0.1951803    0.2041706  0.2007138
#>  [183,]    0.1970959    0.2061745  0.2026838
#>  [184,]    0.1365476    0.1428374  0.1404190
#>  [185,]    0.2182789    0.2283331  0.2244673
#>  [186,]    0.2614872    0.2735316  0.2689005
#>  [187,]    0.1842496    0.1927365  0.1894733
#>  [188,]    0.2131150    0.2229314  0.2191570
#>  [189,]    0.2388487    0.2498504  0.2456202
#>  [190,]    0.2108952    0.2206094  0.2168743
#>  [191,]    0.1822162    0.1906094  0.1873822
#>  [192,]    0.2060576    0.2155490  0.2118996
#>  [193,]    0.1783177    0.1865314  0.1833732
#>  [194,]    0.1887761    0.1974714  0.1941281
#>  [195,]    0.2063609    0.2158662  0.2122114
#>  [196,]    0.2241291    0.2344529  0.2304834
#>  [197,]    0.2025374    0.2118667  0.2082796
#>  [198,]    0.2775914    0.2903777  0.2854614
#>  [199,]    0.2089249    0.2185484  0.2148482
#>  [200,]    0.2067284    0.2162508  0.2125895
#>  [201,]    0.2078035    0.2173753  0.2136950
#>  [202,]    0.1723778    0.1803179  0.1772649
#>  [203,]    0.2641016    0.2762664  0.2715891
#>  [204,]    0.1594124    0.1667553  0.1639320
#>  [205,]    0.2635946    0.2757362  0.2710677
#>  [206,]    0.2603981    0.2723923  0.2677806
#>  [207,]    0.2594377    0.2713875  0.2667928
#>  [208,]    0.2806259    0.2935518  0.2885819
#>  [209,]    0.2012304    0.2104994  0.2069355
#>  [210,]    0.2150481    0.2249535  0.2211449
#>  [211,]    0.2137384    0.2235836  0.2197982
#>  [212,]    0.2135765    0.2234143  0.2196317
#>  [213,]    0.1700375    0.1778697  0.1748582
#>  [214,]    0.1801589    0.1884575  0.1852667
#>  [215,]    0.2344688    0.2452689  0.2411163
#>  [216,]    0.2290773    0.2396290  0.2355719
#>  [217,]    0.2066954    0.2162161  0.2125554
#>  [218,]    0.2294893    0.2400599  0.2359956
#>  [219,]    0.1993100    0.2084906  0.2049607
#>  [220,]    0.2376043    0.2485487  0.2443406
#>  [221,]    0.2038061    0.2131938  0.2095842
#>  [222,]    0.2153060    0.2252233  0.2214101
#>  [223,]    0.2868340    0.3000460  0.2949660
#>  [224,]    0.2071818    0.2167250  0.2130557
#>  [225,]    0.2093702    0.2190142  0.2153061
#>  [226,]    0.2060676    0.2155595  0.2119099
#>  [227,]    0.2044326    0.2138491  0.2102285
#>  [228,]    0.2497445    0.2612481  0.2568250
#>  [229,]    0.2157949    0.2257348  0.2219129
#>  [230,]    0.1743114    0.1823407  0.1792534
#>  [231,]    0.2674365    0.2797549  0.2750185
#>  [232,]    0.2264507    0.2368811  0.2328707
#>  [233,]    0.1979654    0.2070841  0.2035780
#>  [234,]    0.2699379    0.2823716  0.2775908
#>  [235,]    0.2633213    0.2754503  0.2707867
#>  [236,]    0.2508974    0.2624540  0.2580105
#>  [237,]    0.3402934    0.3559674  0.3499408
#>  [238,]    0.2637545    0.2759034  0.2712322
#>  [239,]    0.3031808    0.3171455  0.3117761
#>  [240,]    0.2950850    0.3086768  0.3034508
#>  [241,]    0.2075224    0.2170813  0.2134059
#>  [242,]    0.1848786    0.1933945  0.1901201
#>  [243,]    0.2698706    0.2823009  0.2775215
#>  [244,]    0.2809638    0.2939052  0.2889293
#>  [245,]    0.1632759    0.1707968  0.1679050
#>  [246,]    0.2958856    0.3095144  0.3042741
#>  [247,]    0.2735435    0.2861432  0.2812986
#>  [248,]    0.1916504    0.2004783  0.1970840
#>  [249,]    0.2054995    0.2149653  0.2113257
#>  [250,]    0.2207562    0.2309246  0.2270148
#>  [251,]    0.2366100    0.2475086  0.2433182
#>  [252,]    0.1685450    0.1763086  0.1733235
#>  [253,]    0.1947999    0.2037727  0.2003227
#>  [254,]    0.2456517    0.2569668  0.2526162
#>  [255,]    0.2666297    0.2789107  0.2741887
#>  [256,]    0.1713367    0.1792289  0.1761943
#>  [257,]    0.2069940    0.2165285  0.2128625
#>  [258,]    0.1928140    0.2016954  0.1982805
#>  [259,]    0.2496610    0.2611608  0.2567392
#>  [260,]    0.2224326    0.2326782  0.2287388
#>  [261,]    0.2295281    0.2401006  0.2360355
#>  [262,]    0.3025709    0.3165074  0.3111489
#>  [263,]    0.2988401    0.3126047  0.3073122
#>  [264,]    0.1664004    0.1740652  0.1711181
#>  [265,]    0.1478936    0.1547059  0.1520866
#>  [266,]    0.2130318    0.2228444  0.2190715
#>  [267,]    0.1648101    0.1724016  0.1694827
#>  [268,]    0.2286278    0.2391587  0.2351096
#>  [269,]    0.2663671    0.2786362  0.2739187
#>  [270,]    0.1748342    0.1828874  0.1797909
#>  [271,]    0.1936885    0.2026102  0.1991798
#>  [272,]    0.2951178    0.3087113  0.3034846
#>  [273,]    0.1722763    0.1802117  0.1771605
#>  [274,]    0.1843289    0.1928195  0.1895549
#>  [275,]    0.2449839    0.2562682  0.2519294
#>  [276,]    0.2752668    0.2879458  0.2830708
#>  [277,]    0.2915521    0.3049811  0.2998177
#>  [278,]    0.1818222    0.1901974  0.1869772
#>  [279,]    0.1901562    0.1989153  0.1955474
#>  [280,]    0.2764522    0.2891859  0.2842898
#>  [281,]    0.1820226    0.1904069  0.1871832
#>  [282,]    0.1857904    0.1943483  0.1910578
#>  [283,]    0.1910184    0.1998171  0.1964341
#>  [284,]    0.2585144    0.2704218  0.2658434
#>  [285,]    0.2829001    0.2959308  0.2909205
#>  [286,]    0.1731435    0.1811188  0.1780523
#>  [287,]    0.2063919    0.2158986  0.2122433
#>  [288,]    0.2258048    0.2362058  0.2322066
#>  [289,]    0.2361405    0.2470175  0.2428353
#>  [290,]    0.2397837    0.2508285  0.2465818
#>  [291,]    0.1966619    0.2057206  0.2022376
#>  [292,]    0.2151769    0.2250883  0.2212774
#>  [293,]    0.2103997    0.2200910  0.2163647
#>  [294,]    0.2138096    0.2236579  0.2198713
#>  [295,]    0.2031695    0.2125278  0.2089296
#>  [296,]    0.2024370    0.2117617  0.2081764
#>  [297,]    0.2632435    0.2753688  0.2707066
#>  [298,]    0.2069286    0.2164602  0.2127953
#>  [299,]    0.2140376    0.2238965  0.2201058
#>  [300,]    0.2411865    0.2522958  0.2480243
#>  [301,]    0.3102981    0.3245904  0.3190951
#>  [302,]    0.1437397    0.1503608  0.1478150
#>  [303,]    0.2724397    0.2849886  0.2801636
#>  [304,]    0.2241443    0.2344687  0.2304990
#>  [305,]    0.1907761    0.1995637  0.1961849
#>  [306,]    0.1987277    0.2078815  0.2043619
#>  [307,]    0.2847606    0.2978768  0.2928336
#>  [308,]    0.2580037    0.2698876  0.2653183
#>  [309,]    0.2355691    0.2464198  0.2422477
#>  [310,]    0.1927765    0.2016562  0.1982420
#>  [311,]    0.2930619    0.3065604  0.3013702
#>  [312,]    0.1912866    0.2000976  0.1967098
#>  [313,]    0.2469351    0.2583093  0.2539359
#>  [314,]    0.2026138    0.2119465  0.2083581
#>  [315,]    0.1623976    0.1698780  0.1670018
#>  [316,]    0.1820522    0.1904378  0.1872136
#>  [317,]    0.2026817    0.2120177  0.2084280
#>  [318,]    0.1655813    0.1732083  0.1702757
#>  [319,]    0.1198428    0.1253630  0.1232405
#>  [320,]    0.2076375    0.2172017  0.2135243
#>  [321,]    0.1993134    0.2084941  0.2049641
#>  [322,]    0.2618590    0.2739204  0.2692829
#>  [323,]    0.2289789    0.2395261  0.2354707
#>  [324,]    0.2066925    0.2162130  0.2125524
#>  [325,]    0.2171892    0.2271932  0.2233467
#>  [326,]    0.1896973    0.1984352  0.1950755
#>  [327,]    0.1631673    0.1706831  0.1677933
#>  [328,]    0.2291008    0.2396535  0.2355960
#>  [329,]    0.2247349    0.2350866  0.2311064
#>  [330,]    0.1730768    0.1810491  0.1779838
#>  [331,]    0.1663965    0.1740612  0.1711141
#>  [332,]    0.2186531    0.2287246  0.2248521
#>  [333,]    0.2971592    0.3108466  0.3055838
#>  [334,]    0.2293684    0.2399335  0.2358712
#>  [335,]    0.2020897    0.2113983  0.2078192
#>  [336,]    0.2499409    0.2614535  0.2570269
#>  [337,]    0.1498360    0.1567378  0.1540840
#>  [338,]    0.1871302    0.1957499  0.1924356
#>  [339,]    0.1926964    0.2015724  0.1981596
#>  [340,]    0.2134065    0.2232365  0.2194569
#>  [341,]    0.1621365    0.1696049  0.1667334
#>  [342,]    0.2207468    0.2309147  0.2270052
#>  [343,]    0.2479455    0.2593662  0.2549750
#>  [344,]    0.2537214    0.2654081  0.2609146
#>  [345,]    0.2316717    0.2423430  0.2382399
#>  [346,]    0.2298440    0.2404311  0.2363604
#>  [347,]    0.1970860    0.2061642  0.2026736
#>  [348,]    0.2017565    0.2110498  0.2074765
#>  [349,]    0.2441171    0.2553614  0.2510380
#>  [350,]    0.2357123    0.2465695  0.2423949
#>  [351,]    0.1700342    0.1778664  0.1748549
#>  [352,]    0.1934759    0.2023877  0.1989611
#>  [353,]    0.2331755    0.2439159  0.2397863
#>  [354,]    0.2850106    0.2981385  0.2930908
#>  [355,]    0.2452922    0.2565908  0.2522465
#>  [356,]    0.1708618    0.1787320  0.1757059
#>  [357,]    0.2133052    0.2231303  0.2193526
#>  [358,]    0.1993077    0.2084883  0.2049584
#>  [359,]    0.1706911    0.1785535  0.1755305
#>  [360,]    0.1401515    0.1466073  0.1441250
#>  [361,]    0.2325851    0.2432983  0.2391791
#>  [362,]    0.2421711    0.2533259  0.2490369
#>  [363,]    0.2435652    0.2547842  0.2504705
#>  [364,]    0.2116413    0.2213899  0.2176416
#>  [365,]    0.2109393    0.2206556  0.2169197
#>  [366,]    0.2567627    0.2685895  0.2640421
#>  [367,]    0.1440500    0.1506853  0.1481340
#>  [368,]    0.2180418    0.2280852  0.2242235
#>  [369,]    0.1570043    0.1642362  0.1614556
#>  [370,]    0.2130631    0.2228773  0.2191038
#>  [371,]    0.2785864    0.2914181  0.2864844
#>  [372,]    0.1315260    0.1375844  0.1352550
#>  [373,]    0.2591104    0.2710453  0.2664564
#>  [374,]    0.1956582    0.2046707  0.2012054
#>  [375,]    0.2061373    0.2156324  0.2119815
#>  [376,]    0.2425713    0.2537443  0.2494484
#>  [377,]    0.2542853    0.2659979  0.2614945
#>  [378,]    0.1542271    0.1613312  0.1585997
#>  [379,]    0.2208906    0.2310652  0.2271530
#>  [380,]    0.2237951    0.2341035  0.2301399
#>  [381,]    0.2681359    0.2804866  0.2757378
#>  [382,]    0.1468311    0.1535945  0.1509940
#>  [383,]    0.2053384    0.2147967  0.2111600
#>  [384,]    0.2080110    0.2175922  0.2139083
#>  [385,]    0.3183160    0.3329777  0.3273403
#>  [386,]    0.1556252    0.1627937  0.1600374
#>  [387,]    0.3291058    0.3442643  0.3384359
#>  [388,]    0.2094775    0.2191264  0.2154164
#>  [389,]    0.2132828    0.2231070  0.2193296
#>  [390,]    0.3028867    0.3168380  0.3114738
#>  [391,]    0.2141947    0.2240608  0.2202673
#>  [392,]    0.2891189    0.3024361  0.2973157
#>  [393,]    0.2047140    0.2141435  0.2105178
#>  [394,]    0.2221370    0.2323691  0.2284348
#>  [395,]    0.1727793    0.1807378  0.1776778
#>  [396,]    0.3017055    0.3156021  0.3102589
#>  [397,]    0.1983876    0.2075257  0.2040121
#>  [398,]    0.1886856    0.1973769  0.1940351
#>  [399,]    0.2072985    0.2168471  0.2131756
#>  [400,]    0.1784122    0.1866303  0.1834705
#>  [401,]    0.2011201    0.2103840  0.2068221
#>  [402,]    0.1902199    0.1989817  0.1956128
#>  [403,]    0.1767315    0.1848721  0.1817421
#>  [404,]    0.2741687    0.2867971  0.2819415
#>  [405,]    0.1968247    0.2058908  0.2024049
#>  [406,]    0.1694224    0.1772263  0.1742257
#>  [407,]    0.2177431    0.2277727  0.2239163
#>  [408,]    0.1858883    0.1944508  0.1911585
#>  [409,]    0.2052314    0.2146847  0.2110499
#>  [410,]    0.1695077    0.1773156  0.1743135
#>  [411,]    0.1882366    0.1969072  0.1935734
#>  [412,]    0.2074377    0.2169926  0.2133188
#>  [413,]    0.2187809    0.2288583  0.2249836
#>  [414,]    0.1782843    0.1864965  0.1833389
#>  [415,]    0.3139584    0.3284195  0.3228592
#>  [416,]    0.1434008    0.1500062  0.1474664
#>  [417,]    0.1616698    0.1691167  0.1662534
#>  [418,]    0.1865937    0.1951886  0.1918839
#>  [419,]    0.2001275    0.2093458  0.2058014
#>  [420,]    0.1726671    0.1806205  0.1775624
#>  [421,]    0.2258806    0.2362851  0.2322846
#>  [422,]    0.2455609    0.2568718  0.2525228
#>  [423,]    0.2430315    0.2542259  0.2499217
#>  [424,]    0.2034170    0.2127868  0.2091841
#>  [425,]    0.2205987    0.2307599  0.2268529
#>  [426,]    0.2209627    0.2311406  0.2272272
#>  [427,]    0.1775604    0.1857391  0.1825944
#>  [428,]    0.1557870    0.1629628  0.1602037
#>  [429,]    0.2591462    0.2710827  0.2664931
#>  [430,]    0.1535453    0.1606179  0.1578985
#>  [431,]    0.3066777    0.3208034  0.3153721
#>  [432,]    0.2159491    0.2258961  0.2220715
#>  [433,]    0.2939218    0.3074602  0.3022547
#>  [434,]    0.2631130    0.2752321  0.2705723
#>  [435,]    0.2907532    0.3041455  0.2989962
#>  [436,]    0.1981441    0.2072710  0.2037617
#>  [437,]    0.2727338    0.2852962  0.2804660
#>  [438,]    0.2148698    0.2247672  0.2209617
#>  [439,]    0.2123698    0.2221520  0.2183908
#>  [440,]    0.2059438    0.2154299  0.2117825
#>  [441,]    0.2145758    0.2244595  0.2206592
#>  [442,]    0.2125568    0.2223475  0.2185830
#>  [443,]    0.2830886    0.2961281  0.2911144
#>  [444,]    0.1712597    0.1791483  0.1761152
#>  [445,]    0.2116601    0.2214096  0.2176609
#>  [446,]    0.2143060    0.2241773  0.2203818
#>  [447,]    0.1974612    0.2065566  0.2030594
#>  [448,]    0.3128151    0.3272235  0.3216835
#>  [449,]    0.3238145    0.3387294  0.3329947
#>  [450,]    0.1751353    0.1832023  0.1801006
#>  [451,]    0.1826925    0.1911078  0.1878721
#>  [452,]    0.1771730    0.1853341  0.1821962
#>  [453,]    0.1880405    0.1967020  0.1933717
#>  [454,]    0.1726821    0.1806363  0.1775779
#>  [455,]    0.2113120    0.2210455  0.2173030
#>  [456,]    0.2085943    0.2182025  0.2145082
#>  [457,]    0.1648676    0.1724618  0.1695418
#>  [458,]    0.2161444    0.2261003  0.2222722
#>  [459,]    0.2466919    0.2580547  0.2536857
#>  [460,]    0.2354388    0.2462835  0.2421137
#>  [461,]    0.2170026    0.2269981  0.2231548
#>  [462,]    0.1589747    0.1662974  0.1634818
#>  [463,]    0.1659475    0.1735915  0.1706524
#>  [464,]    0.2027006    0.2120374  0.2084474
#>  [465,]    0.2044594    0.2138773  0.2102562
#>  [466,]    0.2586981    0.2706141  0.2660324
#>  [467,]    0.2056495    0.2151220  0.2114798
#>  [468,]    0.1830692    0.1915017  0.1882594
#>  [469,]    0.1849860    0.1935068  0.1902306
#>  [470,]    0.2948498    0.3084308  0.3032090
#>  [471,]    0.2454526    0.2567586  0.2524115
#>  [472,]    0.1294264    0.1353881  0.1330959
#>  [473,]    0.1894688    0.1981961  0.1948405
#>  [474,]    0.1602351    0.1676159  0.1647780
#>  [475,]    0.2312314    0.2418824  0.2377871
#>  [476,]    0.1946450    0.2036107  0.2001634
#>  [477,]    0.2426908    0.2538694  0.2495712
#>  [478,]    0.2057618    0.2152395  0.2115954
#>  [479,]    0.1944282    0.2033840  0.1999405
#>  [480,]    0.3358261    0.3512943  0.3453468
#>  [481,]    0.1807180    0.1890422  0.1858416
#>  [482,]    0.1872573    0.1958828  0.1925663
#>  [483,]    0.2374466    0.2483838  0.2441785
#>  [484,]    0.1553799    0.1625371  0.1597852
#>  [485,]    0.1318496    0.1379230  0.1355878
#>  [486,]    0.1986626    0.2078134  0.2042949
#>  [487,]    0.1522487    0.1592616  0.1565651
#>  [488,]    0.2105573    0.2202558  0.2165268
#>  [489,]    0.2725524    0.2851065  0.2802795
#>  [490,]    0.2004292    0.2096613  0.2061116
#>  [491,]    0.1756120    0.1837010  0.1805908
#>  [492,]    0.2327495    0.2434703  0.2393481
#>  [493,]    0.1843006    0.1927899  0.1895258
#>  [494,]    0.2247174    0.2350684  0.2310884
#>  [495,]    0.2709408    0.2834205  0.2786221
#>  [496,]    0.1892307    0.1979470  0.1945956
#>  [497,]    0.2227428    0.2330026  0.2290577
#>  [498,]    0.2313316    0.2419870  0.2378900
#>  [499,]    0.2459592    0.2572884  0.2529324
#>  [500,]    0.1856517    0.1942033  0.1909152
#>  [501,]    0.2718043    0.2843237  0.2795100
#>  [502,]    0.1971787    0.2062612  0.2027690
#>  [503,]    0.2293714    0.2399365  0.2358742
#>  [504,]    0.1718586    0.1797748  0.1767311
#>  [505,]    0.1513190    0.1582891  0.1556091
#>  [506,]    0.1852203    0.1937518  0.1904714
#>  [507,]    0.2360540    0.2469269  0.2427463
#>  [508,]    0.1670357    0.1747297  0.1717714
#>  [509,]    0.1807579    0.1890840  0.1858826
#>  [510,]    0.2401862    0.2512495  0.2469957
#>  [511,]    0.2287770    0.2393149  0.2352631
#>  [512,]    0.1806763    0.1889986  0.1857987
#>  [513,]    0.2101036    0.2197813  0.2160603
#>  [514,]    0.2000108    0.2092235  0.2056812
#>  [515,]    0.1688597    0.1766378  0.1736472
#>  [516,]    0.2756903    0.2883888  0.2835063
#>  [517,]    0.2188822    0.2289643  0.2250877
#>  [518,]    0.1669978    0.1746901  0.1717325
#>  [519,]    0.2029596    0.2123082  0.2087137
#>  [520,]    0.1591725    0.1665043  0.1636852
#>  [521,]    0.2996981    0.3135022  0.3081946
#>  [522,]    0.2455734    0.2568850  0.2525357
#>  [523,]    0.2045120    0.2139323  0.2103102
#>  [524,]    0.2079214    0.2174986  0.2138162
#>  [525,]    0.2672230    0.2795317  0.2747990
#>  [526,]    0.1954356    0.2044378  0.2009764
#>  [527,]    0.2162512    0.2262121  0.2223822
#>  [528,]    0.2050591    0.2145045  0.2108728
#>  [529,]    0.1431004    0.1496919  0.1471575
#>  [530,]    0.2141284    0.2239916  0.2201992
#>  [531,]    0.1902256    0.1989878  0.1956187
#>  [532,]    0.1936782    0.2025993  0.1991692
#>  [533,]    0.1817387    0.1901100  0.1868912
#>  [534,]    0.2240572    0.2343776  0.2304094
#>  [535,]    0.2240445    0.2343643  0.2303963
#>  [536,]    0.1922035    0.2010567  0.1976527
#>  [537,]    0.1940557    0.2029942  0.1995574
#>  [538,]    0.2186127    0.2286824  0.2248106
#>  [539,]    0.1897596    0.1985004  0.1951396
#>  [540,]    0.1987958    0.2079527  0.2044319
#>  [541,]    0.1732625    0.1812434  0.1781748
#>  [542,]    0.2460111    0.2573427  0.2529857
#>  [543,]    0.2487769    0.2602360  0.2558300
#>  [544,]    0.2466698    0.2580316  0.2536630
#>  [545,]    0.2223573    0.2325995  0.2286614
#>  [546,]    0.1948621    0.2038377  0.2003866
#>  [547,]    0.2486581    0.2601115  0.2557077
#>  [548,]    0.1859747    0.1945410  0.1912473
#>  [549,]    0.2420089    0.2531561  0.2488700
#>  [550,]    0.2725035    0.2850552  0.2802291
#>  [551,]    0.3115735    0.3259247  0.3204067
#>  [552,]    0.3053922    0.3194586  0.3140501
#>  [553,]    0.1726648    0.1806181  0.1775600
#>  [554,]    0.1852297    0.1937618  0.1904812
#>  [555,]    0.2323503    0.2430527  0.2389376
#>  [556,]    0.2130536    0.2228672  0.2190939
#>  [557,]    0.2422511    0.2534095  0.2491191
#>  [558,]    0.2382602    0.2492346  0.2450149
#>  [559,]    0.2650074    0.2772139  0.2725205
#>  [560,]    0.1968638    0.2059317  0.2024452
#>  [561,]    0.2120995    0.2218693  0.2181128
#>  [562,]    0.1853271    0.1938637  0.1905814
#>  [563,]    0.1670294    0.1747232  0.1717649
#>  [564,]    0.1818158    0.1901907  0.1869706
#>  [565,]    0.2270136    0.2374703  0.2334497
#>  [566,]    0.1553187    0.1624730  0.1597222
#>  [567,]    0.1989352    0.2080985  0.2045752
#>  [568,]    0.2418735    0.2530144  0.2487308
#>  [569,]    0.2754331    0.2881197  0.2832417
#>  [570,]    0.1599049    0.1672706  0.1644385
#>  [571,]    0.2233203    0.2336069  0.2296517
#>  [572,]    0.2078358    0.2174089  0.2137280
#>  [573,]    0.1860192    0.1945876  0.1912931
#>  [574,]    0.2554813    0.2672491  0.2627244
#>  [575,]    0.2561955    0.2679962  0.2634588
#>  [576,]    0.1998041    0.2090075  0.2054688
#>  [577,]    0.2296210    0.2401977  0.2361310
#>  [578,]    0.2072988    0.2168473  0.2131759
#>  [579,]    0.2298236    0.2404096  0.2363393
#>  [580,]    0.2084185    0.2180185  0.2143273
#>  [581,]    0.1803494    0.1886567  0.1854625
#>  [582,]    0.2052704    0.2147254  0.2110900
#>  [583,]    0.2213698    0.2315664  0.2276458
#>  [584,]    0.2226868    0.2329441  0.2290002
#>  [585,]    0.1856211    0.1941712  0.1908837
#>  [586,]    0.2255479    0.2359370  0.2319424
#>  [587,]    0.2518952    0.2634977  0.2590365
#>  [588,]    0.1922833    0.2011403  0.1977348
#>  [589,]    0.2498224    0.2613296  0.2569051
#>  [590,]    0.1568849    0.1641114  0.1613328
#>  [591,]    0.2168816    0.2268716  0.2230305
#>  [592,]    0.1737599    0.1817636  0.1786862
#>  [593,]    0.2644889    0.2766715  0.2719873
#>  [594,]    0.1609878    0.1684032  0.1655520
#>  [595,]    0.1868918    0.1955004  0.1921904
#>  [596,]    0.1961834    0.2052199  0.2017454
#>  [597,]    0.2068216    0.2163483  0.2126853
#>  [598,]    0.2459765    0.2573066  0.2529502
#>  [599,]    0.1825903    0.1910008  0.1877670
#>  [600,]    0.1504191    0.1573478  0.1546837
#>  [601,]    0.1436834    0.1503018  0.1477571
#>  [602,]    0.1940975    0.2030381  0.1996005
#>  [603,]    0.2588899    0.2708146  0.2662296
#>  [604,]    0.1909505    0.1997460  0.1963641
#>  [605,]    0.1920702    0.2009173  0.1975156
#>  [606,]    0.1902671    0.1990311  0.1956614
#>  [607,]    0.2055095    0.2149757  0.2113359
#>  [608,]    0.2021456    0.2114567  0.2078766
#>  [609,]    0.2118509    0.2216092  0.2178571
#>  [610,]    0.1880511    0.1967131  0.1933826
#>  [611,]    0.1675142    0.1752302  0.1722634
#>  [612,]    0.1507643    0.1577089  0.1550387
#>  [613,]    0.2000364    0.2092505  0.2057077
#>  [614,]    0.2360211    0.2468926  0.2427126
#>  [615,]    0.2194071    0.2295134  0.2256276
#>  [616,]    0.1606587    0.1680590  0.1652136
#>  [617,]    0.1705135    0.1783679  0.1753479
#>  [618,]    0.1923572    0.2012175  0.1978107
#>  [619,]    0.1920418    0.2008876  0.1974864
#>  [620,]    0.2023732    0.2116949  0.2081107
#>  [621,]    0.2133467    0.2231738  0.2193953
#>  [622,]    0.2243925    0.2347284  0.2307542
#>  [623,]    0.2101576    0.2198379  0.2161158
#>  [624,]    0.1935585    0.2024742  0.1990461
#>  [625,]    0.2045622    0.2139848  0.2103618
#>  [626,]    0.2280728    0.2385781  0.2345388
#>  [627,]    0.2618852    0.2739480  0.2693099
#>  [628,]    0.1731505    0.1811262  0.1780596
#>  [629,]    0.2122217    0.2219971  0.2182385
#>  [630,]    0.2683672    0.2807284  0.2759755
#>  [631,]    0.2056207    0.2150920  0.2114503
#>  [632,]    0.2622065    0.2742840  0.2696402
#>  [633,]    0.2627507    0.2748533  0.2701999
#>  [634,]    0.1932899    0.2021932  0.1987699
#>  [635,]    0.1711849    0.1790701  0.1760383
#>  [636,]    0.2006565    0.2098992  0.2063454
#>  [637,]    0.2118864    0.2216464  0.2178937
#>  [638,]    0.1935246    0.2024387  0.1990113
#>  [639,]    0.2390698    0.2500817  0.2458476
#>  [640,]    0.1591673    0.1664989  0.1636799
#>  [641,]    0.2885511    0.3018421  0.2967317
#>  [642,]    0.1662440    0.1739016  0.1709573
#>  [643,]    0.1884422    0.1971223  0.1937848
#>  [644,]    0.2840173    0.2970994  0.2920693
#>  [645,]    0.2578097    0.2696847  0.2651188
#>  [646,]    0.2754344    0.2881212  0.2832431
#>  [647,]    0.1269940    0.1328437  0.1305945
#>  [648,]    0.2279926    0.2384944  0.2344565
#>  [649,]    0.1567988    0.1640214  0.1612443
#>  [650,]    0.2173621    0.2273741  0.2235245
#>  [651,]    0.1647115    0.1722985  0.1693813
#>  [652,]    0.2252024    0.2355756  0.2315871
#>  [653,]    0.2048139    0.2142480  0.2106206
#>  [654,]    0.1876000    0.1962412  0.1929187
#>  [655,]    0.1913190    0.2001316  0.1967431
#>  [656,]    0.2371563    0.2480800  0.2438799
#>  [657,]    0.1991420    0.2083149  0.2047879
#>  [658,]    0.2181225    0.2281696  0.2243065
#>  [659,]    0.1967788    0.2058429  0.2023578
#>  [660,]    0.2104193    0.2201115  0.2163849
#>  [661,]    0.1795677    0.1878390  0.1846587
#>  [662,]    0.2623880    0.2744738  0.2698268
#>  [663,]    0.1705999    0.1784581  0.1754367
#>  [664,]    0.2070242    0.2165601  0.2128936
#>  [665,]    0.1593040    0.1666419  0.1638205
#>  [666,]    0.2160320    0.2259828  0.2221567
#>  [667,]    0.2387827    0.2497814  0.2455525
#>  [668,]    0.1644235    0.1719973  0.1690852
#>  [669,]    0.2209750    0.2311535  0.2272399
#>  [670,]    0.2654759    0.2777038  0.2730022
#>  [671,]    0.2213306    0.2315255  0.2276056
#>  [672,]    0.3110902    0.3254192  0.3199097
#>  [673,]    0.2204876    0.2306438  0.2267387
#>  [674,]    0.2779988    0.2908036  0.2858802
#>  [675,]    0.2367081    0.2476112  0.2434190
#>  [676,]    0.2042118    0.2136180  0.2100013
#>  [677,]    0.2188162    0.2288952  0.2250199
#>  [678,]    0.2380220    0.2489855  0.2447701
#>  [679,]    0.1873514    0.1959811  0.1926630
#>  [680,]    0.1944954    0.2034541  0.2000095
#>  [681,]    0.2259261    0.2363326  0.2323314
#>  [682,]    0.1863139    0.1948959  0.1915961
#>  [683,]    0.2050915    0.2145383  0.2109060
#>  [684,]    0.1729793    0.1809471  0.1778835
#>  [685,]    0.2120880    0.2218572  0.2181010
#>  [686,]    0.2343236    0.2451169  0.2409669
#>  [687,]    0.2099088    0.2195776  0.2158600
#>  [688,]    0.1854365    0.1939781  0.1906939
#>  [689,]    0.2416656    0.2527971  0.2485171
#>  [690,]    0.1794961    0.1877641  0.1845851
#>  [691,]    0.1538621    0.1609494  0.1582243
#>  [692,]    0.2300564    0.2406532  0.2365787
#>  [693,]    0.2445242    0.2557874  0.2514567
#>  [694,]    0.2620414    0.2741113  0.2694705
#>  [695,]    0.1999681    0.2091790  0.2056375
#>  [696,]    0.1817096    0.1900795  0.1868613
#>  [697,]    0.1645803    0.1721613  0.1692464
#>  [698,]    0.3018689    0.3157732  0.3104270
#>  [699,]    0.1842465    0.1927333  0.1894702
#>  [700,]    0.1913665    0.2001812  0.1967920
#>  [701,]    0.1782457    0.1864561  0.1832992
#>  [702,]    0.2377362    0.2486868  0.2444763
#>  [703,]    0.1688159    0.1765919  0.1736020
#>  [704,]    0.1831812    0.1916190  0.1883747
#>  [705,]    0.1661938    0.1738491  0.1709056
#>  [706,]    0.1312580    0.1373040  0.1349793
#>  [707,]    0.1816774    0.1900458  0.1868282
#>  [708,]    0.1896690    0.1984055  0.1950463
#>  [709,]    0.1691291    0.1769197  0.1739242
#>  [710,]    0.3005188    0.3143609  0.3090386
#>  [711,]    0.2049201    0.2143592  0.2107299
#>  [712,]    0.1842483    0.1927352  0.1894720
#>  [713,]    0.1957812    0.2047993  0.2013319
#>  [714,]    0.2943462    0.3079041  0.3026911
#>  [715,]    0.2376268    0.2485722  0.2443637
#>  [716,]    0.2936985    0.3072264  0.3020249
#>  [717,]    0.1954281    0.2044299  0.2009688
#>  [718,]    0.1844170    0.1929117  0.1896455
#>  [719,]    0.2025179    0.2118464  0.2082596
#>  [720,]    0.2571289    0.2689725  0.2644186
#>  [721,]    0.1763544    0.1844777  0.1813543
#>  [722,]    0.1998461    0.2090515  0.2055120
#>  [723,]    0.2206382    0.2308012  0.2268935
#>  [724,]    0.2350497    0.2458763  0.2417135
#>  [725,]    0.1884632    0.1971442  0.1938064
#>  [726,]    0.2343111    0.2451038  0.2409540
#>  [727,]    0.2614437    0.2734860  0.2688558
#>  [728,]    0.1891012    0.1978115  0.1944624
#>  [729,]    0.2021130    0.2114227  0.2078431
#>  [730,]    0.3467599    0.3627318  0.3565906
#>  [731,]    0.2456534    0.2569684  0.2526178
#>  [732,]    0.2635268    0.2756650  0.2709979
#>  [733,]    0.2126629    0.2224586  0.2186922
#>  [734,]    0.1766125    0.1847476  0.1816197
#>  [735,]    0.1803098    0.1886153  0.1854218
#>  [736,]    0.1917310    0.2005625  0.1971668
#>  [737,]    0.1452023    0.1518907  0.1493190
#>  [738,]    0.1775171    0.1856939  0.1825499
#>  [739,]    0.1844225    0.1929174  0.1896511
#>  [740,]    0.1990970    0.2082677  0.2047416
#>  [741,]    0.1965328    0.2055855  0.2021047
#>  [742,]    0.2206843    0.2308494  0.2269410
#>  [743,]    0.1642778    0.1718449  0.1689354
#>  [744,]    0.2445280    0.2557914  0.2514607
#>  [745,]    0.2132986    0.2231235  0.2193459
#>  [746,]    0.1962265    0.2052650  0.2017897
#>  [747,]    0.2843567    0.2974544  0.2924184
#>  [748,]    0.2954961    0.3091067  0.3038734
#>  [749,]    0.1764806    0.1846096  0.1814840
#>  [750,]    0.2481623    0.2595929  0.2551979
#>  [751,]    0.1832823    0.1917247  0.1884786
#>  [752,]    0.2303162    0.2409249  0.2368459
#>  [753,]    0.1536182    0.1606942  0.1579735
#>  [754,]    0.2202841    0.2304307  0.2265293
#>  [755,]    0.1468570    0.1536216  0.1510206
#>  [756,]    0.1623038    0.1697799  0.1669054
#>  [757,]    0.2191855    0.2292816  0.2253997
#>  [758,]    0.1740711    0.1820892  0.1790063
#>  [759,]    0.1845101    0.1930090  0.1897412
#>  [760,]    0.2662792    0.2785443  0.2738284
#>  [761,]    0.1660135    0.1736605  0.1707202
#>  [762,]    0.1810410    0.1893801  0.1861737
#>  [763,]    0.2403240    0.2513936  0.2471374
#>  [764,]    0.2315069    0.2421704  0.2380703
#>  [765,]    0.1879416    0.1965987  0.1932700
#>  [766,]    0.2310185    0.2416596  0.2375681
#>  [767,]    0.2298469    0.2404341  0.2363633
#>  [768,]    0.1661143    0.1737659  0.1708239
#>  [769,]    0.1820983    0.1904861  0.1872610
#>  [770,]    0.2370968    0.2480178  0.2438187
#>  [771,]    0.2008291    0.2100796  0.2065228
#>  [772,]    0.3142937    0.3287701  0.3232039
#>  [773,]    0.1790740    0.1873226  0.1841510
#>  [774,]    0.2082216    0.2178127  0.2141249
#>  [775,]    0.1843425    0.1928338  0.1895689
#>  [776,]    0.2145989    0.2244836  0.2206830
#>  [777,]    0.2161399    0.2260957  0.2222677
#>  [778,]    0.2208181    0.2309893  0.2270785
#>  [779,]    0.1621980    0.1696693  0.1667966
#>  [780,]    0.2442676    0.2555188  0.2511927
#>  [781,]    0.1918869    0.2007256  0.1973271
#>  [782,]    0.2015407    0.2108241  0.2072546
#>  [783,]    0.2307094    0.2413362  0.2372503
#>  [784,]    0.3091283    0.3233670  0.3178923
#>  [785,]    0.1767648    0.1849069  0.1817763
#>  [786,]    0.2028407    0.2121839  0.2085915
#>  [787,]    0.3201138    0.3348582  0.3291890
#>  [788,]    0.2174511    0.2274672  0.2236160
#>  [789,]    0.2153461    0.2252654  0.2214515
#>  [790,]    0.2729975    0.2855721  0.2807372
#>  [791,]    0.1688683    0.1766467  0.1736559
#>  [792,]    0.1849359    0.1934544  0.1901791
#>  [793,]    0.1520327    0.1590357  0.1563431
#>  [794,]    0.2671112    0.2794146  0.2746840
#>  [795,]    0.2163080    0.2262716  0.2224406
#>  [796,]    0.2933120    0.3068221  0.3016275
#>  [797,]    0.1446059    0.1512668  0.1487057
#>  [798,]    0.1821848    0.1905766  0.1873499
#>  [799,]    0.1941425    0.2030850  0.1996466
#>  [800,]    0.2248387    0.2351950  0.2312130
#>  [801,]    0.2214521    0.2316525  0.2277305
#>  [802,]    0.2133887    0.2232177  0.2194385
#>  [803,]    0.2470701    0.2584504  0.2540747
#>  [804,]    0.1864345    0.1950220  0.1917202
#>  [805,]    0.1907237    0.1995087  0.1961309
#>  [806,]    0.1735460    0.1815400  0.1784663
#>  [807,]    0.2012326    0.2105017  0.2069378
#>  [808,]    0.2061959    0.2156937  0.2120418
#>  [809,]    0.1632905    0.1708120  0.1679200
#>  [810,]    0.2210167    0.2311969  0.2272826
#>  [811,]    0.2358403    0.2467034  0.2425266
#>  [812,]    0.1939803    0.2029154  0.1994799
#>  [813,]    0.1866443    0.1952415  0.1919359
#>  [814,]    0.1992110    0.2083871  0.2048589
#>  [815,]    0.2521794    0.2637952  0.2593290
#>  [816,]    0.1383047    0.1446753  0.1422258
#>  [817,]    0.2150308    0.2249356  0.2211272
#>  [818,]    0.1701020    0.1779373  0.1749247
#>  [819,]    0.2489566    0.2604240  0.2560148
#>  [820,]    0.1903290    0.1990960  0.1957251
#>  [821,]    0.2011892    0.2104563  0.2068931
#>  [822,]    0.2159596    0.2259072  0.2220824
#>  [823,]    0.1709738    0.1788492  0.1758211
#>  [824,]    0.2437881    0.2550174  0.2506998
#>  [825,]    0.2648230    0.2770210  0.2723309
#>  [826,]    0.3209531    0.3357364  0.3300523
#>  [827,]    0.1835214    0.1919748  0.1887245
#>  [828,]    0.2200754    0.2302124  0.2263147
#>  [829,]    0.1851369    0.1936648  0.1903858
#>  [830,]    0.2162948    0.2262577  0.2224270
#>  [831,]    0.2330615    0.2437966  0.2396690
#>  [832,]    0.2420015    0.2531484  0.2488624
#>  [833,]    0.1961868    0.2052235  0.2017489
#>  [834,]    0.1790791    0.1873279  0.1841563
#>  [835,]    0.1432689    0.1498681  0.1473307
#>  [836,]    0.1531241    0.1601774  0.1574654
#>  [837,]    0.1530168    0.1600651  0.1573550
#>  [838,]    0.1805481    0.1888645  0.1856669
#>  [839,]    0.2194572    0.2295658  0.2256791
#>  [840,]    0.2097700    0.2194324  0.2157173
#>  [841,]    0.1983663    0.2075034  0.2039902
#>  [842,]    0.2429000    0.2540883  0.2497865
#>  [843,]    0.2138549    0.2237054  0.2199179
#>  [844,]    0.2707089    0.2831779  0.2783836
#>  [845,]    0.2947065    0.3082809  0.3030616
#>  [846,]    0.1615206    0.1689606  0.1660999
#>  [847,]    0.2090667    0.2186967  0.2149939
#>  [848,]    0.2048391    0.2142744  0.2106466
#>  [849,]    0.2540208    0.2657213  0.2612225
#>  [850,]    0.2445412    0.2558051  0.2514742
#>  [851,]    0.2053768    0.2148369  0.2111995
#>  [852,]    0.2050521    0.2144972  0.2108656
#>  [853,]    0.1798392    0.1881229  0.1849378
#>  [854,]    0.2143183    0.2241903  0.2203945
#>  [855,]    0.1901603    0.1989195  0.1955516
#>  [856,]    0.2054715    0.2149358  0.2112968
#>  [857,]    0.2863209    0.2995090  0.2944382
#>  [858,]    0.2200120    0.2301461  0.2262496
#>  [859,]    0.3195271    0.3342447  0.3285858
#>  [860,]    0.1315113    0.1375691  0.1352399
#>  [861,]    0.2024638    0.2117897  0.2082039
#>  [862,]    0.1924744    0.2013401  0.1979313
#>  [863,]    0.1541951    0.1612977  0.1585668
#>  [864,]    0.2526083    0.2642438  0.2597700
#>  [865,]    0.2828879    0.2959180  0.2909080
#>  [866,]    0.1581197    0.1654031  0.1626027
#>  [867,]    0.2691870    0.2815860  0.2768186
#>  [868,]    0.1837293    0.1921922  0.1889382
#>  [869,]    0.1835539    0.1920088  0.1887579
#>  [870,]    0.1850201    0.1935425  0.1902656
#>  [871,]    0.1644398    0.1720142  0.1691019
#>  [872,]    0.2090594    0.2186891  0.2149865
#>  [873,]    0.2175476    0.2275682  0.2237153
#>  [874,]    0.2238035    0.2341123  0.2301486
#>  [875,]    0.1456614    0.1523709  0.1497911
#>  [876,]    0.3006364    0.3144838  0.3091595
#>  [877,]    0.2046073    0.2140319  0.2104081
#>  [878,]    0.2285961    0.2391256  0.2350770
#>  [879,]    0.2005747    0.2098137  0.2062613
#>  [880,]    0.2527646    0.2644072  0.2599306
#>  [881,]    0.2415321    0.2526571  0.2483796
#>  [882,]    0.2252075    0.2355808  0.2315923
#>  [883,]    0.2297914    0.2403759  0.2363062
#>  [884,]    0.2273250    0.2377959  0.2337699
#>  [885,]    0.1828923    0.1913167  0.1880776
#>  [886,]    0.2188589    0.2289399  0.2250638
#>  [887,]    0.2182545    0.2283078  0.2244423
#>  [888,]    0.1955649    0.2045730  0.2011094
#>  [889,]    0.1943048    0.2032548  0.1998136
#>  [890,]    0.2062687    0.2157697  0.2121166
#>  [891,]    0.2337153    0.2444806  0.2403414
#>  [892,]    0.1782832    0.1864954  0.1833378
#>  [893,]    0.2127743    0.2225751  0.2188067
#>  [894,]    0.1702847    0.1781284  0.1751125
#>  [895,]    0.2613898    0.2734296  0.2688003
#>  [896,]    0.2794048    0.2922745  0.2873262
#>  [897,]    0.2403883    0.2514610  0.2472036
#>  [898,]    0.1456602    0.1523698  0.1497900
#>  [899,]    0.1733569    0.1813422  0.1782719
#>  [900,]    0.1652371    0.1728483  0.1699218
#>  [901,]    0.1801369    0.1884345  0.1852441
#>  [902,]    0.1728245    0.1807852  0.1777243
#>  [903,]    0.2239977    0.2343153  0.2303482
#>  [904,]    0.2301355    0.2407359  0.2366601
#>  [905,]    0.2465280    0.2578835  0.2535174
#>  [906,]    0.1748425    0.1828961  0.1797995
#>  [907,]    0.2643989    0.2765774  0.2718948
#>  [908,]    0.2594563    0.2714072  0.2668121
#>  [909,]    0.1746403    0.1826846  0.1795916
#>  [910,]    0.1748786    0.1829339  0.1798366
#>  [911,]    0.2085759    0.2181833  0.2144893
#>  [912,]    0.1945203    0.2034802  0.2000352
#>  [913,]    0.3328075    0.3481368  0.3422428
#>  [914,]    0.1627373    0.1702334  0.1673512
#>  [915,]    0.1328800    0.1390008  0.1366474
#>  [916,]    0.2309547    0.2415929  0.2375025
#>  [917,]    0.1793333    0.1875937  0.1844176
#>  [918,]    0.1967087    0.2057694  0.2022856
#>  [919,]    0.2216032    0.2318103  0.2278857
#>  [920,]    0.1762453    0.1843636  0.1812422
#>  [921,]    0.2216366    0.2318456  0.2279203
#>  [922,]    0.2253767    0.2357577  0.2317663
#>  [923,]    0.1774253    0.1855980  0.1824556
#>  [924,]    0.3188550    0.3335416  0.3278946
#>  [925,]    0.2401263    0.2511869  0.2469341
#>  [926,]    0.2595031    0.2714562  0.2668603
#>  [927,]    0.2130520    0.2228655  0.2190922
#>  [928,]    0.1908819    0.1996743  0.1962936
#>  [929,]    0.2443931    0.2556501  0.2513218
#>  [930,]    0.2601744    0.2721583  0.2675505
#>  [931,]    0.2160828    0.2260360  0.2222090
#>  [932,]    0.2213440    0.2315395  0.2276194
#>  [933,]    0.2057976    0.2152771  0.2116322
#>  [934,]    0.1763594    0.1844830  0.1813595
#>  [935,]    0.2158248    0.2257661  0.2219437
#>  [936,]    0.3453176    0.3612228  0.3551073
#>  [937,]    0.2015050    0.2107867  0.2072179
#>  [938,]    0.1922847    0.2011417  0.1977362
#>  [939,]    0.1857844    0.1943420  0.1910516
#>  [940,]    0.2120596    0.2218274  0.2180717
#>  [941,]    0.2214224    0.2316214  0.2276999
#>  [942,]    0.1978946    0.2070100  0.2035052
#>  [943,]    0.2055093    0.2149755  0.2113357
#>  [944,]    0.1968542    0.2059217  0.2024353
#>  [945,]    0.1960309    0.2050605  0.2015886
#>  [946,]    0.2187455    0.2288213  0.2249472
#>  [947,]    0.2237933    0.2341016  0.2301381
#>  [948,]    0.2085533    0.2181595  0.2144660
#>  [949,]    0.2459035    0.2572301  0.2528750
#>  [950,]    0.1836590    0.1921188  0.1888661
#>  [951,]    0.1673044    0.1750108  0.1720477
#>  [952,]    0.1751635    0.1832318  0.1801296
#>  [953,]    0.1832680    0.1917096  0.1884638
#>  [954,]    0.2045505    0.2139725  0.2103498
#>  [955,]    0.3017720    0.3156719  0.3103274
#>  [956,]    0.2204493    0.2306036  0.2266993
#>  [957,]    0.2013526    0.2106274  0.2070612
#>  [958,]    0.2065343    0.2160476  0.2123898
#>  [959,]    0.1539112    0.1610007  0.1582748
#>  [960,]    0.2162187    0.2261780  0.2223487
#>  [961,]    0.2551006    0.2668509  0.2623329
#>  [962,]    0.2105088    0.2202051  0.2164769
#>  [963,]    0.1902661    0.1990301  0.1956603
#>  [964,]    0.2598041    0.2717708  0.2671696
#>  [965,]    0.1958604    0.2048821  0.2014133
#>  [966,]    0.2741177    0.2867437  0.2818890
#>  [967,]    0.2175797    0.2276018  0.2237484
#>  [968,]    0.3209575    0.3357404  0.3300564
#>  [969,]    0.2318072    0.2424845  0.2383791
#>  [970,]    0.2673854    0.2797014  0.2749659
#>  [971,]    0.1699620    0.1777908  0.1747806
#>  [972,]    0.2621209    0.2741947  0.2695524
#>  [973,]    0.1810824    0.1894235  0.1862163
#>  [974,]    0.1887194    0.1974122  0.1940698
#>  [975,]    0.2870779    0.3003010  0.2952168
#>  [976,]    0.1879439    0.1966009  0.1932723
#>  [977,]    0.2167389    0.2267223  0.2228837
#>  [978,]    0.2332270    0.2439698  0.2398392
#>  [979,]    0.2035933    0.2129712  0.2093654
#>  [980,]    0.1758826    0.1839842  0.1808691
#>  [981,]    0.1794567    0.1877229  0.1845446
#>  [982,]    0.2456754    0.2569915  0.2526405
#>  [983,]    0.3020433    0.3159557  0.3106064
#>  [984,]    0.2184593    0.2285218  0.2246528
#>  [985,]    0.2699433    0.2823771  0.2775963
#>  [986,]    0.1762211    0.1843382  0.1812172
#>  [987,]    0.3146381    0.3291303  0.3235581
#>  [988,]    0.1779460    0.1861425  0.1829910
#>  [989,]    0.1594823    0.1668284  0.1640039
#>  [990,]    0.2030510    0.2124039  0.2088078
#>  [991,]    0.2030778    0.2124320  0.2088353
#>  [992,]    0.1765648    0.1846977  0.1815706
#>  [993,]    0.2898109    0.3031598  0.2980272
#>  [994,]    0.3274217    0.3425029  0.3367042
#>  [995,]    0.2201928    0.2303352  0.2264354
#>  [996,]    0.2169022    0.2268931  0.2230517
#>  [997,]    0.1850527    0.1935766  0.1902992
#>  [998,]    0.2109488    0.2206654  0.2169294
#>  [999,]    0.2504010    0.2619347  0.2575000
#> [1000,]    0.1931156    0.2020109  0.1985907