Stochastic migration collision risk model

Run migration stochastic collision risk model for a single species and one turbine scenario

Usage

mig_stoch_crm(
  wing_span_pars,
  flt_speed_pars,
  body_lt_pars,
  prop_crh_pars,
  avoid_bsc_pars,
  n_turbines,
  n_blades = 3,
  rtn_speed_pars,
  bld_pitch_pars,
  rtr_radius_pars,
  bld_width_pars,
  wf_width,
  wf_latitude,
  flight_type,
  prop_upwind = 0.5,
  popn_estim_pars,
  season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl,
  trb_downtime_pars,
  n_iter = 10,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = NULL,
  verbose = TRUE
)

Arguments

wing_span_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species wingspan, in metres. Assumed to follow a tnorm-lw0 distribution.

flt_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species flying speed, in metres/sec. Assumed to follow a Truncated Normal with lower bound at 0 (tnorm-lw0).

body_lt_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the species body length, in metres. Assumed to follow a tnorm-lw0 distribution.

prop_crh_pars

A single row data frame with columns mean and sd. The mean and standard deviation of the proportion of flights at collision risk height.

avoid_bsc_pars

Single row data frame with columns mean and sd, the mean and standard deviation of the species avoidance rate to be used. Avoidance rate expresses the probability that a bird flying on a collision course with a turbine will take evading action to avoid collision, and it is assumed to follow a Beta distribution.

n_turbines

An integer. The number of turbines expected to be installed

n_blades

An integer. The number of blades per turbine (defaults to 3)

rtn_speed_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the rotation speed.

bld_pitch_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the blade pitch angle, i.e. the angle between the blade surface and the rotor plane, in degrees. Assumed to follow a tnorm-lw0 distribution.

rtr_radius_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the radius of the rotor, in metres. Assumed to follow a tnorm-lw0 distribution.

bld_width_pars

A single row data frame with columns mean and sd, the mean and standard deviation of the maximum blade width, in metres. Assumed to be tnorm-lw0 distribution.

wf_width

Numeric value, the approximate longitudinal width of the wind farm, in kilometres (\(w\)).

wf_latitude

A decimal value. The latitude of the centroid of the windfarm, in degrees.

flight_type

A character. Either "flying" or "gliding" representing the type of flight most commonly used by the species

prop_upwind

Numeric value between 0-1 giving the proportion of flights upwind - defaults to 0.5.

popn_estim_pars

A single row data frame with columns mean and sd. The population estimate of the species expected to fly through the wind farm area.

season_specs

A data frame defining the seasons for aggregating over collision estimates. It must comprise the following columns:

season_id, (unique) season identifier,

start_month, name of the season's first month,

end_month, name of the season's last month.

chord_profile

A data frame with the chord taper profile of the rotor blade. It must contain the columns:

pp_radius, equidistant intervals of radius at bird passage point, as a proportion of rotor_radius, within the range \([0, 1]\).

chord, the chord width at pp_radius, as a proportion of blade_width. Defaults to a generic profile for a typical modern 5MW turbine. See chord_prof_5MW() for details.

trb_wind_avbl

A data frame with the monthly estimates of operational wind availability. It must contain the columns:

month, (unique) month names,

pctg, the percentage of time wind conditions allow for turbine operation per month.

trb_downtime_pars

A data frame with monthly estimates of maintenance downtime, assumed to follow a tnorm-lw0 distribution. It must contain the following columns:

month, (unique) month names,

mean, numeric, the mean percentage of time in each month when turbines are not operating due to maintenance,

sd, the standard deviation of monthly maintenance downtime.

n_iter

An integer > 0. The number of iterations for the model simulation.

LargeArrayCorrection

A boolean. Should the large array correction be calculated

log_file

Path to log file to store session info and main model run options. If set to NULL (default value), log file is not created.

seed

Integer, the random seed for random number generation, for analysis reproducibility.

verbose

boolean. TRUE for a verbose output

Value

Estimates of number of collisions per migratory season for the n number of iterations specified

Details

This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in offshore windfarm sites and is a further adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function stoch_crm

function will run only option 1 for migratory species

Examples

# ------------------------------------------------------
# Run with arbitrary parameter values, for illustration
# ------------------------------------------------------
season_specs <- data.frame(
  season_id = c("PrBMigration", "PoBMigration","OMigration"),
  start_month = c("Mar", "May", "Oct"), end_month = c("Apr", "Sep", "Feb")
)

# wind availability
windavb <- data.frame(
  month = month.abb,
  pctg = runif(12, 85, 98)
)
head(windavb)
#>   month     pctg
#> 1   Jan 85.19122
#> 2   Feb 94.28646
#> 3   Mar 91.63482
#> 4   Apr 90.99521
#> 5   May 96.71911
#> 6   Jun 90.14690

# maintenance downtime
dwntm <- data.frame(
  month = month.abb,
  mean = runif(12, 6, 10),
  sd = rep(2, 12))
head(dwntm)
#>   month     mean sd
#> 1   Jan 8.329697  2
#> 2   Feb 6.645504  2
#> 3   Mar 6.937749  2
#> 4   Apr 6.934621  2
#> 5   May 7.177843  2
#> 6   Jun 7.072924  2

mig_stoch_crm(
  # Wing span in m,
  wing_span_pars = data.frame(mean = 1.08, sd = 0.04),
  # Flight speed in m/s
  flt_speed_pars = data.frame(mean = 7.26, sd = 1.5),
  # Body length in m,
  body_lt_pars = data.frame(mean = 0.39, sd = 0.005),
  # Proportion of birds at CRH
  prop_crh_pars = data.frame(mean = 0.06, sd = 0.009),
  # avoidance rate
  avoid_bsc_pars = data.frame(mean = 0.99, sd = 0.001),
  n_turbines = 150,
  n_blades = 3,
  # rotation speed in m/s of turbine blades
  rtn_speed_pars = data.frame(mean = 13.1, sd = 4),
  # pitch in degrees of turbine blades
  bld_pitch_pars = data.frame(mean = 3, sd = 0.3),
  # sd = 0, rotor radius is fixed
  rtr_radius_pars = data.frame(mean = 80, sd = 0),
  # sd = 0, blade width is fixed
  bld_width_pars = data.frame(mean = 8, sd = 0),
  wf_width = 100,
  wf_latitude = 54.1,
  prop_upwind = 0.5,
  flight_type = "flapping",
  # population flying through windfarm,
  popn_estim_pars = data.frame(mean = 21584, sd = 2023),
  season_specs = season_specs,
  chord_profile = chord_prof_5MW,
  trb_wind_avbl = windavb,
  trb_downtime_pars = dwntm,
  n_iter = 1000,
  LargeArrayCorrection = TRUE,
  log_file = NULL,
  seed = 1234,
  verbose = TRUE)
#> 
#> ── Stochastic CRM for Migratory birds ──
#> 
#> ℹ Checking for manditory arguments
#> ✔ Checking for manditory arguments [12ms]
#> 
#> ℹ Data validation
#> ✔ Data validation [23ms]
#> 
#> ℹ Sampling data
#> ✔ Sampling data [69ms]
#> 
#> ℹ Running simulation...
#> ✔ Running simulation... [761ms]
#> 
#> ℹ Creating outputs
#> ✔ Creating outputs [22ms]
#> 
#>         PrBMigration PoBMigration OMigration
#>    [1,]    0.1983137    0.2013006  0.2000403
#>    [2,]    0.2050883    0.2081772  0.2068739
#>    [3,]    0.2283906    0.2318304  0.2303791
#>    [4,]    0.1920731    0.1949659  0.1937453
#>    [5,]    0.2098002    0.2129600  0.2116267
#>    [6,]    0.2751861    0.2793306  0.2775818
#>    [7,]    0.2277920    0.2312228  0.2297752
#>    [8,]    0.1770457    0.1797123  0.1785871
#>    [9,]    0.2493863    0.2531423  0.2515575
#>   [10,]    0.3856694    0.3914777  0.3890270
#>   [11,]    0.3066533    0.3112716  0.3093230
#>   [12,]    0.2399059    0.2435191  0.2419946
#>   [13,]    0.1973621    0.2003347  0.1990805
#>   [14,]    0.1839169    0.1866870  0.1855182
#>   [15,]    0.2695917    0.2736519  0.2719387
#>   [16,]    0.2202432    0.2235603  0.2221607
#>   [17,]    0.2493148    0.2530697  0.2514853
#>   [18,]    0.2369463    0.2405150  0.2390092
#>   [19,]    0.1481851    0.1504170  0.1494752
#>   [20,]    0.2538248    0.2576477  0.2560347
#>   [21,]    0.1812574    0.1839873  0.1828355
#>   [22,]    0.2496363    0.2533961  0.2518097
#>   [23,]    0.2137075    0.2169262  0.2155681
#>   [24,]    0.2068697    0.2099854  0.2086708
#>   [25,]    0.2288598    0.2323067  0.2308523
#>   [26,]    0.1900768    0.1929396  0.1917316
#>   [27,]    0.2759571    0.2801133  0.2783596
#>   [28,]    0.2416198    0.2452589  0.2437234
#>   [29,]    0.2052159    0.2083067  0.2070025
#>   [30,]    0.1731089    0.1757161  0.1746160
#>   [31,]    0.2592456    0.2631501  0.2615026
#>   [32,]    0.2454633    0.2491603  0.2476004
#>   [33,]    0.2494250    0.2531815  0.2515965
#>   [34,]    0.3148747    0.3196169  0.3176160
#>   [35,]    0.2076461    0.2107735  0.2094540
#>   [36,]    0.2212726    0.2246051  0.2231990
#>   [37,]    0.2244309    0.2278111  0.2263849
#>   [38,]    0.1515739    0.1538568  0.1528935
#>   [39,]    0.2336800    0.2371995  0.2357145
#>   [40,]    0.3268081    0.3317300  0.3296533
#>   [41,]    0.2274458    0.2308714  0.2294260
#>   [42,]    0.2022775    0.2053240  0.2040386
#>   [43,]    0.3561370    0.3615006  0.3592375
#>   [44,]    0.1891669    0.1920160  0.1908139
#>   [45,]    0.2551331    0.2589757  0.2573544
#>   [46,]    0.1494755    0.1517268  0.1507769
#>   [47,]    0.2956483    0.3001010  0.2982222
#>   [48,]    0.1777770    0.1804545  0.1793248
#>   [49,]    0.2191471    0.2224477  0.2210550
#>   [50,]    0.1963524    0.1993097  0.1980619
#>   [51,]    0.2048235    0.2079084  0.2066067
#>   [52,]    0.2231367    0.2264974  0.2250794
#>   [53,]    0.2264027    0.2298125  0.2283738
#>   [54,]    0.2641174    0.2680952  0.2664168
#>   [55,]    0.2687429    0.2727905  0.2710827
#>   [56,]    0.2953817    0.2998304  0.2979533
#>   [57,]    0.2628597    0.2668186  0.2651482
#>   [58,]    0.1855029    0.1882968  0.1871179
#>   [59,]    0.2413331    0.2449678  0.2434342
#>   [60,]    0.1737167    0.1763331  0.1752291
#>   [61,]    0.2622928    0.2662432  0.2645763
#>   [62,]    0.2043317    0.2074092  0.2061107
#>   [63,]    0.1706602    0.1732306  0.1721460
#>   [64,]    0.2520013    0.2557967  0.2541953
#>   [65,]    0.1972024    0.2001726  0.1989193
#>   [66,]    0.2435149    0.2471825  0.2456350
#>   [67,]    0.2351930    0.2387353  0.2372406
#>   [68,]    0.2358827    0.2394353  0.2379363
#>   [69,]    0.1981396    0.2011239  0.1998647
#>   [70,]    0.2565196    0.2603830  0.2587529
#>   [71,]    0.1880140    0.1908457  0.1896509
#>   [72,]    0.1885206    0.1913599  0.1901619
#>   [73,]    0.2345105    0.2380426  0.2365523
#>   [74,]    0.2343533    0.2378829  0.2363936
#>   [75,]    0.2763471    0.2805092  0.2787530
#>   [76,]    0.1964202    0.1993785  0.1981303
#>   [77,]    0.2533001    0.2571151  0.2555054
#>   [78,]    0.1883684    0.1912054  0.1900084
#>   [79,]    0.2165813    0.2198432  0.2184669
#>   [80,]    0.1998411    0.2028509  0.2015810
#>   [81,]    0.1913059    0.1941872  0.1929715
#>   [82,]    0.2867786    0.2910977  0.2892753
#>   [83,]    0.3160992    0.3208600  0.3188512
#>   [84,]    0.1881154    0.1909487  0.1897532
#>   [85,]    0.1896543    0.1925107  0.1913055
#>   [86,]    0.1792651    0.1819651  0.1808259
#>   [87,]    0.2446640    0.2483488  0.2467940
#>   [88,]    0.1937983    0.1967171  0.1954856
#>   [89,]    0.3013003    0.3058382  0.3039235
#>   [90,]    0.1426832    0.1448322  0.1439255
#>   [91,]    0.3131057    0.3178213  0.3158316
#>   [92,]    0.1562985    0.1586526  0.1576593
#>   [93,]    0.2291767    0.2326284  0.2311720
#>   [94,]    0.2461202    0.2498269  0.2482629
#>   [95,]    0.2488502    0.2525982  0.2510168
#>   [96,]    0.2262125    0.2296195  0.2281819
#>   [97,]    0.1995846    0.2025906  0.2013222
#>   [98,]    0.2471771    0.2508998  0.2493291
#>   [99,]    0.2058822    0.2089830  0.2076747
#>  [100,]    0.2198887    0.2232005  0.2218031
#>  [101,]    0.2330940    0.2366046  0.2351234
#>  [102,]    0.1914747    0.1943586  0.1931417
#>  [103,]    0.2135117    0.2167274  0.2153705
#>  [104,]    0.1291152    0.1310598  0.1302393
#>  [105,]    0.2637779    0.2677507  0.2660744
#>  [106,]    0.2218776    0.2252193  0.2238093
#>  [107,]    0.1887332    0.1915757  0.1903764
#>  [108,]    0.1153823    0.1171201  0.1163868
#>  [109,]    0.2608963    0.2648257  0.2631678
#>  [110,]    0.2241837    0.2275601  0.2261355
#>  [111,]    0.2940690    0.2984979  0.2966292
#>  [112,]    0.3543968    0.3597341  0.3574821
#>  [113,]    0.2136143    0.2168316  0.2154741
#>  [114,]    0.2132419    0.2164536  0.2150984
#>  [115,]    0.2003532    0.2033708  0.2020976
#>  [116,]    0.1435839    0.1457465  0.1448340
#>  [117,]    0.1933934    0.1963062  0.1950772
#>  [118,]    0.1707144    0.1732856  0.1722007
#>  [119,]    0.1997774    0.2027863  0.2015167
#>  [120,]    0.1825029    0.1852517  0.1840919
#>  [121,]    0.2181501    0.2214357  0.2200493
#>  [122,]    0.1780492    0.1807308  0.1795993
#>  [123,]    0.1590028    0.1613976  0.1603872
#>  [124,]    0.2617893    0.2657320  0.2640685
#>  [125,]    0.3134211    0.3181414  0.3161497
#>  [126,]    0.2625776    0.2665323  0.2648636
#>  [127,]    0.2215428    0.2248795  0.2234716
#>  [128,]    0.2757588    0.2799120  0.2781596
#>  [129,]    0.5094944    0.5171674  0.5139299
#>  [130,]    0.2259359    0.2293388  0.2279030
#>  [131,]    0.2185493    0.2218408  0.2204520
#>  [132,]    0.2169855    0.2202535  0.2188746
#>  [133,]    0.1823442    0.1850906  0.1839318
#>  [134,]    0.1552797    0.1576184  0.1566316
#>  [135,]    0.2364487    0.2400099  0.2385073
#>  [136,]    0.2206772    0.2240009  0.2225985
#>  [137,]    0.2668858    0.2709054  0.2692094
#>  [138,]    0.2200173    0.2233311  0.2219329
#>  [139,]    0.1738027    0.1764204  0.1753159
#>  [140,]    0.1599463    0.1623553  0.1613388
#>  [141,]    0.2195178    0.2228239  0.2214289
#>  [142,]    0.1536368    0.1559508  0.1549744
#>  [143,]    0.1506834    0.1529529  0.1519953
#>  [144,]    0.1629484    0.1654026  0.1643671
#>  [145,]    0.3011332    0.3056685  0.3037549
#>  [146,]    0.2525111    0.2563141  0.2547094
#>  [147,]    0.2465603    0.2502738  0.2487069
#>  [148,]    0.1834080    0.1861704  0.1850049
#>  [149,]    0.2010144    0.2040419  0.2027645
#>  [150,]    0.2909126    0.2952940  0.2934453
#>  [151,]    0.2290110    0.2324601  0.2310048
#>  [152,]    0.1794061    0.1821082  0.1809681
#>  [153,]    0.2101404    0.2133054  0.2119700
#>  [154,]    0.2871845    0.2915097  0.2896847
#>  [155,]    0.2117498    0.2149391  0.2135934
#>  [156,]    0.3413209    0.3464614  0.3442924
#>  [157,]    0.1426757    0.1448246  0.1439179
#>  [158,]    0.2137649    0.2169844  0.2156260
#>  [159,]    0.2000340    0.2030468  0.2017756
#>  [160,]    0.2481073    0.2518440  0.2502673
#>  [161,]    0.1580389    0.1604192  0.1594149
#>  [162,]    0.2578032    0.2616860  0.2600477
#>  [163,]    0.3219243    0.3267727  0.3247270
#>  [164,]    0.2079883    0.2111208  0.2097991
#>  [165,]    0.1871435    0.1899621  0.1887728
#>  [166,]    0.3162705    0.3210338  0.3190240
#>  [167,]    0.2421852    0.2458328  0.2442937
#>  [168,]    0.2208276    0.2241535  0.2227502
#>  [169,]    0.2449945    0.2486844  0.2471275
#>  [170,]    0.2009410    0.2039674  0.2026905
#>  [171,]    0.2870230    0.2913459  0.2895219
#>  [172,]    0.1869138    0.1897289  0.1885411
#>  [173,]    0.2338371    0.2373590  0.2358730
#>  [174,]    0.2370838    0.2406545  0.2391479
#>  [175,]    0.2454470    0.2491437  0.2475839
#>  [176,]    0.2730684    0.2771810  0.2754457
#>  [177,]    0.2346480    0.2381820  0.2366909
#>  [178,]    0.3021031    0.3066530  0.3047332
#>  [179,]    0.2081429    0.2112778  0.2099550
#>  [180,]    0.1988273    0.2018219  0.2005583
#>  [181,]    0.1927142    0.1956167  0.1943920
#>  [182,]    0.1990918    0.2020903  0.2008251
#>  [183,]    0.2010458    0.2040738  0.2027962
#>  [184,]    0.1392841    0.1413820  0.1404968
#>  [185,]    0.2226533    0.2260066  0.2245917
#>  [186,]    0.2667275    0.2707446  0.2690496
#>  [187,]    0.1879421    0.1907727  0.1895783
#>  [188,]    0.2173859    0.2206600  0.2192785
#>  [189,]    0.2436353    0.2473047  0.2457564
#>  [190,]    0.2151216    0.2183616  0.2169945
#>  [191,]    0.1858679    0.1886673  0.1874861
#>  [192,]    0.2101871    0.2133528  0.2120171
#>  [193,]    0.1818913    0.1846308  0.1834749
#>  [194,]    0.1925593    0.1954594  0.1942357
#>  [195,]    0.2104965    0.2136668  0.2123291
#>  [196,]    0.2286208    0.2320641  0.2306112
#>  [197,]    0.2065964    0.2097080  0.2083951
#>  [198,]    0.2831545    0.2874191  0.2856197
#>  [199,]    0.2131119    0.2163216  0.2149673
#>  [200,]    0.2108714    0.2140474  0.2127073
#>  [201,]    0.2119680    0.2151605  0.2138135
#>  [202,]    0.1758323    0.1784806  0.1773632
#>  [203,]    0.2693942    0.2734516  0.2717396
#>  [204,]    0.1626072    0.1650562  0.1640229
#>  [205,]    0.2688771    0.2729267  0.2712180
#>  [206,]    0.2656165    0.2696170  0.2679290
#>  [207,]    0.2646368    0.2686224  0.2669407
#>  [208,]    0.2862497    0.2905609  0.2887418
#>  [209,]    0.2052631    0.2083546  0.2070502
#>  [210,]    0.2193577    0.2226615  0.2212675
#>  [211,]    0.2180219    0.2213055  0.2199200
#>  [212,]    0.2178567    0.2211379  0.2197535
#>  [213,]    0.1734451    0.1760574  0.1749552
#>  [214,]    0.1837695    0.1865373  0.1853694
#>  [215,]    0.2391677    0.2427698  0.2412499
#>  [216,]    0.2336681    0.2371874  0.2357025
#>  [217,]    0.2108377    0.2140131  0.2126733
#>  [218,]    0.2340884    0.2376140  0.2361264
#>  [219,]    0.2033043    0.2063663  0.2050743
#>  [220,]    0.2423660    0.2460163  0.2444761
#>  [221,]    0.2078905    0.2110215  0.2097004
#>  [222,]    0.2196208    0.2229286  0.2215329
#>  [223,]    0.2925823    0.2969889  0.2951296
#>  [224,]    0.2113338    0.2145168  0.2131738
#>  [225,]    0.2135661    0.2167827  0.2154255
#>  [226,]    0.2101973    0.2133632  0.2120274
#>  [227,]    0.2085295    0.2116702  0.2103450
#>  [228,]    0.2547495    0.2585863  0.2569674
#>  [229,]    0.2201195    0.2234348  0.2220360
#>  [230,]    0.1778048    0.1804828  0.1793528
#>  [231,]    0.2727960    0.2769045  0.2751710
#>  [232,]    0.2309888    0.2344676  0.2329998
#>  [233,]    0.2019328    0.2049741  0.2036909
#>  [234,]    0.2753475    0.2794945  0.2777447
#>  [235,]    0.2685984    0.2726438  0.2709368
#>  [236,]    0.2559254    0.2597799  0.2581536
#>  [237,]    0.3471128    0.3523406  0.3501348
#>  [238,]    0.2690402    0.2730922  0.2713825
#>  [239,]    0.3092565    0.3139142  0.3119490
#>  [240,]    0.3009985    0.3055318  0.3036190
#>  [241,]    0.2116813    0.2148694  0.2135242
#>  [242,]    0.1885837    0.1914240  0.1902255
#>  [243,]    0.2752787    0.2794246  0.2776753
#>  [244,]    0.2865943    0.2909107  0.2890894
#>  [245,]    0.1665480    0.1690565  0.1679981
#>  [246,]    0.3018152    0.3063608  0.3044428
#>  [247,]    0.2790254    0.2832277  0.2814546
#>  [248,]    0.1954912    0.1984356  0.1971932
#>  [249,]    0.2096179    0.2127750  0.2114429
#>  [250,]    0.2251802    0.2285717  0.2271407
#>  [251,]    0.2413518    0.2449868  0.2434531
#>  [252,]    0.1719227    0.1745121  0.1734196
#>  [253,]    0.1987038    0.2016965  0.2004338
#>  [254,]    0.2505747    0.2543486  0.2527562
#>  [255,]    0.2719729    0.2760690  0.2743407
#>  [256,]    0.1747704    0.1774027  0.1762920
#>  [257,]    0.2111423    0.2143223  0.2129806
#>  [258,]    0.1966781    0.1996403  0.1983905
#>  [259,]    0.2546643    0.2584998  0.2568815
#>  [260,]    0.2268902    0.2303075  0.2288656
#>  [261,]    0.2341280    0.2376542  0.2361664
#>  [262,]    0.3086344    0.3132827  0.3113214
#>  [263,]    0.3048288    0.3094197  0.3074826
#>  [264,]    0.1697352    0.1722916  0.1712130
#>  [265,]    0.1508575    0.1531296  0.1521709
#>  [266,]    0.2173011    0.2205739  0.2191930
#>  [267,]    0.1681130    0.1706450  0.1695767
#>  [268,]    0.2332096    0.2367220  0.2352399
#>  [269,]    0.2717051    0.2757972  0.2740706
#>  [270,]    0.1783380    0.1810239  0.1798906
#>  [271,]    0.1975702    0.2005458  0.1992903
#>  [272,]    0.3010321    0.3055659  0.3036529
#>  [273,]    0.1757288    0.1783755  0.1772588
#>  [274,]    0.1880230    0.1908549  0.1896600
#>  [275,]    0.2498935    0.2536571  0.2520691
#>  [276,]    0.2807832    0.2850120  0.2832277
#>  [277,]    0.2973948    0.3018738  0.2999839
#>  [278,]    0.1854661    0.1882595  0.1870808
#>  [279,]    0.1939671    0.1968885  0.1956558
#>  [280,]    0.2819924    0.2862394  0.2844474
#>  [281,]    0.1856705    0.1884669  0.1872870
#>  [282,]    0.1895137    0.1923680  0.1911637
#>  [283,]    0.1948466    0.1977812  0.1965430
#>  [284,]    0.2636951    0.2676666  0.2659908
#>  [285,]    0.2885695    0.2929156  0.2910818
#>  [286,]    0.1766134    0.1792734  0.1781510
#>  [287,]    0.2105281    0.2136988  0.2123610
#>  [288,]    0.2303301    0.2337991  0.2323354
#>  [289,]    0.2408728    0.2445007  0.2429699
#>  [290,]    0.2445891    0.2482728  0.2467185
#>  [291,]    0.2006032    0.2036245  0.2023497
#>  [292,]    0.2194892    0.2227949  0.2214001
#>  [293,]    0.2146162    0.2178485  0.2164847
#>  [294,]    0.2180944    0.2213791  0.2199932
#>  [295,]    0.2072411    0.2103624  0.2090454
#>  [296,]    0.2064940    0.2096041  0.2082918
#>  [297,]    0.2685189    0.2725631  0.2708567
#>  [298,]    0.2110756    0.2142546  0.2129133
#>  [299,]    0.2183270    0.2216153  0.2202278
#>  [300,]    0.2460200    0.2497252  0.2481618
#>  [301,]    0.3165164    0.3212833  0.3192720
#>  [302,]    0.1466204    0.1488287  0.1478970
#>  [303,]    0.2778995    0.2820849  0.2803189
#>  [304,]    0.2286362    0.2320797  0.2306268
#>  [305,]    0.1945994    0.1975304  0.1962937
#>  [306,]    0.2027103    0.2057634  0.2044752
#>  [307,]    0.2904672    0.2948418  0.2929960
#>  [308,]    0.2631742    0.2671378  0.2654654
#>  [309,]    0.2402900    0.2439091  0.2423821
#>  [310,]    0.1966399    0.1996015  0.1983519
#>  [311,]    0.2989348    0.3034369  0.3015373
#>  [312,]    0.1951201    0.1980588  0.1968188
#>  [313,]    0.2518837    0.2556774  0.2540767
#>  [314,]    0.2066743    0.2097870  0.2084736
#>  [315,]    0.1656521    0.1681471  0.1670944
#>  [316,]    0.1857006    0.1884975  0.1873174
#>  [317,]    0.2067436    0.2098574  0.2085436
#>  [318,]    0.1688996    0.1714435  0.1703701
#>  [319,]    0.1222445    0.1240857  0.1233088
#>  [320,]    0.2117987    0.2149886  0.2136427
#>  [321,]    0.2033077    0.2063697  0.2050778
#>  [322,]    0.2671067    0.2711295  0.2694321
#>  [323,]    0.2335678    0.2370856  0.2356013
#>  [324,]    0.2108347    0.2140101  0.2126702
#>  [325,]    0.2215418    0.2248784  0.2234705
#>  [326,]    0.1934990    0.1964133  0.1951836
#>  [327,]    0.1664373    0.1689440  0.1678863
#>  [328,]    0.2336920    0.2372117  0.2357266
#>  [329,]    0.2292387    0.2326913  0.2312346
#>  [330,]    0.1765454    0.1792044  0.1780825
#>  [331,]    0.1697313    0.1722877  0.1712090
#>  [332,]    0.2230350    0.2263942  0.2249768
#>  [333,]    0.3031143    0.3076795  0.3057532
#>  [334,]    0.2339651    0.2374888  0.2360020
#>  [335,]    0.2061397    0.2092444  0.2079344
#>  [336,]    0.2549498    0.2587896  0.2571694
#>  [337,]    0.1528388    0.1551408  0.1541695
#>  [338,]    0.1908805    0.1937554  0.1925423
#>  [339,]    0.1965582    0.1995186  0.1982695
#>  [340,]    0.2176833    0.2209619  0.2195786
#>  [341,]    0.1653859    0.1678768  0.1668258
#>  [342,]    0.2251706    0.2285619  0.2271310
#>  [343,]    0.2529144    0.2567236  0.2551164
#>  [344,]    0.2588061    0.2627039  0.2610593
#>  [345,]    0.2363146    0.2398738  0.2383720
#>  [346,]    0.2344502    0.2379813  0.2364914
#>  [347,]    0.2010357    0.2040636  0.2027860
#>  [348,]    0.2057998    0.2088994  0.2075916
#>  [349,]    0.2490093    0.2527596  0.2511772
#>  [350,]    0.2404360    0.2440572  0.2425293
#>  [351,]    0.1734418    0.1760541  0.1749519
#>  [352,]    0.1973532    0.2003256  0.1990714
#>  [353,]    0.2378484    0.2414307  0.2399192
#>  [354,]    0.2907222    0.2951008  0.2932533
#>  [355,]    0.2502080    0.2539764  0.2523864
#>  [356,]    0.1742860    0.1769109  0.1758034
#>  [357,]    0.2175799    0.2208569  0.2194742
#>  [358,]    0.2033020    0.2063640  0.2050720
#>  [359,]    0.1741119    0.1767343  0.1756278
#>  [360,]    0.1429603    0.1451135  0.1442049
#>  [361,]    0.2372462    0.2408193  0.2393117
#>  [362,]    0.2470244    0.2507448  0.2491750
#>  [363,]    0.2484464    0.2521882  0.2506094
#>  [364,]    0.2158827    0.2191342  0.2177623
#>  [365,]    0.2151667    0.2184073  0.2170400
#>  [366,]    0.2619083    0.2658529  0.2641885
#>  [367,]    0.1469369    0.1491499  0.1482162
#>  [368,]    0.2224115    0.2257612  0.2243478
#>  [369,]    0.1601507    0.1625628  0.1615451
#>  [370,]    0.2173331    0.2206064  0.2192253
#>  [371,]    0.2841692    0.2884490  0.2866432
#>  [372,]    0.1341619    0.1361826  0.1353300
#>  [373,]    0.2643030    0.2682837  0.2666041
#>  [374,]    0.1995794    0.2025853  0.2013170
#>  [375,]    0.2102684    0.2134353  0.2120990
#>  [376,]    0.2474325    0.2511590  0.2495867
#>  [377,]    0.2593812    0.2632877  0.2616394
#>  [378,]    0.1573180    0.1596874  0.1586877
#>  [379,]    0.2253173    0.2287108  0.2272790
#>  [380,]    0.2282800    0.2317182  0.2302675
#>  [381,]    0.2735094    0.2776287  0.2758906
#>  [382,]    0.1497737    0.1520295  0.1510777
#>  [383,]    0.2094535    0.2126081  0.2112771
#>  [384,]    0.2121796    0.2153752  0.2140269
#>  [385,]    0.3246950    0.3295851  0.3275218
#>  [386,]    0.1587441    0.1611350  0.1601262
#>  [387,]    0.3357009    0.3407567  0.3386235
#>  [388,]    0.2136755    0.2168937  0.2155358
#>  [389,]    0.2175571    0.2208338  0.2194512
#>  [390,]    0.3089566    0.3136098  0.3116464
#>  [391,]    0.2184872    0.2217779  0.2203894
#>  [392,]    0.2949129    0.2993546  0.2974805
#>  [393,]    0.2088165    0.2119616  0.2106346
#>  [394,]    0.2265887    0.2300014  0.2285615
#>  [395,]    0.1762419    0.1788963  0.1777763
#>  [396,]    0.3077516    0.3123866  0.3104309
#>  [397,]    0.2023634    0.2054112  0.2041252
#>  [398,]    0.1924670    0.1953658  0.1941427
#>  [399,]    0.2114529    0.2146376  0.2132938
#>  [400,]    0.1819877    0.1847287  0.1835722
#>  [401,]    0.2051506    0.2082404  0.2069367
#>  [402,]    0.1940320    0.1969543  0.1957213
#>  [403,]    0.1802733    0.1829885  0.1818428
#>  [404,]    0.2796630    0.2838750  0.2820978
#>  [405,]    0.2007692    0.2037930  0.2025171
#>  [406,]    0.1728177    0.1754205  0.1743223
#>  [407,]    0.2221068    0.2254519  0.2240405
#>  [408,]    0.1896137    0.1924695  0.1912645
#>  [409,]    0.2093443    0.2124973  0.2111669
#>  [410,]    0.1729048    0.1755089  0.1744101
#>  [411,]    0.1920090    0.1949009  0.1936807
#>  [412,]    0.2115949    0.2147817  0.2134371
#>  [413,]    0.2231654    0.2265265  0.2251083
#>  [414,]    0.1818572    0.1845962  0.1834405
#>  [415,]    0.3202501    0.3250734  0.3230382
#>  [416,]    0.1462746    0.1484777  0.1475482
#>  [417,]    0.1649098    0.1673936  0.1663456
#>  [418,]    0.1903332    0.1931998  0.1919903
#>  [419,]    0.2041382    0.2072128  0.2059155
#>  [420,]    0.1761275    0.1787802  0.1776609
#>  [421,]    0.2304074    0.2338776  0.2324134
#>  [422,]    0.2504820    0.2542546  0.2526628
#>  [423,]    0.2479020    0.2516356  0.2500603
#>  [424,]    0.2074936    0.2106187  0.2093001
#>  [425,]    0.2250196    0.2284087  0.2269787
#>  [426,]    0.2253909    0.2287855  0.2273532
#>  [427,]    0.1811187    0.1838466  0.1826956
#>  [428,]    0.1589090    0.1613024  0.1602925
#>  [429,]    0.2643395    0.2683207  0.2666409
#>  [430,]    0.1566225    0.1589814  0.1579861
#>  [431,]    0.3128235    0.3175349  0.3155470
#>  [432,]    0.2202769    0.2235945  0.2221947
#>  [433,]    0.2998121    0.3043275  0.3024223
#>  [434,]    0.2683857    0.2724278  0.2707223
#>  [435,]    0.2965799    0.3010466  0.2991619
#>  [436,]    0.2021150    0.2051591  0.2038747
#>  [437,]    0.2781994    0.2823894  0.2806215
#>  [438,]    0.2191759    0.2224770  0.2210842
#>  [439,]    0.2166259    0.2198885  0.2185119
#>  [440,]    0.2100710    0.2132349  0.2118999
#>  [441,]    0.2188760    0.2221725  0.2207816
#>  [442,]    0.2168165    0.2200820  0.2187042
#>  [443,]    0.2887618    0.2931109  0.2912758
#>  [444,]    0.1746919    0.1773230  0.1762128
#>  [445,]    0.2159019    0.2191536  0.2177816
#>  [446,]    0.2186008    0.2218932  0.2205040
#>  [447,]    0.2014184    0.2044520  0.2031720
#>  [448,]    0.3190839    0.3238896  0.3218619
#>  [449,]    0.3303037    0.3352782  0.3331793
#>  [450,]    0.1786451    0.1813357  0.1802004
#>  [451,]    0.1863538    0.1891606  0.1879763
#>  [452,]    0.1807237    0.1834457  0.1822972
#>  [453,]    0.1918090    0.1946978  0.1934789
#>  [454,]    0.1761428    0.1787957  0.1776764
#>  [455,]    0.2155469    0.2187933  0.2174235
#>  [456,]    0.2127747    0.2159793  0.2146271
#>  [457,]    0.1681717    0.1707046  0.1696358
#>  [458,]    0.2204760    0.2237966  0.2223955
#>  [459,]    0.2516356    0.2554255  0.2538264
#>  [460,]    0.2401571    0.2437742  0.2422480
#>  [461,]    0.2213514    0.2246852  0.2232785
#>  [462,]    0.1621607    0.1646030  0.1635725
#>  [463,]    0.1692732    0.1718227  0.1707470
#>  [464,]    0.2067628    0.2098769  0.2085630
#>  [465,]    0.2085570    0.2116981  0.2103727
#>  [466,]    0.2638825    0.2678568  0.2661799
#>  [467,]    0.2097708    0.2129302  0.2115971
#>  [468,]    0.1867380    0.1895505  0.1883638
#>  [469,]    0.1886932    0.1915352  0.1903360
#>  [470,]    0.3007586    0.3052883  0.3033771
#>  [471,]    0.2503716    0.2541425  0.2525514
#>  [472,]    0.1320202    0.1340086  0.1331696
#>  [473,]    0.1932659    0.1961767  0.1949485
#>  [474,]    0.1634463    0.1659080  0.1648694
#>  [475,]    0.2358654    0.2394178  0.2379189
#>  [476,]    0.1985458    0.2015361  0.2002744
#>  [477,]    0.2475544    0.2512828  0.2497096
#>  [478,]    0.2098854    0.2130465  0.2117127
#>  [479,]    0.1983247    0.2013117  0.2000514
#>  [480,]    0.3425560    0.3477151  0.3455383
#>  [481,]    0.1843397    0.1871161  0.1859446
#>  [482,]    0.1910101    0.1938869  0.1926731
#>  [483,]    0.2422052    0.2458531  0.2443139
#>  [484,]    0.1584939    0.1608810  0.1598738
#>  [485,]    0.1344920    0.1365177  0.1356630
#>  [486,]    0.2026439    0.2056960  0.2044082
#>  [487,]    0.1552998    0.1576389  0.1566519
#>  [488,]    0.2147769    0.2180117  0.2166468
#>  [489,]    0.2780144    0.2822016  0.2804349
#>  [490,]    0.2044459    0.2075251  0.2062258
#>  [491,]    0.1791314    0.1818293  0.1806909
#>  [492,]    0.2374139    0.2409896  0.2394808
#>  [493,]    0.1879941    0.1908256  0.1896309
#>  [494,]    0.2292209    0.2326732  0.2312166
#>  [495,]    0.2763704    0.2805328  0.2787765
#>  [496,]    0.1930230    0.1959301  0.1947035
#>  [497,]    0.2272066    0.2306286  0.2291847
#>  [498,]    0.2359675    0.2395214  0.2380219
#>  [499,]    0.2508883    0.2546669  0.2530726
#>  [500,]    0.1893723    0.1922245  0.1910211
#>  [501,]    0.2772512    0.2814269  0.2796650
#>  [502,]    0.2011303    0.2041596  0.2028814
#>  [503,]    0.2339680    0.2374918  0.2360050
#>  [504,]    0.1753028    0.1779431  0.1768290
#>  [505,]    0.1543516    0.1566763  0.1556954
#>  [506,]    0.1889322    0.1917777  0.1905770
#>  [507,]    0.2407846    0.2444110  0.2428809
#>  [508,]    0.1703832    0.1729494  0.1718666
#>  [509,]    0.1843804    0.1871574  0.1859857
#>  [510,]    0.2449996    0.2486895  0.2471326
#>  [511,]    0.2333618    0.2368765  0.2353935
#>  [512,]    0.1842971    0.1870729  0.1859017
#>  [513,]    0.2143142    0.2175420  0.2161800
#>  [514,]    0.2040190    0.2070918  0.2057953
#>  [515,]    0.1722438    0.1748380  0.1737434
#>  [516,]    0.2812152    0.2854505  0.2836635
#>  [517,]    0.2232687    0.2266314  0.2252125
#>  [518,]    0.1703446    0.1729102  0.1718277
#>  [519,]    0.2070270    0.2101450  0.2088294
#>  [520,]    0.1623624    0.1648078  0.1637760
#>  [521,]    0.3057040    0.3103081  0.3083654
#>  [522,]    0.2504948    0.2542676  0.2526757
#>  [523,]    0.2086106    0.2117525  0.2104268
#>  [524,]    0.2120883    0.2152826  0.2139348
#>  [525,]    0.2725783    0.2766836  0.2749514
#>  [526,]    0.1993522    0.2023547  0.2010879
#>  [527,]    0.2205850    0.2239073  0.2225055
#>  [528,]    0.2091686    0.2123189  0.2109897
#>  [529,]    0.1459682    0.1481667  0.1472391
#>  [530,]    0.2184197    0.2217094  0.2203213
#>  [531,]    0.1940378    0.1969603  0.1957272
#>  [532,]    0.1975596    0.2005350  0.1992796
#>  [533,]    0.1853809    0.1881729  0.1869948
#>  [534,]    0.2285474    0.2319895  0.2305372
#>  [535,]    0.2285344    0.2319764  0.2305241
#>  [536,]    0.1960553    0.1990081  0.1977622
#>  [537,]    0.1979447    0.2009259  0.1996680
#>  [538,]    0.2229938    0.2263524  0.2249353
#>  [539,]    0.1935625    0.1964778  0.1952478
#>  [540,]    0.2027798    0.2058339  0.2045452
#>  [541,]    0.1767348    0.1793967  0.1782735
#>  [542,]    0.2509413    0.2547207  0.2531260
#>  [543,]    0.2537625    0.2575844  0.2559718
#>  [544,]    0.2516131    0.2554026  0.2538037
#>  [545,]    0.2268135    0.2302296  0.2287882
#>  [546,]    0.1987672    0.2017608  0.2004977
#>  [547,]    0.2536412    0.2574613  0.2558495
#>  [548,]    0.1897017    0.1925589  0.1913533
#>  [549,]    0.2468588    0.2505767  0.2490080
#>  [550,]    0.2779645    0.2821509  0.2803845
#>  [551,]    0.3178174    0.3226040  0.3205843
#>  [552,]    0.3115122    0.3162038  0.3142242
#>  [553,]    0.1761251    0.1787778  0.1776585
#>  [554,]    0.1889419    0.1917876  0.1905868
#>  [555,]    0.2370067    0.2405762  0.2390701
#>  [556,]    0.2173233    0.2205964  0.2192154
#>  [557,]    0.2471059    0.2508276  0.2492572
#>  [558,]    0.2430349    0.2466952  0.2451508
#>  [559,]    0.2703182    0.2743894  0.2726716
#>  [560,]    0.2008091    0.2038335  0.2025574
#>  [561,]    0.2163502    0.2196087  0.2182338
#>  [562,]    0.1890412    0.1918884  0.1906871
#>  [563,]    0.1703768    0.1729429  0.1718602
#>  [564,]    0.1854596    0.1882529  0.1870743
#>  [565,]    0.2315631    0.2350507  0.2335791
#>  [566,]    0.1584314    0.1608175  0.1598107
#>  [567,]    0.2029220    0.2059782  0.2046887
#>  [568,]    0.2467207    0.2504365  0.2488687
#>  [569,]    0.2809528    0.2851841  0.2833987
#>  [570,]    0.1631095    0.1655662  0.1645296
#>  [571,]    0.2277958    0.2312267  0.2297791
#>  [572,]    0.2120008    0.2151938  0.2138465
#>  [573,]    0.1897471    0.1926049  0.1913991
#>  [574,]    0.2606012    0.2645261  0.2628701
#>  [575,]    0.2613297    0.2652656  0.2636049
#>  [576,]    0.2038083    0.2068779  0.2055827
#>  [577,]    0.2342227    0.2377503  0.2362619
#>  [578,]    0.2114532    0.2146379  0.2132941
#>  [579,]    0.2344294    0.2379601  0.2364704
#>  [580,]    0.2125953    0.2157972  0.2144462
#>  [581,]    0.1839637    0.1867344  0.1855654
#>  [582,]    0.2093841    0.2125376  0.2112070
#>  [583,]    0.2258061    0.2292070  0.2277721
#>  [584,]    0.2271495    0.2305706  0.2291271
#>  [585,]    0.1893410    0.1921927  0.1909895
#>  [586,]    0.2300680    0.2335331  0.2320710
#>  [587,]    0.2569432    0.2608130  0.2591802
#>  [588,]    0.1961368    0.1990908  0.1978444
#>  [589,]    0.2548289    0.2586669  0.2570475
#>  [590,]    0.1600290    0.1624392  0.1614222
#>  [591,]    0.2212280    0.2245600  0.2231541
#>  [592,]    0.1772422    0.1799117  0.1787853
#>  [593,]    0.2697893    0.2738525  0.2721381
#>  [594,]    0.1642141    0.1666874  0.1656438
#>  [595,]    0.1906372    0.1935085  0.1922970
#>  [596,]    0.2001150    0.2031290  0.2018573
#>  [597,]    0.2109665    0.2141439  0.2128032
#>  [598,]    0.2509060    0.2546849  0.2530904
#>  [599,]    0.1862495    0.1890547  0.1878711
#>  [600,]    0.1534336    0.1557446  0.1547695
#>  [601,]    0.1465630    0.1487704  0.1478390
#>  [602,]    0.1979874    0.2009693  0.1997111
#>  [603,]    0.2640781    0.2680553  0.2663772
#>  [604,]    0.1947772    0.1977108  0.1964730
#>  [605,]    0.1959194    0.1988702  0.1976251
#>  [606,]    0.1940801    0.1970032  0.1957699
#>  [607,]    0.2096280    0.2127853  0.2114531
#>  [608,]    0.2061966    0.2093022  0.2079918
#>  [609,]    0.2160965    0.2193512  0.2179779
#>  [610,]    0.1918198    0.1947088  0.1934898
#>  [611,]    0.1708713    0.1734448  0.1723589
#>  [612,]    0.1537858    0.1561020  0.1551247
#>  [613,]    0.2040453    0.2071184  0.2058218
#>  [614,]    0.2407511    0.2443771  0.2428471
#>  [615,]    0.2238042    0.2271749  0.2257527
#>  [616,]    0.1638785    0.1663467  0.1653052
#>  [617,]    0.1739308    0.1765504  0.1754451
#>  [618,]    0.1962121    0.1991673  0.1979204
#>  [619,]    0.1958904    0.1988408  0.1975959
#>  [620,]    0.2064289    0.2095380  0.2082261
#>  [621,]    0.2176223    0.2208999  0.2195170
#>  [622,]    0.2288894    0.2323367  0.2308822
#>  [623,]    0.2143693    0.2175980  0.2162357
#>  [624,]    0.1974375    0.2004111  0.1991565
#>  [625,]    0.2086618    0.2118045  0.2104785
#>  [626,]    0.2326434    0.2361473  0.2346689
#>  [627,]    0.2671335    0.2711568  0.2694592
#>  [628,]    0.1766206    0.1792807  0.1781583
#>  [629,]    0.2164748    0.2197351  0.2183595
#>  [630,]    0.2737453    0.2778681  0.2761286
#>  [631,]    0.2097415    0.2129004  0.2115675
#>  [632,]    0.2674612    0.2714894  0.2697897
#>  [633,]    0.2680163    0.2720529  0.2703497
#>  [634,]    0.1971635    0.2001330  0.1988801
#>  [635,]    0.1746156    0.1772456  0.1761359
#>  [636,]    0.2046778    0.2077606  0.2064598
#>  [637,]    0.2161328    0.2193880  0.2180145
#>  [638,]    0.1974029    0.2003761  0.1991216
#>  [639,]    0.2438608    0.2475336  0.2459839
#>  [640,]    0.1623571    0.1648024  0.1637707
#>  [641,]    0.2943337    0.2987667  0.2968962
#>  [642,]    0.1695757    0.1721297  0.1710521
#>  [643,]    0.1922187    0.1951138  0.1938923
#>  [644,]    0.2897090    0.2940723  0.2922313
#>  [645,]    0.2629763    0.2669370  0.2652658
#>  [646,]    0.2809541    0.2851856  0.2834002
#>  [647,]    0.1295391    0.1314901  0.1306669
#>  [648,]    0.2325617    0.2360644  0.2345865
#>  [649,]    0.1599412    0.1623502  0.1613337
#>  [650,]    0.2217181    0.2250574  0.2236484
#>  [651,]    0.1680124    0.1705429  0.1694752
#>  [652,]    0.2297155    0.2331753  0.2317155
#>  [653,]    0.2089185    0.2120650  0.2107374
#>  [654,]    0.1913596    0.1942417  0.1930256
#>  [655,]    0.1951532    0.1980924  0.1968522
#>  [656,]    0.2419090    0.2455524  0.2440151
#>  [657,]    0.2031329    0.2061924  0.2049015
#>  [658,]    0.2224938    0.2258448  0.2244309
#>  [659,]    0.2007224    0.2037455  0.2024700
#>  [660,]    0.2146362    0.2178688  0.2165048
#>  [661,]    0.1831664    0.1859251  0.1847611
#>  [662,]    0.2676463    0.2716772  0.2699764
#>  [663,]    0.1740189    0.1766398  0.1755339
#>  [664,]    0.2111731    0.2143536  0.2130116
#>  [665,]    0.1624966    0.1649440  0.1639113
#>  [666,]    0.2203614    0.2236803  0.2222799
#>  [667,]    0.2435680    0.2472364  0.2456886
#>  [668,]    0.1677187    0.1702448  0.1691789
#>  [669,]    0.2254035    0.2287983  0.2273659
#>  [670,]    0.2707960    0.2748744  0.2731536
#>  [671,]    0.2257662    0.2291665  0.2277318
#>  [672,]    0.3173244    0.3221036  0.3200871
#>  [673,]    0.2249063    0.2282937  0.2268645
#>  [674,]    0.2835699    0.2878407  0.2860387
#>  [675,]    0.2414518    0.2450883  0.2435540
#>  [676,]    0.2083042    0.2114415  0.2101178
#>  [677,]    0.2232014    0.2265630  0.2251446
#>  [678,]    0.2427920    0.2464486  0.2449058
#>  [679,]    0.1911060    0.1939843  0.1927698
#>  [680,]    0.1983931    0.2013812  0.2001204
#>  [681,]    0.2304538    0.2339247  0.2324602
#>  [682,]    0.1900478    0.1929101  0.1917024
#>  [683,]    0.2092016    0.2123524  0.2110229
#>  [684,]    0.1764459    0.1791034  0.1779821
#>  [685,]    0.2163384    0.2195967  0.2182219
#>  [686,]    0.2390195    0.2426194  0.2411005
#>  [687,]    0.2141155    0.2173403  0.2159797
#>  [688,]    0.1891528    0.1920017  0.1907996
#>  [689,]    0.2465087    0.2502214  0.2486549
#>  [690,]    0.1830934    0.1858510  0.1846874
#>  [691,]    0.1569456    0.1593095  0.1583121
#>  [692,]    0.2346668    0.2382012  0.2367099
#>  [693,]    0.2494246    0.2531812  0.2515961
#>  [694,]    0.2672928    0.2713184  0.2696199
#>  [695,]    0.2039756    0.2070477  0.2057515
#>  [696,]    0.1853512    0.1881428  0.1869649
#>  [697,]    0.1678786    0.1704071  0.1693402
#>  [698,]    0.3079184    0.3125559  0.3105991
#>  [699,]    0.1879390    0.1907696  0.1895752
#>  [700,]    0.1952016    0.1981416  0.1969011
#>  [701,]    0.1818179    0.1845563  0.1834009
#>  [702,]    0.2425006    0.2461529  0.2446118
#>  [703,]    0.1721990    0.1747926  0.1736983
#>  [704,]    0.1868523    0.1896666  0.1884791
#>  [705,]    0.1695244    0.1720777  0.1710004
#>  [706,]    0.1338885    0.1359050  0.1350542
#>  [707,]    0.1853183    0.1881094  0.1869318
#>  [708,]    0.1934700    0.1963839  0.1951545
#>  [709,]    0.1725186    0.1751170  0.1740206
#>  [710,]    0.3065412    0.3111580  0.3092100
#>  [711,]    0.2090269    0.2121751  0.2108467
#>  [712,]    0.1879408    0.1907714  0.1895771
#>  [713,]    0.1997048    0.2027126  0.2014435
#>  [714,]    0.3002449    0.3047669  0.3028589
#>  [715,]    0.2423890    0.2460396  0.2444992
#>  [716,]    0.2995842    0.3040962  0.3021924
#>  [717,]    0.1993446    0.2023470  0.2010802
#>  [718,]    0.1881129    0.1909461  0.1897507
#>  [719,]    0.2065765    0.2096879  0.2083751
#>  [720,]    0.2622818    0.2662320  0.2645652
#>  [721,]    0.1798887    0.1825981  0.1814549
#>  [722,]    0.2038512    0.2069214  0.2056260
#>  [723,]    0.2250599    0.2284496  0.2270193
#>  [724,]    0.2397602    0.2433711  0.2418475
#>  [725,]    0.1922401    0.1951355  0.1939138
#>  [726,]    0.2390068    0.2426064  0.2410876
#>  [727,]    0.2666831    0.2706996  0.2690049
#>  [728,]    0.1928909    0.1957961  0.1945703
#>  [729,]    0.2061634    0.2092685  0.2079584
#>  [730,]    0.3537089    0.3590360  0.3567883
#>  [731,]    0.2505763    0.2543502  0.2527578
#>  [732,]    0.2688079    0.2728563  0.2711481
#>  [733,]    0.2169248    0.2201920  0.2188134
#>  [734,]    0.1801519    0.1828652  0.1817204
#>  [735,]    0.1839234    0.1866935  0.1855247
#>  [736,]    0.1955734    0.1985190  0.1972761
#>  [737,]    0.1481123    0.1503431  0.1494018
#>  [738,]    0.1810746    0.1838018  0.1826511
#>  [739,]    0.1881185    0.1909518  0.1897563
#>  [740,]    0.2030870    0.2061457  0.2048551
#>  [741,]    0.2004714    0.2034908  0.2022168
#>  [742,]    0.2251070    0.2284973  0.2270668
#>  [743,]    0.1675701    0.1700939  0.1690290
#>  [744,]    0.2494285    0.2531852  0.2516001
#>  [745,]    0.2175732    0.2208501  0.2194675
#>  [746,]    0.2001589    0.2031735  0.2019016
#>  [747,]    0.2900552    0.2944237  0.2925805
#>  [748,]    0.3014177    0.3059573  0.3040419
#>  [749,]    0.1800173    0.1827286  0.1815846
#>  [750,]    0.2531355    0.2569480  0.2553394
#>  [751,]    0.1869554    0.1897712  0.1885831
#>  [752,]    0.2349318    0.2384701  0.2369772
#>  [753,]    0.1566968    0.1590569  0.1580611
#>  [754,]    0.2246987    0.2280829  0.2266549
#>  [755,]    0.1498001    0.1520563  0.1511043
#>  [756,]    0.1655565    0.1680500  0.1669979
#>  [757,]    0.2235781    0.2269455  0.2255246
#>  [758,]    0.1775596    0.1802339  0.1791055
#>  [759,]    0.1882078    0.1910425  0.1898464
#>  [760,]    0.2716155    0.2757063  0.2739802
#>  [761,]    0.1693405    0.1718910  0.1708149
#>  [762,]    0.1846692    0.1874505  0.1862769
#>  [763,]    0.2451401    0.2488322  0.2472744
#>  [764,]    0.2361464    0.2397030  0.2382023
#>  [765,]    0.1917081    0.1945955  0.1933772
#>  [766,]    0.2356482    0.2391973  0.2376998
#>  [767,]    0.2344532    0.2379843  0.2364944
#>  [768,]    0.1694434    0.1719954  0.1709186
#>  [769,]    0.1857477    0.1885452  0.1873648
#>  [770,]    0.2418483    0.2454907  0.2439538
#>  [771,]    0.2048538    0.2079391  0.2066373
#>  [772,]    0.3205921    0.3254204  0.3233831
#>  [773,]    0.1826628    0.1854139  0.1842531
#>  [774,]    0.2123945    0.2155934  0.2142436
#>  [775,]    0.1880369    0.1908690  0.1896740
#>  [776,]    0.2188995    0.2221964  0.2208053
#>  [777,]    0.2204715    0.2237920  0.2223909
#>  [778,]    0.2252433    0.2286358  0.2272044
#>  [779,]    0.1654486    0.1679405  0.1668891
#>  [780,]    0.2491627    0.2529153  0.2513320
#>  [781,]    0.1957324    0.1986804  0.1974365
#>  [782,]    0.2055797    0.2086760  0.2073695
#>  [783,]    0.2353329    0.2388773  0.2373818
#>  [784,]    0.3153233    0.3200723  0.3180685
#>  [785,]    0.1803072    0.1830229  0.1818770
#>  [786,]    0.2069057    0.2100220  0.2087071
#>  [787,]    0.3265288    0.3314465  0.3293715
#>  [788,]    0.2218089    0.2251496  0.2237400
#>  [789,]    0.2196618    0.2229702  0.2215742
#>  [790,]    0.2784685    0.2826624  0.2808928
#>  [791,]    0.1722525    0.1748469  0.1737522
#>  [792,]    0.1886421    0.1914833  0.1902845
#>  [793,]    0.1550796    0.1574153  0.1564298
#>  [794,]    0.2724642    0.2765677  0.2748363
#>  [795,]    0.2206429    0.2239661  0.2225639
#>  [796,]    0.2991900    0.3036960  0.3017947
#>  [797,]    0.1475039    0.1497255  0.1487881
#>  [798,]    0.1858359    0.1886348  0.1874538
#>  [799,]    0.1980332    0.2010158  0.1997573
#>  [800,]    0.2293445    0.2327987  0.2313412
#>  [801,]    0.2258901    0.2292922  0.2278567
#>  [802,]    0.2176651    0.2209434  0.2195601
#>  [803,]    0.2520214    0.2558171  0.2542156
#>  [804,]    0.1901708    0.1930350  0.1918265
#>  [805,]    0.1945458    0.1974759  0.1962396
#>  [806,]    0.1770240    0.1796902  0.1785652
#>  [807,]    0.2052654    0.2083569  0.2070525
#>  [808,]    0.2103282    0.2134960  0.2121594
#>  [809,]    0.1665629    0.1690716  0.1680131
#>  [810,]    0.2254459    0.2288413  0.2274086
#>  [811,]    0.2405666    0.2441898  0.2426610
#>  [812,]    0.1978678    0.2008479  0.1995905
#>  [813,]    0.1903848    0.1932522  0.1920423
#>  [814,]    0.2032034    0.2062639  0.2049725
#>  [815,]    0.2572332    0.2611074  0.2594727
#>  [816,]    0.1410764    0.1432012  0.1423047
#>  [817,]    0.2193402    0.2226437  0.2212498
#>  [818,]    0.1735110    0.1761243  0.1750216
#>  [819,]    0.2539459    0.2577706  0.2561568
#>  [820,]    0.1941434    0.1970674  0.1958336
#>  [821,]    0.2052211    0.2083120  0.2070078
#>  [822,]    0.2202876    0.2236054  0.2222055
#>  [823,]    0.1744002    0.1770269  0.1759186
#>  [824,]    0.2486738    0.2524191  0.2508388
#>  [825,]    0.2701301    0.2741985  0.2724819
#>  [826,]    0.3273850    0.3323157  0.3302352
#>  [827,]    0.1871993    0.1900188  0.1888291
#>  [828,]    0.2244858    0.2278668  0.2264402
#>  [829,]    0.1888472    0.1916915  0.1904914
#>  [830,]    0.2206294    0.2239524  0.2225503
#>  [831,]    0.2377321    0.2413126  0.2398019
#>  [832,]    0.2468513    0.2505691  0.2490004
#>  [833,]    0.2001185    0.2031325  0.2018608
#>  [834,]    0.1826680    0.1854192  0.1842584
#>  [835,]    0.1461401    0.1483411  0.1474124
#>  [836,]    0.1561928    0.1585453  0.1575527
#>  [837,]    0.1560833    0.1584342  0.1574423
#>  [838,]    0.1841664    0.1869402  0.1857698
#>  [839,]    0.2238553    0.2272268  0.2258042
#>  [840,]    0.2139740    0.2171967  0.2158369
#>  [841,]    0.2023416    0.2053891  0.2041033
#>  [842,]    0.2477678    0.2514995  0.2499250
#>  [843,]    0.2181406    0.2214261  0.2200398
#>  [844,]    0.2761339    0.2802927  0.2785380
#>  [845,]    0.3006125    0.3051399  0.3032296
#>  [846,]    0.1647576    0.1672390  0.1661920
#>  [847,]    0.2132565    0.2164684  0.2151131
#>  [848,]    0.2089442    0.2120912  0.2107633
#>  [849,]    0.2591114    0.2630139  0.2613673
#>  [850,]    0.2494419    0.2531987  0.2516136
#>  [851,]    0.2094927    0.2126479  0.2113166
#>  [852,]    0.2091615    0.2123117  0.2109825
#>  [853,]    0.1834432    0.1862061  0.1850403
#>  [854,]    0.2186134    0.2219060  0.2205167
#>  [855,]    0.1939712    0.1968927  0.1956600
#>  [856,]    0.2095892    0.2127458  0.2114139
#>  [857,]    0.2920588    0.2964574  0.2946014
#>  [858,]    0.2244212    0.2278012  0.2263750
#>  [859,]    0.3259304    0.3308392  0.3287680
#>  [860,]    0.1341469    0.1361674  0.1353148
#>  [861,]    0.2065213    0.2096317  0.2083193
#>  [862,]    0.1963317    0.1992887  0.1980410
#>  [863,]    0.1572853    0.1596543  0.1586547
#>  [864,]    0.2576706    0.2615514  0.2599140
#>  [865,]    0.2885570    0.2929030  0.2910693
#>  [866,]    0.1612886    0.1637178  0.1626928
#>  [867,]    0.2745816    0.2787170  0.2769721
#>  [868,]    0.1874113    0.1902339  0.1890430
#>  [869,]    0.1872325    0.1900524  0.1888626
#>  [870,]    0.1887280    0.1915705  0.1903711
#>  [871,]    0.1677353    0.1702616  0.1691956
#>  [872,]    0.2132491    0.2164609  0.2151057
#>  [873,]    0.2219074    0.2252495  0.2238394
#>  [874,]    0.2282886    0.2317269  0.2302762
#>  [875,]    0.1485806    0.1508184  0.1498742
#>  [876,]    0.3066611    0.3112797  0.3093309
#>  [877,]    0.2087077    0.2118511  0.2105248
#>  [878,]    0.2331773    0.2366892  0.2352073
#>  [879,]    0.2045944    0.2076759  0.2063757
#>  [880,]    0.2578301    0.2617132  0.2600748
#>  [881,]    0.2463724    0.2500829  0.2485173
#>  [882,]    0.2297207    0.2331805  0.2317207
#>  [883,]    0.2343965    0.2379267  0.2364372
#>  [884,]    0.2318807    0.2353731  0.2338995
#>  [885,]    0.1865576    0.1893674  0.1881818
#>  [886,]    0.2232449    0.2266073  0.2251886
#>  [887,]    0.2226285    0.2259816  0.2245668
#>  [888,]    0.1994841    0.2024886  0.2012209
#>  [889,]    0.1981988    0.2011839  0.1999243
#>  [890,]    0.2104024    0.2135713  0.2122342
#>  [891,]    0.2383991    0.2419896  0.2404746
#>  [892,]    0.1818562    0.1845952  0.1834395
#>  [893,]    0.2170384    0.2203073  0.2189280
#>  [894,]    0.1736973    0.1763134  0.1752096
#>  [895,]    0.2666280    0.2706437  0.2689493
#>  [896,]    0.2850042    0.2892966  0.2874855
#>  [897,]    0.2452058    0.2488989  0.2473406
#>  [898,]    0.1485794    0.1508173  0.1498730
#>  [899,]    0.1768312    0.1794945  0.1783707
#>  [900,]    0.1685486    0.1710871  0.1700160
#>  [901,]    0.1837470    0.1865145  0.1853468
#>  [902,]    0.1762880    0.1789432  0.1778229
#>  [903,]    0.2284867    0.2319279  0.2304759
#>  [904,]    0.2347475    0.2382831  0.2367913
#>  [905,]    0.2514685    0.2552560  0.2536579
#>  [906,]    0.1783464    0.1810326  0.1798992
#>  [907,]    0.2696975    0.2737594  0.2720455
#>  [908,]    0.2646559    0.2686419  0.2669601
#>  [909,]    0.1781402    0.1808232  0.1796911
#>  [910,]    0.1783833    0.1810700  0.1799363
#>  [911,]    0.2127559    0.2159603  0.2146082
#>  [912,]    0.1984186    0.2014070  0.2001461
#>  [913,]    0.3394770    0.3445898  0.3424325
#>  [914,]    0.1659987    0.1684989  0.1674440
#>  [915,]    0.1355430    0.1375845  0.1367231
#>  [916,]    0.2355832    0.2391313  0.2376342
#>  [917,]    0.1829272    0.1856823  0.1845198
#>  [918,]    0.2006508    0.2036729  0.2023977
#>  [919,]    0.2260441    0.2294485  0.2280121
#>  [920,]    0.1797774    0.1824851  0.1813426
#>  [921,]    0.2260783    0.2294833  0.2280466
#>  [922,]    0.2298933    0.2333557  0.2318948
#>  [923,]    0.1809811    0.1837069  0.1825568
#>  [924,]    0.3252448    0.3301433  0.3280764
#>  [925,]    0.2449385    0.2486276  0.2470710
#>  [926,]    0.2647036    0.2686904  0.2670082
#>  [927,]    0.2173216    0.2205947  0.2192137
#>  [928,]    0.1947073    0.1976398  0.1964025
#>  [929,]    0.2492908    0.2530454  0.2514612
#>  [930,]    0.2653883    0.2693853  0.2676988
#>  [931,]    0.2204132    0.2237329  0.2223322
#>  [932,]    0.2257799    0.2291804  0.2277456
#>  [933,]    0.2099219    0.2130836  0.2117496
#>  [934,]    0.1798938    0.1826033  0.1814600
#>  [935,]    0.2201500    0.2234658  0.2220667
#>  [936,]    0.3522377    0.3575425  0.3553042
#>  [937,]    0.2055433    0.2086390  0.2073328
#>  [938,]    0.1961382    0.1990923  0.1978458
#>  [939,]    0.1895076    0.1923618  0.1911575
#>  [940,]    0.2163094    0.2195672  0.2181926
#>  [941,]    0.2258598    0.2292615  0.2278262
#>  [942,]    0.2018606    0.2049008  0.2036180
#>  [943,]    0.2096278    0.2127851  0.2114529
#>  [944,]    0.2007993    0.2038235  0.2025475
#>  [945,]    0.1999595    0.2029711  0.2017004
#>  [946,]    0.2231293    0.2264898  0.2250719
#>  [947,]    0.2282782    0.2317164  0.2302657
#>  [948,]    0.2127328    0.2159367  0.2145849
#>  [949,]    0.2508314    0.2546092  0.2530152
#>  [950,]    0.1873397    0.1901613  0.1889708
#>  [951,]    0.1706573    0.1732276  0.1721431
#>  [952,]    0.1786739    0.1813649  0.1802294
#>  [953,]    0.1869408    0.1897563  0.1885683
#>  [954,]    0.2086499    0.2117924  0.2104664
#>  [955,]    0.3078196    0.3124556  0.3104995
#>  [956,]    0.2248672    0.2282540  0.2268250
#>  [957,]    0.2053879    0.2084813  0.2071760
#>  [958,]    0.2106734    0.2138464  0.2125076
#>  [959,]    0.1569957    0.1593602  0.1583625
#>  [960,]    0.2205518    0.2238735  0.2224719
#>  [961,]    0.2602129    0.2641320  0.2624784
#>  [962,]    0.2147274    0.2179614  0.2165969
#>  [963,]    0.1940791    0.1970022  0.1957688
#>  [964,]    0.2650105    0.2690018  0.2673177
#>  [965,]    0.1997856    0.2027946  0.2015250
#>  [966,]    0.2796110    0.2838222  0.2820453
#>  [967,]    0.2219401    0.2252828  0.2238724
#>  [968,]    0.3273892    0.3323198  0.3302394
#>  [969,]    0.2364526    0.2400139  0.2385112
#>  [970,]    0.2727438    0.2768516  0.2751184
#>  [971,]    0.1733682    0.1759793  0.1748776
#>  [972,]    0.2673740    0.2714009  0.2697018
#>  [973,]    0.1847114    0.1874934  0.1863196
#>  [974,]    0.1925014    0.1954007  0.1941774
#>  [975,]    0.2928310    0.2972413  0.2953805
#>  [976,]    0.1917104    0.1945978  0.1933795
#>  [977,]    0.2210825    0.2244123  0.2230073
#>  [978,]    0.2379010    0.2414840  0.2399722
#>  [979,]    0.2076734    0.2108012  0.2094815
#>  [980,]    0.1794074    0.1821095  0.1809694
#>  [981,]    0.1830532    0.1858102  0.1846469
#>  [982,]    0.2505988    0.2543731  0.2527806
#>  [983,]    0.3080963    0.3127365  0.3107786
#>  [984,]    0.2228373    0.2261934  0.2247774
#>  [985,]    0.2753530    0.2795000  0.2777502
#>  [986,]    0.1797527    0.1824600  0.1813177
#>  [987,]    0.3209433    0.3257769  0.3237375
#>  [988,]    0.1815121    0.1842459  0.1830924
#>  [989,]    0.1626785    0.1651286  0.1640948
#>  [990,]    0.2071203    0.2102398  0.2089235
#>  [991,]    0.2071476    0.2102675  0.2089511
#>  [992,]    0.1801032    0.1828158  0.1816712
#>  [993,]    0.2956188    0.3000710  0.2981924
#>  [994,]    0.3339832    0.3390132  0.3368909
#>  [995,]    0.2246055    0.2279883  0.2265610
#>  [996,]    0.2212491    0.2245813  0.2231753
#>  [997,]    0.1887613    0.1916042  0.1904047
#>  [998,]    0.2151763    0.2184170  0.2170496
#>  [999,]    0.2554191    0.2592659  0.2576428
#> [1000,]    0.1969858    0.1999526  0.1987008